% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_operations.R
\name{athena_batch_get_prepared_statement}
\alias{athena_batch_get_prepared_statement}
\title{Returns the details of a single prepared statement or a list of up to
256 prepared statements for the array of prepared statement names that
you provide}
\usage{
athena_batch_get_prepared_statement(PreparedStatementNames, WorkGroup)
}
\arguments{
\item{PreparedStatementNames}{[required] A list of prepared statement names to return.}

\item{WorkGroup}{[required] The name of the workgroup to which the prepared statements belong.}
}
\description{
Returns the details of a single prepared statement or a list of up to 256 prepared statements for the array of prepared statement names that you provide. Requires you to have access to the workgroup to which the prepared statements belong. If a prepared statement cannot be retrieved for the name specified, the statement is listed in \code{UnprocessedPreparedStatementNames}.

See \url{https://www.paws-r-sdk.com/docs/athena_batch_get_prepared_statement/} for full documentation.
}
\keyword{internal}
