% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_operations.R
\name{athena_cancel_capacity_reservation}
\alias{athena_cancel_capacity_reservation}
\title{Cancels the capacity reservation with the specified name}
\usage{
athena_cancel_capacity_reservation(Name)
}
\arguments{
\item{Name}{[required] The name of the capacity reservation to cancel.}
}
\description{
Cancels the capacity reservation with the specified name.

See \url{https://paws-r.github.io/docs/athena/cancel_capacity_reservation.html} for full documentation.
}
\keyword{internal}
