% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_operations.R
\name{athena_create_capacity_reservation}
\alias{athena_create_capacity_reservation}
\title{Creates a capacity reservation with the specified name and number of
requested data processing units}
\usage{
athena_create_capacity_reservation(TargetDpus, Name, Tags = NULL)
}
\arguments{
\item{TargetDpus}{[required] The number of requested data processing units.}

\item{Name}{[required] The name of the capacity reservation to create.}

\item{Tags}{The tags for the capacity reservation.}
}
\description{
Creates a capacity reservation with the specified name and number of requested data processing units.

See \url{https://www.paws-r-sdk.com/docs/athena_create_capacity_reservation/} for full documentation.
}
\keyword{internal}
