% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_operations.R
\name{athena_create_prepared_statement}
\alias{athena_create_prepared_statement}
\title{Creates a prepared statement for use with SQL queries in Athena}
\usage{
athena_create_prepared_statement(StatementName, WorkGroup,
  QueryStatement, Description)
}
\arguments{
\item{StatementName}{[required] The name of the prepared statement.}

\item{WorkGroup}{[required] The name of the workgroup to which the prepared statement belongs.}

\item{QueryStatement}{[required] The query string for the prepared statement.}

\item{Description}{The description of the prepared statement.}
}
\value{
An empty list.
}
\description{
Creates a prepared statement for use with SQL queries in Athena.
}
\section{Request syntax}{
\preformatted{svc$create_prepared_statement(
  StatementName = "string",
  WorkGroup = "string",
  QueryStatement = "string",
  Description = "string"
)
}
}

\keyword{internal}
