% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_operations.R
\name{athena_delete_data_catalog}
\alias{athena_delete_data_catalog}
\title{Deletes a data catalog}
\usage{
athena_delete_data_catalog(Name)
}
\arguments{
\item{Name}{[required] The name of the data catalog to delete.}
}
\description{
Deletes a data catalog.

See \url{https://paws-r.github.io/docs/athena/delete_data_catalog.html} for full documentation.
}
\keyword{internal}
