% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_operations.R
\name{athena_delete_named_query}
\alias{athena_delete_named_query}
\title{Deletes the named query if you have access to the workgroup in which the
query was saved}
\usage{
athena_delete_named_query(NamedQueryId)
}
\arguments{
\item{NamedQueryId}{[required] The unique ID of the query to delete.}
}
\description{
Deletes the named query if you have access to the workgroup in which the query was saved.

See \url{https://paws-r.github.io/docs/athena/delete_named_query.html} for full documentation.
}
\keyword{internal}
