% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_operations.R
\name{athena_delete_notebook}
\alias{athena_delete_notebook}
\title{Deletes the specified notebook}
\usage{
athena_delete_notebook(NotebookId)
}
\arguments{
\item{NotebookId}{[required] The ID of the notebook to delete.}
}
\description{
Deletes the specified notebook.

See \url{https://paws-r.github.io/docs/athena/delete_notebook.html} for full documentation.
}
\keyword{internal}
