% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_operations.R
\name{athena_delete_work_group}
\alias{athena_delete_work_group}
\title{Deletes the workgroup with the specified name}
\usage{
athena_delete_work_group(WorkGroup, RecursiveDeleteOption = NULL)
}
\arguments{
\item{WorkGroup}{[required] The unique name of the workgroup to delete.}

\item{RecursiveDeleteOption}{The option to delete the workgroup and its contents even if the
workgroup contains any named queries, query executions, or notebooks.}
}
\description{
Deletes the workgroup with the specified name. The primary workgroup cannot be deleted.

See \url{https://www.paws-r-sdk.com/docs/athena_delete_work_group/} for full documentation.
}
\keyword{internal}
