% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_operations.R
\name{athena_export_notebook}
\alias{athena_export_notebook}
\title{Exports the specified notebook and its metadata}
\usage{
athena_export_notebook(NotebookId)
}
\arguments{
\item{NotebookId}{[required] The ID of the notebook to export.}
}
\description{
Exports the specified notebook and its metadata.

See \url{https://paws-r.github.io/docs/athena/export_notebook.html} for full documentation.
}
\keyword{internal}
