% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_operations.R
\name{athena_get_calculation_execution}
\alias{athena_get_calculation_execution}
\title{Describes a previously submitted calculation execution}
\usage{
athena_get_calculation_execution(CalculationExecutionId)
}
\arguments{
\item{CalculationExecutionId}{[required] The calculation execution UUID.}
}
\description{
Describes a previously submitted calculation execution.

See \url{https://www.paws-r-sdk.com/docs/athena_get_calculation_execution/} for full documentation.
}
\keyword{internal}
