% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_operations.R
\name{athena_get_calculation_execution_code}
\alias{athena_get_calculation_execution_code}
\title{Retrieves the unencrypted code that was executed for the calculation}
\usage{
athena_get_calculation_execution_code(CalculationExecutionId)
}
\arguments{
\item{CalculationExecutionId}{[required] The calculation execution UUID.}
}
\description{
Retrieves the unencrypted code that was executed for the calculation.

See \url{https://paws-r.github.io/docs/athena/get_calculation_execution_code.html} for full documentation.
}
\keyword{internal}
