% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_operations.R
\name{athena_get_capacity_assignment_configuration}
\alias{athena_get_capacity_assignment_configuration}
\title{Gets the capacity assignment configuration for a capacity reservation,
if one exists}
\usage{
athena_get_capacity_assignment_configuration(CapacityReservationName)
}
\arguments{
\item{CapacityReservationName}{[required] The name of the capacity reservation to retrieve the capacity assignment
configuration for.}
}
\description{
Gets the capacity assignment configuration for a capacity reservation, if one exists.

See \url{https://www.paws-r-sdk.com/docs/athena_get_capacity_assignment_configuration/} for full documentation.
}
\keyword{internal}
