% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_operations.R
\name{athena_get_capacity_reservation}
\alias{athena_get_capacity_reservation}
\title{Returns information about the capacity reservation with the specified
name}
\usage{
athena_get_capacity_reservation(Name)
}
\arguments{
\item{Name}{[required] The name of the capacity reservation.}
}
\description{
Returns information about the capacity reservation with the specified name.

See \url{https://www.paws-r-sdk.com/docs/athena_get_capacity_reservation/} for full documentation.
}
\keyword{internal}
