% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_operations.R
\name{athena_get_named_query}
\alias{athena_get_named_query}
\title{Returns information about a single query}
\usage{
athena_get_named_query(NamedQueryId)
}
\arguments{
\item{NamedQueryId}{[required] The unique ID of the query. Use
\code{\link[=athena_list_named_queries]{list_named_queries}} to get query IDs.}
}
\description{
Returns information about a single query. Requires that you have access to the workgroup in which the query was saved.

See \url{https://paws-r.github.io/docs/athena/get_named_query.html} for full documentation.
}
\keyword{internal}
