% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_operations.R
\name{athena_get_notebook_metadata}
\alias{athena_get_notebook_metadata}
\title{Retrieves notebook metadata for the specified notebook ID}
\usage{
athena_get_notebook_metadata(NotebookId)
}
\arguments{
\item{NotebookId}{[required] The ID of the notebook whose metadata is to be retrieved.}
}
\description{
Retrieves notebook metadata for the specified notebook ID.

See \url{https://paws-r.github.io/docs/athena/get_notebook_metadata.html} for full documentation.
}
\keyword{internal}
