% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_operations.R
\name{athena_get_prepared_statement}
\alias{athena_get_prepared_statement}
\title{Retrieves the prepared statement with the specified name from the
specified workgroup}
\usage{
athena_get_prepared_statement(StatementName, WorkGroup)
}
\arguments{
\item{StatementName}{[required] The name of the prepared statement to retrieve.}

\item{WorkGroup}{[required] The workgroup to which the statement to be retrieved belongs.}
}
\description{
Retrieves the prepared statement with the specified name from the specified workgroup.

See \url{https://paws-r.github.io/docs/athena/get_prepared_statement.html} for full documentation.
}
\keyword{internal}
