% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_operations.R
\name{athena_get_query_runtime_statistics}
\alias{athena_get_query_runtime_statistics}
\title{Returns query execution runtime statistics related to a single execution
of a query if you have access to the workgroup in which the query ran}
\usage{
athena_get_query_runtime_statistics(QueryExecutionId)
}
\arguments{
\item{QueryExecutionId}{[required] The unique ID of the query execution.}
}
\description{
Returns query execution runtime statistics related to a single execution of a query if you have access to the workgroup in which the query ran. Query execution runtime statistics are returned only when QueryExecutionStatus$State is in a SUCCEEDED or FAILED state. Stage-level input and output row count and data size statistics are not shown when a query has row-level filters defined in Lake Formation.

See \url{https://www.paws-r-sdk.com/docs/athena_get_query_runtime_statistics/} for full documentation.
}
\keyword{internal}
