% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_operations.R
\name{athena_get_resource_dashboard}
\alias{athena_get_resource_dashboard}
\title{Gets the Live UI/Persistence UI for a session}
\usage{
athena_get_resource_dashboard(ResourceARN)
}
\arguments{
\item{ResourceARN}{[required] The The Amazon Resource Name (ARN) for a session.}
}
\description{
Gets the Live UI/Persistence UI for a session.

See \url{https://www.paws-r-sdk.com/docs/athena_get_resource_dashboard/} for full documentation.
}
\keyword{internal}
