% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_operations.R
\name{athena_get_session}
\alias{athena_get_session}
\title{Gets the full details of a previously created session, including the
session status and configuration}
\usage{
athena_get_session(SessionId)
}
\arguments{
\item{SessionId}{[required] The session ID.}
}
\description{
Gets the full details of a previously created session, including the session status and configuration.

See \url{https://www.paws-r-sdk.com/docs/athena_get_session/} for full documentation.
}
\keyword{internal}
