% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_operations.R
\name{athena_get_session_endpoint}
\alias{athena_get_session_endpoint}
\title{Gets a connection endpoint and authentication token for a given session
Id}
\usage{
athena_get_session_endpoint(SessionId)
}
\arguments{
\item{SessionId}{[required] The session ID.}
}
\description{
Gets a connection endpoint and authentication token for a given session Id.

See \url{https://www.paws-r-sdk.com/docs/athena_get_session_endpoint/} for full documentation.
}
\keyword{internal}
