% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_operations.R
\name{athena_get_session_status}
\alias{athena_get_session_status}
\title{Gets the current status of a session}
\usage{
athena_get_session_status(SessionId)
}
\arguments{
\item{SessionId}{[required] The session ID.}
}
\description{
Gets the current status of a session.

See \url{https://paws-r.github.io/docs/athena/get_session_status.html} for full documentation.
}
\keyword{internal}
