% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_operations.R
\name{athena_list_application_dpu_sizes}
\alias{athena_list_application_dpu_sizes}
\title{Returns the supported DPU sizes for the supported application runtimes
(for example, Athena notebook version 1)}
\usage{
athena_list_application_dpu_sizes(MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{Specifies the maximum number of results to return.}

\item{NextToken}{A token generated by the Athena service that specifies where to continue
pagination if a previous request was truncated.}
}
\description{
Returns the supported DPU sizes for the supported application runtimes (for example, \verb{Athena notebook version 1}).

See \url{https://www.paws-r-sdk.com/docs/athena_list_application_dpu_sizes/} for full documentation.
}
\keyword{internal}
