% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_operations.R
\name{athena_list_capacity_reservations}
\alias{athena_list_capacity_reservations}
\title{Lists the capacity reservations for the current account}
\usage{
athena_list_capacity_reservations(NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{NextToken}{A token generated by the Athena service that specifies where to continue
pagination if a previous request was truncated.}

\item{MaxResults}{Specifies the maximum number of results to return.}
}
\description{
Lists the capacity reservations for the current account.

See \url{https://www.paws-r-sdk.com/docs/athena_list_capacity_reservations/} for full documentation.
}
\keyword{internal}
