% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_operations.R
\name{athena_list_data_catalogs}
\alias{athena_list_data_catalogs}
\title{Lists the data catalogs in the current Amazon Web Services account}
\usage{
athena_list_data_catalogs(NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{NextToken}{A token generated by the Athena service that specifies where to continue
pagination if a previous request was truncated. To obtain the next set
of pages, pass in the NextToken from the response object of the previous
page call.}

\item{MaxResults}{Specifies the maximum number of data catalogs to return.}
}
\description{
Lists the data catalogs in the current Amazon Web Services account.

See \url{https://www.paws-r-sdk.com/docs/athena_list_data_catalogs/} for full documentation.
}
\keyword{internal}
