% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_operations.R
\name{athena_list_engine_versions}
\alias{athena_list_engine_versions}
\title{Returns a list of engine versions that are available to choose from,
including the Auto option}
\usage{
athena_list_engine_versions(NextToken, MaxResults)
}
\arguments{
\item{NextToken}{A token generated by the Athena service that specifies where to continue
pagination if a previous request was truncated. To obtain the next set
of pages, pass in the \code{NextToken} from the response object of the
previous page call.}

\item{MaxResults}{The maximum number of engine versions to return in this request.}
}
\value{
A list with the following syntax:\preformatted{list(
  EngineVersions = list(
    list(
      SelectedEngineVersion = "string",
      EffectiveEngineVersion = "string"
    )
  ),
  NextToken = "string"
)
}
}
\description{
Returns a list of engine versions that are available to choose from,
including the Auto option.
}
\section{Request syntax}{
\preformatted{svc$list_engine_versions(
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
