% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_operations.R
\name{athena_list_notebook_metadata}
\alias{athena_list_notebook_metadata}
\title{Displays the notebook files for the specified workgroup in paginated
format}
\usage{
athena_list_notebook_metadata(
  Filters = NULL,
  NextToken = NULL,
  MaxResults = NULL,
  WorkGroup
)
}
\arguments{
\item{Filters}{Search filter string.}

\item{NextToken}{A token generated by the Athena service that specifies where to continue
pagination if a previous request was truncated.}

\item{MaxResults}{Specifies the maximum number of results to return.}

\item{WorkGroup}{[required] The name of the Spark enabled workgroup to retrieve notebook metadata
for.}
}
\description{
Displays the notebook files for the specified workgroup in paginated format.

See \url{https://paws-r.github.io/docs/athena/list_notebook_metadata.html} for full documentation.
}
\keyword{internal}
