% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_operations.R
\name{athena_put_capacity_assignment_configuration}
\alias{athena_put_capacity_assignment_configuration}
\title{Puts a new capacity assignment configuration for a specified capacity
reservation}
\usage{
athena_put_capacity_assignment_configuration(
  CapacityReservationName,
  CapacityAssignments
)
}
\arguments{
\item{CapacityReservationName}{[required] The name of the capacity reservation to put a capacity assignment
configuration for.}

\item{CapacityAssignments}{[required] The list of assignments for the capacity assignment configuration.}
}
\description{
Puts a new capacity assignment configuration for a specified capacity reservation. If a capacity assignment configuration already exists for the capacity reservation, replaces the existing capacity assignment configuration.

See \url{https://www.paws-r-sdk.com/docs/athena_put_capacity_assignment_configuration/} for full documentation.
}
\keyword{internal}
