% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_operations.R
\name{athena_stop_query_execution}
\alias{athena_stop_query_execution}
\title{Stops a query execution}
\usage{
athena_stop_query_execution(QueryExecutionId)
}
\arguments{
\item{QueryExecutionId}{[required] The unique ID of the query execution to stop.}
}
\description{
Stops a query execution. Requires you to have access to the workgroup in which the query ran.

See \url{https://www.paws-r-sdk.com/docs/athena_stop_query_execution/} for full documentation.
}
\keyword{internal}
