% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_operations.R
\name{athena_terminate_session}
\alias{athena_terminate_session}
\title{Terminates an active session}
\usage{
athena_terminate_session(SessionId)
}
\arguments{
\item{SessionId}{[required] The session ID.}
}
\description{
Terminates an active session. A \code{\link[=athena_terminate_session]{terminate_session}} call on a session that is already inactive (for example, in a \code{FAILED}, \code{TERMINATED} or \code{TERMINATING} state) succeeds but has no effect. Calculations running in the session when \code{\link[=athena_terminate_session]{terminate_session}} is called are forcefully stopped, but may display as \code{FAILED} instead of \code{STOPPED}.

See \url{https://www.paws-r-sdk.com/docs/athena_terminate_session/} for full documentation.
}
\keyword{internal}
