% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_operations.R
\name{athena_update_capacity_reservation}
\alias{athena_update_capacity_reservation}
\title{Updates the number of requested data processing units for the capacity
reservation with the specified name}
\usage{
athena_update_capacity_reservation(TargetDpus, Name)
}
\arguments{
\item{TargetDpus}{[required] The new number of requested data processing units.}

\item{Name}{[required] The name of the capacity reservation.}
}
\description{
Updates the number of requested data processing units for the capacity reservation with the specified name.

See \url{https://paws-r.github.io/docs/athena/update_capacity_reservation.html} for full documentation.
}
\keyword{internal}
