% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_operations.R
\name{athena_update_notebook_metadata}
\alias{athena_update_notebook_metadata}
\title{Updates the metadata for a notebook}
\usage{
athena_update_notebook_metadata(NotebookId, ClientRequestToken = NULL, Name)
}
\arguments{
\item{NotebookId}{[required] The ID of the notebook to update the metadata for.}

\item{ClientRequestToken}{A unique case-sensitive string used to ensure the request to create the
notebook is idempotent (executes only once).

This token is listed as not required because Amazon Web Services SDKs
(for example the Amazon Web Services SDK for Java) auto-generate the
token for you. If you are not using the Amazon Web Services SDK or the
Amazon Web Services CLI, you must provide this token or the action will
fail.}

\item{Name}{[required] The name to update the notebook to.}
}
\description{
Updates the metadata for a notebook.

See \url{https://www.paws-r-sdk.com/docs/athena_update_notebook_metadata/} for full documentation.
}
\keyword{internal}
