% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_operations.R
\name{athena_update_work_group}
\alias{athena_update_work_group}
\title{Updates the workgroup with the specified name}
\usage{
athena_update_work_group(
  WorkGroup,
  Description = NULL,
  ConfigurationUpdates = NULL,
  State = NULL
)
}
\arguments{
\item{WorkGroup}{[required] The specified workgroup that will be updated.}

\item{Description}{The workgroup description.}

\item{ConfigurationUpdates}{The workgroup configuration that will be updated for the given
workgroup.}

\item{State}{The workgroup state that will be updated for the given workgroup.}
}
\description{
Updates the workgroup with the specified name. The workgroup's name cannot be changed.

See \url{https://paws-r.github.io/docs/athena/update_work_group.html} for full documentation.
}
\keyword{internal}
