% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudsearch_operations.R
\name{cloudsearch_create_domain}
\alias{cloudsearch_create_domain}
\title{Creates a new search domain}
\usage{
cloudsearch_create_domain(DomainName)
}
\arguments{
\item{DomainName}{[required] A name for the domain you are creating. Allowed characters are a-z
(lower-case letters), 0-9, and hyphen (-). Domain names must start with
a letter or number and be at least 3 and no more than 28 characters
long.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  DomainStatus = list(
    DomainId = "string",
    DomainName = "string",
    ARN = "string",
    Created = TRUE|FALSE,
    Deleted = TRUE|FALSE,
    DocService = list(
      Endpoint = "string"
    ),
    SearchService = list(
      Endpoint = "string"
    ),
    RequiresIndexDocuments = TRUE|FALSE,
    Processing = TRUE|FALSE,
    SearchInstanceType = "string",
    SearchPartitionCount = 123,
    SearchInstanceCount = 123,
    Limits = list(
      MaximumReplicationCount = 123,
      MaximumPartitionCount = 123
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Creates a new search domain. For more information, see <a
href="https://docs.aws.amazon.com/cloudsearch/latest/developerguide/creating-domains.html"
target="_blank">Creating a Search Domain</a> in the \emph{Amazon CloudSearch
Developer Guide}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_domain(
  DomainName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
