% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudsearch_operations.R
\name{cloudsearch_define_suggester}
\alias{cloudsearch_define_suggester}
\title{Configures a suggester for a domain}
\usage{
cloudsearch_define_suggester(DomainName, Suggester)
}
\arguments{
\item{DomainName}{[required] A string that represents the name of a domain. Domain names are unique across the domains owned by an account within an AWS region. Domain names start with a letter or number and can contain the following characters: a-z (lowercase), 0-9, and - (hyphen).}

\item{Suggester}{[required] Configuration information for a search suggester. Each suggester has a unique name and specifies the text field you want to use for suggestions. The following options can be configured for a suggester: \code{FuzzyMatching}, \code{SortExpression}.}
}
\description{
Configures a suggester for a domain. A suggester enables you to display possible matches before users finish typing their queries. When you configure a suggester, you must specify the name of the text field you want to search for possible matches and a unique name for the suggester. For more information, see \if{html}{\out{<a href="https://docs.aws.amazon.com/cloudsearch/latest/developerguide/getting-suggestions.html" target="_blank">}}Getting Search Suggestions\if{html}{\out{</a>}} in the \emph{Amazon CloudSearch Developer Guide}.

See \url{https://www.paws-r-sdk.com/docs/cloudsearch_define_suggester/} for full documentation.
}
\keyword{internal}
