% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudsearch_operations.R
\name{cloudsearch_delete_analysis_scheme}
\alias{cloudsearch_delete_analysis_scheme}
\title{Deletes an analysis scheme}
\usage{
cloudsearch_delete_analysis_scheme(DomainName, AnalysisSchemeName)
}
\arguments{
\item{DomainName}{[required] A string that represents the name of a domain. Domain names are unique across the domains owned by an account within an AWS region. Domain names start with a letter or number and can contain the following characters: a-z (lowercase), 0-9, and - (hyphen).}

\item{AnalysisSchemeName}{[required] The name of the analysis scheme you want to delete.}
}
\description{
Deletes an analysis scheme. For more information, see \if{html}{\out{<a href="https://docs.aws.amazon.com/cloudsearch/latest/developerguide/configuring-analysis-schemes.html" target="_blank">}}Configuring Analysis Schemes\if{html}{\out{</a>}} in the \emph{Amazon CloudSearch Developer Guide}.

See \url{https://www.paws-r-sdk.com/docs/cloudsearch_delete_analysis_scheme/} for full documentation.
}
\keyword{internal}
