% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudsearch_operations.R
\name{cloudsearch_delete_expression}
\alias{cloudsearch_delete_expression}
\title{Removes an Expression from the search domain}
\usage{
cloudsearch_delete_expression(DomainName, ExpressionName)
}
\arguments{
\item{DomainName}{[required]}

\item{ExpressionName}{[required] The name of the \code{Expression} to delete.}
}
\description{
Removes an \code{Expression} from the search domain. For more information, see \if{html}{\out{<a href="https://docs.aws.amazon.com/cloudsearch/latest/developerguide/configuring-expressions.html" target="_blank">}}Configuring Expressions\if{html}{\out{</a>}} in the \emph{Amazon CloudSearch Developer Guide}.

See \url{https://www.paws-r-sdk.com/docs/cloudsearch_delete_expression/} for full documentation.
}
\keyword{internal}
