% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudsearch_operations.R
\name{cloudsearch_delete_index_field}
\alias{cloudsearch_delete_index_field}
\title{Removes an IndexField from the search domain}
\usage{
cloudsearch_delete_index_field(DomainName, IndexFieldName)
}
\arguments{
\item{DomainName}{[required]}

\item{IndexFieldName}{[required] The name of the index field your want to remove from the domain's
indexing options.}
}
\description{
Removes an \code{IndexField} from the search domain. For more information, see \if{html}{\out{<a href="https://docs.aws.amazon.com/cloudsearch/latest/developerguide/configuring-index-fields.html" target="_blank">}}Configuring Index Fields\if{html}{\out{</a>}} in the \emph{Amazon CloudSearch Developer Guide}.

See \url{https://paws-r.github.io/docs/cloudsearch/delete_index_field.html} for full documentation.
}
\keyword{internal}
