% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudsearch_operations.R
\name{cloudsearch_delete_suggester}
\alias{cloudsearch_delete_suggester}
\title{Deletes a suggester}
\usage{
cloudsearch_delete_suggester(DomainName, SuggesterName)
}
\arguments{
\item{DomainName}{[required] A string that represents the name of a domain. Domain names are unique across the domains owned by an account within an AWS region. Domain names start with a letter or number and can contain the following characters: a-z (lowercase), 0-9, and - (hyphen).}

\item{SuggesterName}{[required] Specifies the name of the suggester you want to delete.}
}
\description{
Deletes a suggester. For more information, see \if{html}{\out{<a href="https://docs.aws.amazon.com/cloudsearch/latest/developerguide/getting-suggestions.html" target="_blank">}}Getting Search Suggestions\if{html}{\out{</a>}} in the \emph{Amazon CloudSearch Developer Guide}.

See \url{https://www.paws-r-sdk.com/docs/cloudsearch_delete_suggester/} for full documentation.
}
\keyword{internal}
