% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudsearch_operations.R
\name{cloudsearch_describe_expressions}
\alias{cloudsearch_describe_expressions}
\title{Gets the expressions configured for the search domain}
\usage{
cloudsearch_describe_expressions(DomainName, ExpressionNames, Deployed)
}
\arguments{
\item{DomainName}{[required] The name of the domain you want to describe.}

\item{ExpressionNames}{Limits the \code{\link[=cloudsearch_describe_expressions]{describe_expressions}}
response to the specified expressions. If not specified, all expressions
are shown.}

\item{Deployed}{Whether to display the deployed configuration (\code{true}) or include any
pending changes (\code{false}). Defaults to \code{false}.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Expressions = list(
    list(
      Options = list(
        ExpressionName = "string",
        ExpressionValue = "string"
      ),
      Status = list(
        CreationDate = as.POSIXct(
          "2015-01-01"
        ),
        UpdateDate = as.POSIXct(
          "2015-01-01"
        ),
        UpdateVersion = 123,
        State = "RequiresIndexDocuments"|"Processing"|"Active"|"FailedToValidate",
        PendingDeletion = TRUE|FALSE
      )
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Gets the expressions configured for the search domain. Can be limited to
specific expressions by name. By default, shows all expressions and
includes any pending changes to the configuration. Set the \code{Deployed}
option to \code{true} to show the active configuration and exclude pending
changes. For more information, see <a
href="https://docs.aws.amazon.com/cloudsearch/latest/developerguide/configuring-expressions.html"
target="_blank">Configuring Expressions</a> in the \emph{Amazon CloudSearch
Developer Guide}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_expressions(
  DomainName = "string",
  ExpressionNames = list(
    "string"
  ),
  Deployed = TRUE|FALSE
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
