% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudsearch_operations.R
\name{cloudsearch_describe_scaling_parameters}
\alias{cloudsearch_describe_scaling_parameters}
\title{Gets the scaling parameters configured for a domain}
\usage{
cloudsearch_describe_scaling_parameters(DomainName)
}
\arguments{
\item{DomainName}{[required]}
}
\description{
Gets the scaling parameters configured for a domain. A domain's scaling parameters specify the desired search instance type and replication count. For more information, see \if{html}{\out{<a href="https://docs.aws.amazon.com/cloudsearch/latest/developerguide/configuring-scaling-options.html" target="_blank">}}Configuring Scaling Options\if{html}{\out{</a>}} in the \emph{Amazon CloudSearch Developer Guide}.

See \url{https://www.paws-r-sdk.com/docs/cloudsearch_describe_scaling_parameters/} for full documentation.
}
\keyword{internal}
