% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudsearch_operations.R
\name{cloudsearch_index_documents}
\alias{cloudsearch_index_documents}
\title{Tells the search domain to start indexing its documents using the latest
indexing options}
\usage{
cloudsearch_index_documents(DomainName)
}
\arguments{
\item{DomainName}{[required] A string that represents the name of a domain. Domain names are unique across the domains owned by an account within an AWS region. Domain names start with a letter or number and can contain the following characters: a-z (lowercase), 0-9, and - (hyphen).}
}
\description{
Tells the search domain to start indexing its documents using the latest indexing options. This operation must be invoked to activate options whose OptionStatus is \code{RequiresIndexDocuments}.

See \url{https://www.paws-r-sdk.com/docs/cloudsearch_index_documents/} for full documentation.
}
\keyword{internal}
