% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudsearch_operations.R
\name{cloudsearch_list_domain_names}
\alias{cloudsearch_list_domain_names}
\title{Lists all search domains owned by an account}
\usage{
cloudsearch_list_domain_names()
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  DomainNames = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Lists all search domains owned by an account.
}
\keyword{internal}
