% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapipeline_operations.R
\name{datapipeline_add_tags}
\alias{datapipeline_add_tags}
\title{Adds or modifies tags for the specified pipeline}
\usage{
datapipeline_add_tags(pipelineId, tags)
}
\arguments{
\item{pipelineId}{[required] The ID of the pipeline.}

\item{tags}{[required] The tags to add, as key/value pairs.}
}
\description{
Adds or modifies tags for the specified pipeline.

See \url{https://www.paws-r-sdk.com/docs/datapipeline_add_tags/} for full documentation.
}
\keyword{internal}
