% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapipeline_operations.R
\name{datapipeline_delete_pipeline}
\alias{datapipeline_delete_pipeline}
\title{Deletes a pipeline, its pipeline definition, and its run history}
\usage{
datapipeline_delete_pipeline(pipelineId)
}
\arguments{
\item{pipelineId}{[required] The ID of the pipeline.}
}
\description{
Deletes a pipeline, its pipeline definition, and its run history. AWS Data Pipeline attempts to cancel instances associated with the pipeline that are currently being processed by task runners.

See \url{https://paws-r.github.io/docs/datapipeline/delete_pipeline.html} for full documentation.
}
\keyword{internal}
