% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapipeline_operations.R
\name{datapipeline_get_pipeline_definition}
\alias{datapipeline_get_pipeline_definition}
\title{Gets the definition of the specified pipeline}
\usage{
datapipeline_get_pipeline_definition(pipelineId, version)
}
\arguments{
\item{pipelineId}{[required] The ID of the pipeline.}

\item{version}{The version of the pipeline definition to retrieve. Set this parameter
to \code{latest} (default) to use the last definition saved to the pipeline
or \code{active} to use the last definition that was activated.}
}
\value{
A list with the following syntax:\preformatted{list(
  pipelineObjects = list(
    list(
      id = "string",
      name = "string",
      fields = list(
        list(
          key = "string",
          stringValue = "string",
          refValue = "string"
        )
      )
    )
  ),
  parameterObjects = list(
    list(
      id = "string",
      attributes = list(
        list(
          key = "string",
          stringValue = "string"
        )
      )
    )
  ),
  parameterValues = list(
    list(
      id = "string",
      stringValue = "string"
    )
  )
)
}
}
\description{
Gets the definition of the specified pipeline. You can call
\code{\link[=datapipeline_get_pipeline_definition]{get_pipeline_definition}} to
retrieve the pipeline definition that you provided using
\code{\link[=datapipeline_put_pipeline_definition]{put_pipeline_definition}}.
}
\section{Request syntax}{
\preformatted{svc$get_pipeline_definition(
  pipelineId = "string",
  version = "string"
)
}
}

\keyword{internal}
