% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_add_entity_owner}
\alias{datazone_add_entity_owner}
\title{Adds the owner of an entity (a domain unit)}
\usage{
datazone_add_entity_owner(
  domainIdentifier,
  entityType,
  entityIdentifier,
  owner,
  clientToken = NULL
)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the domain in which you want to add the entity owner.}

\item{entityType}{[required] The type of an entity.}

\item{entityIdentifier}{[required] The ID of the entity to which you want to add an owner.}

\item{owner}{[required] The owner that you want to add to the entity.}

\item{clientToken}{A unique, case-sensitive identifier that is provided to ensure the
idempotency of the request.}
}
\description{
Adds the owner of an entity (a domain unit).

See \url{https://www.paws-r-sdk.com/docs/datazone_add_entity_owner/} for full documentation.
}
\keyword{internal}
