% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_associate_environment_role}
\alias{datazone_associate_environment_role}
\title{Associates the environment role in Amazon DataZone}
\usage{
datazone_associate_environment_role(
  domainIdentifier,
  environmentIdentifier,
  environmentRoleArn
)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the Amazon DataZone domain in which the environment role is
associated.}

\item{environmentIdentifier}{[required] The ID of the Amazon DataZone environment.}

\item{environmentRoleArn}{[required] The ARN of the environment role.}
}
\description{
Associates the environment role in Amazon DataZone.

See \url{https://www.paws-r-sdk.com/docs/datazone_associate_environment_role/} for full documentation.
}
\keyword{internal}
