% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_associate_governed_terms}
\alias{datazone_associate_governed_terms}
\title{Associates governed terms with an asset}
\usage{
datazone_associate_governed_terms(
  domainIdentifier,
  entityIdentifier,
  entityType,
  governedGlossaryTerms
)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the domain where governed terms are to be associated with an
asset.}

\item{entityIdentifier}{[required] The ID of the asset with which you want to associate a governed term.}

\item{entityType}{[required] The type of the asset with which you want to associate a governed term.}

\item{governedGlossaryTerms}{[required] The glossary terms in a restricted glossary.}
}
\description{
Associates governed terms with an asset.

See \url{https://www.paws-r-sdk.com/docs/datazone_associate_governed_terms/} for full documentation.
}
\keyword{internal}
