% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_batch_get_attributes_metadata}
\alias{datazone_batch_get_attributes_metadata}
\title{Gets the attribute metadata}
\usage{
datazone_batch_get_attributes_metadata(
  attributeIdentifiers,
  domainIdentifier,
  entityIdentifier,
  entityRevision = NULL,
  entityType
)
}
\arguments{
\item{attributeIdentifiers}{[required] The attribute identifier.}

\item{domainIdentifier}{[required] The domain ID where you want to get the attribute metadata.}

\item{entityIdentifier}{[required] The entity ID for which you want to get attribute metadata.}

\item{entityRevision}{The entity revision for which you want to get attribute metadata.}

\item{entityType}{[required] The entity type for which you want to get attribute metadata.}
}
\description{
Gets the attribute metadata.

See \url{https://www.paws-r-sdk.com/docs/datazone_batch_get_attributes_metadata/} for full documentation.
}
\keyword{internal}
