% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_batch_put_attributes_metadata}
\alias{datazone_batch_put_attributes_metadata}
\title{Writes the attribute metadata}
\usage{
datazone_batch_put_attributes_metadata(
  attributes,
  clientToken = NULL,
  domainIdentifier,
  entityIdentifier,
  entityType
)
}
\arguments{
\item{attributes}{[required] The attributes of the metadata.}

\item{clientToken}{A unique, case-sensitive identifier to ensure idempotency of the
request. This field is automatically populated if not provided.}

\item{domainIdentifier}{[required] The domain ID where you want to write the attribute metadata.}

\item{entityIdentifier}{[required] The entity ID for which you want to write the attribute metadata.}

\item{entityType}{[required] The entity type for which you want to write the attribute metadata.}
}
\description{
Writes the attribute metadata.

See \url{https://www.paws-r-sdk.com/docs/datazone_batch_put_attributes_metadata/} for full documentation.
}
\keyword{internal}
