% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_cancel_metadata_generation_run}
\alias{datazone_cancel_metadata_generation_run}
\title{Cancels the metadata generation run}
\usage{
datazone_cancel_metadata_generation_run(domainIdentifier, identifier)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the Amazon DataZone domain in which the metadata generation
run is to be cancelled.}

\item{identifier}{[required] The ID of the metadata generation run.}
}
\description{
Cancels the metadata generation run.

See \url{https://www.paws-r-sdk.com/docs/datazone_cancel_metadata_generation_run/} for full documentation.
}
\keyword{internal}
