% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_cancel_subscription}
\alias{datazone_cancel_subscription}
\title{Cancels the subscription to the specified asset}
\usage{
datazone_cancel_subscription(domainIdentifier, identifier)
}
\arguments{
\item{domainIdentifier}{[required] The unique identifier of the Amazon DataZone domain where the
subscription request is being cancelled.}

\item{identifier}{[required] The unique identifier of the subscription that is being cancelled.}
}
\description{
Cancels the subscription to the specified asset.

See \url{https://www.paws-r-sdk.com/docs/datazone_cancel_subscription/} for full documentation.
}
\keyword{internal}
