% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_create_domain_unit}
\alias{datazone_create_domain_unit}
\title{Creates a domain unit in Amazon DataZone}
\usage{
datazone_create_domain_unit(
  clientToken = NULL,
  description = NULL,
  domainIdentifier,
  name,
  parentDomainUnitIdentifier
)
}
\arguments{
\item{clientToken}{A unique, case-sensitive identifier that is provided to ensure the
idempotency of the request.}

\item{description}{The description of the domain unit.}

\item{domainIdentifier}{[required] The ID of the domain where you want to crate a domain unit.}

\item{name}{[required] The name of the domain unit.}

\item{parentDomainUnitIdentifier}{[required] The ID of the parent domain unit.}
}
\description{
Creates a domain unit in Amazon DataZone.

See \url{https://www.paws-r-sdk.com/docs/datazone_create_domain_unit/} for full documentation.
}
\keyword{internal}
