% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_create_environment_action}
\alias{datazone_create_environment_action}
\title{Creates an action for the environment, for example, creates a console
link for an analytics tool that is available in this environment}
\usage{
datazone_create_environment_action(
  description = NULL,
  domainIdentifier,
  environmentIdentifier,
  name,
  parameters
)
}
\arguments{
\item{description}{The description of the environment action that is being created in the
environment.}

\item{domainIdentifier}{[required] The ID of the Amazon DataZone domain in which the environment action is
created.}

\item{environmentIdentifier}{[required] The ID of the environment in which the environment action is created.}

\item{name}{[required] The name of the environment action.}

\item{parameters}{[required] The parameters of the environment action.}
}
\description{
Creates an action for the environment, for example, creates a console link for an analytics tool that is available in this environment.

See \url{https://www.paws-r-sdk.com/docs/datazone_create_environment_action/} for full documentation.
}
\keyword{internal}
