% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_create_environment_profile}
\alias{datazone_create_environment_profile}
\title{Creates an Amazon DataZone environment profile}
\usage{
datazone_create_environment_profile(
  domainIdentifier,
  name,
  description = NULL,
  environmentBlueprintIdentifier,
  projectIdentifier,
  userParameters = NULL,
  awsAccountId = NULL,
  awsAccountRegion = NULL
)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the Amazon DataZone domain in which this environment profile
is created.}

\item{name}{[required] The name of this Amazon DataZone environment profile.}

\item{description}{The description of this Amazon DataZone environment profile.}

\item{environmentBlueprintIdentifier}{[required] The ID of the blueprint with which this environment profile is created.}

\item{projectIdentifier}{[required] The identifier of the project in which to create the environment
profile.}

\item{userParameters}{The user parameters of this Amazon DataZone environment profile.}

\item{awsAccountId}{The Amazon Web Services account in which the Amazon DataZone environment
is created.}

\item{awsAccountRegion}{The Amazon Web Services region in which this environment profile is
created.}
}
\description{
Creates an Amazon DataZone environment profile.

See \url{https://www.paws-r-sdk.com/docs/datazone_create_environment_profile/} for full documentation.
}
\keyword{internal}
