% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_create_glossary}
\alias{datazone_create_glossary}
\title{Creates an Amazon DataZone business glossary}
\usage{
datazone_create_glossary(
  clientToken = NULL,
  description = NULL,
  domainIdentifier,
  name,
  owningProjectIdentifier,
  status = NULL,
  usageRestrictions = NULL
)
}
\arguments{
\item{clientToken}{A unique, case-sensitive identifier that is provided to ensure the
idempotency of the request.}

\item{description}{The description of this business glossary.}

\item{domainIdentifier}{[required] The ID of the Amazon DataZone domain in which this business glossary is
created.}

\item{name}{[required] The name of this business glossary.}

\item{owningProjectIdentifier}{[required] The ID of the project that currently owns business glossary.}

\item{status}{The status of this business glossary.}

\item{usageRestrictions}{The usage restriction of the restricted glossary.}
}
\description{
Creates an Amazon DataZone business glossary.

See \url{https://www.paws-r-sdk.com/docs/datazone_create_glossary/} for full documentation.
}
\keyword{internal}
