% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_create_group_profile}
\alias{datazone_create_group_profile}
\title{Creates a group profile in Amazon DataZone}
\usage{
datazone_create_group_profile(
  domainIdentifier,
  groupIdentifier = NULL,
  rolePrincipalArn = NULL,
  clientToken = NULL
)
}
\arguments{
\item{domainIdentifier}{[required] The identifier of the Amazon DataZone domain in which the group profile
is created.}

\item{groupIdentifier}{The identifier of the group for which the group profile is created.}

\item{rolePrincipalArn}{The ARN of the IAM role that will be associated with the group profile.
This role defines the permissions that group members will assume when
accessing Amazon DataZone resources.}

\item{clientToken}{A unique, case-sensitive identifier that is provided to ensure the
idempotency of the request.}
}
\description{
Creates a group profile in Amazon DataZone.

See \url{https://www.paws-r-sdk.com/docs/datazone_create_group_profile/} for full documentation.
}
\keyword{internal}
