% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_create_listing_change_set}
\alias{datazone_create_listing_change_set}
\title{Publishes a listing (a record of an asset at a given time) or removes a
listing from the catalog}
\usage{
datazone_create_listing_change_set(
  domainIdentifier,
  entityIdentifier,
  entityType,
  entityRevision = NULL,
  action,
  clientToken = NULL
)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the Amazon DataZone domain.}

\item{entityIdentifier}{[required] The ID of the asset.}

\item{entityType}{[required] The type of an entity.}

\item{entityRevision}{The revision of an asset.}

\item{action}{[required] Specifies whether to publish or unpublish a listing.}

\item{clientToken}{A unique, case-sensitive identifier that is provided to ensure the
idempotency of the request.}
}
\description{
Publishes a listing (a record of an asset at a given time) or removes a listing from the catalog.

See \url{https://www.paws-r-sdk.com/docs/datazone_create_listing_change_set/} for full documentation.
}
\keyword{internal}
