% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_create_notebook}
\alias{datazone_create_notebook}
\title{Creates a notebook in Amazon SageMaker Unified Studio}
\usage{
datazone_create_notebook(
  domainIdentifier,
  owningProjectIdentifier,
  name,
  description = NULL,
  type = NULL,
  metadata = NULL,
  parameters = NULL,
  clientToken = NULL
)
}
\arguments{
\item{domainIdentifier}{[required] The identifier of the Amazon SageMaker Unified Studio domain in which to create the notebook.}

\item{owningProjectIdentifier}{[required] The identifier of the project that owns the notebook.}

\item{name}{[required] The name of the notebook. The name must be between 1 and 256 characters.}

\item{description}{The description of the notebook.}

\item{type}{The type of the notebook.}

\item{metadata}{The metadata for the notebook, specified as key-value pairs. You can specify up to 50 entries, with keys up to 128 characters and values up to 1024 characters.}

\item{parameters}{The sensitive parameters for the notebook, specified as key-value pairs. You can specify up to 50 entries, with keys up to 128 characters and values up to 1024 characters.}

\item{clientToken}{A unique, case-sensitive identifier to ensure idempotency of the request. This field is automatically populated if not provided.}
}
\description{
Creates a \href{https://docs.aws.amazon.com/sagemaker-unified-studio/latest/userguide/notebooks.html}{notebook} in Amazon SageMaker Unified Studio. A notebook is a collaborative document within a project that contains code cells for interactive computing.

See \url{https://www.paws-r-sdk.com/docs/datazone_create_notebook/} for full documentation.
}
\keyword{internal}
