% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_create_subscription_request}
\alias{datazone_create_subscription_request}
\title{Creates a subscription request in Amazon DataZone}
\usage{
datazone_create_subscription_request(
  domainIdentifier,
  subscribedPrincipals,
  subscribedListings,
  requestReason,
  clientToken = NULL,
  metadataForms = NULL,
  assetPermissions = NULL,
  assetScopes = NULL
)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the Amazon DataZone domain in which the subscription request
is created.}

\item{subscribedPrincipals}{[required] The Amazon DataZone principals for whom the subscription request is
created.}

\item{subscribedListings}{[required] The published asset for which the subscription grant is to be created.}

\item{requestReason}{[required] The reason for the subscription request.}

\item{clientToken}{A unique, case-sensitive identifier that is provided to ensure the
idempotency of the request.}

\item{metadataForms}{The metadata form included in the subscription request.}

\item{assetPermissions}{The asset permissions of the subscription request.}

\item{assetScopes}{The asset scopes of the subscription request.}
}
\description{
Creates a subscription request in Amazon DataZone.

See \url{https://www.paws-r-sdk.com/docs/datazone_create_subscription_request/} for full documentation.
}
\keyword{internal}
