% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_create_user_profile}
\alias{datazone_create_user_profile}
\title{Creates a user profile in Amazon DataZone}
\usage{
datazone_create_user_profile(
  clientToken = NULL,
  domainIdentifier,
  userIdentifier,
  userType = NULL
)
}
\arguments{
\item{clientToken}{A unique, case-sensitive identifier that is provided to ensure the
idempotency of the request.}

\item{domainIdentifier}{[required] The identifier of the Amazon DataZone domain in which a user profile is
created.}

\item{userIdentifier}{[required] The identifier of the user for which the user profile is created.}

\item{userType}{The user type of the user for which the user profile is created.}
}
\description{
Creates a user profile in Amazon DataZone.

See \url{https://www.paws-r-sdk.com/docs/datazone_create_user_profile/} for full documentation.
}
\keyword{internal}
