% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_delete_account_pool}
\alias{datazone_delete_account_pool}
\title{Deletes an account pool}
\usage{
datazone_delete_account_pool(domainIdentifier, identifier)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the domain where the account pool is deleted.}

\item{identifier}{[required] The ID of the account pool to be deleted.}
}
\description{
Deletes an account pool.

See \url{https://www.paws-r-sdk.com/docs/datazone_delete_account_pool/} for full documentation.
}
\keyword{internal}
