% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_delete_asset}
\alias{datazone_delete_asset}
\title{Delets an asset in Amazon DataZone}
\usage{
datazone_delete_asset(domainIdentifier, identifier)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the Amazon DataZone domain in which the asset is deleted.}

\item{identifier}{[required] The identifier of the asset that is deleted.}
}
\description{
Delets an asset in Amazon DataZone.

See \url{https://www.paws-r-sdk.com/docs/datazone_delete_asset/} for full documentation.
}
\keyword{internal}
