% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_delete_asset_filter}
\alias{datazone_delete_asset_filter}
\title{Deletes an asset filter}
\usage{
datazone_delete_asset_filter(domainIdentifier, assetIdentifier, identifier)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the domain where you want to delete an asset filter.}

\item{assetIdentifier}{[required] The ID of the data asset.}

\item{identifier}{[required] The ID of the asset filter that you want to delete.}
}
\description{
Deletes an asset filter.

See \url{https://www.paws-r-sdk.com/docs/datazone_delete_asset_filter/} for full documentation.
}
\keyword{internal}
