% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_delete_asset_type}
\alias{datazone_delete_asset_type}
\title{Deletes an asset type in Amazon DataZone}
\usage{
datazone_delete_asset_type(domainIdentifier, identifier)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the Amazon DataZone domain in which the asset type is deleted.}

\item{identifier}{[required] The identifier of the asset type that is deleted.}
}
\description{
Deletes an asset type in Amazon DataZone.

See \url{https://www.paws-r-sdk.com/docs/datazone_delete_asset_type/} for full documentation.
}
\keyword{internal}
