% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_delete_connection}
\alias{datazone_delete_connection}
\title{Deletes and connection}
\usage{
datazone_delete_connection(domainIdentifier, identifier)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the domain where the connection is deleted.}

\item{identifier}{[required] The ID of the connection that is deleted.}
}
\description{
Deletes and connection. In Amazon DataZone, a connection enables you to connect your resources (domains, projects, and environments) to external resources and services.

See \url{https://www.paws-r-sdk.com/docs/datazone_delete_connection/} for full documentation.
}
\keyword{internal}
