% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_delete_data_export_configuration}
\alias{datazone_delete_data_export_configuration}
\title{Deletes data export configuration for a domain}
\usage{
datazone_delete_data_export_configuration(domainIdentifier)
}
\arguments{
\item{domainIdentifier}{[required] The domain ID for which you want to delete the data export
configuration.}
}
\description{
Deletes data export configuration for a domain.

See \url{https://www.paws-r-sdk.com/docs/datazone_delete_data_export_configuration/} for full documentation.
}
\keyword{internal}
