% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_delete_data_product}
\alias{datazone_delete_data_product}
\title{Deletes a data product in Amazon DataZone}
\usage{
datazone_delete_data_product(domainIdentifier, identifier)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the Amazon DataZone domain in which the data product is
deleted.}

\item{identifier}{[required] The identifier of the data product that is deleted.}
}
\description{
Deletes a data product in Amazon DataZone.

See \url{https://www.paws-r-sdk.com/docs/datazone_delete_data_product/} for full documentation.
}
\keyword{internal}
