% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_delete_data_source}
\alias{datazone_delete_data_source}
\title{Deletes a data source in Amazon DataZone}
\usage{
datazone_delete_data_source(
  clientToken = NULL,
  domainIdentifier,
  identifier,
  retainPermissionsOnRevokeFailure = NULL
)
}
\arguments{
\item{clientToken}{A unique, case-sensitive identifier that is provided to ensure the
idempotency of the request.}

\item{domainIdentifier}{[required] The ID of the Amazon DataZone domain in which the data source is
deleted.}

\item{identifier}{[required] The identifier of the data source that is deleted.}

\item{retainPermissionsOnRevokeFailure}{Specifies that the granted permissions are retained in case of a
self-subscribe functionality failure for a data source.}
}
\description{
Deletes a data source in Amazon DataZone.

See \url{https://www.paws-r-sdk.com/docs/datazone_delete_data_source/} for full documentation.
}
\keyword{internal}
