% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_delete_domain_unit}
\alias{datazone_delete_domain_unit}
\title{Deletes a domain unit}
\usage{
datazone_delete_domain_unit(domainIdentifier, identifier)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the domain where you want to delete a domain unit.}

\item{identifier}{[required] The ID of the domain unit that you want to delete.}
}
\description{
Deletes a domain unit.

See \url{https://www.paws-r-sdk.com/docs/datazone_delete_domain_unit/} for full documentation.
}
\keyword{internal}
