% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_delete_environment}
\alias{datazone_delete_environment}
\title{Deletes an environment in Amazon DataZone}
\usage{
datazone_delete_environment(domainIdentifier, identifier)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the Amazon DataZone domain in which the environment is
deleted.}

\item{identifier}{[required] The identifier of the environment that is to be deleted.}
}
\description{
Deletes an environment in Amazon DataZone.

See \url{https://www.paws-r-sdk.com/docs/datazone_delete_environment/} for full documentation.
}
\keyword{internal}
