% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_delete_environment_blueprint}
\alias{datazone_delete_environment_blueprint}
\title{Deletes a blueprint in Amazon DataZone}
\usage{
datazone_delete_environment_blueprint(domainIdentifier, identifier)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the Amazon DataZone domain in which the blueprint is deleted.}

\item{identifier}{[required] The ID of the blueprint that is deleted.}
}
\description{
Deletes a blueprint in Amazon DataZone.

See \url{https://www.paws-r-sdk.com/docs/datazone_delete_environment_blueprint/} for full documentation.
}
\keyword{internal}
