% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_delete_environment_blueprint_configuration}
\alias{datazone_delete_environment_blueprint_configuration}
\title{Deletes the blueprint configuration in Amazon DataZone}
\usage{
datazone_delete_environment_blueprint_configuration(
  domainIdentifier,
  environmentBlueprintIdentifier
)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the Amazon DataZone domain in which the blueprint
configuration is deleted.}

\item{environmentBlueprintIdentifier}{[required] The ID of the blueprint the configuration of which is deleted.}
}
\description{
Deletes the blueprint configuration in Amazon DataZone.

See \url{https://www.paws-r-sdk.com/docs/datazone_delete_environment_blueprint_configuration/} for full documentation.
}
\keyword{internal}
