% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_delete_environment_profile}
\alias{datazone_delete_environment_profile}
\title{Deletes an environment profile in Amazon DataZone}
\usage{
datazone_delete_environment_profile(domainIdentifier, identifier)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the Amazon DataZone domain in which the environment profile is
deleted.}

\item{identifier}{[required] The ID of the environment profile that is deleted.}
}
\description{
Deletes an environment profile in Amazon DataZone.

See \url{https://www.paws-r-sdk.com/docs/datazone_delete_environment_profile/} for full documentation.
}
\keyword{internal}
