% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_delete_form_type}
\alias{datazone_delete_form_type}
\title{Deletes and metadata form type in Amazon DataZone}
\usage{
datazone_delete_form_type(domainIdentifier, formTypeIdentifier)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the Amazon DataZone domain in which the metadata form type is
deleted.}

\item{formTypeIdentifier}{[required] The ID of the metadata form type that is deleted.}
}
\description{
Deletes and metadata form type in Amazon DataZone.

See \url{https://www.paws-r-sdk.com/docs/datazone_delete_form_type/} for full documentation.
}
\keyword{internal}
