% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_delete_glossary}
\alias{datazone_delete_glossary}
\title{Deletes a business glossary in Amazon DataZone}
\usage{
datazone_delete_glossary(domainIdentifier, identifier)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the Amazon DataZone domain in which the business glossary is
deleted.}

\item{identifier}{[required] The ID of the business glossary that is deleted.}
}
\description{
Deletes a business glossary in Amazon DataZone.

See \url{https://www.paws-r-sdk.com/docs/datazone_delete_glossary/} for full documentation.
}
\keyword{internal}
