% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_delete_lineage_event}
\alias{datazone_delete_lineage_event}
\title{Deletes the specified lineage event}
\usage{
datazone_delete_lineage_event(domainIdentifier, identifier)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the domain.}

\item{identifier}{[required] The ID of the lineage event.}
}
\description{
Deletes the specified lineage event.

See \url{https://www.paws-r-sdk.com/docs/datazone_delete_lineage_event/} for full documentation.
}
\keyword{internal}
