% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_delete_listing}
\alias{datazone_delete_listing}
\title{Delete listing}
\usage{
datazone_delete_listing(domainIdentifier, identifier)
}
\arguments{
\item{domainIdentifier}{[required]}

\item{identifier}{[required]}
}
\description{
Delete listing

See \url{https://www.paws-r-sdk.com/docs/datazone_delete_listing/} for full documentation.
}
\keyword{internal}
