% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_delete_project}
\alias{datazone_delete_project}
\title{Deletes a project in Amazon DataZone}
\usage{
datazone_delete_project(domainIdentifier, identifier)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the Amazon DataZone domain in which the project is deleted.}

\item{identifier}{[required] The identifier of the project that is to be deleted.}
}
\description{
Deletes a project in Amazon DataZone.

See \url{https://www.paws-r-sdk.com/docs/datazone_delete_project/} for full documentation.
}
\keyword{internal}
