% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_delete_project_membership}
\alias{datazone_delete_project_membership}
\title{Deletes project membership in Amazon DataZone}
\usage{
datazone_delete_project_membership(domainIdentifier, member, projectIdentifier)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the Amazon DataZone domain where project membership is
deleted.}

\item{member}{[required] The project member whose project membership is deleted.}

\item{projectIdentifier}{[required] The ID of the Amazon DataZone project the membership to which is
deleted.}
}
\description{
Deletes project membership in Amazon DataZone.

See \url{https://www.paws-r-sdk.com/docs/datazone_delete_project_membership/} for full documentation.
}
\keyword{internal}
