% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_delete_project_profile}
\alias{datazone_delete_project_profile}
\title{Deletes a project profile}
\usage{
datazone_delete_project_profile(domainIdentifier, identifier)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the domain where a project profile is deleted.}

\item{identifier}{[required] The ID of the project profile that is deleted.}
}
\description{
Deletes a project profile.

See \url{https://www.paws-r-sdk.com/docs/datazone_delete_project_profile/} for full documentation.
}
\keyword{internal}
