% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_delete_rule}
\alias{datazone_delete_rule}
\title{Deletes a rule in Amazon DataZone}
\usage{
datazone_delete_rule(domainIdentifier, identifier)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the domain that where the rule is to be deleted.}

\item{identifier}{[required] The ID of the rule that is to be deleted.}
}
\description{
Deletes a rule in Amazon DataZone. A rule is a formal agreement that enforces specific requirements across user workflows (e.g., publishing assets to the catalog, requesting subscriptions, creating projects) within the Amazon DataZone data portal. These rules help maintain consistency, ensure compliance, and uphold governance standards in data management processes. For instance, a metadata enforcement rule can specify the required information for creating a subscription request or publishing a data asset to the catalog, ensuring alignment with organizational standards.

See \url{https://www.paws-r-sdk.com/docs/datazone_delete_rule/} for full documentation.
}
\keyword{internal}
