% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_delete_subscription_grant}
\alias{datazone_delete_subscription_grant}
\title{Deletes and subscription grant in Amazon DataZone}
\usage{
datazone_delete_subscription_grant(domainIdentifier, identifier)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the Amazon DataZone domain where the subscription grant is
deleted.}

\item{identifier}{[required] The ID of the subscription grant that is deleted.}
}
\description{
Deletes and subscription grant in Amazon DataZone.

See \url{https://www.paws-r-sdk.com/docs/datazone_delete_subscription_grant/} for full documentation.
}
\keyword{internal}
