% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_delete_subscription_request}
\alias{datazone_delete_subscription_request}
\title{Deletes a subscription request in Amazon DataZone}
\usage{
datazone_delete_subscription_request(domainIdentifier, identifier)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the Amazon DataZone domain in which the subscription request
is deleted.}

\item{identifier}{[required] The ID of the subscription request that is deleted.}
}
\description{
Deletes a subscription request in Amazon DataZone.

See \url{https://www.paws-r-sdk.com/docs/datazone_delete_subscription_request/} for full documentation.
}
\keyword{internal}
