% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_delete_subscription_target}
\alias{datazone_delete_subscription_target}
\title{Deletes a subscription target in Amazon DataZone}
\usage{
datazone_delete_subscription_target(
  domainIdentifier,
  environmentIdentifier,
  identifier
)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the Amazon DataZone domain in which the subscription target is
deleted.}

\item{environmentIdentifier}{[required] The ID of the Amazon DataZone environment in which the subscription
target is deleted.}

\item{identifier}{[required] The ID of the subscription target that is deleted.}
}
\description{
Deletes a subscription target in Amazon DataZone.

See \url{https://www.paws-r-sdk.com/docs/datazone_delete_subscription_target/} for full documentation.
}
\keyword{internal}
