% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_disassociate_governed_terms}
\alias{datazone_disassociate_governed_terms}
\title{Disassociates restricted terms from an asset}
\usage{
datazone_disassociate_governed_terms(
  domainIdentifier,
  entityIdentifier,
  entityType,
  governedGlossaryTerms
)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the domain where you want to disassociate restricted terms
from an asset.}

\item{entityIdentifier}{[required] The ID of an asset from which you want to disassociate restricted terms.}

\item{entityType}{[required] The type of the asset from which you want to disassociate restricted
terms.}

\item{governedGlossaryTerms}{[required] The restricted glossary terms that you want to disassociate from an
asset.}
}
\description{
Disassociates restricted terms from an asset.

See \url{https://www.paws-r-sdk.com/docs/datazone_disassociate_governed_terms/} for full documentation.
}
\keyword{internal}
