% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_get_account_pool}
\alias{datazone_get_account_pool}
\title{Gets the details of the account pool}
\usage{
datazone_get_account_pool(domainIdentifier, identifier)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the domain in which the account pool lives whose details are
to be displayed.}

\item{identifier}{[required] The ID of the account pool whose details are to be displayed.}
}
\description{
Gets the details of the account pool.

See \url{https://www.paws-r-sdk.com/docs/datazone_get_account_pool/} for full documentation.
}
\keyword{internal}
