% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_get_asset}
\alias{datazone_get_asset}
\title{Gets an Amazon DataZone asset}
\usage{
datazone_get_asset(domainIdentifier, identifier, revision = NULL)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the Amazon DataZone domain to which the asset belongs.}

\item{identifier}{[required] The ID of the Amazon DataZone asset.

This parameter supports either the value of \code{assetId} or
\code{externalIdentifier} as input. If you are passing the value of
\code{externalIdentifier}, you must prefix this value with
\verb{externalIdentifer\%2F}.}

\item{revision}{The revision of the Amazon DataZone asset.}
}
\description{
Gets an Amazon DataZone asset.

See \url{https://www.paws-r-sdk.com/docs/datazone_get_asset/} for full documentation.
}
\keyword{internal}
