% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_get_asset_type}
\alias{datazone_get_asset_type}
\title{Gets an Amazon DataZone asset type}
\usage{
datazone_get_asset_type(domainIdentifier, identifier, revision = NULL)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the Amazon DataZone domain in which the asset type exists.}

\item{identifier}{[required] The ID of the asset type.}

\item{revision}{The revision of the asset type.}
}
\description{
Gets an Amazon DataZone asset type.

See \url{https://www.paws-r-sdk.com/docs/datazone_get_asset_type/} for full documentation.
}
\keyword{internal}
