% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_get_connection}
\alias{datazone_get_connection}
\title{Gets a connection}
\usage{
datazone_get_connection(domainIdentifier, identifier, withSecret = NULL)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the domain where we get the connection.}

\item{identifier}{[required] The connection ID.}

\item{withSecret}{Specifies whether a connection has a secret.}
}
\description{
Gets a connection. In Amazon DataZone, a connection enables you to connect your resources (domains, projects, and environments) to external resources and services.

See \url{https://www.paws-r-sdk.com/docs/datazone_get_connection/} for full documentation.
}
\keyword{internal}
