% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_get_data_export_configuration}
\alias{datazone_get_data_export_configuration}
\title{Gets data export configuration details}
\usage{
datazone_get_data_export_configuration(domainIdentifier)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the domain where you want to get the data export configuration
details.}
}
\description{
Gets data export configuration details.

See \url{https://www.paws-r-sdk.com/docs/datazone_get_data_export_configuration/} for full documentation.
}
\keyword{internal}
