% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_get_data_product}
\alias{datazone_get_data_product}
\title{Gets the data product}
\usage{
datazone_get_data_product(domainIdentifier, identifier, revision = NULL)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the domain where the data product lives.}

\item{identifier}{[required] The ID of the data product.}

\item{revision}{The revision of the data product.}
}
\description{
Gets the data product.

See \url{https://www.paws-r-sdk.com/docs/datazone_get_data_product/} for full documentation.
}
\keyword{internal}
