% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_get_data_source}
\alias{datazone_get_data_source}
\title{Gets an Amazon DataZone data source}
\usage{
datazone_get_data_source(domainIdentifier, identifier)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the Amazon DataZone domain in which the data source exists.}

\item{identifier}{[required] The ID of the Amazon DataZone data source.}
}
\description{
Gets an Amazon DataZone data source.

See \url{https://www.paws-r-sdk.com/docs/datazone_get_data_source/} for full documentation.
}
\keyword{internal}
