% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_get_data_source_run}
\alias{datazone_get_data_source_run}
\title{Gets an Amazon DataZone data source run}
\usage{
datazone_get_data_source_run(domainIdentifier, identifier)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the domain in which this data source run was performed.}

\item{identifier}{[required] The ID of the data source run.}
}
\description{
Gets an Amazon DataZone data source run.

See \url{https://www.paws-r-sdk.com/docs/datazone_get_data_source_run/} for full documentation.
}
\keyword{internal}
