% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_get_domain}
\alias{datazone_get_domain}
\title{Gets an Amazon DataZone domain}
\usage{
datazone_get_domain(identifier)
}
\arguments{
\item{identifier}{[required] The identifier of the specified Amazon DataZone domain.}
}
\description{
Gets an Amazon DataZone domain.

See \url{https://www.paws-r-sdk.com/docs/datazone_get_domain/} for full documentation.
}
\keyword{internal}
