% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_get_domain_unit}
\alias{datazone_get_domain_unit}
\title{Gets the details of the specified domain unit}
\usage{
datazone_get_domain_unit(domainIdentifier, identifier)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the domain where you want to get a domain unit.}

\item{identifier}{[required] The identifier of the domain unit that you want to get.}
}
\description{
Gets the details of the specified domain unit.

See \url{https://www.paws-r-sdk.com/docs/datazone_get_domain_unit/} for full documentation.
}
\keyword{internal}
