% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_get_environment}
\alias{datazone_get_environment}
\title{Gets an Amazon DataZone environment}
\usage{
datazone_get_environment(domainIdentifier, identifier)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the Amazon DataZone domain where the environment exists.}

\item{identifier}{[required] The ID of the Amazon DataZone environment.}
}
\description{
Gets an Amazon DataZone environment.

See \url{https://www.paws-r-sdk.com/docs/datazone_get_environment/} for full documentation.
}
\keyword{internal}
