% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_get_environment_action}
\alias{datazone_get_environment_action}
\title{Gets the specified environment action}
\usage{
datazone_get_environment_action(
  domainIdentifier,
  environmentIdentifier,
  identifier
)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the Amazon DataZone domain in which the
\code{\link[=datazone_get_environment_action]{get_environment_action}} API is
invoked.}

\item{environmentIdentifier}{[required] The environment ID of the environment action.}

\item{identifier}{[required] The ID of the environment action}
}
\description{
Gets the specified environment action.

See \url{https://www.paws-r-sdk.com/docs/datazone_get_environment_action/} for full documentation.
}
\keyword{internal}
