% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_get_environment_blueprint}
\alias{datazone_get_environment_blueprint}
\title{Gets an Amazon DataZone blueprint}
\usage{
datazone_get_environment_blueprint(domainIdentifier, identifier)
}
\arguments{
\item{domainIdentifier}{[required] The identifier of the domain in which this blueprint exists.}

\item{identifier}{[required] The ID of this Amazon DataZone blueprint.}
}
\description{
Gets an Amazon DataZone blueprint.

See \url{https://www.paws-r-sdk.com/docs/datazone_get_environment_blueprint/} for full documentation.
}
\keyword{internal}
