% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_get_environment_blueprint_configuration}
\alias{datazone_get_environment_blueprint_configuration}
\title{Gets the blueprint configuration in Amazon DataZone}
\usage{
datazone_get_environment_blueprint_configuration(
  domainIdentifier,
  environmentBlueprintIdentifier
)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the Amazon DataZone domain where this blueprint exists.}

\item{environmentBlueprintIdentifier}{[required] He ID of the blueprint.}
}
\description{
Gets the blueprint configuration in Amazon DataZone.

See \url{https://www.paws-r-sdk.com/docs/datazone_get_environment_blueprint_configuration/} for full documentation.
}
\keyword{internal}
