% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_get_environment_credentials}
\alias{datazone_get_environment_credentials}
\title{Gets the credentials of an environment in Amazon DataZone}
\usage{
datazone_get_environment_credentials(domainIdentifier, environmentIdentifier)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the Amazon DataZone domain in which this environment and its
credentials exist.}

\item{environmentIdentifier}{[required] The ID of the environment whose credentials this operation gets.}
}
\description{
Gets the credentials of an environment in Amazon DataZone.

See \url{https://www.paws-r-sdk.com/docs/datazone_get_environment_credentials/} for full documentation.
}
\keyword{internal}
