% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_get_environment_profile}
\alias{datazone_get_environment_profile}
\title{Gets an evinronment profile in Amazon DataZone}
\usage{
datazone_get_environment_profile(domainIdentifier, identifier)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the Amazon DataZone domain in which this environment profile
exists.}

\item{identifier}{[required] The ID of the environment profile.}
}
\description{
Gets an evinronment profile in Amazon DataZone.

See \url{https://www.paws-r-sdk.com/docs/datazone_get_environment_profile/} for full documentation.
}
\keyword{internal}
