% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_get_form_type}
\alias{datazone_get_form_type}
\title{Gets a metadata form type in Amazon DataZone}
\usage{
datazone_get_form_type(domainIdentifier, formTypeIdentifier, revision = NULL)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the Amazon DataZone domain in which this metadata form type
exists.}

\item{formTypeIdentifier}{[required] The ID of the metadata form type.}

\item{revision}{The revision of this metadata form type.}
}
\description{
Gets a metadata form type in Amazon DataZone.

See \url{https://www.paws-r-sdk.com/docs/datazone_get_form_type/} for full documentation.
}
\keyword{internal}
