% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_get_glossary}
\alias{datazone_get_glossary}
\title{Gets a business glossary in Amazon DataZone}
\usage{
datazone_get_glossary(domainIdentifier, identifier)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the Amazon DataZone domain in which this business glossary
exists.}

\item{identifier}{[required] The ID of the business glossary.}
}
\description{
Gets a business glossary in Amazon DataZone.

See \url{https://www.paws-r-sdk.com/docs/datazone_get_glossary/} for full documentation.
}
\keyword{internal}
