% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_get_glossary_term}
\alias{datazone_get_glossary_term}
\title{Gets a business glossary term in Amazon DataZone}
\usage{
datazone_get_glossary_term(domainIdentifier, identifier)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the Amazon DataZone domain in which this business glossary
term exists.}

\item{identifier}{[required] The ID of the business glossary term.}
}
\description{
Gets a business glossary term in Amazon DataZone.

See \url{https://www.paws-r-sdk.com/docs/datazone_get_glossary_term/} for full documentation.
}
\keyword{internal}
