% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_get_group_profile}
\alias{datazone_get_group_profile}
\title{Gets a group profile in Amazon DataZone}
\usage{
datazone_get_group_profile(domainIdentifier, groupIdentifier)
}
\arguments{
\item{domainIdentifier}{[required] The identifier of the Amazon DataZone domain in which the group profile
exists.}

\item{groupIdentifier}{[required] The identifier of the group profile.}
}
\description{
Gets a group profile in Amazon DataZone.

See \url{https://www.paws-r-sdk.com/docs/datazone_get_group_profile/} for full documentation.
}
\keyword{internal}
