% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_get_iam_portal_login_url}
\alias{datazone_get_iam_portal_login_url}
\title{Gets the data portal URL for the specified Amazon DataZone domain}
\usage{
datazone_get_iam_portal_login_url(domainIdentifier)
}
\arguments{
\item{domainIdentifier}{[required] the ID of the Amazon DataZone domain the data portal of which you want
to get.}
}
\description{
Gets the data portal URL for the specified Amazon DataZone domain.

See \url{https://www.paws-r-sdk.com/docs/datazone_get_iam_portal_login_url/} for full documentation.
}
\keyword{internal}
