% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_get_job_run}
\alias{datazone_get_job_run}
\title{The details of the job run}
\usage{
datazone_get_job_run(domainIdentifier, identifier)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the domain.}

\item{identifier}{[required] The ID of the job run.}
}
\description{
The details of the job run.

See \url{https://www.paws-r-sdk.com/docs/datazone_get_job_run/} for full documentation.
}
\keyword{internal}
