% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_get_lineage_node}
\alias{datazone_get_lineage_node}
\title{Gets the data lineage node}
\usage{
datazone_get_lineage_node(domainIdentifier, eventTimestamp = NULL, identifier)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the domain in which you want to get the data lineage node.}

\item{eventTimestamp}{The event time stamp for which you want to get the data lineage node.}

\item{identifier}{[required] The ID of the data lineage node that you want to get.

Both, a lineage node identifier generated by Amazon DataZone and a
\code{sourceIdentifier} of the lineage node are supported. If
\code{sourceIdentifier} is greater than 1800 characters, you can use lineage
node identifier generated by Amazon DataZone to get the node details.}
}
\description{
Gets the data lineage node.

See \url{https://www.paws-r-sdk.com/docs/datazone_get_lineage_node/} for full documentation.
}
\keyword{internal}
