% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_get_listing}
\alias{datazone_get_listing}
\title{Get listing}
\usage{
datazone_get_listing(domainIdentifier, identifier, listingRevision = NULL)
}
\arguments{
\item{domainIdentifier}{[required]}

\item{identifier}{[required]}

\item{listingRevision}{}
}
\description{
Get listing

See \url{https://www.paws-r-sdk.com/docs/datazone_get_listing/} for full documentation.
}
\keyword{internal}
