% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_get_notebook}
\alias{datazone_get_notebook}
\title{Gets the details of a notebook in Amazon SageMaker Unified Studio}
\usage{
datazone_get_notebook(domainIdentifier, identifier)
}
\arguments{
\item{domainIdentifier}{[required] The identifier of the Amazon SageMaker Unified Studio domain in which the notebook exists.}

\item{identifier}{[required] The identifier of the notebook.}
}
\description{
Gets the details of a \href{https://docs.aws.amazon.com/sagemaker-unified-studio/latest/userguide/notebooks.html}{notebook} in Amazon SageMaker Unified Studio.

See \url{https://www.paws-r-sdk.com/docs/datazone_get_notebook/} for full documentation.
}
\keyword{internal}
