% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_get_notebook_export}
\alias{datazone_get_notebook_export}
\title{Gets the details of a notebook export in Amazon SageMaker Unified Studio}
\usage{
datazone_get_notebook_export(domainIdentifier, identifier)
}
\arguments{
\item{domainIdentifier}{[required] The identifier of the Amazon SageMaker Unified Studio domain in which the notebook export exists.}

\item{identifier}{[required] The identifier of the notebook export.}
}
\description{
Gets the details of a notebook export in Amazon SageMaker Unified Studio.

See \url{https://www.paws-r-sdk.com/docs/datazone_get_notebook_export/} for full documentation.
}
\keyword{internal}
