% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_get_project}
\alias{datazone_get_project}
\title{Gets a project in Amazon DataZone}
\usage{
datazone_get_project(domainIdentifier, identifier)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the Amazon DataZone domain in which the project exists.}

\item{identifier}{[required] The ID of the project.}
}
\description{
Gets a project in Amazon DataZone.

See \url{https://www.paws-r-sdk.com/docs/datazone_get_project/} for full documentation.
}
\keyword{internal}
