% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_get_project_profile}
\alias{datazone_get_project_profile}
\title{The details of the project profile}
\usage{
datazone_get_project_profile(domainIdentifier, identifier)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the domain.}

\item{identifier}{[required] The ID of the project profile.}
}
\description{
The details of the project profile.

See \url{https://www.paws-r-sdk.com/docs/datazone_get_project_profile/} for full documentation.
}
\keyword{internal}
