% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_get_rule}
\alias{datazone_get_rule}
\title{Gets the details of a rule in Amazon DataZone}
\usage{
datazone_get_rule(domainIdentifier, identifier, revision = NULL)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the domain where the \code{\link[=datazone_get_rule]{get_rule}} action is
to be invoked.}

\item{identifier}{[required] The ID of the rule.}

\item{revision}{The revision of the rule.}
}
\description{
Gets the details of a rule in Amazon DataZone. A rule is a formal agreement that enforces specific requirements across user workflows (e.g., publishing assets to the catalog, requesting subscriptions, creating projects) within the Amazon DataZone data portal. These rules help maintain consistency, ensure compliance, and uphold governance standards in data management processes. For instance, a metadata enforcement rule can specify the required information for creating a subscription request or publishing a data asset to the catalog, ensuring alignment with organizational standards.

See \url{https://www.paws-r-sdk.com/docs/datazone_get_rule/} for full documentation.
}
\keyword{internal}
