% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_get_subscription}
\alias{datazone_get_subscription}
\title{Gets a subscription in Amazon DataZone}
\usage{
datazone_get_subscription(domainIdentifier, identifier)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the Amazon DataZone domain in which the subscription exists.}

\item{identifier}{[required] The ID of the subscription.}
}
\description{
Gets a subscription in Amazon DataZone.

See \url{https://www.paws-r-sdk.com/docs/datazone_get_subscription/} for full documentation.
}
\keyword{internal}
