% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_get_subscription_grant}
\alias{datazone_get_subscription_grant}
\title{Gets the subscription grant in Amazon DataZone}
\usage{
datazone_get_subscription_grant(domainIdentifier, identifier)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the Amazon DataZone domain in which the subscription grant
exists.}

\item{identifier}{[required] The ID of the subscription grant.}
}
\description{
Gets the subscription grant in Amazon DataZone.

See \url{https://www.paws-r-sdk.com/docs/datazone_get_subscription_grant/} for full documentation.
}
\keyword{internal}
