% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_get_subscription_request_details}
\alias{datazone_get_subscription_request_details}
\title{Gets the details of the specified subscription request}
\usage{
datazone_get_subscription_request_details(domainIdentifier, identifier)
}
\arguments{
\item{domainIdentifier}{[required] The identifier of the Amazon DataZone domain in which to get the
subscription request details.}

\item{identifier}{[required] The identifier of the subscription request the details of which to get.}
}
\description{
Gets the details of the specified subscription request.

See \url{https://www.paws-r-sdk.com/docs/datazone_get_subscription_request_details/} for full documentation.
}
\keyword{internal}
