% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_get_subscription_target}
\alias{datazone_get_subscription_target}
\title{Gets the subscription target in Amazon DataZone}
\usage{
datazone_get_subscription_target(
  domainIdentifier,
  environmentIdentifier,
  identifier
)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the Amazon DataZone domain in which the subscription target
exists.}

\item{environmentIdentifier}{[required] The ID of the environment associated with the subscription target.}

\item{identifier}{[required] The ID of the subscription target.}
}
\description{
Gets the subscription target in Amazon DataZone.

See \url{https://www.paws-r-sdk.com/docs/datazone_get_subscription_target/} for full documentation.
}
\keyword{internal}
