% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_get_time_series_data_point}
\alias{datazone_get_time_series_data_point}
\title{Gets the existing data point for the asset}
\usage{
datazone_get_time_series_data_point(
  domainIdentifier,
  entityIdentifier,
  entityType,
  formName,
  identifier
)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the Amazon DataZone domain that houses the asset for which you
want to get the data point.}

\item{entityIdentifier}{[required] The ID of the asset for which you want to get the data point.}

\item{entityType}{[required] The type of the asset for which you want to get the data point.}

\item{formName}{[required] The name of the time series form that houses the data point that you
want to get.}

\item{identifier}{[required] The ID of the data point that you want to get.}
}
\description{
Gets the existing data point for the asset.

See \url{https://www.paws-r-sdk.com/docs/datazone_get_time_series_data_point/} for full documentation.
}
\keyword{internal}
