% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_list_account_pools}
\alias{datazone_list_account_pools}
\title{Lists existing account pools}
\usage{
datazone_list_account_pools(
  domainIdentifier,
  maxResults = NULL,
  name = NULL,
  nextToken = NULL,
  sortBy = NULL,
  sortOrder = NULL
)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the domain where exsting account pools are to be listed.}

\item{maxResults}{The maximum number of account pools to return in a single call to
ListAccountPools. When the number of account pools to be listed is
greater than the value of MaxResults, the response contains a NextToken
value that you can use in a subsequent call to ListAccountPools to list
the next set of account pools.}

\item{name}{The name of the account pool to be listed.}

\item{nextToken}{When the number of account pools is greater than the default value for
the MaxResults parameter, or if you explicitly specify a value for
MaxResults that is less than the number of account pools, the response
includes a pagination token named NextToken. You can specify this
NextToken value in a subsequent call to ListAccountPools to list the
next set of account pools.}

\item{sortBy}{The sort by mechanism in which the existing account pools are to be
listed.}

\item{sortOrder}{The sort order in which the existing account pools are to be listed.}
}
\description{
Lists existing account pools.

See \url{https://www.paws-r-sdk.com/docs/datazone_list_account_pools/} for full documentation.
}
\keyword{internal}
