% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_list_asset_revisions}
\alias{datazone_list_asset_revisions}
\title{Lists the revisions for the asset}
\usage{
datazone_list_asset_revisions(
  domainIdentifier,
  identifier,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{domainIdentifier}{[required] The identifier of the domain.}

\item{identifier}{[required] The identifier of the asset.}

\item{maxResults}{The maximum number of revisions to return in a single call to
\code{\link[=datazone_list_asset_revisions]{list_asset_revisions}}. When the number
of revisions to be listed is greater than the value of \code{MaxResults}, the
response contains a \code{NextToken} value that you can use in a subsequent
call to \code{\link[=datazone_list_asset_revisions]{list_asset_revisions}} to list
the next set of revisions.}

\item{nextToken}{When the number of revisions is greater than the default value for the
\code{MaxResults} parameter, or if you explicitly specify a value for
\code{MaxResults} that is less than the number of revisions, the response
includes a pagination token named \code{NextToken}. You can specify this
\code{NextToken} value in a subsequent call to
\code{\link[=datazone_list_asset_revisions]{list_asset_revisions}} to list the next
set of revisions.}
}
\description{
Lists the revisions for the asset.

See \url{https://www.paws-r-sdk.com/docs/datazone_list_asset_revisions/} for full documentation.
}
\keyword{internal}
