% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_list_data_product_revisions}
\alias{datazone_list_data_product_revisions}
\title{Lists data product revisions}
\usage{
datazone_list_data_product_revisions(
  domainIdentifier,
  identifier,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the domain of the data product revisions that you want to
list.}

\item{identifier}{[required] The ID of the data product revision.}

\item{maxResults}{The maximum number of asset filters to return in a single call to
\code{\link[=datazone_list_data_product_revisions]{list_data_product_revisions}}.
When the number of data product revisions to be listed is greater than
the value of \code{MaxResults}, the response contains a \code{NextToken} value
that you can use in a subsequent call to
\code{\link[=datazone_list_data_product_revisions]{list_data_product_revisions}} to
list the next set of data product revisions.}

\item{nextToken}{When the number of data product revisions is greater than the default
value for the \code{MaxResults} parameter, or if you explicitly specify a
value for \code{MaxResults} that is less than the number of data product
revisions, the response includes a pagination token named \code{NextToken}.
You can specify this \code{NextToken} value in a subsequent call to
\code{\link[=datazone_list_data_product_revisions]{list_data_product_revisions}} to
list the next set of data product revisions.}
}
\description{
Lists data product revisions.

See \url{https://www.paws-r-sdk.com/docs/datazone_list_data_product_revisions/} for full documentation.
}
\keyword{internal}
