% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_list_data_source_run_activities}
\alias{datazone_list_data_source_run_activities}
\title{Lists data source run activities}
\usage{
datazone_list_data_source_run_activities(
  domainIdentifier,
  identifier,
  maxResults = NULL,
  nextToken = NULL,
  status = NULL
)
}
\arguments{
\item{domainIdentifier}{[required] The identifier of the Amazon DataZone domain in which to list data
source run activities.}

\item{identifier}{[required] The identifier of the data source run.}

\item{maxResults}{The maximum number of activities to return in a single call to
\code{\link[=datazone_list_data_source_run_activities]{list_data_source_run_activities}}.
When the number of activities to be listed is greater than the value of
\code{MaxResults}, the response contains a \code{NextToken} value that you can use
in a subsequent call to
\code{\link[=datazone_list_data_source_run_activities]{list_data_source_run_activities}}
to list the next set of activities.}

\item{nextToken}{When the number of activities is greater than the default value for the
\code{MaxResults} parameter, or if you explicitly specify a value for
\code{MaxResults} that is less than the number of activities, the response
includes a pagination token named \code{NextToken}. You can specify this
\code{NextToken} value in a subsequent call to
\code{\link[=datazone_list_data_source_run_activities]{list_data_source_run_activities}}
to list the next set of activities.}

\item{status}{The status of the data source run.}
}
\description{
Lists data source run activities.

See \url{https://www.paws-r-sdk.com/docs/datazone_list_data_source_run_activities/} for full documentation.
}
\keyword{internal}
