% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_list_domain_units_for_parent}
\alias{datazone_list_domain_units_for_parent}
\title{Lists child domain units for the specified parent domain unit}
\usage{
datazone_list_domain_units_for_parent(
  domainIdentifier,
  maxResults = NULL,
  nextToken = NULL,
  parentDomainUnitIdentifier
)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the domain in which you want to list domain units for a parent
domain unit.}

\item{maxResults}{The maximum number of domain units to return in a single call to
ListDomainUnitsForParent. When the number of domain units to be listed
is greater than the value of MaxResults, the response contains a
NextToken value that you can use in a subsequent call to
ListDomainUnitsForParent to list the next set of domain units.}

\item{nextToken}{When the number of domain units is greater than the default value for
the MaxResults parameter, or if you explicitly specify a value for
MaxResults that is less than the number of domain units, the response
includes a pagination token named NextToken. You can specify this
NextToken value in a subsequent call to ListDomainUnitsForParent to list
the next set of domain units.}

\item{parentDomainUnitIdentifier}{[required] The ID of the parent domain unit.}
}
\description{
Lists child domain units for the specified parent domain unit.

See \url{https://www.paws-r-sdk.com/docs/datazone_list_domain_units_for_parent/} for full documentation.
}
\keyword{internal}
