% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_list_domains}
\alias{datazone_list_domains}
\title{Lists Amazon DataZone domains}
\usage{
datazone_list_domains(maxResults = NULL, nextToken = NULL, status = NULL)
}
\arguments{
\item{maxResults}{The maximum number of domains to return in a single call to
\code{\link[=datazone_list_domains]{list_domains}}. When the number of domains to
be listed is greater than the value of \code{MaxResults}, the response
contains a \code{NextToken} value that you can use in a subsequent call to
\code{\link[=datazone_list_domains]{list_domains}} to list the next set of domains.}

\item{nextToken}{When the number of domains is greater than the default value for the
\code{MaxResults} parameter, or if you explicitly specify a value for
\code{MaxResults} that is less than the number of domains, the response
includes a pagination token named \code{NextToken}. You can specify this
\code{NextToken} value in a subsequent call to
\code{\link[=datazone_list_domains]{list_domains}} to list the next set of domains.}

\item{status}{The status of the data source.}
}
\description{
Lists Amazon DataZone domains.

See \url{https://www.paws-r-sdk.com/docs/datazone_list_domains/} for full documentation.
}
\keyword{internal}
