% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_list_metadata_generation_runs}
\alias{datazone_list_metadata_generation_runs}
\title{Lists all metadata generation runs}
\usage{
datazone_list_metadata_generation_runs(
  domainIdentifier,
  maxResults = NULL,
  nextToken = NULL,
  status = NULL,
  targetIdentifier = NULL,
  type = NULL
)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the Amazon DataZone domain where you want to list metadata
generation runs.}

\item{maxResults}{The maximum number of metadata generation runs to return in a single
call to ListMetadataGenerationRuns. When the number of metadata
generation runs to be listed is greater than the value of MaxResults,
the response contains a NextToken value that you can use in a subsequent
call to ListMetadataGenerationRuns to list the next set of revisions.}

\item{nextToken}{When the number of metadata generation runs is greater than the default
value for the MaxResults parameter, or if you explicitly specify a value
for MaxResults that is less than the number of metadata generation runs,
the response includes a pagination token named NextToken. You can
specify this NextToken value in a subsequent call to
ListMetadataGenerationRuns to list the next set of revisions.}

\item{status}{The status of the metadata generation runs.}

\item{targetIdentifier}{The target ID for which you want to list metadata generation runs.}

\item{type}{The type of the metadata generation runs.}
}
\description{
Lists all metadata generation runs.

See \url{https://www.paws-r-sdk.com/docs/datazone_list_metadata_generation_runs/} for full documentation.
}
\keyword{internal}
