% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_list_notebook_runs}
\alias{datazone_list_notebook_runs}
\title{Lists notebook runs in an Amazon DataZone domain}
\usage{
datazone_list_notebook_runs(
  domainIdentifier,
  owningProjectIdentifier,
  notebookIdentifier = NULL,
  status = NULL,
  scheduleIdentifier = NULL,
  maxResults = NULL,
  sortOrder = NULL,
  nextToken = NULL
)
}
\arguments{
\item{domainIdentifier}{[required] The identifier of the Amazon DataZone domain in which to list notebook
runs.}

\item{owningProjectIdentifier}{[required] The identifier of the project that owns the notebook runs.}

\item{notebookIdentifier}{The identifier of the notebook to filter runs by.}

\item{status}{The status to filter notebook runs by.}

\item{scheduleIdentifier}{The identifier of the schedule to filter notebook runs by.}

\item{maxResults}{The maximum number of notebook runs to return in a single call. When the
number of notebook runs exceeds the value of \code{MaxResults}, the response
contains a \code{NextToken} value.}

\item{sortOrder}{The sort order for the results.}

\item{nextToken}{When the number of notebook runs is greater than the default value for
the \code{MaxResults} parameter, or if you explicitly specify a value for
\code{MaxResults} that is less than the number of notebook runs, the response
includes a pagination token named \code{NextToken}. You can specify this
\code{NextToken} value in a subsequent call to
\code{\link[=datazone_list_notebook_runs]{list_notebook_runs}} to list the next set
of notebook runs.}
}
\description{
Lists notebook runs in an Amazon DataZone domain.

See \url{https://www.paws-r-sdk.com/docs/datazone_list_notebook_runs/} for full documentation.
}
\keyword{internal}
