% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_list_notebooks}
\alias{datazone_list_notebooks}
\title{Lists notebooks in Amazon SageMaker Unified Studio}
\usage{
datazone_list_notebooks(
  domainIdentifier,
  owningProjectIdentifier,
  maxResults = NULL,
  sortOrder = NULL,
  sortBy = NULL,
  status = NULL,
  type = NULL,
  nextToken = NULL
)
}
\arguments{
\item{domainIdentifier}{[required] The identifier of the Amazon SageMaker Unified Studio domain in which to list notebooks.}

\item{owningProjectIdentifier}{[required] The identifier of the project that owns the notebooks.}

\item{maxResults}{The maximum number of notebooks to return in a single call. When the number of notebooks exceeds the value of \code{MaxResults}, the response contains a \code{NextToken} value.}

\item{sortOrder}{The sort order for the results.}

\item{sortBy}{The field to sort the results by.}

\item{status}{The status to filter notebooks by.}

\item{type}{The type to filter notebooks by.}

\item{nextToken}{When the number of notebooks is greater than the default value for the \code{MaxResults} parameter, or if you explicitly specify a value for \code{MaxResults} that is less than the number of notebooks, the response includes a pagination token named \code{NextToken}. You can specify this \code{NextToken} value in a subsequent call to \code{\link[=datazone_list_notebooks]{list_notebooks}} to list the next set of notebooks.}
}
\description{
Lists \href{https://docs.aws.amazon.com/sagemaker-unified-studio/latest/userguide/notebooks.html}{notebooks} in Amazon SageMaker Unified Studio.

See \url{https://www.paws-r-sdk.com/docs/datazone_list_notebooks/} for full documentation.
}
\keyword{internal}
