% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_list_project_profiles}
\alias{datazone_list_project_profiles}
\title{Lists project profiles}
\usage{
datazone_list_project_profiles(
  domainIdentifier,
  maxResults = NULL,
  name = NULL,
  nextToken = NULL,
  sortBy = NULL,
  sortOrder = NULL
)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the domain where you want to list project profiles.}

\item{maxResults}{The maximum number of project profiles to return in a single call to
ListProjectProfiles. When the number of project profiles to be listed is
greater than the value of MaxResults, the response contains a NextToken
value that you can use in a subsequent call to ListProjectProfiles to
list the next set of project profiles.}

\item{name}{The name of a project profile.}

\item{nextToken}{When the number of project profiles is greater than the default value
for the MaxResults parameter, or if you explicitly specify a value for
MaxResults that is less than the number of project profiles, the
response includes a pagination token named NextToken. You can specify
this NextToken value in a subsequent call to ListProjectProfiles to list
the next set of project profiles.}

\item{sortBy}{Specifies by what to sort project profiles.}

\item{sortOrder}{Specifies the sort order of the project profiles.}
}
\description{
Lists project profiles.

See \url{https://www.paws-r-sdk.com/docs/datazone_list_project_profiles/} for full documentation.
}
\keyword{internal}
