% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_list_tags_for_resource}
\alias{datazone_list_tags_for_resource}
\title{Lists tags for the specified resource in Amazon DataZone}
\usage{
datazone_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The ARN of the resource whose tags you want to list.}
}
\description{
Lists tags for the specified resource in Amazon DataZone.

See \url{https://www.paws-r-sdk.com/docs/datazone_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
