% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_post_lineage_event}
\alias{datazone_post_lineage_event}
\title{Posts a data lineage event}
\usage{
datazone_post_lineage_event(domainIdentifier, event, clientToken = NULL)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the domain where you want to post a data lineage event.}

\item{event}{[required] The data lineage event that you want to post. Only open-lineage run
event are supported as events.}

\item{clientToken}{A unique, case-sensitive identifier that is provided to ensure the
idempotency of the request.}
}
\description{
Posts a data lineage event.

See \url{https://www.paws-r-sdk.com/docs/datazone_post_lineage_event/} for full documentation.
}
\keyword{internal}
