% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_post_time_series_data_points}
\alias{datazone_post_time_series_data_points}
\title{Posts time series data points to Amazon DataZone for the specified asset}
\usage{
datazone_post_time_series_data_points(
  clientToken = NULL,
  domainIdentifier,
  entityIdentifier,
  entityType,
  forms
)
}
\arguments{
\item{clientToken}{A unique, case-sensitive identifier that is provided to ensure the
idempotency of the request.}

\item{domainIdentifier}{[required] The ID of the Amazon DataZone domain in which you want to post time
series data points.}

\item{entityIdentifier}{[required] The ID of the asset for which you want to post time series data points.}

\item{entityType}{[required] The type of the asset for which you want to post data points.}

\item{forms}{[required] The forms that contain the data points that you want to post.}
}
\description{
Posts time series data points to Amazon DataZone for the specified asset.

See \url{https://www.paws-r-sdk.com/docs/datazone_post_time_series_data_points/} for full documentation.
}
\keyword{internal}
