% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_put_data_export_configuration}
\alias{datazone_put_data_export_configuration}
\title{Creates data export configuration details}
\usage{
datazone_put_data_export_configuration(
  clientToken = NULL,
  domainIdentifier,
  enableExport,
  encryptionConfiguration = NULL
)
}
\arguments{
\item{clientToken}{A unique, case-sensitive identifier to ensure idempotency of the
request. This field is automatically populated if not provided.}

\item{domainIdentifier}{[required] The domain ID where you want to create data export configuration
details.}

\item{enableExport}{[required] Specifies that the export is to be enabled as part of creating data
export configuration details.}

\item{encryptionConfiguration}{The encryption configuration as part of creating data export
configuration details.

The KMS key provided here as part of encryptionConfiguration must have
the required permissions as described in \href{https://docs.aws.amazon.com/sagemaker-unified-studio/latest/adminguide/sagemaker-unified-studio-export-asset-metadata-kms-permissions.html}{KMS permissions for exporting asset metadata in Amazon SageMaker Unified Studio}.}
}
\description{
Creates data export configuration details.

See \url{https://www.paws-r-sdk.com/docs/datazone_put_data_export_configuration/} for full documentation.
}
\keyword{internal}
