% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_put_environment_blueprint_configuration}
\alias{datazone_put_environment_blueprint_configuration}
\title{Writes the configuration for the specified environment blueprint in
Amazon DataZone}
\usage{
datazone_put_environment_blueprint_configuration(
  domainIdentifier,
  enabledRegions,
  environmentBlueprintIdentifier,
  environmentRolePermissionBoundary = NULL,
  globalParameters = NULL,
  manageAccessRoleArn = NULL,
  provisioningConfigurations = NULL,
  provisioningRoleArn = NULL,
  regionalParameters = NULL
)
}
\arguments{
\item{domainIdentifier}{[required] The identifier of the Amazon DataZone domain.}

\item{enabledRegions}{[required] Specifies the enabled Amazon Web Services Regions.}

\item{environmentBlueprintIdentifier}{[required] The identifier of the environment blueprint.}

\item{environmentRolePermissionBoundary}{The environment role permissions boundary.}

\item{globalParameters}{Region-agnostic environment blueprint parameters.}

\item{manageAccessRoleArn}{The ARN of the manage access role.}

\item{provisioningConfigurations}{The provisioning configuration of a blueprint.}

\item{provisioningRoleArn}{The ARN of the provisioning role.}

\item{regionalParameters}{The regional parameters in the environment blueprint.}
}
\description{
Writes the configuration for the specified environment blueprint in Amazon DataZone.

See \url{https://www.paws-r-sdk.com/docs/datazone_put_environment_blueprint_configuration/} for full documentation.
}
\keyword{internal}
