% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_query_graph}
\alias{datazone_query_graph}
\title{Queries entities in the graph store}
\usage{
datazone_query_graph(
  domainIdentifier,
  match,
  maxResults = NULL,
  nextToken = NULL,
  additionalAttributes = NULL
)
}
\arguments{
\item{domainIdentifier}{[required] The identifier of the Amazon DataZone domain.}

\item{match}{[required] List of query match clauses.}

\item{maxResults}{The maximum number of entities to return in a single call to
\code{\link[=datazone_query_graph]{query_graph}}. When the number of entities to be
listed is greater than the value of \code{MaxResults}, the response contains
a \code{NextToken} value that you can use in a subsequent call to
\code{\link[=datazone_query_graph]{query_graph}} to list the next set of entities.}

\item{nextToken}{When the number of entities is greater than the default value for the
\code{MaxResults} parameter, or if you explicitly specify a value for
\code{MaxResults} that is less than the number of entities, the response
includes a pagination token named \code{NextToken}. You can specify this
\code{NextToken} value in a subsequent call to
\code{\link[=datazone_query_graph]{query_graph}} to list the next set of entities.}

\item{additionalAttributes}{Additional details on the queried entity that can be requested in the
response.}
}
\description{
Queries entities in the graph store.

See \url{https://www.paws-r-sdk.com/docs/datazone_query_graph/} for full documentation.
}
\keyword{internal}
