% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_reject_subscription_request}
\alias{datazone_reject_subscription_request}
\title{Rejects the specified subscription request}
\usage{
datazone_reject_subscription_request(
  domainIdentifier,
  identifier,
  decisionComment = NULL
)
}
\arguments{
\item{domainIdentifier}{[required] The identifier of the Amazon DataZone domain in which the subscription
request was rejected.}

\item{identifier}{[required] The identifier of the subscription request that was rejected.}

\item{decisionComment}{The decision comment of the rejected subscription request.}
}
\description{
Rejects the specified subscription request.

See \url{https://www.paws-r-sdk.com/docs/datazone_reject_subscription_request/} for full documentation.
}
\keyword{internal}
