% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_remove_entity_owner}
\alias{datazone_remove_entity_owner}
\title{Removes an owner from an entity}
\usage{
datazone_remove_entity_owner(
  clientToken = NULL,
  domainIdentifier,
  entityIdentifier,
  entityType,
  owner
)
}
\arguments{
\item{clientToken}{A unique, case-sensitive identifier that is provided to ensure the
idempotency of the request.}

\item{domainIdentifier}{[required] The ID of the domain where you want to remove an owner from an entity.}

\item{entityIdentifier}{[required] The ID of the entity from which you want to remove an owner.}

\item{entityType}{[required] The type of the entity from which you want to remove an owner.}

\item{owner}{[required] The owner that you want to remove from an entity.}
}
\description{
Removes an owner from an entity.

See \url{https://www.paws-r-sdk.com/docs/datazone_remove_entity_owner/} for full documentation.
}
\keyword{internal}
