% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_remove_policy_grant}
\alias{datazone_remove_policy_grant}
\title{Removes a policy grant}
\usage{
datazone_remove_policy_grant(
  clientToken = NULL,
  domainIdentifier,
  entityIdentifier,
  entityType,
  policyType,
  principal
)
}
\arguments{
\item{clientToken}{A unique, case-sensitive identifier that is provided to ensure the
idempotency of the request.}

\item{domainIdentifier}{[required] The ID of the domain where you want to remove a policy grant.}

\item{entityIdentifier}{[required] The ID of the entity from which you want to remove a policy grant.}

\item{entityType}{[required] The type of the entity from which you want to remove a policy grant.}

\item{policyType}{[required] The type of the policy that you want to remove.}

\item{principal}{[required] The principal from which you want to remove a policy grant.}
}
\description{
Removes a policy grant.

See \url{https://www.paws-r-sdk.com/docs/datazone_remove_policy_grant/} for full documentation.
}
\keyword{internal}
