% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_start_notebook_export}
\alias{datazone_start_notebook_export}
\title{Starts a notebook export in Amazon SageMaker Unified Studio}
\usage{
datazone_start_notebook_export(
  domainIdentifier,
  notebookIdentifier,
  owningProjectIdentifier,
  fileFormat,
  clientToken = NULL
)
}
\arguments{
\item{domainIdentifier}{[required] The identifier of the Amazon SageMaker Unified Studio domain in which to export the notebook.}

\item{notebookIdentifier}{[required] The identifier of the notebook to export.}

\item{owningProjectIdentifier}{[required] The identifier of the project that owns the notebook.}

\item{fileFormat}{[required] The file format for the notebook export. Valid values are \code{PDF} and \code{IPYNB}.}

\item{clientToken}{A unique, case-sensitive identifier to ensure idempotency of the request. This field is automatically populated if not provided.}
}
\description{
Starts a notebook export in Amazon SageMaker Unified Studio. This operation exports a notebook to a specified file format and stores the output in Amazon Simple Storage Service.

See \url{https://www.paws-r-sdk.com/docs/datazone_start_notebook_export/} for full documentation.
}
\keyword{internal}
