% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_start_notebook_import}
\alias{datazone_start_notebook_import}
\title{Starts a notebook import in Amazon SageMaker Unified Studio}
\usage{
datazone_start_notebook_import(
  domainIdentifier,
  owningProjectIdentifier,
  sourceLocation,
  name,
  description = NULL,
  clientToken = NULL
)
}
\arguments{
\item{domainIdentifier}{[required] The identifier of the Amazon SageMaker Unified Studio domain in which to import the notebook.}

\item{owningProjectIdentifier}{[required] The identifier of the project that will own the imported notebook.}

\item{sourceLocation}{[required] The source location of the notebook to import. This specifies the Amazon Simple Storage Service URI of the notebook file.}

\item{name}{[required] The name of the imported notebook. The name must be between 1 and 256 characters.}

\item{description}{The description of the imported notebook.}

\item{clientToken}{A unique, case-sensitive identifier to ensure idempotency of the request. This field is automatically populated if not provided.}
}
\description{
Starts a notebook import in Amazon SageMaker Unified Studio. This operation imports a notebook from an Amazon Simple Storage Service location into a project.

See \url{https://www.paws-r-sdk.com/docs/datazone_start_notebook_import/} for full documentation.
}
\keyword{internal}
