% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_start_notebook_sync}
\alias{datazone_start_notebook_sync}
\title{Starts a notebook sync in Amazon SageMaker Unified Studio}
\usage{
datazone_start_notebook_sync(
  domainIdentifier,
  owningProjectIdentifier,
  sourceLocation,
  gitMetadata = NULL,
  notebookId = NULL,
  name = NULL,
  description = NULL,
  clientToken = NULL
)
}
\arguments{
\item{domainIdentifier}{[required] The identifier of the Amazon SageMaker Unified Studio domain in which to sync the notebook.}

\item{owningProjectIdentifier}{[required] The identifier of the project that will own the synced notebook.}

\item{sourceLocation}{[required] The source location of the notebook to sync. This specifies the Amazon Simple Storage Service URI of the notebook file.}

\item{gitMetadata}{The Git metadata for the notebook sync, including repository, branch, and commit information.}

\item{notebookId}{The identifier of an existing notebook to sync. If not specified, a new notebook is created.}

\item{name}{The name of the notebook. The name must be between 1 and 256 characters.}

\item{description}{The description of the notebook.}

\item{clientToken}{A unique, case-sensitive identifier to ensure idempotency of the request. This field is automatically populated if not provided.}
}
\description{
Starts a notebook sync in Amazon SageMaker Unified Studio. This operation syncs a notebook from a Git repository into a project.

See \url{https://www.paws-r-sdk.com/docs/datazone_start_notebook_sync/} for full documentation.
}
\keyword{internal}
