% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_stop_notebook_run}
\alias{datazone_stop_notebook_run}
\title{Stops a running notebook run in an Amazon DataZone domain}
\usage{
datazone_stop_notebook_run(domainIdentifier, identifier, clientToken = NULL)
}
\arguments{
\item{domainIdentifier}{[required] The identifier of the Amazon DataZone domain in which the notebook run
is stopped.}

\item{identifier}{[required] The identifier of the notebook run to stop.}

\item{clientToken}{A unique, case-sensitive identifier to ensure idempotency of the
request. This field is automatically populated if not provided.}
}
\description{
Stops a running notebook run in an Amazon DataZone domain.

See \url{https://www.paws-r-sdk.com/docs/datazone_stop_notebook_run/} for full documentation.
}
\keyword{internal}
