% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_tag_resource}
\alias{datazone_tag_resource}
\title{Tags a resource in Amazon DataZone}
\usage{
datazone_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The ARN of the resource to be tagged in Amazon DataZone.}

\item{tags}{[required] Specifies the tags for the \code{\link[=datazone_tag_resource]{tag_resource}}
action.}
}
\description{
Tags a resource in Amazon DataZone.

See \url{https://www.paws-r-sdk.com/docs/datazone_tag_resource/} for full documentation.
}
\keyword{internal}
