% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_untag_resource}
\alias{datazone_untag_resource}
\title{Untags a resource in Amazon DataZone}
\usage{
datazone_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The ARN of the resource to be untagged in Amazon DataZone.}

\item{tagKeys}{[required] Specifies the tag keys for the
\code{\link[=datazone_untag_resource]{untag_resource}} action.}
}
\description{
Untags a resource in Amazon DataZone.

See \url{https://www.paws-r-sdk.com/docs/datazone_untag_resource/} for full documentation.
}
\keyword{internal}
