% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_update_account_pool}
\alias{datazone_update_account_pool}
\title{Updates the account pool}
\usage{
datazone_update_account_pool(
  domainIdentifier,
  identifier,
  name = NULL,
  description = NULL,
  resolutionStrategy = NULL,
  accountSource = NULL
)
}
\arguments{
\item{domainIdentifier}{[required] The domain ID where the account pool that is to be updated lives.}

\item{identifier}{[required] The ID of the account pool that is to be updated.}

\item{name}{The name of the account pool that is to be updated.}

\item{description}{The description of the account pool that is to be udpated.}

\item{resolutionStrategy}{The mechanism used to resolve the account selection from the account
pool.}

\item{accountSource}{The source of accounts for the account pool. In the current release,
it's either a static list of accounts provided by the customer or a
custom Amazon Web Services Lambda handler.}
}
\description{
Updates the account pool.

See \url{https://www.paws-r-sdk.com/docs/datazone_update_account_pool/} for full documentation.
}
\keyword{internal}
