% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_update_connection}
\alias{datazone_update_connection}
\title{Updates a connection}
\usage{
datazone_update_connection(
  awsLocation = NULL,
  description = NULL,
  domainIdentifier,
  identifier,
  props = NULL
)
}
\arguments{
\item{awsLocation}{The location where a connection is to be updated.}

\item{description}{The description of a connection.}

\item{domainIdentifier}{[required] The ID of the domain where a connection is to be updated.}

\item{identifier}{[required] The ID of the connection to be updated.}

\item{props}{The connection props.}
}
\description{
Updates a connection. In Amazon DataZone, a connection enables you to connect your resources (domains, projects, and environments) to external resources and services.

See \url{https://www.paws-r-sdk.com/docs/datazone_update_connection/} for full documentation.
}
\keyword{internal}
