% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_update_domain}
\alias{datazone_update_domain}
\title{Updates a Amazon DataZone domain}
\usage{
datazone_update_domain(
  clientToken = NULL,
  description = NULL,
  domainExecutionRole = NULL,
  identifier,
  name = NULL,
  singleSignOn = NULL
)
}
\arguments{
\item{clientToken}{A unique, case-sensitive identifier that is provided to ensure the
idempotency of the request.}

\item{description}{The description to be updated as part of the
\code{\link[=datazone_update_domain]{update_domain}} action.}

\item{domainExecutionRole}{The domain execution role to be updated as part of the
\code{\link[=datazone_update_domain]{update_domain}} action.}

\item{identifier}{[required] The ID of the Amazon Web Services domain that is to be updated.}

\item{name}{The name to be updated as part of the
\code{\link[=datazone_update_domain]{update_domain}} action.}

\item{singleSignOn}{The single sign-on option to be updated as part of the
\code{\link[=datazone_update_domain]{update_domain}} action.}
}
\description{
Updates a Amazon DataZone domain.

See \url{https://www.paws-r-sdk.com/docs/datazone_update_domain/} for full documentation.
}
\keyword{internal}
