% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_update_domain_unit}
\alias{datazone_update_domain_unit}
\title{Updates the domain unit}
\usage{
datazone_update_domain_unit(
  description = NULL,
  domainIdentifier,
  identifier,
  name = NULL
)
}
\arguments{
\item{description}{The description of the domain unit that you want to update.}

\item{domainIdentifier}{[required] The ID of the domain where you want to update a domain unit.}

\item{identifier}{[required] The ID of the domain unit that you want to update.}

\item{name}{The name of the domain unit that you want to update.}
}
\description{
Updates the domain unit.

See \url{https://www.paws-r-sdk.com/docs/datazone_update_domain_unit/} for full documentation.
}
\keyword{internal}
