% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_update_environment}
\alias{datazone_update_environment}
\title{Updates the specified environment in Amazon DataZone}
\usage{
datazone_update_environment(
  domainIdentifier,
  identifier,
  name = NULL,
  description = NULL,
  glossaryTerms = NULL,
  blueprintVersion = NULL,
  userParameters = NULL,
  environmentConfigurationName = NULL
)
}
\arguments{
\item{domainIdentifier}{[required] The identifier of the domain in which the environment is to be updated.}

\item{identifier}{[required] The identifier of the environment that is to be updated.}

\item{name}{The name to be updated as part of the
\code{\link[=datazone_update_environment]{update_environment}} action.}

\item{description}{The description to be updated as part of the
\code{\link[=datazone_update_environment]{update_environment}} action.}

\item{glossaryTerms}{The glossary terms to be updated as part of the
\code{\link[=datazone_update_environment]{update_environment}} action.}

\item{blueprintVersion}{The blueprint version to which the environment should be updated. You
can only specify the following string for this parameter: \code{latest}.}

\item{userParameters}{The user parameters of the environment.}

\item{environmentConfigurationName}{The configuration name of the environment.}
}
\description{
Updates the specified environment in Amazon DataZone.

See \url{https://www.paws-r-sdk.com/docs/datazone_update_environment/} for full documentation.
}
\keyword{internal}
