% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_update_environment_action}
\alias{datazone_update_environment_action}
\title{Updates an environment action}
\usage{
datazone_update_environment_action(
  description = NULL,
  domainIdentifier,
  environmentIdentifier,
  identifier,
  name = NULL,
  parameters = NULL
)
}
\arguments{
\item{description}{The description of the environment action.}

\item{domainIdentifier}{[required] The domain ID of the environment action.}

\item{environmentIdentifier}{[required] The environment ID of the environment action.}

\item{identifier}{[required] The ID of the environment action.}

\item{name}{The name of the environment action.}

\item{parameters}{The parameters of the environment action.}
}
\description{
Updates an environment action.

See \url{https://www.paws-r-sdk.com/docs/datazone_update_environment_action/} for full documentation.
}
\keyword{internal}
