% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_update_environment_profile}
\alias{datazone_update_environment_profile}
\title{Updates the specified environment profile in Amazon DataZone}
\usage{
datazone_update_environment_profile(
  awsAccountId = NULL,
  awsAccountRegion = NULL,
  description = NULL,
  domainIdentifier,
  identifier,
  name = NULL,
  userParameters = NULL
)
}
\arguments{
\item{awsAccountId}{The Amazon Web Services account in which a specified environment profile
is to be udpated.}

\item{awsAccountRegion}{The Amazon Web Services Region in which a specified environment profile
is to be updated.}

\item{description}{The description to be updated as part of the
\code{\link[=datazone_update_environment_profile]{update_environment_profile}}
action.}

\item{domainIdentifier}{[required] The identifier of the Amazon DataZone domain in which an environment
profile is to be updated.}

\item{identifier}{[required] The identifier of the environment profile that is to be updated.}

\item{name}{The name to be updated as part of the
\code{\link[=datazone_update_environment_profile]{update_environment_profile}}
action.}

\item{userParameters}{The user parameters to be updated as part of the
\code{\link[=datazone_update_environment_profile]{update_environment_profile}}
action.}
}
\description{
Updates the specified environment profile in Amazon DataZone.

See \url{https://www.paws-r-sdk.com/docs/datazone_update_environment_profile/} for full documentation.
}
\keyword{internal}
