% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_update_glossary}
\alias{datazone_update_glossary}
\title{Updates the business glossary in Amazon DataZone}
\usage{
datazone_update_glossary(
  clientToken = NULL,
  description = NULL,
  domainIdentifier,
  identifier,
  name = NULL,
  status = NULL
)
}
\arguments{
\item{clientToken}{A unique, case-sensitive identifier that is provided to ensure the
idempotency of the request.}

\item{description}{The description to be updated as part of the
\code{\link[=datazone_update_glossary]{update_glossary}} action.}

\item{domainIdentifier}{[required] The identifier of the Amazon DataZone domain in which a business
glossary is to be updated.}

\item{identifier}{[required] The identifier of the business glossary to be updated.}

\item{name}{The name to be updated as part of the
\code{\link[=datazone_update_glossary]{update_glossary}} action.}

\item{status}{The status to be updated as part of the
\code{\link[=datazone_update_glossary]{update_glossary}} action.}
}
\description{
Updates the business glossary in Amazon DataZone.

See \url{https://www.paws-r-sdk.com/docs/datazone_update_glossary/} for full documentation.
}
\keyword{internal}
