% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_update_group_profile}
\alias{datazone_update_group_profile}
\title{Updates the specified group profile in Amazon DataZone}
\usage{
datazone_update_group_profile(domainIdentifier, groupIdentifier, status)
}
\arguments{
\item{domainIdentifier}{[required] The identifier of the Amazon DataZone domain in which a group profile is
updated.}

\item{groupIdentifier}{[required] The identifier of the group profile that is updated.}

\item{status}{[required] The status of the group profile that is updated.}
}
\description{
Updates the specified group profile in Amazon DataZone.

See \url{https://www.paws-r-sdk.com/docs/datazone_update_group_profile/} for full documentation.
}
\keyword{internal}
