% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_update_project}
\alias{datazone_update_project}
\title{Updates the specified project in Amazon DataZone}
\usage{
datazone_update_project(
  domainIdentifier,
  identifier,
  name = NULL,
  description = NULL,
  resourceTags = NULL,
  glossaryTerms = NULL,
  domainUnitId = NULL,
  environmentDeploymentDetails = NULL,
  userParameters = NULL,
  projectProfileVersion = NULL
)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the Amazon DataZone domain where a project is being updated.}

\item{identifier}{[required] The identifier of the project that is to be updated.}

\item{name}{The name to be updated as part of the
\code{\link[=datazone_update_project]{update_project}} action.}

\item{description}{The description to be updated as part of the
\code{\link[=datazone_update_project]{update_project}} action.}

\item{resourceTags}{The resource tags of the project.}

\item{glossaryTerms}{The glossary terms to be updated as part of the
\code{\link[=datazone_update_project]{update_project}} action.}

\item{domainUnitId}{The ID of the domain unit.}

\item{environmentDeploymentDetails}{The environment deployment details of the project.}

\item{userParameters}{The user parameters of the project.}

\item{projectProfileVersion}{The project profile version to which the project should be updated. You
can only specify the following string for this parameter: \code{latest}.}
}
\description{
Updates the specified project in Amazon DataZone.

See \url{https://www.paws-r-sdk.com/docs/datazone_update_project/} for full documentation.
}
\keyword{internal}
