% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_update_root_domain_unit_owner}
\alias{datazone_update_root_domain_unit_owner}
\title{Updates the owner of the root domain unit}
\usage{
datazone_update_root_domain_unit_owner(
  clientToken = NULL,
  currentOwner,
  domainIdentifier,
  newOwner
)
}
\arguments{
\item{clientToken}{A unique, case-sensitive identifier to ensure idempotency of the
request. This field is automatically populated if not provided.}

\item{currentOwner}{[required] The current owner of the root domain unit.}

\item{domainIdentifier}{[required] The ID of the domain where the root domain unit owner is to be updated.}

\item{newOwner}{[required] The new owner of the root domain unit.}
}
\description{
Updates the owner of the root domain unit.

See \url{https://www.paws-r-sdk.com/docs/datazone_update_root_domain_unit_owner/} for full documentation.
}
\keyword{internal}
