% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_update_subscription_grant_status}
\alias{datazone_update_subscription_grant_status}
\title{Updates the status of the specified subscription grant status in Amazon
DataZone}
\usage{
datazone_update_subscription_grant_status(
  assetIdentifier,
  domainIdentifier,
  failureCause = NULL,
  identifier,
  status,
  targetName = NULL
)
}
\arguments{
\item{assetIdentifier}{[required] The identifier of the asset the subscription grant status of which is to
be updated.}

\item{domainIdentifier}{[required] The identifier of the Amazon DataZone domain in which a subscription
grant status is to be updated.}

\item{failureCause}{Specifies the error message that is returned if the operation cannot be
successfully completed.}

\item{identifier}{[required] The identifier of the subscription grant the status of which is to be
updated.}

\item{status}{[required] The status to be updated as part of the
\code{\link[=datazone_update_subscription_grant_status]{update_subscription_grant_status}}
action.}

\item{targetName}{The target name to be updated as part of the
\code{\link[=datazone_update_subscription_grant_status]{update_subscription_grant_status}}
action.}
}
\description{
Updates the status of the specified subscription grant status in Amazon DataZone.

See \url{https://www.paws-r-sdk.com/docs/datazone_update_subscription_grant_status/} for full documentation.
}
\keyword{internal}
