% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_update_subscription_request}
\alias{datazone_update_subscription_request}
\title{Updates a specified subscription request in Amazon DataZone}
\usage{
datazone_update_subscription_request(
  domainIdentifier,
  identifier,
  requestReason
)
}
\arguments{
\item{domainIdentifier}{[required] The identifier of the Amazon DataZone domain in which a subscription
request is to be updated.}

\item{identifier}{[required] The identifier of the subscription request that is to be updated.}

\item{requestReason}{[required] The reason for the
\code{\link[=datazone_update_subscription_request]{update_subscription_request}}
action.}
}
\description{
Updates a specified subscription request in Amazon DataZone.

See \url{https://www.paws-r-sdk.com/docs/datazone_update_subscription_request/} for full documentation.
}
\keyword{internal}
