% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_update_user_profile}
\alias{datazone_update_user_profile}
\title{Updates the specified user profile in Amazon DataZone}
\usage{
datazone_update_user_profile(
  domainIdentifier,
  status,
  type = NULL,
  userIdentifier
)
}
\arguments{
\item{domainIdentifier}{[required] The identifier of the Amazon DataZone domain in which a user profile is
updated.}

\item{status}{[required] The status of the user profile that are to be updated.}

\item{type}{The type of the user profile that are to be updated.}

\item{userIdentifier}{[required] The identifier of the user whose user profile is to be updated.}
}
\description{
Updates the specified user profile in Amazon DataZone.

See \url{https://www.paws-r-sdk.com/docs/datazone_update_user_profile/} for full documentation.
}
\keyword{internal}
