% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticsearchservice_operations.R
\name{elasticsearchservice_cancel_elasticsearch_service_software_update}
\alias{elasticsearchservice_cancel_elasticsearch_service_software_update}
\title{Cancels a scheduled service software update for an Amazon ES domain}
\usage{
elasticsearchservice_cancel_elasticsearch_service_software_update(DomainName)
}
\arguments{
\item{DomainName}{[required] The name of the domain that you want to stop the latest service software
update on.}
}
\description{
Cancels a scheduled service software update for an Amazon ES domain. You can only perform this operation before the \code{AutomatedUpdateDate} and when the \code{UpdateStatus} is in the \code{PENDING_UPDATE} state.

See \url{https://paws-r.github.io/docs/elasticsearchservice/cancel_elasticsearch_service_software_update.html} for full documentation.
}
\keyword{internal}
