% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticsearchservice_operations.R
\name{elasticsearchservice_cancel_domain_config_change}
\alias{elasticsearchservice_cancel_domain_config_change}
\title{Cancels a pending configuration change on an Amazon OpenSearch Service
domain}
\usage{
elasticsearchservice_cancel_domain_config_change(DomainName, DryRun = NULL)
}
\arguments{
\item{DomainName}{[required] Name of the OpenSearch Service domain configuration request to cancel.}

\item{DryRun}{When set to \strong{True}, returns the list of change IDs and properties that
will be cancelled without actually cancelling the change.}
}
\description{
Cancels a pending configuration change on an Amazon OpenSearch Service domain.

See \url{https://www.paws-r-sdk.com/docs/elasticsearchservice_cancel_domain_config_change/} for full documentation.
}
\keyword{internal}
