% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticsearchservice_operations.R
\name{elasticsearchservice_create_outbound_cross_cluster_search_connection}
\alias{elasticsearchservice_create_outbound_cross_cluster_search_connection}
\title{Creates a new cross-cluster search connection from a source domain to a
destination domain}
\usage{
elasticsearchservice_create_outbound_cross_cluster_search_connection(
  SourceDomainInfo, DestinationDomainInfo, ConnectionAlias)
}
\arguments{
\item{SourceDomainInfo}{[required] Specifies the \code{DomainInformation} for the source Elasticsearch domain.}

\item{DestinationDomainInfo}{[required] Specifies the \code{DomainInformation} for the destination Elasticsearch
domain.}

\item{ConnectionAlias}{[required] Specifies the connection alias that will be used by the customer for
this connection.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  SourceDomainInfo = list(
    OwnerId = "string",
    DomainName = "string",
    Region = "string"
  ),
  DestinationDomainInfo = list(
    OwnerId = "string",
    DomainName = "string",
    Region = "string"
  ),
  ConnectionAlias = "string",
  ConnectionStatus = list(
    StatusCode = "PENDING_ACCEPTANCE"|"VALIDATING"|"VALIDATION_FAILED"|"PROVISIONING"|"ACTIVE"|"REJECTED"|"DELETING"|"DELETED",
    Message = "string"
  ),
  CrossClusterSearchConnectionId = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Creates a new cross-cluster search connection from a source domain to a
destination domain.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_outbound_cross_cluster_search_connection(
  SourceDomainInfo = list(
    OwnerId = "string",
    DomainName = "string",
    Region = "string"
  ),
  DestinationDomainInfo = list(
    OwnerId = "string",
    DomainName = "string",
    Region = "string"
  ),
  ConnectionAlias = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
