% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticsearchservice_operations.R
\name{elasticsearchservice_create_package}
\alias{elasticsearchservice_create_package}
\title{Create a package for use with Amazon ES domains}
\usage{
elasticsearchservice_create_package(PackageName, PackageType,
  PackageDescription, PackageSource)
}
\arguments{
\item{PackageName}{[required] Unique identifier for the package.}

\item{PackageType}{[required] Type of package. Currently supports only TXT-DICTIONARY.}

\item{PackageDescription}{Description of the package.}

\item{PackageSource}{[required] The customer S3 location \code{PackageSource} for importing the package.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  PackageDetails = list(
    PackageID = "string",
    PackageName = "string",
    PackageType = "TXT-DICTIONARY",
    PackageDescription = "string",
    PackageStatus = "COPYING"|"COPY_FAILED"|"VALIDATING"|"VALIDATION_FAILED"|"AVAILABLE"|"DELETING"|"DELETED"|"DELETE_FAILED",
    CreatedAt = as.POSIXct(
      "2015-01-01"
    ),
    LastUpdatedAt = as.POSIXct(
      "2015-01-01"
    ),
    AvailablePackageVersion = "string",
    ErrorDetails = list(
      ErrorType = "string",
      ErrorMessage = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Create a package for use with Amazon ES domains.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_package(
  PackageName = "string",
  PackageType = "TXT-DICTIONARY",
  PackageDescription = "string",
  PackageSource = list(
    S3BucketName = "string",
    S3Key = "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
