% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticsearchservice_operations.R
\name{elasticsearchservice_delete_outbound_cross_cluster_search_connection}
\alias{elasticsearchservice_delete_outbound_cross_cluster_search_connection}
\title{Allows the source domain owner to delete an existing outbound
cross-cluster search connection}
\usage{
elasticsearchservice_delete_outbound_cross_cluster_search_connection(
  CrossClusterSearchConnectionId
)
}
\arguments{
\item{CrossClusterSearchConnectionId}{[required] The id of the outbound connection that you want to permanently delete.}
}
\description{
Allows the source domain owner to delete an existing outbound cross-cluster search connection.

See \url{https://www.paws-r-sdk.com/docs/elasticsearchservice_delete_outbound_cross_cluster_search_connection/} for full documentation.
}
\keyword{internal}
