% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticsearchservice_operations.R
\name{elasticsearchservice_delete_elasticsearch_domain}
\alias{elasticsearchservice_delete_elasticsearch_domain}
\title{Permanently deletes the specified Elasticsearch domain and all of its
data}
\usage{
elasticsearchservice_delete_elasticsearch_domain(DomainName)
}
\arguments{
\item{DomainName}{[required] The name of the Elasticsearch domain that you want to permanently
delete.}
}
\description{
Permanently deletes the specified Elasticsearch domain and all of its data. Once a domain is deleted, it cannot be recovered.

See \url{https://www.paws-r-sdk.com/docs/elasticsearchservice_delete_elasticsearch_domain/} for full documentation.
}
\keyword{internal}
