% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticsearchservice_operations.R
\name{elasticsearchservice_delete_elasticsearch_service_role}
\alias{elasticsearchservice_delete_elasticsearch_service_role}
\title{Deletes the service-linked role that Elasticsearch Service uses to
manage and maintain VPC domains}
\usage{
elasticsearchservice_delete_elasticsearch_service_role()
}
\description{
Deletes the service-linked role that Elasticsearch Service uses to manage and maintain VPC domains. Role deletion will fail if any existing VPC domains use the role. You must delete any such Elasticsearch domains before deleting the role. See \if{html}{\out{<a href="https://docs.aws.amazon.com/opensearch-service/latest/developerguide/vpc.html#es-enabling-slr" target="_blank">}}Deleting Elasticsearch Service Role\if{html}{\out{</a>}} in \emph{VPC Endpoints for Amazon Elasticsearch Service Domains}.

See \url{https://paws-r.github.io/docs/elasticsearchservice/delete_elasticsearch_service_role.html} for full documentation.
}
\keyword{internal}
