% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticsearchservice_operations.R
\name{elasticsearchservice_delete_package}
\alias{elasticsearchservice_delete_package}
\title{Delete the package}
\usage{
elasticsearchservice_delete_package(PackageID)
}
\arguments{
\item{PackageID}{[required] Internal ID of the package that you want to delete. Use
\code{\link[=elasticsearchservice_describe_packages]{describe_packages}} to find
this value.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  PackageDetails = list(
    PackageID = "string",
    PackageName = "string",
    PackageType = "TXT-DICTIONARY",
    PackageDescription = "string",
    PackageStatus = "COPYING"|"COPY_FAILED"|"VALIDATING"|"VALIDATION_FAILED"|"AVAILABLE"|"DELETING"|"DELETED"|"DELETE_FAILED",
    CreatedAt = as.POSIXct(
      "2015-01-01"
    ),
    LastUpdatedAt = as.POSIXct(
      "2015-01-01"
    ),
    AvailablePackageVersion = "string",
    ErrorDetails = list(
      ErrorType = "string",
      ErrorMessage = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Delete the package.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_package(
  PackageID = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
