% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticsearchservice_operations.R
\name{elasticsearchservice_describe_elasticsearch_instance_type_limits}
\alias{elasticsearchservice_describe_elasticsearch_instance_type_limits}
\title{Describe Elasticsearch Limits for a given InstanceType and
ElasticsearchVersion}
\usage{
elasticsearchservice_describe_elasticsearch_instance_type_limits(
  DomainName = NULL,
  InstanceType,
  ElasticsearchVersion
)
}
\arguments{
\item{DomainName}{DomainName represents the name of the Domain that we are trying to
modify. This should be present only if we are querying for Elasticsearch
\code{Limits} for existing domain.}

\item{InstanceType}{[required] The instance type for an Elasticsearch cluster for which Elasticsearch
\code{Limits} are needed.}

\item{ElasticsearchVersion}{[required] Version of Elasticsearch for which \code{Limits} are needed.}
}
\description{
Describe Elasticsearch Limits for a given InstanceType and ElasticsearchVersion. When modifying existing Domain, specify the \code{DomainName} to know what Limits are supported for modifying.

See \url{https://paws-r.github.io/docs/elasticsearchservice/describe_elasticsearch_instance_type_limits.html} for full documentation.
}
\keyword{internal}
