% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticsearchservice_operations.R
\name{elasticsearchservice_describe_domain_auto_tunes}
\alias{elasticsearchservice_describe_domain_auto_tunes}
\title{Provides scheduled Auto-Tune action details for the Elasticsearch
domain, such as Auto-Tune action type, description, severity, and
scheduled date}
\usage{
elasticsearchservice_describe_domain_auto_tunes(DomainName, MaxResults,
  NextToken)
}
\arguments{
\item{DomainName}{[required] Specifies the domain name for which you want Auto-Tune action details.}

\item{MaxResults}{Set this value to limit the number of results returned. If not
specified, defaults to 100.}

\item{NextToken}{NextToken is sent in case the earlier API call results contain the
NextToken. It is used for pagination.}
}
\value{
A list with the following syntax:\preformatted{list(
  AutoTunes = list(
    list(
      AutoTuneType = "SCHEDULED_ACTION",
      AutoTuneDetails = list(
        ScheduledAutoTuneDetails = list(
          Date = as.POSIXct(
            "2015-01-01"
          ),
          ActionType = "JVM_HEAP_SIZE_TUNING"|"JVM_YOUNG_GEN_TUNING",
          Action = "string",
          Severity = "LOW"|"MEDIUM"|"HIGH"
        )
      )
    )
  ),
  NextToken = "string"
)
}
}
\description{
Provides scheduled Auto-Tune action details for the Elasticsearch
domain, such as Auto-Tune action type, description, severity, and
scheduled date.
}
\section{Request syntax}{
\preformatted{svc$describe_domain_auto_tunes(
  DomainName = "string",
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
