% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticsearchservice_operations.R
\name{elasticsearchservice_describe_domain_change_progress}
\alias{elasticsearchservice_describe_domain_change_progress}
\title{Returns information about the current blue/green deployment happening on
a domain, including a change ID, status, and progress stages}
\usage{
elasticsearchservice_describe_domain_change_progress(
  DomainName,
  ChangeId = NULL
)
}
\arguments{
\item{DomainName}{[required] The domain you want to get the progress information about.}

\item{ChangeId}{The specific change ID for which you want to get progress information.
This is an optional parameter. If omitted, the service returns
information about the most recent configuration change.}
}
\description{
Returns information about the current blue/green deployment happening on a domain, including a change ID, status, and progress stages.

See \url{https://www.paws-r-sdk.com/docs/elasticsearchservice_describe_domain_change_progress/} for full documentation.
}
\keyword{internal}
