% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticsearchservice_operations.R
\name{elasticsearchservice_describe_elasticsearch_domain}
\alias{elasticsearchservice_describe_elasticsearch_domain}
\title{Returns domain configuration information about the specified
Elasticsearch domain, including the domain ID, domain endpoint, and
domain ARN}
\usage{
elasticsearchservice_describe_elasticsearch_domain(DomainName)
}
\arguments{
\item{DomainName}{[required] The name of the Elasticsearch domain for which you want information.}
}
\description{
Returns domain configuration information about the specified Elasticsearch domain, including the domain ID, domain endpoint, and domain ARN.

See \url{https://www.paws-r-sdk.com/docs/elasticsearchservice_describe_elasticsearch_domain/} for full documentation.
}
\keyword{internal}
