% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticsearchservice_operations.R
\name{elasticsearchservice_describe_elasticsearch_domains}
\alias{elasticsearchservice_describe_elasticsearch_domains}
\title{Returns domain configuration information about the specified
Elasticsearch domains, including the domain ID, domain endpoint, and
domain ARN}
\usage{
elasticsearchservice_describe_elasticsearch_domains(DomainNames)
}
\arguments{
\item{DomainNames}{[required] The Elasticsearch domains for which you want information.}
}
\description{
Returns domain configuration information about the specified Elasticsearch domains, including the domain ID, domain endpoint, and domain ARN.

See \url{https://paws-r.github.io/docs/elasticsearchservice/describe_elasticsearch_domains.html} for full documentation.
}
\keyword{internal}
