% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticsearchservice_operations.R
\name{elasticsearchservice_describe_outbound_cross_cluster_search_connections}
\alias{elasticsearchservice_describe_outbound_cross_cluster_search_connections}
\title{Lists all the outbound cross-cluster search connections for a source
domain}
\usage{
elasticsearchservice_describe_outbound_cross_cluster_search_connections(
  Filters = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{Filters}{A list of filters used to match properties for outbound cross-cluster
search connection. Available \code{Filter} names for this operation are:
\itemize{
\item cross-cluster-search-connection-id
\item destination-domain-info.domain-name
\item destination-domain-info.owner-id
\item destination-domain-info.region
\item source-domain-info.domain-name
}}

\item{MaxResults}{Set this value to limit the number of results returned. If not
specified, defaults to 100.}

\item{NextToken}{NextToken is sent in case the earlier API call results contain the
NextToken. It is used for pagination.}
}
\description{
Lists all the outbound cross-cluster search connections for a source domain.

See \url{https://paws-r.github.io/docs/elasticsearchservice/describe_outbound_cross_cluster_search_connections.html} for full documentation.
}
\keyword{internal}
