% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticsearchservice_operations.R
\name{elasticsearchservice_describe_vpc_endpoints}
\alias{elasticsearchservice_describe_vpc_endpoints}
\title{Describes one or more Amazon OpenSearch Service-managed VPC endpoints}
\usage{
elasticsearchservice_describe_vpc_endpoints(VpcEndpointIds)
}
\arguments{
\item{VpcEndpointIds}{[required] The unique identifiers of the endpoints to get information about.}
}
\description{
Describes one or more Amazon OpenSearch Service-managed VPC endpoints.

See \url{https://www.paws-r-sdk.com/docs/elasticsearchservice_describe_vpc_endpoints/} for full documentation.
}
\keyword{internal}
