% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticsearchservice_operations.R
\name{elasticsearchservice_get_compatible_elasticsearch_versions}
\alias{elasticsearchservice_get_compatible_elasticsearch_versions}
\title{Returns a list of upgrade compatible Elastisearch versions}
\usage{
elasticsearchservice_get_compatible_elasticsearch_versions(DomainName = NULL)
}
\arguments{
\item{DomainName}{The name of an Elasticsearch domain. Domain names are unique across the domains owned by an account within an AWS region. Domain names start with a letter or number and can contain the following characters: a-z (lowercase), 0-9, and - (hyphen).}
}
\description{
Returns a list of upgrade compatible Elastisearch versions. You can optionally pass a \code{DomainName} to get all upgrade compatible Elasticsearch versions for that specific domain.

See \url{https://www.paws-r-sdk.com/docs/elasticsearchservice_get_compatible_elasticsearch_versions/} for full documentation.
}
\keyword{internal}
