% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticsearchservice_operations.R
\name{elasticsearchservice_get_upgrade_history}
\alias{elasticsearchservice_get_upgrade_history}
\title{Retrieves the complete history of the last 10 upgrades that were
performed on the domain}
\usage{
elasticsearchservice_get_upgrade_history(
  DomainName,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{DomainName}{[required] The name of an Elasticsearch domain. Domain names are unique across the domains owned by an account within an AWS region. Domain names start with a letter or number and can contain the following characters: a-z (lowercase), 0-9, and - (hyphen).}

\item{MaxResults}{Set this value to limit the number of results returned.}

\item{NextToken}{Paginated APIs accepts NextToken input to returns next page results and provides a NextToken output in the response which can be used by the client to retrieve more results.}
}
\description{
Retrieves the complete history of the last 10 upgrades that were performed on the domain.

See \url{https://www.paws-r-sdk.com/docs/elasticsearchservice_get_upgrade_history/} for full documentation.
}
\keyword{internal}
