% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticsearchservice_operations.R
\name{elasticsearchservice_list_elasticsearch_versions}
\alias{elasticsearchservice_list_elasticsearch_versions}
\title{List all supported Elasticsearch versions}
\usage{
elasticsearchservice_list_elasticsearch_versions(
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{MaxResults}{Set this value to limit the number of results returned. Value provided must be greater than 10 else it wont be honored.}

\item{NextToken}{Paginated APIs accepts NextToken input to returns next page results and provides a NextToken output in the response which can be used by the client to retrieve more results.}
}
\description{
List all supported Elasticsearch versions

See \url{https://www.paws-r-sdk.com/docs/elasticsearchservice_list_elasticsearch_versions/} for full documentation.
}
\keyword{internal}
