% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticsearchservice_operations.R
\name{elasticsearchservice_list_vpc_endpoint_access}
\alias{elasticsearchservice_list_vpc_endpoint_access}
\title{Retrieves information about each principal that is allowed to access a
given Amazon OpenSearch Service domain through the use of an interface
VPC endpoint}
\usage{
elasticsearchservice_list_vpc_endpoint_access(DomainName, NextToken = NULL)
}
\arguments{
\item{DomainName}{[required] The name of the OpenSearch Service domain to retrieve access information
for.}

\item{NextToken}{Provides an identifier to allow retrieval of paginated results.}
}
\description{
Retrieves information about each principal that is allowed to access a given Amazon OpenSearch Service domain through the use of an interface VPC endpoint.

See \url{https://paws-r.github.io/docs/elasticsearchservice/list_vpc_endpoint_access.html} for full documentation.
}
\keyword{internal}
