% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticsearchservice_operations.R
\name{elasticsearchservice_list_vpc_endpoints}
\alias{elasticsearchservice_list_vpc_endpoints}
\title{Retrieves all Amazon OpenSearch Service-managed VPC endpoints in the
current account and Region}
\usage{
elasticsearchservice_list_vpc_endpoints(NextToken = NULL)
}
\arguments{
\item{NextToken}{Identifier to allow retrieval of paginated results.}
}
\description{
Retrieves all Amazon OpenSearch Service-managed VPC endpoints in the current account and Region.

See \url{https://paws-r.github.io/docs/elasticsearchservice/list_vpc_endpoints.html} for full documentation.
}
\keyword{internal}
