% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticsearchservice_operations.R
\name{elasticsearchservice_purchase_reserved_elasticsearch_instance_offering}
\alias{elasticsearchservice_purchase_reserved_elasticsearch_instance_offering}
\title{Allows you to purchase reserved Elasticsearch instances}
\usage{
elasticsearchservice_purchase_reserved_elasticsearch_instance_offering(
  ReservedElasticsearchInstanceOfferingId, ReservationName, InstanceCount)
}
\arguments{
\item{ReservedElasticsearchInstanceOfferingId}{[required] The ID of the reserved Elasticsearch instance offering to purchase.}

\item{ReservationName}{[required] A customer-specified identifier to track this reservation.}

\item{InstanceCount}{The number of Elasticsearch instances to reserve.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ReservedElasticsearchInstanceId = "string",
  ReservationName = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Allows you to purchase reserved Elasticsearch instances.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$purchase_reserved_elasticsearch_instance_offering(
  ReservedElasticsearchInstanceOfferingId = "string",
  ReservationName = "string",
  InstanceCount = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
