% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticsearchservice_operations.R
\name{elasticsearchservice_reject_inbound_cross_cluster_search_connection}
\alias{elasticsearchservice_reject_inbound_cross_cluster_search_connection}
\title{Allows the destination domain owner to reject an inbound cross-cluster
search connection request}
\usage{
elasticsearchservice_reject_inbound_cross_cluster_search_connection(
  CrossClusterSearchConnectionId
)
}
\arguments{
\item{CrossClusterSearchConnectionId}{[required] The id of the inbound connection that you want to reject.}
}
\description{
Allows the destination domain owner to reject an inbound cross-cluster search connection request.

See \url{https://www.paws-r-sdk.com/docs/elasticsearchservice_reject_inbound_cross_cluster_search_connection/} for full documentation.
}
\keyword{internal}
