% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticsearchservice_operations.R
\name{elasticsearchservice_remove_tags}
\alias{elasticsearchservice_remove_tags}
\title{Removes the specified set of tags from the specified Elasticsearch
domain}
\usage{
elasticsearchservice_remove_tags(ARN, TagKeys)
}
\arguments{
\item{ARN}{[required] Specifies the \code{ARN} for the Elasticsearch domain from which you want to
delete the specified tags.}

\item{TagKeys}{[required] Specifies the \code{TagKey} list which you want to remove from the
Elasticsearch domain.}
}
\value{
An empty list.
}
\description{
Removes the specified set of tags from the specified Elasticsearch
domain.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$remove_tags(
  ARN = "string",
  TagKeys = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
