% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticsearchservice_operations.R
\name{elasticsearchservice_revoke_vpc_endpoint_access}
\alias{elasticsearchservice_revoke_vpc_endpoint_access}
\title{Revokes access to an Amazon OpenSearch Service domain that was provided
through an interface VPC endpoint}
\usage{
elasticsearchservice_revoke_vpc_endpoint_access(DomainName, Account)
}
\arguments{
\item{DomainName}{[required] The name of the OpenSearch Service domain.}

\item{Account}{[required] The account ID to revoke access from.}
}
\description{
Revokes access to an Amazon OpenSearch Service domain that was provided through an interface VPC endpoint.

See \url{https://www.paws-r-sdk.com/docs/elasticsearchservice_revoke_vpc_endpoint_access/} for full documentation.
}
\keyword{internal}
