% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr_operations.R
\name{emr_add_instance_fleet}
\alias{emr_add_instance_fleet}
\title{Adds an instance fleet to a running cluster}
\usage{
emr_add_instance_fleet(ClusterId, InstanceFleet)
}
\arguments{
\item{ClusterId}{[required] The unique identifier of the cluster.}

\item{InstanceFleet}{[required] Specifies the configuration of the instance fleet.}
}
\description{
Adds an instance fleet to a running cluster.

See \url{https://www.paws-r-sdk.com/docs/emr_add_instance_fleet/} for full documentation.
}
\keyword{internal}
