% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr_operations.R
\name{emr_add_instance_groups}
\alias{emr_add_instance_groups}
\title{Adds one or more instance groups to a running cluster}
\usage{
emr_add_instance_groups(InstanceGroups, JobFlowId)
}
\arguments{
\item{InstanceGroups}{[required] Instance groups to add.}

\item{JobFlowId}{[required] Job flow in which to add the instance groups.}
}
\description{
Adds one or more instance groups to a running cluster.

See \url{https://www.paws-r-sdk.com/docs/emr_add_instance_groups/} for full documentation.
}
\keyword{internal}
