% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr_operations.R
\name{emr_create_persistent_app_ui}
\alias{emr_create_persistent_app_ui}
\title{Creates a persistent application user interface}
\usage{
emr_create_persistent_app_ui(
  TargetResourceArn,
  EMRContainersConfig = NULL,
  Tags = NULL,
  XReferer = NULL,
  ProfilerType = NULL
)
}
\arguments{
\item{TargetResourceArn}{[required] The unique Amazon Resource Name (ARN) of the target resource.}

\item{EMRContainersConfig}{The EMR containers configuration.}

\item{Tags}{Tags for the persistent application user interface.}

\item{XReferer}{The cross reference for the persistent application user interface.}

\item{ProfilerType}{The profiler type for the persistent application user interface.}
}
\description{
Creates a persistent application user interface.

See \url{https://www.paws-r-sdk.com/docs/emr_create_persistent_app_ui/} for full documentation.
}
\keyword{internal}
