% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr_operations.R
\name{emr_create_security_configuration}
\alias{emr_create_security_configuration}
\title{Creates a security configuration, which is stored in the service and can
be specified when a cluster is created}
\usage{
emr_create_security_configuration(Name, SecurityConfiguration)
}
\arguments{
\item{Name}{[required] The name of the security configuration.}

\item{SecurityConfiguration}{[required] The security configuration details in JSON format. For JSON parameters
and examples, see \href{https://docs.aws.amazon.com/emr/latest/ManagementGuide/emr-security-configurations.html}{Use Security Configurations to Set Up Cluster Security}
in the \emph{Amazon EMR Management Guide}.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Name = "string",
  CreationDateTime = as.POSIXct(
    "2015-01-01"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Creates a security configuration, which is stored in the service and can
be specified when a cluster is created.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_security_configuration(
  Name = "string",
  SecurityConfiguration = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
