% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr_operations.R
\name{emr_delete_security_configuration}
\alias{emr_delete_security_configuration}
\title{Deletes a security configuration}
\usage{
emr_delete_security_configuration(Name)
}
\arguments{
\item{Name}{[required] The name of the security configuration.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Deletes a security configuration.
}
\section{Request syntax}{
\preformatted{svc$delete_security_configuration(
  Name = "string"
)
}
}

\keyword{internal}
