% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr_operations.R
\name{emr_delete_studio}
\alias{emr_delete_studio}
\title{The Amazon EMR Studio APIs are in preview release for Amazon EMR and are
subject to change}
\usage{
emr_delete_studio(StudioId)
}
\arguments{
\item{StudioId}{[required] The ID of the Amazon EMR Studio.}
}
\value{
An empty list.
}
\description{
The Amazon EMR Studio APIs are in preview release for Amazon EMR and are
subject to change.

Removes an Amazon EMR Studio from the Studio metadata store.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_studio(
  StudioId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
