% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr_operations.R
\name{emr_describe_job_flows}
\alias{emr_describe_job_flows}
\title{This API is no longer supported and will eventually be removed}
\usage{
emr_describe_job_flows(
  CreatedAfter = NULL,
  CreatedBefore = NULL,
  JobFlowIds = NULL,
  JobFlowStates = NULL
)
}
\arguments{
\item{CreatedAfter}{Return only job flows created after this date and time.}

\item{CreatedBefore}{Return only job flows created before this date and time.}

\item{JobFlowIds}{Return only job flows whose job flow ID is contained in this list.}

\item{JobFlowStates}{Return only job flows whose state is contained in this list.}
}
\description{
This API is no longer supported and will eventually be removed. We recommend you use \code{\link[=emr_list_clusters]{list_clusters}}, \code{\link[=emr_describe_cluster]{describe_cluster}}, \code{\link[=emr_list_steps]{list_steps}}, \code{\link[=emr_list_instance_groups]{list_instance_groups}} and \code{\link[=emr_list_bootstrap_actions]{list_bootstrap_actions}} instead.

See \url{https://www.paws-r-sdk.com/docs/emr_describe_job_flows/} for full documentation.
}
\keyword{internal}
