% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr_operations.R
\name{emr_describe_notebook_execution}
\alias{emr_describe_notebook_execution}
\title{Provides details of a notebook execution}
\usage{
emr_describe_notebook_execution(NotebookExecutionId)
}
\arguments{
\item{NotebookExecutionId}{[required] The unique identifier of the notebook execution.}
}
\description{
Provides details of a notebook execution.

See \url{https://paws-r.github.io/docs/emr/describe_notebook_execution.html} for full documentation.
}
\keyword{internal}
