% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr_operations.R
\name{emr_describe_persistent_app_ui}
\alias{emr_describe_persistent_app_ui}
\title{Describes a persistent application user interface}
\usage{
emr_describe_persistent_app_ui(PersistentAppUIId)
}
\arguments{
\item{PersistentAppUIId}{[required] The identifier for the persistent application user interface.}
}
\description{
Describes a persistent application user interface.

See \url{https://www.paws-r-sdk.com/docs/emr_describe_persistent_app_ui/} for full documentation.
}
\keyword{internal}
