% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr_operations.R
\name{emr_describe_release_label}
\alias{emr_describe_release_label}
\title{Provides EMR release label details, such as releases available the
region where the API request is run, and the available applications for
a specific EMR release label}
\usage{
emr_describe_release_label(
  ReleaseLabel = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{ReleaseLabel}{The target release label to be described.}

\item{NextToken}{The pagination token. Reserved for future use. Currently set to null.}

\item{MaxResults}{Reserved for future use. Currently set to null.}
}
\description{
Provides EMR release label details, such as releases available the region where the API request is run, and the available applications for a specific EMR release label. Can also list EMR release versions that support a specified version of Spark.

See \url{https://paws-r.github.io/docs/emr/describe_release_label.html} for full documentation.
}
\keyword{internal}
