% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr_operations.R
\name{emr_describe_security_configuration}
\alias{emr_describe_security_configuration}
\title{Provides the details of a security configuration by returning the
configuration JSON}
\usage{
emr_describe_security_configuration(Name)
}
\arguments{
\item{Name}{[required] The name of the security configuration.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Name = "string",
  SecurityConfiguration = "string",
  CreationDateTime = as.POSIXct(
    "2015-01-01"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Provides the details of a security configuration by returning the
configuration JSON.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_security_configuration(
  Name = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
