% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr_operations.R
\name{emr_describe_studio}
\alias{emr_describe_studio}
\title{The Amazon EMR Studio APIs are in preview release for Amazon EMR and are
subject to change}
\usage{
emr_describe_studio(StudioId)
}
\arguments{
\item{StudioId}{[required] The Amazon EMR Studio ID.}
}
\description{
The Amazon EMR Studio APIs are in preview release for Amazon EMR and are
subject to change.
}
\details{
Returns details for the specified Amazon EMR Studio including ID, Name,
VPC, Studio access URL, and so on.
}
\section{Request syntax}{
\preformatted{svc$describe_studio(
  StudioId = "string"
)
}
}

\keyword{internal}
