% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr_operations.R
\name{emr_get_auto_termination_policy}
\alias{emr_get_auto_termination_policy}
\title{Returns the auto-termination policy for an Amazon EMR cluster}
\usage{
emr_get_auto_termination_policy(ClusterId)
}
\arguments{
\item{ClusterId}{[required] Specifies the ID of the Amazon EMR cluster for which the
auto-termination policy will be fetched.}
}
\description{
Returns the auto-termination policy for an Amazon EMR cluster.

See \url{https://paws-r.github.io/docs/emr/get_auto_termination_policy.html} for full documentation.
}
\keyword{internal}
