% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr_operations.R
\name{emr_get_managed_scaling_policy}
\alias{emr_get_managed_scaling_policy}
\title{Fetches the attached managed scaling policy for an Amazon EMR cluster}
\usage{
emr_get_managed_scaling_policy(ClusterId)
}
\arguments{
\item{ClusterId}{[required] Specifies the ID of the cluster for which the managed scaling policy
will be fetched.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ManagedScalingPolicy = list(
    ComputeLimits = list(
      UnitType = "InstanceFleetUnits"|"Instances"|"VCPU",
      MinimumCapacityUnits = 123,
      MaximumCapacityUnits = 123,
      MaximumOnDemandCapacityUnits = 123,
      MaximumCoreCapacityUnits = 123
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Fetches the attached managed scaling policy for an Amazon EMR cluster.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_managed_scaling_policy(
  ClusterId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
