% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr_operations.R
\name{emr_get_on_cluster_app_ui_presigned_url}
\alias{emr_get_on_cluster_app_ui_presigned_url}
\title{The presigned URL properties for the cluster's application user
interface}
\usage{
emr_get_on_cluster_app_ui_presigned_url(
  ClusterId,
  OnClusterAppUIType = NULL,
  ApplicationId = NULL,
  DryRun = NULL,
  ExecutionRoleArn = NULL
)
}
\arguments{
\item{ClusterId}{[required] The cluster ID associated with the cluster's application user interface
presigned URL.}

\item{OnClusterAppUIType}{The application UI type associated with the cluster's application user
interface presigned URL.}

\item{ApplicationId}{The application ID associated with the cluster's application user
interface presigned URL.}

\item{DryRun}{Determines if the user interface presigned URL is for a dry run.}

\item{ExecutionRoleArn}{The execution role ARN associated with the cluster's application user
interface presigned URL.}
}
\description{
The presigned URL properties for the cluster's application user interface.

See \url{https://www.paws-r-sdk.com/docs/emr_get_on_cluster_app_ui_presigned_url/} for full documentation.
}
\keyword{internal}
