% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr_operations.R
\name{emr_get_persistent_app_ui_presigned_url}
\alias{emr_get_persistent_app_ui_presigned_url}
\title{The presigned URL properties for the cluster's application user
interface}
\usage{
emr_get_persistent_app_ui_presigned_url(
  PersistentAppUIId,
  PersistentAppUIType = NULL,
  ApplicationId = NULL,
  AuthProxyCall = NULL,
  ExecutionRoleArn = NULL
)
}
\arguments{
\item{PersistentAppUIId}{[required] The persistent application user interface ID associated with the
presigned URL.}

\item{PersistentAppUIType}{The persistent application user interface type associated with the
presigned URL.}

\item{ApplicationId}{The application ID associated with the presigned URL.}

\item{AuthProxyCall}{A boolean that represents if the caller is an authentication proxy call.}

\item{ExecutionRoleArn}{The execution role ARN associated with the presigned URL.}
}
\description{
The presigned URL properties for the cluster's application user interface.

See \url{https://www.paws-r-sdk.com/docs/emr_get_persistent_app_ui_presigned_url/} for full documentation.
}
\keyword{internal}
