% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr_operations.R
\name{emr_get_session}
\alias{emr_get_session}
\title{Returns detailed information about a session}
\usage{
emr_get_session(ClusterId, SessionId)
}
\arguments{
\item{ClusterId}{[required] The ID of the cluster that the session belongs to.}

\item{SessionId}{[required] The ID of the session.}
}
\description{
Returns detailed information about a session.

See \url{https://www.paws-r-sdk.com/docs/emr_get_session/} for full documentation.
}
\keyword{internal}
