% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr_operations.R
\name{emr_list_bootstrap_actions}
\alias{emr_list_bootstrap_actions}
\title{Provides information about the bootstrap actions associated with a
cluster}
\usage{
emr_list_bootstrap_actions(ClusterId, Marker = NULL)
}
\arguments{
\item{ClusterId}{[required] The cluster identifier for the bootstrap actions to list.}

\item{Marker}{The pagination token that indicates the next set of results to retrieve.}
}
\description{
Provides information about the bootstrap actions associated with a cluster.

See \url{https://www.paws-r-sdk.com/docs/emr_list_bootstrap_actions/} for full documentation.
}
\keyword{internal}
