% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr_operations.R
\name{emr_list_clusters}
\alias{emr_list_clusters}
\title{Provides the status of all clusters visible to this Amazon Web Services
account}
\usage{
emr_list_clusters(
  CreatedAfter = NULL,
  CreatedBefore = NULL,
  ClusterStates = NULL,
  Marker = NULL
)
}
\arguments{
\item{CreatedAfter}{The creation date and time beginning value filter for listing clusters.}

\item{CreatedBefore}{The creation date and time end value filter for listing clusters.}

\item{ClusterStates}{The cluster state filters to apply when listing clusters. Clusters that
change state while this action runs may be not be returned as expected
in the list of clusters.}

\item{Marker}{The pagination token that indicates the next set of results to retrieve.}
}
\description{
Provides the status of all clusters visible to this Amazon Web Services account. Allows you to filter the list of clusters based on certain criteria; for example, filtering by cluster creation date and time or by status. This call returns a maximum of 50 clusters in unsorted order per call, but returns a marker to track the paging of the cluster list across multiple ListClusters calls.

See \url{https://paws-r.github.io/docs/emr/list_clusters.html} for full documentation.
}
\keyword{internal}
