% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr_operations.R
\name{emr_list_instance_fleets}
\alias{emr_list_instance_fleets}
\title{Lists all available details about the instance fleets in a cluster}
\usage{
emr_list_instance_fleets(ClusterId, Marker = NULL)
}
\arguments{
\item{ClusterId}{[required] The unique identifier of the cluster.}

\item{Marker}{The pagination token that indicates the next set of results to retrieve.}
}
\description{
Lists all available details about the instance fleets in a cluster.

See \url{https://www.paws-r-sdk.com/docs/emr_list_instance_fleets/} for full documentation.
}
\keyword{internal}
