% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr_operations.R
\name{emr_list_sessions}
\alias{emr_list_sessions}
\title{Lists the sessions on a cluster}
\usage{
emr_list_sessions(
  ClusterId,
  SessionStates = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{ClusterId}{[required] The ID of the cluster to list sessions for.}

\item{SessionStates}{An optional filter that limits the results to sessions in the specified states.}

\item{NextToken}{The pagination token returned by a previous \code{\link[=emr_list_sessions]{list_sessions}} call. Use it to retrieve the next page of results.}

\item{MaxResults}{The maximum number of sessions to return in each page of results.}
}
\description{
Lists the sessions on a cluster. You can filter the results by session state. Newer sessions are returned first.

See \url{https://www.paws-r-sdk.com/docs/emr_list_sessions/} for full documentation.
}
\keyword{internal}
