% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr_operations.R
\name{emr_list_studios}
\alias{emr_list_studios}
\title{Returns a list of all Amazon EMR Studios associated with the Amazon Web
Services account}
\usage{
emr_list_studios(Marker = NULL)
}
\arguments{
\item{Marker}{The pagination token that indicates the set of results to retrieve.}
}
\description{
Returns a list of all Amazon EMR Studios associated with the Amazon Web Services account. The list includes details such as ID, Studio Access URL, and creation time for each Studio.

See \url{https://www.paws-r-sdk.com/docs/emr_list_studios/} for full documentation.
}
\keyword{internal}
