% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr_operations.R
\name{emr_modify_instance_fleet}
\alias{emr_modify_instance_fleet}
\title{Modifies the target On-Demand and target Spot capacities for the
instance fleet with the specified InstanceFleetID within the cluster
specified using ClusterID}
\usage{
emr_modify_instance_fleet(ClusterId, InstanceFleet)
}
\arguments{
\item{ClusterId}{[required] The unique identifier of the cluster.}

\item{InstanceFleet}{[required] The configuration parameters of the instance fleet.}
}
\description{
Modifies the target On-Demand and target Spot capacities for the instance fleet with the specified InstanceFleetID within the cluster specified using ClusterID. The call either succeeds or fails atomically.

See \url{https://www.paws-r-sdk.com/docs/emr_modify_instance_fleet/} for full documentation.
}
\keyword{internal}
