% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr_operations.R
\name{emr_modify_instance_groups}
\alias{emr_modify_instance_groups}
\title{ModifyInstanceGroups modifies the number of nodes and configuration
settings of an instance group}
\usage{
emr_modify_instance_groups(ClusterId = NULL, InstanceGroups = NULL)
}
\arguments{
\item{ClusterId}{The ID of the cluster to which the instance group belongs.}

\item{InstanceGroups}{Instance groups to change.}
}
\description{
ModifyInstanceGroups modifies the number of nodes and configuration settings of an instance group. The input parameters include the new target instance count for the group and the instance group ID. The call will either succeed or fail atomically.

See \url{https://www.paws-r-sdk.com/docs/emr_modify_instance_groups/} for full documentation.
}
\keyword{internal}
