% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr_operations.R
\name{emr_remove_auto_termination_policy}
\alias{emr_remove_auto_termination_policy}
\title{Removes an auto-termination policy from an Amazon EMR cluster}
\usage{
emr_remove_auto_termination_policy(ClusterId)
}
\arguments{
\item{ClusterId}{[required] Specifies the ID of the Amazon EMR cluster from which the
auto-termination policy will be removed.}
}
\description{
Removes an auto-termination policy from an Amazon EMR cluster.

See \url{https://paws-r.github.io/docs/emr/remove_auto_termination_policy.html} for full documentation.
}
\keyword{internal}
