% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr_operations.R
\name{emr_remove_managed_scaling_policy}
\alias{emr_remove_managed_scaling_policy}
\title{Removes a managed scaling policy from a specified Amazon EMR cluster}
\usage{
emr_remove_managed_scaling_policy(ClusterId)
}
\arguments{
\item{ClusterId}{[required] Specifies the ID of the cluster from which the managed scaling policy
will be removed.}
}
\description{
Removes a managed scaling policy from a specified Amazon EMR cluster.

See \url{https://www.paws-r-sdk.com/docs/emr_remove_managed_scaling_policy/} for full documentation.
}
\keyword{internal}
