% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr_operations.R
\name{emr_terminate_session}
\alias{emr_terminate_session}
\title{Terminates an active session}
\usage{
emr_terminate_session(ClusterId, SessionId)
}
\arguments{
\item{ClusterId}{[required] The ID of the cluster that the session belongs to.}

\item{SessionId}{[required] The ID of the session to terminate.}
}
\description{
Terminates an active session. After you call this operation, the session enters the \code{TERMINATING} state and then transitions to \code{TERMINATED}.

See \url{https://www.paws-r-sdk.com/docs/emr_terminate_session/} for full documentation.
}
\keyword{internal}
