% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entityresolution_operations.R
\name{entityresolution_create_matching_workflow}
\alias{entityresolution_create_matching_workflow}
\title{Creates a matching workflow that defines the configuration for a data
processing job}
\usage{
entityresolution_create_matching_workflow(
  workflowName,
  description = NULL,
  inputSourceConfig,
  outputSourceConfig,
  resolutionTechniques,
  incrementalRunConfig = NULL,
  roleArn,
  tags = NULL
)
}
\arguments{
\item{workflowName}{[required] The name of the workflow. There can't be multiple \code{MatchingWorkflows} with the same name.}

\item{description}{A description of the workflow.}

\item{inputSourceConfig}{[required] A list of \code{InputSource} objects, which have the fields \code{InputSourceARN} and \code{SchemaName}.}

\item{outputSourceConfig}{[required] A list of \code{OutputSource} objects, each of which contains fields \code{outputS3Path}, \code{applyNormalization}, \code{KMSArn}, and \code{output}.}

\item{resolutionTechniques}{[required] An object which defines the \code{resolutionType} and the \code{ruleBasedProperties}.}

\item{incrementalRunConfig}{Optional. An object that defines the incremental run type. This object contains only the \code{incrementalRunType} field, which appears as "Automatic" in the console.

For workflows where \code{resolutionType} is \code{PROVIDER}, incremental processing is not supported.}

\item{roleArn}{[required] The Amazon Resource Name (ARN) of the IAM role. Entity Resolution assumes this role to create resources on your behalf as part of workflow execution.}

\item{tags}{The tags used to organize, track, or control access for this resource.}
}
\description{
Creates a matching workflow that defines the configuration for a data processing job. The workflow name must be unique. To modify an existing workflow, use \code{\link[=entityresolution_update_matching_workflow]{update_matching_workflow}}.

See \url{https://www.paws-r-sdk.com/docs/entityresolution_create_matching_workflow/} for full documentation.
}
\keyword{internal}
