% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entityresolution_operations.R
\name{entityresolution_delete_id_mapping_workflow}
\alias{entityresolution_delete_id_mapping_workflow}
\title{Deletes the IdMappingWorkflow with a given name}
\usage{
entityresolution_delete_id_mapping_workflow(workflowName)
}
\arguments{
\item{workflowName}{[required] The name of the workflow to be deleted.}
}
\description{
Deletes the \code{IdMappingWorkflow} with a given name. This operation returns a \code{ResourceNotFoundException} if a workflow with the given name does not exist.

See \url{https://www.paws-r-sdk.com/docs/entityresolution_delete_id_mapping_workflow/} for full documentation.
}
\keyword{internal}
