% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entityresolution_operations.R
\name{entityresolution_delete_policy_statement}
\alias{entityresolution_delete_policy_statement}
\title{Deletes the policy statement}
\usage{
entityresolution_delete_policy_statement(arn, statementId)
}
\arguments{
\item{arn}{[required] The ARN of the resource for which the policy need to be deleted.}

\item{statementId}{[required] A statement identifier that differentiates the statement from others in
the same policy.}
}
\description{
Deletes the policy statement.

See \url{https://www.paws-r-sdk.com/docs/entityresolution_delete_policy_statement/} for full documentation.
}
\keyword{internal}
