% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entityresolution_operations.R
\name{entityresolution_get_id_mapping_job}
\alias{entityresolution_get_id_mapping_job}
\title{Returns the status, metrics, and errors (if there are any) that are
associated with a job}
\usage{
entityresolution_get_id_mapping_job(workflowName, jobId)
}
\arguments{
\item{workflowName}{[required] The name of the workflow.}

\item{jobId}{[required] The ID of the job.}
}
\description{
Returns the status, metrics, and errors (if there are any) that are associated with a job.

See \url{https://www.paws-r-sdk.com/docs/entityresolution_get_id_mapping_job/} for full documentation.
}
\keyword{internal}
