% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entityresolution_operations.R
\name{entityresolution_get_id_namespace}
\alias{entityresolution_get_id_namespace}
\title{Returns the IdNamespace with a given name, if it exists}
\usage{
entityresolution_get_id_namespace(idNamespaceName)
}
\arguments{
\item{idNamespaceName}{[required] The name of the ID namespace.}
}
\description{
Returns the \code{IdNamespace} with a given name, if it exists.

See \url{https://www.paws-r-sdk.com/docs/entityresolution_get_id_namespace/} for full documentation.
}
\keyword{internal}
