% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entityresolution_operations.R
\name{entityresolution_get_matching_workflow}
\alias{entityresolution_get_matching_workflow}
\title{Returns the MatchingWorkflow with a given name, if it exists}
\usage{
entityresolution_get_matching_workflow(workflowName)
}
\arguments{
\item{workflowName}{[required] The name of the workflow.}
}
\description{
Returns the \code{MatchingWorkflow} with a given name, if it exists.

See \url{https://www.paws-r-sdk.com/docs/entityresolution_get_matching_workflow/} for full documentation.
}
\keyword{internal}
