% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entityresolution_operations.R
\name{entityresolution_get_policy}
\alias{entityresolution_get_policy}
\title{Returns the resource-based policy}
\usage{
entityresolution_get_policy(arn)
}
\arguments{
\item{arn}{[required] The Amazon Resource Name (ARN) of the resource for which the policy need
to be returned.}
}
\description{
Returns the resource-based policy.

See \url{https://www.paws-r-sdk.com/docs/entityresolution_get_policy/} for full documentation.
}
\keyword{internal}
