% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entityresolution_operations.R
\name{entityresolution_get_provider_service}
\alias{entityresolution_get_provider_service}
\title{Returns the ProviderService of a given name}
\usage{
entityresolution_get_provider_service(providerName, providerServiceName)
}
\arguments{
\item{providerName}{[required] The name of the provider. This name is typically the company name.}

\item{providerServiceName}{[required] The ARN (Amazon Resource Name) of the product that the provider service
provides.}
}
\description{
Returns the \code{ProviderService} of a given name.

See \url{https://www.paws-r-sdk.com/docs/entityresolution_get_provider_service/} for full documentation.
}
\keyword{internal}
