% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entityresolution_operations.R
\name{entityresolution_get_schema_mapping}
\alias{entityresolution_get_schema_mapping}
\title{Returns the SchemaMapping of a given name}
\usage{
entityresolution_get_schema_mapping(schemaName)
}
\arguments{
\item{schemaName}{[required] The name of the schema to be retrieved.}
}
\description{
Returns the SchemaMapping of a given name.

See \url{https://www.paws-r-sdk.com/docs/entityresolution_get_schema_mapping/} for full documentation.
}
\keyword{internal}
