% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entityresolution_operations.R
\name{entityresolution_list_id_namespaces}
\alias{entityresolution_list_id_namespaces}
\title{Returns a list of all ID namespaces}
\usage{
entityresolution_list_id_namespaces(maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{maxResults}{The maximum number of \code{IdNamespace} objects returned per page.}

\item{nextToken}{The pagination token from the previous API call.}
}
\description{
Returns a list of all ID namespaces.

See \url{https://www.paws-r-sdk.com/docs/entityresolution_list_id_namespaces/} for full documentation.
}
\keyword{internal}
