% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entityresolution_operations.R
\name{entityresolution_list_matching_workflows}
\alias{entityresolution_list_matching_workflows}
\title{Returns a list of all the MatchingWorkflows that have been created for
an Amazon Web Services account}
\usage{
entityresolution_list_matching_workflows(maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{maxResults}{The maximum number of objects returned per page.}

\item{nextToken}{The pagination token from the previous API call.}
}
\description{
Returns a list of all the \code{MatchingWorkflows} that have been created for an Amazon Web Services account.

See \url{https://www.paws-r-sdk.com/docs/entityresolution_list_matching_workflows/} for full documentation.
}
\keyword{internal}
