% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entityresolution_operations.R
\name{entityresolution_list_provider_services}
\alias{entityresolution_list_provider_services}
\title{Returns a list of all the ProviderServices that are available in this
Amazon Web Services Region}
\usage{
entityresolution_list_provider_services(
  maxResults = NULL,
  nextToken = NULL,
  providerName = NULL
)
}
\arguments{
\item{maxResults}{The maximum number of objects returned per page.}

\item{nextToken}{The pagination token from the previous API call.}

\item{providerName}{The name of the provider. This name is typically the company name.}
}
\description{
Returns a list of all the \code{ProviderServices} that are available in this Amazon Web Services Region.

See \url{https://www.paws-r-sdk.com/docs/entityresolution_list_provider_services/} for full documentation.
}
\keyword{internal}
