% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entityresolution_operations.R
\name{entityresolution_list_schema_mappings}
\alias{entityresolution_list_schema_mappings}
\title{Returns a list of all the SchemaMappings that have been created for an
Amazon Web Services account}
\usage{
entityresolution_list_schema_mappings(nextToken = NULL, maxResults = NULL)
}
\arguments{
\item{nextToken}{The pagination token from the previous API call.}

\item{maxResults}{The maximum number of objects returned per page.}
}
\description{
Returns a list of all the \code{SchemaMappings} that have been created for an Amazon Web Services account.

See \url{https://www.paws-r-sdk.com/docs/entityresolution_list_schema_mappings/} for full documentation.
}
\keyword{internal}
