% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entityresolution_operations.R
\name{entityresolution_list_tags_for_resource}
\alias{entityresolution_list_tags_for_resource}
\title{Displays the tags associated with an Entity Resolution resource}
\usage{
entityresolution_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The ARN of the resource for which you want to view tags.}
}
\description{
Displays the tags associated with an Entity Resolution resource. In Entity Resolution, \code{SchemaMapping}, and \code{MatchingWorkflow} can be tagged.

See \url{https://www.paws-r-sdk.com/docs/entityresolution_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
