% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entityresolution_operations.R
\name{entityresolution_put_policy}
\alias{entityresolution_put_policy}
\title{Updates the resource-based policy}
\usage{
entityresolution_put_policy(arn, token = NULL, policy)
}
\arguments{
\item{arn}{[required] The Amazon Resource Name (ARN) of the resource for which the policy
needs to be updated.}

\item{token}{A unique identifier for the current revision of the policy.}

\item{policy}{[required] The resource-based policy.

If you set the value of the \code{effect} parameter in the \code{policy} to \code{Deny}
for the \code{\link[=entityresolution_put_policy]{put_policy}} operation, you must
also set the value of the \code{effect} parameter to \code{Deny} for the
\code{\link[=entityresolution_add_policy_statement]{add_policy_statement}}
operation.}
}
\description{
Updates the resource-based policy.

See \url{https://www.paws-r-sdk.com/docs/entityresolution_put_policy/} for full documentation.
}
\keyword{internal}
