% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entityresolution_operations.R
\name{entityresolution_start_id_mapping_job}
\alias{entityresolution_start_id_mapping_job}
\title{Starts the IdMappingJob of a workflow}
\usage{
entityresolution_start_id_mapping_job(outputSourceConfig = NULL, workflowName)
}
\arguments{
\item{outputSourceConfig}{A list of \code{OutputSource} objects.}

\item{workflowName}{[required] The name of the ID mapping job to be retrieved.}
}
\description{
Starts the \code{IdMappingJob} of a workflow. The workflow must have previously been created using the \code{\link[=entityresolution_create_id_mapping_workflow]{create_id_mapping_workflow}} endpoint.

See \url{https://www.paws-r-sdk.com/docs/entityresolution_start_id_mapping_job/} for full documentation.
}
\keyword{internal}
