% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entityresolution_operations.R
\name{entityresolution_start_matching_job}
\alias{entityresolution_start_matching_job}
\title{Starts the MatchingJob of a workflow}
\usage{
entityresolution_start_matching_job(workflowName)
}
\arguments{
\item{workflowName}{[required] The name of the matching job to be retrieved.}
}
\description{
Starts the \code{MatchingJob} of a workflow. The workflow must have previously been created using the \code{\link[=entityresolution_create_matching_workflow]{create_matching_workflow}} endpoint.

See \url{https://www.paws-r-sdk.com/docs/entityresolution_start_matching_job/} for full documentation.
}
\keyword{internal}
