% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entityresolution_operations.R
\name{entityresolution_untag_resource}
\alias{entityresolution_untag_resource}
\title{Removes one or more tags from the specified Entity Resolution resource}
\usage{
entityresolution_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The ARN of the resource for which you want to untag.}

\item{tagKeys}{[required] The list of tag keys to remove from the resource.}
}
\description{
Removes one or more tags from the specified Entity Resolution resource. In Entity Resolution, \code{SchemaMapping}, and \code{MatchingWorkflow} can be tagged.

See \url{https://www.paws-r-sdk.com/docs/entityresolution_untag_resource/} for full documentation.
}
\keyword{internal}
