% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entityresolution_operations.R
\name{entityresolution_update_id_namespace}
\alias{entityresolution_update_id_namespace}
\title{Updates an existing ID namespace}
\usage{
entityresolution_update_id_namespace(
  idNamespaceName,
  description = NULL,
  inputSourceConfig = NULL,
  idMappingWorkflowProperties = NULL,
  roleArn = NULL
)
}
\arguments{
\item{idNamespaceName}{[required] The name of the ID namespace.}

\item{description}{The description of the ID namespace.}

\item{inputSourceConfig}{A list of \code{InputSource} objects, which have the fields \code{InputSourceARN}
and \code{SchemaName}.}

\item{idMappingWorkflowProperties}{Determines the properties of \code{IdMappingWorkflow} where this
\code{IdNamespace} can be used as a \code{Source} or a \code{Target}.}

\item{roleArn}{The Amazon Resource Name (ARN) of the IAM role. Entity Resolution
assumes this role to access the resources defined in this \code{IdNamespace}
on your behalf as part of a workflow run.}
}
\description{
Updates an existing ID namespace.

See \url{https://www.paws-r-sdk.com/docs/entityresolution_update_id_namespace/} for full documentation.
}
\keyword{internal}
