% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entityresolution_operations.R
\name{entityresolution_update_schema_mapping}
\alias{entityresolution_update_schema_mapping}
\title{Updates a schema mapping}
\usage{
entityresolution_update_schema_mapping(
  description = NULL,
  mappedInputFields,
  schemaName
)
}
\arguments{
\item{description}{A description of the schema.}

\item{mappedInputFields}{[required] A list of \code{MappedInputFields}. Each \code{MappedInputField} corresponds to a
column the source data table, and contains column name plus additional
information that Entity Resolution uses for matching.}

\item{schemaName}{[required] The name of the schema. There can't be multiple \code{SchemaMappings} with
the same name.}
}
\description{
Updates a schema mapping.

See \url{https://www.paws-r-sdk.com/docs/entityresolution_update_schema_mapping/} for full documentation.
}
\keyword{internal}
