% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firehose_operations.R
\name{firehose_put_record}
\alias{firehose_put_record}
\title{Writes a single data record into an Amazon Kinesis Data Firehose
delivery stream}
\usage{
firehose_put_record(DeliveryStreamName, Record)
}
\arguments{
\item{DeliveryStreamName}{[required] The name of the delivery stream.}

\item{Record}{[required] The record.}
}
\description{
Writes a single data record into an Amazon Kinesis Data Firehose delivery stream. To write multiple data records into a delivery stream, use \code{\link[=firehose_put_record_batch]{put_record_batch}}. Applications using these operations are referred to as producers.

See \url{https://paws-r.github.io/docs/firehose/put_record.html} for full documentation.
}
\keyword{internal}
