% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_associate_glossary_terms}
\alias{glue_associate_glossary_terms}
\title{Associates one or more glossary terms with an asset in Glue Data Catalog}
\usage{
glue_associate_glossary_terms(
  AssetIdentifier,
  IterableFormName = NULL,
  ItemIdentifier = NULL,
  GlossaryTermIdentifiers,
  ClientToken = NULL
)
}
\arguments{
\item{AssetIdentifier}{[required] The unique identifier of the asset to associate glossary terms with.}

\item{IterableFormName}{The name of the iterable form. When specified along with \code{itemIdentifier}, the glossary terms are associated with an item within the iterable form rather than the asset itself.}

\item{ItemIdentifier}{The identifier of the item within the iterable form. Required when \code{iterableFormName} is specified.}

\item{GlossaryTermIdentifiers}{[required] The list of glossary term identifiers to associate with the asset.}

\item{ClientToken}{A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.}
}
\description{
Associates one or more glossary terms with an asset in Glue Data Catalog.

See \url{https://www.paws-r-sdk.com/docs/glue_associate_glossary_terms/} for full documentation.
}
\keyword{internal}
