% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_batch_create_partition}
\alias{glue_batch_create_partition}
\title{Creates one or more partitions in a batch operation}
\usage{
glue_batch_create_partition(
  CatalogId = NULL,
  DatabaseName,
  TableName,
  PartitionInputList
)
}
\arguments{
\item{CatalogId}{The ID of the catalog in which the partition is to be created.
Currently, this should be the Amazon Web Services account ID.}

\item{DatabaseName}{[required] The name of the metadata database in which the partition is to be
created.}

\item{TableName}{[required] The name of the metadata table in which the partition is to be created.}

\item{PartitionInputList}{[required] A list of \code{PartitionInput} structures that define the partitions to be
created.}
}
\description{
Creates one or more partitions in a batch operation.

See \url{https://www.paws-r-sdk.com/docs/glue_batch_create_partition/} for full documentation.
}
\keyword{internal}
