% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_batch_delete_connection}
\alias{glue_batch_delete_connection}
\title{Deletes a list of connection definitions from the Data Catalog}
\usage{
glue_batch_delete_connection(CatalogId, ConnectionNameList)
}
\arguments{
\item{CatalogId}{The ID of the Data Catalog in which the connections reside. If none is
provided, the AWS account ID is used by default.}

\item{ConnectionNameList}{[required] A list of names of the connections to delete.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Succeeded = list(
    "string"
  ),
  Errors = list(
    list(
      ErrorCode = "string",
      ErrorMessage = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Deletes a list of connection definitions from the Data Catalog.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$batch_delete_connection(
  CatalogId = "string",
  ConnectionNameList = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
