% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_batch_delete_table}
\alias{glue_batch_delete_table}
\title{Deletes multiple tables at once}
\usage{
glue_batch_delete_table(
  CatalogId = NULL,
  DatabaseName,
  TablesToDelete,
  TransactionId = NULL
)
}
\arguments{
\item{CatalogId}{The ID of the Data Catalog where the table resides. If none is provided,
the Amazon Web Services account ID is used by default.}

\item{DatabaseName}{[required] The name of the catalog database in which the tables to delete reside.
For Hive compatibility, this name is entirely lowercase.}

\item{TablesToDelete}{[required] A list of the table to delete.}

\item{TransactionId}{The transaction ID at which to delete the table contents.}
}
\description{
Deletes multiple tables at once.

See \url{https://paws-r.github.io/docs/glue/batch_delete_table.html} for full documentation.
}
\keyword{internal}
