% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_batch_get_blueprints}
\alias{glue_batch_get_blueprints}
\title{Retrieves information about a list of blueprints}
\usage{
glue_batch_get_blueprints(
  Names,
  IncludeBlueprint = NULL,
  IncludeParameterSpec = NULL
)
}
\arguments{
\item{Names}{[required] A list of blueprint names.}

\item{IncludeBlueprint}{Specifies whether or not to include the blueprint in the response.}

\item{IncludeParameterSpec}{Specifies whether or not to include the parameters, as a JSON string,
for the blueprint in the response.}
}
\description{
Retrieves information about a list of blueprints.

See \url{https://www.paws-r-sdk.com/docs/glue_batch_get_blueprints/} for full documentation.
}
\keyword{internal}
