% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_batch_get_custom_entity_types}
\alias{glue_batch_get_custom_entity_types}
\title{Retrieves the details for the custom patterns specified by a list of
names}
\usage{
glue_batch_get_custom_entity_types(Names)
}
\arguments{
\item{Names}{[required] A list of names of the custom patterns that you want to retrieve.}
}
\description{
Retrieves the details for the custom patterns specified by a list of names.

See \url{https://www.paws-r-sdk.com/docs/glue_batch_get_custom_entity_types/} for full documentation.
}
\keyword{internal}
