% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_batch_get_data_quality_result}
\alias{glue_batch_get_data_quality_result}
\title{Retrieves a list of data quality results for the specified result IDs}
\usage{
glue_batch_get_data_quality_result(ResultIds)
}
\arguments{
\item{ResultIds}{[required] A list of unique result IDs for the data quality results.}
}
\description{
Retrieves a list of data quality results for the specified result IDs.

See \url{https://paws-r.github.io/docs/glue/batch_get_data_quality_result.html} for full documentation.
}
\keyword{internal}
