% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_batch_get_data_quality_ruleset_evaluation_run}
\alias{glue_batch_get_data_quality_ruleset_evaluation_run}
\title{Retrieves the details of multiple evaluation runs in a single request}
\usage{
glue_batch_get_data_quality_ruleset_evaluation_run(RunIds)
}
\arguments{
\item{RunIds}{[required] A list of unique run identifiers for the evaluation runs to retrieve.}
}
\description{
Retrieves the details of multiple evaluation runs in a single request.

See \url{https://www.paws-r-sdk.com/docs/glue_batch_get_data_quality_ruleset_evaluation_run/} for full documentation.
}
\keyword{internal}
