% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_batch_get_iterable_forms}
\alias{glue_batch_get_iterable_forms}
\title{Retrieves multiple items from an iterable form on an asset in Glue Data
Catalog in a single request}
\usage{
glue_batch_get_iterable_forms(
  AssetIdentifier,
  IterableFormName,
  ItemIdentifiers
)
}
\arguments{
\item{AssetIdentifier}{[required] The unique identifier of the asset.}

\item{IterableFormName}{[required] The name of the iterable form to retrieve items from.}

\item{ItemIdentifiers}{[required] The list of item identifiers to retrieve. Each identifier can be an item ID or item name.}
}
\description{
Retrieves multiple items from an iterable form on an asset in Glue Data Catalog in a single request.

See \url{https://www.paws-r-sdk.com/docs/glue_batch_get_iterable_forms/} for full documentation.
}
\keyword{internal}
