% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_batch_get_table_optimizer}
\alias{glue_batch_get_table_optimizer}
\title{Returns the configuration for the specified table optimizers}
\usage{
glue_batch_get_table_optimizer(Entries)
}
\arguments{
\item{Entries}{[required] A list of \code{BatchGetTableOptimizerEntry} objects specifying the table
optimizers to retrieve.}
}
\description{
Returns the configuration for the specified table optimizers.

See \url{https://www.paws-r-sdk.com/docs/glue_batch_get_table_optimizer/} for full documentation.
}
\keyword{internal}
