% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_batch_get_workflows}
\alias{glue_batch_get_workflows}
\title{Returns a list of resource metadata for a given list of workflow names}
\usage{
glue_batch_get_workflows(Names, IncludeGraph = NULL)
}
\arguments{
\item{Names}{[required] A list of workflow names, which may be the names returned from the
\code{\link[=glue_list_workflows]{list_workflows}} operation.}

\item{IncludeGraph}{Specifies whether to include a graph when returning the workflow
resource metadata.}
}
\description{
Returns a list of resource metadata for a given list of workflow names. After calling the \code{\link[=glue_list_workflows]{list_workflows}} operation, you can call this operation to access the data to which you have been granted permissions. This operation supports all IAM permissions, including permission conditions that uses tags.

See \url{https://www.paws-r-sdk.com/docs/glue_batch_get_workflows/} for full documentation.
}
\keyword{internal}
