% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_batch_put_data_quality_statistic_annotation}
\alias{glue_batch_put_data_quality_statistic_annotation}
\title{Annotate datapoints over time for a specific data quality statistic}
\usage{
glue_batch_put_data_quality_statistic_annotation(
  InclusionAnnotations,
  ClientToken = NULL
)
}
\arguments{
\item{InclusionAnnotations}{[required] A list of \code{DatapointInclusionAnnotation}'s. The InclusionAnnotations
must contain a profileId and statisticId. If there are multiple
InclusionAnnotations, the list must refer to a single statisticId across
multiple profileIds.}

\item{ClientToken}{Client Token.}
}
\description{
Annotate datapoints over time for a specific data quality statistic. The API requires both profileID and statisticID as part of the InclusionAnnotation input. The API only works for a single statisticId across multiple profiles.

See \url{https://www.paws-r-sdk.com/docs/glue_batch_put_data_quality_statistic_annotation/} for full documentation.
}
\keyword{internal}
