% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_batch_update_partition}
\alias{glue_batch_update_partition}
\title{Updates one or more partitions in a batch operation}
\usage{
glue_batch_update_partition(CatalogId = NULL, DatabaseName, TableName, Entries)
}
\arguments{
\item{CatalogId}{The ID of the catalog in which the partition is to be updated.
Currently, this should be the Amazon Web Services account ID.}

\item{DatabaseName}{[required] The name of the metadata database in which the partition is to be
updated.}

\item{TableName}{[required] The name of the metadata table in which the partition is to be updated.}

\item{Entries}{[required] A list of up to 100 \code{BatchUpdatePartitionRequestEntry} objects to
update.}
}
\description{
Updates one or more partitions in a batch operation.

See \url{https://paws-r.github.io/docs/glue/batch_update_partition.html} for full documentation.
}
\keyword{internal}
