% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_cancel_data_quality_rule_recommendation_run}
\alias{glue_cancel_data_quality_rule_recommendation_run}
\title{Cancels the specified recommendation run that was being used to generate
rules}
\usage{
glue_cancel_data_quality_rule_recommendation_run(RunId)
}
\arguments{
\item{RunId}{[required] The unique run identifier associated with this run.}
}
\description{
Cancels the specified recommendation run that was being used to generate rules.

See \url{https://paws-r.github.io/docs/glue/cancel_data_quality_rule_recommendation_run.html} for full documentation.
}
\keyword{internal}
