% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_cancel_ml_task_run}
\alias{glue_cancel_ml_task_run}
\title{Cancels (stops) a task run}
\usage{
glue_cancel_ml_task_run(TransformId, TaskRunId)
}
\arguments{
\item{TransformId}{[required] The unique identifier of the machine learning transform.}

\item{TaskRunId}{[required] A unique identifier for the task run.}
}
\description{
Cancels (stops) a task run. Machine learning task runs are asynchronous tasks that Glue runs on your behalf as part of various machine learning workflows. You can cancel a machine learning task run at any time by calling \code{\link[=glue_cancel_ml_task_run]{cancel_ml_task_run}} with a task run's parent transform's \code{TransformID} and the task run's \code{TaskRunId}.

See \url{https://paws-r.github.io/docs/glue/cancel_ml_task_run.html} for full documentation.
}
\keyword{internal}
