% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_cancel_statement}
\alias{glue_cancel_statement}
\title{Cancels the statement}
\usage{
glue_cancel_statement(SessionId, Id, RequestOrigin = NULL)
}
\arguments{
\item{SessionId}{[required] The Session ID of the statement to be cancelled.}

\item{Id}{[required] The ID of the statement to be cancelled.}

\item{RequestOrigin}{The origin of the request to cancel the statement.}
}
\description{
Cancels the statement.

See \url{https://www.paws-r-sdk.com/docs/glue_cancel_statement/} for full documentation.
}
\keyword{internal}
