% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_create_blueprint}
\alias{glue_create_blueprint}
\title{Registers a blueprint with Glue}
\usage{
glue_create_blueprint(Name, Description = NULL, BlueprintLocation, Tags = NULL)
}
\arguments{
\item{Name}{[required] The name of the blueprint.}

\item{Description}{A description of the blueprint.}

\item{BlueprintLocation}{[required] Specifies a path in Amazon S3 where the blueprint is published.}

\item{Tags}{The tags to be applied to this blueprint.}
}
\description{
Registers a blueprint with Glue.

See \url{https://paws-r.github.io/docs/glue/create_blueprint.html} for full documentation.
}
\keyword{internal}
