% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_create_connection}
\alias{glue_create_connection}
\title{Creates a connection definition in the Data Catalog}
\usage{
glue_create_connection(CatalogId = NULL, ConnectionInput, Tags = NULL)
}
\arguments{
\item{CatalogId}{The ID of the Data Catalog in which to create the connection. If none is
provided, the Amazon Web Services account ID is used by default.}

\item{ConnectionInput}{[required] A \code{ConnectionInput} object defining the connection to create.}

\item{Tags}{The tags you assign to the connection.}
}
\description{
Creates a connection definition in the Data Catalog.

See \url{https://www.paws-r-sdk.com/docs/glue_create_connection/} for full documentation.
}
\keyword{internal}
