% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_create_glossary}
\alias{glue_create_glossary}
\title{Creates a business glossary in Glue Data Catalog}
\usage{
glue_create_glossary(Name, Description = NULL, ClientToken = NULL)
}
\arguments{
\item{Name}{[required] The name of the glossary.}

\item{Description}{The description of the glossary.}

\item{ClientToken}{A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.}
}
\description{
Creates a business glossary in Glue Data Catalog. A glossary is a container for glossary terms that define business concepts.

See \url{https://www.paws-r-sdk.com/docs/glue_create_glossary/} for full documentation.
}
\keyword{internal}
