% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_create_glossary_term}
\alias{glue_create_glossary_term}
\title{Creates a glossary term within a business glossary in Glue Data Catalog}
\usage{
glue_create_glossary_term(
  GlossaryIdentifier,
  Name,
  ShortDescription = NULL,
  LongDescription = NULL,
  ClientToken = NULL
)
}
\arguments{
\item{GlossaryIdentifier}{[required] The unique identifier of the glossary in which to create the term.}

\item{Name}{[required] The name of the glossary term.}

\item{ShortDescription}{A short description of the glossary term.}

\item{LongDescription}{A long description of the glossary term.}

\item{ClientToken}{A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.}
}
\description{
Creates a glossary term within a business glossary in Glue Data Catalog.

See \url{https://www.paws-r-sdk.com/docs/glue_create_glossary_term/} for full documentation.
}
\keyword{internal}
