% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_create_glue_identity_center_configuration}
\alias{glue_create_glue_identity_center_configuration}
\title{Creates a new Glue Identity Center configuration to enable integration
between Glue and Amazon Web Services IAM Identity Center for
authentication and authorization}
\usage{
glue_create_glue_identity_center_configuration(
  InstanceArn,
  Scopes = NULL,
  UserBackgroundSessionsEnabled = NULL
)
}
\arguments{
\item{InstanceArn}{[required] The Amazon Resource Name (ARN) of the Identity Center instance to be
associated with the Glue configuration.}

\item{Scopes}{A list of Identity Center scopes that define the permissions and access
levels for the Glue configuration.}

\item{UserBackgroundSessionsEnabled}{Specifies whether users can run background sessions when using Identity
Center authentication with Glue services.}
}
\description{
Creates a new Glue Identity Center configuration to enable integration between Glue and Amazon Web Services IAM Identity Center for authentication and authorization.

See \url{https://www.paws-r-sdk.com/docs/glue_create_glue_identity_center_configuration/} for full documentation.
}
\keyword{internal}
