% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_create_integration_resource_property}
\alias{glue_create_integration_resource_property}
\title{This API can be used for setting up the ResourceProperty of the Glue
connection (for the source) or Glue database ARN (for the target)}
\usage{
glue_create_integration_resource_property(
  ResourceArn,
  SourceProcessingProperties = NULL,
  TargetProcessingProperties = NULL,
  Tags = NULL
)
}
\arguments{
\item{ResourceArn}{[required] The connection ARN of the source, or the database ARN of the target.}

\item{SourceProcessingProperties}{The resource properties associated with the integration source.}

\item{TargetProcessingProperties}{The resource properties associated with the integration target.}

\item{Tags}{Metadata assigned to the resource consisting of a list of key-value
pairs.}
}
\description{
This API can be used for setting up the \code{ResourceProperty} of the Glue connection (for the source) or Glue database ARN (for the target). These properties can include the role to access the connection or database. To set both source and target properties the same API needs to be invoked with the Glue connection ARN as \code{ResourceArn} with \code{SourceProcessingProperties} and the Glue database ARN as \code{ResourceArn} with \code{TargetProcessingProperties} respectively.

See \url{https://www.paws-r-sdk.com/docs/glue_create_integration_resource_property/} for full documentation.
}
\keyword{internal}
