% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_create_partition}
\alias{glue_create_partition}
\title{Creates a new partition}
\usage{
glue_create_partition(
  CatalogId = NULL,
  DatabaseName,
  TableName,
  PartitionInput
)
}
\arguments{
\item{CatalogId}{The Amazon Web Services account ID of the catalog in which the partition
is to be created.}

\item{DatabaseName}{[required] The name of the metadata database in which the partition is to be
created.}

\item{TableName}{[required] The name of the metadata table in which the partition is to be created.}

\item{PartitionInput}{[required] A \code{PartitionInput} structure defining the partition to be created.}
}
\description{
Creates a new partition.

See \url{https://www.paws-r-sdk.com/docs/glue_create_partition/} for full documentation.
}
\keyword{internal}
