% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_create_registry}
\alias{glue_create_registry}
\title{Creates a new registry which may be used to hold a collection of schemas}
\usage{
glue_create_registry(RegistryName, Description, Tags)
}
\arguments{
\item{RegistryName}{[required] Name of the registry to be created of max length of 255, and may only
contain letters, numbers, hyphen, underscore, dollar sign, or hash mark.
No whitespace.}

\item{Description}{A description of the registry. If description is not provided, there
will not be any default value for this.}

\item{Tags}{AWS tags that contain a key value pair and may be searched by console,
command line, or API.}
}
\value{
A list with the following syntax:\preformatted{list(
  RegistryArn = "string",
  RegistryName = "string",
  Description = "string",
  Tags = list(
    "string"
  )
)
}
}
\description{
Creates a new registry which may be used to hold a collection of
schemas.
}
\section{Request syntax}{
\preformatted{svc$create_registry(
  RegistryName = "string",
  Description = "string",
  Tags = list(
    "string"
  )
)
}
}

\keyword{internal}
