% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_create_script}
\alias{glue_create_script}
\title{Transforms a directed acyclic graph (DAG) into code}
\usage{
glue_create_script(DagNodes = NULL, DagEdges = NULL, Language = NULL)
}
\arguments{
\item{DagNodes}{A list of the nodes in the DAG.}

\item{DagEdges}{A list of the edges in the DAG.}

\item{Language}{The programming language of the resulting code from the DAG.}
}
\description{
Transforms a directed acyclic graph (DAG) into code.

See \url{https://paws-r.github.io/docs/glue/create_script.html} for full documentation.
}
\keyword{internal}
