% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_create_security_configuration}
\alias{glue_create_security_configuration}
\title{Creates a new security configuration}
\usage{
glue_create_security_configuration(Name, EncryptionConfiguration)
}
\arguments{
\item{Name}{[required] The name for the new security configuration.}

\item{EncryptionConfiguration}{[required] The encryption configuration for the new security configuration.}
}
\description{
Creates a new security configuration. A security configuration is a set of security properties that can be used by Glue. You can use a security configuration to encrypt data at rest. For information about using security configurations in Glue, see \href{https://docs.aws.amazon.com/glue/latest/dg/encryption-security-configuration.html}{Encrypting Data Written by Crawlers, Jobs, and Development Endpoints}.

See \url{https://www.paws-r-sdk.com/docs/glue_create_security_configuration/} for full documentation.
}
\keyword{internal}
