% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_create_table_optimizer}
\alias{glue_create_table_optimizer}
\title{Creates a new table optimizer for a specific function}
\usage{
glue_create_table_optimizer(
  CatalogId,
  DatabaseName,
  TableName,
  Type,
  TableOptimizerConfiguration
)
}
\arguments{
\item{CatalogId}{[required] The Catalog ID of the table.}

\item{DatabaseName}{[required] The name of the database in the catalog in which the table resides.}

\item{TableName}{[required] The name of the table.}

\item{Type}{[required] The type of table optimizer.}

\item{TableOptimizerConfiguration}{[required] A \code{TableOptimizerConfiguration} object representing the configuration of
a table optimizer.}
}
\description{
Creates a new table optimizer for a specific function.

See \url{https://www.paws-r-sdk.com/docs/glue_create_table_optimizer/} for full documentation.
}
\keyword{internal}
