% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_create_usage_profile}
\alias{glue_create_usage_profile}
\title{Creates an Glue usage profile}
\usage{
glue_create_usage_profile(Name, Description = NULL, Configuration, Tags = NULL)
}
\arguments{
\item{Name}{[required] The name of the usage profile.}

\item{Description}{A description of the usage profile.}

\item{Configuration}{[required] A \code{ProfileConfiguration} object specifying the job and session values
for the profile.}

\item{Tags}{A list of tags applied to the usage profile.}
}
\description{
Creates an Glue usage profile.

See \url{https://www.paws-r-sdk.com/docs/glue_create_usage_profile/} for full documentation.
}
\keyword{internal}
