% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_delete_asset}
\alias{glue_delete_asset}
\title{Deletes an asset from Glue Data Catalog}
\usage{
glue_delete_asset(Identifier)
}
\arguments{
\item{Identifier}{[required] The unique identifier of the asset to delete.}
}
\description{
Deletes an asset from Glue Data Catalog.

See \url{https://www.paws-r-sdk.com/docs/glue_delete_asset/} for full documentation.
}
\keyword{internal}
