% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_delete_asset_type}
\alias{glue_delete_asset_type}
\title{Deletes an asset type from Glue Data Catalog}
\usage{
glue_delete_asset_type(Identifier)
}
\arguments{
\item{Identifier}{[required] The identifier of the asset type to delete.}
}
\description{
Deletes an asset type from Glue Data Catalog.

See \url{https://www.paws-r-sdk.com/docs/glue_delete_asset_type/} for full documentation.
}
\keyword{internal}
