% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_delete_attachment}
\alias{glue_delete_attachment}
\title{Deletes a form attachment from an asset in Glue Data Catalog}
\usage{
glue_delete_attachment(
  AssetIdentifier,
  IterableFormName = NULL,
  ItemIdentifier = NULL,
  AttachmentName
)
}
\arguments{
\item{AssetIdentifier}{[required] The unique identifier of the asset from which to delete the attachment.}

\item{IterableFormName}{The name of the iterable form. When specified along with \code{itemIdentifier}, the attachment is deleted from an item within the iterable form rather than from the asset itself.}

\item{ItemIdentifier}{The identifier of the item within the iterable form. Required when \code{iterableFormName} is specified.}

\item{AttachmentName}{[required] The name of the attachment to delete.}
}
\description{
Deletes a form attachment from an asset in Glue Data Catalog.

See \url{https://www.paws-r-sdk.com/docs/glue_delete_attachment/} for full documentation.
}
\keyword{internal}
