% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_delete_blueprint}
\alias{glue_delete_blueprint}
\title{Deletes an existing blueprint}
\usage{
glue_delete_blueprint(Name)
}
\arguments{
\item{Name}{[required] The name of the blueprint to delete.}
}
\description{
Deletes an existing blueprint.

See \url{https://paws-r.github.io/docs/glue/delete_blueprint.html} for full documentation.
}
\keyword{internal}
