% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_delete_catalog}
\alias{glue_delete_catalog}
\title{Removes the specified catalog from the Glue Data Catalog}
\usage{
glue_delete_catalog(CatalogId)
}
\arguments{
\item{CatalogId}{[required] The ID of the catalog.}
}
\description{
Removes the specified catalog from the Glue Data Catalog.

See \url{https://www.paws-r-sdk.com/docs/glue_delete_catalog/} for full documentation.
}
\keyword{internal}
