% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_delete_column_statistics_for_partition}
\alias{glue_delete_column_statistics_for_partition}
\title{Delete the partition column statistics of a column}
\usage{
glue_delete_column_statistics_for_partition(
  CatalogId = NULL,
  DatabaseName,
  TableName,
  PartitionValues,
  ColumnName
)
}
\arguments{
\item{CatalogId}{The ID of the Data Catalog where the partitions in question reside. If
none is supplied, the Amazon Web Services account ID is used by default.}

\item{DatabaseName}{[required] The name of the catalog database where the partitions reside.}

\item{TableName}{[required] The name of the partitions' table.}

\item{PartitionValues}{[required] A list of partition values identifying the partition.}

\item{ColumnName}{[required] Name of the column.}
}
\description{
Delete the partition column statistics of a column.

See \url{https://paws-r.github.io/docs/glue/delete_column_statistics_for_partition.html} for full documentation.
}
\keyword{internal}
