% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_delete_column_statistics_for_table}
\alias{glue_delete_column_statistics_for_table}
\title{Retrieves table statistics of columns}
\usage{
glue_delete_column_statistics_for_table(
  CatalogId = NULL,
  DatabaseName,
  TableName,
  ColumnName
)
}
\arguments{
\item{CatalogId}{The ID of the Data Catalog where the partitions in question reside. If
none is supplied, the Amazon Web Services account ID is used by default.}

\item{DatabaseName}{[required] The name of the catalog database where the partitions reside.}

\item{TableName}{[required] The name of the partitions' table.}

\item{ColumnName}{[required] The name of the column.}
}
\description{
Retrieves table statistics of columns.

See \url{https://www.paws-r-sdk.com/docs/glue_delete_column_statistics_for_table/} for full documentation.
}
\keyword{internal}
