% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_delete_connection}
\alias{glue_delete_connection}
\title{Deletes a connection from the Data Catalog}
\usage{
glue_delete_connection(CatalogId, ConnectionName)
}
\arguments{
\item{CatalogId}{The ID of the Data Catalog in which the connection resides. If none is
provided, the AWS account ID is used by default.}

\item{ConnectionName}{[required] The name of the connection to delete.}
}
\value{
An empty list.
}
\description{
Deletes a connection from the Data Catalog.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_connection(
  CatalogId = "string",
  ConnectionName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
