% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_delete_connection_type}
\alias{glue_delete_connection_type}
\title{Deletes a custom connection type in Glue}
\usage{
glue_delete_connection_type(ConnectionType)
}
\arguments{
\item{ConnectionType}{[required] The name of the connection type to delete. Must reference an existing
registered connection type.}
}
\description{
Deletes a custom connection type in Glue.

See \url{https://www.paws-r-sdk.com/docs/glue_delete_connection_type/} for full documentation.
}
\keyword{internal}
