% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_delete_crawler}
\alias{glue_delete_crawler}
\title{Removes a specified crawler from the Glue Data Catalog, unless the
crawler state is RUNNING}
\usage{
glue_delete_crawler(Name)
}
\arguments{
\item{Name}{[required] The name of the crawler to remove.}
}
\description{
Removes a specified crawler from the Glue Data Catalog, unless the crawler state is \code{RUNNING}.

See \url{https://paws-r.github.io/docs/glue/delete_crawler.html} for full documentation.
}
\keyword{internal}
