% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_delete_custom_entity_type}
\alias{glue_delete_custom_entity_type}
\title{Deletes a custom pattern by specifying its name}
\usage{
glue_delete_custom_entity_type(Name)
}
\arguments{
\item{Name}{[required] The name of the custom pattern that you want to delete.}
}
\description{
Deletes a custom pattern by specifying its name.

See \url{https://www.paws-r-sdk.com/docs/glue_delete_custom_entity_type/} for full documentation.
}
\keyword{internal}
