% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_delete_data_quality_ruleset}
\alias{glue_delete_data_quality_ruleset}
\title{Deletes a data quality ruleset}
\usage{
glue_delete_data_quality_ruleset(Name)
}
\arguments{
\item{Name}{[required] A name for the data quality ruleset.}
}
\description{
Deletes a data quality ruleset.

See \url{https://paws-r.github.io/docs/glue/delete_data_quality_ruleset.html} for full documentation.
}
\keyword{internal}
