% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_delete_database}
\alias{glue_delete_database}
\title{Removes a specified database from a Data Catalog}
\usage{
glue_delete_database(CatalogId = NULL, Name)
}
\arguments{
\item{CatalogId}{The ID of the Data Catalog in which the database resides. If none is
provided, the Amazon Web Services account ID is used by default.}

\item{Name}{[required] The name of the database to delete. For Hive compatibility, this must be
all lowercase.}
}
\description{
Removes a specified database from a Data Catalog.

See \url{https://paws-r.github.io/docs/glue/delete_database.html} for full documentation.
}
\keyword{internal}
