% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_delete_dev_endpoint}
\alias{glue_delete_dev_endpoint}
\title{Deletes a specified development endpoint}
\usage{
glue_delete_dev_endpoint(EndpointName)
}
\arguments{
\item{EndpointName}{[required] The name of the \code{DevEndpoint}.}
}
\value{
An empty list.
}
\description{
Deletes a specified development endpoint.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_dev_endpoint(
  EndpointName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
