% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_delete_form_type}
\alias{glue_delete_form_type}
\title{Deletes a form type from Glue Data Catalog}
\usage{
glue_delete_form_type(Identifier)
}
\arguments{
\item{Identifier}{[required] The identifier of the form type to delete.}
}
\description{
Deletes a form type from Glue Data Catalog. A form type cannot be deleted if it is still referenced by an asset type.

See \url{https://www.paws-r-sdk.com/docs/glue_delete_form_type/} for full documentation.
}
\keyword{internal}
