% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_delete_glossary}
\alias{glue_delete_glossary}
\title{Deletes a business glossary from Glue Data Catalog}
\usage{
glue_delete_glossary(Identifier)
}
\arguments{
\item{Identifier}{[required] The unique identifier of the glossary to delete.}
}
\description{
Deletes a business glossary from Glue Data Catalog. A glossary cannot be deleted if it still contains glossary terms.

See \url{https://www.paws-r-sdk.com/docs/glue_delete_glossary/} for full documentation.
}
\keyword{internal}
