% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_delete_glossary_term}
\alias{glue_delete_glossary_term}
\title{Deletes a glossary term from Glue Data Catalog}
\usage{
glue_delete_glossary_term(Identifier)
}
\arguments{
\item{Identifier}{[required] The unique identifier of the glossary term to delete.}
}
\description{
Deletes a glossary term from Glue Data Catalog.

See \url{https://www.paws-r-sdk.com/docs/glue_delete_glossary_term/} for full documentation.
}
\keyword{internal}
