% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_delete_integration}
\alias{glue_delete_integration}
\title{Deletes the specified Zero-ETL integration}
\usage{
glue_delete_integration(IntegrationIdentifier)
}
\arguments{
\item{IntegrationIdentifier}{[required] The Amazon Resource Name (ARN) for the integration.}
}
\description{
Deletes the specified Zero-ETL integration.

See \url{https://www.paws-r-sdk.com/docs/glue_delete_integration/} for full documentation.
}
\keyword{internal}
