% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_delete_integration_resource_property}
\alias{glue_delete_integration_resource_property}
\title{This API is used for deleting the ResourceProperty of the Glue
connection (for the source) or Glue database ARN (for the target)}
\usage{
glue_delete_integration_resource_property(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The connection ARN of the source, or the database ARN of the target.}
}
\description{
This API is used for deleting the \code{ResourceProperty} of the Glue connection (for the source) or Glue database ARN (for the target).

See \url{https://www.paws-r-sdk.com/docs/glue_delete_integration_resource_property/} for full documentation.
}
\keyword{internal}
