% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_delete_integration_table_properties}
\alias{glue_delete_integration_table_properties}
\title{Deletes the table properties that have been created for the tables that
need to be replicated}
\usage{
glue_delete_integration_table_properties(ResourceArn, TableName)
}
\arguments{
\item{ResourceArn}{[required] The connection ARN of the source, or the database ARN of the target.}

\item{TableName}{[required] The name of the table to be replicated.}
}
\description{
Deletes the table properties that have been created for the tables that need to be replicated.

See \url{https://www.paws-r-sdk.com/docs/glue_delete_integration_table_properties/} for full documentation.
}
\keyword{internal}
