% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_delete_job}
\alias{glue_delete_job}
\title{Deletes a specified job definition}
\usage{
glue_delete_job(JobName)
}
\arguments{
\item{JobName}{[required] The name of the job definition to delete.}
}
\description{
Deletes a specified job definition. If the job definition is not found, no exception is thrown.

See \url{https://paws-r.github.io/docs/glue/delete_job.html} for full documentation.
}
\keyword{internal}
