% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_delete_partition}
\alias{glue_delete_partition}
\title{Deletes a specified partition}
\usage{
glue_delete_partition(
  CatalogId = NULL,
  DatabaseName,
  TableName,
  PartitionValues
)
}
\arguments{
\item{CatalogId}{The ID of the Data Catalog where the partition to be deleted resides. If
none is provided, the Amazon Web Services account ID is used by default.}

\item{DatabaseName}{[required] The name of the catalog database in which the table in question resides.}

\item{TableName}{[required] The name of the table that contains the partition to be deleted.}

\item{PartitionValues}{[required] The values that define the partition.}
}
\description{
Deletes a specified partition.

See \url{https://paws-r.github.io/docs/glue/delete_partition.html} for full documentation.
}
\keyword{internal}
