% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_delete_partition_index}
\alias{glue_delete_partition_index}
\title{Deletes a specified partition index from an existing table}
\usage{
glue_delete_partition_index(
  CatalogId = NULL,
  DatabaseName,
  TableName,
  IndexName
)
}
\arguments{
\item{CatalogId}{The catalog ID where the table resides.}

\item{DatabaseName}{[required] Specifies the name of a database from which you want to delete a
partition index.}

\item{TableName}{[required] Specifies the name of a table from which you want to delete a partition
index.}

\item{IndexName}{[required] The name of the partition index to be deleted.}
}
\description{
Deletes a specified partition index from an existing table.

See \url{https://paws-r.github.io/docs/glue/delete_partition_index.html} for full documentation.
}
\keyword{internal}
