% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_delete_resource_policy}
\alias{glue_delete_resource_policy}
\title{Deletes a specified policy}
\usage{
glue_delete_resource_policy(PolicyHashCondition = NULL, ResourceArn = NULL)
}
\arguments{
\item{PolicyHashCondition}{The hash value returned when this policy was set.}

\item{ResourceArn}{The ARN of the Glue resource for the resource policy to be deleted.}
}
\description{
Deletes a specified policy.

See \url{https://paws-r.github.io/docs/glue/delete_resource_policy.html} for full documentation.
}
\keyword{internal}
