% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_delete_session}
\alias{glue_delete_session}
\title{Deletes the session}
\usage{
glue_delete_session(Id, RequestOrigin = NULL)
}
\arguments{
\item{Id}{[required] The ID of the session to be deleted.}

\item{RequestOrigin}{The name of the origin of the delete session request.}
}
\description{
Deletes the session.

See \url{https://paws-r.github.io/docs/glue/delete_session.html} for full documentation.
}
\keyword{internal}
