% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_delete_table}
\alias{glue_delete_table}
\title{Removes a table definition from the Data Catalog}
\usage{
glue_delete_table(CatalogId = NULL, DatabaseName, Name, TransactionId = NULL)
}
\arguments{
\item{CatalogId}{The ID of the Data Catalog where the table resides. If none is provided,
the Amazon Web Services account ID is used by default.}

\item{DatabaseName}{[required] The name of the catalog database in which the table resides. For Hive
compatibility, this name is entirely lowercase.}

\item{Name}{[required] The name of the table to be deleted. For Hive compatibility, this name
is entirely lowercase.}

\item{TransactionId}{The transaction ID at which to delete the table contents.}
}
\description{
Removes a table definition from the Data Catalog.

See \url{https://paws-r.github.io/docs/glue/delete_table.html} for full documentation.
}
\keyword{internal}
