% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_delete_table_optimizer}
\alias{glue_delete_table_optimizer}
\title{Deletes an optimizer and all associated metadata for a table}
\usage{
glue_delete_table_optimizer(CatalogId, DatabaseName, TableName, Type)
}
\arguments{
\item{CatalogId}{[required] The Catalog ID of the table.}

\item{DatabaseName}{[required] The name of the database in the catalog in which the table resides.}

\item{TableName}{[required] The name of the table.}

\item{Type}{[required] The type of table optimizer.}
}
\description{
Deletes an optimizer and all associated metadata for a table. The optimization will no longer be performed on the table.

See \url{https://www.paws-r-sdk.com/docs/glue_delete_table_optimizer/} for full documentation.
}
\keyword{internal}
