% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_delete_table_version}
\alias{glue_delete_table_version}
\title{Deletes a specified version of a table}
\usage{
glue_delete_table_version(CatalogId, DatabaseName, TableName, VersionId)
}
\arguments{
\item{CatalogId}{The ID of the Data Catalog where the tables reside. If none is provided,
the AWS account ID is used by default.}

\item{DatabaseName}{[required] The database in the catalog in which the table resides. For Hive
compatibility, this name is entirely lowercase.}

\item{TableName}{[required] The name of the table. For Hive compatibility, this name is entirely
lowercase.}

\item{VersionId}{[required] The ID of the table version to be deleted. A \code{VersionID} is a string
representation of an integer. Each version is incremented by 1.}
}
\value{
An empty list.
}
\description{
Deletes a specified version of a table.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_table_version(
  CatalogId = "string",
  DatabaseName = "string",
  TableName = "string",
  VersionId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
