% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_delete_usage_profile}
\alias{glue_delete_usage_profile}
\title{Deletes the Glue specified usage profile}
\usage{
glue_delete_usage_profile(Name)
}
\arguments{
\item{Name}{[required] The name of the usage profile to delete.}
}
\description{
Deletes the Glue specified usage profile.

See \url{https://www.paws-r-sdk.com/docs/glue_delete_usage_profile/} for full documentation.
}
\keyword{internal}
