% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_delete_user_defined_function}
\alias{glue_delete_user_defined_function}
\title{Deletes an existing function definition from the Data Catalog}
\usage{
glue_delete_user_defined_function(CatalogId = NULL, DatabaseName, FunctionName)
}
\arguments{
\item{CatalogId}{The ID of the Data Catalog where the function to be deleted is located.
If none is supplied, the Amazon Web Services account ID is used by
default.}

\item{DatabaseName}{[required] The name of the catalog database where the function is located.}

\item{FunctionName}{[required] The name of the function definition to be deleted.}
}
\description{
Deletes an existing function definition from the Data Catalog.

See \url{https://www.paws-r-sdk.com/docs/glue_delete_user_defined_function/} for full documentation.
}
\keyword{internal}
