% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_delete_workflow}
\alias{glue_delete_workflow}
\title{Deletes a workflow}
\usage{
glue_delete_workflow(Name)
}
\arguments{
\item{Name}{[required] Name of the workflow to be deleted.}
}
\description{
Deletes a workflow.

See \url{https://www.paws-r-sdk.com/docs/glue_delete_workflow/} for full documentation.
}
\keyword{internal}
