% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_describe_connection_type}
\alias{glue_describe_connection_type}
\title{The DescribeConnectionType API provides full details of the supported
options for a given connection type in Glue}
\usage{
glue_describe_connection_type(ConnectionType)
}
\arguments{
\item{ConnectionType}{[required] The name of the connection type to be described.}
}
\description{
The \code{\link[=glue_describe_connection_type]{describe_connection_type}} API provides full details of the supported options for a given connection type in Glue.

See \url{https://www.paws-r-sdk.com/docs/glue_describe_connection_type/} for full documentation.
}
\keyword{internal}
