% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_describe_entity}
\alias{glue_describe_entity}
\title{Provides details regarding the entity used with the connection type,
with a description of the data model for each field in the selected
entity}
\usage{
glue_describe_entity(
  ConnectionName,
  CatalogId = NULL,
  EntityName,
  NextToken = NULL,
  DataStoreApiVersion = NULL
)
}
\arguments{
\item{ConnectionName}{[required] The name of the connection that contains the connection type
credentials.}

\item{CatalogId}{The catalog ID of the catalog that contains the connection. This can be
null, By default, the Amazon Web Services Account ID is the catalog ID.}

\item{EntityName}{[required] The name of the entity that you want to describe from the connection
type.}

\item{NextToken}{A continuation token, included if this is a continuation call.}

\item{DataStoreApiVersion}{The version of the API used for the data store.}
}
\description{
Provides details regarding the entity used with the connection type, with a description of the data model for each field in the selected entity.

See \url{https://www.paws-r-sdk.com/docs/glue_describe_entity/} for full documentation.
}
\keyword{internal}
