% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_describe_inbound_integrations}
\alias{glue_describe_inbound_integrations}
\title{Returns a list of inbound integrations for the specified integration}
\usage{
glue_describe_inbound_integrations(
  IntegrationArn = NULL,
  Marker = NULL,
  MaxRecords = NULL,
  TargetArn = NULL
)
}
\arguments{
\item{IntegrationArn}{The Amazon Resource Name (ARN) of the integration.}

\item{Marker}{A token to specify where to start paginating. This is the marker from a
previously truncated response.}

\item{MaxRecords}{The total number of items to return in the output.}

\item{TargetArn}{The Amazon Resource Name (ARN) of the target resource in the
integration.}
}
\description{
Returns a list of inbound integrations for the specified integration.

See \url{https://www.paws-r-sdk.com/docs/glue_describe_inbound_integrations/} for full documentation.
}
\keyword{internal}
