% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_describe_integrations}
\alias{glue_describe_integrations}
\title{The API is used to retrieve a list of integrations}
\usage{
glue_describe_integrations(
  IntegrationIdentifier = NULL,
  Marker = NULL,
  MaxRecords = NULL,
  Filters = NULL
)
}
\arguments{
\item{IntegrationIdentifier}{The Amazon Resource Name (ARN) for the integration.}

\item{Marker}{A value that indicates the starting point for the next set of response
records in a subsequent request.}

\item{MaxRecords}{The total number of items to return in the output.}

\item{Filters}{A list of key and values, to filter down the results. Supported keys are
"Status", "IntegrationName", and "SourceArn". IntegrationName is limited
to only one value.}
}
\description{
The API is used to retrieve a list of integrations.

See \url{https://www.paws-r-sdk.com/docs/glue_describe_integrations/} for full documentation.
}
\keyword{internal}
