% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_asset}
\alias{glue_get_asset}
\title{Retrieves the metadata for an asset in Glue Data Catalog, including its
forms, additional attachments, and associated glossary terms}
\usage{
glue_get_asset(Identifier)
}
\arguments{
\item{Identifier}{[required] The unique identifier of the asset to retrieve.}
}
\description{
Retrieves the metadata for an asset in Glue Data Catalog, including its forms, additional attachments, and associated glossary terms.

See \url{https://www.paws-r-sdk.com/docs/glue_get_asset/} for full documentation.
}
\keyword{internal}
