% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_asset_type}
\alias{glue_get_asset_type}
\title{Retrieves an asset type in Glue Data Catalog by its identifier}
\usage{
glue_get_asset_type(Identifier)
}
\arguments{
\item{Identifier}{[required] The identifier of the asset type to retrieve.}
}
\description{
Retrieves an asset type in Glue Data Catalog by its identifier.

See \url{https://www.paws-r-sdk.com/docs/glue_get_asset_type/} for full documentation.
}
\keyword{internal}
