% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_blueprint}
\alias{glue_get_blueprint}
\title{Retrieves the details of a blueprint}
\usage{
glue_get_blueprint(Name, IncludeBlueprint = NULL, IncludeParameterSpec = NULL)
}
\arguments{
\item{Name}{[required] The name of the blueprint.}

\item{IncludeBlueprint}{Specifies whether or not to include the blueprint in the response.}

\item{IncludeParameterSpec}{Specifies whether or not to include the parameter specification.}
}
\description{
Retrieves the details of a blueprint.

See \url{https://paws-r.github.io/docs/glue/get_blueprint.html} for full documentation.
}
\keyword{internal}
