% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_blueprint_run}
\alias{glue_get_blueprint_run}
\title{Retrieves the details of a blueprint run}
\usage{
glue_get_blueprint_run(BlueprintName, RunId)
}
\arguments{
\item{BlueprintName}{[required] The name of the blueprint.}

\item{RunId}{[required] The run ID for the blueprint run you want to retrieve.}
}
\description{
Retrieves the details of a blueprint run.

See \url{https://www.paws-r-sdk.com/docs/glue_get_blueprint_run/} for full documentation.
}
\keyword{internal}
