% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_blueprint_runs}
\alias{glue_get_blueprint_runs}
\title{Retrieves the details of blueprint runs for a specified blueprint}
\usage{
glue_get_blueprint_runs(BlueprintName, NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{BlueprintName}{[required] The name of the blueprint.}

\item{NextToken}{A continuation token, if this is a continuation request.}

\item{MaxResults}{The maximum size of a list to return.}
}
\description{
Retrieves the details of blueprint runs for a specified blueprint.

See \url{https://www.paws-r-sdk.com/docs/glue_get_blueprint_runs/} for full documentation.
}
\keyword{internal}
