% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_catalog}
\alias{glue_get_catalog}
\title{The name of the Catalog to retrieve}
\usage{
glue_get_catalog(CatalogId)
}
\arguments{
\item{CatalogId}{[required] The ID of the parent catalog in which the catalog resides. If none is
provided, the Amazon Web Services Account Number is used by default.}
}
\description{
The name of the Catalog to retrieve. This should be all lowercase.

See \url{https://www.paws-r-sdk.com/docs/glue_get_catalog/} for full documentation.
}
\keyword{internal}
