% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_catalog_import_status}
\alias{glue_get_catalog_import_status}
\title{Retrieves the status of a migration operation}
\usage{
glue_get_catalog_import_status(CatalogId)
}
\arguments{
\item{CatalogId}{The ID of the catalog to migrate. Currently, this should be the AWS
account ID.}
}
\value{
A list with the following syntax:\preformatted{list(
  ImportStatus = list(
    ImportCompleted = TRUE|FALSE,
    ImportTime = as.POSIXct(
      "2015-01-01"
    ),
    ImportedBy = "string"
  )
)
}
}
\description{
Retrieves the status of a migration operation.
}
\section{Request syntax}{
\preformatted{svc$get_catalog_import_status(
  CatalogId = "string"
)
}
}

\keyword{internal}
