% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_column_statistics_task_runs}
\alias{glue_get_column_statistics_task_runs}
\title{Retrieves information about all runs associated with the specified table}
\usage{
glue_get_column_statistics_task_runs(
  DatabaseName,
  TableName,
  MaxResults = NULL,
  NextToken = NULL,
  CatalogID = NULL
)
}
\arguments{
\item{DatabaseName}{[required] The name of the database where the table resides.}

\item{TableName}{[required] The name of the table.}

\item{MaxResults}{The maximum size of the response.}

\item{NextToken}{A continuation token, if this is a continuation call.}

\item{CatalogID}{The ID of the Data Catalog where the table resides. If none is supplied, the Amazon Web Services account ID is used by default.}
}
\description{
Retrieves information about all runs associated with the specified table.

See \url{https://www.paws-r-sdk.com/docs/glue_get_column_statistics_task_runs/} for full documentation.
}
\keyword{internal}
