% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_column_statistics_task_settings}
\alias{glue_get_column_statistics_task_settings}
\title{Gets settings for a column statistics task}
\usage{
glue_get_column_statistics_task_settings(
  DatabaseName,
  TableName,
  CatalogID = NULL
)
}
\arguments{
\item{DatabaseName}{[required] The name of the database where the table resides.}

\item{TableName}{[required] The name of the table for which to retrieve column statistics.}

\item{CatalogID}{The ID of the Data Catalog where the table resides. If none is supplied, the Amazon Web Services account ID is used by default.}
}
\description{
Gets settings for a column statistics task.

See \url{https://www.paws-r-sdk.com/docs/glue_get_column_statistics_task_settings/} for full documentation.
}
\keyword{internal}
