% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_crawlers}
\alias{glue_get_crawlers}
\title{Retrieves metadata for all crawlers defined in the customer account}
\usage{
glue_get_crawlers(MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{The number of crawlers to return on each call.}

\item{NextToken}{A continuation token, if this is a continuation request.}
}
\description{
Retrieves metadata for all crawlers defined in the customer account.

See \url{https://paws-r.github.io/docs/glue/get_crawlers.html} for full documentation.
}
\keyword{internal}
