% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_custom_entity_type}
\alias{glue_get_custom_entity_type}
\title{Retrieves the details of a custom pattern by specifying its name}
\usage{
glue_get_custom_entity_type(Name)
}
\arguments{
\item{Name}{[required] The name of the custom pattern that you want to retrieve.}
}
\description{
Retrieves the details of a custom pattern by specifying its name.

See \url{https://paws-r.github.io/docs/glue/get_custom_entity_type.html} for full documentation.
}
\keyword{internal}
