% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_dashboard_url}
\alias{glue_get_dashboard_url}
\title{Retrieves the URL for the Spark monitoring dashboard for a Glue resource}
\usage{
glue_get_dashboard_url(ResourceId, ResourceType, RequestOrigin = NULL)
}
\arguments{
\item{ResourceId}{[required] The unique identifier of the resource for which to retrieve the dashboard URL.}

\item{ResourceType}{[required] The type of the resource. Valid values are \code{SESSION} and \code{JOB}.}

\item{RequestOrigin}{The origin of the request.}
}
\description{
Retrieves the URL for the Spark monitoring dashboard for a Glue resource.

See \url{https://www.paws-r-sdk.com/docs/glue_get_dashboard_url/} for full documentation.
}
\keyword{internal}
