% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_data_catalog_encryption_settings}
\alias{glue_get_data_catalog_encryption_settings}
\title{Retrieves the security configuration for a specified catalog}
\usage{
glue_get_data_catalog_encryption_settings(CatalogId)
}
\arguments{
\item{CatalogId}{The ID of the Data Catalog to retrieve the security configuration for.
If none is provided, the AWS account ID is used by default.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  DataCatalogEncryptionSettings = list(
    EncryptionAtRest = list(
      CatalogEncryptionMode = "DISABLED"|"SSE-KMS",
      SseAwsKmsKeyId = "string"
    ),
    ConnectionPasswordEncryption = list(
      ReturnConnectionPasswordEncrypted = TRUE|FALSE,
      AwsKmsKeyId = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves the security configuration for a specified catalog.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_data_catalog_encryption_settings(
  CatalogId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
