% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_data_catalog_export_configuration}
\alias{glue_get_data_catalog_export_configuration}
\title{Retrieves the current export configuration for the Glue Data Catalog}
\usage{
glue_get_data_catalog_export_configuration()
}
\description{
Retrieves the current export configuration for the Glue Data Catalog. The export configuration controls whether catalog metadata is exported to S3 Tables.

See \url{https://www.paws-r-sdk.com/docs/glue_get_data_catalog_export_configuration/} for full documentation.
}
\keyword{internal}
