% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_data_quality_model}
\alias{glue_get_data_quality_model}
\title{Retrieve the training status of the model along with more information
(CompletedOn, StartedOn, FailureReason)}
\usage{
glue_get_data_quality_model(StatisticId = NULL, ProfileId)
}
\arguments{
\item{StatisticId}{The Statistic ID.}

\item{ProfileId}{[required] The Profile ID.}
}
\description{
Retrieve the training status of the model along with more information (CompletedOn, StartedOn, FailureReason).

See \url{https://www.paws-r-sdk.com/docs/glue_get_data_quality_model/} for full documentation.
}
\keyword{internal}
