% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_data_quality_model_result}
\alias{glue_get_data_quality_model_result}
\title{Retrieve a statistic's predictions for a given Profile ID}
\usage{
glue_get_data_quality_model_result(StatisticId, ProfileId)
}
\arguments{
\item{StatisticId}{[required] The Statistic ID.}

\item{ProfileId}{[required] The Profile ID.}
}
\description{
Retrieve a statistic's predictions for a given Profile ID.

See \url{https://www.paws-r-sdk.com/docs/glue_get_data_quality_model_result/} for full documentation.
}
\keyword{internal}
