% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_data_quality_result}
\alias{glue_get_data_quality_result}
\title{Retrieves the result of a data quality rule evaluation}
\usage{
glue_get_data_quality_result(ResultId)
}
\arguments{
\item{ResultId}{[required] A unique result ID for the data quality result.}
}
\description{
Retrieves the result of a data quality rule evaluation.

See \url{https://www.paws-r-sdk.com/docs/glue_get_data_quality_result/} for full documentation.
}
\keyword{internal}
