% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_data_quality_rule_recommendation_run}
\alias{glue_get_data_quality_rule_recommendation_run}
\title{Gets the specified recommendation run that was used to generate rules}
\usage{
glue_get_data_quality_rule_recommendation_run(RunId)
}
\arguments{
\item{RunId}{[required] The unique run identifier associated with this run.}
}
\description{
Gets the specified recommendation run that was used to generate rules.

See \url{https://www.paws-r-sdk.com/docs/glue_get_data_quality_rule_recommendation_run/} for full documentation.
}
\keyword{internal}
