% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_data_quality_ruleset}
\alias{glue_get_data_quality_ruleset}
\title{Returns an existing ruleset by identifier or name}
\usage{
glue_get_data_quality_ruleset(Name)
}
\arguments{
\item{Name}{[required] The name of the ruleset.}
}
\description{
Returns an existing ruleset by identifier or name.

See \url{https://paws-r.github.io/docs/glue/get_data_quality_ruleset.html} for full documentation.
}
\keyword{internal}
