% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_data_quality_ruleset_evaluation_run}
\alias{glue_get_data_quality_ruleset_evaluation_run}
\title{Retrieves a specific run where a ruleset is evaluated against a data
source}
\usage{
glue_get_data_quality_ruleset_evaluation_run(RunId)
}
\arguments{
\item{RunId}{[required] The unique run identifier associated with this run.}
}
\description{
Retrieves a specific run where a ruleset is evaluated against a data source.

See \url{https://www.paws-r-sdk.com/docs/glue_get_data_quality_ruleset_evaluation_run/} for full documentation.
}
\keyword{internal}
