% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_dataflow_graph}
\alias{glue_get_dataflow_graph}
\title{Transforms a Python script into a directed acyclic graph (DAG)}
\usage{
glue_get_dataflow_graph(PythonScript = NULL)
}
\arguments{
\item{PythonScript}{The Python script to transform.}
}
\description{
Transforms a Python script into a directed acyclic graph (DAG).

See \url{https://www.paws-r-sdk.com/docs/glue_get_dataflow_graph/} for full documentation.
}
\keyword{internal}
