% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_dev_endpoints}
\alias{glue_get_dev_endpoints}
\title{Retrieves all the development endpoints in this Amazon Web Services
account}
\usage{
glue_get_dev_endpoints(MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{The maximum size of information to return.}

\item{NextToken}{A continuation token, if this is a continuation call.}
}
\description{
Retrieves all the development endpoints in this Amazon Web Services account.

See \url{https://paws-r.github.io/docs/glue/get_dev_endpoints.html} for full documentation.
}
\keyword{internal}
