% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_form_type}
\alias{glue_get_form_type}
\title{Retrieves a form type in Glue Data Catalog by its identifier}
\usage{
glue_get_form_type(Identifier)
}
\arguments{
\item{Identifier}{[required] The identifier of the form type to retrieve.}
}
\description{
Retrieves a form type in Glue Data Catalog by its identifier.

See \url{https://www.paws-r-sdk.com/docs/glue_get_form_type/} for full documentation.
}
\keyword{internal}
