% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_glossary}
\alias{glue_get_glossary}
\title{Retrieves a business glossary in Glue Data Catalog by its identifier}
\usage{
glue_get_glossary(Identifier)
}
\arguments{
\item{Identifier}{[required] The unique identifier of the glossary to retrieve.}
}
\description{
Retrieves a business glossary in Glue Data Catalog by its identifier.

See \url{https://www.paws-r-sdk.com/docs/glue_get_glossary/} for full documentation.
}
\keyword{internal}
