% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_glossary_term}
\alias{glue_get_glossary_term}
\title{Retrieves a glossary term in Glue Data Catalog by its identifier}
\usage{
glue_get_glossary_term(Identifier)
}
\arguments{
\item{Identifier}{[required] The unique identifier of the glossary term to retrieve.}
}
\description{
Retrieves a glossary term in Glue Data Catalog by its identifier.

See \url{https://www.paws-r-sdk.com/docs/glue_get_glossary_term/} for full documentation.
}
\keyword{internal}
