% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_glue_identity_center_configuration}
\alias{glue_get_glue_identity_center_configuration}
\title{Retrieves the current Glue Identity Center configuration details,
including the associated Identity Center instance and application
information}
\usage{
glue_get_glue_identity_center_configuration()
}
\description{
Retrieves the current Glue Identity Center configuration details, including the associated Identity Center instance and application information.

See \url{https://www.paws-r-sdk.com/docs/glue_get_glue_identity_center_configuration/} for full documentation.
}
\keyword{internal}
