% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_job_bookmark}
\alias{glue_get_job_bookmark}
\title{Returns information on a job bookmark entry}
\usage{
glue_get_job_bookmark(JobName, RunId)
}
\arguments{
\item{JobName}{[required] The name of the job in question.}

\item{RunId}{The unique run identifier associated with this job run.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  JobBookmarkEntry = list(
    JobName = "string",
    Version = 123,
    Run = 123,
    Attempt = 123,
    PreviousRunId = "string",
    RunId = "string",
    JobBookmark = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Returns information on a job bookmark entry.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_job_bookmark(
  JobName = "string",
  RunId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
