% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_job_runs}
\alias{glue_get_job_runs}
\title{Retrieves metadata for all runs of a given job definition}
\usage{
glue_get_job_runs(JobName, NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{JobName}{[required] The name of the job definition for which to retrieve all job runs.}

\item{NextToken}{A continuation token, if this is a continuation call.}

\item{MaxResults}{The maximum size of the response.}
}
\description{
Retrieves metadata for all runs of a given job definition.

See \url{https://paws-r.github.io/docs/glue/get_job_runs.html} for full documentation.
}
\keyword{internal}
