% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_materialized_view_refresh_task_run}
\alias{glue_get_materialized_view_refresh_task_run}
\title{Get the associated metadata/information for a task run, given a task run
ID}
\usage{
glue_get_materialized_view_refresh_task_run(
  CatalogId,
  MaterializedViewRefreshTaskRunId
)
}
\arguments{
\item{CatalogId}{[required] The ID of the Data Catalog where the table resides. If none is supplied,
the account ID is used by default.}

\item{MaterializedViewRefreshTaskRunId}{[required] The identifier for the particular materialized view refresh task run.}
}
\description{
Get the associated metadata/information for a task run, given a task run ID.

See \url{https://www.paws-r-sdk.com/docs/glue_get_materialized_view_refresh_task_run/} for full documentation.
}
\keyword{internal}
