% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_ml_task_run}
\alias{glue_get_ml_task_run}
\title{Gets details for a specific task run on a machine learning transform}
\usage{
glue_get_ml_task_run(TransformId, TaskRunId)
}
\arguments{
\item{TransformId}{[required] The unique identifier of the machine learning transform.}

\item{TaskRunId}{[required] The unique identifier of the task run.}
}
\description{
Gets details for a specific task run on a machine learning transform.
Machine learning task runs are asynchronous tasks that AWS Glue runs on
your behalf as part of various machine learning workflows. You can check
the stats of any task run by calling
\code{\link[=glue_get_ml_task_run]{get_ml_task_run}} with the \code{TaskRunID} and its
parent transform's \code{TransformID}.
}
\section{Request syntax}{
\preformatted{svc$get_ml_task_run(
  TransformId = "string",
  TaskRunId = "string"
)
}
}

\keyword{internal}
