% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_ml_transforms}
\alias{glue_get_ml_transforms}
\title{Gets a sortable, filterable list of existing Glue machine learning
transforms}
\usage{
glue_get_ml_transforms(
  NextToken = NULL,
  MaxResults = NULL,
  Filter = NULL,
  Sort = NULL
)
}
\arguments{
\item{NextToken}{A paginated token to offset the results.}

\item{MaxResults}{The maximum number of results to return.}

\item{Filter}{The filter transformation criteria.}

\item{Sort}{The sorting criteria.}
}
\description{
Gets a sortable, filterable list of existing Glue machine learning transforms. Machine learning transforms are a special type of transform that use machine learning to learn the details of the transformation to be performed by learning from examples provided by humans. These transformations are then saved by Glue, and you can retrieve their metadata by calling \code{\link[=glue_get_ml_transforms]{get_ml_transforms}}.

See \url{https://www.paws-r-sdk.com/docs/glue_get_ml_transforms/} for full documentation.
}
\keyword{internal}
