% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_partition_indexes}
\alias{glue_get_partition_indexes}
\title{Retrieves the partition indexes associated with a table}
\usage{
glue_get_partition_indexes(
  CatalogId = NULL,
  DatabaseName,
  TableName,
  NextToken = NULL
)
}
\arguments{
\item{CatalogId}{The catalog ID where the table resides.}

\item{DatabaseName}{[required] Specifies the name of a database from which you want to retrieve
partition indexes.}

\item{TableName}{[required] Specifies the name of a table for which you want to retrieve the
partition indexes.}

\item{NextToken}{A continuation token, included if this is a continuation call.}
}
\description{
Retrieves the partition indexes associated with a table.

See \url{https://www.paws-r-sdk.com/docs/glue_get_partition_indexes/} for full documentation.
}
\keyword{internal}
