% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_registry}
\alias{glue_get_registry}
\title{Describes the specified registry in detail}
\usage{
glue_get_registry(RegistryId)
}
\arguments{
\item{RegistryId}{[required] This is a wrapper structure that may contain the registry name and
Amazon Resource Name (ARN).}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  RegistryName = "string",
  RegistryArn = "string",
  Description = "string",
  Status = "AVAILABLE"|"DELETING",
  CreatedTime = "string",
  UpdatedTime = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Describes the specified registry in detail.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_registry(
  RegistryId = list(
    RegistryName = "string",
    RegistryArn = "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
