% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_schema}
\alias{glue_get_schema}
\title{Describes the specified schema in detail}
\usage{
glue_get_schema(SchemaId)
}
\arguments{
\item{SchemaId}{[required] This is a wrapper structure to contain schema identity fields. The
structure contains:
\itemize{
\item SchemaId$SchemaArn: The Amazon Resource Name (ARN) of the schema.
Either \code{SchemaArn} or \code{SchemaName} and \code{RegistryName} has to be
provided.
\item SchemaId$SchemaName: The name of the schema. Either \code{SchemaArn} or
\code{SchemaName} and \code{RegistryName} has to be provided.
}}
}
\value{
A list with the following syntax:\preformatted{list(
  RegistryName = "string",
  RegistryArn = "string",
  SchemaName = "string",
  SchemaArn = "string",
  Description = "string",
  DataFormat = "AVRO",
  Compatibility = "NONE"|"DISABLED"|"BACKWARD"|"BACKWARD_ALL"|"FORWARD"|"FORWARD_ALL"|"FULL"|"FULL_ALL",
  SchemaCheckpoint = 123,
  LatestSchemaVersion = 123,
  NextSchemaVersion = 123,
  SchemaStatus = "AVAILABLE"|"PENDING"|"DELETING",
  CreatedTime = "string",
  UpdatedTime = "string"
)
}
}
\description{
Describes the specified schema in detail.
}
\section{Request syntax}{
\preformatted{svc$get_schema(
  SchemaId = list(
    SchemaArn = "string",
    SchemaName = "string",
    RegistryName = "string"
  )
)
}
}

\keyword{internal}
