% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_session}
\alias{glue_get_session}
\title{Retrieves the session}
\usage{
glue_get_session(Id, RequestOrigin = NULL)
}
\arguments{
\item{Id}{[required] The ID of the session.}

\item{RequestOrigin}{The origin of the request.}
}
\description{
Retrieves the session.

See \url{https://paws-r.github.io/docs/glue/get_session.html} for full documentation.
}
\keyword{internal}
