% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_session_endpoint}
\alias{glue_get_session_endpoint}
\title{Returns the Spark Connect endpoint URL and authentication token for an
interactive session}
\usage{
glue_get_session_endpoint(SessionId)
}
\arguments{
\item{SessionId}{[required] The unique identifier of the interactive session.}
}
\description{
Returns the Spark Connect endpoint URL and authentication token for an interactive session.

See \url{https://www.paws-r-sdk.com/docs/glue_get_session_endpoint/} for full documentation.
}
\keyword{internal}
