% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_statement}
\alias{glue_get_statement}
\title{Retrieves the statement}
\usage{
glue_get_statement(SessionId, Id, RequestOrigin = NULL)
}
\arguments{
\item{SessionId}{[required] The Session ID of the statement.}

\item{Id}{[required] The Id of the statement.}

\item{RequestOrigin}{The origin of the request.}
}
\description{
Retrieves the statement.

See \url{https://www.paws-r-sdk.com/docs/glue_get_statement/} for full documentation.
}
\keyword{internal}
