% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_table_optimizer}
\alias{glue_get_table_optimizer}
\title{Returns the configuration of all optimizers associated with a specified
table}
\usage{
glue_get_table_optimizer(CatalogId, DatabaseName, TableName, Type)
}
\arguments{
\item{CatalogId}{[required] The Catalog ID of the table.}

\item{DatabaseName}{[required] The name of the database in the catalog in which the table resides.}

\item{TableName}{[required] The name of the table.}

\item{Type}{[required] The type of table optimizer.}
}
\description{
Returns the configuration of all optimizers associated with a specified table.

See \url{https://www.paws-r-sdk.com/docs/glue_get_table_optimizer/} for full documentation.
}
\keyword{internal}
