% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_tags}
\alias{glue_get_tags}
\title{Retrieves a list of tags associated with a resource}
\usage{
glue_get_tags(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource for which to retrieve
tags.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Tags = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves a list of tags associated with a resource.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_tags(
  ResourceArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
