% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_unfiltered_partition_metadata}
\alias{glue_get_unfiltered_partition_metadata}
\title{Get unfiltered partition metadata}
\usage{
glue_get_unfiltered_partition_metadata(
  CatalogId,
  DatabaseName,
  TableName,
  PartitionValues,
  AuditContext = NULL,
  SupportedPermissionTypes
)
}
\arguments{
\item{CatalogId}{[required]}

\item{DatabaseName}{[required]}

\item{TableName}{[required]}

\item{PartitionValues}{[required]}

\item{AuditContext}{}

\item{SupportedPermissionTypes}{[required]}
}
\description{
Get unfiltered partition metadata

See \url{https://paws-r.github.io/docs/glue/get_unfiltered_partition_metadata.html} for full documentation.
}
\keyword{internal}
