% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_unfiltered_partitions_metadata}
\alias{glue_get_unfiltered_partitions_metadata}
\title{Get unfiltered partitions metadata}
\usage{
glue_get_unfiltered_partitions_metadata(
  CatalogId,
  DatabaseName,
  TableName,
  Expression = NULL,
  AuditContext = NULL,
  SupportedPermissionTypes,
  NextToken = NULL,
  Segment = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{CatalogId}{[required]}

\item{DatabaseName}{[required]}

\item{TableName}{[required]}

\item{Expression}{}

\item{AuditContext}{}

\item{SupportedPermissionTypes}{[required]}

\item{NextToken}{}

\item{Segment}{}

\item{MaxResults}{}
}
\description{
Get unfiltered partitions metadata

See \url{https://paws-r.github.io/docs/glue/get_unfiltered_partitions_metadata.html} for full documentation.
}
\keyword{internal}
