% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_usage_profile}
\alias{glue_get_usage_profile}
\title{Retrieves information about the specified Glue usage profile}
\usage{
glue_get_usage_profile(Name)
}
\arguments{
\item{Name}{[required] The name of the usage profile to retrieve.}
}
\description{
Retrieves information about the specified Glue usage profile.

See \url{https://www.paws-r-sdk.com/docs/glue_get_usage_profile/} for full documentation.
}
\keyword{internal}
