% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_user_defined_function}
\alias{glue_get_user_defined_function}
\title{Retrieves a specified function definition from the Data Catalog}
\usage{
glue_get_user_defined_function(CatalogId, DatabaseName, FunctionName)
}
\arguments{
\item{CatalogId}{The ID of the Data Catalog where the function to be retrieved is
located. If none is provided, the AWS account ID is used by default.}

\item{DatabaseName}{[required] The name of the catalog database where the function is located.}

\item{FunctionName}{[required] The name of the function.}
}
\value{
A list with the following syntax:\preformatted{list(
  UserDefinedFunction = list(
    FunctionName = "string",
    DatabaseName = "string",
    ClassName = "string",
    OwnerName = "string",
    OwnerType = "USER"|"ROLE"|"GROUP",
    CreateTime = as.POSIXct(
      "2015-01-01"
    ),
    ResourceUris = list(
      list(
        ResourceType = "JAR"|"FILE"|"ARCHIVE",
        Uri = "string"
      )
    ),
    CatalogId = "string"
  )
)
}
}
\description{
Retrieves a specified function definition from the Data Catalog.
}
\section{Request syntax}{
\preformatted{svc$get_user_defined_function(
  CatalogId = "string",
  DatabaseName = "string",
  FunctionName = "string"
)
}
}

\keyword{internal}
