% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_user_defined_functions}
\alias{glue_get_user_defined_functions}
\title{Retrieves multiple function definitions from the Data Catalog}
\usage{
glue_get_user_defined_functions(CatalogId, DatabaseName, Pattern,
  NextToken, MaxResults)
}
\arguments{
\item{CatalogId}{The ID of the Data Catalog where the functions to be retrieved are
located. If none is provided, the AWS account ID is used by default.}

\item{DatabaseName}{The name of the catalog database where the functions are located. If
none is provided, functions from all the databases across the catalog
will be returned.}

\item{Pattern}{[required] An optional function-name pattern string that filters the function
definitions returned.}

\item{NextToken}{A continuation token, if this is a continuation call.}

\item{MaxResults}{The maximum number of functions to return in one response.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  UserDefinedFunctions = list(
    list(
      FunctionName = "string",
      DatabaseName = "string",
      ClassName = "string",
      OwnerName = "string",
      OwnerType = "USER"|"ROLE"|"GROUP",
      CreateTime = as.POSIXct(
        "2015-01-01"
      ),
      ResourceUris = list(
        list(
          ResourceType = "JAR"|"FILE"|"ARCHIVE",
          Uri = "string"
        )
      ),
      CatalogId = "string"
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves multiple function definitions from the Data Catalog.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_user_defined_functions(
  CatalogId = "string",
  DatabaseName = "string",
  Pattern = "string",
  NextToken = "string",
  MaxResults = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
