% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_workflow}
\alias{glue_get_workflow}
\title{Retrieves resource metadata for a workflow}
\usage{
glue_get_workflow(Name, IncludeGraph = NULL)
}
\arguments{
\item{Name}{[required] The name of the workflow to retrieve.}

\item{IncludeGraph}{Specifies whether to include a graph when returning the workflow
resource metadata.}
}
\description{
Retrieves resource metadata for a workflow.

See \url{https://paws-r.github.io/docs/glue/get_workflow.html} for full documentation.
}
\keyword{internal}
