% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_workflow_run}
\alias{glue_get_workflow_run}
\title{Retrieves the metadata for a given workflow run}
\usage{
glue_get_workflow_run(Name, RunId, IncludeGraph = NULL)
}
\arguments{
\item{Name}{[required] Name of the workflow being run.}

\item{RunId}{[required] The ID of the workflow run.}

\item{IncludeGraph}{Specifies whether to include the workflow graph in response or not.}
}
\description{
Retrieves the metadata for a given workflow run.

See \url{https://paws-r.github.io/docs/glue/get_workflow_run.html} for full documentation.
}
\keyword{internal}
