% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_import_catalog_to_glue}
\alias{glue_import_catalog_to_glue}
\title{Imports an existing Amazon Athena Data Catalog to AWS Glue}
\usage{
glue_import_catalog_to_glue(CatalogId)
}
\arguments{
\item{CatalogId}{The ID of the catalog to import. Currently, this should be the AWS
account ID.}
}
\value{
An empty list.
}
\description{
Imports an existing Amazon Athena Data Catalog to AWS Glue
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$import_catalog_to_glue(
  CatalogId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
