% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_list_column_statistics_task_runs}
\alias{glue_list_column_statistics_task_runs}
\title{List all task runs for a particular account}
\usage{
glue_list_column_statistics_task_runs(MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{The maximum size of the response.}

\item{NextToken}{A continuation token, if this is a continuation call.}
}
\description{
List all task runs for a particular account.

See \url{https://www.paws-r-sdk.com/docs/glue_list_column_statistics_task_runs/} for full documentation.
}
\keyword{internal}
