% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_list_connection_types}
\alias{glue_list_connection_types}
\title{The ListConnectionTypes API provides a discovery mechanism to learn
available connection types in Glue}
\usage{
glue_list_connection_types(MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{The maximum number of results to return.}

\item{NextToken}{A continuation token, if this is a continuation call.}
}
\description{
The \code{\link[=glue_list_connection_types]{list_connection_types}} API provides a discovery mechanism to learn available connection types in Glue. The response contains a list of connection types with high-level details of what is supported for each connection type. The connection types listed are the set of supported options for the \code{ConnectionType} value in the \code{\link[=glue_create_connection]{create_connection}} API.

See \url{https://www.paws-r-sdk.com/docs/glue_list_connection_types/} for full documentation.
}
\keyword{internal}
