% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_list_custom_entity_types}
\alias{glue_list_custom_entity_types}
\title{Lists all the custom patterns that have been created}
\usage{
glue_list_custom_entity_types(NextToken = NULL, MaxResults = NULL, Tags = NULL)
}
\arguments{
\item{NextToken}{A paginated token to offset the results.}

\item{MaxResults}{The maximum number of results to return.}

\item{Tags}{A list of key-value pair tags.}
}
\description{
Lists all the custom patterns that have been created.

See \url{https://paws-r.github.io/docs/glue/list_custom_entity_types.html} for full documentation.
}
\keyword{internal}
