% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_list_data_quality_results}
\alias{glue_list_data_quality_results}
\title{Returns all data quality execution results for your account}
\usage{
glue_list_data_quality_results(
  Filter = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{Filter}{The filter criteria.}

\item{NextToken}{A paginated token to offset the results.}

\item{MaxResults}{The maximum number of results to return.}
}
\description{
Returns all data quality execution results for your account.

See \url{https://paws-r.github.io/docs/glue/list_data_quality_results.html} for full documentation.
}
\keyword{internal}
