% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_list_data_quality_ruleset_evaluation_runs}
\alias{glue_list_data_quality_ruleset_evaluation_runs}
\title{Lists all the runs meeting the filter criteria, where a ruleset is
evaluated against a data source}
\usage{
glue_list_data_quality_ruleset_evaluation_runs(
  Filter = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{Filter}{The filter criteria.}

\item{NextToken}{A paginated token to offset the results.}

\item{MaxResults}{The maximum number of results to return.}
}
\description{
Lists all the runs meeting the filter criteria, where a ruleset is evaluated against a data source.

See \url{https://paws-r.github.io/docs/glue/list_data_quality_ruleset_evaluation_runs.html} for full documentation.
}
\keyword{internal}
