% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_list_data_quality_rulesets}
\alias{glue_list_data_quality_rulesets}
\title{Returns a paginated list of rulesets for the specified list of Glue
tables}
\usage{
glue_list_data_quality_rulesets(
  NextToken = NULL,
  MaxResults = NULL,
  Filter = NULL,
  Tags = NULL
)
}
\arguments{
\item{NextToken}{A paginated token to offset the results.}

\item{MaxResults}{The maximum number of results to return.}

\item{Filter}{The filter criteria.}

\item{Tags}{A list of key-value pair tags.}
}
\description{
Returns a paginated list of rulesets for the specified list of Glue tables.

See \url{https://www.paws-r-sdk.com/docs/glue_list_data_quality_rulesets/} for full documentation.
}
\keyword{internal}
