% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_list_data_quality_statistic_annotations}
\alias{glue_list_data_quality_statistic_annotations}
\title{Retrieve annotations for a data quality statistic}
\usage{
glue_list_data_quality_statistic_annotations(
  StatisticId = NULL,
  ProfileId = NULL,
  TimestampFilter = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{StatisticId}{The Statistic ID.}

\item{ProfileId}{The Profile ID.}

\item{TimestampFilter}{A timestamp filter.}

\item{MaxResults}{The maximum number of results to return in this request.}

\item{NextToken}{A pagination token to retrieve the next set of results.}
}
\description{
Retrieve annotations for a data quality statistic.

See \url{https://www.paws-r-sdk.com/docs/glue_list_data_quality_statistic_annotations/} for full documentation.
}
\keyword{internal}
