% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_list_data_quality_statistics}
\alias{glue_list_data_quality_statistics}
\title{Retrieves a list of data quality statistics}
\usage{
glue_list_data_quality_statistics(
  StatisticId = NULL,
  ProfileId = NULL,
  TimestampFilter = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{StatisticId}{The Statistic ID.}

\item{ProfileId}{The Profile ID.}

\item{TimestampFilter}{A timestamp filter.}

\item{MaxResults}{The maximum number of results to return in this request.}

\item{NextToken}{A pagination token to request the next page of results.}
}
\description{
Retrieves a list of data quality statistics.

See \url{https://www.paws-r-sdk.com/docs/glue_list_data_quality_statistics/} for full documentation.
}
\keyword{internal}
