% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_list_dev_endpoints}
\alias{glue_list_dev_endpoints}
\title{Retrieves the names of all DevEndpoint resources in this AWS account, or
the resources with the specified tag}
\usage{
glue_list_dev_endpoints(NextToken, MaxResults, Tags)
}
\arguments{
\item{NextToken}{A continuation token, if this is a continuation request.}

\item{MaxResults}{The maximum size of a list to return.}

\item{Tags}{Specifies to return only these tagged resources.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  DevEndpointNames = list(
    "string"
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves the names of all \code{DevEndpoint} resources in this AWS account,
or the resources with the specified tag. This operation allows you to
see which resources are available in your account, and their names.

This operation takes the optional \code{Tags} field, which you can use as a
filter on the response so that tagged resources can be retrieved as a
group. If you choose to use tags filtering, only resources with the tag
are retrieved.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_dev_endpoints(
  NextToken = "string",
  MaxResults = 123,
  Tags = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
