% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_list_entities}
\alias{glue_list_entities}
\title{Returns the available entities supported by the connection type}
\usage{
glue_list_entities(
  ConnectionName = NULL,
  CatalogId = NULL,
  ParentEntityName = NULL,
  NextToken = NULL,
  DataStoreApiVersion = NULL
)
}
\arguments{
\item{ConnectionName}{A name for the connection that has required credentials to query any
connection type.}

\item{CatalogId}{The catalog ID of the catalog that contains the connection. This can be
null, By default, the Amazon Web Services Account ID is the catalog ID.}

\item{ParentEntityName}{Name of the parent entity for which you want to list the children. This
parameter takes a fully-qualified path of the entity in order to list
the child entities.}

\item{NextToken}{A continuation token, included if this is a continuation call.}

\item{DataStoreApiVersion}{The API version of the SaaS connector.}
}
\description{
Returns the available entities supported by the connection type.

See \url{https://www.paws-r-sdk.com/docs/glue_list_entities/} for full documentation.
}
\keyword{internal}
