% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_list_glossaries}
\alias{glue_list_glossaries}
\title{Lists business glossaries in Glue Data Catalog}
\usage{
glue_list_glossaries(MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{The maximum number of results to return in the response.}

\item{NextToken}{A continuation token, if this is a continuation call.}
}
\description{
Lists business glossaries in Glue Data Catalog.

See \url{https://www.paws-r-sdk.com/docs/glue_list_glossaries/} for full documentation.
}
\keyword{internal}
