% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_list_glossary_terms}
\alias{glue_list_glossary_terms}
\title{Lists glossary terms within a business glossary in Glue Data Catalog}
\usage{
glue_list_glossary_terms(
  GlossaryIdentifier,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{GlossaryIdentifier}{[required] The unique identifier of the glossary whose terms to list.}

\item{MaxResults}{The maximum number of results to return in the response.}

\item{NextToken}{A continuation token, if this is a continuation call.}
}
\description{
Lists glossary terms within a business glossary in Glue Data Catalog.

See \url{https://www.paws-r-sdk.com/docs/glue_list_glossary_terms/} for full documentation.
}
\keyword{internal}
