% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_list_integration_resource_properties}
\alias{glue_list_integration_resource_properties}
\title{List integration resource properties for a single customer}
\usage{
glue_list_integration_resource_properties(
  Marker = NULL,
  Filters = NULL,
  MaxRecords = NULL
)
}
\arguments{
\item{Marker}{This is the pagination token for next page, initial value is \code{null}.}

\item{Filters}{A list of filters, supported filter Key is \code{SourceArn} and \code{TargetArn}.}

\item{MaxRecords}{This is total number of items to be evaluated.}
}
\description{
List integration resource properties for a single customer. It supports the filters, maxRecords and markers.

See \url{https://www.paws-r-sdk.com/docs/glue_list_integration_resource_properties/} for full documentation.
}
\keyword{internal}
