% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_list_integration_table_properties}
\alias{glue_list_integration_table_properties}
\title{Lists the integration table properties in your account}
\usage{
glue_list_integration_table_properties(
  Marker = NULL,
  Filters = NULL,
  MaxRecords = NULL
)
}
\arguments{
\item{Marker}{The pagination token for the next page of results. The initial value is \code{null}.}

\item{Filters}{A list of filters. Supported filter keys are \code{SourceArn}, \code{TargetArn}, \code{SourceTableName}, and \code{TargetTableName}.}

\item{MaxRecords}{The maximum number of records to return in the response.}
}
\description{
Lists the integration table properties in your account. This operation supports filtering and pagination.

See \url{https://www.paws-r-sdk.com/docs/glue_list_integration_table_properties/} for full documentation.
}
\keyword{internal}
