% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_list_iterable_forms}
\alias{glue_list_iterable_forms}
\title{Lists the items in an iterable form on an asset in Glue Data Catalog}
\usage{
glue_list_iterable_forms(
  AssetIdentifier,
  IterableFormName,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{AssetIdentifier}{[required] The unique identifier of the asset.}

\item{IterableFormName}{[required] The name of the iterable form to list items from.}

\item{MaxResults}{The maximum number of results to return in the response.}

\item{NextToken}{A continuation token, if this is a continuation call.}
}
\description{
Lists the items in an iterable form on an asset in Glue Data Catalog. For example, lists the columns of a table asset.

See \url{https://www.paws-r-sdk.com/docs/glue_list_iterable_forms/} for full documentation.
}
\keyword{internal}
