% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_list_jobs}
\alias{glue_list_jobs}
\title{Retrieves the names of all job resources in this Amazon Web Services
account, or the resources with the specified tag}
\usage{
glue_list_jobs(NextToken = NULL, MaxResults = NULL, Tags = NULL)
}
\arguments{
\item{NextToken}{A continuation token, if this is a continuation request.}

\item{MaxResults}{The maximum size of a list to return.}

\item{Tags}{Specifies to return only these tagged resources.}
}
\description{
Retrieves the names of all job resources in this Amazon Web Services account, or the resources with the specified tag. This operation allows you to see which resources are available in your account, and their names.

See \url{https://www.paws-r-sdk.com/docs/glue_list_jobs/} for full documentation.
}
\keyword{internal}
