% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_list_sessions}
\alias{glue_list_sessions}
\title{Retrieve a list of sessions}
\usage{
glue_list_sessions(
  NextToken = NULL,
  MaxResults = NULL,
  Tags = NULL,
  RequestOrigin = NULL
)
}
\arguments{
\item{NextToken}{The token for the next set of results, or null if there are no more
result.}

\item{MaxResults}{The maximum number of results.}

\item{Tags}{Tags belonging to the session.}

\item{RequestOrigin}{The origin of the request.}
}
\description{
Retrieve a list of sessions.

See \url{https://paws-r.github.io/docs/glue/list_sessions.html} for full documentation.
}
\keyword{internal}
