% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_list_statements}
\alias{glue_list_statements}
\title{Lists statements for the session}
\usage{
glue_list_statements(SessionId, RequestOrigin = NULL, NextToken = NULL)
}
\arguments{
\item{SessionId}{[required] The Session ID of the statements.}

\item{RequestOrigin}{The origin of the request to list statements.}

\item{NextToken}{A continuation token, if this is a continuation call.}
}
\description{
Lists statements for the session.

See \url{https://www.paws-r-sdk.com/docs/glue_list_statements/} for full documentation.
}
\keyword{internal}
