% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_list_usage_profiles}
\alias{glue_list_usage_profiles}
\title{List all the Glue usage profiles}
\usage{
glue_list_usage_profiles(NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{NextToken}{A continuation token, included if this is a continuation call.}

\item{MaxResults}{The maximum number of usage profiles to return in a single response.}
}
\description{
List all the Glue usage profiles.

See \url{https://www.paws-r-sdk.com/docs/glue_list_usage_profiles/} for full documentation.
}
\keyword{internal}
