% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_list_workflows}
\alias{glue_list_workflows}
\title{Lists names of workflows created in the account}
\usage{
glue_list_workflows(NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{NextToken}{A continuation token, if this is a continuation request.}

\item{MaxResults}{The maximum size of a list to return.}
}
\description{
Lists names of workflows created in the account.

See \url{https://paws-r.github.io/docs/glue/list_workflows.html} for full documentation.
}
\keyword{internal}
