% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_modify_integration}
\alias{glue_modify_integration}
\title{Modifies a Zero-ETL integration in the caller's account}
\usage{
glue_modify_integration(
  IntegrationIdentifier,
  Description = NULL,
  DataFilter = NULL,
  IntegrationConfig = NULL,
  IntegrationName = NULL
)
}
\arguments{
\item{IntegrationIdentifier}{[required] The Amazon Resource Name (ARN) for the integration.}

\item{Description}{A description of the integration.}

\item{DataFilter}{Selects source tables for the integration using Maxwell filter syntax.}

\item{IntegrationConfig}{The configuration settings for the integration. Currently, only the
RefreshInterval can be modified.}

\item{IntegrationName}{A unique name for an integration in Glue.}
}
\description{
Modifies a Zero-ETL integration in the caller's account.

See \url{https://www.paws-r-sdk.com/docs/glue_modify_integration/} for full documentation.
}
\keyword{internal}
