% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_put_attachment}
\alias{glue_put_attachment}
\title{Attaches a form to an asset or an iterable form item in Glue Data
Catalog}
\usage{
glue_put_attachment(
  AssetIdentifier,
  IterableFormName = NULL,
  ItemIdentifier = NULL,
  AttachmentName,
  Content,
  FormTypeId,
  ClientToken = NULL
)
}
\arguments{
\item{AssetIdentifier}{[required] The unique identifier of the asset to attach the form to.}

\item{IterableFormName}{The name of the iterable form. When specified along with \code{itemIdentifier}, the attachment targets an item within the iterable form rather than the asset itself.}

\item{ItemIdentifier}{The identifier of the item within the iterable form. Required when \code{iterableFormName} is specified.}

\item{AttachmentName}{[required] The name of the attachment.}

\item{Content}{[required] The JSON content of the form, conforming to the schema of the specified form type.}

\item{FormTypeId}{[required] The identifier of the form type for this attachment.}

\item{ClientToken}{A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.}
}
\description{
Attaches a form to an asset or an iterable form item in Glue Data Catalog. If an attachment with the same name already exists, it is overwritten.

See \url{https://www.paws-r-sdk.com/docs/glue_put_attachment/} for full documentation.
}
\keyword{internal}
