% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_put_data_catalog_encryption_settings}
\alias{glue_put_data_catalog_encryption_settings}
\title{Sets the security configuration for a specified catalog}
\usage{
glue_put_data_catalog_encryption_settings(
  CatalogId = NULL,
  DataCatalogEncryptionSettings
)
}
\arguments{
\item{CatalogId}{The ID of the Data Catalog to set the security configuration for. If
none is provided, the Amazon Web Services account ID is used by default.}

\item{DataCatalogEncryptionSettings}{[required] The security configuration to set.}
}
\description{
Sets the security configuration for a specified catalog. After the configuration has been set, the specified encryption is applied to every catalog write thereafter.

See \url{https://paws-r.github.io/docs/glue/put_data_catalog_encryption_settings.html} for full documentation.
}
\keyword{internal}
