% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_put_data_catalog_export_configuration}
\alias{glue_put_data_catalog_export_configuration}
\title{Creates or updates the export configuration for the Glue Data Catalog}
\usage{
glue_put_data_catalog_export_configuration(
  ExportSetting,
  EncryptionConfiguration = NULL,
  ClientToken = NULL
)
}
\arguments{
\item{ExportSetting}{[required] The export setting for the data catalog. Specify \code{ENABLED} to start exporting catalog metadata to S3 Tables, or \code{DISABLED} to stop exporting. This field is required.}

\item{EncryptionConfiguration}{The encryption configuration for the exported data. If not specified, the default encryption settings are used.}

\item{ClientToken}{A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.}
}
\description{
Creates or updates the export configuration for the Glue Data Catalog. Use this operation to enable or disable the export of catalog metadata to S3 Tables.

See \url{https://www.paws-r-sdk.com/docs/glue_put_data_catalog_export_configuration/} for full documentation.
}
\keyword{internal}
