% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_put_data_quality_profile_annotation}
\alias{glue_put_data_quality_profile_annotation}
\title{Annotate all datapoints for a Profile}
\usage{
glue_put_data_quality_profile_annotation(ProfileId, InclusionAnnotation)
}
\arguments{
\item{ProfileId}{[required] The ID of the data quality monitoring profile to annotate.}

\item{InclusionAnnotation}{[required] The inclusion annotation value to apply to the profile.}
}
\description{
Annotate all datapoints for a Profile.

See \url{https://www.paws-r-sdk.com/docs/glue_put_data_quality_profile_annotation/} for full documentation.
}
\keyword{internal}
