% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_put_form_type}
\alias{glue_put_form_type}
\title{Creates or updates a form type in Glue Data Catalog}
\usage{
glue_put_form_type(Name, Schema, ClientToken = NULL)
}
\arguments{
\item{Name}{[required] The name of the form type. Must start with an uppercase letter.}

\item{Schema}{[required] The Smithy IDL schema definition for the form type.}

\item{ClientToken}{A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.}
}
\description{
Creates or updates a form type in Glue Data Catalog. A form type defines the schema for structured metadata that can be attached to assets.

See \url{https://www.paws-r-sdk.com/docs/glue_put_form_type/} for full documentation.
}
\keyword{internal}
