% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_put_schema_version_metadata}
\alias{glue_put_schema_version_metadata}
\title{Puts the metadata key value pair for a specified schema version ID}
\usage{
glue_put_schema_version_metadata(
  SchemaId = NULL,
  SchemaVersionNumber = NULL,
  SchemaVersionId = NULL,
  MetadataKeyValue
)
}
\arguments{
\item{SchemaId}{The unique ID for the schema.}

\item{SchemaVersionNumber}{The version number of the schema.}

\item{SchemaVersionId}{The unique version ID of the schema version.}

\item{MetadataKeyValue}{[required] The metadata key's corresponding value.}
}
\description{
Puts the metadata key value pair for a specified schema version ID. A maximum of 10 key value pairs will be allowed per schema version. They can be added over one or more calls.

See \url{https://www.paws-r-sdk.com/docs/glue_put_schema_version_metadata/} for full documentation.
}
\keyword{internal}
