% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_put_workflow_run_properties}
\alias{glue_put_workflow_run_properties}
\title{Puts the specified workflow run properties for the given workflow run}
\usage{
glue_put_workflow_run_properties(Name, RunId, RunProperties)
}
\arguments{
\item{Name}{[required] Name of the workflow which was run.}

\item{RunId}{[required] The ID of the workflow run for which the run properties should be
updated.}

\item{RunProperties}{[required] The properties to put for the specified run.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Puts the specified workflow run properties for the given workflow run.
If a property already exists for the specified run, then it overrides
the value otherwise adds the property to existing properties.
}
\section{Request syntax}{
\preformatted{svc$put_workflow_run_properties(
  Name = "string",
  RunId = "string",
  RunProperties = list(
    "string"
  )
)
}
}

\keyword{internal}
