% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_register_connection_type}
\alias{glue_register_connection_type}
\title{Registers a custom connection type in Glue based on the configuration
provided}
\usage{
glue_register_connection_type(
  ConnectionType,
  IntegrationType,
  Description = NULL,
  ConnectionProperties,
  ConnectorAuthenticationConfiguration,
  RestConfiguration,
  Tags = NULL
)
}
\arguments{
\item{ConnectionType}{[required] The name of the connection type. Must be between 1 and 255 characters
and must be prefixed with "REST-" to indicate it is a REST-based
connector.}

\item{IntegrationType}{[required] The integration type for the connection. Currently only "REST" protocol
is supported.}

\item{Description}{A description of the connection type. Can be up to 2048 characters and
provides details about the purpose and functionality of the connection
type.}

\item{ConnectionProperties}{[required] Defines the base URL and additional request parameters needed during
connection creation for this connection type.}

\item{ConnectorAuthenticationConfiguration}{[required] Defines the supported authentication types and required properties for
this connection type, including Basic, OAuth2, and Custom authentication
methods.}

\item{RestConfiguration}{[required] Defines the HTTP request and response configuration, validation
endpoint, and entity configurations for REST API interactions.}

\item{Tags}{The tags you assign to the connection type.}
}
\description{
Registers a custom connection type in Glue based on the configuration provided. This operation enables customers to configure custom connectors for any data source with REST-based APIs, eliminating the need for building custom Lambda connectors.

See \url{https://www.paws-r-sdk.com/docs/glue_register_connection_type/} for full documentation.
}
\keyword{internal}
