% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_remove_schema_version_metadata}
\alias{glue_remove_schema_version_metadata}
\title{Removes a key value pair from the schema version metadata for the
specified schema version ID}
\usage{
glue_remove_schema_version_metadata(
  SchemaId = NULL,
  SchemaVersionNumber = NULL,
  SchemaVersionId = NULL,
  MetadataKeyValue
)
}
\arguments{
\item{SchemaId}{A wrapper structure that may contain the schema name and Amazon Resource
Name (ARN).}

\item{SchemaVersionNumber}{The version number of the schema.}

\item{SchemaVersionId}{The unique version ID of the schema version.}

\item{MetadataKeyValue}{[required] The value of the metadata key.}
}
\description{
Removes a key value pair from the schema version metadata for the specified schema version ID.

See \url{https://paws-r.github.io/docs/glue/remove_schema_version_metadata.html} for full documentation.
}
\keyword{internal}
