% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_run_statement}
\alias{glue_run_statement}
\title{Executes the statement}
\usage{
glue_run_statement(SessionId, Code, RequestOrigin = NULL)
}
\arguments{
\item{SessionId}{[required] The Session Id of the statement to be run.}

\item{Code}{[required] The statement code to be run.}

\item{RequestOrigin}{The origin of the request.}
}
\description{
Executes the statement.

See \url{https://paws-r.github.io/docs/glue/run_statement.html} for full documentation.
}
\keyword{internal}
