% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_start_blueprint_run}
\alias{glue_start_blueprint_run}
\title{Starts a new run of the specified blueprint}
\usage{
glue_start_blueprint_run(BlueprintName, Parameters = NULL, RoleArn)
}
\arguments{
\item{BlueprintName}{[required] The name of the blueprint.}

\item{Parameters}{Specifies the parameters as a \code{BlueprintParameters} object.}

\item{RoleArn}{[required] Specifies the IAM role used to create the workflow.}
}
\description{
Starts a new run of the specified blueprint.

See \url{https://www.paws-r-sdk.com/docs/glue_start_blueprint_run/} for full documentation.
}
\keyword{internal}
