% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_start_crawler}
\alias{glue_start_crawler}
\title{Starts a crawl using the specified crawler, regardless of what is
scheduled}
\usage{
glue_start_crawler(Name)
}
\arguments{
\item{Name}{[required] Name of the crawler to start.}
}
\description{
Starts a crawl using the specified crawler, regardless of what is scheduled. If the crawler is already running, returns a \href{https://docs.aws.amazon.com/glue/latest/dg/aws-glue-api-exceptions.html#aws-glue-api-exceptions-CrawlerRunningException}{CrawlerRunningException}.

See \url{https://paws-r.github.io/docs/glue/start_crawler.html} for full documentation.
}
\keyword{internal}
