% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_start_materialized_view_refresh_task_run}
\alias{glue_start_materialized_view_refresh_task_run}
\title{Starts a materialized view refresh task run for a specified materialized
view}
\usage{
glue_start_materialized_view_refresh_task_run(
  CatalogId,
  DatabaseName,
  TableName,
  FullRefresh = NULL
)
}
\arguments{
\item{CatalogId}{[required] The ID of the Data Catalog where the table reside. If none is supplied, the account ID is used by default.}

\item{DatabaseName}{[required] The name of the database where the table resides.}

\item{TableName}{[required] The name of the materialized view to run the refresh task for.}

\item{FullRefresh}{Specifies whether this is a full refresh of the task run.}
}
\description{
Starts a materialized view refresh task run for a specified materialized view.

See \url{https://www.paws-r-sdk.com/docs/glue_start_materialized_view_refresh_task_run/} for full documentation.
}
\keyword{internal}
