% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_start_ml_evaluation_task_run}
\alias{glue_start_ml_evaluation_task_run}
\title{Starts a task to estimate the quality of the transform}
\usage{
glue_start_ml_evaluation_task_run(TransformId)
}
\arguments{
\item{TransformId}{[required] The unique identifier of the machine learning transform.}
}
\description{
Starts a task to estimate the quality of the transform.

See \url{https://www.paws-r-sdk.com/docs/glue_start_ml_evaluation_task_run/} for full documentation.
}
\keyword{internal}
