% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_start_ml_labeling_set_generation_task_run}
\alias{glue_start_ml_labeling_set_generation_task_run}
\title{Starts the active learning workflow for your machine learning transform
to improve the transform's quality by generating label sets and adding
labels}
\usage{
glue_start_ml_labeling_set_generation_task_run(TransformId, OutputS3Path)
}
\arguments{
\item{TransformId}{[required] The unique identifier of the machine learning transform.}

\item{OutputS3Path}{[required] The Amazon Simple Storage Service (Amazon S3) path where you generate
the labeling set.}
}
\description{
Starts the active learning workflow for your machine learning transform to improve the transform's quality by generating label sets and adding labels.

See \url{https://www.paws-r-sdk.com/docs/glue_start_ml_labeling_set_generation_task_run/} for full documentation.
}
\keyword{internal}
