% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_start_trigger}
\alias{glue_start_trigger}
\title{Starts an existing trigger}
\usage{
glue_start_trigger(Name)
}
\arguments{
\item{Name}{[required] The name of the trigger to start.}
}
\value{
A list with the following syntax:\preformatted{list(
  Name = "string"
)
}
}
\description{
Starts an existing trigger. See \href{https://docs.aws.amazon.com/glue/latest/dg/trigger-job.html}{Triggering Jobs} for
information about how different types of trigger are started.
}
\section{Request syntax}{
\preformatted{svc$start_trigger(
  Name = "string"
)
}
}

\keyword{internal}
