% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_start_workflow_run}
\alias{glue_start_workflow_run}
\title{Starts a new run of the specified workflow}
\usage{
glue_start_workflow_run(Name)
}
\arguments{
\item{Name}{[required] The name of the workflow to start.}
}
\description{
Starts a new run of the specified workflow.
}
\section{Request syntax}{
\preformatted{svc$start_workflow_run(
  Name = "string"
)
}
}

\keyword{internal}
