% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_stop_column_statistics_task_run}
\alias{glue_stop_column_statistics_task_run}
\title{Stops a task run for the specified table}
\usage{
glue_stop_column_statistics_task_run(DatabaseName, TableName, CatalogID = NULL)
}
\arguments{
\item{DatabaseName}{[required] The name of the database where the table resides.}

\item{TableName}{[required] The name of the table.}

\item{CatalogID}{The ID of the Data Catalog where the table resides. If none is supplied, the Amazon Web Services account ID is used by default.}
}
\description{
Stops a task run for the specified table.

See \url{https://www.paws-r-sdk.com/docs/glue_stop_column_statistics_task_run/} for full documentation.
}
\keyword{internal}
