% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_stop_crawler}
\alias{glue_stop_crawler}
\title{If the specified crawler is running, stops the crawl}
\usage{
glue_stop_crawler(Name)
}
\arguments{
\item{Name}{[required] Name of the crawler to stop.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
If the specified crawler is running, stops the crawl.
}
\section{Request syntax}{
\preformatted{svc$stop_crawler(
  Name = "string"
)
}
}

\keyword{internal}
