% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_stop_trigger}
\alias{glue_stop_trigger}
\title{Stops a specified trigger}
\usage{
glue_stop_trigger(Name)
}
\arguments{
\item{Name}{[required] The name of the trigger to stop.}
}
\description{
Stops a specified trigger.

See \url{https://www.paws-r-sdk.com/docs/glue_stop_trigger/} for full documentation.
}
\keyword{internal}
