% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_stop_workflow_run}
\alias{glue_stop_workflow_run}
\title{Stops the execution of the specified workflow run}
\usage{
glue_stop_workflow_run(Name, RunId)
}
\arguments{
\item{Name}{[required] The name of the workflow to stop.}

\item{RunId}{[required] The ID of the workflow run to stop.}
}
\description{
Stops the execution of the specified workflow run.

See \url{https://paws-r.github.io/docs/glue/stop_workflow_run.html} for full documentation.
}
\keyword{internal}
