% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_test_connection}
\alias{glue_test_connection}
\title{Tests a connection to a service to validate the service credentials that
you provide}
\usage{
glue_test_connection(
  ConnectionName = NULL,
  CatalogId = NULL,
  TestConnectionInput = NULL
)
}
\arguments{
\item{ConnectionName}{Optional. The name of the connection to test. If only name is provided,
the operation will get the connection and use that for testing.}

\item{CatalogId}{The catalog ID where the connection resides.}

\item{TestConnectionInput}{A structure that is used to specify testing a connection to a service.}
}
\description{
Tests a connection to a service to validate the service credentials that you provide.

See \url{https://www.paws-r-sdk.com/docs/glue_test_connection/} for full documentation.
}
\keyword{internal}
