% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_untag_resource}
\alias{glue_untag_resource}
\title{Removes tags from a resource}
\usage{
glue_untag_resource(ResourceArn, TagsToRemove)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource from which to remove the
tags.}

\item{TagsToRemove}{[required] Tags to remove from this resource.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Removes tags from a resource.
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  ResourceArn = "string",
  TagsToRemove = list(
    "string"
  )
)
}
}

\keyword{internal}
