% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_update_asset}
\alias{glue_update_asset}
\title{Updates the name and description of an existing asset in Glue Data
Catalog}
\usage{
glue_update_asset(
  Identifier,
  Name = NULL,
  Description = NULL,
  ClientToken = NULL
)
}
\arguments{
\item{Identifier}{[required] The unique identifier of the asset to update.}

\item{Name}{The new name of the asset.}

\item{Description}{The new description of the asset.}

\item{ClientToken}{A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.}
}
\description{
Updates the name and description of an existing asset in Glue Data Catalog. Only the fields that you provide are updated.

See \url{https://www.paws-r-sdk.com/docs/glue_update_asset/} for full documentation.
}
\keyword{internal}
