% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_update_blueprint}
\alias{glue_update_blueprint}
\title{Updates a registered blueprint}
\usage{
glue_update_blueprint(Name, Description = NULL, BlueprintLocation)
}
\arguments{
\item{Name}{[required] The name of the blueprint.}

\item{Description}{A description of the blueprint.}

\item{BlueprintLocation}{[required] Specifies a path in Amazon S3 where the blueprint is published.}
}
\description{
Updates a registered blueprint.

See \url{https://www.paws-r-sdk.com/docs/glue_update_blueprint/} for full documentation.
}
\keyword{internal}
