% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_update_catalog}
\alias{glue_update_catalog}
\title{Updates an existing catalog's properties in the Glue Data Catalog}
\usage{
glue_update_catalog(CatalogId, CatalogInput)
}
\arguments{
\item{CatalogId}{[required] The ID of the catalog.}

\item{CatalogInput}{[required] A \code{CatalogInput} object specifying the new properties of an existing
catalog.}
}
\description{
Updates an existing catalog's properties in the Glue Data Catalog.

See \url{https://www.paws-r-sdk.com/docs/glue_update_catalog/} for full documentation.
}
\keyword{internal}
