% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_update_classifier}
\alias{glue_update_classifier}
\title{Modifies an existing classifier (a GrokClassifier, an XMLClassifier, a
JsonClassifier, or a CsvClassifier, depending on which field is present)}
\usage{
glue_update_classifier(
  GrokClassifier = NULL,
  XMLClassifier = NULL,
  JsonClassifier = NULL,
  CsvClassifier = NULL
)
}
\arguments{
\item{GrokClassifier}{A \code{GrokClassifier} object with updated fields.}

\item{XMLClassifier}{An \code{XMLClassifier} object with updated fields.}

\item{JsonClassifier}{A \code{JsonClassifier} object with updated fields.}

\item{CsvClassifier}{A \code{CsvClassifier} object with updated fields.}
}
\description{
Modifies an existing classifier (a \code{GrokClassifier}, an \code{XMLClassifier}, a \code{JsonClassifier}, or a \code{CsvClassifier}, depending on which field is present).

See \url{https://www.paws-r-sdk.com/docs/glue_update_classifier/} for full documentation.
}
\keyword{internal}
