% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_update_connection}
\alias{glue_update_connection}
\title{Updates a connection definition in the Data Catalog}
\usage{
glue_update_connection(CatalogId = NULL, Name, ConnectionInput)
}
\arguments{
\item{CatalogId}{The ID of the Data Catalog in which the connection resides. If none is
provided, the Amazon Web Services account ID is used by default.}

\item{Name}{[required] The name of the connection definition to update.}

\item{ConnectionInput}{[required] A \code{ConnectionInput} object that redefines the connection in question.}
}
\description{
Updates a connection definition in the Data Catalog.

See \url{https://www.paws-r-sdk.com/docs/glue_update_connection/} for full documentation.
}
\keyword{internal}
