% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_update_database}
\alias{glue_update_database}
\title{Updates an existing database definition in a Data Catalog}
\usage{
glue_update_database(CatalogId, Name, DatabaseInput)
}
\arguments{
\item{CatalogId}{The ID of the Data Catalog in which the metadata database resides. If
none is provided, the AWS account ID is used by default.}

\item{Name}{[required] The name of the database to update in the catalog. For Hive
compatibility, this is folded to lowercase.}

\item{DatabaseInput}{[required] A \code{DatabaseInput} object specifying the new definition of the metadata
database in the catalog.}
}
\value{
An empty list.
}
\description{
Updates an existing database definition in a Data Catalog.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_database(
  CatalogId = "string",
  Name = "string",
  DatabaseInput = list(
    Name = "string",
    Description = "string",
    LocationUri = "string",
    Parameters = list(
      "string"
    ),
    CreateTableDefaultPermissions = list(
      list(
        Principal = list(
          DataLakePrincipalIdentifier = "string"
        ),
        Permissions = list(
          "ALL"|"SELECT"|"ALTER"|"DROP"|"DELETE"|"INSERT"|"CREATE_DATABASE"|"CREATE_TABLE"|"DATA_LOCATION_ACCESS"
        )
      )
    ),
    TargetDatabase = list(
      CatalogId = "string",
      DatabaseName = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
