% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_update_glossary}
\alias{glue_update_glossary}
\title{Updates a business glossary in Glue Data Catalog}
\usage{
glue_update_glossary(
  Identifier,
  Name = NULL,
  Description = NULL,
  ClientToken = NULL
)
}
\arguments{
\item{Identifier}{[required] The unique identifier of the glossary to update.}

\item{Name}{The updated name of the glossary.}

\item{Description}{The updated description of the glossary.}

\item{ClientToken}{A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.}
}
\description{
Updates a business glossary in Glue Data Catalog.

See \url{https://www.paws-r-sdk.com/docs/glue_update_glossary/} for full documentation.
}
\keyword{internal}
