% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_update_glossary_term}
\alias{glue_update_glossary_term}
\title{Updates a glossary term in Glue Data Catalog}
\usage{
glue_update_glossary_term(
  Identifier,
  Name = NULL,
  ShortDescription = NULL,
  LongDescription = NULL,
  ClientToken = NULL
)
}
\arguments{
\item{Identifier}{[required] The unique identifier of the glossary term to update.}

\item{Name}{The updated name of the glossary term.}

\item{ShortDescription}{The updated short description of the glossary term.}

\item{LongDescription}{The updated long description of the glossary term.}

\item{ClientToken}{A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.}
}
\description{
Updates a glossary term in Glue Data Catalog.

See \url{https://www.paws-r-sdk.com/docs/glue_update_glossary_term/} for full documentation.
}
\keyword{internal}
