% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_update_glue_identity_center_configuration}
\alias{glue_update_glue_identity_center_configuration}
\title{Updates the existing Glue Identity Center configuration, allowing
modification of scopes and permissions for the integration}
\usage{
glue_update_glue_identity_center_configuration(
  Scopes = NULL,
  UserBackgroundSessionsEnabled = NULL
)
}
\arguments{
\item{Scopes}{A list of Identity Center scopes that define the updated permissions and
access levels for the Glue configuration.}

\item{UserBackgroundSessionsEnabled}{Specifies whether users can run background sessions when using Identity
Center authentication with Glue services.}
}
\description{
Updates the existing Glue Identity Center configuration, allowing modification of scopes and permissions for the integration.

See \url{https://www.paws-r-sdk.com/docs/glue_update_glue_identity_center_configuration/} for full documentation.
}
\keyword{internal}
