% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_update_integration_resource_property}
\alias{glue_update_integration_resource_property}
\title{This API can be used for updating the ResourceProperty of the Glue
connection (for the source) or Glue database ARN (for the target)}
\usage{
glue_update_integration_resource_property(
  ResourceArn,
  SourceProcessingProperties = NULL,
  TargetProcessingProperties = NULL
)
}
\arguments{
\item{ResourceArn}{[required] The connection ARN of the source, or the database ARN of the target.}

\item{SourceProcessingProperties}{The resource properties associated with the integration source.}

\item{TargetProcessingProperties}{The resource properties associated with the integration target.}
}
\description{
This API can be used for updating the \code{ResourceProperty} of the Glue connection (for the source) or Glue database ARN (for the target). These properties can include the role to access the connection or database. Since the same resource can be used across multiple integrations, updating resource properties will impact all the integrations using it.

See \url{https://www.paws-r-sdk.com/docs/glue_update_integration_resource_property/} for full documentation.
}
\keyword{internal}
