% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_update_job}
\alias{glue_update_job}
\title{Updates an existing job definition}
\usage{
glue_update_job(JobName, JobUpdate)
}
\arguments{
\item{JobName}{[required] The name of the job definition to update.}

\item{JobUpdate}{[required] Specifies the values with which to update the job definition.
Unspecified configuration is removed or reset to default values.}
}
\description{
Updates an existing job definition. The previous job definition is completely overwritten by this information.

See \url{https://www.paws-r-sdk.com/docs/glue_update_job/} for full documentation.
}
\keyword{internal}
