% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_update_partition}
\alias{glue_update_partition}
\title{Updates a partition}
\usage{
glue_update_partition(
  CatalogId = NULL,
  DatabaseName,
  TableName,
  PartitionValueList,
  PartitionInput
)
}
\arguments{
\item{CatalogId}{The ID of the Data Catalog where the partition to be updated resides. If
none is provided, the Amazon Web Services account ID is used by default.}

\item{DatabaseName}{[required] The name of the catalog database in which the table in question resides.}

\item{TableName}{[required] The name of the table in which the partition to be updated is located.}

\item{PartitionValueList}{[required] List of partition key values that define the partition to update.}

\item{PartitionInput}{[required] The new partition object to update the partition to.

The \code{Values} property can't be changed. If you want to change the
partition key values for a partition, delete and recreate the partition.}
}
\description{
Updates a partition.

See \url{https://www.paws-r-sdk.com/docs/glue_update_partition/} for full documentation.
}
\keyword{internal}
