% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_update_usage_profile}
\alias{glue_update_usage_profile}
\title{Update an Glue usage profile}
\usage{
glue_update_usage_profile(Name, Description = NULL, Configuration)
}
\arguments{
\item{Name}{[required] The name of the usage profile.}

\item{Description}{A description of the usage profile.}

\item{Configuration}{[required] A \code{ProfileConfiguration} object specifying the job and session values
for the profile.}
}
\description{
Update an Glue usage profile.

See \url{https://www.paws-r-sdk.com/docs/glue_update_usage_profile/} for full documentation.
}
\keyword{internal}
