% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_update_user_defined_function}
\alias{glue_update_user_defined_function}
\title{Updates an existing function definition in the Data Catalog}
\usage{
glue_update_user_defined_function(
  CatalogId = NULL,
  DatabaseName,
  FunctionName,
  FunctionInput
)
}
\arguments{
\item{CatalogId}{The ID of the Data Catalog where the function to be updated is located.
If none is provided, the Amazon Web Services account ID is used by
default.}

\item{DatabaseName}{[required] The name of the catalog database where the function to be updated is
located.}

\item{FunctionName}{[required] The name of the function.}

\item{FunctionInput}{[required] A \code{FunctionInput} object that redefines the function in the Data
Catalog.}
}
\description{
Updates an existing function definition in the Data Catalog.

See \url{https://paws-r.github.io/docs/glue/update_user_defined_function.html} for full documentation.
}
\keyword{internal}
