% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluedatabrew_operations.R
\name{gluedatabrew_batch_delete_recipe_version}
\alias{gluedatabrew_batch_delete_recipe_version}
\title{Deletes one or more versions of a recipe at a time}
\usage{
gluedatabrew_batch_delete_recipe_version(Name, RecipeVersions)
}
\arguments{
\item{Name}{[required] The name of the recipe whose versions are to be deleted.}

\item{RecipeVersions}{[required] An array of version identifiers, for the recipe versions to be deleted.
You can specify numeric versions (\code{X.Y}) or \code{LATEST_WORKING}.
\code{LATEST_PUBLISHED} is not supported.}
}
\description{
Deletes one or more versions of a recipe at a time.

See \url{https://paws-r.github.io/docs/gluedatabrew/batch_delete_recipe_version.html} for full documentation.
}
\keyword{internal}
