% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluedatabrew_operations.R
\name{gluedatabrew_create_recipe}
\alias{gluedatabrew_create_recipe}
\title{Creates a new DataBrew recipe}
\usage{
gluedatabrew_create_recipe(Description = NULL, Name, Steps, Tags = NULL)
}
\arguments{
\item{Description}{A description for the recipe.}

\item{Name}{[required] A unique name for the recipe. Valid characters are alphanumeric (A-Z,
a-z, 0-9), hyphen (-), period (.), and space.}

\item{Steps}{[required] An array containing the steps to be performed by the recipe. Each recipe
step consists of one recipe action and (optionally) an array of
condition expressions.}

\item{Tags}{Metadata tags to apply to this recipe.}
}
\description{
Creates a new DataBrew recipe.

See \url{https://www.paws-r-sdk.com/docs/gluedatabrew_create_recipe/} for full documentation.
}
\keyword{internal}
