% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluedatabrew_operations.R
\name{gluedatabrew_delete_dataset}
\alias{gluedatabrew_delete_dataset}
\title{Deletes a dataset from DataBrew}
\usage{
gluedatabrew_delete_dataset(Name)
}
\arguments{
\item{Name}{[required] The name of the dataset to be deleted.}
}
\description{
Deletes a dataset from DataBrew.

See \url{https://www.paws-r-sdk.com/docs/gluedatabrew_delete_dataset/} for full documentation.
}
\keyword{internal}
