% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluedatabrew_operations.R
\name{gluedatabrew_delete_job}
\alias{gluedatabrew_delete_job}
\title{Deletes the specified DataBrew job}
\usage{
gluedatabrew_delete_job(Name)
}
\arguments{
\item{Name}{[required] The name of the job to be deleted.}
}
\description{
Deletes the specified DataBrew job.

See \url{https://www.paws-r-sdk.com/docs/gluedatabrew_delete_job/} for full documentation.
}
\keyword{internal}
