% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluedatabrew_operations.R
\name{gluedatabrew_delete_project}
\alias{gluedatabrew_delete_project}
\title{Deletes an existing DataBrew project}
\usage{
gluedatabrew_delete_project(Name)
}
\arguments{
\item{Name}{[required] The name of the project to be deleted.}
}
\description{
Deletes an existing DataBrew project.

See \url{https://paws-r.github.io/docs/gluedatabrew/delete_project.html} for full documentation.
}
\keyword{internal}
