% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluedatabrew_operations.R
\name{gluedatabrew_delete_recipe_version}
\alias{gluedatabrew_delete_recipe_version}
\title{Deletes a single version of a DataBrew recipe}
\usage{
gluedatabrew_delete_recipe_version(Name, RecipeVersion)
}
\arguments{
\item{Name}{[required] The name of the recipe.}

\item{RecipeVersion}{[required] The version of the recipe to be deleted. You can specify a numeric
versions (\code{X.Y}) or \code{LATEST_WORKING}. \code{LATEST_PUBLISHED} is not
supported.}
}
\description{
Deletes a single version of a DataBrew recipe.

See \url{https://paws-r.github.io/docs/gluedatabrew/delete_recipe_version.html} for full documentation.
}
\keyword{internal}
