% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluedatabrew_operations.R
\name{gluedatabrew_delete_ruleset}
\alias{gluedatabrew_delete_ruleset}
\title{Deletes a ruleset}
\usage{
gluedatabrew_delete_ruleset(Name)
}
\arguments{
\item{Name}{[required] The name of the ruleset to be deleted.}
}
\description{
Deletes a ruleset.

See \url{https://paws-r.github.io/docs/gluedatabrew/delete_ruleset.html} for full documentation.
}
\keyword{internal}
