% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluedatabrew_operations.R
\name{gluedatabrew_delete_schedule}
\alias{gluedatabrew_delete_schedule}
\title{Deletes the specified DataBrew schedule}
\usage{
gluedatabrew_delete_schedule(Name)
}
\arguments{
\item{Name}{[required] The name of the schedule to be deleted.}
}
\description{
Deletes the specified DataBrew schedule.

See \url{https://www.paws-r-sdk.com/docs/gluedatabrew_delete_schedule/} for full documentation.
}
\keyword{internal}
