% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluedatabrew_operations.R
\name{gluedatabrew_describe_job}
\alias{gluedatabrew_describe_job}
\title{Returns the definition of a specific DataBrew job}
\usage{
gluedatabrew_describe_job(Name)
}
\arguments{
\item{Name}{[required] The name of the job to be described.}
}
\description{
Returns the definition of a specific DataBrew job.

See \url{https://paws-r.github.io/docs/gluedatabrew/describe_job.html} for full documentation.
}
\keyword{internal}
