% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluedatabrew_operations.R
\name{gluedatabrew_describe_job_run}
\alias{gluedatabrew_describe_job_run}
\title{Represents one run of a DataBrew job}
\usage{
gluedatabrew_describe_job_run(Name, RunId)
}
\arguments{
\item{Name}{[required] The name of the job being processed during this run.}

\item{RunId}{[required] The unique identifier of the job run.}
}
\description{
Represents one run of a DataBrew job.

See \url{https://www.paws-r-sdk.com/docs/gluedatabrew_describe_job_run/} for full documentation.
}
\keyword{internal}
