% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluedatabrew_operations.R
\name{gluedatabrew_describe_recipe}
\alias{gluedatabrew_describe_recipe}
\title{Returns the definition of a specific DataBrew recipe corresponding to a
particular version}
\usage{
gluedatabrew_describe_recipe(Name, RecipeVersion = NULL)
}
\arguments{
\item{Name}{[required] The name of the recipe to be described.}

\item{RecipeVersion}{The recipe version identifier. If this parameter isn't specified, then
the latest published version is returned.}
}
\description{
Returns the definition of a specific DataBrew recipe corresponding to a particular version.

See \url{https://www.paws-r-sdk.com/docs/gluedatabrew_describe_recipe/} for full documentation.
}
\keyword{internal}
