% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluedatabrew_operations.R
\name{gluedatabrew_describe_ruleset}
\alias{gluedatabrew_describe_ruleset}
\title{Retrieves detailed information about the ruleset}
\usage{
gluedatabrew_describe_ruleset(Name)
}
\arguments{
\item{Name}{[required] The name of the ruleset to be described.}
}
\description{
Retrieves detailed information about the ruleset.

See \url{https://paws-r.github.io/docs/gluedatabrew/describe_ruleset.html} for full documentation.
}
\keyword{internal}
