% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluedatabrew_operations.R
\name{gluedatabrew_list_job_runs}
\alias{gluedatabrew_list_job_runs}
\title{Lists all of the previous runs of a particular DataBrew job}
\usage{
gluedatabrew_list_job_runs(Name, MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{Name}{[required] The name of the job.}

\item{MaxResults}{The maximum number of results to return in this request.}

\item{NextToken}{The token returned by a previous call to retrieve the next set of
results.}
}
\description{
Lists all of the previous runs of a particular DataBrew job.

See \url{https://www.paws-r-sdk.com/docs/gluedatabrew_list_job_runs/} for full documentation.
}
\keyword{internal}
