% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluedatabrew_operations.R
\name{gluedatabrew_list_projects}
\alias{gluedatabrew_list_projects}
\title{Lists all of the DataBrew projects that are defined}
\usage{
gluedatabrew_list_projects(NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{NextToken}{The token returned by a previous call to retrieve the next set of
results.}

\item{MaxResults}{The maximum number of results to return in this request.}
}
\description{
Lists all of the DataBrew projects that are defined.

See \url{https://paws-r.github.io/docs/gluedatabrew/list_projects.html} for full documentation.
}
\keyword{internal}
