% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluedatabrew_operations.R
\name{gluedatabrew_list_recipe_versions}
\alias{gluedatabrew_list_recipe_versions}
\title{Lists the versions of a particular DataBrew recipe, except for
LATEST_WORKING}
\usage{
gluedatabrew_list_recipe_versions(MaxResults = NULL, NextToken = NULL, Name)
}
\arguments{
\item{MaxResults}{The maximum number of results to return in this request.}

\item{NextToken}{The token returned by a previous call to retrieve the next set of
results.}

\item{Name}{[required] The name of the recipe for which to return version information.}
}
\description{
Lists the versions of a particular DataBrew recipe, except for \code{LATEST_WORKING}.

See \url{https://www.paws-r-sdk.com/docs/gluedatabrew_list_recipe_versions/} for full documentation.
}
\keyword{internal}
