% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluedatabrew_operations.R
\name{gluedatabrew_list_recipes}
\alias{gluedatabrew_list_recipes}
\title{Lists all of the DataBrew recipes that are defined}
\usage{
gluedatabrew_list_recipes(
  MaxResults = NULL,
  NextToken = NULL,
  RecipeVersion = NULL
)
}
\arguments{
\item{MaxResults}{The maximum number of results to return in this request.}

\item{NextToken}{The token returned by a previous call to retrieve the next set of
results.}

\item{RecipeVersion}{Return only those recipes with a version identifier of \code{LATEST_WORKING}
or \code{LATEST_PUBLISHED}. If \code{RecipeVersion} is omitted,
\code{\link[=gluedatabrew_list_recipes]{list_recipes}} returns all of the
\code{LATEST_PUBLISHED} recipe versions.

Valid values: \code{LATEST_WORKING} | \code{LATEST_PUBLISHED}}
}
\description{
Lists all of the DataBrew recipes that are defined.

See \url{https://paws-r.github.io/docs/gluedatabrew/list_recipes.html} for full documentation.
}
\keyword{internal}
