% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluedatabrew_operations.R
\name{gluedatabrew_list_rulesets}
\alias{gluedatabrew_list_rulesets}
\title{List all rulesets available in the current account or rulesets
associated with a specific resource (dataset)}
\usage{
gluedatabrew_list_rulesets(
  TargetArn = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{TargetArn}{The Amazon Resource Name (ARN) of a resource (dataset). Using this
parameter indicates to return only those rulesets that are associated
with the specified resource.}

\item{MaxResults}{The maximum number of results to return in this request.}

\item{NextToken}{A token generated by DataBrew that specifies where to continue
pagination if a previous request was truncated. To get the next set of
pages, pass in the NextToken value from the response object of the
previous page call.}
}
\description{
List all rulesets available in the current account or rulesets associated with a specific resource (dataset).

See \url{https://www.paws-r-sdk.com/docs/gluedatabrew_list_rulesets/} for full documentation.
}
\keyword{internal}
