% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluedatabrew_operations.R
\name{gluedatabrew_list_schedules}
\alias{gluedatabrew_list_schedules}
\title{Lists the DataBrew schedules that are defined}
\usage{
gluedatabrew_list_schedules(
  JobName = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{JobName}{The name of the job that these schedules apply to.}

\item{MaxResults}{The maximum number of results to return in this request.}

\item{NextToken}{The token returned by a previous call to retrieve the next set of
results.}
}
\description{
Lists the DataBrew schedules that are defined.

See \url{https://www.paws-r-sdk.com/docs/gluedatabrew_list_schedules/} for full documentation.
}
\keyword{internal}
