% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluedatabrew_operations.R
\name{gluedatabrew_publish_recipe}
\alias{gluedatabrew_publish_recipe}
\title{Publishes a new version of a DataBrew recipe}
\usage{
gluedatabrew_publish_recipe(Description = NULL, Name)
}
\arguments{
\item{Description}{A description of the recipe to be published, for this version of the
recipe.}

\item{Name}{[required] The name of the recipe to be published.}
}
\description{
Publishes a new version of a DataBrew recipe.

See \url{https://www.paws-r-sdk.com/docs/gluedatabrew_publish_recipe/} for full documentation.
}
\keyword{internal}
