% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluedatabrew_operations.R
\name{gluedatabrew_start_job_run}
\alias{gluedatabrew_start_job_run}
\title{Runs a DataBrew job}
\usage{
gluedatabrew_start_job_run(Name)
}
\arguments{
\item{Name}{[required] The name of the job to be run.}
}
\description{
Runs a DataBrew job.

See \url{https://www.paws-r-sdk.com/docs/gluedatabrew_start_job_run/} for full documentation.
}
\keyword{internal}
