% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluedatabrew_operations.R
\name{gluedatabrew_stop_job_run}
\alias{gluedatabrew_stop_job_run}
\title{Stops a particular run of a job}
\usage{
gluedatabrew_stop_job_run(Name, RunId)
}
\arguments{
\item{Name}{[required] The name of the job to be stopped.}

\item{RunId}{[required] The ID of the job run to be stopped.}
}
\description{
Stops a particular run of a job.

See \url{https://www.paws-r-sdk.com/docs/gluedatabrew_stop_job_run/} for full documentation.
}
\keyword{internal}
