% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluedatabrew_operations.R
\name{gluedatabrew_tag_resource}
\alias{gluedatabrew_tag_resource}
\title{Adds metadata tags to a DataBrew resource, such as a dataset, project,
recipe, job, or schedule}
\usage{
gluedatabrew_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] The DataBrew resource to which tags should be added. The value for this
parameter is an Amazon Resource Name (ARN). For DataBrew, you can tag a
dataset, a job, a project, or a recipe.}

\item{Tags}{[required] One or more tags to be assigned to the resource.}
}
\description{
Adds metadata tags to a DataBrew resource, such as a dataset, project, recipe, job, or schedule.

See \url{https://paws-r.github.io/docs/gluedatabrew/tag_resource.html} for full documentation.
}
\keyword{internal}
