% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluedatabrew_operations.R
\name{gluedatabrew_update_dataset}
\alias{gluedatabrew_update_dataset}
\title{Modifies the definition of an existing DataBrew dataset}
\usage{
gluedatabrew_update_dataset(
  Name,
  Format = NULL,
  FormatOptions = NULL,
  Input,
  PathOptions = NULL
)
}
\arguments{
\item{Name}{[required] The name of the dataset to be updated.}

\item{Format}{The file format of a dataset that is created from an Amazon S3 file or
folder.}

\item{FormatOptions}{}

\item{Input}{[required]}

\item{PathOptions}{A set of options that defines how DataBrew interprets an Amazon S3 path
of the dataset.}
}
\description{
Modifies the definition of an existing DataBrew dataset.

See \url{https://paws-r.github.io/docs/gluedatabrew/update_dataset.html} for full documentation.
}
\keyword{internal}
