% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluedatabrew_operations.R
\name{gluedatabrew_update_project}
\alias{gluedatabrew_update_project}
\title{Modifies the definition of an existing DataBrew project}
\usage{
gluedatabrew_update_project(Sample = NULL, RoleArn, Name)
}
\arguments{
\item{Sample}{}

\item{RoleArn}{[required] The Amazon Resource Name (ARN) of the IAM role to be assumed for this
request.}

\item{Name}{[required] The name of the project to be updated.}
}
\description{
Modifies the definition of an existing DataBrew project.

See \url{https://www.paws-r-sdk.com/docs/gluedatabrew_update_project/} for full documentation.
}
\keyword{internal}
