% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluedatabrew_operations.R
\name{gluedatabrew_update_recipe}
\alias{gluedatabrew_update_recipe}
\title{Modifies the definition of the LATEST_WORKING version of a DataBrew
recipe}
\usage{
gluedatabrew_update_recipe(Description = NULL, Name, Steps = NULL)
}
\arguments{
\item{Description}{A description of the recipe.}

\item{Name}{[required] The name of the recipe to be updated.}

\item{Steps}{One or more steps to be performed by the recipe. Each step consists of
an action, and the conditions under which the action should succeed.}
}
\description{
Modifies the definition of the \code{LATEST_WORKING} version of a DataBrew recipe.

See \url{https://www.paws-r-sdk.com/docs/gluedatabrew_update_recipe/} for full documentation.
}
\keyword{internal}
