% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluedatabrew_operations.R
\name{gluedatabrew_update_ruleset}
\alias{gluedatabrew_update_ruleset}
\title{Updates specified ruleset}
\usage{
gluedatabrew_update_ruleset(Name, Description = NULL, Rules)
}
\arguments{
\item{Name}{[required] The name of the ruleset to be updated.}

\item{Description}{The description of the ruleset.}

\item{Rules}{[required] A list of rules that are defined with the ruleset. A rule includes one
or more checks to be validated on a DataBrew dataset.}
}
\description{
Updates specified ruleset.

See \url{https://paws-r.github.io/docs/gluedatabrew/update_ruleset.html} for full documentation.
}
\keyword{internal}
