% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluedatabrew_operations.R
\name{gluedatabrew_update_schedule}
\alias{gluedatabrew_update_schedule}
\title{Modifies the definition of an existing DataBrew schedule}
\usage{
gluedatabrew_update_schedule(JobNames = NULL, CronExpression, Name)
}
\arguments{
\item{JobNames}{The name or names of one or more jobs to be run for this schedule.}

\item{CronExpression}{[required] The date or dates and time or times when the jobs are to be run. For
more information, see \href{https://docs.aws.amazon.com/databrew/latest/dg/}{Cron expressions} in the
\emph{Glue DataBrew Developer Guide}.}

\item{Name}{[required] The name of the schedule to update.}
}
\description{
Modifies the definition of an existing DataBrew schedule.

See \url{https://paws-r.github.io/docs/gluedatabrew/update_schedule.html} for full documentation.
}
\keyword{internal}
