% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthlake_operations.R
\name{healthlake_create_fhir_datastore}
\alias{healthlake_create_fhir_datastore}
\title{Create a FHIR-enabled data store}
\usage{
healthlake_create_fhir_datastore(
  DatastoreName = NULL,
  DatastoreTypeVersion,
  SseConfiguration = NULL,
  PreloadDataConfig = NULL,
  ClientToken = NULL,
  Tags = NULL,
  IdentityProviderConfiguration = NULL,
  AnalyticsConfiguration = NULL,
  NlpConfiguration = NULL,
  ProfileConfiguration = NULL,
  BackupConfiguration = NULL
)
}
\arguments{
\item{DatastoreName}{The data store name (user-generated).}

\item{DatastoreTypeVersion}{[required] The FHIR release version supported by the data store. Current support is for version \code{R4}.}

\item{SseConfiguration}{The server-side encryption key configuration for a customer-provided encryption key specified for creating a data store.}

\item{PreloadDataConfig}{An optional parameter to preload (import) open source Synthea FHIR data upon creation of the data store.}

\item{ClientToken}{An optional user-provided token to ensure API idempotency.}

\item{Tags}{The resource tags applied to a data store when it is created.}

\item{IdentityProviderConfiguration}{The identity provider configuration to use for the data store.}

\item{AnalyticsConfiguration}{The analytics configuration for the data store.}

\item{NlpConfiguration}{The natural language processing (NLP) configuration for the data store.}

\item{ProfileConfiguration}{The profile configuration for the data store.}

\item{BackupConfiguration}{The backup configuration for the data store.}
}
\description{
Create a FHIR-enabled data store.

See \url{https://www.paws-r-sdk.com/docs/healthlake_create_fhir_datastore/} for full documentation.
}
\keyword{internal}
