% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthlake_operations.R
\name{healthlake_delete_data_transformation_profile}
\alias{healthlake_delete_data_transformation_profile}
\title{Deletes a data transformation profile and all its versions, including
the DRAFT and all published versions}
\usage{
healthlake_delete_data_transformation_profile(ProfileId)
}
\arguments{
\item{ProfileId}{[required] The unique identifier of the profile to delete.}
}
\description{
Deletes a data transformation profile and all its versions, including the DRAFT and all published versions.

See \url{https://www.paws-r-sdk.com/docs/healthlake_delete_data_transformation_profile/} for full documentation.
}
\keyword{internal}
