% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthlake_operations.R
\name{healthlake_describe_data_transformation_job}
\alias{healthlake_describe_data_transformation_job}
\title{Describes a data transformation job, including its current status,
configuration, and progress information}
\usage{
healthlake_describe_data_transformation_job(JobId)
}
\arguments{
\item{JobId}{[required] The unique identifier of the data transformation job to describe.}
}
\description{
Describes a data transformation job, including its current status, configuration, and progress information.

See \url{https://www.paws-r-sdk.com/docs/healthlake_describe_data_transformation_job/} for full documentation.
}
\keyword{internal}
