% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthlake_operations.R
\name{healthlake_describe_fhir_export_job}
\alias{healthlake_describe_fhir_export_job}
\title{Displays the properties of a FHIR export job, including the ID, ARN,
name, and the status of the job}
\usage{
healthlake_describe_fhir_export_job(DatastoreId, JobId)
}
\arguments{
\item{DatastoreId}{[required] The AWS generated ID for the Data Store from which files are being
exported from for an export job.}

\item{JobId}{[required] The AWS generated ID for an export job.}
}
\description{
Displays the properties of a FHIR export job, including the ID, ARN, name, and the status of the job.

See \url{https://paws-r.github.io/docs/healthlake/describe_fhir_export_job.html} for full documentation.
}
\keyword{internal}
