% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthlake_operations.R
\name{healthlake_describe_fhir_import_job}
\alias{healthlake_describe_fhir_import_job}
\title{Displays the properties of a FHIR import job, including the ID, ARN,
name, and the status of the job}
\usage{
healthlake_describe_fhir_import_job(DatastoreId, JobId)
}
\arguments{
\item{DatastoreId}{[required] The AWS-generated ID of the Data Store.}

\item{JobId}{[required] The AWS-generated job ID.}
}
\description{
Displays the properties of a FHIR import job, including the ID, ARN, name, and the status of the job.

See \url{https://www.paws-r-sdk.com/docs/healthlake_describe_fhir_import_job/} for full documentation.
}
\keyword{internal}
