% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthlake_operations.R
\name{healthlake_get_data_transformation_profile}
\alias{healthlake_get_data_transformation_profile}
\title{Retrieves a data transformation profile's metadata and profile content
at a specific version}
\usage{
healthlake_get_data_transformation_profile(ProfileId, ProfileVersion = NULL)
}
\arguments{
\item{ProfileId}{[required] The unique identifier of the profile to retrieve.}

\item{ProfileVersion}{The version number to retrieve. Specify 0 to retrieve the DRAFT version. If you omit this parameter, the service returns the latest published version.}
}
\description{
Retrieves a data transformation profile's metadata and profile content at a specific version. Specify version 0 to retrieve the DRAFT, a version number between 1 and 99 to retrieve a specific published version, or omit the version to retrieve the latest published version.

See \url{https://www.paws-r-sdk.com/docs/healthlake_get_data_transformation_profile/} for full documentation.
}
\keyword{internal}
