% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthlake_operations.R
\name{healthlake_list_data_transformation_jobs}
\alias{healthlake_list_data_transformation_jobs}
\title{Lists data transformation jobs for your Amazon Web Services account}
\usage{
healthlake_list_data_transformation_jobs(
  MaxResults = NULL,
  NextToken = NULL,
  JobStatus = NULL,
  JobName = NULL,
  SubmittedAfter = NULL,
  SubmittedBefore = NULL
)
}
\arguments{
\item{MaxResults}{The maximum number of jobs to return per page. If you don't specify a value, the service returns up to 100 results.}

\item{NextToken}{The pagination token from a previous response. Pass this value to retrieve the next page of results.}

\item{JobStatus}{Filters the results to include only jobs with the specified status.}

\item{JobName}{Filters the results to include only jobs with the specified name.}

\item{SubmittedAfter}{Filters the results to include only jobs submitted at or after this timestamp.}

\item{SubmittedBefore}{Filters the results to include only jobs submitted at or before this timestamp.}
}
\description{
Lists data transformation jobs for your Amazon Web Services account. Results can be filtered by status, job name, and submit time window. Results are paginated. Use the \code{NextToken} parameter to retrieve additional results.

See \url{https://www.paws-r-sdk.com/docs/healthlake_list_data_transformation_jobs/} for full documentation.
}
\keyword{internal}
