% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthlake_operations.R
\name{healthlake_list_data_transformation_profile_versions}
\alias{healthlake_list_data_transformation_profile_versions}
\title{Lists all versions of a specific data transformation profile (DRAFT and
published), in reverse chronological order (newest first)}
\usage{
healthlake_list_data_transformation_profile_versions(
  ProfileId,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{ProfileId}{[required] The unique identifier of the profile whose versions to list.}

\item{MaxResults}{The maximum number of profile versions to return per page. If you don't specify a value, the service returns up to 100 results.}

\item{NextToken}{The pagination token from a previous response. Pass this value to retrieve the next page of results.}
}
\description{
Lists all versions of a specific data transformation profile (DRAFT and published), in reverse chronological order (newest first). Use \code{\link[=healthlake_get_data_transformation_profile]{get_data_transformation_profile}} to retrieve profile content. Results are paginated. Use the \code{NextToken} parameter to retrieve additional results.

See \url{https://www.paws-r-sdk.com/docs/healthlake_list_data_transformation_profile_versions/} for full documentation.
}
\keyword{internal}
