% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthlake_operations.R
\name{healthlake_list_data_transformation_profiles}
\alias{healthlake_list_data_transformation_profiles}
\title{Lists all data transformation profiles in your account, returning the
latest version summary for each}
\usage{
healthlake_list_data_transformation_profiles(
  SourceFormat,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{SourceFormat}{[required] Filters the results by source data format.}

\item{MaxResults}{The maximum number of profiles to return per page. If you don't specify a value, the service returns up to 100 results.}

\item{NextToken}{The pagination token from a previous response. Pass this value to retrieve the next page of results.}
}
\description{
Lists all data transformation profiles in your account, returning the latest version summary for each. Use \code{\link[=healthlake_get_data_transformation_profile]{get_data_transformation_profile}} to retrieve profile content. Results are paginated. Use the \code{NextToken} parameter to retrieve additional results.

See \url{https://www.paws-r-sdk.com/docs/healthlake_list_data_transformation_profiles/} for full documentation.
}
\keyword{internal}
