% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthlake_operations.R
\name{healthlake_list_fhir_datastores}
\alias{healthlake_list_fhir_datastores}
\title{Lists all FHIR Data Stores that are in the user’s account, regardless of
Data Store status}
\usage{
healthlake_list_fhir_datastores(
  Filter = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{Filter}{Lists all filters associated with a FHIR Data Store request.}

\item{NextToken}{Fetches the next page of Data Stores when results are paginated.}

\item{MaxResults}{The maximum number of Data Stores returned in a single page of a
ListFHIRDatastoresRequest call.}
}
\description{
Lists all FHIR Data Stores that are in the user’s account, regardless of Data Store status.

See \url{https://paws-r.github.io/docs/healthlake/list_fhir_datastores.html} for full documentation.
}
\keyword{internal}
