% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthlake_operations.R
\name{healthlake_list_fhir_export_jobs}
\alias{healthlake_list_fhir_export_jobs}
\title{Lists all FHIR export jobs associated with an account and their statuses}
\usage{
healthlake_list_fhir_export_jobs(
  DatastoreId,
  NextToken = NULL,
  MaxResults = NULL,
  JobName = NULL,
  JobStatus = NULL,
  SubmittedBefore = NULL,
  SubmittedAfter = NULL
)
}
\arguments{
\item{DatastoreId}{[required] This parameter limits the response to the export job with the specified
Data Store ID.}

\item{NextToken}{A pagination token used to identify the next page of results to return
for a ListFHIRExportJobs query.}

\item{MaxResults}{This parameter limits the number of results returned for a
ListFHIRExportJobs to a maximum quantity specified by the user.}

\item{JobName}{This parameter limits the response to the export job with the specified
job name.}

\item{JobStatus}{This parameter limits the response to the export jobs with the specified
job status.}

\item{SubmittedBefore}{This parameter limits the response to FHIR export jobs submitted before
a user specified date.}

\item{SubmittedAfter}{This parameter limits the response to FHIR export jobs submitted after a
user specified date.}
}
\description{
Lists all FHIR export jobs associated with an account and their statuses.

See \url{https://paws-r.github.io/docs/healthlake/list_fhir_export_jobs.html} for full documentation.
}
\keyword{internal}
