% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthlake_operations.R
\name{healthlake_publish_data_transformation_profile}
\alias{healthlake_publish_data_transformation_profile}
\title{Promotes the current DRAFT version of a data transformation profile to a
new immutable published version}
\usage{
healthlake_publish_data_transformation_profile(
  ProfileId,
  SourceFormat,
  FromExistingVersion = NULL,
  ChangeDescription = NULL
)
}
\arguments{
\item{ProfileId}{[required] The unique identifier of the profile to publish.}

\item{SourceFormat}{[required] The source data format of the profile.}

\item{FromExistingVersion}{The version number of a previously published version to republish as the new latest version. Use this parameter for rollback scenarios. If you omit this parameter, the service publishes the current DRAFT version.}

\item{ChangeDescription}{A description of what changed or why this version is being published.}
}
\description{
Promotes the current DRAFT version of a data transformation profile to a new immutable published version. Also supports rollback by publishing from a previously published version.

See \url{https://www.paws-r-sdk.com/docs/healthlake_publish_data_transformation_profile/} for full documentation.
}
\keyword{internal}
