% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthlake_operations.R
\name{healthlake_start_data_transformation_job}
\alias{healthlake_start_data_transformation_job}
\title{Starts an asynchronous data transformation job that converts source
files from Amazon Simple Storage Service (Amazon S3) and writes the
output to Amazon S3 or HealthLake}
\usage{
healthlake_start_data_transformation_job(
  InputDataConfig,
  OutputDataConfig,
  DataAccessRoleArn,
  ClientToken,
  JobName = NULL,
  ProfileId,
  DriftDetectionEnabled = NULL,
  ProvenanceEnabled = NULL
)
}
\arguments{
\item{InputDataConfig}{[required] The Amazon S3 location and format of the source files to transform.}

\item{OutputDataConfig}{[required] The Amazon S3 output location and Amazon Web Services Key Management Service (Amazon Web Services KMS) encryption configuration.}

\item{DataAccessRoleArn}{[required] The Amazon Resource Name (ARN) of the Amazon Web Services Identity and Access Management (IAM) role that HealthLake assumes to read from and write to the specified Amazon S3 locations.}

\item{ClientToken}{[required] A unique, case-sensitive identifier to ensure that the operation completes no more than one time. If this token matches a previous request, the service ignores the request but does not return an error.}

\item{JobName}{A descriptive name for the data transformation job.}

\item{ProfileId}{[required] The unique identifier of the data transformation profile to use for conversion.}

\item{DriftDetectionEnabled}{Specifies whether drift detection is enabled for this job. When enabled, HealthLake writes a drift report to the output Amazon S3 location alongside the converted files.}

\item{ProvenanceEnabled}{Specifies whether FHIR R4 Provenance resource generation is enabled for this transformation job. When provenance is enabled, the service also generates related DocumentReference and Device resources. If you don't specify a value, the default is \code{true}. To disable provenance output, set this parameter to \code{false}.}
}
\description{
Starts an asynchronous data transformation job that converts source files from Amazon Simple Storage Service (Amazon S3) and writes the output to Amazon S3 or HealthLake.

See \url{https://www.paws-r-sdk.com/docs/healthlake_start_data_transformation_job/} for full documentation.
}
\keyword{internal}
