% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthlake_operations.R
\name{healthlake_start_fhir_export_job}
\alias{healthlake_start_fhir_export_job}
\title{Begins a FHIR export job}
\usage{
healthlake_start_fhir_export_job(
  JobName = NULL,
  OutputDataConfig,
  DatastoreId,
  DataAccessRoleArn,
  ClientToken = NULL
)
}
\arguments{
\item{JobName}{The user generated name for an export job.}

\item{OutputDataConfig}{[required] The output data configuration that was supplied when the export job was
created.}

\item{DatastoreId}{[required] The AWS generated ID for the data store from which files are being
exported for an export job.}

\item{DataAccessRoleArn}{[required] The Amazon Resource Name used during the initiation of the job.}

\item{ClientToken}{An optional user provided token used for ensuring idempotency.}
}
\description{
Begins a FHIR export job.

See \url{https://www.paws-r-sdk.com/docs/healthlake_start_fhir_export_job/} for full documentation.
}
\keyword{internal}
