% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthlake_operations.R
\name{healthlake_start_fhir_import_job}
\alias{healthlake_start_fhir_import_job}
\title{Start importing bulk FHIR data into an ACTIVE data store}
\usage{
healthlake_start_fhir_import_job(
  JobName = NULL,
  InputDataConfig,
  JobOutputDataConfig,
  DatastoreId,
  DataAccessRoleArn,
  ClientToken = NULL,
  ValidationLevel = NULL,
  ProfileId = NULL,
  InputFormat = NULL,
  DriftDetectionEnabled = NULL,
  ProvenanceEnabled = NULL
)
}
\arguments{
\item{JobName}{The import job name.}

\item{InputDataConfig}{[required] The input properties for the import job request.}

\item{JobOutputDataConfig}{[required] The output data configuration supplied when the export job was created.}

\item{DatastoreId}{[required] The data store identifier.}

\item{DataAccessRoleArn}{[required] The Amazon Resource Name (ARN) that grants access permission to HealthLake.}

\item{ClientToken}{The optional user-provided token used for ensuring API idempotency.}

\item{ValidationLevel}{The validation level of the import job.}

\item{ProfileId}{The data transformation profile identifier to use for the import job.}

\item{InputFormat}{The input format of the data to be imported.}

\item{DriftDetectionEnabled}{Specifies whether to enable drift detection for the import job.}

\item{ProvenanceEnabled}{Specifies whether to enable provenance for the import job.}
}
\description{
Start importing bulk FHIR data into an ACTIVE data store. The import job imports FHIR data found in the \code{InputDataConfig} object and stores processing results in the \code{JobOutputDataConfig} object.

See \url{https://www.paws-r-sdk.com/docs/healthlake_start_fhir_import_job/} for full documentation.
}
\keyword{internal}
