% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthlake_operations.R
\name{healthlake_untag_resource}
\alias{healthlake_untag_resource}
\title{Removes tags from a Data Store}
\usage{
healthlake_untag_resource(ResourceARN, TagKeys)
}
\arguments{
\item{ResourceARN}{[required] "The Amazon Resource Name(ARN) of the Data Store for which tags are
being removed}

\item{TagKeys}{[required] The keys for the tags to be removed from the Healthlake Data Store.}
}
\description{
Removes tags from a Data Store.

See \url{https://www.paws-r-sdk.com/docs/healthlake_untag_resource/} for full documentation.
}
\keyword{internal}
