% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthlake_operations.R
\name{healthlake_update_data_transformation_profile}
\alias{healthlake_update_data_transformation_profile}
\title{Updates the DRAFT version (version 0) of a data transformation profile
with new profile content}
\usage{
healthlake_update_data_transformation_profile(
  ProfileId,
  ProfileMapping,
  ChangeDescription = NULL
)
}
\arguments{
\item{ProfileId}{[required] The unique identifier of the profile to update.}

\item{ProfileMapping}{[required] The new profile content for the DRAFT version. This is a full replacement of all profile files.}

\item{ChangeDescription}{A description of what changed in this update.}
}
\description{
Updates the DRAFT version (version 0) of a data transformation profile with new profile content. The update replaces all existing DRAFT content.

See \url{https://www.paws-r-sdk.com/docs/healthlake_update_data_transformation_profile/} for full documentation.
}
\keyword{internal}
