% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthlake_operations.R
\name{healthlake_update_fhir_datastore}
\alias{healthlake_update_fhir_datastore}
\title{Update the properties of a FHIR-enabled data store}
\usage{
healthlake_update_fhir_datastore(
  DatastoreId,
  DatastoreName = NULL,
  AnalyticsConfiguration = NULL,
  NlpConfiguration = NULL,
  ProfileConfiguration = NULL,
  IdentityProviderConfiguration = NULL,
  BackupConfiguration = NULL
)
}
\arguments{
\item{DatastoreId}{[required] The data store identifier.}

\item{DatastoreName}{The data store name.}

\item{AnalyticsConfiguration}{The analytics configuration for the data store.}

\item{NlpConfiguration}{The natural language processing (NLP) configuration for the data store.}

\item{ProfileConfiguration}{The profile configuration for the data store.}

\item{IdentityProviderConfiguration}{The identity provider configuration for the data store.}

\item{BackupConfiguration}{The backup configuration for the data store.}
}
\description{
Update the properties of a FHIR-enabled data store.

See \url{https://www.paws-r-sdk.com/docs/healthlake_update_fhir_datastore/} for full documentation.
}
\keyword{internal}
