% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivs_operations.R
\name{ivs_batch_get_stream_key}
\alias{ivs_batch_get_stream_key}
\title{Performs GetStreamKey on multiple ARNs simultaneously}
\usage{
ivs_batch_get_stream_key(arns)
}
\arguments{
\item{arns}{[required] Array of ARNs, one per stream key.}
}
\description{
Performs \code{\link[=ivs_get_stream_key]{get_stream_key}} on multiple ARNs simultaneously.

See \url{https://www.paws-r-sdk.com/docs/ivs_batch_get_stream_key/} for full documentation.
}
\keyword{internal}
