% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivs_operations.R
\name{ivs_batch_start_viewer_session_revocation}
\alias{ivs_batch_start_viewer_session_revocation}
\title{Performs StartViewerSessionRevocation on multiple channel ARN and viewer
ID pairs simultaneously}
\usage{
ivs_batch_start_viewer_session_revocation(viewerSessions)
}
\arguments{
\item{viewerSessions}{[required] Array of viewer sessions, one per channel-ARN and viewer-ID pair.}
}
\description{
Performs \code{\link[=ivs_start_viewer_session_revocation]{start_viewer_session_revocation}} on multiple channel ARN and viewer ID pairs simultaneously.

See \url{https://www.paws-r-sdk.com/docs/ivs_batch_start_viewer_session_revocation/} for full documentation.
}
\keyword{internal}
