% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivs_operations.R
\name{ivs_delete_ad_configuration}
\alias{ivs_delete_ad_configuration}
\title{Deletes the specified ad configuration}
\usage{
ivs_delete_ad_configuration(arn)
}
\arguments{
\item{arn}{[required] ARN of the ad configuration to be deleted.}
}
\description{
Deletes the specified ad configuration.

See \url{https://www.paws-r-sdk.com/docs/ivs_delete_ad_configuration/} for full documentation.
}
\keyword{internal}
