% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivs_operations.R
\name{ivs_delete_recording_configuration}
\alias{ivs_delete_recording_configuration}
\title{Deletes the recording configuration for the specified ARN}
\usage{
ivs_delete_recording_configuration(arn)
}
\arguments{
\item{arn}{[required] ARN of the recording configuration to be deleted.}
}
\description{
Deletes the recording configuration for the specified ARN.

See \url{https://www.paws-r-sdk.com/docs/ivs_delete_recording_configuration/} for full documentation.
}
\keyword{internal}
