% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivs_operations.R
\name{ivs_delete_stream_key}
\alias{ivs_delete_stream_key}
\title{Deletes the stream key for the specified ARN, so it can no longer be
used to stream}
\usage{
ivs_delete_stream_key(arn)
}
\arguments{
\item{arn}{[required] ARN of the stream key to be deleted.}
}
\description{
Deletes the stream key for the specified ARN, so it can no longer be used to stream.

See \url{https://paws-r.github.io/docs/ivs/delete_stream_key.html} for full documentation.
}
\keyword{internal}
