% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivs_operations.R
\name{ivs_get_ad_configuration}
\alias{ivs_get_ad_configuration}
\title{Gets the ad configuration represented by the specified ARN}
\usage{
ivs_get_ad_configuration(arn)
}
\arguments{
\item{arn}{[required] ARN of the ad configuration to be retrieved.}
}
\description{
Gets the ad configuration represented by the specified ARN.

See \url{https://www.paws-r-sdk.com/docs/ivs_get_ad_configuration/} for full documentation.
}
\keyword{internal}
