% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivs_operations.R
\name{ivs_get_channel}
\alias{ivs_get_channel}
\title{Gets the channel configuration for the specified channel ARN}
\usage{
ivs_get_channel(arn)
}
\arguments{
\item{arn}{[required] ARN of the channel for which the configuration is to be retrieved.}
}
\description{
Gets the channel configuration for the specified channel ARN. See also \code{\link[=ivs_batch_get_channel]{batch_get_channel}}.

See \url{https://www.paws-r-sdk.com/docs/ivs_get_channel/} for full documentation.
}
\keyword{internal}
