% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivs_operations.R
\name{ivs_get_playback_restriction_policy}
\alias{ivs_get_playback_restriction_policy}
\title{Gets the specified playback restriction policy}
\usage{
ivs_get_playback_restriction_policy(arn)
}
\arguments{
\item{arn}{[required] ARN of the playback restriction policy to be returned.}
}
\description{
Gets the specified playback restriction policy.

See \url{https://www.paws-r-sdk.com/docs/ivs_get_playback_restriction_policy/} for full documentation.
}
\keyword{internal}
