% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivs_operations.R
\name{ivs_get_stream_session}
\alias{ivs_get_stream_session}
\title{Gets metadata on a specified stream}
\usage{
ivs_get_stream_session(channelArn, streamId = NULL)
}
\arguments{
\item{channelArn}{[required] ARN of the channel resource}

\item{streamId}{Unique identifier for a live or previously live stream in the specified
channel. If no \code{streamId} is provided, this returns the most recent
stream session for the channel, if it exists.}
}
\description{
Gets metadata on a specified stream.

See \url{https://paws-r.github.io/docs/ivs/get_stream_session.html} for full documentation.
}
\keyword{internal}
