% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivs_operations.R
\name{ivs_list_playback_key_pairs}
\alias{ivs_list_playback_key_pairs}
\title{Gets summary information about playback key pairs}
\usage{
ivs_list_playback_key_pairs(maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{maxResults}{The first key pair to retrieve. This is used for pagination; see the
\code{nextToken} response field. Default: 50.}

\item{nextToken}{Maximum number of key pairs to return.}
}
\description{
Gets summary information about playback key pairs. For more information, see \href{https://docs.aws.amazon.com/ivs/latest/userguide/private-channels.html}{Setting Up Private Channels} in the \emph{Amazon IVS User Guide}.

See \url{https://paws-r.github.io/docs/ivs/list_playback_key_pairs.html} for full documentation.
}
\keyword{internal}
