% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivs_operations.R
\name{ivs_list_playback_restriction_policies}
\alias{ivs_list_playback_restriction_policies}
\title{Gets summary information about playback restriction policies}
\usage{
ivs_list_playback_restriction_policies(nextToken = NULL, maxResults = NULL)
}
\arguments{
\item{nextToken}{The first policy to retrieve. This is used for pagination; see the
\code{nextToken} response field.}

\item{maxResults}{Maximum number of policies to return. Default: 1.}
}
\description{
Gets summary information about playback restriction policies.

See \url{https://www.paws-r-sdk.com/docs/ivs_list_playback_restriction_policies/} for full documentation.
}
\keyword{internal}
