% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivs_operations.R
\name{ivs_list_stream_keys}
\alias{ivs_list_stream_keys}
\title{Gets summary information about stream keys for the specified channel}
\usage{
ivs_list_stream_keys(channelArn, maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{channelArn}{[required] Channel ARN used to filter the list.}

\item{maxResults}{Maximum number of streamKeys to return. Default: 1.}

\item{nextToken}{The first stream key to retrieve. This is used for pagination; see the
\code{nextToken} response field.}
}
\description{
Gets summary information about stream keys for the specified channel.

See \url{https://paws-r.github.io/docs/ivs/list_stream_keys.html} for full documentation.
}
\keyword{internal}
