% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivs_operations.R
\name{ivs_list_streams}
\alias{ivs_list_streams}
\title{Gets summary information about live streams in your account, in the
Amazon Web Services region where the API request is processed}
\usage{
ivs_list_streams(filterBy = NULL, nextToken = NULL, maxResults = NULL)
}
\arguments{
\item{filterBy}{Filters the stream list to match the specified criterion.}

\item{nextToken}{The first stream to retrieve. This is used for pagination; see the
\code{nextToken} response field.}

\item{maxResults}{Maximum number of streams to return. Default: 100.}
}
\description{
Gets summary information about live streams in your account, in the Amazon Web Services region where the API request is processed.

See \url{https://www.paws-r-sdk.com/docs/ivs_list_streams/} for full documentation.
}
\keyword{internal}
