% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivs_operations.R
\name{ivs_list_tags_for_resource}
\alias{ivs_list_tags_for_resource}
\title{Gets information about Amazon Web Services tags for the specified ARN}
\usage{
ivs_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The ARN of the resource to be retrieved. The ARN must be URL-encoded.}
}
\description{
Gets information about Amazon Web Services tags for the specified ARN.

See \url{https://www.paws-r-sdk.com/docs/ivs_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
