% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivs_operations.R
\name{ivs_put_metadata}
\alias{ivs_put_metadata}
\title{Inserts metadata into the active stream of the specified channel}
\usage{
ivs_put_metadata(channelArn, metadata)
}
\arguments{
\item{channelArn}{[required] ARN of the channel into which metadata is inserted. This channel must
have an active stream.}

\item{metadata}{[required] Metadata to insert into the stream. Maximum: 1 KB per request.}
}
\description{
Inserts metadata into the active stream of the specified channel. At most 5 requests per second per channel are allowed, each with a maximum 1 KB payload. (If 5 TPS is not sufficient for your needs, we recommend batching your data into a single PutMetadata call.) At most 155 requests per second per account are allowed. Also see \href{https://docs.aws.amazon.com/ivs/latest/userguide/metadata.html}{Embedding Metadata within a Video Stream} in the \emph{Amazon IVS User Guide}.

See \url{https://paws-r.github.io/docs/ivs/put_metadata.html} for full documentation.
}
\keyword{internal}
