% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivs_operations.R
\name{ivs_tag_resource}
\alias{ivs_tag_resource}
\title{Adds or updates tags for the Amazon Web Services resource with the
specified ARN}
\usage{
ivs_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] ARN of the resource for which tags are to be added or updated.}

\item{tags}{[required] Array of tags to be added or updated.}
}
\description{
Adds or updates tags for the Amazon Web Services resource with the specified ARN.

See \url{https://paws-r.github.io/docs/ivs/tag_resource.html} for full documentation.
}
\keyword{internal}
