% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivs_operations.R
\name{ivs_update_ad_configuration}
\alias{ivs_update_ad_configuration}
\title{Updates a specified ad configuration}
\usage{
ivs_update_ad_configuration(
  arn,
  name = NULL,
  mediaTailorPlaybackConfigurations = NULL,
  postRollConfiguration = NULL
)
}
\arguments{
\item{arn}{[required] ARN of the ad configuration to be updated.}

\item{name}{Ad configuration name. The value does not need to be unique.}

\item{mediaTailorPlaybackConfigurations}{List of integration configurations with MediaTailor resources. The first item in the list is the default playback configuration used for the ad configuration. To select a different configuration per viewing session, see \href{https://docs.aws.amazon.com/ivs/latest/LowLatencyUserGuide/private-channels-generate-tokens.html}{Generate and Sign IVS Playback Tokens}.}

\item{postRollConfiguration}{Configuration for the post-roll ad break to use for this ad configuration.}
}
\description{
Updates a specified ad configuration.

See \url{https://www.paws-r-sdk.com/docs/ivs_update_ad_configuration/} for full documentation.
}
\keyword{internal}
