% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivsrealtime_operations.R
\name{ivsrealtime_create_stage}
\alias{ivsrealtime_create_stage}
\title{Creates a new stage (and optionally participant tokens)}
\usage{
ivsrealtime_create_stage(
  name = NULL,
  participantTokenConfigurations = NULL,
  tags = NULL,
  autoParticipantRecordingConfiguration = NULL
)
}
\arguments{
\item{name}{Optional name that can be specified for the stage being created.}

\item{participantTokenConfigurations}{Array of participant token configuration objects to attach to the new
stage.}

\item{tags}{Tags attached to the resource. Array of maps, each of the form
\code{string:string (key:value)}. See \href{https://docs.aws.amazon.com/tag-editor/latest/userguide/best-practices-and-strats.html}{Best practices and strategies}
in \emph{Tagging AWS Resources and Tag Editor} for details, including
restrictions that apply to tags and "Tag naming limits and
requirements"; Amazon IVS has no constraints on tags beyond what is
documented there.}

\item{autoParticipantRecordingConfiguration}{Configuration object for individual participant recording, to attach to
the new stage.}
}
\description{
Creates a new stage (and optionally participant tokens).

See \url{https://www.paws-r-sdk.com/docs/ivsrealtime_create_stage/} for full documentation.
}
\keyword{internal}
