% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivsrealtime_operations.R
\name{ivsrealtime_delete_encoder_configuration}
\alias{ivsrealtime_delete_encoder_configuration}
\title{Deletes an EncoderConfiguration resource}
\usage{
ivsrealtime_delete_encoder_configuration(arn)
}
\arguments{
\item{arn}{[required] ARN of the EncoderConfiguration.}
}
\description{
Deletes an EncoderConfiguration resource. Ensures that no Compositions are using this template; otherwise, returns an error.

See \url{https://www.paws-r-sdk.com/docs/ivsrealtime_delete_encoder_configuration/} for full documentation.
}
\keyword{internal}
