% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivsrealtime_operations.R
\name{ivsrealtime_delete_ingest_configuration}
\alias{ivsrealtime_delete_ingest_configuration}
\title{Deletes a specified IngestConfiguration, so it can no longer be used to
broadcast}
\usage{
ivsrealtime_delete_ingest_configuration(arn, force = NULL)
}
\arguments{
\item{arn}{[required] ARN of the IngestConfiguration.}

\item{force}{Optional field to force deletion of the IngestConfiguration. If this is
set to \code{true} when a participant is actively publishing, the participant
is disconnected from the stage, followed by deletion of the
IngestConfiguration. Default: \code{false}.}
}
\description{
Deletes a specified IngestConfiguration, so it can no longer be used to broadcast. An IngestConfiguration cannot be deleted if the publisher is actively streaming to a stage, unless \code{force} is set to \code{true}.

See \url{https://www.paws-r-sdk.com/docs/ivsrealtime_delete_ingest_configuration/} for full documentation.
}
\keyword{internal}
