% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivsrealtime_operations.R
\name{ivsrealtime_delete_public_key}
\alias{ivsrealtime_delete_public_key}
\title{Deletes the specified public key used to sign stage participant tokens}
\usage{
ivsrealtime_delete_public_key(arn)
}
\arguments{
\item{arn}{[required] ARN of the public key to be deleted.}
}
\description{
Deletes the specified public key used to sign stage participant tokens. This invalidates future participant tokens generated using the key pair’s private key.

See \url{https://www.paws-r-sdk.com/docs/ivsrealtime_delete_public_key/} for full documentation.
}
\keyword{internal}
