% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivsrealtime_operations.R
\name{ivsrealtime_delete_stage}
\alias{ivsrealtime_delete_stage}
\title{Shuts down and deletes the specified stage (disconnecting all
participants)}
\usage{
ivsrealtime_delete_stage(arn)
}
\arguments{
\item{arn}{[required] ARN of the stage to be deleted.}
}
\description{
Shuts down and deletes the specified stage (disconnecting all participants). This operation also removes the \code{stageArn} from the associated IngestConfiguration, if there are participants using the IngestConfiguration to publish to the stage.

See \url{https://www.paws-r-sdk.com/docs/ivsrealtime_delete_stage/} for full documentation.
}
\keyword{internal}
