% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivsrealtime_operations.R
\name{ivsrealtime_get_composition}
\alias{ivsrealtime_get_composition}
\title{Get information about the specified Composition resource}
\usage{
ivsrealtime_get_composition(arn)
}
\arguments{
\item{arn}{[required] ARN of the Composition resource.}
}
\description{
Get information about the specified Composition resource.

See \url{https://www.paws-r-sdk.com/docs/ivsrealtime_get_composition/} for full documentation.
}
\keyword{internal}
