% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivsrealtime_operations.R
\name{ivsrealtime_get_encoder_configuration}
\alias{ivsrealtime_get_encoder_configuration}
\title{Gets information about the specified EncoderConfiguration resource}
\usage{
ivsrealtime_get_encoder_configuration(arn)
}
\arguments{
\item{arn}{[required] ARN of the EncoderConfiguration resource.}
}
\description{
Gets information about the specified EncoderConfiguration resource.

See \url{https://www.paws-r-sdk.com/docs/ivsrealtime_get_encoder_configuration/} for full documentation.
}
\keyword{internal}
