% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivsrealtime_operations.R
\name{ivsrealtime_get_ingest_configuration}
\alias{ivsrealtime_get_ingest_configuration}
\title{Gets information about the specified IngestConfiguration}
\usage{
ivsrealtime_get_ingest_configuration(arn)
}
\arguments{
\item{arn}{[required] ARN of the ingest for which the information is to be retrieved.}
}
\description{
Gets information about the specified IngestConfiguration.

See \url{https://www.paws-r-sdk.com/docs/ivsrealtime_get_ingest_configuration/} for full documentation.
}
\keyword{internal}
