% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivsrealtime_operations.R
\name{ivsrealtime_get_participant}
\alias{ivsrealtime_get_participant}
\title{Gets information about the specified participant token}
\usage{
ivsrealtime_get_participant(participantId, sessionId, stageArn)
}
\arguments{
\item{participantId}{[required] Unique identifier for the participant. This is assigned by IVS and
returned by
\code{\link[=ivsrealtime_create_participant_token]{create_participant_token}}.}

\item{sessionId}{[required] ID of a session within the stage.}

\item{stageArn}{[required] Stage ARN.}
}
\description{
Gets information about the specified participant token.

See \url{https://www.paws-r-sdk.com/docs/ivsrealtime_get_participant/} for full documentation.
}
\keyword{internal}
