% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivsrealtime_operations.R
\name{ivsrealtime_get_public_key}
\alias{ivsrealtime_get_public_key}
\title{Gets information for the specified public key}
\usage{
ivsrealtime_get_public_key(arn)
}
\arguments{
\item{arn}{[required] ARN of the public key for which the information is to be retrieved.}
}
\description{
Gets information for the specified public key.

See \url{https://www.paws-r-sdk.com/docs/ivsrealtime_get_public_key/} for full documentation.
}
\keyword{internal}
