% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivsrealtime_operations.R
\name{ivsrealtime_get_stage}
\alias{ivsrealtime_get_stage}
\title{Gets information for the specified stage}
\usage{
ivsrealtime_get_stage(arn)
}
\arguments{
\item{arn}{[required] ARN of the stage for which the information is to be retrieved.}
}
\description{
Gets information for the specified stage.

See \url{https://www.paws-r-sdk.com/docs/ivsrealtime_get_stage/} for full documentation.
}
\keyword{internal}
