% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivsrealtime_operations.R
\name{ivsrealtime_get_stage_session}
\alias{ivsrealtime_get_stage_session}
\title{Gets information for the specified stage session}
\usage{
ivsrealtime_get_stage_session(stageArn, sessionId)
}
\arguments{
\item{stageArn}{[required] ARN of the stage for which the information is to be retrieved.}

\item{sessionId}{[required] ID of a session within the stage.}
}
\description{
Gets information for the specified stage session.

See \url{https://www.paws-r-sdk.com/docs/ivsrealtime_get_stage_session/} for full documentation.
}
\keyword{internal}
