% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivsrealtime_operations.R
\name{ivsrealtime_get_storage_configuration}
\alias{ivsrealtime_get_storage_configuration}
\title{Gets the storage configuration for the specified ARN}
\usage{
ivsrealtime_get_storage_configuration(arn)
}
\arguments{
\item{arn}{[required] ARN of the storage configuration to be retrieved.}
}
\description{
Gets the storage configuration for the specified ARN.

See \url{https://www.paws-r-sdk.com/docs/ivsrealtime_get_storage_configuration/} for full documentation.
}
\keyword{internal}
