% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivsrealtime_operations.R
\name{ivsrealtime_list_ingest_configurations}
\alias{ivsrealtime_list_ingest_configurations}
\title{Lists all IngestConfigurations in your account, in the AWS region where
the API request is processed}
\usage{
ivsrealtime_list_ingest_configurations(
  filterByStageArn = NULL,
  filterByState = NULL,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{filterByStageArn}{Filters the response list to match the specified stage ARN. Only one
filter (by stage ARN or by state) can be used at a time.}

\item{filterByState}{Filters the response list to match the specified state. Only one filter
(by stage ARN or by state) can be used at a time.}

\item{nextToken}{The first IngestConfiguration to retrieve. This is used for pagination;
see the \code{nextToken} response field.}

\item{maxResults}{Maximum number of results to return. Default: 50.}
}
\description{
Lists all IngestConfigurations in your account, in the AWS region where the API request is processed.

See \url{https://www.paws-r-sdk.com/docs/ivsrealtime_list_ingest_configurations/} for full documentation.
}
\keyword{internal}
