% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivsrealtime_operations.R
\name{ivsrealtime_list_participant_replicas}
\alias{ivsrealtime_list_participant_replicas}
\title{Lists all the replicas for a participant from a source stage}
\usage{
ivsrealtime_list_participant_replicas(
  sourceStageArn,
  participantId,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{sourceStageArn}{[required] ARN of the stage where the participant is publishing.}

\item{participantId}{[required] Participant ID of the publisher that has been replicated. This is
assigned by IVS and returned by
\code{\link[=ivsrealtime_create_participant_token]{create_participant_token}} or
the \code{jti} (JWT ID) used to \href{https://docs.aws.amazon.com/ivs/latest/RealTimeUserGuide/getting-started-distribute-tokens.html#getting-started-distribute-tokens-self-signed}{create a self signed token}.}

\item{nextToken}{The first participant to retrieve. This is used for pagination; see the
\code{nextToken} response field.}

\item{maxResults}{Maximum number of results to return. Default: 50.}
}
\description{
Lists all the replicas for a participant from a source stage.

See \url{https://www.paws-r-sdk.com/docs/ivsrealtime_list_participant_replicas/} for full documentation.
}
\keyword{internal}
