% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivsrealtime_operations.R
\name{ivsrealtime_list_public_keys}
\alias{ivsrealtime_list_public_keys}
\title{Gets summary information about all public keys in your account, in the
AWS region where the API request is processed}
\usage{
ivsrealtime_list_public_keys(nextToken = NULL, maxResults = NULL)
}
\arguments{
\item{nextToken}{The first public key to retrieve. This is used for pagination; see the
\code{nextToken} response field.}

\item{maxResults}{Maximum number of results to return. Default: 50.}
}
\description{
Gets summary information about all public keys in your account, in the AWS region where the API request is processed.

See \url{https://www.paws-r-sdk.com/docs/ivsrealtime_list_public_keys/} for full documentation.
}
\keyword{internal}
