% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivsrealtime_operations.R
\name{ivsrealtime_list_stage_sessions}
\alias{ivsrealtime_list_stage_sessions}
\title{Gets all sessions for a specified stage}
\usage{
ivsrealtime_list_stage_sessions(maxResults = NULL, nextToken = NULL, stageArn)
}
\arguments{
\item{maxResults}{Maximum number of results to return. Default: 50.}

\item{nextToken}{The first stage session to retrieve. This is used for pagination; see
the \code{nextToken} response field.}

\item{stageArn}{[required] Stage ARN.}
}
\description{
Gets all sessions for a specified stage.

See \url{https://www.paws-r-sdk.com/docs/ivsrealtime_list_stage_sessions/} for full documentation.
}
\keyword{internal}
