% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivsrealtime_operations.R
\name{ivsrealtime_list_storage_configurations}
\alias{ivsrealtime_list_storage_configurations}
\title{Gets summary information about all storage configurations in your
account, in the AWS region where the API request is processed}
\usage{
ivsrealtime_list_storage_configurations(maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{maxResults}{Maximum number of storage configurations to return. Default: your
service quota or 100, whichever is smaller.}

\item{nextToken}{The first storage configuration to retrieve. This is used for
pagination; see the \code{nextToken} response field.}
}
\description{
Gets summary information about all storage configurations in your account, in the AWS region where the API request is processed.

See \url{https://www.paws-r-sdk.com/docs/ivsrealtime_list_storage_configurations/} for full documentation.
}
\keyword{internal}
