% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivsrealtime_operations.R
\name{ivsrealtime_start_composition}
\alias{ivsrealtime_start_composition}
\title{Starts a Composition from a stage based on the configuration provided in
the request}
\usage{
ivsrealtime_start_composition(
  destinations,
  idempotencyToken = NULL,
  layout = NULL,
  stageArn,
  tags = NULL
)
}
\arguments{
\item{destinations}{[required] Array of destination configuration.}

\item{idempotencyToken}{Idempotency token.}

\item{layout}{Layout object to configure composition parameters.}

\item{stageArn}{[required] ARN of the stage to be used for compositing.}

\item{tags}{Tags attached to the resource. Array of maps, each of the form
\code{string:string (key:value)}. See \href{https://docs.aws.amazon.com/tag-editor/latest/userguide/tagging.html}{Tagging AWS Resources}
for details, including restrictions that apply to tags and "Tag naming
limits and requirements"; Amazon IVS has no constraints on tags beyond
what is documented there.}
}
\description{
Starts a Composition from a stage based on the configuration provided in the request.

See \url{https://www.paws-r-sdk.com/docs/ivsrealtime_start_composition/} for full documentation.
}
\keyword{internal}
