% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivsrealtime_operations.R
\name{ivsrealtime_stop_composition}
\alias{ivsrealtime_stop_composition}
\title{Stops and deletes a Composition resource}
\usage{
ivsrealtime_stop_composition(arn)
}
\arguments{
\item{arn}{[required] ARN of the Composition.}
}
\description{
Stops and deletes a Composition resource. Any broadcast from the Composition resource is stopped.

See \url{https://www.paws-r-sdk.com/docs/ivsrealtime_stop_composition/} for full documentation.
}
\keyword{internal}
