% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivsrealtime_operations.R
\name{ivsrealtime_stop_participant_replication}
\alias{ivsrealtime_stop_participant_replication}
\title{Stops a replicated participant session}
\usage{
ivsrealtime_stop_participant_replication(
  sourceStageArn,
  destinationStageArn,
  participantId
)
}
\arguments{
\item{sourceStageArn}{[required] ARN of the stage where the participant is publishing.}

\item{destinationStageArn}{[required] ARN of the stage where the participant has been replicated.}

\item{participantId}{[required] Participant ID of the publisher that has been replicated. This is
assigned by IVS and returned by
\code{\link[=ivsrealtime_create_participant_token]{create_participant_token}} or
the \code{jti} (JWT ID) used to \href{https://docs.aws.amazon.com/ivs/latest/RealTimeUserGuide/getting-started-distribute-tokens.html#getting-started-distribute-tokens-self-signed}{create a self signed token}.}
}
\description{
Stops a replicated participant session.

See \url{https://www.paws-r-sdk.com/docs/ivsrealtime_stop_participant_replication/} for full documentation.
}
\keyword{internal}
