% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivsrealtime_operations.R
\name{ivsrealtime_untag_resource}
\alias{ivsrealtime_untag_resource}
\title{Removes tags from the resource with the specified ARN}
\usage{
ivsrealtime_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The ARN of the resource to be untagged. The ARN must be URL-encoded.}

\item{tagKeys}{[required] Array of tag keys (strings) for the tags to be removed. See \href{https://docs.aws.amazon.com/tag-editor/latest/userguide/best-practices-and-strats.html}{Best practices and strategies}
in \emph{Tagging AWS Resources and Tag Editor} for details, including
restrictions that apply to tags and "Tag naming limits and
requirements"; Amazon IVS has no constraints on tags beyond what is
documented there.}
}
\description{
Removes tags from the resource with the specified ARN.

See \url{https://www.paws-r-sdk.com/docs/ivsrealtime_untag_resource/} for full documentation.
}
\keyword{internal}
