% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivsrealtime_operations.R
\name{ivsrealtime_update_ingest_configuration}
\alias{ivsrealtime_update_ingest_configuration}
\title{Updates a specified IngestConfiguration}
\usage{
ivsrealtime_update_ingest_configuration(arn, stageArn = NULL)
}
\arguments{
\item{arn}{[required] ARN of the IngestConfiguration, for which the related stage ARN needs to
be updated.}

\item{stageArn}{Stage ARN that needs to be updated.}
}
\description{
Updates a specified IngestConfiguration. Only the stage ARN attached to the IngestConfiguration can be updated. An IngestConfiguration that is active cannot be updated.

See \url{https://www.paws-r-sdk.com/docs/ivsrealtime_update_ingest_configuration/} for full documentation.
}
\keyword{internal}
