% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_batch_associate_scram_secret}
\alias{kafka_batch_associate_scram_secret}
\title{Associates one or more Scram Secrets with an Amazon MSK cluster}
\usage{
kafka_batch_associate_scram_secret(ClusterArn, SecretArnList)
}
\arguments{
\item{ClusterArn}{[required]
<p>The Amazon Resource Name (ARN) of the cluster to be updated.</p>}

\item{SecretArnList}{[required]
<p>List of AWS Secrets Manager secret ARNs.</p>}
}
\description{
\preformatted{        <p>Associates one or more Scram Secrets with an Amazon MSK cluster.</p>
}
}
\section{Request syntax}{
\preformatted{svc$batch_associate_scram_secret(
  ClusterArn = "string",
  SecretArnList = list(
    "string"
  )
)
}
}

\keyword{internal}
