% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_batch_disassociate_scram_secret}
\alias{kafka_batch_disassociate_scram_secret}
\title{Disassociates one or more Scram Secrets from an Amazon MSK cluster}
\usage{
kafka_batch_disassociate_scram_secret(ClusterArn, SecretArnList)
}
\arguments{
\item{ClusterArn}{[required] The Amazon Resource Name (ARN) of the cluster to be updated.}

\item{SecretArnList}{[required] List of AWS Secrets Manager secret ARNs.}
}
\description{
Disassociates one or more Scram Secrets from an Amazon MSK cluster.

See \url{https://www.paws-r-sdk.com/docs/kafka_batch_disassociate_scram_secret/} for full documentation.
}
\keyword{internal}
