% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_create_channel}
\alias{kafka_create_channel}
\title{Creates a Channel that streams records from an Amazon MSK Express
cluster topic to Amazon S3 or Apache Iceberg}
\usage{
kafka_create_channel(
  ChannelName,
  ClusterArn,
  EncryptionConfiguration = NULL,
  IcebergDestinationConfiguration = NULL,
  S3DestinationConfiguration = NULL,
  Tags = NULL,
  TopicConfigurationList,
  LoggingInfo = NULL
)
}
\arguments{
\item{ChannelName}{[required] The name of the channel. Must be unique within the cluster.}

\item{ClusterArn}{[required] The Amazon Resource Name (ARN) that uniquely identifies the cluster.}

\item{EncryptionConfiguration}{The encryption configuration applied to the channel.}

\item{IcebergDestinationConfiguration}{The Apache Iceberg destination for the channel. Mutually exclusive with s3DestinationConfiguration.}

\item{S3DestinationConfiguration}{The Amazon S3 destination for the channel. Mutually exclusive with icebergDestinationConfiguration.}

\item{Tags}{The tags attached to the channel.}

\item{TopicConfigurationList}{[required] The list of topic configurations for the channel. Currently exactly one topic must be specified.}

\item{LoggingInfo}{The destinations to which the channel publishes operational logs.}
}
\description{
Creates a Channel that streams records from an Amazon MSK Express cluster topic to Amazon S3 or Apache Iceberg.

See \url{https://www.paws-r-sdk.com/docs/kafka_create_channel/} for full documentation.
}
\keyword{internal}
