% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_create_cluster_v2}
\alias{kafka_create_cluster_v2}
\title{Creates a new MSK cluster}
\usage{
kafka_create_cluster_v2(
  ClusterName,
  Tags = NULL,
  Provisioned = NULL,
  Serverless = NULL
)
}
\arguments{
\item{ClusterName}{[required] The name of the cluster.}

\item{Tags}{A map of tags that you want the cluster to have.}

\item{Provisioned}{Information about the provisioned cluster.}

\item{Serverless}{Information about the serverless cluster.}
}
\description{
Creates a new MSK cluster.

See \url{https://www.paws-r-sdk.com/docs/kafka_create_cluster_v2/} for full documentation.
}
\keyword{internal}
