% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_create_configuration}
\alias{kafka_create_configuration}
\title{Creates a new MSK configuration}
\usage{
kafka_create_configuration(
  Description = NULL,
  KafkaVersions = NULL,
  Name,
  ServerProperties
)
}
\arguments{
\item{Description}{The description of the configuration.}

\item{KafkaVersions}{The versions of Apache Kafka with which you can use this MSK
configuration.}

\item{Name}{[required] The name of the configuration.}

\item{ServerProperties}{[required] Contents of the server.properties file. When using the API, you must
ensure that the contents of the file are base64 encoded. When using the
AWS Management Console, the SDK, or the AWS CLI, the contents of
server.properties can be in plaintext.}
}
\description{
Creates a new MSK configuration.

See \url{https://www.paws-r-sdk.com/docs/kafka_create_configuration/} for full documentation.
}
\keyword{internal}
