% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_create_replicator}
\alias{kafka_create_replicator}
\title{Creates the replicator}
\usage{
kafka_create_replicator(
  Description = NULL,
  KafkaClusters,
  ReplicationInfoList,
  ReplicatorName,
  ServiceExecutionRoleArn,
  Tags = NULL
)
}
\arguments{
\item{Description}{A summary description of the replicator.}

\item{KafkaClusters}{[required] Kafka Clusters to use in setting up sources / targets for replication.}

\item{ReplicationInfoList}{[required] A list of replication configurations, where each configuration targets a
given source cluster to target cluster replication flow.}

\item{ReplicatorName}{[required] The name of the replicator. Alpha-numeric characters with '-' are
allowed.}

\item{ServiceExecutionRoleArn}{[required] The ARN of the IAM role used by the replicator to access resources in
the customer's account (e.g source and target clusters)}

\item{Tags}{List of tags to attach to created Replicator.}
}
\description{
Creates the replicator.

See \url{https://www.paws-r-sdk.com/docs/kafka_create_replicator/} for full documentation.
}
\keyword{internal}
