% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_create_topic}
\alias{kafka_create_topic}
\title{Creates a topic in the specified MSK cluster}
\usage{
kafka_create_topic(
  ClusterArn,
  TopicName,
  PartitionCount,
  ReplicationFactor,
  Configs = NULL
)
}
\arguments{
\item{ClusterArn}{[required] The Amazon Resource Name (ARN) that uniquely identifies the cluster.}

\item{TopicName}{[required] The name of the topic to create.}

\item{PartitionCount}{[required] The number of partitions for the topic.}

\item{ReplicationFactor}{[required] The replication factor for the topic.}

\item{Configs}{Topic configurations encoded as a Base64 string.}
}
\description{
Creates a topic in the specified MSK cluster.

See \url{https://www.paws-r-sdk.com/docs/kafka_create_topic/} for full documentation.
}
\keyword{internal}
