% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_create_vpc_connection}
\alias{kafka_create_vpc_connection}
\title{Creates a new MSK VPC connection}
\usage{
kafka_create_vpc_connection(
  TargetClusterArn,
  Authentication,
  VpcId,
  ClientSubnets,
  SecurityGroups,
  Tags = NULL
)
}
\arguments{
\item{TargetClusterArn}{[required] The cluster Amazon Resource Name (ARN) for the VPC connection.}

\item{Authentication}{[required] The authentication type of VPC connection.}

\item{VpcId}{[required] The VPC ID of VPC connection.}

\item{ClientSubnets}{[required] The list of client subnets.}

\item{SecurityGroups}{[required] The list of security groups.}

\item{Tags}{A map of tags for the VPC connection.}
}
\description{
Creates a new MSK VPC connection.

See \url{https://paws-r.github.io/docs/kafka/create_vpc_connection.html} for full documentation.
}
\keyword{internal}
