% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_delete_channel}
\alias{kafka_delete_channel}
\title{Deletes the channel specified by channelArn from the cluster specified
by clusterArn}
\usage{
kafka_delete_channel(ChannelArn, ClusterArn)
}
\arguments{
\item{ChannelArn}{[required] The Amazon Resource Name (ARN) that uniquely identifies the channel.}

\item{ClusterArn}{[required] The Amazon Resource Name (ARN) that uniquely identifies the cluster.}
}
\description{
Deletes the channel specified by channelArn from the cluster specified by clusterArn. The channel transitions through DELETING and is removed when the asynchronous delete completes.

See \url{https://www.paws-r-sdk.com/docs/kafka_delete_channel/} for full documentation.
}
\keyword{internal}
