% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_delete_cluster_policy}
\alias{kafka_delete_cluster_policy}
\title{Deletes the MSK cluster policy specified by the Amazon Resource Name
(ARN) in the request}
\usage{
kafka_delete_cluster_policy(ClusterArn)
}
\arguments{
\item{ClusterArn}{[required] The Amazon Resource Name (ARN) of the cluster.}
}
\description{
Deletes the MSK cluster policy specified by the Amazon Resource Name (ARN) in the request.

See \url{https://paws-r.github.io/docs/kafka/delete_cluster_policy.html} for full documentation.
}
\keyword{internal}
