% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_delete_configuration}
\alias{kafka_delete_configuration}
\title{Deletes an MSK Configuration}
\usage{
kafka_delete_configuration(Arn)
}
\arguments{
\item{Arn}{[required] The Amazon Resource Name (ARN) that uniquely identifies an MSK
configuration.}
}
\value{
A list with the following syntax:\preformatted{list(
  Arn = "string",
  State = "ACTIVE"|"DELETING"|"DELETE_FAILED"
)
}
}
\description{
Deletes an MSK Configuration.
}
\section{Request syntax}{
\preformatted{svc$delete_configuration(
  Arn = "string"
)
}
}

\keyword{internal}
