% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_delete_replicator}
\alias{kafka_delete_replicator}
\title{Deletes a replicator}
\usage{
kafka_delete_replicator(CurrentVersion = NULL, ReplicatorArn)
}
\arguments{
\item{CurrentVersion}{The current version of the replicator.}

\item{ReplicatorArn}{[required] The Amazon Resource Name (ARN) of the replicator to be deleted.}
}
\description{
Deletes a replicator.

See \url{https://www.paws-r-sdk.com/docs/kafka_delete_replicator/} for full documentation.
}
\keyword{internal}
