% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_delete_topic}
\alias{kafka_delete_topic}
\title{Deletes a topic in the specified MSK cluster}
\usage{
kafka_delete_topic(ClusterArn, TopicName)
}
\arguments{
\item{ClusterArn}{[required] The Amazon Resource Name (ARN) that uniquely identifies the cluster.}

\item{TopicName}{[required] The name of the topic to delete.}
}
\description{
Deletes a topic in the specified MSK cluster.

See \url{https://www.paws-r-sdk.com/docs/kafka_delete_topic/} for full documentation.
}
\keyword{internal}
