% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_describe_channel}
\alias{kafka_describe_channel}
\title{Returns the current configuration and state of a channel}
\usage{
kafka_describe_channel(ChannelArn, ClusterArn)
}
\arguments{
\item{ChannelArn}{[required] The Amazon Resource Name (ARN) that uniquely identifies the channel.}

\item{ClusterArn}{[required] The Amazon Resource Name (ARN) that uniquely identifies the cluster.}
}
\description{
Returns the current configuration and state of a channel.

See \url{https://www.paws-r-sdk.com/docs/kafka_describe_channel/} for full documentation.
}
\keyword{internal}
