% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_describe_cluster_operation}
\alias{kafka_describe_cluster_operation}
\title{Returns a description of the cluster operation specified by the ARN}
\usage{
kafka_describe_cluster_operation(ClusterOperationArn)
}
\arguments{
\item{ClusterOperationArn}{[required] The Amazon Resource Name (ARN) that uniquely identifies the MSK cluster
operation.}
}
\description{
Returns a description of the cluster operation specified by the ARN.

See \url{https://paws-r.github.io/docs/kafka/describe_cluster_operation.html} for full documentation.
}
\keyword{internal}
