% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_describe_cluster_operation_v2}
\alias{kafka_describe_cluster_operation_v2}
\title{Returns a description of the cluster operation specified by the ARN}
\usage{
kafka_describe_cluster_operation_v2(ClusterOperationArn)
}
\arguments{
\item{ClusterOperationArn}{[required] ARN of the cluster operation to describe.}
}
\description{
Returns a description of the cluster operation specified by the ARN.

See \url{https://www.paws-r-sdk.com/docs/kafka_describe_cluster_operation_v2/} for full documentation.
}
\keyword{internal}
