% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_describe_configuration}
\alias{kafka_describe_configuration}
\title{Returns a description of this MSK configuration}
\usage{
kafka_describe_configuration(Arn)
}
\arguments{
\item{Arn}{[required] The Amazon Resource Name (ARN) that uniquely identifies an MSK
configuration and all of its revisions.}
}
\description{
Returns a description of this MSK configuration.

See \url{https://paws-r.github.io/docs/kafka/describe_configuration.html} for full documentation.
}
\keyword{internal}
