% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_describe_configuration_revision}
\alias{kafka_describe_configuration_revision}
\title{Returns a description of this revision of the configuration}
\usage{
kafka_describe_configuration_revision(Arn, Revision)
}
\arguments{
\item{Arn}{[required] The Amazon Resource Name (ARN) that uniquely identifies an MSK
configuration and all of its revisions.}

\item{Revision}{[required] A string that uniquely identifies a revision of an MSK configuration.}
}
\description{
Returns a description of this revision of the configuration.

See \url{https://www.paws-r-sdk.com/docs/kafka_describe_configuration_revision/} for full documentation.
}
\keyword{internal}
