% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_describe_replicator}
\alias{kafka_describe_replicator}
\title{Describes a replicator}
\usage{
kafka_describe_replicator(ReplicatorArn)
}
\arguments{
\item{ReplicatorArn}{[required] The Amazon Resource Name (ARN) of the replicator to be described.}
}
\description{
Describes a replicator.

See \url{https://www.paws-r-sdk.com/docs/kafka_describe_replicator/} for full documentation.
}
\keyword{internal}
