% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_describe_topic}
\alias{kafka_describe_topic}
\title{Returns topic details of this topic on a MSK cluster}
\usage{
kafka_describe_topic(ClusterArn, TopicName)
}
\arguments{
\item{ClusterArn}{[required] The Amazon Resource Name (ARN) that uniquely identifies the cluster.}

\item{TopicName}{[required] The Kafka topic name that uniquely identifies the topic.}
}
\description{
Returns topic details of this topic on a MSK cluster.

See \url{https://www.paws-r-sdk.com/docs/kafka_describe_topic/} for full documentation.
}
\keyword{internal}
