% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_describe_topic_partitions}
\alias{kafka_describe_topic_partitions}
\title{Returns partition details of this topic on a MSK cluster}
\usage{
kafka_describe_topic_partitions(
  ClusterArn,
  TopicName,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{ClusterArn}{[required] The Amazon Resource Name (ARN) that uniquely identifies the cluster.}

\item{TopicName}{[required] The Kafka topic name that uniquely identifies the topic.}

\item{MaxResults}{The maximum number of results to return in the response. If there are
more results, the response includes a NextToken parameter.}

\item{NextToken}{The paginated results marker. When the result of the operation is
truncated, the call returns NextToken in the response. To get the next
batch, provide this token in your next request.}
}
\description{
Returns partition details of this topic on a MSK cluster.

See \url{https://www.paws-r-sdk.com/docs/kafka_describe_topic_partitions/} for full documentation.
}
\keyword{internal}
