% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_get_cluster_policy}
\alias{kafka_get_cluster_policy}
\title{Get the MSK cluster policy specified by the Amazon Resource Name (ARN)
in the request}
\usage{
kafka_get_cluster_policy(ClusterArn)
}
\arguments{
\item{ClusterArn}{[required] The Amazon Resource Name (ARN) of the cluster.}
}
\description{
Get the MSK cluster policy specified by the Amazon Resource Name (ARN) in the request.

See \url{https://paws-r.github.io/docs/kafka/get_cluster_policy.html} for full documentation.
}
\keyword{internal}
