% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_get_compatible_kafka_versions}
\alias{kafka_get_compatible_kafka_versions}
\title{Gets the Apache Kafka versions to which you can update the MSK cluster}
\usage{
kafka_get_compatible_kafka_versions(ClusterArn = NULL)
}
\arguments{
\item{ClusterArn}{The Amazon Resource Name (ARN) of the cluster check.}
}
\description{
Gets the Apache Kafka versions to which you can update the MSK cluster.

See \url{https://paws-r.github.io/docs/kafka/get_compatible_kafka_versions.html} for full documentation.
}
\keyword{internal}
