% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_list_channels}
\alias{kafka_list_channels}
\title{Returns the list of channels in a cluster}
\usage{
kafka_list_channels(
  ClusterArn,
  MaxResults = NULL,
  NextToken = NULL,
  TopicNameFilter = NULL
)
}
\arguments{
\item{ClusterArn}{[required] The Amazon Resource Name (ARN) that uniquely identifies the cluster.}

\item{MaxResults}{Maximum number of channels to return in a single response.}

\item{NextToken}{If the response of ListChannels is truncated, it returns a nextToken in the response. This nextToken should be sent in the subsequent request to ListChannels.}

\item{TopicNameFilter}{Filters results to channels whose topic name matches the specified value.}
}
\description{
Returns the list of channels in a cluster.

See \url{https://www.paws-r-sdk.com/docs/kafka_list_channels/} for full documentation.
}
\keyword{internal}
