% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_list_cluster_operations_v2}
\alias{kafka_list_cluster_operations_v2}
\title{Returns a list of all the operations that have been performed on the
specified MSK cluster}
\usage{
kafka_list_cluster_operations_v2(
  ClusterArn,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{ClusterArn}{[required] The arn of the cluster whose operations are being requested.}

\item{MaxResults}{The maxResults of the query.}

\item{NextToken}{The nextToken of the query.}
}
\description{
Returns a list of all the operations that have been performed on the specified MSK cluster.

See \url{https://www.paws-r-sdk.com/docs/kafka_list_cluster_operations_v2/} for full documentation.
}
\keyword{internal}
