% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_list_replicators}
\alias{kafka_list_replicators}
\title{Lists the replicators}
\usage{
kafka_list_replicators(
  MaxResults = NULL,
  NextToken = NULL,
  ReplicatorNameFilter = NULL
)
}
\arguments{
\item{MaxResults}{The maximum number of results to return in the response. If there are
more results, the response includes a NextToken parameter.}

\item{NextToken}{If the response of ListReplicators is truncated, it returns a NextToken
in the response. This NextToken should be sent in the subsequent request
to ListReplicators.}

\item{ReplicatorNameFilter}{Returns replicators starting with given name.}
}
\description{
Lists the replicators.

See \url{https://www.paws-r-sdk.com/docs/kafka_list_replicators/} for full documentation.
}
\keyword{internal}
