% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_list_topics}
\alias{kafka_list_topics}
\title{List topics in a MSK cluster}
\usage{
kafka_list_topics(
  ClusterArn,
  MaxResults = NULL,
  NextToken = NULL,
  TopicNameFilter = NULL
)
}
\arguments{
\item{ClusterArn}{[required] The Amazon Resource Name (ARN) that uniquely identifies the cluster.}

\item{MaxResults}{The maximum number of results to return in the response. If there are
more results, the response includes a NextToken parameter.}

\item{NextToken}{The paginated results marker. When the result of the operation is
truncated, the call returns NextToken in the response. To get the next
batch, provide this token in your next request.}

\item{TopicNameFilter}{Returns topics starting with given name.}
}
\description{
List topics in a MSK cluster.

See \url{https://www.paws-r-sdk.com/docs/kafka_list_topics/} for full documentation.
}
\keyword{internal}
