% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_list_vpc_connections}
\alias{kafka_list_vpc_connections}
\title{Returns a list of all the VPC connections in this Region}
\usage{
kafka_list_vpc_connections(MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{The maximum number of results to return in the response. If there are
more results, the response includes a NextToken parameter.}

\item{NextToken}{The paginated results marker. When the result of the operation is
truncated, the call returns NextToken in the response. To get the next
batch, provide this token in your next request.}
}
\description{
Returns a list of all the VPC connections in this Region.

See \url{https://www.paws-r-sdk.com/docs/kafka_list_vpc_connections/} for full documentation.
}
\keyword{internal}
