% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_reboot_broker}
\alias{kafka_reboot_broker}
\title{Reboots brokers}
\usage{
kafka_reboot_broker(BrokerIds, ClusterArn)
}
\arguments{
\item{BrokerIds}{[required] The list of broker IDs to be rebooted. The reboot-broker operation
supports rebooting one broker at a time.}

\item{ClusterArn}{[required] The Amazon Resource Name (ARN) of the cluster to be updated.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ClusterArn = "string",
  ClusterOperationArn = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Reboots brokers.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$reboot_broker(
  BrokerIds = list(
    "string"
  ),
  ClusterArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
