% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_tag_resource}
\alias{kafka_tag_resource}
\title{Adds tags to the specified MSK resource}
\usage{
kafka_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) that uniquely identifies the resource
that's associated with the tags.}

\item{Tags}{[required] The key-value pair for the resource tag.}
}
\description{
Adds tags to the specified MSK resource.

See \url{https://www.paws-r-sdk.com/docs/kafka_tag_resource/} for full documentation.
}
\keyword{internal}
