% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_update_broker_count}
\alias{kafka_update_broker_count}
\title{Updates the number of broker nodes in the cluster}
\usage{
kafka_update_broker_count(ClusterArn, CurrentVersion,
  TargetNumberOfBrokerNodes)
}
\arguments{
\item{ClusterArn}{[required] The Amazon Resource Name (ARN) that uniquely identifies the cluster.}

\item{CurrentVersion}{[required] The version of cluster to update from. A successful operation will then
generate a new version.}

\item{TargetNumberOfBrokerNodes}{[required] The number of broker nodes that you want the cluster to have after this
operation completes successfully.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ClusterArn = "string",
  ClusterOperationArn = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Updates the number of broker nodes in the cluster.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_broker_count(
  ClusterArn = "string",
  CurrentVersion = "string",
  TargetNumberOfBrokerNodes = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
