% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_update_broker_storage}
\alias{kafka_update_broker_storage}
\title{Updates the EBS storage associated with MSK brokers}
\usage{
kafka_update_broker_storage(
  ClusterArn,
  CurrentVersion,
  TargetBrokerEBSVolumeInfo
)
}
\arguments{
\item{ClusterArn}{[required] The Amazon Resource Name (ARN) that uniquely identifies the cluster.}

\item{CurrentVersion}{[required] The version of cluster to update from. A successful operation will then
generate a new version.}

\item{TargetBrokerEBSVolumeInfo}{[required] Describes the target volume size and the ID of the broker to apply the
update to.}
}
\description{
Updates the EBS storage associated with MSK brokers.

See \url{https://www.paws-r-sdk.com/docs/kafka_update_broker_storage/} for full documentation.
}
\keyword{internal}
