% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_update_broker_type}
\alias{kafka_update_broker_type}
\title{Updates EC2 instance type}
\usage{
kafka_update_broker_type(ClusterArn, CurrentVersion, TargetInstanceType)
}
\arguments{
\item{ClusterArn}{[required] The Amazon Resource Name (ARN) that uniquely identifies the cluster.}

\item{CurrentVersion}{[required] The cluster version that you want to change. After this operation
completes successfully, the cluster will have a new version.}

\item{TargetInstanceType}{[required] The Amazon MSK broker type that you want all of the brokers in this
cluster to be.}
}
\description{
Updates EC2 instance type.

See \url{https://paws-r.github.io/docs/kafka/update_broker_type.html} for full documentation.
}
\keyword{internal}
