% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_update_channel}
\alias{kafka_update_channel}
\title{Updates the destination configuration of an existing channel}
\usage{
kafka_update_channel(
  ChannelArn,
  ClusterArn,
  IcebergDestinationUpdate = NULL,
  S3DestinationUpdate = NULL
)
}
\arguments{
\item{ChannelArn}{[required] The Amazon Resource Name (ARN) that uniquely identifies the channel.}

\item{ClusterArn}{[required] The Amazon Resource Name (ARN) that uniquely identifies the cluster.}

\item{IcebergDestinationUpdate}{Updates fields on an Apache Iceberg destination. Use only when the channel was created with an Iceberg destination.}

\item{S3DestinationUpdate}{Updates fields on an Amazon S3 destination. Use only when the channel was created with an Amazon S3 destination.}
}
\description{
Updates the destination configuration of an existing channel. Exactly one of icebergDestinationUpdate or s3DestinationUpdate must be supplied.

See \url{https://www.paws-r-sdk.com/docs/kafka_update_channel/} for full documentation.
}
\keyword{internal}
