% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_update_cluster_kafka_version}
\alias{kafka_update_cluster_kafka_version}
\title{Updates the Apache Kafka version for the cluster}
\usage{
kafka_update_cluster_kafka_version(ClusterArn, ConfigurationInfo,
  CurrentVersion, TargetKafkaVersion)
}
\arguments{
\item{ClusterArn}{[required]
<p>The Amazon Resource Name (ARN) of the cluster to be updated.</p>}

\item{ConfigurationInfo}{<p>The custom configuration that should be applied on the new version of cluster.</p>}

\item{CurrentVersion}{[required]
<p>Current cluster version.</p>}

\item{TargetKafkaVersion}{[required]
<p>Target Kafka version.</p>}
}
\description{
\preformatted{        <p>Updates the Apache Kafka version for the cluster.</p>
}
}
\section{Request syntax}{
\preformatted{svc$update_cluster_kafka_version(
  ClusterArn = "string",
  ConfigurationInfo = list(
    Arn = "string",
    Revision = 123
  ),
  CurrentVersion = "string",
  TargetKafkaVersion = "string"
)
}
}

\keyword{internal}
