% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_update_connectivity}
\alias{kafka_update_connectivity}
\title{Updates the cluster's connectivity configuration}
\usage{
kafka_update_connectivity(ClusterArn, ConnectivityInfo, CurrentVersion)
}
\arguments{
\item{ClusterArn}{[required] The Amazon Resource Name (ARN) of the configuration.}

\item{ConnectivityInfo}{[required] Information about the broker access configuration.}

\item{CurrentVersion}{[required] The version of the MSK cluster to update. Cluster versions aren't simple
numbers. You can describe an MSK cluster to find its version. When this
update operation is successful, it generates a new cluster version.}
}
\description{
Updates the cluster's connectivity configuration.

See \url{https://paws-r.github.io/docs/kafka/update_connectivity.html} for full documentation.
}
\keyword{internal}
