% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_update_rebalancing}
\alias{kafka_update_rebalancing}
\title{Use this resource to update the intelligent rebalancing status of an
Amazon MSK Provisioned cluster with Express brokers}
\usage{
kafka_update_rebalancing(ClusterArn, CurrentVersion, Rebalancing)
}
\arguments{
\item{ClusterArn}{[required] The Amazon Resource Name (ARN) of the cluster.}

\item{CurrentVersion}{[required] The current version of the cluster.}

\item{Rebalancing}{[required] Specifies if intelligent rebalancing should be turned on for your
cluster. The default intelligent rebalancing status is ACTIVE for all
new MSK Provisioned clusters that you create with Express brokers.}
}
\description{
Use this resource to update the intelligent rebalancing status of an Amazon MSK Provisioned cluster with Express brokers.

See \url{https://www.paws-r-sdk.com/docs/kafka_update_rebalancing/} for full documentation.
}
\keyword{internal}
