% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_update_security}
\alias{kafka_update_security}
\title{Updates the security settings for the cluster}
\usage{
kafka_update_security(
  ClientAuthentication = NULL,
  ClusterArn,
  CurrentVersion,
  EncryptionInfo = NULL
)
}
\arguments{
\item{ClientAuthentication}{Includes all client authentication related information.}

\item{ClusterArn}{[required] The Amazon Resource Name (ARN) that uniquely identifies the cluster.}

\item{CurrentVersion}{[required] The version of the MSK cluster to update. Cluster versions aren't simple
numbers. You can describe an MSK cluster to find its version. When this
update operation is successful, it generates a new cluster version.}

\item{EncryptionInfo}{Includes all encryption-related information.}
}
\description{
Updates the security settings for the cluster. You can use this operation to specify encryption and authentication on existing clusters.

See \url{https://paws-r.github.io/docs/kafka/update_security.html} for full documentation.
}
\keyword{internal}
