% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_update_storage}
\alias{kafka_update_storage}
\title{Updates cluster broker volume size (or) sets cluster storage mode to
TIERED}
\usage{
kafka_update_storage(
  ClusterArn,
  CurrentVersion,
  ProvisionedThroughput = NULL,
  StorageMode = NULL,
  VolumeSizeGB = NULL
)
}
\arguments{
\item{ClusterArn}{[required] The Amazon Resource Name (ARN) of the cluster to be updated.}

\item{CurrentVersion}{[required] The version of cluster to update from. A successful operation will then
generate a new version.}

\item{ProvisionedThroughput}{EBS volume provisioned throughput information.}

\item{StorageMode}{Controls storage mode for supported storage tiers.}

\item{VolumeSizeGB}{size of the EBS volume to update.}
}
\description{
Updates cluster broker volume size (or) sets cluster storage mode to TIERED.

See \url{https://paws-r.github.io/docs/kafka/update_storage.html} for full documentation.
}
\keyword{internal}
