% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_update_topic}
\alias{kafka_update_topic}
\title{Updates the configuration of the specified topic}
\usage{
kafka_update_topic(
  ClusterArn,
  TopicName,
  Configs = NULL,
  PartitionCount = NULL
)
}
\arguments{
\item{ClusterArn}{[required] The Amazon Resource Name (ARN) that uniquely identifies the cluster.}

\item{TopicName}{[required] The name of the topic to update configuration for.}

\item{Configs}{The new topic configurations encoded as a Base64 string.}

\item{PartitionCount}{The new total number of partitions for the topic.}
}
\description{
Updates the configuration of the specified topic.

See \url{https://www.paws-r-sdk.com/docs/kafka_update_topic/} for full documentation.
}
\keyword{internal}
