% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafkaconnect_operations.R
\name{kafkaconnect_create_custom_plugin}
\alias{kafkaconnect_create_custom_plugin}
\title{Creates a custom plugin using the specified properties}
\usage{
kafkaconnect_create_custom_plugin(
  contentType,
  description = NULL,
  location,
  name
)
}
\arguments{
\item{contentType}{[required] The type of the plugin file.}

\item{description}{A summary description of the custom plugin.}

\item{location}{[required] Information about the location of a custom plugin.}

\item{name}{[required] The name of the custom plugin.}
}
\description{
Creates a custom plugin using the specified properties.

See \url{https://www.paws-r-sdk.com/docs/kafkaconnect_create_custom_plugin/} for full documentation.
}
\keyword{internal}
