% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafkaconnect_operations.R
\name{kafkaconnect_create_worker_configuration}
\alias{kafkaconnect_create_worker_configuration}
\title{Creates a worker configuration using the specified properties}
\usage{
kafkaconnect_create_worker_configuration(
  description = NULL,
  name,
  propertiesFileContent
)
}
\arguments{
\item{description}{A summary description of the worker configuration.}

\item{name}{[required] The name of the worker configuration.}

\item{propertiesFileContent}{[required] Base64 encoded contents of connect-distributed.properties file.}
}
\description{
Creates a worker configuration using the specified properties.

See \url{https://www.paws-r-sdk.com/docs/kafkaconnect_create_worker_configuration/} for full documentation.
}
\keyword{internal}
