% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafkaconnect_operations.R
\name{kafkaconnect_delete_connector}
\alias{kafkaconnect_delete_connector}
\title{Deletes the specified connector}
\usage{
kafkaconnect_delete_connector(connectorArn, currentVersion = NULL)
}
\arguments{
\item{connectorArn}{[required] The Amazon Resource Name (ARN) of the connector that you want to delete.}

\item{currentVersion}{The current version of the connector that you want to delete.}
}
\description{
Deletes the specified connector.

See \url{https://www.paws-r-sdk.com/docs/kafkaconnect_delete_connector/} for full documentation.
}
\keyword{internal}
