% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafkaconnect_operations.R
\name{kafkaconnect_delete_custom_plugin}
\alias{kafkaconnect_delete_custom_plugin}
\title{Deletes a custom plugin}
\usage{
kafkaconnect_delete_custom_plugin(customPluginArn)
}
\arguments{
\item{customPluginArn}{[required] The Amazon Resource Name (ARN) of the custom plugin that you want to
delete.}
}
\description{
Deletes a custom plugin.

See \url{https://paws-r.github.io/docs/kafkaconnect/delete_custom_plugin.html} for full documentation.
}
\keyword{internal}
