% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafkaconnect_operations.R
\name{kafkaconnect_delete_worker_configuration}
\alias{kafkaconnect_delete_worker_configuration}
\title{Deletes the specified worker configuration}
\usage{
kafkaconnect_delete_worker_configuration(workerConfigurationArn)
}
\arguments{
\item{workerConfigurationArn}{[required] The Amazon Resource Name (ARN) of the worker configuration that you want
to delete.}
}
\description{
Deletes the specified worker configuration.

See \url{https://www.paws-r-sdk.com/docs/kafkaconnect_delete_worker_configuration/} for full documentation.
}
\keyword{internal}
