% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafkaconnect_operations.R
\name{kafkaconnect_describe_connector}
\alias{kafkaconnect_describe_connector}
\title{Returns summary information about the connector}
\usage{
kafkaconnect_describe_connector(connectorArn)
}
\arguments{
\item{connectorArn}{[required] The Amazon Resource Name (ARN) of the connector that you want to
describe.}
}
\description{
Returns summary information about the connector.

See \url{https://www.paws-r-sdk.com/docs/kafkaconnect_describe_connector/} for full documentation.
}
\keyword{internal}
