% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafkaconnect_operations.R
\name{kafkaconnect_describe_connector_operation}
\alias{kafkaconnect_describe_connector_operation}
\title{Returns information about the specified connector's operations}
\usage{
kafkaconnect_describe_connector_operation(connectorOperationArn)
}
\arguments{
\item{connectorOperationArn}{[required] ARN of the connector operation to be described.}
}
\description{
Returns information about the specified connector's operations.

See \url{https://www.paws-r-sdk.com/docs/kafkaconnect_describe_connector_operation/} for full documentation.
}
\keyword{internal}
