% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafkaconnect_operations.R
\name{kafkaconnect_describe_custom_plugin}
\alias{kafkaconnect_describe_custom_plugin}
\title{A summary description of the custom plugin}
\usage{
kafkaconnect_describe_custom_plugin(customPluginArn)
}
\arguments{
\item{customPluginArn}{[required] Returns information about a custom plugin.}
}
\description{
A summary description of the custom plugin.

See \url{https://www.paws-r-sdk.com/docs/kafkaconnect_describe_custom_plugin/} for full documentation.
}
\keyword{internal}
