% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafkaconnect_operations.R
\name{kafkaconnect_describe_worker_configuration}
\alias{kafkaconnect_describe_worker_configuration}
\title{Returns information about a worker configuration}
\usage{
kafkaconnect_describe_worker_configuration(workerConfigurationArn)
}
\arguments{
\item{workerConfigurationArn}{[required] The Amazon Resource Name (ARN) of the worker configuration that you want
to get information about.}
}
\description{
Returns information about a worker configuration.

See \url{https://paws-r.github.io/docs/kafkaconnect/describe_worker_configuration.html} for full documentation.
}
\keyword{internal}
