% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafkaconnect_operations.R
\name{kafkaconnect_list_connector_operations}
\alias{kafkaconnect_list_connector_operations}
\title{Lists information about a connector's operation(s)}
\usage{
kafkaconnect_list_connector_operations(
  connectorArn,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{connectorArn}{[required] The Amazon Resource Name (ARN) of the connector for which to list
operations.}

\item{maxResults}{Maximum number of connector operations to fetch in one get request.}

\item{nextToken}{If the response is truncated, it includes a NextToken. Send this
NextToken in a subsequent request to continue listing from where it left
off.}
}
\description{
Lists information about a connector's operation(s).

See \url{https://www.paws-r-sdk.com/docs/kafkaconnect_list_connector_operations/} for full documentation.
}
\keyword{internal}
