% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafkaconnect_operations.R
\name{kafkaconnect_list_connectors}
\alias{kafkaconnect_list_connectors}
\title{Returns a list of all the connectors in this account and Region}
\usage{
kafkaconnect_list_connectors(
  connectorNamePrefix = NULL,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{connectorNamePrefix}{The name prefix that you want to use to search for and list connectors.}

\item{maxResults}{The maximum number of connectors to list in one response.}

\item{nextToken}{If the response of a ListConnectors operation is truncated, it will
include a NextToken. Send this NextToken in a subsequent request to
continue listing from where the previous operation left off.}
}
\description{
Returns a list of all the connectors in this account and Region. The list is limited to connectors whose name starts with the specified prefix. The response also includes a description of each of the listed connectors.

See \url{https://www.paws-r-sdk.com/docs/kafkaconnect_list_connectors/} for full documentation.
}
\keyword{internal}
