% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafkaconnect_operations.R
\name{kafkaconnect_list_custom_plugins}
\alias{kafkaconnect_list_custom_plugins}
\title{Returns a list of all of the custom plugins in this account and Region}
\usage{
kafkaconnect_list_custom_plugins(maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{maxResults}{The maximum number of custom plugins to list in one response.}

\item{nextToken}{If the response of a ListCustomPlugins operation is truncated, it will
include a NextToken. Send this NextToken in a subsequent request to
continue listing from where the previous operation left off.}
}
\description{
Returns a list of all of the custom plugins in this account and Region.

See \url{https://paws-r.github.io/docs/kafkaconnect/list_custom_plugins.html} for full documentation.
}
\keyword{internal}
