% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafkaconnect_operations.R
\name{kafkaconnect_list_tags_for_resource}
\alias{kafkaconnect_list_tags_for_resource}
\title{Lists all the tags attached to the specified resource}
\usage{
kafkaconnect_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource for which you want to
list all attached tags.}
}
\description{
Lists all the tags attached to the specified resource.

See \url{https://www.paws-r-sdk.com/docs/kafkaconnect_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
