% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafkaconnect_operations.R
\name{kafkaconnect_restart_connector}
\alias{kafkaconnect_restart_connector}
\title{Restarts the specified connector}
\usage{
kafkaconnect_restart_connector(connectorArn, onlyFailedTasks = NULL)
}
\arguments{
\item{connectorArn}{[required] The Amazon Resource Name (ARN) of the connector that you want to restart.}

\item{onlyFailedTasks}{Specifies whether to restart only the connector's failed tasks. If \code{true}, the operation restarts only the tasks that are currently in a failed state, and healthy tasks continue running. If \code{false} or not specified, the operation restarts the connector and all of its tasks.}
}
\description{
Restarts the specified connector. By default, this operation restarts the connector and all of its tasks. This operation is asynchronous and returns a connector operation ARN that you can pass to \code{\link[=kafkaconnect_describe_connector_operation]{describe_connector_operation}} to track the state of the restart.

See \url{https://www.paws-r-sdk.com/docs/kafkaconnect_restart_connector/} for full documentation.
}
\keyword{internal}
