% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafkaconnect_operations.R
\name{kafkaconnect_tag_resource}
\alias{kafkaconnect_tag_resource}
\title{Attaches tags to the specified resource}
\usage{
kafkaconnect_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource to which you want to
attach tags.}

\item{tags}{[required] The tags that you want to attach to the resource.}
}
\description{
Attaches tags to the specified resource.

See \url{https://www.paws-r-sdk.com/docs/kafkaconnect_tag_resource/} for full documentation.
}
\keyword{internal}
