% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafkaconnect_operations.R
\name{kafkaconnect_untag_resource}
\alias{kafkaconnect_untag_resource}
\title{Removes tags from the specified resource}
\usage{
kafkaconnect_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource from which you want to
remove tags.}

\item{tagKeys}{[required] The keys of the tags that you want to remove from the resource.}
}
\description{
Removes tags from the specified resource.

See \url{https://www.paws-r-sdk.com/docs/kafkaconnect_untag_resource/} for full documentation.
}
\keyword{internal}
