% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafkaconnect_operations.R
\name{kafkaconnect_update_connector}
\alias{kafkaconnect_update_connector}
\title{Updates the specified connector}
\usage{
kafkaconnect_update_connector(capacity, connectorArn, currentVersion)
}
\arguments{
\item{capacity}{[required] The target capacity.}

\item{connectorArn}{[required] The Amazon Resource Name (ARN) of the connector that you want to update.}

\item{currentVersion}{[required] The current version of the connector that you want to update.}
}
\description{
Updates the specified connector.

See \url{https://paws-r.github.io/docs/kafkaconnect/update_connector.html} for full documentation.
}
\keyword{internal}
