% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendra_operations.R
\name{kendra_batch_delete_document}
\alias{kendra_batch_delete_document}
\title{Removes one or more documents from an index}
\usage{
kendra_batch_delete_document(
  IndexId,
  DocumentIdList,
  DataSourceSyncJobMetricTarget = NULL
)
}
\arguments{
\item{IndexId}{[required] The identifier of the index that contains the documents to delete.}

\item{DocumentIdList}{[required] One or more identifiers for documents to delete from the index.}

\item{DataSourceSyncJobMetricTarget}{}
}
\description{
Removes one or more documents from an index. The documents must have been added with the \code{\link[=kendra_batch_put_document]{batch_put_document}} API.

See \url{https://www.paws-r-sdk.com/docs/kendra_batch_delete_document/} for full documentation.
}
\keyword{internal}
