% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendra_operations.R
\name{kendra_batch_delete_featured_results_set}
\alias{kendra_batch_delete_featured_results_set}
\title{Removes one or more sets of featured results}
\usage{
kendra_batch_delete_featured_results_set(IndexId, FeaturedResultsSetIds)
}
\arguments{
\item{IndexId}{[required] The identifier of the index used for featuring results.}

\item{FeaturedResultsSetIds}{[required] The identifiers of the featured results sets that you want to delete.}
}
\description{
Removes one or more sets of featured results. Features results are placed above all other results for certain queries. If there's an exact match of a query, then one or more specific documents are featured in the search results.

See \url{https://paws-r.github.io/docs/kendra/batch_delete_featured_results_set.html} for full documentation.
}
\keyword{internal}
