% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendra_operations.R
\name{kendra_delete_access_control_configuration}
\alias{kendra_delete_access_control_configuration}
\title{Deletes an access control configuration that you created for your
documents in an index}
\usage{
kendra_delete_access_control_configuration(IndexId, Id)
}
\arguments{
\item{IndexId}{[required] The identifier of the index for an access control configuration.}

\item{Id}{[required] The identifier of the access control configuration you want to delete.}
}
\description{
Deletes an access control configuration that you created for your documents in an index. This includes user and group access information for your documents. This is useful for user context filtering, where search results are filtered based on the user or their group access to documents.

See \url{https://www.paws-r-sdk.com/docs/kendra_delete_access_control_configuration/} for full documentation.
}
\keyword{internal}
