% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendra_operations.R
\name{kendra_delete_data_source}
\alias{kendra_delete_data_source}
\title{Deletes an Amazon Kendra data source connector}
\usage{
kendra_delete_data_source(Id, IndexId)
}
\arguments{
\item{Id}{[required] The identifier of the data source connector you want to delete.}

\item{IndexId}{[required] The identifier of the index used with the data source connector.}
}
\description{
Deletes an Amazon Kendra data source connector. An exception is not thrown if the data source is already being deleted. While the data source is being deleted, the \code{Status} field returned by a call to the \code{\link[=kendra_describe_data_source]{describe_data_source}} API is set to \code{DELETING}. For more information, see \href{https://docs.aws.amazon.com/kendra/latest/dg/delete-data-source.html}{Deleting Data Sources}.

See \url{https://www.paws-r-sdk.com/docs/kendra_delete_data_source/} for full documentation.
}
\keyword{internal}
