% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendra_operations.R
\name{kendra_delete_experience}
\alias{kendra_delete_experience}
\title{Deletes your Amazon Kendra experience such as a search application}
\usage{
kendra_delete_experience(Id, IndexId)
}
\arguments{
\item{Id}{[required] The identifier of your Amazon Kendra experience you want to delete.}

\item{IndexId}{[required] The identifier of the index for your Amazon Kendra experience.}
}
\description{
Deletes your Amazon Kendra experience such as a search application. For more information on creating a search application experience, see \href{https://docs.aws.amazon.com/kendra/latest/dg/deploying-search-experience-no-code.html}{Building a search experience with no code}.

See \url{https://www.paws-r-sdk.com/docs/kendra_delete_experience/} for full documentation.
}
\keyword{internal}
