% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendra_operations.R
\name{kendra_delete_query_suggestions_block_list}
\alias{kendra_delete_query_suggestions_block_list}
\title{Deletes a block list used for query suggestions for an index}
\usage{
kendra_delete_query_suggestions_block_list(IndexId, Id)
}
\arguments{
\item{IndexId}{[required] The identifier of the index for the block list.}

\item{Id}{[required] The identifier of the block list you want to delete.}
}
\description{
Deletes a block list used for query suggestions for an index.

See \url{https://www.paws-r-sdk.com/docs/kendra_delete_query_suggestions_block_list/} for full documentation.
}
\keyword{internal}
