% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendra_operations.R
\name{kendra_delete_thesaurus}
\alias{kendra_delete_thesaurus}
\title{Deletes an existing Amazon Kendra thesaurus}
\usage{
kendra_delete_thesaurus(Id, IndexId)
}
\arguments{
\item{Id}{[required] The identifier of the thesaurus you want to delete.}

\item{IndexId}{[required] The identifier of the index for the thesaurus.}
}
\description{
Deletes an existing Amazon Kendra thesaurus.

See \url{https://paws-r.github.io/docs/kendra/delete_thesaurus.html} for full documentation.
}
\keyword{internal}
