% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendra_operations.R
\name{kendra_describe_access_control_configuration}
\alias{kendra_describe_access_control_configuration}
\title{Gets information about an access control configuration that you created
for your documents in an index}
\usage{
kendra_describe_access_control_configuration(IndexId, Id)
}
\arguments{
\item{IndexId}{[required] The identifier of the index for an access control configuration.}

\item{Id}{[required] The identifier of the access control configuration you want to get
information on.}
}
\description{
Gets information about an access control configuration that you created for your documents in an index. This includes user and group access information for your documents. This is useful for user context filtering, where search results are filtered based on the user or their group access to documents.

See \url{https://paws-r.github.io/docs/kendra/describe_access_control_configuration.html} for full documentation.
}
\keyword{internal}
