% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendra_operations.R
\name{kendra_describe_data_source}
\alias{kendra_describe_data_source}
\title{Gets information about an Amazon Kendra data source connector}
\usage{
kendra_describe_data_source(Id, IndexId)
}
\arguments{
\item{Id}{[required] The identifier of the data source connector.}

\item{IndexId}{[required] The identifier of the index used with the data source connector.}
}
\description{
Gets information about an Amazon Kendra data source connector.

See \url{https://www.paws-r-sdk.com/docs/kendra_describe_data_source/} for full documentation.
}
\keyword{internal}
