% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendra_operations.R
\name{kendra_describe_faq}
\alias{kendra_describe_faq}
\title{Gets information about an FAQ list}
\usage{
kendra_describe_faq(Id, IndexId)
}
\arguments{
\item{Id}{[required] The identifier of the FAQ you want to get information on.}

\item{IndexId}{[required] The identifier of the index for the FAQ.}
}
\description{
Gets information about an FAQ list.

See \url{https://www.paws-r-sdk.com/docs/kendra_describe_faq/} for full documentation.
}
\keyword{internal}
