% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendra_operations.R
\name{kendra_describe_index}
\alias{kendra_describe_index}
\title{Gets information about an Amazon Kendra index}
\usage{
kendra_describe_index(Id)
}
\arguments{
\item{Id}{[required] The identifier of the index you want to get information on.}
}
\description{
Gets information about an Amazon Kendra index.

See \url{https://www.paws-r-sdk.com/docs/kendra_describe_index/} for full documentation.
}
\keyword{internal}
