% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendra_operations.R
\name{kendra_describe_thesaurus}
\alias{kendra_describe_thesaurus}
\title{Gets information about an Amazon Kendra thesaurus}
\usage{
kendra_describe_thesaurus(Id, IndexId)
}
\arguments{
\item{Id}{[required] The identifier of the thesaurus you want to get information on.}

\item{IndexId}{[required] The identifier of the index for the thesaurus.}
}
\description{
Gets information about an Amazon Kendra thesaurus.

See \url{https://www.paws-r-sdk.com/docs/kendra_describe_thesaurus/} for full documentation.
}
\keyword{internal}
