% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendra_operations.R
\name{kendra_list_access_control_configurations}
\alias{kendra_list_access_control_configurations}
\title{Lists one or more access control configurations for an index}
\usage{
kendra_list_access_control_configurations(
  IndexId,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{IndexId}{[required] The identifier of the index for the access control configuration.}

\item{NextToken}{If the previous response was incomplete (because there's more data to
retrieve), Amazon Kendra returns a pagination token in the response. You
can use this pagination token to retrieve the next set of access control
configurations.}

\item{MaxResults}{The maximum number of access control configurations to return.}
}
\description{
Lists one or more access control configurations for an index. This includes user and group access information for your documents. This is useful for user context filtering, where search results are filtered based on the user or their group access to documents.

See \url{https://paws-r.github.io/docs/kendra/list_access_control_configurations.html} for full documentation.
}
\keyword{internal}
