% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendra_operations.R
\name{kendra_list_entity_personas}
\alias{kendra_list_entity_personas}
\title{Lists specific permissions of users and groups with access to your
Amazon Kendra experience}
\usage{
kendra_list_entity_personas(Id, IndexId, NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{Id}{[required] The identifier of your Amazon Kendra experience.}

\item{IndexId}{[required] The identifier of the index for your Amazon Kendra experience.}

\item{NextToken}{If the previous response was incomplete (because there is more data to
retrieve), Amazon Kendra returns a pagination token in the response. You
can use this pagination token to retrieve the next set of users or
groups.}

\item{MaxResults}{The maximum number of returned users or groups.}
}
\description{
Lists specific permissions of users and groups with access to your Amazon Kendra experience.

See \url{https://paws-r.github.io/docs/kendra/list_entity_personas.html} for full documentation.
}
\keyword{internal}
