% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendra_operations.R
\name{kendra_list_experiences}
\alias{kendra_list_experiences}
\title{Lists one or more Amazon Kendra experiences}
\usage{
kendra_list_experiences(IndexId, NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{IndexId}{[required] The identifier of the index for your Amazon Kendra experience.}

\item{NextToken}{If the previous response was incomplete (because there is more data to
retrieve), Amazon Kendra returns a pagination token in the response. You
can use this pagination token to retrieve the next set of Amazon Kendra
experiences.}

\item{MaxResults}{The maximum number of returned Amazon Kendra experiences.}
}
\description{
Lists one or more Amazon Kendra experiences. You can create an Amazon Kendra experience such as a search application. For more information on creating a search application experience, see \href{https://docs.aws.amazon.com/kendra/latest/dg/deploying-search-experience-no-code.html}{Building a search experience with no code}.

See \url{https://www.paws-r-sdk.com/docs/kendra_list_experiences/} for full documentation.
}
\keyword{internal}
