% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendra_operations.R
\name{kendra_list_faqs}
\alias{kendra_list_faqs}
\title{Gets a list of FAQs associated with an index}
\usage{
kendra_list_faqs(IndexId, NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{IndexId}{[required] The index for the FAQs.}

\item{NextToken}{If the previous response was incomplete (because there is more data to
retrieve), Amazon Kendra returns a pagination token in the response. You
can use this pagination token to retrieve the next set of FAQs.}

\item{MaxResults}{The maximum number of FAQs to return in the response. If there are fewer
results in the list, this response contains only the actual results.}
}
\description{
Gets a list of FAQs associated with an index.

See \url{https://www.paws-r-sdk.com/docs/kendra_list_faqs/} for full documentation.
}
\keyword{internal}
