% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendra_operations.R
\name{kendra_list_featured_results_sets}
\alias{kendra_list_featured_results_sets}
\title{Lists all your sets of featured results for a given index}
\usage{
kendra_list_featured_results_sets(IndexId, NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{IndexId}{[required] The identifier of the index used for featuring results.}

\item{NextToken}{If the response is truncated, Amazon Kendra returns a pagination token
in the response. You can use this pagination token to retrieve the next
set of featured results sets.}

\item{MaxResults}{The maximum number of featured results sets to return.}
}
\description{
Lists all your sets of featured results for a given index. Features results are placed above all other results for certain queries. If there's an exact match of a query, then one or more specific documents are featured in the search results.

See \url{https://www.paws-r-sdk.com/docs/kendra_list_featured_results_sets/} for full documentation.
}
\keyword{internal}
