% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendra_operations.R
\name{kendra_list_query_suggestions_block_lists}
\alias{kendra_list_query_suggestions_block_lists}
\title{Lists the block lists used for query suggestions for an index}
\usage{
kendra_list_query_suggestions_block_lists(
  IndexId,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{IndexId}{[required] The identifier of the index for a list of all block lists that exist for
that index.

For information on the current quota limits for block lists, see \href{https://docs.aws.amazon.com/kendra/latest/dg/quotas.html}{Quotas for Amazon Kendra}.}

\item{NextToken}{If the previous response was incomplete (because there is more data to
retrieve), Amazon Kendra returns a pagination token in the response. You
can use this pagination token to retrieve the next set of block lists
(\code{BlockListSummaryItems}).}

\item{MaxResults}{The maximum number of block lists to return.}
}
\description{
Lists the block lists used for query suggestions for an index.

See \url{https://www.paws-r-sdk.com/docs/kendra_list_query_suggestions_block_lists/} for full documentation.
}
\keyword{internal}
