% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendra_operations.R
\name{kendra_list_thesauri}
\alias{kendra_list_thesauri}
\title{Lists the thesauri for an index}
\usage{
kendra_list_thesauri(IndexId, NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{IndexId}{[required] The identifier of the index with one or more thesauri.}

\item{NextToken}{If the previous response was incomplete (because there is more data to
retrieve), Amazon Kendra returns a pagination token in the response. You
can use this pagination token to retrieve the next set of thesauri
(\code{ThesaurusSummaryItems}).}

\item{MaxResults}{The maximum number of thesauri to return.}
}
\description{
Lists the thesauri for an index.

See \url{https://www.paws-r-sdk.com/docs/kendra_list_thesauri/} for full documentation.
}
\keyword{internal}
