% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendra_operations.R
\name{kendra_start_data_source_sync_job}
\alias{kendra_start_data_source_sync_job}
\title{Starts a synchronization job for a data source connector}
\usage{
kendra_start_data_source_sync_job(Id, IndexId)
}
\arguments{
\item{Id}{[required] The identifier of the data source connector to synchronize.}

\item{IndexId}{[required] The identifier of the index used with the data source connector.}
}
\description{
Starts a synchronization job for a data source connector. If a synchronization job is already in progress, Amazon Kendra returns a \code{ResourceInUseException} exception.

See \url{https://www.paws-r-sdk.com/docs/kendra_start_data_source_sync_job/} for full documentation.
}
\keyword{internal}
