% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendra_operations.R
\name{kendra_stop_data_source_sync_job}
\alias{kendra_stop_data_source_sync_job}
\title{Stops a synchronization job that is currently running}
\usage{
kendra_stop_data_source_sync_job(Id, IndexId)
}
\arguments{
\item{Id}{[required] The identifier of the data source for which to stop the synchronization
jobs.}

\item{IndexId}{[required] The identifier of the index that contains the data source.}
}
\description{
Stops a synchronization job that is currently running. You can't stop a scheduled synchronization job.

See \url{https://paws-r.github.io/docs/kendra/stop_data_source_sync_job.html} for full documentation.
}
\keyword{internal}
