% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendra_operations.R
\name{kendra_submit_feedback}
\alias{kendra_submit_feedback}
\title{Enables you to provide feedback to Amazon Kendra to improve the
performance of your index}
\usage{
kendra_submit_feedback(
  IndexId,
  QueryId,
  ClickFeedbackItems = NULL,
  RelevanceFeedbackItems = NULL
)
}
\arguments{
\item{IndexId}{[required] The identifier of the index that was queried.}

\item{QueryId}{[required] The identifier of the specific query for which you are submitting
feedback. The query ID is returned in the response to the
\code{\link[=kendra_query]{query}} API.}

\item{ClickFeedbackItems}{Tells Amazon Kendra that a particular search result link was chosen by
the user.}

\item{RelevanceFeedbackItems}{Provides Amazon Kendra with relevant or not relevant feedback for
whether a particular item was relevant to the search.}
}
\description{
Enables you to provide feedback to Amazon Kendra to improve the performance of your index.

See \url{https://www.paws-r-sdk.com/docs/kendra_submit_feedback/} for full documentation.
}
\keyword{internal}
