% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendraranking_operations.R
\name{kendraranking_delete_rescore_execution_plan}
\alias{kendraranking_delete_rescore_execution_plan}
\title{Deletes a rescore execution plan}
\usage{
kendraranking_delete_rescore_execution_plan(Id)
}
\arguments{
\item{Id}{[required] The identifier of the rescore execution plan that you want to delete.}
}
\description{
Deletes a rescore execution plan. A rescore execution plan is an Amazon Kendra Intelligent Ranking resource used for provisioning the \code{\link[=kendraranking_rescore]{rescore}} API.

See \url{https://www.paws-r-sdk.com/docs/kendraranking_delete_rescore_execution_plan/} for full documentation.
}
\keyword{internal}
