% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendraranking_operations.R
\name{kendraranking_list_rescore_execution_plans}
\alias{kendraranking_list_rescore_execution_plans}
\title{Lists your rescore execution plans}
\usage{
kendraranking_list_rescore_execution_plans(NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{NextToken}{If the response is truncated, Amazon Kendra Intelligent Ranking returns
a pagination token in the response. You can use this pagination token to
retrieve the next set of rescore execution plans.}

\item{MaxResults}{The maximum number of rescore execution plans to return.}
}
\description{
Lists your rescore execution plans. A rescore execution plan is an Amazon Kendra Intelligent Ranking resource used for provisioning the \code{\link[=kendraranking_rescore]{rescore}} API.

See \url{https://www.paws-r-sdk.com/docs/kendraranking_list_rescore_execution_plans/} for full documentation.
}
\keyword{internal}
