% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendraranking_operations.R
\name{kendraranking_list_tags_for_resource}
\alias{kendraranking_list_tags_for_resource}
\title{Gets a list of tags associated with a specified resource}
\usage{
kendraranking_list_tags_for_resource(ResourceARN)
}
\arguments{
\item{ResourceARN}{[required] The Amazon Resource Name (ARN) of the rescore execution plan to get a
list of tags for.}
}
\description{
Gets a list of tags associated with a specified resource. A rescore execution plan is an example of a resource that can have tags associated with it.

See \url{https://www.paws-r-sdk.com/docs/kendraranking_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
