% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendraranking_operations.R
\name{kendraranking_rescore}
\alias{kendraranking_rescore}
\title{Rescores or re-ranks search results from a search service such as
OpenSearch (self managed)}
\usage{
kendraranking_rescore(RescoreExecutionPlanId, SearchQuery, Documents)
}
\arguments{
\item{RescoreExecutionPlanId}{[required] The identifier of the rescore execution plan. A rescore execution plan
is an Amazon Kendra Intelligent Ranking resource used for provisioning
the \code{\link[=kendraranking_rescore]{rescore}} API.}

\item{SearchQuery}{[required] The input query from the search service.}

\item{Documents}{[required] The list of documents for Amazon Kendra Intelligent Ranking to rescore
or rank on.}
}
\description{
Rescores or re-ranks search results from a search service such as OpenSearch (self managed). You use the semantic search capabilities of Amazon Kendra Intelligent Ranking to improve the search service's results.

See \url{https://www.paws-r-sdk.com/docs/kendraranking_rescore/} for full documentation.
}
\keyword{internal}
