% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendraranking_operations.R
\name{kendraranking_untag_resource}
\alias{kendraranking_untag_resource}
\title{Removes a tag from a rescore execution plan}
\usage{
kendraranking_untag_resource(ResourceARN, TagKeys)
}
\arguments{
\item{ResourceARN}{[required] The Amazon Resource Name (ARN) of the rescore execution plan to remove
the tag.}

\item{TagKeys}{[required] A list of tag keys to remove from the rescore execution plan. If a tag
key does not exist on the resource, it is ignored.}
}
\description{
Removes a tag from a rescore execution plan. A rescore execution plan is an Amazon Kendra Intelligent Ranking resource used for provisioning the \code{\link[=kendraranking_rescore]{rescore}} operation.

See \url{https://www.paws-r-sdk.com/docs/kendraranking_untag_resource/} for full documentation.
}
\keyword{internal}
