% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendraranking_operations.R
\name{kendraranking_update_rescore_execution_plan}
\alias{kendraranking_update_rescore_execution_plan}
\title{Updates a rescore execution plan}
\usage{
kendraranking_update_rescore_execution_plan(
  Id,
  Name = NULL,
  Description = NULL,
  CapacityUnits = NULL
)
}
\arguments{
\item{Id}{[required] The identifier of the rescore execution plan that you want to update.}

\item{Name}{A new name for the rescore execution plan.}

\item{Description}{A new description for the rescore execution plan.}

\item{CapacityUnits}{You can set additional capacity units to meet the needs of your rescore
execution plan. You are given a single capacity unit by default. If you
want to use the default capacity, you don't set additional capacity
units. For more information on the default capacity and additional
capacity units, see \href{https://docs.aws.amazon.com/kendra/latest/dg/adjusting-capacity.html}{Adjusting capacity}.}
}
\description{
Updates a rescore execution plan. A rescore execution plan is an Amazon Kendra Intelligent Ranking resource used for provisioning the \code{\link[=kendraranking_rescore]{rescore}} API. You can update the number of capacity units you require for Amazon Kendra Intelligent Ranking to rescore or re-rank a search service's results.

See \url{https://www.paws-r-sdk.com/docs/kendraranking_update_rescore_execution_plan/} for full documentation.
}
\keyword{internal}
