% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesis_operations.R
\name{kinesis_add_tags_to_stream}
\alias{kinesis_add_tags_to_stream}
\title{Adds or updates tags for the specified Kinesis data stream}
\usage{
kinesis_add_tags_to_stream(StreamName = NULL, Tags, StreamARN = NULL)
}
\arguments{
\item{StreamName}{The name of the stream.}

\item{Tags}{[required] A set of up to 50 key-value pairs to use to create the tags. A tag
consists of a required key and an optional value. You can add up to 50
tags per resource.}

\item{StreamARN}{The ARN of the stream.}
}
\description{
Adds or updates tags for the specified Kinesis data stream. You can assign up to 50 tags to a data stream.

See \url{https://www.paws-r-sdk.com/docs/kinesis_add_tags_to_stream/} for full documentation.
}
\keyword{internal}
