% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesis_operations.R
\name{kinesis_create_stream}
\alias{kinesis_create_stream}
\title{Creates a Kinesis data stream}
\usage{
kinesis_create_stream(
  StreamName,
  ShardCount = NULL,
  StreamModeDetails = NULL,
  Tags = NULL,
  WarmThroughputMiBps = NULL,
  MaxRecordSizeInKiB = NULL,
  RecordDistributionStrategy = NULL
)
}
\arguments{
\item{StreamName}{[required] A name to identify the stream. The stream name is scoped to the Amazon Web Services account used by the application that creates the stream. It is also scoped by Amazon Web Services Region. That is, two streams in two different Amazon Web Services accounts can have the same name. Two streams in the same Amazon Web Services account but in two different Regions can also have the same name.}

\item{ShardCount}{The number of shards that the stream will use. The throughput of the stream is a function of the number of shards; more shards are required for greater provisioned throughput.}

\item{StreamModeDetails}{Indicates the capacity mode of the data stream. Currently, in Kinesis Data Streams, you can choose between an \strong{on-demand} capacity mode and a \strong{provisioned} capacity mode for your data streams.}

\item{Tags}{A set of up to 50 key-value pairs to use to create the tags. A tag consists of a required key and an optional value.}

\item{WarmThroughputMiBps}{The target warm throughput in MB/s that the stream should be scaled to handle. This represents the throughput capacity that will be immediately available for write operations.}

\item{MaxRecordSizeInKiB}{The maximum record size of a single record in kibibyte (KiB) that you can write to, and read from a stream.}

\item{RecordDistributionStrategy}{The record distribution strategy for the stream, which determines how Amazon Kinesis Data Streams distributes records across shards. Specify one of the following values:
\itemize{
\item \code{AUTO} – Amazon Kinesis Data Streams distributes records evenly across shards and ignores any partition key and \code{ExplicitHashKey} that producers supply. Use this value for stateless workloads that do not require partition-key ordering.
\item \code{USER_PARTITION_KEY} – Producers must supply a partition key, which Amazon Kinesis Data Streams uses to determine shard placement. This is the default.
}

The record distribution strategy is only supported for streams that use the on-demand capacity mode. If you do not specify this parameter, the stream uses \code{USER_PARTITION_KEY}.}
}
\description{
Creates a Kinesis data stream. A stream captures and transports data records that are continuously emitted from different data sources or \emph{producers}. Scale-out within a stream is explicitly supported by means of shards, which are uniquely identified groups of data records in a stream.

See \url{https://www.paws-r-sdk.com/docs/kinesis_create_stream/} for full documentation.
}
\keyword{internal}
