% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesis_operations.R
\name{kinesis_delete_channel}
\alias{kinesis_delete_channel}
\title{Deletes the specified channel}
\usage{
kinesis_delete_channel(ChannelARN)
}
\arguments{
\item{ChannelARN}{[required] The Amazon Resource Name (ARN) of the channel to delete.}
}
\description{
Deletes the specified channel. Deleting a channel stops delivery from the source stream to the destination. Data already delivered to the destination is not deleted.

See \url{https://www.paws-r-sdk.com/docs/kinesis_delete_channel/} for full documentation.
}
\keyword{internal}
