% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesis_operations.R
\name{kinesis_delete_resource_policy}
\alias{kinesis_delete_resource_policy}
\title{Delete a policy for the specified data stream or consumer}
\usage{
kinesis_delete_resource_policy(ResourceARN)
}
\arguments{
\item{ResourceARN}{[required] The Amazon Resource Name (ARN) of the data stream or consumer.}
}
\description{
Delete a policy for the specified data stream or consumer. Request patterns can be one of the following:

See \url{https://www.paws-r-sdk.com/docs/kinesis_delete_resource_policy/} for full documentation.
}
\keyword{internal}
