% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesis_operations.R
\name{kinesis_describe_account_settings}
\alias{kinesis_describe_account_settings}
\title{Describes the account-level settings for Amazon Kinesis Data Streams}
\usage{
kinesis_describe_account_settings()
}
\description{
Describes the account-level settings for Amazon Kinesis Data Streams. This operation returns information about the minimum throughput billing commitments and other account-level configurations.

See \url{https://www.paws-r-sdk.com/docs/kinesis_describe_account_settings/} for full documentation.
}
\keyword{internal}
