% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesis_operations.R
\name{kinesis_describe_channel}
\alias{kinesis_describe_channel}
\title{Describes the specified channel, including its configuration and current
status}
\usage{
kinesis_describe_channel(ChannelARN)
}
\arguments{
\item{ChannelARN}{[required] The Amazon Resource Name (ARN) of the channel to describe.}
}
\description{
Describes the specified channel, including its configuration and current status.

See \url{https://www.paws-r-sdk.com/docs/kinesis_describe_channel/} for full documentation.
}
\keyword{internal}
