% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesis_operations.R
\name{kinesis_describe_stream_summary}
\alias{kinesis_describe_stream_summary}
\title{Provides a summarized description of the specified Kinesis data stream
without the shard list}
\usage{
kinesis_describe_stream_summary(StreamName)
}
\arguments{
\item{StreamName}{[required] The name of the stream to describe.}
}
\description{
Provides a summarized description of the specified Kinesis data stream without the shard list.

See \url{https://paws-r.github.io/docs/kinesis/describe_stream_summary.html} for full documentation.
}
\keyword{internal}
