% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesis_operations.R
\name{kinesis_get_records}
\alias{kinesis_get_records}
\title{Gets data records from a Kinesis data stream's shard}
\usage{
kinesis_get_records(
  ShardIterator,
  Limit = NULL,
  StreamARN = NULL,
  StreamId = NULL,
  DryRun = NULL
)
}
\arguments{
\item{ShardIterator}{[required] The position in the shard from which you want to start sequentially reading data records. A shard iterator specifies this position using the sequence number of a data record in the shard.}

\item{Limit}{The maximum number of records to return. Specify a value of up to 10,000. If you specify a value that is greater than 10,000, \code{\link[=kinesis_get_records]{get_records}} throws \code{InvalidArgumentException}. The default value is 10,000.}

\item{StreamARN}{The ARN of the stream.}

\item{StreamId}{Not Implemented. Reserved for future use.}

\item{DryRun}{Checks if your request will succeed. \code{DryRun} is an optional parameter.}
}
\description{
Gets data records from a Kinesis data stream's shard.

See \url{https://www.paws-r-sdk.com/docs/kinesis_get_records/} for full documentation.
}
\keyword{internal}
