% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesis_operations.R
\name{kinesis_list_channels}
\alias{kinesis_list_channels}
\title{Lists the channels in your account}
\usage{
kinesis_list_channels(StreamFilter = NULL, MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{StreamFilter}{Filters the results to channels associated with the specified streams.}

\item{MaxResults}{The maximum number of channels to return in a single call. The default value is 100. If you specify a value greater than 100, at most 100 results are returned.}

\item{NextToken}{The pagination token returned by a previous call. Specify this token to retrieve the next page of results.}
}
\description{
Lists the channels in your account. You can filter the results by source stream. The results are paginated. Use the \code{NextToken} value returned in the response to retrieve additional results.

See \url{https://www.paws-r-sdk.com/docs/kinesis_list_channels/} for full documentation.
}
\keyword{internal}
