% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesis_operations.R
\name{kinesis_list_tags_for_resource}
\alias{kinesis_list_tags_for_resource}
\title{List all tags added to the specified Kinesis resource}
\usage{
kinesis_list_tags_for_resource(ResourceARN)
}
\arguments{
\item{ResourceARN}{[required] The Amazon Resource Name (ARN) of the Kinesis resource for which to list
tags.}
}
\description{
List all tags added to the specified Kinesis resource. Each tag is a label consisting of a user-defined key and value. Tags can help you manage, identify, organize, search for, and filter resources.

See \url{https://www.paws-r-sdk.com/docs/kinesis_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
