% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesis_operations.R
\name{kinesis_list_tags_for_stream}
\alias{kinesis_list_tags_for_stream}
\title{Lists the tags for the specified Kinesis data stream}
\usage{
kinesis_list_tags_for_stream(
  StreamName = NULL,
  ExclusiveStartTagKey = NULL,
  Limit = NULL,
  StreamARN = NULL
)
}
\arguments{
\item{StreamName}{The name of the stream.}

\item{ExclusiveStartTagKey}{The key to use as the starting point for the list of tags. If this
parameter is set, \code{\link[=kinesis_list_tags_for_stream]{list_tags_for_stream}}
gets all tags that occur after \code{ExclusiveStartTagKey}.}

\item{Limit}{The number of tags to return. If this number is less than the total
number of tags associated with the stream, \code{HasMoreTags} is set to
\code{true}. To list additional tags, set \code{ExclusiveStartTagKey} to the last
key in the response.}

\item{StreamARN}{The ARN of the stream.}
}
\description{
Lists the tags for the specified Kinesis data stream. This operation has a limit of five transactions per second per account.

See \url{https://www.paws-r-sdk.com/docs/kinesis_list_tags_for_stream/} for full documentation.
}
\keyword{internal}
