% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesis_operations.R
\name{kinesis_update_account_settings}
\alias{kinesis_update_account_settings}
\title{Updates the account-level settings for Amazon Kinesis Data Streams}
\usage{
kinesis_update_account_settings(MinimumThroughputBillingCommitment)
}
\arguments{
\item{MinimumThroughputBillingCommitment}{[required] Specifies the minimum throughput billing commitment configuration for
your account.}
}
\description{
Updates the account-level settings for Amazon Kinesis Data Streams.

See \url{https://www.paws-r-sdk.com/docs/kinesis_update_account_settings/} for full documentation.
}
\keyword{internal}
