% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesis_operations.R
\name{kinesis_update_channel}
\alias{kinesis_update_channel}
\title{Updates the data freshness interval or the Amazon CloudWatch Logs
configuration of an existing channel}
\usage{
kinesis_update_channel(
  ChannelARN,
  S3DestinationConfiguration = NULL,
  S3TablesDestinationConfiguration = NULL,
  LoggingConfiguration = NULL
)
}
\arguments{
\item{ChannelARN}{[required] The Amazon Resource Name (ARN) of the channel to update.}

\item{S3DestinationConfiguration}{The updated configuration for a general purpose Amazon S3 destination. Specify this parameter when the channel delivers to a general purpose Amazon S3 bucket. Only \code{DataFreshnessInSeconds} can be updated.}

\item{S3TablesDestinationConfiguration}{The updated configuration for a streaming table destination. Specify this parameter when the channel delivers to streaming tables on Apache Iceberg in Amazon S3 Tables. Only \code{DataFreshnessInSeconds} can be updated.}

\item{LoggingConfiguration}{The updated Amazon CloudWatch Logs configuration for the channel.}
}
\description{
Updates the data freshness interval or the Amazon CloudWatch Logs configuration of an existing channel. You cannot change the destination, source stream, record format, schema, encryption configuration, or service execution role of an existing channel. To change any other setting, delete the channel and create a new one.

See \url{https://www.paws-r-sdk.com/docs/kinesis_update_channel/} for full documentation.
}
\keyword{internal}
