% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesis_operations.R
\name{kinesis_update_stream_mode}
\alias{kinesis_update_stream_mode}
\title{Updates the capacity mode of the data stream}
\usage{
kinesis_update_stream_mode(StreamARN, StreamModeDetails)
}
\arguments{
\item{StreamARN}{[required] Specifies the ARN of the data stream whose capacity mode you want to
update.}

\item{StreamModeDetails}{[required] Specifies the capacity mode to which you want to set your data stream.
Currently, in Kinesis Data Streams, you can choose between an
\strong{on-demand} capacity mode and a \strong{provisioned} capacity mode for your
data streams.}
}
\description{
Updates the capacity mode of the data stream. Currently, in Kinesis Data Streams, you can choose between an \strong{on-demand} capacity mode and a \strong{provisioned} capacity mode for your data stream.

See \url{https://paws-r.github.io/docs/kinesis/update_stream_mode.html} for full documentation.
}
\keyword{internal}
