% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesis_operations.R
\name{kinesis_update_stream_record_distribution_strategy}
\alias{kinesis_update_stream_record_distribution_strategy}
\title{Updates the record distribution strategy for the specified Amazon
Kinesis Data Streams on-demand data stream}
\usage{
kinesis_update_stream_record_distribution_strategy(
  StreamARN,
  StreamId = NULL,
  RecordDistributionStrategy
)
}
\arguments{
\item{StreamARN}{[required] The Amazon Resource Name (ARN) of the stream to update.}

\item{StreamId}{Not Implemented. Reserved for future use.}

\item{RecordDistributionStrategy}{[required] The record distribution strategy to apply to the stream. Specify one of the following values:
\itemize{
\item \code{AUTO} – Amazon Kinesis Data Streams distributes records evenly across shards and ignores any partition key and \code{ExplicitHashKey} that producers supply.
\item \code{USER_PARTITION_KEY} – Producers must supply a partition key, which Amazon Kinesis Data Streams uses to determine shard placement. This is the default.
}}
}
\description{
Updates the record distribution strategy for the specified Amazon Kinesis Data Streams on-demand data stream. The record distribution strategy determines how Amazon Kinesis Data Streams distributes records across the shards in a stream.

See \url{https://www.paws-r-sdk.com/docs/kinesis_update_stream_record_distribution_strategy/} for full documentation.
}
\keyword{internal}
