% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesis_operations.R
\name{kinesis_update_stream_warm_throughput}
\alias{kinesis_update_stream_warm_throughput}
\title{Updates the warm throughput configuration for the specified Amazon
Kinesis Data Streams on-demand data stream}
\usage{
kinesis_update_stream_warm_throughput(
  StreamARN = NULL,
  StreamName = NULL,
  StreamId = NULL,
  WarmThroughputMiBps
)
}
\arguments{
\item{StreamARN}{The ARN of the stream to be updated.}

\item{StreamName}{The name of the stream to be updated.}

\item{StreamId}{Not Implemented. Reserved for future use.}

\item{WarmThroughputMiBps}{[required] The target warm throughput in MB/s that the stream should be scaled to handle. This represents the throughput capacity that will be immediately available for write operations.}
}
\description{
Updates the warm throughput configuration for the specified Amazon Kinesis Data Streams on-demand data stream. Updates the warm throughput configuration for the specified on-demand data stream. Use this operation to scale your stream to a specified throughput level before anticipated traffic spikes, or to release excess capacity after traffic has decreased.

See \url{https://www.paws-r-sdk.com/docs/kinesis_update_stream_warm_throughput/} for full documentation.
}
\keyword{internal}
