% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisanalytics_operations.R
\name{kinesisanalytics_delete_application_cloud_watch_logging_option}
\alias{kinesisanalytics_delete_application_cloud_watch_logging_option}
\title{This documentation is for version 1 of the Amazon Kinesis Data Analytics
API, which only supports SQL applications}
\usage{
kinesisanalytics_delete_application_cloud_watch_logging_option(
  ApplicationName,
  CurrentApplicationVersionId,
  CloudWatchLoggingOptionId
)
}
\arguments{
\item{ApplicationName}{[required] The Kinesis Analytics application name.}

\item{CurrentApplicationVersionId}{[required] The version ID of the Kinesis Analytics application.}

\item{CloudWatchLoggingOptionId}{[required] The \code{CloudWatchLoggingOptionId} of the CloudWatch logging option to
delete. You can get the \code{CloudWatchLoggingOptionId} by using the
\code{\link[=kinesisanalytics_describe_application]{describe_application}}
operation.}
}
\description{
This documentation is for version 1 of the Amazon Kinesis Data Analytics API, which only supports SQL applications. Version 2 of the API supports SQL and Java applications. For more information about version 2, see Amazon Kinesis Data Analytics API V2 Documentation.

See \url{https://www.paws-r-sdk.com/docs/kinesisanalytics_delete_application_cloud_watch_logging_option/} for full documentation.
}
\keyword{internal}
