% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisanalytics_operations.R
\name{kinesisanalytics_describe_application}
\alias{kinesisanalytics_describe_application}
\title{This documentation is for version 1 of the Amazon Kinesis Data Analytics
API, which only supports SQL applications}
\usage{
kinesisanalytics_describe_application(ApplicationName)
}
\arguments{
\item{ApplicationName}{[required] Name of the application.}
}
\description{
This documentation is for version 1 of the Amazon Kinesis Data Analytics
API, which only supports SQL applications. Version 2 of the API supports
SQL and Java applications. For more information about version 2, see
Amazon Kinesis Data Analytics API V2 Documentation.

Returns information about a specific Amazon Kinesis Analytics
application.

If you want to retrieve a list of all applications in your account, use
the
\href{https://docs.aws.amazon.com/kinesisanalytics/latest/dev/API_ListApplications.html}{ListApplications}
operation.

This operation requires permissions to perform the
\code{kinesisanalytics:DescribeApplication} action. You can use
\code{DescribeApplication} to get the current application versionId, which
you need to call other operations such as \code{Update}.
}
\section{Request syntax}{
\preformatted{svc$describe_application(
  ApplicationName = "string"
)
}
}

\keyword{internal}
