% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisanalytics_operations.R
\name{kinesisanalytics_discover_input_schema}
\alias{kinesisanalytics_discover_input_schema}
\title{This documentation is for version 1 of the Amazon Kinesis Data Analytics
API, which only supports SQL applications}
\usage{
kinesisanalytics_discover_input_schema(
  ResourceARN = NULL,
  RoleARN = NULL,
  InputStartingPositionConfiguration = NULL,
  S3Configuration = NULL,
  InputProcessingConfiguration = NULL
)
}
\arguments{
\item{ResourceARN}{Amazon Resource Name (ARN) of the streaming source.}

\item{RoleARN}{ARN of the IAM role that Amazon Kinesis Analytics can assume to access
the stream on your behalf.}

\item{InputStartingPositionConfiguration}{Point at which you want Amazon Kinesis Analytics to start reading
records from the specified streaming source discovery purposes.}

\item{S3Configuration}{Specify this parameter to discover a schema from data in an Amazon S3
object.}

\item{InputProcessingConfiguration}{The
\href{https://docs.aws.amazon.com/kinesisanalytics/latest/dev/API_InputProcessingConfiguration.html}{InputProcessingConfiguration}
to use to preprocess the records before discovering the schema of the
records.}
}
\description{
This documentation is for version 1 of the Amazon Kinesis Data Analytics API, which only supports SQL applications. Version 2 of the API supports SQL and Java applications. For more information about version 2, see Amazon Kinesis Data Analytics API V2 Documentation.

See \url{https://paws-r.github.io/docs/kinesisanalytics/discover_input_schema.html} for full documentation.
}
\keyword{internal}
