% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisanalytics_operations.R
\name{kinesisanalytics_update_application}
\alias{kinesisanalytics_update_application}
\title{This documentation is for version 1 of the Amazon Kinesis Data Analytics
API, which only supports SQL applications}
\usage{
kinesisanalytics_update_application(
  ApplicationName,
  CurrentApplicationVersionId,
  ApplicationUpdate
)
}
\arguments{
\item{ApplicationName}{[required] Name of the Amazon Kinesis Analytics application to update.}

\item{CurrentApplicationVersionId}{[required] The current application version ID. You can use the
\code{\link[=kinesisanalytics_describe_application]{describe_application}}
operation to get this value.}

\item{ApplicationUpdate}{[required] Describes application updates.}
}
\description{
This documentation is for version 1 of the Amazon Kinesis Data Analytics API, which only supports SQL applications. Version 2 of the API supports SQL and Java applications. For more information about version 2, see Amazon Kinesis Data Analytics API V2 Documentation.

See \url{https://www.paws-r-sdk.com/docs/kinesisanalytics_update_application/} for full documentation.
}
\keyword{internal}
