% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisanalyticsv2_operations.R
\name{kinesisanalyticsv2_add_application_input}
\alias{kinesisanalyticsv2_add_application_input}
\title{Adds a streaming source to your SQL-based Kinesis Data Analytics
application}
\usage{
kinesisanalyticsv2_add_application_input(
  ApplicationName,
  CurrentApplicationVersionId,
  Input
)
}
\arguments{
\item{ApplicationName}{[required] The name of your existing application to which you want to add the
streaming source.}

\item{CurrentApplicationVersionId}{[required] The current version of your application. You must provide the
\code{ApplicationVersionID} or the \code{ConditionalToken}.You can use the
\code{\link[=kinesisanalyticsv2_describe_application]{describe_application}}
operation to find the current application version.}

\item{Input}{[required] The Input to add.}
}
\description{
Adds a streaming source to your SQL-based Kinesis Data Analytics application.

See \url{https://www.paws-r-sdk.com/docs/kinesisanalyticsv2_add_application_input/} for full documentation.
}
\keyword{internal}
