% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisanalyticsv2_operations.R
\name{kinesisanalyticsv2_create_application_presigned_url}
\alias{kinesisanalyticsv2_create_application_presigned_url}
\title{Creates and returns a URL that you can use to connect to an
application's extension}
\usage{
kinesisanalyticsv2_create_application_presigned_url(
  ApplicationName,
  UrlType,
  SessionExpirationDurationInSeconds = NULL
)
}
\arguments{
\item{ApplicationName}{[required] The name of the application.}

\item{UrlType}{[required] The type of the extension for which to create and return a URL.
Currently, the only valid extension URL type is \code{FLINK_DASHBOARD_URL}.}

\item{SessionExpirationDurationInSeconds}{The duration in seconds for which the returned URL will be valid.}
}
\description{
Creates and returns a URL that you can use to connect to an application's extension.

See \url{https://paws-r.github.io/docs/kinesisanalyticsv2/create_application_presigned_url.html} for full documentation.
}
\keyword{internal}
