% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisanalyticsv2_operations.R
\name{kinesisanalyticsv2_delete_application}
\alias{kinesisanalyticsv2_delete_application}
\title{Deletes the specified application}
\usage{
kinesisanalyticsv2_delete_application(ApplicationName, CreateTimestamp)
}
\arguments{
\item{ApplicationName}{[required] The name of the application to delete.}

\item{CreateTimestamp}{[required] Use the
\code{\link[=kinesisanalyticsv2_describe_application]{describe_application}}
operation to get this value.}
}
\value{
An empty list.
}
\description{
Deletes the specified application. Kinesis Data Analytics halts
application execution and deletes the application.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_application(
  ApplicationName = "string",
  CreateTimestamp = as.POSIXct(
    "2015-01-01"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
