% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisanalyticsv2_operations.R
\name{kinesisanalyticsv2_delete_application_cloud_watch_logging_option}
\alias{kinesisanalyticsv2_delete_application_cloud_watch_logging_option}
\title{Deletes an Amazon CloudWatch log stream from an Kinesis Data Analytics
application}
\usage{
kinesisanalyticsv2_delete_application_cloud_watch_logging_option(
  ApplicationName, CurrentApplicationVersionId, CloudWatchLoggingOptionId,
  ConditionalToken)
}
\arguments{
\item{ApplicationName}{[required] The application name.}

\item{CurrentApplicationVersionId}{The version ID of the application. You must provide the
\code{ApplicationVersionID} or the \code{ConditionalToken}. You can retrieve the
application version ID using
\code{\link[=kinesisanalyticsv2_describe_application]{describe_application}}.}

\item{CloudWatchLoggingOptionId}{[required] The \code{CloudWatchLoggingOptionId} of the Amazon CloudWatch logging option
to delete. You can get the \code{CloudWatchLoggingOptionId} by using the
\code{\link[=kinesisanalyticsv2_describe_application]{describe_application}}
operation.}

\item{ConditionalToken}{A value you use to implement strong concurrency for application updates.
You must provide the \code{ApplicationVersionID} or the \code{ConditionalToken}.
You get the application's current \code{ConditionalToken} using
\code{\link[=kinesisanalyticsv2_describe_application]{describe_application}}.}
}
\value{
A list with the following syntax:\preformatted{list(
  ApplicationARN = "string",
  ApplicationVersionId = 123,
  CloudWatchLoggingOptionDescriptions = list(
    list(
      CloudWatchLoggingOptionId = "string",
      LogStreamARN = "string",
      RoleARN = "string"
    )
  )
)
}
}
\description{
Deletes an Amazon CloudWatch log stream from an Kinesis Data Analytics
application.
}
\section{Request syntax}{
\preformatted{svc$delete_application_cloud_watch_logging_option(
  ApplicationName = "string",
  CurrentApplicationVersionId = 123,
  CloudWatchLoggingOptionId = "string",
  ConditionalToken = "string"
)
}
}

\keyword{internal}
