% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisanalyticsv2_operations.R
\name{kinesisanalyticsv2_describe_application}
\alias{kinesisanalyticsv2_describe_application}
\title{Returns information about a specific Kinesis Data Analytics application}
\usage{
kinesisanalyticsv2_describe_application(
  ApplicationName,
  IncludeAdditionalDetails = NULL
)
}
\arguments{
\item{ApplicationName}{[required] The name of the application.}

\item{IncludeAdditionalDetails}{Displays verbose information about a Kinesis Data Analytics application,
including the application's job plan.}
}
\description{
Returns information about a specific Kinesis Data Analytics application.

See \url{https://paws-r.github.io/docs/kinesisanalyticsv2/describe_application.html} for full documentation.
}
\keyword{internal}
