% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisanalyticsv2_operations.R
\name{kinesisanalyticsv2_describe_application_operation}
\alias{kinesisanalyticsv2_describe_application_operation}
\title{Returns information about a specific operation performed on a Managed
Service for Apache Flink application}
\usage{
kinesisanalyticsv2_describe_application_operation(ApplicationName, OperationId)
}
\arguments{
\item{ApplicationName}{[required]}

\item{OperationId}{[required]}
}
\description{
Returns information about a specific operation performed on a Managed Service for Apache Flink application

See \url{https://www.paws-r-sdk.com/docs/kinesisanalyticsv2_describe_application_operation/} for full documentation.
}
\keyword{internal}
