% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisanalyticsv2_operations.R
\name{kinesisanalyticsv2_describe_application_snapshot}
\alias{kinesisanalyticsv2_describe_application_snapshot}
\title{Returns information about a snapshot of application state data}
\usage{
kinesisanalyticsv2_describe_application_snapshot(ApplicationName, SnapshotName)
}
\arguments{
\item{ApplicationName}{[required] The name of an existing application.}

\item{SnapshotName}{[required] The identifier of an application snapshot. You can retrieve this value
using .}
}
\description{
Returns information about a snapshot of application state data.

See \url{https://www.paws-r-sdk.com/docs/kinesisanalyticsv2_describe_application_snapshot/} for full documentation.
}
\keyword{internal}
