% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisanalyticsv2_operations.R
\name{kinesisanalyticsv2_describe_application_version}
\alias{kinesisanalyticsv2_describe_application_version}
\title{Provides a detailed description of a specified version of the
application}
\usage{
kinesisanalyticsv2_describe_application_version(
  ApplicationName,
  ApplicationVersionId
)
}
\arguments{
\item{ApplicationName}{[required] The name of the application for which you want to get the version
description.}

\item{ApplicationVersionId}{[required] The ID of the application version for which you want to get the
description.}
}
\description{
Provides a detailed description of a specified version of the application. To see a list of all the versions of an application, invoke the \code{\link[=kinesisanalyticsv2_list_application_versions]{list_application_versions}} operation.

See \url{https://paws-r.github.io/docs/kinesisanalyticsv2/describe_application_version.html} for full documentation.
}
\keyword{internal}
