% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisanalyticsv2_operations.R
\name{kinesisanalyticsv2_list_application_operations}
\alias{kinesisanalyticsv2_list_application_operations}
\title{Lists all the operations performed for the specified application such as
UpdateApplication, StartApplication etc}
\usage{
kinesisanalyticsv2_list_application_operations(
  ApplicationName,
  Limit = NULL,
  NextToken = NULL,
  Operation = NULL,
  OperationStatus = NULL
)
}
\arguments{
\item{ApplicationName}{[required] The name of the application.}

\item{Limit}{The limit on the number of records to be returned in the response.}

\item{NextToken}{A pagination token that can be used in a subsequent request.}

\item{Operation}{The type of operation that is performed on an application.}

\item{OperationStatus}{The status of the operation.}
}
\description{
Lists all the operations performed for the specified application such as UpdateApplication, StartApplication etc. The response also includes a summary of the operation.

See \url{https://www.paws-r-sdk.com/docs/kinesisanalyticsv2_list_application_operations/} for full documentation.
}
\keyword{internal}
