% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisanalyticsv2_operations.R
\name{kinesisanalyticsv2_list_application_versions}
\alias{kinesisanalyticsv2_list_application_versions}
\title{Lists all the versions for the specified application, including versions
that were rolled back}
\usage{
kinesisanalyticsv2_list_application_versions(ApplicationName, Limit,
  NextToken)
}
\arguments{
\item{ApplicationName}{[required] The name of the application for which you want to list all versions.}

\item{Limit}{The maximum number of versions to list in this invocation of the
operation.}

\item{NextToken}{If a previous invocation of this operation returned a pagination token,
pass it into this value to retrieve the next set of results. For more
information about pagination, see \href{https://docs.aws.amazon.com/cli/latest/userguide/cli-usage-pagination.html}{Using the AWS Command Line Interface's Pagination Options}.}
}
\value{
A list with the following syntax:\preformatted{list(
  ApplicationVersionSummaries = list(
    list(
      ApplicationVersionId = 123,
      ApplicationStatus = "DELETING"|"STARTING"|"STOPPING"|"READY"|"RUNNING"|"UPDATING"|"AUTOSCALING"|"FORCE_STOPPING"|"MAINTENANCE"|"ROLLING_BACK"|"ROLLED_BACK"
    )
  ),
  NextToken = "string"
)
}
}
\description{
Lists all the versions for the specified application, including versions
that were rolled back. The response also includes a summary of the
configuration associated with each version.

To get the complete description of a specific application version,
invoke the
\code{\link[=kinesisanalyticsv2_describe_application_version]{describe_application_version}}
operation.

This operation is supported only for Amazon Kinesis Data Analytics for
Apache Flink.
}
\section{Request syntax}{
\preformatted{svc$list_application_versions(
  ApplicationName = "string",
  Limit = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
