% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisanalyticsv2_operations.R
\name{kinesisanalyticsv2_list_tags_for_resource}
\alias{kinesisanalyticsv2_list_tags_for_resource}
\title{Retrieves the list of key-value tags assigned to the application}
\usage{
kinesisanalyticsv2_list_tags_for_resource(ResourceARN)
}
\arguments{
\item{ResourceARN}{[required] The ARN of the application for which to retrieve tags.}
}
\description{
Retrieves the list of key-value tags assigned to the application.
}
\section{Request syntax}{
\preformatted{svc$list_tags_for_resource(
  ResourceARN = "string"
)
}
}

\keyword{internal}
