% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisanalyticsv2_operations.R
\name{kinesisanalyticsv2_start_application}
\alias{kinesisanalyticsv2_start_application}
\title{Starts the specified Kinesis Data Analytics application}
\usage{
kinesisanalyticsv2_start_application(ApplicationName, RunConfiguration = NULL)
}
\arguments{
\item{ApplicationName}{[required] The name of the application.}

\item{RunConfiguration}{Identifies the run configuration (start parameters) of a Kinesis Data
Analytics application.}
}
\description{
Starts the specified Kinesis Data Analytics application. After creating an application, you must exclusively call this operation to start your application.

See \url{https://paws-r.github.io/docs/kinesisanalyticsv2/start_application.html} for full documentation.
}
\keyword{internal}
