% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisanalyticsv2_operations.R
\name{kinesisanalyticsv2_update_application_maintenance_configuration}
\alias{kinesisanalyticsv2_update_application_maintenance_configuration}
\title{Updates the maintenance configuration of the Kinesis Data Analytics
application}
\usage{
kinesisanalyticsv2_update_application_maintenance_configuration(
  ApplicationName, ApplicationMaintenanceConfigurationUpdate)
}
\arguments{
\item{ApplicationName}{[required] The name of the application for which you want to update the maintenance
configuration.}

\item{ApplicationMaintenanceConfigurationUpdate}{[required] Describes the application maintenance configuration update.}
}
\value{
A list with the following syntax:\preformatted{list(
  ApplicationARN = "string",
  ApplicationMaintenanceConfigurationDescription = list(
    ApplicationMaintenanceWindowStartTime = "string",
    ApplicationMaintenanceWindowEndTime = "string"
  )
)
}
}
\description{
Updates the maintenance configuration of the Kinesis Data Analytics
application.

You can invoke this operation on an application that is in one of the
two following states: \code{READY} or \code{RUNNING}. If you invoke it when the
application is in a state other than these two states, it throws a
\code{ResourceInUseException}. The service makes use of the updated
configuration the next time it schedules maintenance for the
application. If you invoke this operation after the service schedules
maintenance, the service will apply the configuration update the next
time it schedules maintenance for the application. This means that you
might not see the maintenance configuration update applied to the
maintenance process that follows a successful invocation of this
operation, but to the following maintenance process instead.

To see the current maintenance configuration of your application, invoke
the \code{\link[=kinesisanalyticsv2_describe_application]{describe_application}}
operation.

For information about application maintenance, see \href{https://docs.aws.amazon.com/kinesisanalytics/latest/java/maintenance.html}{Kinesis Data Analytics for Apache Flink Maintenance}.

This operation is supported only for Amazon Kinesis Data Analytics for
Apache Flink.
}
\section{Request syntax}{
\preformatted{svc$update_application_maintenance_configuration(
  ApplicationName = "string",
  ApplicationMaintenanceConfigurationUpdate = list(
    ApplicationMaintenanceWindowStartTimeUpdate = "string"
  )
)
}
}

\keyword{internal}
