% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mturk_operations.R
\name{mturk_approve_assignment}
\alias{mturk_approve_assignment}
\title{The ApproveAssignment operation approves the results of a completed
assignment}
\usage{
mturk_approve_assignment(
  AssignmentId,
  RequesterFeedback = NULL,
  OverrideRejection = NULL
)
}
\arguments{
\item{AssignmentId}{[required] The ID of the assignment. The assignment must correspond to a HIT
created by the Requester.}

\item{RequesterFeedback}{A message for the Worker, which the Worker can see in the Status section
of the web site.}

\item{OverrideRejection}{A flag indicating that an assignment should be approved even if it was
previously rejected. Defaults to \code{False}.}
}
\description{
The \code{\link[=mturk_approve_assignment]{approve_assignment}} operation approves the results of a completed assignment.

See \url{https://paws-r.github.io/docs/mturk/approve_assignment.html} for full documentation.
}
\keyword{internal}
