% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mturk_operations.R
\name{mturk_create_additional_assignments_for_hit}
\alias{mturk_create_additional_assignments_for_hit}
\title{The CreateAdditionalAssignmentsForHIT operation increases the maximum
number of assignments of an existing HIT}
\usage{
mturk_create_additional_assignments_for_hit(
  HITId,
  NumberOfAdditionalAssignments,
  UniqueRequestToken = NULL
)
}
\arguments{
\item{HITId}{[required] The ID of the HIT to extend.}

\item{NumberOfAdditionalAssignments}{[required] The number of additional assignments to request for this HIT.}

\item{UniqueRequestToken}{A unique identifier for this request, which allows you to retry the call
on error without extending the HIT multiple times. This is useful in
cases such as network timeouts where it is unclear whether or not the
call succeeded on the server. If the extend HIT already exists in the
system from a previous call using the same \code{UniqueRequestToken},
subsequent calls will return an error with a message containing the
request ID.}
}
\description{
The \code{\link[=mturk_create_additional_assignments_for_hit]{create_additional_assignments_for_hit}} operation increases the maximum number of assignments of an existing HIT.

See \url{https://www.paws-r-sdk.com/docs/mturk_create_additional_assignments_for_hit/} for full documentation.
}
\keyword{internal}
