% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mturk_operations.R
\name{mturk_delete_hit}
\alias{mturk_delete_hit}
\title{The DeleteHIT operation is used to delete HIT that is no longer needed}
\usage{
mturk_delete_hit(HITId)
}
\arguments{
\item{HITId}{[required] The ID of the HIT to be deleted.}
}
\description{
The \code{\link[=mturk_delete_hit]{delete_hit}} operation is used to delete HIT that is no longer needed. Only the Requester who created the HIT can delete it.

See \url{https://paws-r.github.io/docs/mturk/delete_hit.html} for full documentation.
}
\keyword{internal}
