% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mturk_operations.R
\name{mturk_delete_qualification_type}
\alias{mturk_delete_qualification_type}
\title{The DeleteQualificationType deletes a Qualification type and deletes any
HIT types that are associated with the Qualification type}
\usage{
mturk_delete_qualification_type(QualificationTypeId)
}
\arguments{
\item{QualificationTypeId}{[required] The ID of the QualificationType to dispose.}
}
\description{
The \code{\link[=mturk_delete_qualification_type]{delete_qualification_type}} deletes a Qualification type and deletes any HIT types that are associated with the Qualification type.

See \url{https://paws-r.github.io/docs/mturk/delete_qualification_type.html} for full documentation.
}
\keyword{internal}
