% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mturk_operations.R
\name{mturk_get_assignment}
\alias{mturk_get_assignment}
\title{The GetAssignment operation retrieves the details of the specified
Assignment}
\usage{
mturk_get_assignment(AssignmentId)
}
\arguments{
\item{AssignmentId}{[required] The ID of the Assignment to be retrieved.}
}
\description{
The \code{\link[=mturk_get_assignment]{get_assignment}} operation retrieves the details of the specified Assignment.

See \url{https://paws-r.github.io/docs/mturk/get_assignment.html} for full documentation.
}
\keyword{internal}
