% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mturk_operations.R
\name{mturk_list_assignments_for_hit}
\alias{mturk_list_assignments_for_hit}
\title{The ListAssignmentsForHIT operation retrieves completed assignments for
a HIT}
\usage{
mturk_list_assignments_for_hit(
  HITId,
  NextToken = NULL,
  MaxResults = NULL,
  AssignmentStatuses = NULL
)
}
\arguments{
\item{HITId}{[required] The ID of the HIT.}

\item{NextToken}{Pagination token}

\item{MaxResults}{}

\item{AssignmentStatuses}{The status of the assignments to return: Submitted | Approved | Rejected}
}
\description{
The \code{\link[=mturk_list_assignments_for_hit]{list_assignments_for_hit}} operation retrieves completed assignments for a HIT. You can use this operation to retrieve the results for a HIT.

See \url{https://www.paws-r-sdk.com/docs/mturk_list_assignments_for_hit/} for full documentation.
}
\keyword{internal}
