% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mturk_operations.R
\name{mturk_list_workers_with_qualification_type}
\alias{mturk_list_workers_with_qualification_type}
\title{The ListWorkersWithQualificationType operation returns all of the
Workers that have been associated with a given Qualification type}
\usage{
mturk_list_workers_with_qualification_type(
  QualificationTypeId,
  Status = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{QualificationTypeId}{[required] The ID of the Qualification type of the Qualifications to return.}

\item{Status}{The status of the Qualifications to return. Can be \code{Granted | Revoked}.}

\item{NextToken}{Pagination Token}

\item{MaxResults}{Limit the number of results returned.}
}
\description{
The \code{\link[=mturk_list_workers_with_qualification_type]{list_workers_with_qualification_type}} operation returns all of the Workers that have been associated with a given Qualification type.

See \url{https://paws-r.github.io/docs/mturk/list_workers_with_qualification_type.html} for full documentation.
}
\keyword{internal}
