% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mturk_operations.R
\name{mturk_reject_assignment}
\alias{mturk_reject_assignment}
\title{The RejectAssignment operation rejects the results of a completed
assignment}
\usage{
mturk_reject_assignment(AssignmentId, RequesterFeedback)
}
\arguments{
\item{AssignmentId}{[required] The ID of the assignment. The assignment must correspond to a HIT
created by the Requester.}

\item{RequesterFeedback}{[required] A message for the Worker, which the Worker can see in the Status section
of the web site.}
}
\description{
The \code{\link[=mturk_reject_assignment]{reject_assignment}} operation rejects the results of a completed assignment.

See \url{https://paws-r.github.io/docs/mturk/reject_assignment.html} for full documentation.
}
\keyword{internal}
