% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mturk_operations.R
\name{mturk_update_hit_review_status}
\alias{mturk_update_hit_review_status}
\title{The UpdateHITReviewStatus operation updates the status of a HIT}
\usage{
mturk_update_hit_review_status(HITId, Revert)
}
\arguments{
\item{HITId}{[required] The ID of the HIT to update.}

\item{Revert}{Specifies how to update the HIT status. Default is \code{False}.
\itemize{
\item Setting this to false will only transition a HIT from \code{Reviewable}
to \code{Reviewing}
\item Setting this to true will only transition a HIT from \code{Reviewing} to
\code{Reviewable}
}}
}
\value{
An empty list.
}
\description{
The \code{\link[=mturk_update_hit_review_status]{update_hit_review_status}}
operation updates the status of a HIT. If the status is Reviewable, this
operation can update the status to Reviewing, or it can revert a
Reviewing HIT back to the Reviewable status.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_hit_review_status(
  HITId = "string",
  Revert = TRUE|FALSE
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
