% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mturk_operations.R
\name{mturk_update_hit_type_of_hit}
\alias{mturk_update_hit_type_of_hit}
\title{The UpdateHITTypeOfHIT operation allows you to change the HITType
properties of a HIT}
\usage{
mturk_update_hit_type_of_hit(HITId, HITTypeId)
}
\arguments{
\item{HITId}{[required] The HIT to update.}

\item{HITTypeId}{[required] The ID of the new HIT type.}
}
\description{
The \code{\link[=mturk_update_hit_type_of_hit]{update_hit_type_of_hit}} operation allows you to change the HITType properties of a HIT. This operation disassociates the HIT from its old HITType properties and associates it with the new HITType properties. The HIT takes on the properties of the new HITType in place of the old ones.

See \url{https://www.paws-r-sdk.com/docs/mturk_update_hit_type_of_hit/} for full documentation.
}
\keyword{internal}
