% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchingestion_operations.R
\name{opensearchingestion_create_pipeline_endpoint}
\alias{opensearchingestion_create_pipeline_endpoint}
\title{Creates a VPC endpoint for an OpenSearch Ingestion pipeline}
\usage{
opensearchingestion_create_pipeline_endpoint(PipelineArn, VpcOptions)
}
\arguments{
\item{PipelineArn}{[required] The Amazon Resource Name (ARN) of the pipeline to create the endpoint
for.}

\item{VpcOptions}{[required] Container for the VPC configuration for the pipeline endpoint, including
subnet IDs and security group IDs.}
}
\description{
Creates a VPC endpoint for an OpenSearch Ingestion pipeline. Pipeline endpoints allow you to ingest data from your VPC into pipelines that you have access to.

See \url{https://www.paws-r-sdk.com/docs/opensearchingestion_create_pipeline_endpoint/} for full documentation.
}
\keyword{internal}
