% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchingestion_operations.R
\name{opensearchingestion_delete_pipeline}
\alias{opensearchingestion_delete_pipeline}
\title{Deletes an OpenSearch Ingestion pipeline}
\usage{
opensearchingestion_delete_pipeline(PipelineName)
}
\arguments{
\item{PipelineName}{[required] The name of the pipeline to delete.}
}
\description{
Deletes an OpenSearch Ingestion pipeline. For more information, see \href{https://docs.aws.amazon.com/opensearch-service/latest/developerguide/delete-pipeline.html}{Deleting Amazon OpenSearch Ingestion pipelines}.

See \url{https://www.paws-r-sdk.com/docs/opensearchingestion_delete_pipeline/} for full documentation.
}
\keyword{internal}
