% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchingestion_operations.R
\name{opensearchingestion_delete_resource_policy}
\alias{opensearchingestion_delete_resource_policy}
\title{Deletes a resource-based policy from an OpenSearch Ingestion resource}
\usage{
opensearchingestion_delete_resource_policy(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource from which to delete the
policy.}
}
\description{
Deletes a resource-based policy from an OpenSearch Ingestion resource.

See \url{https://www.paws-r-sdk.com/docs/opensearchingestion_delete_resource_policy/} for full documentation.
}
\keyword{internal}
