% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchingestion_operations.R
\name{opensearchingestion_get_pipeline}
\alias{opensearchingestion_get_pipeline}
\title{Retrieves information about an OpenSearch Ingestion pipeline}
\usage{
opensearchingestion_get_pipeline(PipelineName)
}
\arguments{
\item{PipelineName}{[required] The name of the pipeline to get information about.}
}
\description{
Retrieves information about an OpenSearch Ingestion pipeline.

See \url{https://www.paws-r-sdk.com/docs/opensearchingestion_get_pipeline/} for full documentation.
}
\keyword{internal}
