% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchingestion_operations.R
\name{opensearchingestion_get_pipeline_blueprint}
\alias{opensearchingestion_get_pipeline_blueprint}
\title{Retrieves information about a specific blueprint for OpenSearch
Ingestion}
\usage{
opensearchingestion_get_pipeline_blueprint(BlueprintName)
}
\arguments{
\item{BlueprintName}{[required] The name of the blueprint to retrieve.}
}
\description{
Retrieves information about a specific blueprint for OpenSearch Ingestion. Blueprints are templates for the configuration needed for a \code{\link[=opensearchingestion_create_pipeline]{create_pipeline}} request. For more information, see \href{https://docs.aws.amazon.com/opensearch-service/latest/developerguide/creating-pipeline.html#pipeline-blueprint}{Using blueprints to create a pipeline}.

See \url{https://www.paws-r-sdk.com/docs/opensearchingestion_get_pipeline_blueprint/} for full documentation.
}
\keyword{internal}
