% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchingestion_operations.R
\name{opensearchingestion_get_pipeline_change_progress}
\alias{opensearchingestion_get_pipeline_change_progress}
\title{Returns progress information for the current change happening on an
OpenSearch Ingestion pipeline}
\usage{
opensearchingestion_get_pipeline_change_progress(PipelineName)
}
\arguments{
\item{PipelineName}{[required] The name of the pipeline.}
}
\description{
Returns progress information for the current change happening on an OpenSearch Ingestion pipeline. Currently, this operation only returns information when a pipeline is being created.

See \url{https://www.paws-r-sdk.com/docs/opensearchingestion_get_pipeline_change_progress/} for full documentation.
}
\keyword{internal}
