% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchingestion_operations.R
\name{opensearchingestion_get_resource_policy}
\alias{opensearchingestion_get_resource_policy}
\title{Retrieves the resource-based policy attached to an OpenSearch Ingestion
resource}
\usage{
opensearchingestion_get_resource_policy(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource for which to retrieve the
policy.}
}
\description{
Retrieves the resource-based policy attached to an OpenSearch Ingestion resource.

See \url{https://www.paws-r-sdk.com/docs/opensearchingestion_get_resource_policy/} for full documentation.
}
\keyword{internal}
