% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchingestion_operations.R
\name{opensearchingestion_list_pipeline_blueprints}
\alias{opensearchingestion_list_pipeline_blueprints}
\title{Retrieves a list of all available blueprints for Data Prepper}
\usage{
opensearchingestion_list_pipeline_blueprints()
}
\description{
Retrieves a list of all available blueprints for Data Prepper. For more information, see \href{https://docs.aws.amazon.com/opensearch-service/latest/developerguide/creating-pipeline.html#pipeline-blueprint}{Using blueprints to create a pipeline}.

See \url{https://www.paws-r-sdk.com/docs/opensearchingestion_list_pipeline_blueprints/} for full documentation.
}
\keyword{internal}
