% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchingestion_operations.R
\name{opensearchingestion_list_pipeline_endpoint_connections}
\alias{opensearchingestion_list_pipeline_endpoint_connections}
\title{Lists the pipeline endpoints connected to pipelines in your account}
\usage{
opensearchingestion_list_pipeline_endpoint_connections(
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{MaxResults}{The maximum number of pipeline endpoint connections to return in the
response.}

\item{NextToken}{If your initial
\code{\link[=opensearchingestion_list_pipeline_endpoint_connections]{list_pipeline_endpoint_connections}}
operation returns a \code{nextToken}, you can include the returned
\code{nextToken} in subsequent
\code{\link[=opensearchingestion_list_pipeline_endpoint_connections]{list_pipeline_endpoint_connections}}
operations, which returns results in the next page.}
}
\description{
Lists the pipeline endpoints connected to pipelines in your account.

See \url{https://www.paws-r-sdk.com/docs/opensearchingestion_list_pipeline_endpoint_connections/} for full documentation.
}
\keyword{internal}
