% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchingestion_operations.R
\name{opensearchingestion_list_pipeline_endpoints}
\alias{opensearchingestion_list_pipeline_endpoints}
\title{Lists all pipeline endpoints in your account}
\usage{
opensearchingestion_list_pipeline_endpoints(
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{MaxResults}{The maximum number of pipeline endpoints to return in the response.}

\item{NextToken}{If your initial
\code{\link[=opensearchingestion_list_pipeline_endpoints]{list_pipeline_endpoints}}
operation returns a \code{NextToken}, you can include the returned
\code{NextToken} in subsequent
\code{\link[=opensearchingestion_list_pipeline_endpoints]{list_pipeline_endpoints}}
operations, which returns results in the next page.}
}
\description{
Lists all pipeline endpoints in your account.

See \url{https://www.paws-r-sdk.com/docs/opensearchingestion_list_pipeline_endpoints/} for full documentation.
}
\keyword{internal}
