% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchingestion_operations.R
\name{opensearchingestion_list_tags_for_resource}
\alias{opensearchingestion_list_tags_for_resource}
\title{Lists all resource tags associated with an OpenSearch Ingestion pipeline}
\usage{
opensearchingestion_list_tags_for_resource(Arn)
}
\arguments{
\item{Arn}{[required] The Amazon Resource Name (ARN) of the pipeline to retrieve tags for.}
}
\description{
Lists all resource tags associated with an OpenSearch Ingestion pipeline. For more information, see \href{https://docs.aws.amazon.com/opensearch-service/latest/developerguide/tag-pipeline.html}{Tagging Amazon OpenSearch Ingestion pipelines}.

See \url{https://www.paws-r-sdk.com/docs/opensearchingestion_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
