% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchingestion_operations.R
\name{opensearchingestion_put_resource_policy}
\alias{opensearchingestion_put_resource_policy}
\title{Attaches a resource-based policy to an OpenSearch Ingestion resource}
\usage{
opensearchingestion_put_resource_policy(ResourceArn, Policy)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource to attach the policy to.}

\item{Policy}{[required] The resource-based policy document in JSON format.}
}
\description{
Attaches a resource-based policy to an OpenSearch Ingestion resource. Resource-based policies grant permissions to principals to perform actions on the resource.

See \url{https://www.paws-r-sdk.com/docs/opensearchingestion_put_resource_policy/} for full documentation.
}
\keyword{internal}
