% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchingestion_operations.R
\name{opensearchingestion_revoke_pipeline_endpoint_connections}
\alias{opensearchingestion_revoke_pipeline_endpoint_connections}
\title{Revokes pipeline endpoints from specified endpoint IDs}
\usage{
opensearchingestion_revoke_pipeline_endpoint_connections(
  PipelineArn,
  EndpointIds
)
}
\arguments{
\item{PipelineArn}{[required] The Amazon Resource Name (ARN) of the pipeline from which to revoke
endpoint connections.}

\item{EndpointIds}{[required] A list of endpoint IDs for which to revoke access to the pipeline.}
}
\description{
Revokes pipeline endpoints from specified endpoint IDs.

See \url{https://www.paws-r-sdk.com/docs/opensearchingestion_revoke_pipeline_endpoint_connections/} for full documentation.
}
\keyword{internal}
