% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchingestion_operations.R
\name{opensearchingestion_start_pipeline}
\alias{opensearchingestion_start_pipeline}
\title{Starts an OpenSearch Ingestion pipeline}
\usage{
opensearchingestion_start_pipeline(PipelineName)
}
\arguments{
\item{PipelineName}{[required] The name of the pipeline to start.}
}
\description{
Starts an OpenSearch Ingestion pipeline. For more information, see \href{https://docs.aws.amazon.com/opensearch-service/latest/developerguide/pipeline--stop-start.html#pipeline--start}{Starting an OpenSearch Ingestion pipeline}.

See \url{https://www.paws-r-sdk.com/docs/opensearchingestion_start_pipeline/} for full documentation.
}
\keyword{internal}
