% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchingestion_operations.R
\name{opensearchingestion_stop_pipeline}
\alias{opensearchingestion_stop_pipeline}
\title{Stops an OpenSearch Ingestion pipeline}
\usage{
opensearchingestion_stop_pipeline(PipelineName)
}
\arguments{
\item{PipelineName}{[required] The name of the pipeline to stop.}
}
\description{
Stops an OpenSearch Ingestion pipeline. For more information, see \href{https://docs.aws.amazon.com/opensearch-service/latest/developerguide/pipeline--stop-start.html#pipeline--stop}{Stopping an OpenSearch Ingestion pipeline}.

See \url{https://www.paws-r-sdk.com/docs/opensearchingestion_stop_pipeline/} for full documentation.
}
\keyword{internal}
