% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchingestion_operations.R
\name{opensearchingestion_tag_resource}
\alias{opensearchingestion_tag_resource}
\title{Tags an OpenSearch Ingestion pipeline}
\usage{
opensearchingestion_tag_resource(Arn, Tags)
}
\arguments{
\item{Arn}{[required] The Amazon Resource Name (ARN) of the pipeline to tag.}

\item{Tags}{[required] The list of key-value tags to add to the pipeline.}
}
\description{
Tags an OpenSearch Ingestion pipeline. For more information, see \href{https://docs.aws.amazon.com/opensearch-service/latest/developerguide/tag-pipeline.html}{Tagging Amazon OpenSearch Ingestion pipelines}.

See \url{https://www.paws-r-sdk.com/docs/opensearchingestion_tag_resource/} for full documentation.
}
\keyword{internal}
