% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchingestion_operations.R
\name{opensearchingestion_untag_resource}
\alias{opensearchingestion_untag_resource}
\title{Removes one or more tags from an OpenSearch Ingestion pipeline}
\usage{
opensearchingestion_untag_resource(Arn, TagKeys)
}
\arguments{
\item{Arn}{[required] The Amazon Resource Name (ARN) of the pipeline to remove tags from.}

\item{TagKeys}{[required] The tag keys to remove.}
}
\description{
Removes one or more tags from an OpenSearch Ingestion pipeline. For more information, see \href{https://docs.aws.amazon.com/opensearch-service/latest/developerguide/tag-pipeline.html}{Tagging Amazon OpenSearch Ingestion pipelines}.

See \url{https://www.paws-r-sdk.com/docs/opensearchingestion_untag_resource/} for full documentation.
}
\keyword{internal}
