% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchingestion_operations.R
\name{opensearchingestion_validate_pipeline}
\alias{opensearchingestion_validate_pipeline}
\title{Checks whether an OpenSearch Ingestion pipeline configuration is valid
prior to creation}
\usage{
opensearchingestion_validate_pipeline(PipelineConfigurationBody)
}
\arguments{
\item{PipelineConfigurationBody}{[required] The pipeline configuration in YAML format. The command accepts the
pipeline configuration as a string or within a .yaml file. If you
provide the configuration as a string, each new line must be escaped
with \verb{\\\\n}.}
}
\description{
Checks whether an OpenSearch Ingestion pipeline configuration is valid prior to creation. For more information, see \href{https://docs.aws.amazon.com/opensearch-service/latest/developerguide/creating-pipeline.html}{Creating Amazon OpenSearch Ingestion pipelines}.

See \url{https://www.paws-r-sdk.com/docs/opensearchingestion_validate_pipeline/} for full documentation.
}
\keyword{internal}
