% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_accept_inbound_connection}
\alias{opensearchservice_accept_inbound_connection}
\title{Allows the remote domain owner to accept an inbound cross-cluster
connection request}
\usage{
opensearchservice_accept_inbound_connection(ConnectionId)
}
\arguments{
\item{ConnectionId}{[required] The ID of the inbound connection you want to accept.}
}
\description{
Allows the remote domain owner to accept an inbound cross-cluster connection request.

See \url{https://paws-r.github.io/docs/opensearchservice/accept_inbound_connection.html} for full documentation.
}
\keyword{internal}
