% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_add_data_source}
\alias{opensearchservice_add_data_source}
\title{Creates a new direct-query data source to the specified domain}
\usage{
opensearchservice_add_data_source(
  DomainName,
  Name,
  DataSourceType,
  Description = NULL
)
}
\arguments{
\item{DomainName}{[required] The name of the domain to add the data source to.}

\item{Name}{[required] A name for the data source.}

\item{DataSourceType}{[required] The type of data source.}

\item{Description}{A description of the data source.}
}
\description{
Creates a new direct-query data source to the specified domain. For more information, see \href{https://docs.aws.amazon.com/opensearch-service/latest/developerguide/direct-query-s3-creating.html}{Creating Amazon OpenSearch Service data source integrations with Amazon S3}.

See \url{https://www.paws-r-sdk.com/docs/opensearchservice_add_data_source/} for full documentation.
}
\keyword{internal}
