% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_associate_package}
\alias{opensearchservice_associate_package}
\title{Associates a package with an Amazon OpenSearch Service domain}
\usage{
opensearchservice_associate_package(PackageID, DomainName)
}
\arguments{
\item{PackageID}{[required] Internal ID of the package to associate with a domain. Use
\code{\link[=opensearchservice_describe_packages]{describe_packages}} to find this
value.}

\item{DomainName}{[required] Name of the domain to associate the package with.}
}
\description{
Associates a package with an Amazon OpenSearch Service domain. For more information, see \href{https://docs.aws.amazon.com/opensearch-service/latest/developerguide/custom-packages.html}{Custom packages for Amazon OpenSearch Service}.

See \url{https://www.paws-r-sdk.com/docs/opensearchservice_associate_package/} for full documentation.
}
\keyword{internal}
