% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_associate_packages}
\alias{opensearchservice_associate_packages}
\title{Operation in the Amazon OpenSearch Service API for associating multiple
packages with a domain simultaneously}
\usage{
opensearchservice_associate_packages(PackageList, DomainName)
}
\arguments{
\item{PackageList}{[required] A list of packages and their prerequisites to be associated with a
domain.}

\item{DomainName}{[required]}
}
\description{
Operation in the Amazon OpenSearch Service API for associating multiple packages with a domain simultaneously.

See \url{https://www.paws-r-sdk.com/docs/opensearchservice_associate_packages/} for full documentation.
}
\keyword{internal}
