% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_delete_application}
\alias{opensearchservice_delete_application}
\title{Deletes an existing OpenSearch Application}
\usage{
opensearchservice_delete_application(id)
}
\arguments{
\item{id}{[required] Unique identifier for the OpenSearch Application that you want to
delete.}
}
\description{
Deletes an existing OpenSearch Application.

See \url{https://www.paws-r-sdk.com/docs/opensearchservice_delete_application/} for full documentation.
}
\keyword{internal}
