% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_delete_data_source}
\alias{opensearchservice_delete_data_source}
\title{Deletes a direct-query data source}
\usage{
opensearchservice_delete_data_source(DomainName, Name)
}
\arguments{
\item{DomainName}{[required] The name of the domain.}

\item{Name}{[required] The name of the data source to delete.}
}
\description{
Deletes a direct-query data source. For more information, see \href{https://docs.aws.amazon.com/opensearch-service/latest/developerguide/}{Deleting an Amazon OpenSearch Service data source with Amazon S3}.

See \url{https://www.paws-r-sdk.com/docs/opensearchservice_delete_data_source/} for full documentation.
}
\keyword{internal}
