% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_delete_direct_query_data_source}
\alias{opensearchservice_delete_direct_query_data_source}
\title{Deletes a previously configured direct query data source from Amazon
OpenSearch Service}
\usage{
opensearchservice_delete_direct_query_data_source(DataSourceName)
}
\arguments{
\item{DataSourceName}{[required] A unique, user-defined label to identify the data source within your
OpenSearch Service environment.}
}
\description{
Deletes a previously configured direct query data source from Amazon OpenSearch Service.

See \url{https://www.paws-r-sdk.com/docs/opensearchservice_delete_direct_query_data_source/} for full documentation.
}
\keyword{internal}
