% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_delete_domain}
\alias{opensearchservice_delete_domain}
\title{Deletes an Amazon OpenSearch Service domain and all of its data}
\usage{
opensearchservice_delete_domain(DomainName)
}
\arguments{
\item{DomainName}{[required] The name of the domain you want to permanently delete.}
}
\description{
Deletes an Amazon OpenSearch Service domain and all of its data. You can't recover a domain after you delete it.

See \url{https://paws-r.github.io/docs/opensearchservice/delete_domain.html} for full documentation.
}
\keyword{internal}
