% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_delete_inbound_connection}
\alias{opensearchservice_delete_inbound_connection}
\title{Allows the remote domain owner to delete an existing inbound
cross-cluster connection}
\usage{
opensearchservice_delete_inbound_connection(ConnectionId)
}
\arguments{
\item{ConnectionId}{[required] The ID of the inbound connection to permanently delete.}
}
\description{
Allows the remote domain owner to delete an existing inbound cross-cluster connection.

See \url{https://paws-r.github.io/docs/opensearchservice/delete_inbound_connection.html} for full documentation.
}
\keyword{internal}
