% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_delete_index}
\alias{opensearchservice_delete_index}
\title{Deletes an OpenSearch index}
\usage{
opensearchservice_delete_index(DomainName, IndexName)
}
\arguments{
\item{DomainName}{[required]}

\item{IndexName}{[required] The name of the index to delete.}
}
\description{
Deletes an OpenSearch index. This operation permanently removes the index and cannot be undone.

See \url{https://www.paws-r-sdk.com/docs/opensearchservice_delete_index/} for full documentation.
}
\keyword{internal}
