% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_delete_outbound_connection}
\alias{opensearchservice_delete_outbound_connection}
\title{Allows the local domain owner to delete an existing outbound
cross-cluster connection}
\usage{
opensearchservice_delete_outbound_connection(ConnectionId)
}
\arguments{
\item{ConnectionId}{[required] The ID of the outbound connection you want to permanently delete.}
}
\description{
Allows the local domain owner to delete an existing outbound cross-cluster connection.

See \url{https://paws-r.github.io/docs/opensearchservice/delete_outbound_connection.html} for full documentation.
}
\keyword{internal}
