% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_delete_package}
\alias{opensearchservice_delete_package}
\title{Deletes the package}
\usage{
opensearchservice_delete_package(PackageID)
}
\arguments{
\item{PackageID}{[required] The internal ID of the package you want to delete. Use
\code{\link[=opensearchservice_describe_packages]{describe_packages}} to find this
value.}
}
\description{
Deletes the package.

See \url{https://paws-r.github.io/docs/opensearchservice/delete_package.html} for full documentation.
}
\keyword{internal}
