% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_deregister_capability}
\alias{opensearchservice_deregister_capability}
\title{Deregisters a capability from an OpenSearch UI application}
\usage{
opensearchservice_deregister_capability(applicationId, capabilityName)
}
\arguments{
\item{applicationId}{[required] The unique identifier of the OpenSearch UI application to deregister the
capability from.}

\item{capabilityName}{[required] The name of the capability to deregister.}
}
\description{
Deregisters a capability from an OpenSearch UI application. This operation removes the capability and its associated configuration.

See \url{https://www.paws-r-sdk.com/docs/opensearchservice_deregister_capability/} for full documentation.
}
\keyword{internal}
