% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_describe_data_source_attachment}
\alias{opensearchservice_describe_data_source_attachment}
\title{Returns the current status and details of a specific data source
attachment for an OpenSearch application}
\usage{
opensearchservice_describe_data_source_attachment(id, dataSourceArn)
}
\arguments{
\item{id}{[required] The unique identifier or name of the OpenSearch application.}

\item{dataSourceArn}{[required] The Amazon Resource Name (ARN) of the domain. See \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/}{Identifiers for IAM Entities} in \emph{Using Amazon Web Services Identity and Access Management} for more information.}
}
\description{
Returns the current status and details of a specific data source attachment for an OpenSearch application. Throws a \code{ResourceNotFoundException} if no attachment record exists for the specified application and data source combination.

See \url{https://www.paws-r-sdk.com/docs/opensearchservice_describe_data_source_attachment/} for full documentation.
}
\keyword{internal}
