% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_describe_domain}
\alias{opensearchservice_describe_domain}
\title{Returns domain configuration information about the specified domain,
including the domain ID, domain endpoint, and domain ARN}
\usage{
opensearchservice_describe_domain(DomainName)
}
\arguments{
\item{DomainName}{[required] The name of the domain for which you want information.}
}
\description{
Returns domain configuration information about the specified domain, including the domain ID, domain endpoint, and domain ARN.

See \url{https://paws-r.github.io/docs/opensearchservice/describe_domain.html} for full documentation.
}
\keyword{internal}
