% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_describe_domain_auto_tunes}
\alias{opensearchservice_describe_domain_auto_tunes}
\title{Provides scheduled Auto-Tune action details for the domain, such as
Auto-Tune action type, description, severity, and scheduled date}
\usage{
opensearchservice_describe_domain_auto_tunes(
  DomainName,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{DomainName}{[required] The domain name for which you want Auto-Tune action details.}

\item{MaxResults}{Set this value to limit the number of results returned. If not
specified, defaults to 100.}

\item{NextToken}{NextToken is sent in case the earlier API call results contain the
NextToken. Used for pagination.}
}
\description{
Provides scheduled Auto-Tune action details for the domain, such as Auto-Tune action type, description, severity, and scheduled date.

See \url{https://paws-r.github.io/docs/opensearchservice/describe_domain_auto_tunes.html} for full documentation.
}
\keyword{internal}
