% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_describe_domain_change_progress}
\alias{opensearchservice_describe_domain_change_progress}
\title{Returns information about the current blue/green deployment happening on
a domain, including a change ID, status, and progress stages}
\usage{
opensearchservice_describe_domain_change_progress(DomainName, ChangeId = NULL)
}
\arguments{
\item{DomainName}{[required] The domain you want to get the progress information about.}

\item{ChangeId}{The specific change ID for which you want to get progress information.
This is an optional parameter. If omitted, the service returns
information about the most recent configuration change.}
}
\description{
Returns information about the current blue/green deployment happening on a domain, including a change ID, status, and progress stages.

See \url{https://paws-r.github.io/docs/opensearchservice/describe_domain_change_progress.html} for full documentation.
}
\keyword{internal}
