% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_describe_domain_config}
\alias{opensearchservice_describe_domain_config}
\title{Provides cluster configuration information about the specified domain,
such as the state, creation date, update version, and update date for
cluster options}
\usage{
opensearchservice_describe_domain_config(DomainName)
}
\arguments{
\item{DomainName}{[required] The domain you want to get information about.}
}
\description{
Provides cluster configuration information about the specified domain, such as the state, creation date, update version, and update date for cluster options.

See \url{https://paws-r.github.io/docs/opensearchservice/describe_domain_config.html} for full documentation.
}
\keyword{internal}
