% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_describe_domains}
\alias{opensearchservice_describe_domains}
\title{Returns domain configuration information about the specified domains,
including the domain ID, domain endpoint, and domain ARN}
\usage{
opensearchservice_describe_domains(DomainNames)
}
\arguments{
\item{DomainNames}{[required] The domains for which you want information.}
}
\description{
Returns domain configuration information about the specified domains, including the domain ID, domain endpoint, and domain ARN.

See \url{https://paws-r.github.io/docs/opensearchservice/describe_domains.html} for full documentation.
}
\keyword{internal}
