% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_describe_dry_run_progress}
\alias{opensearchservice_describe_dry_run_progress}
\title{Describes the progress of a pre-update dry run analysis on an Amazon
OpenSearch Service domain}
\usage{
opensearchservice_describe_dry_run_progress(
  DomainName,
  DryRunId = NULL,
  LoadDryRunConfig = NULL
)
}
\arguments{
\item{DomainName}{[required] The name of the domain.}

\item{DryRunId}{The unique identifier of the dry run.}

\item{LoadDryRunConfig}{Whether to include the configuration of the dry run in the response. The
configuration specifies the updates that you're planning to make on the
domain.}
}
\description{
Describes the progress of a pre-update dry run analysis on an Amazon OpenSearch Service domain. For more information, see \href{https://docs.aws.amazon.com/opensearch-service/latest/developerguide/managedomains-configuration-changes.html#dryrun}{Determining whether a change will cause a blue/green deployment}.

See \url{https://www.paws-r-sdk.com/docs/opensearchservice_describe_dry_run_progress/} for full documentation.
}
\keyword{internal}
