% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_describe_inbound_connections}
\alias{opensearchservice_describe_inbound_connections}
\title{Lists all the inbound cross-cluster search connections for a destination
(remote) Amazon OpenSearch Service domain}
\usage{
opensearchservice_describe_inbound_connections(
  Filters = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{Filters}{A list of filters used to match properties for inbound cross-cluster
connections.}

\item{MaxResults}{An optional parameter that specifies the maximum number of results to
return. You can use \code{nextToken} to get the next page of results.}

\item{NextToken}{If your initial
\code{\link[=opensearchservice_describe_inbound_connections]{describe_inbound_connections}}
operation returns a \code{nextToken}, you can include the returned
\code{nextToken} in subsequent
\code{\link[=opensearchservice_describe_inbound_connections]{describe_inbound_connections}}
operations, which returns results in the next page.}
}
\description{
Lists all the inbound cross-cluster search connections for a destination (remote) Amazon OpenSearch Service domain. For more information, see \href{https://docs.aws.amazon.com/opensearch-service/latest/developerguide/cross-cluster-search.html}{Cross-cluster search for Amazon OpenSearch Service}.

See \url{https://www.paws-r-sdk.com/docs/opensearchservice_describe_inbound_connections/} for full documentation.
}
\keyword{internal}
