% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_describe_insight_details}
\alias{opensearchservice_describe_insight_details}
\title{Describes the details of an existing insight for an Amazon OpenSearch
Service domain}
\usage{
opensearchservice_describe_insight_details(
  Entity,
  InsightId,
  ShowHtmlContent = NULL
)
}
\arguments{
\item{Entity}{[required] The entity for which to retrieve insight details. Specifies the type and
value of the entity, such as a domain name or Amazon Web Services
account ID.}

\item{InsightId}{[required] The unique identifier of the insight to describe.}

\item{ShowHtmlContent}{Specifies whether to show response with HTML content in response or not.}
}
\description{
Describes the details of an existing insight for an Amazon OpenSearch Service domain. Returns detailed fields associated with the specified insight, such as text descriptions and metric data.

See \url{https://www.paws-r-sdk.com/docs/opensearchservice_describe_insight_details/} for full documentation.
}
\keyword{internal}
