% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_describe_instance_type_limits}
\alias{opensearchservice_describe_instance_type_limits}
\title{Describe the limits for a given instance type and OpenSearch or
Elasticsearch version}
\usage{
opensearchservice_describe_instance_type_limits(
  DomainName = NULL,
  InstanceType,
  EngineVersion
)
}
\arguments{
\item{DomainName}{The name of the domain you want to modify. Only include this value if
you're querying OpenSearch \code{Limits} for an existing domain.}

\item{InstanceType}{[required] The instance type for an OpenSearch cluster for which OpenSearch
\code{Limits} are needed.}

\item{EngineVersion}{[required] Version of OpenSearch for which \code{Limits} are needed.}
}
\description{
Describe the limits for a given instance type and OpenSearch or Elasticsearch version. When modifying an existing domain, specify the \code{DomainName} to see which limits you can modify.

See \url{https://paws-r.github.io/docs/opensearchservice/describe_instance_type_limits.html} for full documentation.
}
\keyword{internal}
