% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_describe_reserved_instance_offerings}
\alias{opensearchservice_describe_reserved_instance_offerings}
\title{Lists available reserved OpenSearch instance offerings}
\usage{
opensearchservice_describe_reserved_instance_offerings(
  ReservedInstanceOfferingId = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{ReservedInstanceOfferingId}{The offering identifier filter value. Use this parameter to show only
the available offering that matches the specified reservation
identifier.}

\item{MaxResults}{Set this value to limit the number of results returned. If not
specified, defaults to 100.}

\item{NextToken}{Provides an identifier to allow retrieval of paginated results.}
}
\description{
Lists available reserved OpenSearch instance offerings.

See \url{https://paws-r.github.io/docs/opensearchservice/describe_reserved_instance_offerings.html} for full documentation.
}
\keyword{internal}
