% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_describe_reserved_instances}
\alias{opensearchservice_describe_reserved_instances}
\title{Returns information about reserved OpenSearch instances for this account}
\usage{
opensearchservice_describe_reserved_instances(
  ReservedInstanceId = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{ReservedInstanceId}{The reserved instance identifier filter value. Use this parameter to
show only the reservation that matches the specified reserved OpenSearch
instance ID.}

\item{MaxResults}{Set this value to limit the number of results returned. If not
specified, defaults to 100.}

\item{NextToken}{Provides an identifier to allow retrieval of paginated results.}
}
\description{
Returns information about reserved OpenSearch instances for this account.

See \url{https://paws-r.github.io/docs/opensearchservice/describe_reserved_instances.html} for full documentation.
}
\keyword{internal}
