% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_describe_vpc_endpoints}
\alias{opensearchservice_describe_vpc_endpoints}
\title{Describes one or more Amazon OpenSearch Service-managed VPC endpoints}
\usage{
opensearchservice_describe_vpc_endpoints(VpcEndpointIds)
}
\arguments{
\item{VpcEndpointIds}{[required] The unique identifiers of the endpoints to get information about.}
}
\description{
Describes one or more Amazon OpenSearch Service-managed VPC endpoints.

See \url{https://paws-r.github.io/docs/opensearchservice/describe_vpc_endpoints.html} for full documentation.
}
\keyword{internal}
