% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_detach_data_source}
\alias{opensearchservice_detach_data_source}
\title{Removes a data source from an OpenSearch application}
\usage{
opensearchservice_detach_data_source(id, dataSourceArn)
}
\arguments{
\item{id}{[required] The unique identifier or name of the OpenSearch application to detach the data source from.}

\item{dataSourceArn}{[required] The Amazon Resource Name (ARN) of the domain. See \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/}{Identifiers for IAM Entities} in \emph{Using Amazon Web Services Identity and Access Management} for more information.}
}
\description{
Removes a data source from an OpenSearch application. The application must be in the \code{ACTIVE} state. This operation removes the data source saved object from the application and deletes the attachment record. Throws a \code{ConflictException} if the specified data source has a \code{PENDING} attachment, and a \code{ResourceNotFoundException} if the data source is not currently attached to the application.

See \url{https://www.paws-r-sdk.com/docs/opensearchservice_detach_data_source/} for full documentation.
}
\keyword{internal}
