% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_dissociate_packages}
\alias{opensearchservice_dissociate_packages}
\title{Dissociates multiple packages from a domain simultaneously}
\usage{
opensearchservice_dissociate_packages(PackageList, DomainName)
}
\arguments{
\item{PackageList}{[required] A list of package IDs to be dissociated from a domain.}

\item{DomainName}{[required]}
}
\description{
Dissociates multiple packages from a domain simultaneously.

See \url{https://www.paws-r-sdk.com/docs/opensearchservice_dissociate_packages/} for full documentation.
}
\keyword{internal}
