% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_get_application}
\alias{opensearchservice_get_application}
\title{Check the configuration and status of an existing OpenSearch Application}
\usage{
opensearchservice_get_application(id)
}
\arguments{
\item{id}{[required] Unique identifier of the checked OpenSearch Application.}
}
\description{
Check the configuration and status of an existing OpenSearch Application.

See \url{https://www.paws-r-sdk.com/docs/opensearchservice_get_application/} for full documentation.
}
\keyword{internal}
