% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_get_capability}
\alias{opensearchservice_get_capability}
\title{Retrieves information about a registered capability for an OpenSearch UI
application, including its configuration and current status}
\usage{
opensearchservice_get_capability(applicationId, capabilityName)
}
\arguments{
\item{applicationId}{[required] The unique identifier of the OpenSearch UI application.}

\item{capabilityName}{[required] The name of the capability to retrieve information about.}
}
\description{
Retrieves information about a registered capability for an OpenSearch UI application, including its configuration and current status.

See \url{https://www.paws-r-sdk.com/docs/opensearchservice_get_capability/} for full documentation.
}
\keyword{internal}
