% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_get_compatible_versions}
\alias{opensearchservice_get_compatible_versions}
\title{Returns a map of OpenSearch or Elasticsearch versions and the versions
you can upgrade them to}
\usage{
opensearchservice_get_compatible_versions(DomainName = NULL)
}
\arguments{
\item{DomainName}{The name of an existing domain. Provide this parameter to limit the
results to a single domain.}
}
\description{
Returns a map of OpenSearch or Elasticsearch versions and the versions you can upgrade them to.

See \url{https://www.paws-r-sdk.com/docs/opensearchservice_get_compatible_versions/} for full documentation.
}
\keyword{internal}
