% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_get_data_source}
\alias{opensearchservice_get_data_source}
\title{Retrieves information about a direct query data source}
\usage{
opensearchservice_get_data_source(DomainName, Name)
}
\arguments{
\item{DomainName}{[required] The name of the domain.}

\item{Name}{[required] The name of the data source to get information about.}
}
\description{
Retrieves information about a direct query data source.

See \url{https://www.paws-r-sdk.com/docs/opensearchservice_get_data_source/} for full documentation.
}
\keyword{internal}
