% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_get_default_application_setting}
\alias{opensearchservice_get_default_application_setting}
\title{Gets the ARN of the current default application}
\usage{
opensearchservice_get_default_application_setting()
}
\description{
Gets the ARN of the current default application.

See \url{https://www.paws-r-sdk.com/docs/opensearchservice_get_default_application_setting/} for full documentation.
}
\keyword{internal}
