% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_get_direct_query_data_source}
\alias{opensearchservice_get_direct_query_data_source}
\title{Returns detailed configuration information for a specific direct query
data source in Amazon OpenSearch Service}
\usage{
opensearchservice_get_direct_query_data_source(DataSourceName)
}
\arguments{
\item{DataSourceName}{[required] A unique, user-defined label that identifies the data source within your
OpenSearch Service environment.}
}
\description{
Returns detailed configuration information for a specific direct query data source in Amazon OpenSearch Service.

See \url{https://www.paws-r-sdk.com/docs/opensearchservice_get_direct_query_data_source/} for full documentation.
}
\keyword{internal}
