% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_get_domain_maintenance_status}
\alias{opensearchservice_get_domain_maintenance_status}
\title{The status of the maintenance action}
\usage{
opensearchservice_get_domain_maintenance_status(DomainName, MaintenanceId)
}
\arguments{
\item{DomainName}{[required] The name of the domain.}

\item{MaintenanceId}{[required] The request ID of the maintenance action.}
}
\description{
The status of the maintenance action.

See \url{https://www.paws-r-sdk.com/docs/opensearchservice_get_domain_maintenance_status/} for full documentation.
}
\keyword{internal}
