% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_get_migration}
\alias{opensearchservice_get_migration}
\title{Retrieves the current status and progress of a migration job, including
the number of exported and imported objects and error details if the
migration failed}
\usage{
opensearchservice_get_migration(migrationId)
}
\arguments{
\item{migrationId}{[required] The unique identifier of the migration job to retrieve.}
}
\description{
Retrieves the current status and progress of a migration job, including the number of exported and imported objects and error details if the migration failed.

See \url{https://www.paws-r-sdk.com/docs/opensearchservice_get_migration/} for full documentation.
}
\keyword{internal}
