% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_get_upgrade_status}
\alias{opensearchservice_get_upgrade_status}
\title{Returns the most recent status of the last upgrade or upgrade
eligibility check performed on an Amazon OpenSearch Service domain}
\usage{
opensearchservice_get_upgrade_status(DomainName)
}
\arguments{
\item{DomainName}{[required] The domain of the domain to get upgrade status information for.}
}
\description{
Returns the most recent status of the last upgrade or upgrade eligibility check performed on an Amazon OpenSearch Service domain.

See \url{https://paws-r.github.io/docs/opensearchservice/get_upgrade_status.html} for full documentation.
}
\keyword{internal}
