% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_insight_feedback}
\alias{opensearchservice_insight_feedback}
\title{Submits feedback for an existing insight in an Amazon OpenSearch Service
domain}
\usage{
opensearchservice_insight_feedback(
  Entity,
  InsightId,
  Thumbs,
  FeedbackText = NULL
)
}
\arguments{
\item{Entity}{[required] The entity for which to submit insight feedback. Specifies the type and value of the entity, such as a domain name.}

\item{InsightId}{[required] The unique identifier of the insight for which to submit feedback.}

\item{Thumbs}{[required] The thumbs up or thumbs down feedback for the insight. Possible values are \code{Up} and \code{Down}.}

\item{FeedbackText}{Optional text feedback providing additional details about the insight. Maximum length is 1000 characters.}
}
\description{
Submits feedback for an existing insight in an Amazon OpenSearch Service domain. Allows users to provide a thumbs up or thumbs down rating and optional text feedback for a specific insight.

See \url{https://www.paws-r-sdk.com/docs/opensearchservice_insight_feedback/} for full documentation.
}
\keyword{internal}
