% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_list_data_source_attachments}
\alias{opensearchservice_list_data_source_attachments}
\title{Returns a paginated list of all data source attachments for an
OpenSearch application, including attachments in all states (PENDING,
ATTACHED, and FAILED)}
\usage{
opensearchservice_list_data_source_attachments(
  id,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{id}{[required] The unique identifier or name of the OpenSearch application to list attachments for.}

\item{nextToken}{The pagination token from a previous call to retrieve the next set of results.}

\item{maxResults}{The maximum number of results to return per page. The default is 50.}
}
\description{
Returns a paginated list of all data source attachments for an OpenSearch application, including attachments in all states (\code{PENDING}, \code{ATTACHED}, and \code{FAILED}).

See \url{https://www.paws-r-sdk.com/docs/opensearchservice_list_data_source_attachments/} for full documentation.
}
\keyword{internal}
