% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_list_data_sources}
\alias{opensearchservice_list_data_sources}
\title{Lists direct-query data sources for a specific domain}
\usage{
opensearchservice_list_data_sources(DomainName)
}
\arguments{
\item{DomainName}{[required] The name of the domain.}
}
\description{
Lists direct-query data sources for a specific domain. For more information, see For more information, see \href{https://docs.aws.amazon.com/opensearch-service/latest/developerguide/direct-query-s3.html}{Working with Amazon OpenSearch Service direct queries with Amazon S3}.

See \url{https://www.paws-r-sdk.com/docs/opensearchservice_list_data_sources/} for full documentation.
}
\keyword{internal}
