% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_list_direct_query_data_sources}
\alias{opensearchservice_list_direct_query_data_sources}
\title{Lists an inventory of all the direct query data sources that you have
configured within Amazon OpenSearch Service}
\usage{
opensearchservice_list_direct_query_data_sources(NextToken = NULL)
}
\arguments{
\item{NextToken}{When \code{nextToken} is returned, there are more results available. The value of \code{nextToken} is a unique pagination token for each page. Send the request again using the returned token to retrieve the next page.}
}
\description{
Lists an inventory of all the direct query data sources that you have configured within Amazon OpenSearch Service.

See \url{https://www.paws-r-sdk.com/docs/opensearchservice_list_direct_query_data_sources/} for full documentation.
}
\keyword{internal}
