% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_list_domain_maintenances}
\alias{opensearchservice_list_domain_maintenances}
\title{A list of maintenance actions for the domain}
\usage{
opensearchservice_list_domain_maintenances(
  DomainName,
  Action = NULL,
  Status = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{DomainName}{[required] The name of the domain.}

\item{Action}{The name of the action.}

\item{Status}{The status of the action.}

\item{MaxResults}{An optional parameter that specifies the maximum number of results to
return. You can use \code{nextToken} to get the next page of results.}

\item{NextToken}{If your initial
\code{\link[=opensearchservice_list_domain_maintenances]{list_domain_maintenances}}
operation returns a \code{nextToken}, include the returned \code{nextToken} in
subsequent
\code{\link[=opensearchservice_list_domain_maintenances]{list_domain_maintenances}}
operations, which returns results in the next page.}
}
\description{
A list of maintenance actions for the domain.

See \url{https://www.paws-r-sdk.com/docs/opensearchservice_list_domain_maintenances/} for full documentation.
}
\keyword{internal}
