% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_list_insights}
\alias{opensearchservice_list_insights}
\title{Lists insights for an Amazon OpenSearch Service domain or Amazon Web
Services account}
\usage{
opensearchservice_list_insights(
  Entity,
  TimeRange = NULL,
  SortOrder = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{Entity}{[required] The entity for which to list insights. Specifies the type and value of
the entity, such as a domain name or Amazon Web Services account ID.}

\item{TimeRange}{The time range for filtering insights, specified as epoch millisecond
timestamps.}

\item{SortOrder}{The sort order for the results. Possible values are \code{ASC} (ascending)
and \code{DESC} (descending).}

\item{MaxResults}{An optional parameter that specifies the maximum number of results to
return. You can use \code{NextToken} to get the next page of results. Valid
values are 1 to 500.}

\item{NextToken}{If your initial \code{\link[=opensearchservice_list_insights]{list_insights}}
operation returns a \code{NextToken}, include the returned \code{NextToken} in
subsequent \code{\link[=opensearchservice_list_insights]{list_insights}} operations
to retrieve the next page of results.}
}
\description{
Lists insights for an Amazon OpenSearch Service domain or Amazon Web Services account. Returns a paginated list of insights based on the specified entity, filters, time range, and sort order.

See \url{https://www.paws-r-sdk.com/docs/opensearchservice_list_insights/} for full documentation.
}
\keyword{internal}
