% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_list_instance_type_details}
\alias{opensearchservice_list_instance_type_details}
\title{List instance type details}
\usage{
opensearchservice_list_instance_type_details(
  EngineVersion,
  DomainName = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{EngineVersion}{[required]}

\item{DomainName}{}

\item{MaxResults}{}

\item{NextToken}{}
}
\description{
List instance type details

See \url{https://paws-r.github.io/docs/opensearchservice/list_instance_type_details.html} for full documentation.
}
\keyword{internal}
