% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_list_migrations}
\alias{opensearchservice_list_migrations}
\title{Lists migration jobs for an Amazon OpenSearch Service application}
\usage{
opensearchservice_list_migrations(
  applicationId,
  status = NULL,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{applicationId}{[required] The unique identifier of the OpenSearch application to list migrations for.}

\item{status}{Filters the results by migration status. Valid values are \code{PENDING}, \code{IN_PROGRESS}, \code{SUCCEEDED}, and \code{FAILED}.}

\item{maxResults}{The maximum number of results to return in a single call.}

\item{nextToken}{The pagination token from a previous call to retrieve the next set of results.}
}
\description{
Lists migration jobs for an Amazon OpenSearch Service application. You can filter results by migration status. Use pagination to ensure that the operation returns quickly and successfully.

See \url{https://www.paws-r-sdk.com/docs/opensearchservice_list_migrations/} for full documentation.
}
\keyword{internal}
