% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_list_versions}
\alias{opensearchservice_list_versions}
\title{List all supported versions of OpenSearch and Elasticsearch}
\usage{
opensearchservice_list_versions(MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{Set this value to limit the number of results returned. Value must be
greater than 10 or it won't be honored.}

\item{NextToken}{}
}
\description{
List all supported versions of OpenSearch and Elasticsearch.

See \url{https://paws-r.github.io/docs/opensearchservice/list_versions.html} for full documentation.
}
\keyword{internal}
