% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_list_vpc_endpoint_access}
\alias{opensearchservice_list_vpc_endpoint_access}
\title{Retrieves information about each Amazon Web Services principal that is
allowed to access a given Amazon OpenSearch Service domain through the
use of an interface VPC endpoint}
\usage{
opensearchservice_list_vpc_endpoint_access(DomainName, NextToken = NULL)
}
\arguments{
\item{DomainName}{[required] The name of the OpenSearch Service domain to retrieve access information
for.}

\item{NextToken}{If your initial
\code{\link[=opensearchservice_list_vpc_endpoint_access]{list_vpc_endpoint_access}}
operation returns a \code{nextToken}, you can include the returned
\code{nextToken} in subsequent
\code{\link[=opensearchservice_list_vpc_endpoint_access]{list_vpc_endpoint_access}}
operations, which returns results in the next page.}
}
\description{
Retrieves information about each Amazon Web Services principal that is allowed to access a given Amazon OpenSearch Service domain through the use of an interface VPC endpoint.

See \url{https://www.paws-r-sdk.com/docs/opensearchservice_list_vpc_endpoint_access/} for full documentation.
}
\keyword{internal}
