% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_list_vpc_endpoints}
\alias{opensearchservice_list_vpc_endpoints}
\title{Retrieves all Amazon OpenSearch Service-managed VPC endpoints in the
current Amazon Web Services account and Region}
\usage{
opensearchservice_list_vpc_endpoints(NextToken = NULL)
}
\arguments{
\item{NextToken}{If your initial
\code{\link[=opensearchservice_list_vpc_endpoints]{list_vpc_endpoints}} operation
returns a \code{nextToken}, you can include the returned \code{nextToken} in
subsequent \code{\link[=opensearchservice_list_vpc_endpoints]{list_vpc_endpoints}}
operations, which returns results in the next page.}
}
\description{
Retrieves all Amazon OpenSearch Service-managed VPC endpoints in the current Amazon Web Services account and Region.

See \url{https://www.paws-r-sdk.com/docs/opensearchservice_list_vpc_endpoints/} for full documentation.
}
\keyword{internal}
