% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_list_vpc_endpoints_for_domain}
\alias{opensearchservice_list_vpc_endpoints_for_domain}
\title{Retrieves all Amazon OpenSearch Service-managed VPC endpoints associated
with a particular domain}
\usage{
opensearchservice_list_vpc_endpoints_for_domain(DomainName, NextToken = NULL)
}
\arguments{
\item{DomainName}{[required] The name of the domain to list associated VPC endpoints for.}

\item{NextToken}{If your initial \code{ListEndpointsForDomain} operation returns a
\code{nextToken}, you can include the returned \code{nextToken} in subsequent
\code{ListEndpointsForDomain} operations, which returns results in the next
page.}
}
\description{
Retrieves all Amazon OpenSearch Service-managed VPC endpoints associated with a particular domain.

See \url{https://paws-r.github.io/docs/opensearchservice/list_vpc_endpoints_for_domain.html} for full documentation.
}
\keyword{internal}
