% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_purchase_reserved_instance_offering}
\alias{opensearchservice_purchase_reserved_instance_offering}
\title{Allows you to purchase Amazon OpenSearch Service Reserved Instances}
\usage{
opensearchservice_purchase_reserved_instance_offering(
  ReservedInstanceOfferingId,
  ReservationName,
  InstanceCount = NULL
)
}
\arguments{
\item{ReservedInstanceOfferingId}{[required] The ID of the Reserved Instance offering to purchase.}

\item{ReservationName}{[required] A customer-specified identifier to track this reservation.}

\item{InstanceCount}{The number of OpenSearch instances to reserve.}
}
\description{
Allows you to purchase Amazon OpenSearch Service Reserved Instances.

See \url{https://www.paws-r-sdk.com/docs/opensearchservice_purchase_reserved_instance_offering/} for full documentation.
}
\keyword{internal}
