% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_put_default_application_setting}
\alias{opensearchservice_put_default_application_setting}
\title{Sets the default application to the application with the specified ARN}
\usage{
opensearchservice_put_default_application_setting(applicationArn, setAsDefault)
}
\arguments{
\item{applicationArn}{[required]}

\item{setAsDefault}{[required] Set to true to set the specified ARN as the default application. Set to
false to clear the default application.}
}
\description{
Sets the default application to the application with the specified ARN.

See \url{https://www.paws-r-sdk.com/docs/opensearchservice_put_default_application_setting/} for full documentation.
}
\keyword{internal}
