% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_reject_inbound_connection}
\alias{opensearchservice_reject_inbound_connection}
\title{Allows the remote Amazon OpenSearch Service domain owner to reject an
inbound cross-cluster connection request}
\usage{
opensearchservice_reject_inbound_connection(ConnectionId)
}
\arguments{
\item{ConnectionId}{[required] The unique identifier of the inbound connection to reject.}
}
\description{
Allows the remote Amazon OpenSearch Service domain owner to reject an inbound cross-cluster connection request.

See \url{https://www.paws-r-sdk.com/docs/opensearchservice_reject_inbound_connection/} for full documentation.
}
\keyword{internal}
