% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_rollback_service_software_update}
\alias{opensearchservice_rollback_service_software_update}
\title{Rolls back a service software update for a domain to the previous
version}
\usage{
opensearchservice_rollback_service_software_update(DomainName)
}
\arguments{
\item{DomainName}{[required] The name of the domain to roll back the service software update on.}
}
\description{
Rolls back a service software update for a domain to the previous version. For more information, see \href{https://docs.aws.amazon.com/opensearch-service/latest/developerguide/service-software.html}{Service software updates in Amazon OpenSearch Service}.

See \url{https://www.paws-r-sdk.com/docs/opensearchservice_rollback_service_software_update/} for full documentation.
}
\keyword{internal}
