% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_start_domain_maintenance}
\alias{opensearchservice_start_domain_maintenance}
\title{Starts the node maintenance process on the data node}
\usage{
opensearchservice_start_domain_maintenance(DomainName, Action, NodeId = NULL)
}
\arguments{
\item{DomainName}{[required] The name of the domain.}

\item{Action}{[required] The name of the action.}

\item{NodeId}{The ID of the data node.}
}
\description{
Starts the node maintenance process on the data node. These processes can include a node reboot, an Opensearch or Elasticsearch process restart, or a Dashboard or Kibana restart.

See \url{https://www.paws-r-sdk.com/docs/opensearchservice_start_domain_maintenance/} for full documentation.
}
\keyword{internal}
