% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_update_application}
\alias{opensearchservice_update_application}
\title{Updates the configuration and settings of an existing OpenSearch
application}
\usage{
opensearchservice_update_application(
  id,
  dataSources = NULL,
  appConfigs = NULL,
  iamIdentityCenterOptions = NULL
)
}
\arguments{
\item{id}{[required] The unique identifier for the OpenSearch application to be updated.}

\item{dataSources}{The data sources to associate with the OpenSearch application.}

\item{appConfigs}{The configuration settings to modify for the OpenSearch application.}

\item{iamIdentityCenterOptions}{Configuration settings for integrating IAM Identity Center with the OpenSearch application.}
}
\description{
Updates the configuration and settings of an existing OpenSearch application.

See \url{https://www.paws-r-sdk.com/docs/opensearchservice_update_application/} for full documentation.
}
\keyword{internal}
