% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_update_data_source}
\alias{opensearchservice_update_data_source}
\title{Updates a direct-query data source}
\usage{
opensearchservice_update_data_source(
  DomainName,
  Name,
  DataSourceType,
  Description = NULL,
  Status = NULL
)
}
\arguments{
\item{DomainName}{[required] The name of the domain.}

\item{Name}{[required] The name of the data source to modify.}

\item{DataSourceType}{[required] The type of data source.}

\item{Description}{A new description of the data source.}

\item{Status}{The status of the data source update.}
}
\description{
Updates a direct-query data source. For more information, see \href{https://docs.aws.amazon.com/opensearch-service/latest/developerguide/direct-query-s3-creating.html}{Working with Amazon OpenSearch Service data source integrations with Amazon S3}.

See \url{https://www.paws-r-sdk.com/docs/opensearchservice_update_data_source/} for full documentation.
}
\keyword{internal}
