% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_update_package}
\alias{opensearchservice_update_package}
\title{Updates a package for use with Amazon OpenSearch Service domains}
\usage{
opensearchservice_update_package(
  PackageID,
  PackageSource,
  PackageDescription = NULL,
  CommitMessage = NULL
)
}
\arguments{
\item{PackageID}{[required] The unique identifier for the package.}

\item{PackageSource}{[required]}

\item{PackageDescription}{A new description of the package.}

\item{CommitMessage}{A commit message for the new version which is shown as part of
\code{GetPackageVersionHistoryResponse}.}
}
\description{
Updates a package for use with Amazon OpenSearch Service domains.

See \url{https://paws-r.github.io/docs/opensearchservice/update_package.html} for full documentation.
}
\keyword{internal}
