% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_update_package_scope}
\alias{opensearchservice_update_package_scope}
\title{Updates the scope of a package}
\usage{
opensearchservice_update_package_scope(PackageID, Operation, PackageUserList)
}
\arguments{
\item{PackageID}{[required] ID of the package whose scope is being updated.}

\item{Operation}{[required] The operation to perform on the package scope (e.g., add/remove/override
users).}

\item{PackageUserList}{[required] List of users to be added or removed from the package scope.}
}
\description{
Updates the scope of a package. Scope of the package defines users who can view and associate a package.

See \url{https://www.paws-r-sdk.com/docs/opensearchservice_update_package_scope/} for full documentation.
}
\keyword{internal}
