% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_update_vpc_endpoint}
\alias{opensearchservice_update_vpc_endpoint}
\title{Modifies an Amazon OpenSearch Service-managed interface VPC endpoint}
\usage{
opensearchservice_update_vpc_endpoint(VpcEndpointId, VpcOptions)
}
\arguments{
\item{VpcEndpointId}{[required] The unique identifier of the endpoint.}

\item{VpcOptions}{[required] The security groups and/or subnets to add, remove, or modify.}
}
\description{
Modifies an Amazon OpenSearch Service-managed interface VPC endpoint.

See \url{https://www.paws-r-sdk.com/docs/opensearchservice_update_vpc_endpoint/} for full documentation.
}
\keyword{internal}
