% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchserviceserverless_operations.R
\name{opensearchserviceserverless_batch_get_effective_lifecycle_policy}
\alias{opensearchserviceserverless_batch_get_effective_lifecycle_policy}
\title{Returns a list of successful and failed retrievals for the OpenSearch
Serverless indexes}
\usage{
opensearchserviceserverless_batch_get_effective_lifecycle_policy(
  resourceIdentifiers
)
}
\arguments{
\item{resourceIdentifiers}{[required] The unique identifiers of policy types and resource names.}
}
\description{
Returns a list of successful and failed retrievals for the OpenSearch Serverless indexes. For more information, see \href{https://docs.aws.amazon.com/opensearch-service/latest/developerguide/serverless-lifecycle.html#serverless-lifecycle-list}{Viewing data lifecycle policies}.

See \url{https://www.paws-r-sdk.com/docs/opensearchserviceserverless_batch_get_effective_lifecycle_policy/} for full documentation.
}
\keyword{internal}
