% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchserviceserverless_operations.R
\name{opensearchserviceserverless_batch_get_collection_group}
\alias{opensearchserviceserverless_batch_get_collection_group}
\title{Returns attributes for one or more collection groups, including capacity
limits and the number of collections in each group}
\usage{
opensearchserviceserverless_batch_get_collection_group(
  ids = NULL,
  names = NULL
)
}
\arguments{
\item{ids}{A list of collection group IDs. You can't provide names and IDs in the
same request.}

\item{names}{A list of collection group names. You can't provide names and IDs in the
same request.}
}
\description{
Returns attributes for one or more collection groups, including capacity limits and the number of collections in each group. For more information, see \href{https://docs.aws.amazon.com/opensearch-service/latest/developerguide/serverless-manage.html}{Creating and managing Amazon OpenSearch Serverless collections}.

See \url{https://www.paws-r-sdk.com/docs/opensearchserviceserverless_batch_get_collection_group/} for full documentation.
}
\keyword{internal}
