% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchserviceserverless_operations.R
\name{opensearchserviceserverless_batch_get_lifecycle_policy}
\alias{opensearchserviceserverless_batch_get_lifecycle_policy}
\title{Returns one or more configured OpenSearch Serverless lifecycle policies}
\usage{
opensearchserviceserverless_batch_get_lifecycle_policy(identifiers)
}
\arguments{
\item{identifiers}{[required] The unique identifiers of policy types and policy names.}
}
\description{
Returns one or more configured OpenSearch Serverless lifecycle policies. For more information, see \href{https://docs.aws.amazon.com/opensearch-service/latest/developerguide/serverless-lifecycle.html#serverless-lifecycle-list}{Viewing data lifecycle policies}.

See \url{https://www.paws-r-sdk.com/docs/opensearchserviceserverless_batch_get_lifecycle_policy/} for full documentation.
}
\keyword{internal}
