% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchserviceserverless_operations.R
\name{opensearchserviceserverless_batch_get_vpc_endpoint}
\alias{opensearchserviceserverless_batch_get_vpc_endpoint}
\title{Returns attributes for one or more VPC endpoints associated with the
current account}
\usage{
opensearchserviceserverless_batch_get_vpc_endpoint(ids)
}
\arguments{
\item{ids}{[required] A list of VPC endpoint identifiers.}
}
\description{
Returns attributes for one or more VPC endpoints associated with the current account. For more information, see \href{https://docs.aws.amazon.com/opensearch-service/latest/developerguide/serverless-vpc.html}{Access Amazon OpenSearch Serverless using an interface endpoint}.

See \url{https://www.paws-r-sdk.com/docs/opensearchserviceserverless_batch_get_vpc_endpoint/} for full documentation.
}
\keyword{internal}
