% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchserviceserverless_operations.R
\name{opensearchserviceserverless_create_collection}
\alias{opensearchserviceserverless_create_collection}
\title{Creates a new OpenSearch Serverless collection}
\usage{
opensearchserviceserverless_create_collection(
  name,
  type = NULL,
  description = NULL,
  tags = NULL,
  standbyReplicas = NULL,
  vectorOptions = NULL,
  collectionGroupName = NULL,
  encryptionConfig = NULL,
  deletionProtection = NULL,
  clientToken = NULL
)
}
\arguments{
\item{name}{[required] Name of the collection.}

\item{type}{The type of collection.}

\item{description}{Description of the collection.}

\item{tags}{An arbitrary set of tags (key–value pairs) to associate with the OpenSearch Serverless collection.}

\item{standbyReplicas}{Indicates whether standby replicas should be used for a collection.}

\item{vectorOptions}{Configuration options for vector search capabilities in the collection.}

\item{collectionGroupName}{The name of the collection group to associate with the collection.}

\item{encryptionConfig}{Encryption settings for the collection.}

\item{deletionProtection}{Indicates whether to enable deletion protection for the collection. When set to \code{ENABLED}, the collection cannot be deleted.}

\item{clientToken}{Unique, case-sensitive identifier to ensure idempotency of the request.}
}
\description{
Creates a new OpenSearch Serverless collection. For more information, see \href{https://docs.aws.amazon.com/opensearch-service/latest/developerguide/serverless-manage.html}{Creating and managing Amazon OpenSearch Serverless collections}.

See \url{https://www.paws-r-sdk.com/docs/opensearchserviceserverless_create_collection/} for full documentation.
}
\keyword{internal}
