% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchserviceserverless_operations.R
\name{opensearchserviceserverless_create_index}
\alias{opensearchserviceserverless_create_index}
\title{Creates an index within an OpenSearch Serverless collection}
\usage{
opensearchserviceserverless_create_index(id, indexName, indexSchema = NULL)
}
\arguments{
\item{id}{[required] The unique identifier of the collection in which to create the index.}

\item{indexName}{[required] The name of the index to create. Index names must be lowercase and can't
begin with underscores (_) or hyphens (-).}

\item{indexSchema}{The JSON schema definition for the index, including field mappings and
settings.}
}
\description{
Creates an index within an OpenSearch Serverless collection. Unlike other OpenSearch indexes, indexes created by this API are automatically configured to conduct automatic semantic enrichment ingestion and search. For more information, see \href{https://docs.aws.amazon.com/opensearch-service/latest/developerguide/serverless-manage.html#serverless-semantic-enrichment}{About automatic semantic enrichment} in the \emph{OpenSearch User Guide}.

See \url{https://www.paws-r-sdk.com/docs/opensearchserviceserverless_create_index/} for full documentation.
}
\keyword{internal}
