% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchserviceserverless_operations.R
\name{opensearchserviceserverless_create_lifecycle_policy}
\alias{opensearchserviceserverless_create_lifecycle_policy}
\title{Creates a lifecyle policy to be applied to OpenSearch Serverless indexes}
\usage{
opensearchserviceserverless_create_lifecycle_policy(
  clientToken = NULL,
  description = NULL,
  name,
  policy,
  type
)
}
\arguments{
\item{clientToken}{A unique, case-sensitive identifier to ensure idempotency of the
request.}

\item{description}{A description of the lifecycle policy.}

\item{name}{[required] The name of the lifecycle policy.}

\item{policy}{[required] The JSON policy document to use as the content for the lifecycle policy.}

\item{type}{[required] The type of lifecycle policy.}
}
\description{
Creates a lifecyle policy to be applied to OpenSearch Serverless indexes. Lifecycle policies define the number of days or hours to retain the data on an OpenSearch Serverless index. For more information, see \href{https://docs.aws.amazon.com/opensearch-service/latest/developerguide/serverless-lifecycle.html#serverless-lifecycle-create}{Creating data lifecycle policies}.

See \url{https://www.paws-r-sdk.com/docs/opensearchserviceserverless_create_lifecycle_policy/} for full documentation.
}
\keyword{internal}
