% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchserviceserverless_operations.R
\name{opensearchserviceserverless_delete_collection}
\alias{opensearchserviceserverless_delete_collection}
\title{Deletes an OpenSearch Serverless collection}
\usage{
opensearchserviceserverless_delete_collection(id, clientToken = NULL)
}
\arguments{
\item{id}{[required] The unique identifier of the collection. For example, \verb{1iu5usc406kd}.
The ID is part of the collection endpoint. You can also retrieve it
using the
\code{\link[=opensearchserviceserverless_list_collections]{list_collections}} API.}

\item{clientToken}{A unique, case-sensitive identifier to ensure idempotency of the
request.}
}
\description{
Deletes an OpenSearch Serverless collection. For more information, see \href{https://docs.aws.amazon.com/opensearch-service/latest/developerguide/serverless-manage.html}{Creating and managing Amazon OpenSearch Serverless collections}.

See \url{https://www.paws-r-sdk.com/docs/opensearchserviceserverless_delete_collection/} for full documentation.
}
\keyword{internal}
