% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchserviceserverless_operations.R
\name{opensearchserviceserverless_delete_collection_group}
\alias{opensearchserviceserverless_delete_collection_group}
\title{Deletes a collection group}
\usage{
opensearchserviceserverless_delete_collection_group(id, clientToken = NULL)
}
\arguments{
\item{id}{[required] The unique identifier of the collection group to delete.}

\item{clientToken}{Unique, case-sensitive identifier to ensure idempotency of the request.}
}
\description{
Deletes a collection group. You can only delete empty collection groups that contain no collections. For more information, see \href{https://docs.aws.amazon.com/opensearch-service/latest/developerguide/serverless-manage.html}{Creating and managing Amazon OpenSearch Serverless collections}.

See \url{https://www.paws-r-sdk.com/docs/opensearchserviceserverless_delete_collection_group/} for full documentation.
}
\keyword{internal}
