% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchserviceserverless_operations.R
\name{opensearchserviceserverless_delete_index}
\alias{opensearchserviceserverless_delete_index}
\title{Deletes an index from an OpenSearch Serverless collection}
\usage{
opensearchserviceserverless_delete_index(id, indexName)
}
\arguments{
\item{id}{[required] The unique identifier of the collection containing the index to delete.}

\item{indexName}{[required] The name of the index to delete.}
}
\description{
Deletes an index from an OpenSearch Serverless collection. Be aware that the index might be configured to conduct automatic semantic enrichment ingestion and search. For more information, see \href{https://docs.aws.amazon.com/opensearch-service/latest/developerguide/serverless-manage.html#serverless-semantic-enrichment}{About automatic semantic enrichment}.

See \url{https://www.paws-r-sdk.com/docs/opensearchserviceserverless_delete_index/} for full documentation.
}
\keyword{internal}
