% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchserviceserverless_operations.R
\name{opensearchserviceserverless_delete_security_config}
\alias{opensearchserviceserverless_delete_security_config}
\title{Deletes a security configuration for OpenSearch Serverless}
\usage{
opensearchserviceserverless_delete_security_config(id, clientToken = NULL)
}
\arguments{
\item{id}{[required] The security configuration identifier. For SAML the ID will be
\verb{saml/<accountId>/<idpProviderName>}. For example,
\code{saml/123456789123/OKTADev}.}

\item{clientToken}{Unique, case-sensitive identifier to ensure idempotency of the request.}
}
\description{
Deletes a security configuration for OpenSearch Serverless. For more information, see \href{https://docs.aws.amazon.com/opensearch-service/latest/developerguide/serverless-saml.html}{SAML authentication for Amazon OpenSearch Serverless}.

See \url{https://www.paws-r-sdk.com/docs/opensearchserviceserverless_delete_security_config/} for full documentation.
}
\keyword{internal}
