% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchserviceserverless_operations.R
\name{opensearchserviceserverless_delete_vpc_endpoint}
\alias{opensearchserviceserverless_delete_vpc_endpoint}
\title{Deletes an OpenSearch Serverless-managed interface endpoint}
\usage{
opensearchserviceserverless_delete_vpc_endpoint(clientToken = NULL, id)
}
\arguments{
\item{clientToken}{Unique, case-sensitive identifier to ensure idempotency of the request.}

\item{id}{[required] The VPC endpoint identifier.}
}
\description{
Deletes an OpenSearch Serverless-managed interface endpoint. For more information, see \href{https://docs.aws.amazon.com/opensearch-service/latest/developerguide/serverless-vpc.html}{Access Amazon OpenSearch Serverless using an interface endpoint}.

See \url{https://www.paws-r-sdk.com/docs/opensearchserviceserverless_delete_vpc_endpoint/} for full documentation.
}
\keyword{internal}
