% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchserviceserverless_operations.R
\name{opensearchserviceserverless_get_access_policy}
\alias{opensearchserviceserverless_get_access_policy}
\title{Returns an OpenSearch Serverless access policy}
\usage{
opensearchserviceserverless_get_access_policy(name, type)
}
\arguments{
\item{name}{[required] The name of the access policy.}

\item{type}{[required] Tye type of policy. Currently, the only supported value is \code{data}.}
}
\description{
Returns an OpenSearch Serverless access policy. For more information, see \href{https://docs.aws.amazon.com/opensearch-service/latest/developerguide/serverless-data-access.html}{Data access control for Amazon OpenSearch Serverless}.

See \url{https://www.paws-r-sdk.com/docs/opensearchserviceserverless_get_access_policy/} for full documentation.
}
\keyword{internal}
