% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchserviceserverless_operations.R
\name{opensearchserviceserverless_get_account_settings}
\alias{opensearchserviceserverless_get_account_settings}
\title{Returns account-level settings related to OpenSearch Serverless}
\usage{
opensearchserviceserverless_get_account_settings()
}
\description{
Returns account-level settings related to OpenSearch Serverless.

See \url{https://www.paws-r-sdk.com/docs/opensearchserviceserverless_get_account_settings/} for full documentation.
}
\keyword{internal}
