% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchserviceserverless_operations.R
\name{opensearchserviceserverless_get_index}
\alias{opensearchserviceserverless_get_index}
\title{Retrieves information about an index in an OpenSearch Serverless
collection, including its schema definition}
\usage{
opensearchserviceserverless_get_index(id, indexName)
}
\arguments{
\item{id}{[required] The unique identifier of the collection containing the index.}

\item{indexName}{[required] The name of the index to retrieve information about.}
}
\description{
Retrieves information about an index in an OpenSearch Serverless collection, including its schema definition. The index might be configured to conduct automatic semantic enrichment ingestion and search. For more information, see \href{https://docs.aws.amazon.com/opensearch-service/latest/developerguide/serverless-manage.html#serverless-semantic-enrichment}{About automatic semantic enrichment}.

See \url{https://www.paws-r-sdk.com/docs/opensearchserviceserverless_get_index/} for full documentation.
}
\keyword{internal}
