% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchserviceserverless_operations.R
\name{opensearchserviceserverless_get_policies_stats}
\alias{opensearchserviceserverless_get_policies_stats}
\title{Returns statistical information about your OpenSearch Serverless access
policies, security configurations, and security policies}
\usage{
opensearchserviceserverless_get_policies_stats()
}
\description{
Returns statistical information about your OpenSearch Serverless access policies, security configurations, and security policies.

See \url{https://www.paws-r-sdk.com/docs/opensearchserviceserverless_get_policies_stats/} for full documentation.
}
\keyword{internal}
