% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchserviceserverless_operations.R
\name{opensearchserviceserverless_get_security_config}
\alias{opensearchserviceserverless_get_security_config}
\title{Returns information about an OpenSearch Serverless security
configuration}
\usage{
opensearchserviceserverless_get_security_config(id)
}
\arguments{
\item{id}{[required] The unique identifier of the security configuration.}
}
\description{
Returns information about an OpenSearch Serverless security configuration. For more information, see \href{https://docs.aws.amazon.com/opensearch-service/latest/developerguide/serverless-saml.html}{SAML authentication for Amazon OpenSearch Serverless}.

See \url{https://www.paws-r-sdk.com/docs/opensearchserviceserverless_get_security_config/} for full documentation.
}
\keyword{internal}
