% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchserviceserverless_operations.R
\name{opensearchserviceserverless_list_tags_for_resource}
\alias{opensearchserviceserverless_list_tags_for_resource}
\title{Returns the tags for an OpenSearch Serverless resource}
\usage{
opensearchserviceserverless_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource. The resource must be
active (not in the \code{DELETING} state), and must be owned by the account
ID included in the request.}
}
\description{
Returns the tags for an OpenSearch Serverless resource. For more information, see \href{https://docs.aws.amazon.com/opensearch-service/latest/developerguide/tag-collection.html}{Tagging Amazon OpenSearch Serverless collections}.

See \url{https://www.paws-r-sdk.com/docs/opensearchserviceserverless_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
