% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchserviceserverless_operations.R
\name{opensearchserviceserverless_list_vpc_endpoints}
\alias{opensearchserviceserverless_list_vpc_endpoints}
\title{Returns the OpenSearch Serverless-managed interface VPC endpoints
associated with the current account}
\usage{
opensearchserviceserverless_list_vpc_endpoints(
  maxResults = NULL,
  nextToken = NULL,
  vpcEndpointFilters = NULL
)
}
\arguments{
\item{maxResults}{An optional parameter that specifies the maximum number of results to
return. You can use \code{nextToken} to get the next page of results. The
default is 20.}

\item{nextToken}{If your initial
\code{\link[=opensearchserviceserverless_list_vpc_endpoints]{list_vpc_endpoints}}
operation returns a \code{nextToken}, you can include the returned
\code{nextToken} in subsequent
\code{\link[=opensearchserviceserverless_list_vpc_endpoints]{list_vpc_endpoints}}
operations, which returns results in the next page.}

\item{vpcEndpointFilters}{Filter the results according to the current status of the VPC endpoint.
Possible statuses are \code{CREATING}, \code{DELETING}, \code{UPDATING}, \code{ACTIVE}, and
\code{FAILED}.}
}
\description{
Returns the OpenSearch Serverless-managed interface VPC endpoints associated with the current account. For more information, see \href{https://docs.aws.amazon.com/opensearch-service/latest/developerguide/serverless-vpc.html}{Access Amazon OpenSearch Serverless using an interface endpoint}.

See \url{https://www.paws-r-sdk.com/docs/opensearchserviceserverless_list_vpc_endpoints/} for full documentation.
}
\keyword{internal}
