% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchserviceserverless_operations.R
\name{opensearchserviceserverless_tag_resource}
\alias{opensearchserviceserverless_tag_resource}
\title{Associates tags with an OpenSearch Serverless resource}
\usage{
opensearchserviceserverless_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource. The resource must be
active (not in the \code{DELETING} state), and must be owned by the account
ID included in the request.}

\item{tags}{[required] A list of tags (key-value pairs) to add to the resource. All tag keys in
the request must be unique.}
}
\description{
Associates tags with an OpenSearch Serverless resource. For more information, see \href{https://docs.aws.amazon.com/opensearch-service/latest/developerguide/tag-collection.html}{Tagging Amazon OpenSearch Serverless collections}.

See \url{https://www.paws-r-sdk.com/docs/opensearchserviceserverless_tag_resource/} for full documentation.
}
\keyword{internal}
