% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchserviceserverless_operations.R
\name{opensearchserviceserverless_update_account_settings}
\alias{opensearchserviceserverless_update_account_settings}
\title{Update the OpenSearch Serverless settings for the current Amazon Web
Services account}
\usage{
opensearchserviceserverless_update_account_settings(capacityLimits = NULL)
}
\arguments{
\item{capacityLimits}{}
}
\description{
Update the OpenSearch Serverless settings for the current Amazon Web Services account. For more information, see \href{https://docs.aws.amazon.com/opensearch-service/latest/developerguide/serverless-scaling.html}{Managing capacity limits for Amazon OpenSearch Serverless}.

See \url{https://www.paws-r-sdk.com/docs/opensearchserviceserverless_update_account_settings/} for full documentation.
}
\keyword{internal}
