% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchserviceserverless_operations.R
\name{opensearchserviceserverless_update_collection}
\alias{opensearchserviceserverless_update_collection}
\title{Updates an OpenSearch Serverless collection}
\usage{
opensearchserviceserverless_update_collection(
  id,
  description = NULL,
  vectorOptions = NULL,
  deletionProtection = NULL,
  clientToken = NULL
)
}
\arguments{
\item{id}{[required] The unique identifier of the collection.}

\item{description}{A description of the collection.}

\item{vectorOptions}{Configuration options for vector search capabilities in the collection.}

\item{deletionProtection}{Indicates whether to enable or disable deletion protection for the collection. When set to \code{ENABLED}, the collection cannot be deleted.}

\item{clientToken}{Unique, case-sensitive identifier to ensure idempotency of the request.}
}
\description{
Updates an OpenSearch Serverless collection.

See \url{https://www.paws-r-sdk.com/docs/opensearchserviceserverless_update_collection/} for full documentation.
}
\keyword{internal}
