% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchserviceserverless_operations.R
\name{opensearchserviceserverless_update_collection_group}
\alias{opensearchserviceserverless_update_collection_group}
\title{Updates the description and capacity limits of a collection group}
\usage{
opensearchserviceserverless_update_collection_group(
  id,
  description = NULL,
  capacityLimits = NULL,
  clientToken = NULL
)
}
\arguments{
\item{id}{[required] The unique identifier of the collection group to update.}

\item{description}{A new description for the collection group.}

\item{capacityLimits}{Updated capacity limits for the collection group, in OpenSearch Compute
Units (OCUs).}

\item{clientToken}{Unique, case-sensitive identifier to ensure idempotency of the request.}
}
\description{
Updates the description and capacity limits of a collection group.

See \url{https://www.paws-r-sdk.com/docs/opensearchserviceserverless_update_collection_group/} for full documentation.
}
\keyword{internal}
