% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchserviceserverless_operations.R
\name{opensearchserviceserverless_update_index}
\alias{opensearchserviceserverless_update_index}
\title{Updates an existing index in an OpenSearch Serverless collection}
\usage{
opensearchserviceserverless_update_index(id, indexName, indexSchema = NULL)
}
\arguments{
\item{id}{[required] The unique identifier of the collection containing the index to update.}

\item{indexName}{[required] The name of the index to update.}

\item{indexSchema}{The updated JSON schema definition for the index, including field
mappings and settings.}
}
\description{
Updates an existing index in an OpenSearch Serverless collection. This operation allows you to modify the index schema, including adding new fields or changing field mappings. You can also enable automatic semantic enrichment ingestion and search. For more information, see \href{https://docs.aws.amazon.com/opensearch-service/latest/developerguide/serverless-manage.html#serverless-semantic-enrichment}{About automatic semantic enrichment}.

See \url{https://www.paws-r-sdk.com/docs/opensearchserviceserverless_update_index/} for full documentation.
}
\keyword{internal}
