% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchserviceserverless_operations.R
\name{opensearchserviceserverless_update_lifecycle_policy}
\alias{opensearchserviceserverless_update_lifecycle_policy}
\title{Updates an OpenSearch Serverless access policy}
\usage{
opensearchserviceserverless_update_lifecycle_policy(
  type,
  name,
  policyVersion,
  description = NULL,
  policy = NULL,
  clientToken = NULL
)
}
\arguments{
\item{type}{[required] The type of lifecycle policy.}

\item{name}{[required] The name of the policy.}

\item{policyVersion}{[required] The version of the policy being updated.}

\item{description}{A description of the lifecycle policy.}

\item{policy}{The JSON policy document to use as the content for the lifecycle policy.}

\item{clientToken}{A unique, case-sensitive identifier to ensure idempotency of the
request.}
}
\description{
Updates an OpenSearch Serverless access policy. For more information, see \href{https://docs.aws.amazon.com/opensearch-service/latest/developerguide/serverless-lifecycle.html#serverless-lifecycle-update}{Updating data lifecycle policies}.

See \url{https://www.paws-r-sdk.com/docs/opensearchserviceserverless_update_lifecycle_policy/} for full documentation.
}
\keyword{internal}
