% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_batch_create_topic_reviewed_answer}
\alias{quicksight_batch_create_topic_reviewed_answer}
\title{Creates new reviewed answers for a Q Topic}
\usage{
quicksight_batch_create_topic_reviewed_answer(AwsAccountId, TopicId, Answers)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that you want to create a
reviewed answer in.}

\item{TopicId}{[required] The ID for the topic reviewed answer that you want to create. This ID is
unique per Amazon Web Services Region for each Amazon Web Services
account.}

\item{Answers}{[required] The definition of the Answers to be created.}
}
\description{
Creates new reviewed answers for a Q Topic.

See \url{https://www.paws-r-sdk.com/docs/quicksight_batch_create_topic_reviewed_answer/} for full documentation.
}
\keyword{internal}
