% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_batch_delete_knowledge_base}
\alias{quicksight_batch_delete_knowledge_base}
\title{Deletes one or more knowledge bases}
\usage{
quicksight_batch_delete_knowledge_base(AwsAccountId, KnowledgeBaseIds)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the knowledge base.}

\item{KnowledgeBaseIds}{[required] A list of knowledge base identifiers to delete.}
}
\description{
Deletes one or more knowledge bases.

See \url{https://www.paws-r-sdk.com/docs/quicksight_batch_delete_knowledge_base/} for full documentation.
}
\keyword{internal}
