% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_batch_delete_topic_reviewed_answer}
\alias{quicksight_batch_delete_topic_reviewed_answer}
\title{Deletes reviewed answers for Q Topic}
\usage{
quicksight_batch_delete_topic_reviewed_answer(
  AwsAccountId,
  TopicId,
  AnswerIds = NULL
)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that you want to delete a
reviewed answers in.}

\item{TopicId}{[required] The ID for the topic reviewed answer that you want to delete. This ID is
unique per Amazon Web Services Region for each Amazon Web Services
account.}

\item{AnswerIds}{The Answer IDs of the Answers to be deleted.}
}
\description{
Deletes reviewed answers for Q Topic.

See \url{https://www.paws-r-sdk.com/docs/quicksight_batch_delete_topic_reviewed_answer/} for full documentation.
}
\keyword{internal}
