% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_batch_describe_user_limits}
\alias{quicksight_batch_describe_user_limits}
\title{Describes the effective resource limits for one or more Amazon Quick
Sight users, including the limits that apply to each user based on their
profile assignments}
\usage{
quicksight_batch_describe_user_limits(
  accountId,
  users = NULL,
  resourceTypes = NULL
)
}
\arguments{
\item{accountId}{[required] The ID of the Amazon Web Services account that contains the users.}

\item{users}{A list of users to describe limits for. Each entry contains a user name and namespace.}

\item{resourceTypes}{An optional filter that limits the results to specific resource types. If you don't specify a value, the operation returns limits for all resource types.}
}
\description{
Describes the effective resource limits for one or more Amazon Quick Sight users, including the limits that apply to each user based on their profile assignments.

See \url{https://www.paws-r-sdk.com/docs/quicksight_batch_describe_user_limits/} for full documentation.
}
\keyword{internal}
