% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_create_agent}
\alias{quicksight_create_agent}
\title{Creates an agent in Amazon QuickSight}
\usage{
quicksight_create_agent(
  Spaces = NULL,
  ActionConnectors = NULL,
  AwsAccountId,
  AgentId,
  Name,
  Description = NULL,
  IconId = NULL,
  StarterPrompts = NULL,
  WelcomeMessage = NULL,
  AgentLifecycle = NULL,
  CustomPromptInput = NULL
)
}
\arguments{
\item{Spaces}{The Amazon Resource Names (ARNs) of the spaces to attach to the agent.}

\item{ActionConnectors}{The Amazon Resource Names (ARNs) of the action connectors to attach to the agent.}

\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the agent.}

\item{AgentId}{[required] A unique identifier for the agent.}

\item{Name}{[required] The name of the agent.}

\item{Description}{A description of the agent.}

\item{IconId}{The icon identifier for the agent.}

\item{StarterPrompts}{A list of starter prompts that are displayed to users when they begin interacting with the agent.}

\item{WelcomeMessage}{The welcome message that is displayed when a user starts a conversation with the agent.}

\item{AgentLifecycle}{The lifecycle state of the agent. Valid values are \code{PREVIEW} and \code{PUBLISHED}.}

\item{CustomPromptInput}{The custom prompt configuration for the agent.}
}
\description{
Creates an agent in Amazon QuickSight.

See \url{https://www.paws-r-sdk.com/docs/quicksight_create_agent/} for full documentation.
}
\keyword{internal}
