% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_create_approval_policy}
\alias{quicksight_create_approval_policy}
\title{Creates an approval policy in Quick Sight}
\usage{
quicksight_create_approval_policy(
  PolicyId,
  Name,
  Description = NULL,
  Actions,
  AssetTypes,
  ApplicableTo,
  ApprovalGroups
)
}
\arguments{
\item{PolicyId}{[required] The unique identifier to assign to the approval policy. You cannot change this value after you create the policy.}

\item{Name}{[required] The name of the approval policy.}

\item{Description}{A description of the approval policy.}

\item{Actions}{[required] The list of governed actions that trigger the approval workflow.}

\item{AssetTypes}{[required] The list of asset types that the approval policy applies to.}

\item{ApplicableTo}{[required] The scoping configuration that determines who the approval policy applies to.}

\item{ApprovalGroups}{[required] The list of group ARNs whose members can approve requests.}
}
\description{
Creates an approval policy in Quick Sight.

See \url{https://www.paws-r-sdk.com/docs/quicksight_create_approval_policy/} for full documentation.
}
\keyword{internal}
