% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_create_custom_permissions}
\alias{quicksight_create_custom_permissions}
\title{Creates a custom permissions profile}
\usage{
quicksight_create_custom_permissions(
  AwsAccountId,
  CustomPermissionsName,
  Capabilities = NULL,
  Governance = NULL,
  Tags = NULL
)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that you want to create the custom permissions profile in.}

\item{CustomPermissionsName}{[required] The name of the custom permissions profile that you want to create.}

\item{Capabilities}{A set of actions to include in the custom permissions profile.}

\item{Governance}{The governance configuration for the custom permissions profile. When governance controls are defined for a category, any capabilities in that category not explicitly set to \code{ALLOW} in \code{Capabilities} are denied. Even newly added capabilities in the category are implicitly disabled when Amazon Quick releases them.}

\item{Tags}{The tags to associate with the custom permissions profile.}
}
\description{
Creates a custom permissions profile.

See \url{https://www.paws-r-sdk.com/docs/quicksight_create_custom_permissions/} for full documentation.
}
\keyword{internal}
