% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_create_flow}
\alias{quicksight_create_flow}
\title{Creates a new flow in the specified Amazon Web Services account}
\usage{
quicksight_create_flow(
  AwsAccountId,
  Name,
  Description = NULL,
  FlowDefinition,
  Permissions = NULL,
  ClientToken = NULL
)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account where you want to create the flow.}

\item{Name}{[required] The display name for the flow.}

\item{Description}{The description for the flow.}

\item{FlowDefinition}{[required] The definition of the flow, specifying the steps and configurations. This is the flow definition in Quick Flow's internal format. The format is subject to change.

Always derive or depend on the flow definition from the \code{\link[=quicksight_describe_flow]{describe_flow}} operation to ensure you are working with the latest format.}

\item{Permissions}{Initial permissions for the flow. If omitted, the flow is created without any permissions.}

\item{ClientToken}{A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.}
}
\description{
Creates a new flow in the specified Amazon Web Services account. Creates both a DRAFT and PUBLISHED (auto-published) version.

See \url{https://www.paws-r-sdk.com/docs/quicksight_create_flow/} for full documentation.
}
\keyword{internal}
