% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_create_folder}
\alias{quicksight_create_folder}
\title{Creates an empty shared folder}
\usage{
quicksight_create_folder(
  AwsAccountId,
  FolderId,
  Name = NULL,
  FolderType = NULL,
  ParentFolderArn = NULL,
  Permissions = NULL,
  Tags = NULL,
  SharingModel = NULL
)
}
\arguments{
\item{AwsAccountId}{[required] The ID for the Amazon Web Services account where you want to create the
folder.}

\item{FolderId}{[required] The ID of the folder.}

\item{Name}{The name of the folder.}

\item{FolderType}{The type of folder. By default, \code{folderType} is \code{SHARED}.}

\item{ParentFolderArn}{The Amazon Resource Name (ARN) for the parent folder.

\code{ParentFolderArn} can be null. An empty \code{parentFolderArn} creates a
root-level folder.}

\item{Permissions}{A structure that describes the principals and the resource-level
permissions of a folder.

To specify no permissions, omit \code{Permissions}.}

\item{Tags}{Tags for the folder.}

\item{SharingModel}{An optional parameter that determines the sharing scope of the folder.
The default value for this parameter is \code{ACCOUNT}.}
}
\description{
Creates an empty shared folder.

See \url{https://www.paws-r-sdk.com/docs/quicksight_create_folder/} for full documentation.
}
\keyword{internal}
