% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_create_ingestion}
\alias{quicksight_create_ingestion}
\title{Creates and starts a new SPICE ingestion for a dataset}
\usage{
quicksight_create_ingestion(
  DataSetId,
  IngestionId,
  AwsAccountId,
  IngestionType = NULL
)
}
\arguments{
\item{DataSetId}{[required] The ID of the dataset used in the ingestion.}

\item{IngestionId}{[required] An ID for the ingestion.}

\item{AwsAccountId}{[required] The Amazon Web Services account ID.}

\item{IngestionType}{The type of ingestion that you want to create.}
}
\description{
Creates and starts a new SPICE ingestion for a dataset. You can manually refresh datasets in an Enterprise edition account 32 times in a 24-hour period. You can manually refresh datasets in a Standard edition account 8 times in a 24-hour period. Each 24-hour period is measured starting 24 hours before the current date and time.

See \url{https://paws-r.github.io/docs/quicksight/create_ingestion.html} for full documentation.
}
\keyword{internal}
