% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_create_limits_profile}
\alias{quicksight_create_limits_profile}
\title{Creates a limits profile that defines resource usage limits for Amazon
Quick Sight users}
\usage{
quicksight_create_limits_profile(
  accountId,
  profileName,
  description = NULL,
  resourceLimits,
  clientToken
)
}
\arguments{
\item{accountId}{[required] The ID of the Amazon Web Services account that contains the limits profile.}

\item{profileName}{[required] A display name for the limits profile.}

\item{description}{A description for the limits profile.}

\item{resourceLimits}{[required] A map of resource types to their limit values for this profile.}

\item{clientToken}{[required] A unique, case-sensitive identifier to ensure that the operation completes no more than one time. If this token matches a previous request, the service ignores the request, but does not return an error.}
}
\description{
Creates a limits profile that defines resource usage limits for Amazon Quick Sight users.

See \url{https://www.paws-r-sdk.com/docs/quicksight_create_limits_profile/} for full documentation.
}
\keyword{internal}
