% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_create_o_auth_client_application}
\alias{quicksight_create_o_auth_client_application}
\title{Creates an OAuthClientApplication}
\usage{
quicksight_create_o_auth_client_application(
  AwsAccountId,
  OAuthClientApplicationId,
  Name,
  OAuthClientAuthenticationType,
  ClientId,
  ClientSecret,
  OAuthTokenEndpointUrl,
  OAuthAuthorizationEndpointUrl = NULL,
  OAuthScopes = NULL,
  DataSourceType = NULL,
  IdentityProviderVpcConnectionProperties = NULL,
  Tags = NULL
)
}
\arguments{
\item{AwsAccountId}{[required] The Amazon Web Services account ID.}

\item{OAuthClientApplicationId}{[required] An ID for the OAuthClientApplication that you want to create. This ID is unique per Amazon Web Services Region for each Amazon Web Services account.}

\item{Name}{[required] The display name for the OAuthClientApplication.}

\item{OAuthClientAuthenticationType}{[required] The authentication type to use for the OAuthClientApplication. This determines the OAuth 2.0 grant flow that is used when the data source connects to the identity provider. Valid values are \code{TOKEN}.}

\item{ClientId}{[required] The client ID of the OAuth application that is registered with the identity provider.}

\item{ClientSecret}{[required] The client secret of the OAuth application that is registered with the identity provider.}

\item{OAuthTokenEndpointUrl}{[required] The token endpoint URL of the identity provider that is used to obtain access tokens.}

\item{OAuthAuthorizationEndpointUrl}{The authorization endpoint URL of the identity provider that is used to obtain authorization codes.}

\item{OAuthScopes}{The OAuth scopes that are requested when the OAuthClientApplication obtains an access token from the identity provider.}

\item{DataSourceType}{The type of data source that the OAuthClientApplication is used with. Valid values are \code{SNOWFLAKE}.}

\item{IdentityProviderVpcConnectionProperties}{VPC connection properties.}

\item{Tags}{Contains a map of the key-value pairs for the resource tag or tags assigned to the OAuthClientApplication.}
}
\description{
Creates an OAuthClientApplication.

See \url{https://www.paws-r-sdk.com/docs/quicksight_create_o_auth_client_application/} for full documentation.
}
\keyword{internal}
