% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_create_space}
\alias{quicksight_create_space}
\title{Creates a new Amazon QuickSight space}
\usage{
quicksight_create_space(AwsAccountId, SpaceId, Name, Description = NULL)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the space.}

\item{SpaceId}{[required] The ID of the space. This ID is unique per Amazon Web Services Region for each Amazon Web Services account.}

\item{Name}{[required] A display name for the space.}

\item{Description}{A description of the space.}
}
\description{
Creates a new Amazon QuickSight space. A space is a collection of resources that can be used to organize and manage QuickSight assets.

See \url{https://www.paws-r-sdk.com/docs/quicksight_create_space/} for full documentation.
}
\keyword{internal}
