% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_create_template}
\alias{quicksight_create_template}
\title{Creates a template either from a TemplateDefinition or from an existing
Quick Sight analysis or template}
\usage{
quicksight_create_template(
  AwsAccountId,
  TemplateId,
  Name = NULL,
  Permissions = NULL,
  SourceEntity = NULL,
  Tags = NULL,
  VersionDescription = NULL,
  Definition = NULL,
  ValidationStrategy = NULL
)
}
\arguments{
\item{AwsAccountId}{[required] The ID for the Amazon Web Services account that the group is in. You use the ID for the Amazon Web Services account that contains your Amazon Quick Sight account.}

\item{TemplateId}{[required] An ID for the template that you want to create. This template is unique per Amazon Web Services Region; in each Amazon Web Services account.}

\item{Name}{A display name for the template.}

\item{Permissions}{A list of resource permissions to be set on the template.}

\item{SourceEntity}{The entity that you are using as a source when you create the template. In \code{SourceEntity}, you specify the type of object you're using as source: \code{SourceTemplate} for a template or \code{SourceAnalysis} for an analysis. Both of these require an Amazon Resource Name (ARN). For \code{SourceTemplate}, specify the ARN of the source template. For \code{SourceAnalysis}, specify the ARN of the source analysis. The \code{SourceTemplate} ARN can contain any Amazon Web Services account and any Quick Sight-supported Amazon Web Services Region.

Use the \code{DataSetReferences} entity within \code{SourceTemplate} or \code{SourceAnalysis} to list the replacement datasets for the placeholders listed in the original. The schema in each dataset must match its placeholder. Use the \code{TopicReferences} entity to list the replacement topics for the topic placeholders listed in the original. The schema in each topic must match its placeholder.

Either a \code{SourceEntity} or a \code{Definition} must be provided in order for the request to be valid.}

\item{Tags}{Contains a map of the key-value pairs for the resource tag or tags assigned to the resource.}

\item{VersionDescription}{A description of the current template version being created. This API operation creates the first version of the template. Every time \code{\link[=quicksight_update_template]{update_template}} is called, a new version is created. Each version of the template maintains a description of the version in the \code{VersionDescription} field.}

\item{Definition}{The definition of a template.

A definition is the data model of all features in a Dashboard, Template, or Analysis.

Either a \code{SourceEntity} or a \code{Definition} must be provided in order for the request to be valid.}

\item{ValidationStrategy}{TThe option to relax the validation needed to create a template with definition objects. This skips the validation step for specific errors.}
}
\description{
Creates a template either from a \code{TemplateDefinition} or from an existing Quick Sight analysis or template. You can use the resulting template to create additional dashboards, templates, or analyses.

See \url{https://www.paws-r-sdk.com/docs/quicksight_create_template/} for full documentation.
}
\keyword{internal}
