% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_create_theme_alias}
\alias{quicksight_create_theme_alias}
\title{Creates a theme alias for a theme}
\usage{
quicksight_create_theme_alias(AwsAccountId, ThemeId, AliasName,
  ThemeVersionNumber)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the AWS account that contains the theme for the new theme
alias.}

\item{ThemeId}{[required] An ID for the theme alias.}

\item{AliasName}{[required] The name that you want to give to the theme alias that you are creating.
The alias name can't begin with a \code{$}. Alias names that start with \code{$}
are reserved by Amazon QuickSight.}

\item{ThemeVersionNumber}{[required] The version number of the theme.}
}
\description{
Creates a theme alias for a theme.
}
\section{Request syntax}{
\preformatted{svc$create_theme_alias(
  AwsAccountId = "string",
  ThemeId = "string",
  AliasName = "string",
  ThemeVersionNumber = 123
)
}
}

\keyword{internal}
