% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_create_topic}
\alias{quicksight_create_topic}
\title{Creates a new Q topic}
\usage{
quicksight_create_topic(
  AwsAccountId,
  TopicId,
  Topic,
  Tags = NULL,
  FolderArns = NULL,
  CustomInstructions = NULL
)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that you want to create a
topic in.}

\item{TopicId}{[required] The ID for the topic that you want to create. This ID is unique per
Amazon Web Services Region for each Amazon Web Services account.}

\item{Topic}{[required] The definition of a topic to create.}

\item{Tags}{Contains a map of the key-value pairs for the resource tag or tags that
are assigned to the dataset.}

\item{FolderArns}{The Folder ARN of the folder that you want the topic to reside in.}

\item{CustomInstructions}{Custom instructions for the topic.}
}
\description{
Creates a new Q topic.

See \url{https://www.paws-r-sdk.com/docs/quicksight_create_topic/} for full documentation.
}
\keyword{internal}
