% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_create_topic_refresh_schedule}
\alias{quicksight_create_topic_refresh_schedule}
\title{Creates a topic refresh schedule}
\usage{
quicksight_create_topic_refresh_schedule(
  AwsAccountId,
  TopicId,
  DatasetArn,
  DatasetName = NULL,
  RefreshSchedule
)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the topic you're
creating a refresh schedule for.}

\item{TopicId}{[required] The ID of the topic that you want to modify. This ID is unique per
Amazon Web Services Region for each Amazon Web Services account.}

\item{DatasetArn}{[required] The Amazon Resource Name (ARN) of the dataset.}

\item{DatasetName}{The name of the dataset.}

\item{RefreshSchedule}{[required] The definition of a refresh schedule.}
}
\description{
Creates a topic refresh schedule.

See \url{https://www.paws-r-sdk.com/docs/quicksight_create_topic_refresh_schedule/} for full documentation.
}
\keyword{internal}
