% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_delete_account_customization}
\alias{quicksight_delete_account_customization}
\title{Deletes all Amazon QuickSight customizations in this Amazon Web Services
Region for the specified Amazon Web Services account and Amazon
QuickSight namespace}
\usage{
quicksight_delete_account_customization(AwsAccountId, Namespace = NULL)
}
\arguments{
\item{AwsAccountId}{[required] The ID for the Amazon Web Services account that you want to delete
Amazon QuickSight customizations from in this Amazon Web Services
Region.}

\item{Namespace}{The Amazon QuickSight namespace that you're deleting the customizations
from.}
}
\description{
Deletes all Amazon QuickSight customizations in this Amazon Web Services Region for the specified Amazon Web Services account and Amazon QuickSight namespace.

See \url{https://paws-r.github.io/docs/quicksight/delete_account_customization.html} for full documentation.
}
\keyword{internal}
