% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_delete_account_subscription}
\alias{quicksight_delete_account_subscription}
\title{Use the DeleteAccountSubscription operation to delete an Amazon
QuickSight account}
\usage{
quicksight_delete_account_subscription(AwsAccountId)
}
\arguments{
\item{AwsAccountId}{[required] The Amazon Web Services account ID of the account that you want to
delete.}
}
\description{
Use the \code{\link[=quicksight_delete_account_subscription]{delete_account_subscription}} operation to delete an Amazon QuickSight account. This operation will result in an error message if you have configured your account termination protection settings to \code{True}. To change this setting and delete your account, call the \code{\link[=quicksight_update_account_settings]{update_account_settings}} API and set the value of the \code{TerminationProtectionEnabled} parameter to \code{False}, then make another call to the \code{\link[=quicksight_delete_account_subscription]{delete_account_subscription}} API.

See \url{https://www.paws-r-sdk.com/docs/quicksight_delete_account_subscription/} for full documentation.
}
\keyword{internal}
