% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_delete_action_connector}
\alias{quicksight_delete_action_connector}
\title{Hard deletes an action connector, making it unrecoverable}
\usage{
quicksight_delete_action_connector(AwsAccountId, ActionConnectorId)
}
\arguments{
\item{AwsAccountId}{[required] The Amazon Web Services account ID that contains the action connector to
delete.}

\item{ActionConnectorId}{[required] The unique identifier of the action connector to delete.}
}
\description{
Hard deletes an action connector, making it unrecoverable. This operation removes the connector and all its associated configurations. Any resources currently using this action connector will no longer be able to perform actions through it.

See \url{https://www.paws-r-sdk.com/docs/quicksight_delete_action_connector/} for full documentation.
}
\keyword{internal}
