% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_delete_agent}
\alias{quicksight_delete_agent}
\title{Deletes an agent}
\usage{
quicksight_delete_agent(AgentId, AwsAccountId)
}
\arguments{
\item{AgentId}{[required] The unique identifier for the agent to delete.}

\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the agent.}
}
\description{
Deletes an agent.

See \url{https://www.paws-r-sdk.com/docs/quicksight_delete_agent/} for full documentation.
}
\keyword{internal}
