% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_delete_app}
\alias{quicksight_delete_app}
\title{Deletes an app}
\usage{
quicksight_delete_app(AwsAccountId, AppId)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the app.}

\item{AppId}{[required] The ID of the app that you want to delete.}
}
\description{
Deletes an app.

See \url{https://www.paws-r-sdk.com/docs/quicksight_delete_app/} for full documentation.
}
\keyword{internal}
