% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_delete_approval_policy}
\alias{quicksight_delete_approval_policy}
\title{Deletes an approval policy in Quick Sight}
\usage{
quicksight_delete_approval_policy(PolicyId)
}
\arguments{
\item{PolicyId}{[required] The unique identifier of the approval policy to delete.}
}
\description{
Deletes an approval policy in Quick Sight.

See \url{https://www.paws-r-sdk.com/docs/quicksight_delete_approval_policy/} for full documentation.
}
\keyword{internal}
