% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_delete_brand}
\alias{quicksight_delete_brand}
\title{This API permanently deletes the specified Quick Sight brand}
\usage{
quicksight_delete_brand(AwsAccountId, BrandId)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that owns the brand.}

\item{BrandId}{[required] The ID of the Quick Suite brand.}
}
\description{
This API permanently deletes the specified Quick Sight brand. When you delete a brand:

See \url{https://www.paws-r-sdk.com/docs/quicksight_delete_brand/} for full documentation.
}
\keyword{internal}
