% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_delete_custom_permissions}
\alias{quicksight_delete_custom_permissions}
\title{Deletes a custom permissions profile}
\usage{
quicksight_delete_custom_permissions(AwsAccountId, CustomPermissionsName)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the custom
permissions profile that you want to delete.}

\item{CustomPermissionsName}{[required] The name of the custom permissions profile that you want to delete.}
}
\description{
Deletes a custom permissions profile.

See \url{https://www.paws-r-sdk.com/docs/quicksight_delete_custom_permissions/} for full documentation.
}
\keyword{internal}
