% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_delete_dashboard}
\alias{quicksight_delete_dashboard}
\title{Deletes a dashboard}
\usage{
quicksight_delete_dashboard(AwsAccountId, DashboardId, VersionNumber)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the AWS account that contains the dashboard that you're
deleting.}

\item{DashboardId}{[required] The ID for the dashboard.}

\item{VersionNumber}{The version number of the dashboard. If the version number property is
provided, only the specified version of the dashboard is deleted.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Status = 123,
  Arn = "string",
  DashboardId = "string",
  RequestId = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Deletes a dashboard.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_dashboard(
  AwsAccountId = "string",
  DashboardId = "string",
  VersionNumber = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
