% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_delete_data_set_refresh_properties}
\alias{quicksight_delete_data_set_refresh_properties}
\title{Deletes the dataset refresh properties of the dataset}
\usage{
quicksight_delete_data_set_refresh_properties(AwsAccountId, DataSetId)
}
\arguments{
\item{AwsAccountId}{[required] The Amazon Web Services account ID.}

\item{DataSetId}{[required] The ID of the dataset.}
}
\description{
Deletes the dataset refresh properties of the dataset.

See \url{https://paws-r.github.io/docs/quicksight/delete_data_set_refresh_properties.html} for full documentation.
}
\keyword{internal}
