% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_delete_dlp_setting}
\alias{quicksight_delete_dlp_setting}
\title{Deletes a DLP setting configuration from an Amazon Web Services account}
\usage{
quicksight_delete_dlp_setting(AwsAccountId, DlpSettingId)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the DLP setting that you want to delete.}

\item{DlpSettingId}{[required] The ID of the DLP setting that you want to delete.}
}
\description{
Deletes a DLP setting configuration from an Amazon Web Services account.

See \url{https://www.paws-r-sdk.com/docs/quicksight_delete_dlp_setting/} for full documentation.
}
\keyword{internal}
