% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_delete_flow}
\alias{quicksight_delete_flow}
\title{Permanently deletes a flow from the specified Amazon Web Services
account}
\usage{
quicksight_delete_flow(AwsAccountId, FlowId)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the flow that you are deleting.}

\item{FlowId}{[required] The unique identifier of the flow to delete.}
}
\description{
Permanently deletes a flow from the specified Amazon Web Services account. This operation cannot be undone.

See \url{https://www.paws-r-sdk.com/docs/quicksight_delete_flow/} for full documentation.
}
\keyword{internal}
