% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_delete_folder_membership}
\alias{quicksight_delete_folder_membership}
\title{Removes an asset, such as a dashboard, analysis, or dataset, from a
folder}
\usage{
quicksight_delete_folder_membership(
  AwsAccountId,
  FolderId,
  MemberId,
  MemberType
)
}
\arguments{
\item{AwsAccountId}{[required] The ID for the Amazon Web Services account that contains the folder.}

\item{FolderId}{[required] The Folder ID.}

\item{MemberId}{[required] The ID of the asset that you want to delete.}

\item{MemberType}{[required] The member type of the asset that you want to delete from a folder.}
}
\description{
Removes an asset, such as a dashboard, analysis, or dataset, from a folder.

See \url{https://www.paws-r-sdk.com/docs/quicksight_delete_folder_membership/} for full documentation.
}
\keyword{internal}
