% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_delete_group}
\alias{quicksight_delete_group}
\title{Removes a user group from Amazon QuickSight}
\usage{
quicksight_delete_group(GroupName, AwsAccountId, Namespace)
}
\arguments{
\item{GroupName}{[required] The name of the group that you want to delete.}

\item{AwsAccountId}{[required] The ID for the AWS account that the group is in. Currently, you use the
ID for the AWS account that contains your Amazon QuickSight account.}

\item{Namespace}{[required] The namespace. Currently, you should set this to \code{default}.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  RequestId = "string",
  Status = 123
)
}\if{html}{\out{</div>}}
}
\description{
Removes a user group from Amazon QuickSight.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_group(
  GroupName = "string",
  AwsAccountId = "string",
  Namespace = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
