% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_delete_group_membership}
\alias{quicksight_delete_group_membership}
\title{Removes a user from a group so that the user is no longer a member of
the group}
\usage{
quicksight_delete_group_membership(
  MemberName,
  GroupName,
  AwsAccountId,
  Namespace
)
}
\arguments{
\item{MemberName}{[required] The name of the user that you want to delete from the group membership.}

\item{GroupName}{[required] The name of the group that you want to delete the user from.}

\item{AwsAccountId}{[required] The ID for the Amazon Web Services account that the group is in.
Currently, you use the ID for the Amazon Web Services account that
contains your Amazon Quick Sight account.}

\item{Namespace}{[required] The namespace of the group that you want to remove a user from.}
}
\description{
Removes a user from a group so that the user is no longer a member of the group.

See \url{https://www.paws-r-sdk.com/docs/quicksight_delete_group_membership/} for full documentation.
}
\keyword{internal}
