% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_delete_iam_policy_assignment}
\alias{quicksight_delete_iam_policy_assignment}
\title{Deletes an existing IAM policy assignment}
\usage{
quicksight_delete_iam_policy_assignment(AwsAccountId, AssignmentName,
  Namespace)
}
\arguments{
\item{AwsAccountId}{[required] The AWS account ID where you want to delete the IAM policy assignment.}

\item{AssignmentName}{[required] The name of the assignment.}

\item{Namespace}{[required] The namespace that contains the assignment.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  AssignmentName = "string",
  RequestId = "string",
  Status = 123
)
}\if{html}{\out{</div>}}
}
\description{
Deletes an existing IAM policy assignment.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_iam_policy_assignment(
  AwsAccountId = "string",
  AssignmentName = "string",
  Namespace = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
