% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_delete_knowledge_base}
\alias{quicksight_delete_knowledge_base}
\title{Deletes a knowledge base}
\usage{
quicksight_delete_knowledge_base(AwsAccountId, KnowledgeBaseId)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the knowledge base.}

\item{KnowledgeBaseId}{[required] The unique identifier for the knowledge base.}
}
\description{
Deletes a knowledge base.

See \url{https://www.paws-r-sdk.com/docs/quicksight_delete_knowledge_base/} for full documentation.
}
\keyword{internal}
