% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_delete_limits_profile}
\alias{quicksight_delete_limits_profile}
\title{Deletes a limits profile}
\usage{
quicksight_delete_limits_profile(profileId, accountId)
}
\arguments{
\item{profileId}{[required] The unique identifier for the limits profile to delete.}

\item{accountId}{[required] The ID of the Amazon Web Services account that contains the limits profile.}
}
\description{
Deletes a limits profile.

See \url{https://www.paws-r-sdk.com/docs/quicksight_delete_limits_profile/} for full documentation.
}
\keyword{internal}
