% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_delete_namespace}
\alias{quicksight_delete_namespace}
\title{Deletes a namespace and the users and groups that are associated with
the namespace}
\usage{
quicksight_delete_namespace(AwsAccountId, Namespace)
}
\arguments{
\item{AwsAccountId}{[required] The ID for the Amazon Web Services account that you want to delete the
Quick Sight namespace from.}

\item{Namespace}{[required] The namespace that you want to delete.}
}
\description{
Deletes a namespace and the users and groups that are associated with the namespace. This is an asynchronous process. Assets including dashboards, analyses, datasets and data sources are not deleted. To delete these assets, you use the API operations for the relevant asset.

See \url{https://www.paws-r-sdk.com/docs/quicksight_delete_namespace/} for full documentation.
}
\keyword{internal}
