% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_delete_o_auth_client_application}
\alias{quicksight_delete_o_auth_client_application}
\title{Deletes an OAuthClientApplication}
\usage{
quicksight_delete_o_auth_client_application(
  AwsAccountId,
  OAuthClientApplicationId
)
}
\arguments{
\item{AwsAccountId}{[required] The Amazon Web Services account ID.}

\item{OAuthClientApplicationId}{[required] The ID of the OAuthClientApplication that you want to delete.}
}
\description{
Deletes an OAuthClientApplication.

See \url{https://www.paws-r-sdk.com/docs/quicksight_delete_o_auth_client_application/} for full documentation.
}
\keyword{internal}
