% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_delete_role_custom_permission}
\alias{quicksight_delete_role_custom_permission}
\title{Removes custom permissions from the role}
\usage{
quicksight_delete_role_custom_permission(Role, AwsAccountId, Namespace)
}
\arguments{
\item{Role}{[required] The role that you want to remove permissions from.}

\item{AwsAccountId}{[required] The ID for the Amazon Web Services account that the group is in.
Currently, you use the ID for the Amazon Web Services account that
contains your Amazon Quick Sight account.}

\item{Namespace}{[required] The namespace that includes the role.}
}
\description{
Removes custom permissions from the role.

See \url{https://www.paws-r-sdk.com/docs/quicksight_delete_role_custom_permission/} for full documentation.
}
\keyword{internal}
