% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_delete_role_membership}
\alias{quicksight_delete_role_membership}
\title{Removes a group from a role}
\usage{
quicksight_delete_role_membership(MemberName, Role, AwsAccountId, Namespace)
}
\arguments{
\item{MemberName}{[required] The name of the group.}

\item{Role}{[required] The role that you want to remove permissions from.}

\item{AwsAccountId}{[required] The ID for the Amazon Web Services account that you want to create a
group in. The Amazon Web Services account ID that you provide must be
the same Amazon Web Services account that contains your Amazon Quick
Sight account.}

\item{Namespace}{[required] The namespace that contains the role.}
}
\description{
Removes a group from a role.

See \url{https://www.paws-r-sdk.com/docs/quicksight_delete_role_membership/} for full documentation.
}
\keyword{internal}
