% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_delete_space}
\alias{quicksight_delete_space}
\title{Deletes an Amazon QuickSight space}
\usage{
quicksight_delete_space(AwsAccountId, SpaceId)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the space.}

\item{SpaceId}{[required] The ID of the space that you want to delete.}
}
\description{
Deletes an Amazon QuickSight space.

See \url{https://www.paws-r-sdk.com/docs/quicksight_delete_space/} for full documentation.
}
\keyword{internal}
