% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_delete_template_alias}
\alias{quicksight_delete_template_alias}
\title{Update template alias of given template}
\usage{
quicksight_delete_template_alias(AwsAccountId, TemplateId, AliasName)
}
\arguments{
\item{AwsAccountId}{[required] AWS account ID that contains the template alias you are deleting.}

\item{TemplateId}{[required] An ID for the template.}

\item{AliasName}{[required] The alias of the template that you want to delete. If you provide a
specific alias, you delete the version that the alias points to. You can
specify the latest version of the template by providing the keyword
\verb{$LATEST} in the \code{AliasName} parameter.}
}
\description{
Update template alias of given template.
}
\section{Request syntax}{
\preformatted{svc$delete_template_alias(
  AwsAccountId = "string",
  TemplateId = "string",
  AliasName = "string"
)
}
}

\keyword{internal}
