% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_delete_theme}
\alias{quicksight_delete_theme}
\title{Deletes a theme}
\usage{
quicksight_delete_theme(AwsAccountId, ThemeId, VersionNumber)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the AWS account that contains the theme that you're deleting.}

\item{ThemeId}{[required] An ID for the theme that you want to delete.}

\item{VersionNumber}{The version of the theme that you want to delete.

\strong{Note:} If you don't provide a version number, you're using this call
to \code{\link[=quicksight_delete_theme]{delete_theme}} to delete all versions of
the theme.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Arn = "string",
  RequestId = "string",
  Status = 123,
  ThemeId = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Deletes a theme.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_theme(
  AwsAccountId = "string",
  ThemeId = "string",
  VersionNumber = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
