% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_delete_topic}
\alias{quicksight_delete_topic}
\title{Deletes a topic}
\usage{
quicksight_delete_topic(AwsAccountId, TopicId)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the topic that
you want to delete.}

\item{TopicId}{[required] The ID of the topic that you want to delete. This ID is unique per
Amazon Web Services Region for each Amazon Web Services account.}
}
\description{
Deletes a topic.

See \url{https://paws-r.github.io/docs/quicksight/delete_topic.html} for full documentation.
}
\keyword{internal}
