% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_delete_topic_refresh_schedule}
\alias{quicksight_delete_topic_refresh_schedule}
\title{Deletes a topic refresh schedule}
\usage{
quicksight_delete_topic_refresh_schedule(AwsAccountId, TopicId, DatasetId)
}
\arguments{
\item{AwsAccountId}{[required] The Amazon Web Services account ID.}

\item{TopicId}{[required] The ID of the topic that you want to modify. This ID is unique per
Amazon Web Services Region for each Amazon Web Services account.}

\item{DatasetId}{[required] The ID of the dataset.}
}
\description{
Deletes a topic refresh schedule.

See \url{https://paws-r.github.io/docs/quicksight/delete_topic_refresh_schedule.html} for full documentation.
}
\keyword{internal}
