% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_delete_topic_v2}
\alias{quicksight_delete_topic_v2}
\title{Deletes a Q topic}
\usage{
quicksight_delete_topic_v2(AwsAccountId, TopicId)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the topic that you want to delete.}

\item{TopicId}{[required] The ID of the topic that you want to delete. This ID is unique per Amazon Web Services Region for each Amazon Web Services account.}
}
\description{
Deletes a Q topic.

See \url{https://www.paws-r-sdk.com/docs/quicksight_delete_topic_v2/} for full documentation.
}
\keyword{internal}
