% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_delete_user_custom_permission}
\alias{quicksight_delete_user_custom_permission}
\title{Deletes a custom permissions profile from a user}
\usage{
quicksight_delete_user_custom_permission(UserName, AwsAccountId, Namespace)
}
\arguments{
\item{UserName}{[required] The username of the user that you want to remove custom permissions
from.}

\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the custom
permission configuration that you want to delete.}

\item{Namespace}{[required] The namespace that the user belongs to.}
}
\description{
Deletes a custom permissions profile from a user.

See \url{https://www.paws-r-sdk.com/docs/quicksight_delete_user_custom_permission/} for full documentation.
}
\keyword{internal}
