% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_delete_vpc_connection}
\alias{quicksight_delete_vpc_connection}
\title{Deletes a VPC connection}
\usage{
quicksight_delete_vpc_connection(AwsAccountId, VPCConnectionId)
}
\arguments{
\item{AwsAccountId}{[required] The Amazon Web Services account ID of the account where you want to
delete a VPC connection.}

\item{VPCConnectionId}{[required] The ID of the VPC connection that you're creating. This ID is a unique
identifier for each Amazon Web Services Region in an Amazon Web Services
account.}
}
\description{
Deletes a VPC connection.

See \url{https://www.paws-r-sdk.com/docs/quicksight_delete_vpc_connection/} for full documentation.
}
\keyword{internal}
