% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_account_custom_permission}
\alias{quicksight_describe_account_custom_permission}
\title{Describes the custom permissions profile that is applied to an account}
\usage{
quicksight_describe_account_custom_permission(AwsAccountId)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account for which you want to describe
the applied custom permissions profile.}
}
\description{
Describes the custom permissions profile that is applied to an account.

See \url{https://www.paws-r-sdk.com/docs/quicksight_describe_account_custom_permission/} for full documentation.
}
\keyword{internal}
