% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_account_settings}
\alias{quicksight_describe_account_settings}
\title{Describes the settings that were used when your QuickSight subscription
was first created in this AWS account}
\usage{
quicksight_describe_account_settings(AwsAccountId)
}
\arguments{
\item{AwsAccountId}{[required] The ID for the AWS account that contains the settings that you want to
list.}
}
\value{
A list with the following syntax:\preformatted{list(
  AccountSettings = list(
    AccountName = "string",
    Edition = "STANDARD"|"ENTERPRISE",
    DefaultNamespace = "string",
    NotificationEmail = "string"
  ),
  RequestId = "string",
  Status = 123
)
}
}
\description{
Describes the settings that were used when your QuickSight subscription
was first created in this AWS account.
}
\section{Request syntax}{
\preformatted{svc$describe_account_settings(
  AwsAccountId = "string"
)
}
}

\keyword{internal}
