% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_account_subscription}
\alias{quicksight_describe_account_subscription}
\title{Use the DescribeAccountSubscription operation to receive a description
of a Amazon QuickSight account's subscription}
\usage{
quicksight_describe_account_subscription(AwsAccountId)
}
\arguments{
\item{AwsAccountId}{[required] The Amazon Web Services account ID associated with your Amazon
QuickSight account.}
}
\description{
Use the DescribeAccountSubscription operation to receive a description of a Amazon QuickSight account's subscription. A successful API call returns an \code{AccountInfo} object that includes an account's name, subscription status, authentication type, edition, and notification email address.

See \url{https://paws-r.github.io/docs/quicksight/describe_account_subscription.html} for full documentation.
}
\keyword{internal}
