% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_action_connector}
\alias{quicksight_describe_action_connector}
\title{Retrieves detailed information about an action connector, including its
configuration, authentication settings, enabled actions, and current
status}
\usage{
quicksight_describe_action_connector(AwsAccountId, ActionConnectorId)
}
\arguments{
\item{AwsAccountId}{[required] The Amazon Web Services account ID that contains the action connector.}

\item{ActionConnectorId}{[required] The unique identifier of the action connector to describe.}
}
\description{
Retrieves detailed information about an action connector, including its configuration, authentication settings, enabled actions, and current status.

See \url{https://www.paws-r-sdk.com/docs/quicksight_describe_action_connector/} for full documentation.
}
\keyword{internal}
