% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_action_connector_permissions}
\alias{quicksight_describe_action_connector_permissions}
\title{Retrieves the permissions configuration for an action connector, showing
which users, groups, and namespaces have access and what operations they
can perform}
\usage{
quicksight_describe_action_connector_permissions(
  AwsAccountId,
  ActionConnectorId
)
}
\arguments{
\item{AwsAccountId}{[required] The Amazon Web Services account ID that contains the action connector.}

\item{ActionConnectorId}{[required] The unique identifier of the action connector whose permissions you want
to describe.}
}
\description{
Retrieves the permissions configuration for an action connector, showing which users, groups, and namespaces have access and what operations they can perform.

See \url{https://www.paws-r-sdk.com/docs/quicksight_describe_action_connector_permissions/} for full documentation.
}
\keyword{internal}
