% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_agent}
\alias{quicksight_describe_agent}
\title{Describes an agent}
\usage{
quicksight_describe_agent(AgentId, AwsAccountId)
}
\arguments{
\item{AgentId}{[required] The unique identifier for the agent.}

\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the agent.}
}
\description{
Describes an agent.

See \url{https://www.paws-r-sdk.com/docs/quicksight_describe_agent/} for full documentation.
}
\keyword{internal}
