% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_agent_permissions}
\alias{quicksight_describe_agent_permissions}
\title{Describes the resource permissions for an agent}
\usage{
quicksight_describe_agent_permissions(AgentId, AwsAccountId)
}
\arguments{
\item{AgentId}{[required] The unique identifier for the agent.}

\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the agent.}
}
\description{
Describes the resource permissions for an agent.

See \url{https://www.paws-r-sdk.com/docs/quicksight_describe_agent_permissions/} for full documentation.
}
\keyword{internal}
