% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_analysis_permissions}
\alias{quicksight_describe_analysis_permissions}
\title{Provides the read and write permissions for an analysis}
\usage{
quicksight_describe_analysis_permissions(AwsAccountId, AnalysisId)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the analysis
whose permissions you're describing. You must be using the Amazon Web
Services account that the analysis is in.}

\item{AnalysisId}{[required] The ID of the analysis whose permissions you're describing. The ID is
part of the analysis URL.}
}
\description{
Provides the read and write permissions for an analysis.

See \url{https://www.paws-r-sdk.com/docs/quicksight_describe_analysis_permissions/} for full documentation.
}
\keyword{internal}
