% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_app}
\alias{quicksight_describe_app}
\title{Describes an app}
\usage{
quicksight_describe_app(AwsAccountId, AppId)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the app.}

\item{AppId}{[required] The ID of the app that you want to describe.}
}
\description{
Describes an app.

See \url{https://www.paws-r-sdk.com/docs/quicksight_describe_app/} for full documentation.
}
\keyword{internal}
