% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_app_permissions}
\alias{quicksight_describe_app_permissions}
\title{Describes the resource permissions for an app}
\usage{
quicksight_describe_app_permissions(AwsAccountId, AppId)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the app.}

\item{AppId}{[required] The ID of the app.}
}
\description{
Describes the resource permissions for an app.

See \url{https://www.paws-r-sdk.com/docs/quicksight_describe_app_permissions/} for full documentation.
}
\keyword{internal}
