% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_approval_policy}
\alias{quicksight_describe_approval_policy}
\title{Describes an approval policy in Quick Sight}
\usage{
quicksight_describe_approval_policy(PolicyId)
}
\arguments{
\item{PolicyId}{[required] The unique identifier of the approval policy to describe.}
}
\description{
Describes an approval policy in Quick Sight.

See \url{https://www.paws-r-sdk.com/docs/quicksight_describe_approval_policy/} for full documentation.
}
\keyword{internal}
