% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_asset_bundle_export_job}
\alias{quicksight_describe_asset_bundle_export_job}
\title{Describes an existing export job}
\usage{
quicksight_describe_asset_bundle_export_job(
  AwsAccountId,
  AssetBundleExportJobId
)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account the export job is executed in.}

\item{AssetBundleExportJobId}{[required] The ID of the job that you want described. The job ID is set when you
start a new job with a
\code{\link[=quicksight_start_asset_bundle_export_job]{start_asset_bundle_export_job}}
API call.}
}
\description{
Describes an existing export job.

See \url{https://paws-r.github.io/docs/quicksight/describe_asset_bundle_export_job.html} for full documentation.
}
\keyword{internal}
