% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_asset_bundle_import_job}
\alias{quicksight_describe_asset_bundle_import_job}
\title{Describes an existing import job}
\usage{
quicksight_describe_asset_bundle_import_job(
  AwsAccountId,
  AssetBundleImportJobId
)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account the import job was executed
in.}

\item{AssetBundleImportJobId}{[required] The ID of the job. The job ID is set when you start a new job with a
\code{\link[=quicksight_start_asset_bundle_import_job]{start_asset_bundle_import_job}}
API call.}
}
\description{
Describes an existing import job.

See \url{https://www.paws-r-sdk.com/docs/quicksight_describe_asset_bundle_import_job/} for full documentation.
}
\keyword{internal}
