% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_brand}
\alias{quicksight_describe_brand}
\title{Describes a brand}
\usage{
quicksight_describe_brand(AwsAccountId, BrandId, VersionId = NULL)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that owns the brand.}

\item{BrandId}{[required] The ID of the Quick Suite brand.}

\item{VersionId}{The ID of the specific version. The default value is the latest version.}
}
\description{
Describes a brand.

See \url{https://www.paws-r-sdk.com/docs/quicksight_describe_brand/} for full documentation.
}
\keyword{internal}
