% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_brand_assignment}
\alias{quicksight_describe_brand_assignment}
\title{Describes a brand assignment}
\usage{
quicksight_describe_brand_assignment(AwsAccountId)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that owns the brand
assignment.}
}
\description{
Describes a brand assignment.

See \url{https://www.paws-r-sdk.com/docs/quicksight_describe_brand_assignment/} for full documentation.
}
\keyword{internal}
