% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_brand_published_version}
\alias{quicksight_describe_brand_published_version}
\title{Describes the published version of the brand}
\usage{
quicksight_describe_brand_published_version(AwsAccountId, BrandId)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that owns the brand.}

\item{BrandId}{[required] The ID of the Quick Suite brand.}
}
\description{
Describes the published version of the brand.

See \url{https://www.paws-r-sdk.com/docs/quicksight_describe_brand_published_version/} for full documentation.
}
\keyword{internal}
