% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_custom_permissions}
\alias{quicksight_describe_custom_permissions}
\title{Describes a custom permissions profile}
\usage{
quicksight_describe_custom_permissions(AwsAccountId, CustomPermissionsName)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the custom
permissions profile that you want described.}

\item{CustomPermissionsName}{[required] The name of the custom permissions profile to describe.}
}
\description{
Describes a custom permissions profile.

See \url{https://www.paws-r-sdk.com/docs/quicksight_describe_custom_permissions/} for full documentation.
}
\keyword{internal}
