% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_dashboard}
\alias{quicksight_describe_dashboard}
\title{Provides a summary for a dashboard}
\usage{
quicksight_describe_dashboard(AwsAccountId, DashboardId, VersionNumber,
  AliasName)
}
\arguments{
\item{AwsAccountId}{[required] AWS account ID that contains the dashboard you are describing.}

\item{DashboardId}{[required] The ID for the dashboard.}

\item{VersionNumber}{The version number for the dashboard. If version number isn't passed the
latest published dashboard version is described.}

\item{AliasName}{The alias name.}
}
\description{
Provides a summary for a dashboard.
}
\section{Request syntax}{
\preformatted{svc$describe_dashboard(
  AwsAccountId = "string",
  DashboardId = "string",
  VersionNumber = 123,
  AliasName = "string"
)
}
}

\keyword{internal}
