% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_dashboard_definition}
\alias{quicksight_describe_dashboard_definition}
\title{Provides a detailed description of the definition of a dashboard}
\usage{
quicksight_describe_dashboard_definition(
  AwsAccountId,
  DashboardId,
  VersionNumber = NULL,
  AliasName = NULL
)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the dashboard
that you're describing.}

\item{DashboardId}{[required] The ID for the dashboard.}

\item{VersionNumber}{The version number for the dashboard. If a version number isn't passed,
the latest published dashboard version is described.}

\item{AliasName}{The alias name.}
}
\description{
Provides a detailed description of the definition of a dashboard.

See \url{https://www.paws-r-sdk.com/docs/quicksight_describe_dashboard_definition/} for full documentation.
}
\keyword{internal}
