% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_dashboard_permissions}
\alias{quicksight_describe_dashboard_permissions}
\title{Describes read and write permissions on a dashboard}
\usage{
quicksight_describe_dashboard_permissions(AwsAccountId, DashboardId)
}
\arguments{
\item{AwsAccountId}{[required] AWS account ID that contains the dashboard you are describing
permissions of.}

\item{DashboardId}{[required] The ID for the dashboard, also added to IAM policy.}
}
\description{
Describes read and write permissions on a dashboard.
}
\section{Request syntax}{
\preformatted{svc$describe_dashboard_permissions(
  AwsAccountId = "string",
  DashboardId = "string"
)
}
}

\keyword{internal}
