% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_dashboards_qa_configuration}
\alias{quicksight_describe_dashboards_qa_configuration}
\title{Describes an existing dashboard QA configuration}
\usage{
quicksight_describe_dashboards_qa_configuration(AwsAccountId)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the dashboard QA
configuration that you want described.}
}
\description{
Describes an existing dashboard QA configuration.

See \url{https://www.paws-r-sdk.com/docs/quicksight_describe_dashboards_qa_configuration/} for full documentation.
}
\keyword{internal}
