% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_data_set_permissions}
\alias{quicksight_describe_data_set_permissions}
\title{Describes the permissions on a dataset}
\usage{
quicksight_describe_data_set_permissions(AwsAccountId, DataSetId)
}
\arguments{
\item{AwsAccountId}{[required] The AWS account ID.}

\item{DataSetId}{[required] The ID for the dataset that you want to create. This ID is unique per
AWS Region for each AWS account.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  DataSetArn = "string",
  DataSetId = "string",
  Permissions = list(
    list(
      Principal = "string",
      Actions = list(
        "string"
      )
    )
  ),
  RequestId = "string",
  Status = 123
)
}\if{html}{\out{</div>}}
}
\description{
Describes the permissions on a dataset.

The permissions resource is
\code{arn:aws:quicksight:region:aws-account-id:dataset/data-set-id}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_data_set_permissions(
  AwsAccountId = "string",
  DataSetId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
