% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_data_set_refresh_properties}
\alias{quicksight_describe_data_set_refresh_properties}
\title{Describes the refresh properties of a dataset}
\usage{
quicksight_describe_data_set_refresh_properties(AwsAccountId, DataSetId)
}
\arguments{
\item{AwsAccountId}{[required] The Amazon Web Services account ID.}

\item{DataSetId}{[required] The ID of the dataset.}
}
\description{
Describes the refresh properties of a dataset.

See \url{https://paws-r.github.io/docs/quicksight/describe_data_set_refresh_properties.html} for full documentation.
}
\keyword{internal}
