% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_data_source_permissions}
\alias{quicksight_describe_data_source_permissions}
\title{Describes the resource permissions for a data source}
\usage{
quicksight_describe_data_source_permissions(AwsAccountId, DataSourceId)
}
\arguments{
\item{AwsAccountId}{[required] The Amazon Web Services account ID.}

\item{DataSourceId}{[required] The ID of the data source. This ID is unique per Amazon Web Services
Region for each Amazon Web Services account.}
}
\description{
Describes the resource permissions for a data source.

See \url{https://www.paws-r-sdk.com/docs/quicksight_describe_data_source_permissions/} for full documentation.
}
\keyword{internal}
