% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_default_q_business_application}
\alias{quicksight_describe_default_q_business_application}
\title{Describes a Amazon Q Business application that is linked to an Quick
Sight account}
\usage{
quicksight_describe_default_q_business_application(
  AwsAccountId,
  Namespace = NULL
)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Quick Sight account that is linked to the Amazon Q
Business application that you want described.}

\item{Namespace}{The Quick Sight namespace that contains the linked Amazon Q Business
application. If this field is left blank, the default namespace is used.
Currently, the default namespace is the only valid value for this
parameter.}
}
\description{
Describes a Amazon Q Business application that is linked to an Quick Sight account.

See \url{https://www.paws-r-sdk.com/docs/quicksight_describe_default_q_business_application/} for full documentation.
}
\keyword{internal}
