% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_dlp_setting}
\alias{quicksight_describe_dlp_setting}
\title{Describes the full configuration of a DLP setting in an Amazon Web
Services account}
\usage{
quicksight_describe_dlp_setting(AwsAccountId, DlpSettingId)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the DLP setting that you want to describe.}

\item{DlpSettingId}{[required] The ID of the DLP setting that you want to describe.}
}
\description{
Describes the full configuration of a DLP setting in an Amazon Web Services account.

See \url{https://www.paws-r-sdk.com/docs/quicksight_describe_dlp_setting/} for full documentation.
}
\keyword{internal}
