% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_flow}
\alias{quicksight_describe_flow}
\title{Returns the full details of a flow for the latest version of the
requested publish state}
\usage{
quicksight_describe_flow(AwsAccountId, FlowId, PublishState)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the flow that you are describing.}

\item{FlowId}{[required] The unique identifier of the flow.}

\item{PublishState}{[required] The publish state of the flow version to describe. Valid values are \code{DRAFT}, \code{PUBLISHED}, or \code{PENDING_APPROVAL}.}
}
\description{
Returns the full details of a flow for the latest version of the requested publish state.

See \url{https://www.paws-r-sdk.com/docs/quicksight_describe_flow/} for full documentation.
}
\keyword{internal}
