% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_folder_resolved_permissions}
\alias{quicksight_describe_folder_resolved_permissions}
\title{Describes the folder resolved permissions}
\usage{
quicksight_describe_folder_resolved_permissions(AwsAccountId, FolderId)
}
\arguments{
\item{AwsAccountId}{[required] The ID for the Amazon Web Services account that contains the folder.}

\item{FolderId}{[required] The ID of the folder.}
}
\description{
Describes the folder resolved permissions. Permissions consists of both folder direct permissions and the inherited permissions from the ancestor folders.

See \url{https://paws-r.github.io/docs/quicksight/describe_folder_resolved_permissions.html} for full documentation.
}
\keyword{internal}
