% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_ingestion}
\alias{quicksight_describe_ingestion}
\title{Describes a SPICE ingestion}
\usage{
quicksight_describe_ingestion(AwsAccountId, DataSetId, IngestionId)
}
\arguments{
\item{AwsAccountId}{[required] The Amazon Web Services account ID.}

\item{DataSetId}{[required] The ID of the dataset used in the ingestion.}

\item{IngestionId}{[required] An ID for the ingestion.}
}
\description{
Describes a SPICE ingestion.

See \url{https://paws-r.github.io/docs/quicksight/describe_ingestion.html} for full documentation.
}
\keyword{internal}
