% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_ip_restriction}
\alias{quicksight_describe_ip_restriction}
\title{Provides a summary and status of IP rules}
\usage{
quicksight_describe_ip_restriction(AwsAccountId)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the IP rules.}
}
\description{
Provides a summary and status of IP rules.

See \url{https://www.paws-r-sdk.com/docs/quicksight_describe_ip_restriction/} for full documentation.
}
\keyword{internal}
