% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_key_registration}
\alias{quicksight_describe_key_registration}
\title{Describes all customer managed key registrations in a Quick Sight
account}
\usage{
quicksight_describe_key_registration(AwsAccountId, DefaultKeyOnly = NULL)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the customer
managed key registration that you want to describe.}

\item{DefaultKeyOnly}{Determines whether the request returns the default key only.}
}
\description{
Describes all customer managed key registrations in a Quick Sight account.

See \url{https://www.paws-r-sdk.com/docs/quicksight_describe_key_registration/} for full documentation.
}
\keyword{internal}
