% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_knowledge_base_permissions}
\alias{quicksight_describe_knowledge_base_permissions}
\title{Describes the resource permissions for a knowledge base}
\usage{
quicksight_describe_knowledge_base_permissions(AwsAccountId, KnowledgeBaseId)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the knowledge base.}

\item{KnowledgeBaseId}{[required] The unique identifier for the knowledge base.}
}
\description{
Describes the resource permissions for a knowledge base.

See \url{https://www.paws-r-sdk.com/docs/quicksight_describe_knowledge_base_permissions/} for full documentation.
}
\keyword{internal}
