% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_limits_profile}
\alias{quicksight_describe_limits_profile}
\title{Describes the properties of an existing limits profile}
\usage{
quicksight_describe_limits_profile(profileId, accountId)
}
\arguments{
\item{profileId}{[required] The unique identifier for the limits profile.}

\item{accountId}{[required] The ID of the Amazon Web Services account that contains the limits profile.}
}
\description{
Describes the properties of an existing limits profile.

See \url{https://www.paws-r-sdk.com/docs/quicksight_describe_limits_profile/} for full documentation.
}
\keyword{internal}
