% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_namespace}
\alias{quicksight_describe_namespace}
\title{Describes the current namespace}
\usage{
quicksight_describe_namespace(AwsAccountId, Namespace)
}
\arguments{
\item{AwsAccountId}{[required] The ID for the AWS account that contains the QuickSight namespace that
you want to describe.}

\item{Namespace}{[required] The namespace that you want to describe.}
}
\description{
Describes the current namespace.
}
\section{Request syntax}{
\preformatted{svc$describe_namespace(
  AwsAccountId = "string",
  Namespace = "string"
)
}
}

\keyword{internal}
