% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_o_auth_client_application}
\alias{quicksight_describe_o_auth_client_application}
\title{Describes an OAuthClientApplication}
\usage{
quicksight_describe_o_auth_client_application(
  AwsAccountId,
  OAuthClientApplicationId
)
}
\arguments{
\item{AwsAccountId}{[required] The Amazon Web Services account ID.}

\item{OAuthClientApplicationId}{[required] The ID of the OAuthClientApplication that you want to describe.}
}
\description{
Describes an OAuthClientApplication.

See \url{https://www.paws-r-sdk.com/docs/quicksight_describe_o_auth_client_application/} for full documentation.
}
\keyword{internal}
