% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_q_personalization_configuration}
\alias{quicksight_describe_q_personalization_configuration}
\title{Describes a personalization configuration}
\usage{
quicksight_describe_q_personalization_configuration(AwsAccountId)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the
personalization configuration that the user wants described.}
}
\description{
Describes a personalization configuration.

See \url{https://www.paws-r-sdk.com/docs/quicksight_describe_q_personalization_configuration/} for full documentation.
}
\keyword{internal}
