% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_refresh_schedule}
\alias{quicksight_describe_refresh_schedule}
\title{Provides a summary of a refresh schedule}
\usage{
quicksight_describe_refresh_schedule(AwsAccountId, DataSetId, ScheduleId)
}
\arguments{
\item{AwsAccountId}{[required] The Amazon Web Services account ID.}

\item{DataSetId}{[required] The ID of the dataset.}

\item{ScheduleId}{[required] The ID of the refresh schedule.}
}
\description{
Provides a summary of a refresh schedule.

See \url{https://www.paws-r-sdk.com/docs/quicksight_describe_refresh_schedule/} for full documentation.
}
\keyword{internal}
