% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_role_custom_permission}
\alias{quicksight_describe_role_custom_permission}
\title{Describes all custom permissions that are mapped to a role}
\usage{
quicksight_describe_role_custom_permission(Role, AwsAccountId, Namespace)
}
\arguments{
\item{Role}{[required] The name of the role whose permissions you want described.}

\item{AwsAccountId}{[required] The ID for the Amazon Web Services account that you want to create a
group in. The Amazon Web Services account ID that you provide must be
the same Amazon Web Services account that contains your Amazon Quick
Sight account.}

\item{Namespace}{[required] The namespace that contains the role.}
}
\description{
Describes all custom permissions that are mapped to a role.

See \url{https://www.paws-r-sdk.com/docs/quicksight_describe_role_custom_permission/} for full documentation.
}
\keyword{internal}
