% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_self_upgrade_configuration}
\alias{quicksight_describe_self_upgrade_configuration}
\title{Describes the self-upgrade configuration for a Quick Suite account}
\usage{
quicksight_describe_self_upgrade_configuration(AwsAccountId, Namespace)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the Quick Suite
self-upgrade configuration.}

\item{Namespace}{[required] The Quick Suite namespace that you want to describe the Quick Suite
self-upgrade configuration for.}
}
\description{
Describes the self-upgrade configuration for a Quick Suite account.

See \url{https://www.paws-r-sdk.com/docs/quicksight_describe_self_upgrade_configuration/} for full documentation.
}
\keyword{internal}
