% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_space}
\alias{quicksight_describe_space}
\title{Describes an Amazon QuickSight space}
\usage{
quicksight_describe_space(AwsAccountId, SpaceId, MaxContributors = NULL)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the space.}

\item{SpaceId}{[required] The ID of the space that you want to describe.}

\item{MaxContributors}{The maximum number of contributors to include in the response.}
}
\description{
Describes an Amazon QuickSight space.

See \url{https://www.paws-r-sdk.com/docs/quicksight_describe_space/} for full documentation.
}
\keyword{internal}
