% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_space_permissions}
\alias{quicksight_describe_space_permissions}
\title{Describes the permissions for an Amazon QuickSight space}
\usage{
quicksight_describe_space_permissions(AwsAccountId, SpaceId)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the space.}

\item{SpaceId}{[required] The ID of the space that you want to describe permissions for.}
}
\description{
Describes the permissions for an Amazon QuickSight space.

See \url{https://www.paws-r-sdk.com/docs/quicksight_describe_space_permissions/} for full documentation.
}
\keyword{internal}
