% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_theme}
\alias{quicksight_describe_theme}
\title{Describes a theme}
\usage{
quicksight_describe_theme(
  AwsAccountId,
  ThemeId,
  VersionNumber = NULL,
  AliasName = NULL
)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the theme that
you're describing.}

\item{ThemeId}{[required] The ID for the theme.}

\item{VersionNumber}{The version number for the version to describe. If a \code{VersionNumber}
parameter value isn't provided, the latest version of the theme is
described.}

\item{AliasName}{The alias of the theme that you want to describe. If you name a specific
alias, you describe the version that the alias points to. You can
specify the latest version of the theme by providing the keyword
\verb{$LATEST} in the \code{AliasName} parameter. The keyword \verb{$PUBLISHED} doesn't
apply to themes.}
}
\description{
Describes a theme.

See \url{https://paws-r.github.io/docs/quicksight/describe_theme.html} for full documentation.
}
\keyword{internal}
