% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_theme_alias}
\alias{quicksight_describe_theme_alias}
\title{Describes the alias for a theme}
\usage{
quicksight_describe_theme_alias(AwsAccountId, ThemeId, AliasName)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the AWS account that contains the theme alias that you're
describing.}

\item{ThemeId}{[required] The ID for the theme.}

\item{AliasName}{[required] The name of the theme alias that you want to describe.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ThemeAlias = list(
    Arn = "string",
    AliasName = "string",
    ThemeVersionNumber = 123
  ),
  Status = 123,
  RequestId = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Describes the alias for a theme.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_theme_alias(
  AwsAccountId = "string",
  ThemeId = "string",
  AliasName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
