% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_theme_permissions}
\alias{quicksight_describe_theme_permissions}
\title{Describes the read and write permissions for a theme}
\usage{
quicksight_describe_theme_permissions(AwsAccountId, ThemeId)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the AWS account that contains the theme that you're
describing.}

\item{ThemeId}{[required] The ID for the theme that you want to describe permissions for.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ThemeId = "string",
  ThemeArn = "string",
  Permissions = list(
    list(
      Principal = "string",
      Actions = list(
        "string"
      )
    )
  ),
  RequestId = "string",
  Status = 123
)
}\if{html}{\out{</div>}}
}
\description{
Describes the read and write permissions for a theme.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_theme_permissions(
  AwsAccountId = "string",
  ThemeId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
