% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_topic}
\alias{quicksight_describe_topic}
\title{Describes a topic}
\usage{
quicksight_describe_topic(AwsAccountId, TopicId)
}
\arguments{
\item{AwsAccountId}{[required] The Amazon Web Services account ID.}

\item{TopicId}{[required] The ID of the topic that you want to describe. This ID is unique per
Amazon Web Services Region for each Amazon Web Services account.}
}
\description{
Describes a topic.

See \url{https://paws-r.github.io/docs/quicksight/describe_topic.html} for full documentation.
}
\keyword{internal}
