% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_topic_permissions}
\alias{quicksight_describe_topic_permissions}
\title{Describes the permissions of a topic}
\usage{
quicksight_describe_topic_permissions(AwsAccountId, TopicId)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the topic that
you want described.}

\item{TopicId}{[required] The ID of the topic that you want to describe. This ID is unique per
Amazon Web Services Region for each Amazon Web Services account.}
}
\description{
Describes the permissions of a topic.

See \url{https://paws-r.github.io/docs/quicksight/describe_topic_permissions.html} for full documentation.
}
\keyword{internal}
