% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_topic_permissions_v2}
\alias{quicksight_describe_topic_permissions_v2}
\title{Describes the permissions of a topic}
\usage{
quicksight_describe_topic_permissions_v2(AwsAccountId, TopicId)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the topic that you want described.}

\item{TopicId}{[required] The ID of the topic that you want to describe. This ID is unique per Amazon Web Services Region for each Amazon Web Services account.}
}
\description{
Describes the permissions of a topic.

See \url{https://www.paws-r-sdk.com/docs/quicksight_describe_topic_permissions_v2/} for full documentation.
}
\keyword{internal}
