% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_topic_refresh_schedule}
\alias{quicksight_describe_topic_refresh_schedule}
\title{Deletes a topic refresh schedule}
\usage{
quicksight_describe_topic_refresh_schedule(AwsAccountId, TopicId, DatasetId)
}
\arguments{
\item{AwsAccountId}{[required] The Amazon Web Services account ID.}

\item{TopicId}{[required] The ID of the topic that contains the refresh schedule that you want to
describe. This ID is unique per Amazon Web Services Region for each
Amazon Web Services account.}

\item{DatasetId}{[required] The ID of the dataset.}
}
\description{
Deletes a topic refresh schedule.

See \url{https://www.paws-r-sdk.com/docs/quicksight_describe_topic_refresh_schedule/} for full documentation.
}
\keyword{internal}
