% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_vpc_connection}
\alias{quicksight_describe_vpc_connection}
\title{Describes a VPC connection}
\usage{
quicksight_describe_vpc_connection(AwsAccountId, VPCConnectionId)
}
\arguments{
\item{AwsAccountId}{[required] The Amazon Web Services account ID of the account that contains the VPC
connection that you want described.}

\item{VPCConnectionId}{[required] The ID of the VPC connection that you're creating. This ID is a unique
identifier for each Amazon Web Services Region in an Amazon Web Services
account.}
}
\description{
Describes a VPC connection.

See \url{https://paws-r.github.io/docs/quicksight/describe_vpc_connection.html} for full documentation.
}
\keyword{internal}
