% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_get_flow_metadata}
\alias{quicksight_get_flow_metadata}
\title{Retrieves the metadata of a flow, not including its definition
specifying the steps}
\usage{
quicksight_get_flow_metadata(AwsAccountId, FlowId)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the flow that
you are getting metadata for.}

\item{FlowId}{[required] The unique identifier of the flow.}
}
\description{
Retrieves the metadata of a flow, not including its definition specifying the steps.

See \url{https://www.paws-r-sdk.com/docs/quicksight_get_flow_metadata/} for full documentation.
}
\keyword{internal}
