% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_get_flow_permissions}
\alias{quicksight_get_flow_permissions}
\title{Get permissions for a flow}
\usage{
quicksight_get_flow_permissions(AwsAccountId, FlowId)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the flow that
you are getting permissions for.}

\item{FlowId}{[required] The unique identifier of the flow to get permissions from.}
}
\description{
Get permissions for a flow.

See \url{https://www.paws-r-sdk.com/docs/quicksight_get_flow_permissions/} for full documentation.
}
\keyword{internal}
