% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_list_action_connectors}
\alias{quicksight_list_action_connectors}
\title{Lists all action connectors in the specified Amazon Web Services account}
\usage{
quicksight_list_action_connectors(
  AwsAccountId,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{AwsAccountId}{[required] The Amazon Web Services account ID for which to list action connectors.}

\item{MaxResults}{The maximum number of action connectors to return in a single response.
Valid range is 1 to 100.}

\item{NextToken}{A pagination token to retrieve the next set of results. Use the token
returned from a previous call to continue listing action connectors.}
}
\description{
Lists all action connectors in the specified Amazon Web Services account. Returns summary information for each connector including its name, type, creation time, and status.

See \url{https://www.paws-r-sdk.com/docs/quicksight_list_action_connectors/} for full documentation.
}
\keyword{internal}
