% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_list_approval_policies}
\alias{quicksight_list_approval_policies}
\title{Lists all approval policies in the specified Quick Sight account}
\usage{
quicksight_list_approval_policies(NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{NextToken}{The token for the next set of results, or null if there are no more results.}

\item{MaxResults}{The maximum number of results to return in a single call. If you don't specify a value, the service returns a default number of results. Use the \code{NextToken} value in the response to retrieve additional results.}
}
\description{
Lists all approval policies in the specified Quick Sight account. The results are paginated. If the response includes a \code{NextToken} value, pass it in a subsequent call to retrieve the next set of results.

See \url{https://www.paws-r-sdk.com/docs/quicksight_list_approval_policies/} for full documentation.
}
\keyword{internal}
