% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_list_apps}
\alias{quicksight_list_apps}
\title{Lists the apps in an Amazon Web Services account}
\usage{
quicksight_list_apps(AwsAccountId, MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the apps.}

\item{MaxResults}{The maximum number of results to return in a single request. Valid range is 1 to 100. If you don't specify a value, the default is 20.}

\item{NextToken}{The token for the next set of results, or null if there are no more results.}
}
\description{
Lists the apps in an Amazon Web Services account. Results are paginated; use the \code{NextToken} parameter to retrieve additional results.

See \url{https://www.paws-r-sdk.com/docs/quicksight_list_apps/} for full documentation.
}
\keyword{internal}
