% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_list_custom_permissions}
\alias{quicksight_list_custom_permissions}
\title{Returns a list of all the custom permissions profiles}
\usage{
quicksight_list_custom_permissions(
  AwsAccountId,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the custom
permissions profiles that you want to list.}

\item{MaxResults}{The maximum number of results to return.}

\item{NextToken}{The token for the next set of results, or null if there are no more
results.}
}
\description{
Returns a list of all the custom permissions profiles.

See \url{https://www.paws-r-sdk.com/docs/quicksight_list_custom_permissions/} for full documentation.
}
\keyword{internal}
