% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_list_dashboard_versions}
\alias{quicksight_list_dashboard_versions}
\title{Lists all the versions of the dashboards in the QuickSight subscription}
\usage{
quicksight_list_dashboard_versions(AwsAccountId, DashboardId, NextToken,
  MaxResults)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the AWS account that contains the dashboard that you're
listing versions for.}

\item{DashboardId}{[required] The ID for the dashboard.}

\item{NextToken}{The token for the next set of results, or null if there are no more
results.}

\item{MaxResults}{The maximum number of results to be returned per request.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  DashboardVersionSummaryList = list(
    list(
      Arn = "string",
      CreatedTime = as.POSIXct(
        "2015-01-01"
      ),
      VersionNumber = 123,
      Status = "CREATION_IN_PROGRESS"|"CREATION_SUCCESSFUL"|"CREATION_FAILED"|"UPDATE_IN_PROGRESS"|"UPDATE_SUCCESSFUL"|"UPDATE_FAILED"|"DELETED",
      SourceEntityArn = "string",
      Description = "string"
    )
  ),
  NextToken = "string",
  Status = 123,
  RequestId = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Lists all the versions of the dashboards in the QuickSight subscription.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_dashboard_versions(
  AwsAccountId = "string",
  DashboardId = "string",
  NextToken = "string",
  MaxResults = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
