% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_list_data_sources}
\alias{quicksight_list_data_sources}
\title{Lists data sources in current Amazon Web Services Region that belong to
this Amazon Web Services account}
\usage{
quicksight_list_data_sources(AwsAccountId, NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{AwsAccountId}{[required] The Amazon Web Services account ID.}

\item{NextToken}{The token for the next set of results, or null if there are no more
results.}

\item{MaxResults}{The maximum number of results to be returned per request.}
}
\description{
Lists data sources in current Amazon Web Services Region that belong to this Amazon Web Services account.

See \url{https://www.paws-r-sdk.com/docs/quicksight_list_data_sources/} for full documentation.
}
\keyword{internal}
