% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_list_dlp_settings}
\alias{quicksight_list_dlp_settings}
\title{Lists all DLP settings in an Amazon Web Services account}
\usage{
quicksight_list_dlp_settings(AwsAccountId, NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the DLP settings that you want to list.}

\item{NextToken}{The token for the next set of results, or null if there are no more results.}

\item{MaxResults}{The maximum number of results to return per request.}
}
\description{
Lists all DLP settings in an Amazon Web Services account.

See \url{https://www.paws-r-sdk.com/docs/quicksight_list_dlp_settings/} for full documentation.
}
\keyword{internal}
