% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_list_flows}
\alias{quicksight_list_flows}
\title{Lists flows in an Amazon Web Services account}
\usage{
quicksight_list_flows(AwsAccountId, NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the flow list
that you are getting.}

\item{NextToken}{The token to request the next set of results, or null if you want to
retrieve the first set.}

\item{MaxResults}{The maximum number of results to be returned per request.}
}
\description{
Lists flows in an Amazon Web Services account.

See \url{https://www.paws-r-sdk.com/docs/quicksight_list_flows/} for full documentation.
}
\keyword{internal}
