% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_list_folder_members}
\alias{quicksight_list_folder_members}
\title{List all assets (DASHBOARD, ANALYSIS, and DATASET) in a folder}
\usage{
quicksight_list_folder_members(
  AwsAccountId,
  FolderId,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{AwsAccountId}{[required] The ID for the Amazon Web Services account that contains the folder.}

\item{FolderId}{[required] The ID of the folder.}

\item{NextToken}{The token for the next set of results, or null if there are no more
results.}

\item{MaxResults}{The maximum number of results to be returned per request.}
}
\description{
List all assets (\code{DASHBOARD}, \code{ANALYSIS}, and \code{DATASET}) in a folder.

See \url{https://paws-r.github.io/docs/quicksight/list_folder_members.html} for full documentation.
}
\keyword{internal}
