% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_list_folders_for_resource}
\alias{quicksight_list_folders_for_resource}
\title{List all folders that a resource is a member of}
\usage{
quicksight_list_folders_for_resource(
  AwsAccountId,
  ResourceArn,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{AwsAccountId}{[required] The ID for the Amazon Web Services account that contains the resource.}

\item{ResourceArn}{[required] The Amazon Resource Name (ARN) the resource whose folders you need to
list.}

\item{NextToken}{The token for the next set of results, or null if there are no more
results.}

\item{MaxResults}{The maximum number of results to be returned per request.}
}
\description{
List all folders that a resource is a member of.

See \url{https://www.paws-r-sdk.com/docs/quicksight_list_folders_for_resource/} for full documentation.
}
\keyword{internal}
