% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_list_iam_policy_assignments_for_user}
\alias{quicksight_list_iam_policy_assignments_for_user}
\title{Lists all the IAM policy assignments, including the Amazon Resource
Names (ARNs) for the IAM policies assigned to the specified user and
group or groups that the user belongs to}
\usage{
quicksight_list_iam_policy_assignments_for_user(
  AwsAccountId,
  UserName,
  NextToken = NULL,
  MaxResults = NULL,
  Namespace
)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the assignments.}

\item{UserName}{[required] The name of the user.}

\item{NextToken}{The token for the next set of results, or null if there are no more
results.}

\item{MaxResults}{The maximum number of results to be returned per request.}

\item{Namespace}{[required] The namespace of the assignment.}
}
\description{
Lists all the IAM policy assignments, including the Amazon Resource Names (ARNs) for the IAM policies assigned to the specified user and group or groups that the user belongs to.

See \url{https://paws-r.github.io/docs/quicksight/list_iam_policy_assignments_for_user.html} for full documentation.
}
\keyword{internal}
