% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_list_identity_propagation_configs}
\alias{quicksight_list_identity_propagation_configs}
\title{Lists all services and authorized targets that the Quick Sight IAM
Identity Center application can access}
\usage{
quicksight_list_identity_propagation_configs(
  AwsAccountId,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contain the identity
propagation configurations of.}

\item{MaxResults}{The maximum number of results to be returned.}

\item{NextToken}{The token for the next set of results, or null if there are no more
results.}
}
\description{
Lists all services and authorized targets that the Quick Sight IAM Identity Center application can access.

See \url{https://www.paws-r-sdk.com/docs/quicksight_list_identity_propagation_configs/} for full documentation.
}
\keyword{internal}
