% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_list_knowledge_bases}
\alias{quicksight_list_knowledge_bases}
\title{Lists all knowledge bases in an Amazon QuickSight account}
\usage{
quicksight_list_knowledge_bases(
  AwsAccountId,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the knowledge base.}

\item{MaxResults}{The maximum number of results to return.}

\item{NextToken}{The token for the next set of results, or null if there are no more results.}
}
\description{
Lists all knowledge bases in an Amazon QuickSight account.

See \url{https://www.paws-r-sdk.com/docs/quicksight_list_knowledge_bases/} for full documentation.
}
\keyword{internal}
