% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_list_limits_profiles}
\alias{quicksight_list_limits_profiles}
\title{Lists all limits profiles in an Amazon Quick Sight account}
\usage{
quicksight_list_limits_profiles(
  accountId,
  resourceType = NULL,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{accountId}{[required] The ID of the Amazon Web Services account that contains the limits profiles.}

\item{resourceType}{An optional filter that limits the results to profiles that contain the specified resource type. If you don't specify a value, the operation returns all profiles.}

\item{maxResults}{The maximum number of results to return in a single call. If you don't specify a value, the service uses the default maximum.}

\item{nextToken}{The token for the next set of results, or null if there are no more results.}
}
\description{
Lists all limits profiles in an Amazon Quick Sight account. Results are paginated. Use the \code{maxResults} parameter to limit the number of results returned in a single call, and use the \code{nextToken} parameter to retrieve the next page of results.

See \url{https://www.paws-r-sdk.com/docs/quicksight_list_limits_profiles/} for full documentation.
}
\keyword{internal}
