% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_list_o_auth_client_applications}
\alias{quicksight_list_o_auth_client_applications}
\title{Lists all OAuthClientApplications in the current Amazon Web Services
Region that belong to this Amazon Web Services account}
\usage{
quicksight_list_o_auth_client_applications(
  AwsAccountId,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{AwsAccountId}{[required] The Amazon Web Services account ID.}

\item{NextToken}{A pagination token that can be used in a subsequent request.}

\item{MaxResults}{The maximum number of results to return.}
}
\description{
Lists all OAuthClientApplications in the current Amazon Web Services Region that belong to this Amazon Web Services account.

See \url{https://www.paws-r-sdk.com/docs/quicksight_list_o_auth_client_applications/} for full documentation.
}
\keyword{internal}
