% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_list_refresh_schedules}
\alias{quicksight_list_refresh_schedules}
\title{Lists the refresh schedules of a dataset}
\usage{
quicksight_list_refresh_schedules(AwsAccountId, DataSetId)
}
\arguments{
\item{AwsAccountId}{[required] The Amazon Web Services account ID.}

\item{DataSetId}{[required] The ID of the dataset.}
}
\description{
Lists the refresh schedules of a dataset. Each dataset can have up to 5 schedules.

See \url{https://www.paws-r-sdk.com/docs/quicksight_list_refresh_schedules/} for full documentation.
}
\keyword{internal}
