% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_list_role_memberships}
\alias{quicksight_list_role_memberships}
\title{Lists all groups that are associated with a role}
\usage{
quicksight_list_role_memberships(
  Role,
  NextToken = NULL,
  MaxResults = NULL,
  AwsAccountId,
  Namespace
)
}
\arguments{
\item{Role}{[required] The name of the role.}

\item{NextToken}{A pagination token that can be used in a subsequent request.}

\item{MaxResults}{The maximum number of results to return.}

\item{AwsAccountId}{[required] The ID for the Amazon Web Services account that you want to create a
group in. The Amazon Web Services account ID that you provide must be
the same Amazon Web Services account that contains your Amazon Quick
Sight account.}

\item{Namespace}{[required] The namespace that includes the role.}
}
\description{
Lists all groups that are associated with a role.

See \url{https://www.paws-r-sdk.com/docs/quicksight_list_role_memberships/} for full documentation.
}
\keyword{internal}
