% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_list_self_upgrades}
\alias{quicksight_list_self_upgrades}
\title{Lists all self-upgrade requests for a Quick Suite account}
\usage{
quicksight_list_self_upgrades(
  AwsAccountId,
  Namespace,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the self-upgrade
requests.}

\item{Namespace}{[required] The Quick Suite namespace for the self-upgrade requests.}

\item{NextToken}{The token for the next set of results, or null if there are no more
results.}

\item{MaxResults}{The maximum number of results to return.}
}
\description{
Lists all self-upgrade requests for a Quick Suite account.

See \url{https://www.paws-r-sdk.com/docs/quicksight_list_self_upgrades/} for full documentation.
}
\keyword{internal}
