% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_list_space_resources}
\alias{quicksight_list_space_resources}
\title{Lists the resources in an Amazon QuickSight space}
\usage{
quicksight_list_space_resources(AwsAccountId, SpaceId)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the space.}

\item{SpaceId}{[required] The ID of the space that you want to list resources for.}
}
\description{
Lists the resources in an Amazon QuickSight space.

See \url{https://www.paws-r-sdk.com/docs/quicksight_list_space_resources/} for full documentation.
}
\keyword{internal}
