% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_list_templates}
\alias{quicksight_list_templates}
\title{Lists all the templates in the current Amazon QuickSight account}
\usage{
quicksight_list_templates(AwsAccountId, NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the templates
that you're listing.}

\item{NextToken}{The token for the next set of results, or null if there are no more
results.}

\item{MaxResults}{The maximum number of results to be returned per request.}
}
\description{
Lists all the templates in the current Amazon QuickSight account.

See \url{https://paws-r.github.io/docs/quicksight/list_templates.html} for full documentation.
}
\keyword{internal}
