% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_list_theme_versions}
\alias{quicksight_list_theme_versions}
\title{Lists all the versions of the themes in the current Amazon Web Services
account}
\usage{
quicksight_list_theme_versions(
  AwsAccountId,
  ThemeId,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the themes that
you're listing.}

\item{ThemeId}{[required] The ID for the theme.}

\item{NextToken}{The token for the next set of results, or null if there are no more
results.}

\item{MaxResults}{The maximum number of results to be returned per request.}
}
\description{
Lists all the versions of the themes in the current Amazon Web Services account.

See \url{https://paws-r.github.io/docs/quicksight/list_theme_versions.html} for full documentation.
}
\keyword{internal}
