% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_list_themes}
\alias{quicksight_list_themes}
\title{Lists all the themes in the current AWS account}
\usage{
quicksight_list_themes(AwsAccountId, NextToken, MaxResults, Type)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the AWS account that contains the themes that you're listing.}

\item{NextToken}{The token for the next set of results, or null if there are no more
results.}

\item{MaxResults}{The maximum number of results to be returned per request.}

\item{Type}{The type of themes that you want to list. Valid options include the
following:
\itemize{
\item \code{ALL (default)}- Display all existing themes.
\item \code{CUSTOM} - Display only the themes created by people using Amazon
QuickSight.
\item \code{QUICKSIGHT} - Display only the starting themes defined by
QuickSight.
}}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ThemeSummaryList = list(
    list(
      Arn = "string",
      Name = "string",
      ThemeId = "string",
      LatestVersionNumber = 123,
      CreatedTime = as.POSIXct(
        "2015-01-01"
      ),
      LastUpdatedTime = as.POSIXct(
        "2015-01-01"
      )
    )
  ),
  NextToken = "string",
  Status = 123,
  RequestId = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Lists all the themes in the current AWS account.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_themes(
  AwsAccountId = "string",
  NextToken = "string",
  MaxResults = 123,
  Type = "QUICKSIGHT"|"CUSTOM"|"ALL"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
