% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_list_topic_refresh_schedules}
\alias{quicksight_list_topic_refresh_schedules}
\title{Lists all of the refresh schedules for a topic}
\usage{
quicksight_list_topic_refresh_schedules(AwsAccountId, TopicId)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the topic whose
refresh schedule you want described.}

\item{TopicId}{[required] The ID for the topic that you want to describe. This ID is unique per
Amazon Web Services Region for each Amazon Web Services account.}
}
\description{
Lists all of the refresh schedules for a topic.

See \url{https://paws-r.github.io/docs/quicksight/list_topic_refresh_schedules.html} for full documentation.
}
\keyword{internal}
