% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_list_topic_reviewed_answers}
\alias{quicksight_list_topic_reviewed_answers}
\title{Lists all reviewed answers for a Q Topic}
\usage{
quicksight_list_topic_reviewed_answers(AwsAccountId, TopicId)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that containd the reviewed
answers that you want listed.}

\item{TopicId}{[required] The ID for the topic that contains the reviewed answer that you want to
list. This ID is unique per Amazon Web Services Region for each Amazon
Web Services account.}
}
\description{
Lists all reviewed answers for a Q Topic.

See \url{https://www.paws-r-sdk.com/docs/quicksight_list_topic_reviewed_answers/} for full documentation.
}
\keyword{internal}
