% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_list_topics_v2}
\alias{quicksight_list_topics_v2}
\title{Lists all of the Q topics in the specified Amazon Web Services account
in an Amazon Web Services Region}
\usage{
quicksight_list_topics_v2(AwsAccountId, NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the topics that you want to list.}

\item{NextToken}{The token for the next set of results, or null if there are no more results.}

\item{MaxResults}{The maximum number of results to be returned per request.}
}
\description{
Lists all of the Q topics in the specified Amazon Web Services account in an Amazon Web Services Region.

See \url{https://www.paws-r-sdk.com/docs/quicksight_list_topics_v2/} for full documentation.
}
\keyword{internal}
