% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_list_users_index_capacity}
\alias{quicksight_list_users_index_capacity}
\title{Lists per-user index capacity consumption for an account}
\usage{
quicksight_list_users_index_capacity(
  awsAccountId,
  namespace = NULL,
  filters = NULL,
  sortBy = NULL,
  sortOrder = NULL,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{awsAccountId}{[required] The ID of the Amazon Web Services account that contains the index capacity data.}

\item{namespace}{The namespace to scope the user search to. Required when the userNameOrEmail filter is present.}

\item{filters}{Filters to apply. Only one filter is supported per request. The userNameOrEmail and totalCapacityBytes filters are mutually exclusive.}

\item{sortBy}{The field to sort results by.}

\item{sortOrder}{The sort order for results. Defaults to DESC if not specified.}

\item{maxResults}{The maximum number of results to return per page.}

\item{nextToken}{The token for the next set of results, received from a previous call.}
}
\description{
Lists per-user index capacity consumption for an account.

See \url{https://www.paws-r-sdk.com/docs/quicksight_list_users_index_capacity/} for full documentation.
}
\keyword{internal}
