% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_list_vpc_connections}
\alias{quicksight_list_vpc_connections}
\title{Lists all of the VPC connections in the current set Amazon Web Services
Region of an Amazon Web Services account}
\usage{
quicksight_list_vpc_connections(
  AwsAccountId,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{AwsAccountId}{[required] The Amazon Web Services account ID of the account that contains the VPC
connections that you want to list.}

\item{NextToken}{The token for the next set of results, or null if there are no more
results.}

\item{MaxResults}{The maximum number of results to be returned per request.}
}
\description{
Lists all of the VPC connections in the current set Amazon Web Services Region of an Amazon Web Services account.

See \url{https://www.paws-r-sdk.com/docs/quicksight_list_vpc_connections/} for full documentation.
}
\keyword{internal}
