% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_put_data_set_refresh_properties}
\alias{quicksight_put_data_set_refresh_properties}
\title{Creates or updates the dataset refresh properties for the dataset}
\usage{
quicksight_put_data_set_refresh_properties(
  AwsAccountId,
  DataSetId,
  DataSetRefreshProperties
)
}
\arguments{
\item{AwsAccountId}{[required] The Amazon Web Services account ID.}

\item{DataSetId}{[required] The ID of the dataset.}

\item{DataSetRefreshProperties}{[required] The dataset refresh properties.}
}
\description{
Creates or updates the dataset refresh properties for the dataset.

See \url{https://www.paws-r-sdk.com/docs/quicksight_put_data_set_refresh_properties/} for full documentation.
}
\keyword{internal}
