% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_register_user}
\alias{quicksight_register_user}
\title{Creates an Amazon QuickSight user, whose identity is associated with the
AWS Identity and Access Management (IAM) identity or role specified in
the request}
\usage{
quicksight_register_user(IdentityType, Email, UserRole, IamArn,
  SessionName, AwsAccountId, Namespace, UserName, CustomPermissionsName)
}
\arguments{
\item{IdentityType}{[required] Amazon QuickSight supports several ways of managing the identity of
users. This parameter accepts two values:
\itemize{
\item \code{IAM}: A user whose identity maps to an existing IAM user or role.
\item \code{QUICKSIGHT}: A user whose identity is owned and managed internally
by Amazon QuickSight.
}}

\item{Email}{[required] The email address of the user that you want to register.}

\item{UserRole}{[required] The Amazon QuickSight role for the user. The user role can be one of the
following:
\itemize{
\item \code{READER}: A user who has read-only access to dashboards.
\item \code{AUTHOR}: A user who can create data sources, datasets, analyses,
and dashboards.
\item \code{ADMIN}: A user who is an author, who can also manage Amazon
QuickSight settings.
\item \code{RESTRICTED_READER}: This role isn't currently available for use.
\item \code{RESTRICTED_AUTHOR}: This role isn't currently available for use.
}}

\item{IamArn}{The ARN of the IAM user or role that you are registering with Amazon
QuickSight.}

\item{SessionName}{You need to use this parameter only when you register one or more users
using an assumed IAM role. You don't need to provide the session name
for other scenarios, for example when you are registering an IAM user or
an Amazon QuickSight user. You can register multiple users using the
same IAM role if each user has a different session name. For more
information on assuming IAM roles, see
\href{https://docs.aws.amazon.com/cli/latest/reference/sts/assume-role.html}{\code{assume-role}}
in the \emph{AWS CLI Reference.}}

\item{AwsAccountId}{[required] The ID for the AWS account that the user is in. Currently, you use the
ID for the AWS account that contains your Amazon QuickSight account.}

\item{Namespace}{[required] The namespace. Currently, you should set this to \code{default}.}

\item{UserName}{The Amazon QuickSight user name that you want to create for the user you
are registering.}

\item{CustomPermissionsName}{(Enterprise edition only) The name of the custom permissions profile
that you want to assign to this user. Customized permissions allows you
to control a user's access by restricting access the following
operations:
\itemize{
\item Create and update data sources
\item Create and update datasets
\item Create and update email reports
\item Subscribe to email reports
}

To add custom permissions to an existing user, use
\code{\link[=quicksight_update_user]{update_user}} instead.

A set of custom permissions includes any combination of these
restrictions. Currently, you need to create the profile names for custom
permission sets by using the QuickSight console. Then, you use the
\code{\link[=quicksight_register_user]{register_user}} API operation to assign the
named set of permissions to a QuickSight user.

QuickSight custom permissions are applied through IAM policies.
Therefore, they override the permissions typically granted by assigning
QuickSight users to one of the default security cohorts in QuickSight
(admin, author, reader).

This feature is available only to QuickSight Enterprise edition
subscriptions that use SAML 2.0-Based Federation for Single Sign-On
(SSO).}
}
\value{
A list with the following syntax:\preformatted{list(
  User = list(
    Arn = "string",
    UserName = "string",
    Email = "string",
    Role = "ADMIN"|"AUTHOR"|"READER"|"RESTRICTED_AUTHOR"|"RESTRICTED_READER",
    IdentityType = "IAM"|"QUICKSIGHT",
    Active = TRUE|FALSE,
    PrincipalId = "string",
    CustomPermissionsName = "string"
  ),
  UserInvitationUrl = "string",
  RequestId = "string",
  Status = 123
)
}
}
\description{
Creates an Amazon QuickSight user, whose identity is associated with the
AWS Identity and Access Management (IAM) identity or role specified in
the request.
}
\section{Request syntax}{
\preformatted{svc$register_user(
  IdentityType = "IAM"|"QUICKSIGHT",
  Email = "string",
  UserRole = "ADMIN"|"AUTHOR"|"READER"|"RESTRICTED_AUTHOR"|"RESTRICTED_READER",
  IamArn = "string",
  SessionName = "string",
  AwsAccountId = "string",
  Namespace = "string",
  UserName = "string",
  CustomPermissionsName = "string"
)
}
}

\keyword{internal}
