% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_restore_analysis}
\alias{quicksight_restore_analysis}
\title{Restores an analysis}
\usage{
quicksight_restore_analysis(AwsAccountId, AnalysisId)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the AWS account that contains the analysis.}

\item{AnalysisId}{[required] The ID of the analysis that you're restoring.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Status = 123,
  Arn = "string",
  AnalysisId = "string",
  RequestId = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Restores an analysis.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$restore_analysis(
  AwsAccountId = "string",
  AnalysisId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
