% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_search_action_connectors}
\alias{quicksight_search_action_connectors}
\title{Searches for action connectors in the specified Amazon Web Services
account using filters}
\usage{
quicksight_search_action_connectors(
  AwsAccountId,
  MaxResults = NULL,
  NextToken = NULL,
  Filters
)
}
\arguments{
\item{AwsAccountId}{[required] The Amazon Web Services account ID in which to search for action
connectors.}

\item{MaxResults}{The maximum number of action connectors to return in a single response.
Valid range is 1 to 100.}

\item{NextToken}{A pagination token to retrieve the next set of results. Use the token
returned from a previous call to continue searching.}

\item{Filters}{[required] The search filters to apply. You can filter by connector name, type, or
user permissions. Maximum of one filter is supported.}
}
\description{
Searches for action connectors in the specified Amazon Web Services account using filters. You can search by connector name, type, or user permissions.

See \url{https://www.paws-r-sdk.com/docs/quicksight_search_action_connectors/} for full documentation.
}
\keyword{internal}
