% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_search_agents}
\alias{quicksight_search_agents}
\title{Searches for agents based on specified filters}
\usage{
quicksight_search_agents(
  AwsAccountId,
  Filters,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the agents.}

\item{Filters}{[required] The filters to apply when searching agents.}

\item{MaxResults}{The maximum number of results to return.}

\item{NextToken}{The token for the next set of results, or null if there are no more results.}
}
\description{
Searches for agents based on specified filters.

See \url{https://www.paws-r-sdk.com/docs/quicksight_search_agents/} for full documentation.
}
\keyword{internal}
