% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_search_analyses}
\alias{quicksight_search_analyses}
\title{Searches for analyses that belong to the user specified in the filter}
\usage{
quicksight_search_analyses(
  AwsAccountId,
  Filters,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the analyses
that you're searching for.}

\item{Filters}{[required] The structure for the search filters that you want to apply to your
search.}

\item{NextToken}{A pagination token that can be used in a subsequent request.}

\item{MaxResults}{The maximum number of results to return.}
}
\description{
Searches for analyses that belong to the user specified in the filter.

See \url{https://paws-r.github.io/docs/quicksight/search_analyses.html} for full documentation.
}
\keyword{internal}
