% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_search_apps}
\alias{quicksight_search_apps}
\title{Searches for apps in an Amazon Web Services account using the specified
filters}
\usage{
quicksight_search_apps(
  AwsAccountId,
  Filters,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the apps to search.}

\item{Filters}{[required] The filters to apply to the search.}

\item{MaxResults}{The maximum number of results to return in a single request. Valid range is 1 to 100. If you don't specify a value, the default is 20.}

\item{NextToken}{The token for the next set of results, or null if there are no more results.}
}
\description{
Searches for apps in an Amazon Web Services account using the specified filters. This operation is eventually consistent; the results might not reflect very recent updates. Results are paginated; use the \code{NextToken} parameter to retrieve additional results.

See \url{https://www.paws-r-sdk.com/docs/quicksight_search_apps/} for full documentation.
}
\keyword{internal}
