% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_search_flows}
\alias{quicksight_search_flows}
\title{Search for the flows in an Amazon Web Services account}
\usage{
quicksight_search_flows(
  AwsAccountId,
  Filters,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account where you are searching for
flows from.}

\item{Filters}{[required] The filters applied to the search when searching for flows in the Amazon
Web Services account.}

\item{NextToken}{The token to request the next set of results, or null if you want to
retrieve the first set.}

\item{MaxResults}{The maximum number of results to be returned per request.}
}
\description{
Search for the flows in an Amazon Web Services account.

See \url{https://www.paws-r-sdk.com/docs/quicksight_search_flows/} for full documentation.
}
\keyword{internal}
