% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_search_folders}
\alias{quicksight_search_folders}
\title{Searches the subfolders in a folder}
\usage{
quicksight_search_folders(
  AwsAccountId,
  Filters,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{AwsAccountId}{[required] The ID for the Amazon Web Services account that contains the folder.}

\item{Filters}{[required] The filters to apply to the search. Currently, you can search only by
the parent folder ARN. For example,
\verb{"Filters": [ \{ "Name": "PARENT_FOLDER_ARN", "Operator": "StringEquals", "Value": "arn:aws:quicksight:us-east-1:1:folder/folderId" \} ]}.}

\item{NextToken}{The token for the next set of results, or null if there are no more
results.}

\item{MaxResults}{The maximum number of results to be returned per request.}
}
\description{
Searches the subfolders in a folder.

See \url{https://paws-r.github.io/docs/quicksight/search_folders.html} for full documentation.
}
\keyword{internal}
