% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_search_groups}
\alias{quicksight_search_groups}
\title{Use the SearchGroups operation to search groups in a specified Quick
Sight namespace using the supplied filters}
\usage{
quicksight_search_groups(
  AwsAccountId,
  NextToken = NULL,
  MaxResults = NULL,
  Namespace,
  Filters
)
}
\arguments{
\item{AwsAccountId}{[required] The ID for the Amazon Web Services account that the group is in.
Currently, you use the ID for the Amazon Web Services account that
contains your Amazon Quick Sight account.}

\item{NextToken}{A pagination token that can be used in a subsequent request.}

\item{MaxResults}{The maximum number of results to return from this request.}

\item{Namespace}{[required] The namespace that you want to search.}

\item{Filters}{[required] The structure for the search filters that you want to apply to your
search.}
}
\description{
Use the \code{\link[=quicksight_search_groups]{search_groups}} operation to search groups in a specified Quick Sight namespace using the supplied filters.

See \url{https://www.paws-r-sdk.com/docs/quicksight_search_groups/} for full documentation.
}
\keyword{internal}
