% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_search_knowledge_bases}
\alias{quicksight_search_knowledge_bases}
\title{Searches for a subset of knowledge bases based on specified filters}
\usage{
quicksight_search_knowledge_bases(
  AwsAccountId,
  NextToken = NULL,
  MaxResults = NULL,
  Filters = NULL,
  SortBy = NULL
)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the knowledge base.}

\item{NextToken}{The token for the next set of results, or null if there are no more results.}

\item{MaxResults}{The maximum number of results to return.}

\item{Filters}{The filters to apply when searching knowledge bases.}

\item{SortBy}{The sort configuration for the search results.}
}
\description{
Searches for a subset of knowledge bases based on specified filters.

See \url{https://www.paws-r-sdk.com/docs/quicksight_search_knowledge_bases/} for full documentation.
}
\keyword{internal}
