% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_search_spaces}
\alias{quicksight_search_spaces}
\title{Searches for Amazon QuickSight spaces that match the specified filters}
\usage{
quicksight_search_spaces(
  AwsAccountId,
  NextToken = NULL,
  MaxResults = NULL,
  Filters
)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the spaces.}

\item{NextToken}{The token for the next set of results, or null if there are no more results.}

\item{MaxResults}{The maximum number of results to return.}

\item{Filters}{[required] The filters to apply to the search.}
}
\description{
Searches for Amazon QuickSight spaces that match the specified filters.

See \url{https://www.paws-r-sdk.com/docs/quicksight_search_spaces/} for full documentation.
}
\keyword{internal}
