% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_search_topics_v2}
\alias{quicksight_search_topics_v2}
\title{Searches for any Q topic that exists in an Amazon Web Services account}
\usage{
quicksight_search_topics_v2(
  AwsAccountId,
  Filters,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the topic that you want to search.}

\item{Filters}{[required] The filters that you want to use to search for the topic.}

\item{NextToken}{The token for the next set of results, or null if there are no more results.}

\item{MaxResults}{The maximum number of results to be returned per request.}
}
\description{
Searches for any Q topic that exists in an Amazon Web Services account.

See \url{https://www.paws-r-sdk.com/docs/quicksight_search_topics_v2/} for full documentation.
}
\keyword{internal}
