% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_start_automation_job}
\alias{quicksight_start_automation_job}
\title{Starts a new job for a specified automation}
\usage{
quicksight_start_automation_job(
  AwsAccountId,
  AutomationGroupId,
  AutomationId,
  InputPayload = NULL
)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the automation.}

\item{AutomationGroupId}{[required] The ID of the automation group that contains the automation to run.}

\item{AutomationId}{[required] The ID of the automation to run.}

\item{InputPayload}{The input payload for the automation job, provided as a JSON string.}
}
\description{
Starts a new job for a specified automation. The job runs the automation with the provided input payload.

See \url{https://www.paws-r-sdk.com/docs/quicksight_start_automation_job/} for full documentation.
}
\keyword{internal}
