% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_tag_resource}
\alias{quicksight_tag_resource}
\title{Assigns one or more tags (key-value pairs) to the specified Amazon
QuickSight resource}
\usage{
quicksight_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource that you want to tag.}

\item{Tags}{[required] Contains a map of the key-value pairs for the resource tag or tags
assigned to the resource.}
}
\description{
Assigns one or more tags (key-value pairs) to the specified Amazon QuickSight resource.

See \url{https://paws-r.github.io/docs/quicksight/tag_resource.html} for full documentation.
}
\keyword{internal}
