% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_untag_resource}
\alias{quicksight_untag_resource}
\title{Removes a tag or tags from a resource}
\usage{
quicksight_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource that you want to untag.}

\item{TagKeys}{[required] The keys of the key-value pairs for the resource tag or tags assigned to
the resource.}
}
\value{
A list with the following syntax:\preformatted{list(
  RequestId = "string",
  Status = 123
)
}
}
\description{
Removes a tag or tags from a resource.
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  ResourceArn = "string",
  TagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
