% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_update_account_custom_permission}
\alias{quicksight_update_account_custom_permission}
\title{Applies a custom permissions profile to an account}
\usage{
quicksight_update_account_custom_permission(
  CustomPermissionsName,
  AwsAccountId
)
}
\arguments{
\item{CustomPermissionsName}{[required] The name of the custom permissions profile that you want to apply to an
account.}

\item{AwsAccountId}{[required] The ID of the Amazon Web Services account for which you want to apply a
custom permissions profile.}
}
\description{
Applies a custom permissions profile to an account.

See \url{https://www.paws-r-sdk.com/docs/quicksight_update_account_custom_permission/} for full documentation.
}
\keyword{internal}
