% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_update_account_customization}
\alias{quicksight_update_account_customization}
\title{Updates Amazon QuickSight customizations for the current Amazon Web
Services Region}
\usage{
quicksight_update_account_customization(
  AwsAccountId,
  Namespace = NULL,
  AccountCustomization
)
}
\arguments{
\item{AwsAccountId}{[required] The ID for the Amazon Web Services account that you want to update
Amazon QuickSight customizations for.}

\item{Namespace}{The namespace that you want to update Amazon QuickSight customizations
for.}

\item{AccountCustomization}{[required] The Amazon QuickSight customizations you're updating in the current
Amazon Web Services Region.}
}
\description{
Updates Amazon QuickSight customizations for the current Amazon Web Services Region. Currently, the only customization that you can use is a theme.

See \url{https://www.paws-r-sdk.com/docs/quicksight_update_account_customization/} for full documentation.
}
\keyword{internal}
