% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_update_account_settings}
\alias{quicksight_update_account_settings}
\title{Updates the Amazon QuickSight settings in your Amazon Web Services
account}
\usage{
quicksight_update_account_settings(
  AwsAccountId,
  DefaultNamespace,
  NotificationEmail = NULL
)
}
\arguments{
\item{AwsAccountId}{[required] The ID for the Amazon Web Services account that contains the Amazon
QuickSight settings that you want to list.}

\item{DefaultNamespace}{[required] The default namespace for this Amazon Web Services account. Currently,
the default is \code{default}. Identity and Access Management (IAM) users
that register for the first time with Amazon QuickSight provide an email
address that becomes associated with the default namespace.}

\item{NotificationEmail}{The email address that you want Amazon QuickSight to send notifications
to regarding your Amazon Web Services account or Amazon QuickSight
subscription.}
}
\description{
Updates the Amazon QuickSight settings in your Amazon Web Services account.

See \url{https://paws-r.github.io/docs/quicksight/update_account_settings.html} for full documentation.
}
\keyword{internal}
