% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_update_action_connector}
\alias{quicksight_update_action_connector}
\title{Updates an existing action connector with new configuration details,
authentication settings, or enabled actions}
\usage{
quicksight_update_action_connector(
  AwsAccountId,
  ActionConnectorId,
  Name,
  AuthenticationConfig,
  Description = NULL,
  VpcConnectionArn = NULL
)
}
\arguments{
\item{AwsAccountId}{[required] The Amazon Web Services account ID that contains the action connector to
update.}

\item{ActionConnectorId}{[required] The unique identifier of the action connector to update.}

\item{Name}{[required] The new name for the action connector.}

\item{AuthenticationConfig}{[required] The updated authentication configuration for connecting to the external
service.}

\item{Description}{The updated description of the action connector.}

\item{VpcConnectionArn}{The updated ARN of the VPC connection to use for secure connectivity.}
}
\description{
Updates an existing action connector with new configuration details, authentication settings, or enabled actions. You can modify the connector's name, description, authentication configuration, and which actions are enabled. For more information, \url{https://docs.aws.amazon.com/quicksuite/latest/userguide/quick-action-auth.html}.

See \url{https://www.paws-r-sdk.com/docs/quicksight_update_action_connector/} for full documentation.
}
\keyword{internal}
