% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_update_agent}
\alias{quicksight_update_agent}
\title{Updates an existing agent}
\usage{
quicksight_update_agent(
  AgentId,
  AwsAccountId,
  Name,
  Description = NULL,
  IconId = NULL,
  StarterPrompts = NULL,
  WelcomeMessage = NULL,
  CustomPromptInput = NULL,
  SpacesToAdd = NULL,
  SpacesToRemove = NULL,
  ActionConnectorsToAdd = NULL,
  ActionConnectorsToRemove = NULL
)
}
\arguments{
\item{AgentId}{[required] The unique identifier for the agent to update.}

\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the agent.}

\item{Name}{[required] The name of the agent.}

\item{Description}{A description of the agent.}

\item{IconId}{The icon identifier for the agent.}

\item{StarterPrompts}{A list of starter prompts that are displayed to users when they begin interacting with the agent.}

\item{WelcomeMessage}{The welcome message that is displayed when a user starts a conversation with the agent.}

\item{CustomPromptInput}{The custom prompt configuration for the agent.}

\item{SpacesToAdd}{The Amazon Resource Names (ARNs) of the spaces to attach to the agent.}

\item{SpacesToRemove}{The Amazon Resource Names (ARNs) of the spaces to detach from the agent.}

\item{ActionConnectorsToAdd}{The Amazon Resource Names (ARNs) of the action connectors to attach to the agent.}

\item{ActionConnectorsToRemove}{The Amazon Resource Names (ARNs) of the action connectors to detach from the agent.}
}
\description{
Updates an existing agent.

See \url{https://www.paws-r-sdk.com/docs/quicksight_update_agent/} for full documentation.
}
\keyword{internal}
