% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_update_agent_permissions}
\alias{quicksight_update_agent_permissions}
\title{Updates the resource permissions for an agent}
\usage{
quicksight_update_agent_permissions(
  AgentId,
  AwsAccountId,
  GrantPermissions = NULL,
  RevokePermissions = NULL
)
}
\arguments{
\item{AgentId}{[required] The unique identifier for the agent.}

\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the agent.}

\item{GrantPermissions}{The resource permissions that you want to grant on the agent.}

\item{RevokePermissions}{The resource permissions that you want to revoke from the agent.}
}
\description{
Updates the resource permissions for an agent.

See \url{https://www.paws-r-sdk.com/docs/quicksight_update_agent_permissions/} for full documentation.
}
\keyword{internal}
