% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_update_app_permissions}
\alias{quicksight_update_app_permissions}
\title{Updates the resource permissions for an app}
\usage{
quicksight_update_app_permissions(
  AwsAccountId,
  AppId,
  GrantPermissions = NULL,
  RevokePermissions = NULL,
  Visibility = NULL
)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the app.}

\item{AppId}{[required] The ID of the app.}

\item{GrantPermissions}{The permissions that you want to grant on the app.}

\item{RevokePermissions}{The permissions that you want to revoke from the app.}

\item{Visibility}{The visibility to set for the app. Currently, only \code{PRIVATE} is accepted, which removes public (anonymous) access from the app. If you don't specify a value, the app's visibility is unchanged. Setting an app to \code{PUBLIC} through this operation is not supported.}
}
\description{
Updates the resource permissions for an app. You can grant or revoke permissions and, optionally, change the app's visibility.

See \url{https://www.paws-r-sdk.com/docs/quicksight_update_app_permissions/} for full documentation.
}
\keyword{internal}
