% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_update_approval_policy}
\alias{quicksight_update_approval_policy}
\title{Updates an approval policy in Quick Sight}
\usage{
quicksight_update_approval_policy(
  PolicyId,
  Name = NULL,
  Description = NULL,
  Actions = NULL,
  AssetTypes = NULL,
  ApplicableTo = NULL,
  ApprovalGroups = NULL
)
}
\arguments{
\item{PolicyId}{[required] The unique identifier of the approval policy to update.}

\item{Name}{The name of the approval policy.}

\item{Description}{A description of the approval policy.}

\item{Actions}{The list of governed actions that trigger the approval workflow.}

\item{AssetTypes}{The list of asset types that the approval policy applies to.}

\item{ApplicableTo}{The scoping configuration that determines who the approval policy applies to.}

\item{ApprovalGroups}{The list of group ARNs whose members can approve requests.}
}
\description{
Updates an approval policy in Quick Sight.

See \url{https://www.paws-r-sdk.com/docs/quicksight_update_approval_policy/} for full documentation.
}
\keyword{internal}
