% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_update_brand}
\alias{quicksight_update_brand}
\title{Updates a brand}
\usage{
quicksight_update_brand(AwsAccountId, BrandId, BrandDefinition = NULL)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that owns the brand.}

\item{BrandId}{[required] The ID of the Quick Suite brand.}

\item{BrandDefinition}{The definition of the brand.}
}
\description{
Updates a brand.

See \url{https://www.paws-r-sdk.com/docs/quicksight_update_brand/} for full documentation.
}
\keyword{internal}
