% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_update_brand_assignment}
\alias{quicksight_update_brand_assignment}
\title{Updates a brand assignment}
\usage{
quicksight_update_brand_assignment(AwsAccountId, BrandArn)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that owns the brand
assignment.}

\item{BrandArn}{[required] The Amazon Resource Name (ARN) of the brand.}
}
\description{
Updates a brand assignment.

See \url{https://www.paws-r-sdk.com/docs/quicksight_update_brand_assignment/} for full documentation.
}
\keyword{internal}
