% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_update_brand_published_version}
\alias{quicksight_update_brand_published_version}
\title{Updates the published version of a brand}
\usage{
quicksight_update_brand_published_version(AwsAccountId, BrandId, VersionId)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that owns the brand.}

\item{BrandId}{[required] The ID of the Quick Suite brand.}

\item{VersionId}{[required] The ID of the published version.}
}
\description{
Updates the published version of a brand.

See \url{https://www.paws-r-sdk.com/docs/quicksight_update_brand_published_version/} for full documentation.
}
\keyword{internal}
