% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_update_custom_permissions}
\alias{quicksight_update_custom_permissions}
\title{Updates a custom permissions profile}
\usage{
quicksight_update_custom_permissions(
  AwsAccountId,
  CustomPermissionsName,
  Capabilities = NULL,
  Governance = NULL
)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the custom permissions profile that you want to update.}

\item{CustomPermissionsName}{[required] The name of the custom permissions profile that you want to update.}

\item{Capabilities}{A set of actions to include in the custom permissions profile.}

\item{Governance}{The governance configuration for the custom permissions profile. The \code{\link[=quicksight_update_custom_permissions]{update_custom_permissions}} operation replaces all existing \code{Capabilities} and \code{Governance} values. If you omit this parameter, Amazon Quick removes governance from the profile and the existing custom permission behavior applies.}
}
\description{
Updates a custom permissions profile.

See \url{https://www.paws-r-sdk.com/docs/quicksight_update_custom_permissions/} for full documentation.
}
\keyword{internal}
