% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_update_dashboard_links}
\alias{quicksight_update_dashboard_links}
\title{Updates the linked analyses on a dashboard}
\usage{
quicksight_update_dashboard_links(AwsAccountId, DashboardId, LinkEntities)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the dashboard
whose links you want to update.}

\item{DashboardId}{[required] The ID for the dashboard.}

\item{LinkEntities}{[required] list of analysis Amazon Resource Names (ARNs) to be linked to the
dashboard.}
}
\description{
Updates the linked analyses on a dashboard.

See \url{https://www.paws-r-sdk.com/docs/quicksight_update_dashboard_links/} for full documentation.
}
\keyword{internal}
