% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_update_dashboard_permissions}
\alias{quicksight_update_dashboard_permissions}
\title{Updates read and write permissions on a dashboard}
\usage{
quicksight_update_dashboard_permissions(AwsAccountId, DashboardId,
  GrantPermissions, RevokePermissions)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the AWS account that contains the dashboard whose permissions
you're updating.}

\item{DashboardId}{[required] The ID for the dashboard.}

\item{GrantPermissions}{The permissions that you want to grant on this resource.}

\item{RevokePermissions}{The permissions that you want to revoke from this resource.}
}
\description{
Updates read and write permissions on a dashboard.
}
\section{Request syntax}{
\preformatted{svc$update_dashboard_permissions(
  AwsAccountId = "string",
  DashboardId = "string",
  GrantPermissions = list(
    list(
      Principal = "string",
      Actions = list(
        "string"
      )
    )
  ),
  RevokePermissions = list(
    list(
      Principal = "string",
      Actions = list(
        "string"
      )
    )
  )
)
}
}

\keyword{internal}
