% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_update_dashboards_qa_configuration}
\alias{quicksight_update_dashboards_qa_configuration}
\title{Updates a Dashboard QA configuration}
\usage{
quicksight_update_dashboards_qa_configuration(AwsAccountId, DashboardsQAStatus)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the dashboard QA
configuration that you want to update.}

\item{DashboardsQAStatus}{[required] The status of dashboards QA configuration that you want to update.}
}
\description{
Updates a Dashboard QA configuration.

See \url{https://www.paws-r-sdk.com/docs/quicksight_update_dashboards_qa_configuration/} for full documentation.
}
\keyword{internal}
