% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_update_data_set_permissions}
\alias{quicksight_update_data_set_permissions}
\title{Updates the permissions on a dataset}
\usage{
quicksight_update_data_set_permissions(
  AwsAccountId,
  DataSetId,
  GrantPermissions = NULL,
  RevokePermissions = NULL
)
}
\arguments{
\item{AwsAccountId}{[required] The Amazon Web Services account ID.}

\item{DataSetId}{[required] The ID for the dataset whose permissions you want to update. This ID is
unique per Amazon Web Services Region for each Amazon Web Services
account.}

\item{GrantPermissions}{The resource permissions that you want to grant to the dataset.}

\item{RevokePermissions}{The resource permissions that you want to revoke from the dataset.}
}
\description{
Updates the permissions on a dataset.

See \url{https://www.paws-r-sdk.com/docs/quicksight_update_data_set_permissions/} for full documentation.
}
\keyword{internal}
