% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_update_dlp_setting}
\alias{quicksight_update_dlp_setting}
\title{Updates an existing DLP setting configuration in an Amazon Web Services
account}
\usage{
quicksight_update_dlp_setting(
  AwsAccountId,
  DlpSettingId,
  Name = NULL,
  ProviderType = NULL,
  ProviderConfig = NULL,
  ProviderOutageAction = NULL,
  Enabled = NULL
)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the DLP setting that you want to update.}

\item{DlpSettingId}{[required] The ID of the DLP setting that you want to update.}

\item{Name}{An updated display name for the DLP setting.}

\item{ProviderType}{An updated DLP provider type. Currently, the only supported value is \code{MICROSOFT_PURVIEW}.}

\item{ProviderConfig}{An updated provider-specific configuration for the DLP integration. This is a union type structure. For this structure to be valid, only one of the attributes can be defined.}

\item{ProviderOutageAction}{An updated behavior to apply when the DLP provider is unreachable. Valid values are \code{ALLOW}, \code{WARN}, and \code{BLOCK}.}

\item{Enabled}{Specifies whether DLP enforcement is active for this setting. Set to \code{true} to enable enforcement, or \code{false} to disable it.}
}
\description{
Updates an existing DLP setting configuration in an Amazon Web Services account. Fields that are omitted from the request retain their current values.

See \url{https://www.paws-r-sdk.com/docs/quicksight_update_dlp_setting/} for full documentation.
}
\keyword{internal}
