% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_update_flow_permissions}
\alias{quicksight_update_flow_permissions}
\title{Updates permissions against principals on a flow}
\usage{
quicksight_update_flow_permissions(
  AwsAccountId,
  FlowId,
  GrantPermissions = NULL,
  RevokePermissions = NULL
)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the flow you are
updating permissions against.}

\item{FlowId}{[required] The unique identifier of the flow to update permissions for.}

\item{GrantPermissions}{The permissions that you want to grant on this flow.}

\item{RevokePermissions}{The permissions that you want to revoke from this flow.}
}
\description{
Updates permissions against principals on a flow.

See \url{https://www.paws-r-sdk.com/docs/quicksight_update_flow_permissions/} for full documentation.
}
\keyword{internal}
