% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_update_folder}
\alias{quicksight_update_folder}
\title{Updates the name of a folder}
\usage{
quicksight_update_folder(AwsAccountId, FolderId, Name)
}
\arguments{
\item{AwsAccountId}{[required] The ID for the Amazon Web Services account that contains the folder to
update.}

\item{FolderId}{[required] The ID of the folder.}

\item{Name}{[required] The name of the folder.}
}
\description{
Updates the name of a folder.

See \url{https://paws-r.github.io/docs/quicksight/update_folder.html} for full documentation.
}
\keyword{internal}
