% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_update_group}
\alias{quicksight_update_group}
\title{Changes a group description}
\usage{
quicksight_update_group(GroupName, Description = NULL, AwsAccountId, Namespace)
}
\arguments{
\item{GroupName}{[required] The name of the group that you want to update.}

\item{Description}{The description for the group that you want to update.}

\item{AwsAccountId}{[required] The ID for the Amazon Web Services account that the group is in.
Currently, you use the ID for the Amazon Web Services account that
contains your Amazon QuickSight account.}

\item{Namespace}{[required] The namespace of the group that you want to update.}
}
\description{
Changes a group description.

See \url{https://paws-r.github.io/docs/quicksight/update_group.html} for full documentation.
}
\keyword{internal}
