% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_update_iam_policy_assignment}
\alias{quicksight_update_iam_policy_assignment}
\title{Updates an existing assignment}
\usage{
quicksight_update_iam_policy_assignment(AwsAccountId, AssignmentName,
  Namespace, AssignmentStatus, PolicyArn, Identities)
}
\arguments{
\item{AwsAccountId}{[required] The AWS account ID that contains the IAM policy assignment.}

\item{AssignmentName}{[required] The name of the assignment. It must be unique within an AWS account.}

\item{Namespace}{[required] The namespace of the assignment.}

\item{AssignmentStatus}{The status of an assignment:
\itemize{
\item ENABLED - Anything specified in this assignment is used while
creating the data source.
\item DISABLED - This assignment isn\'t used while creating the data
source.
\item DRAFT - Assignment is an unfinished draft and isn\'t used while
creating the data source.
}}

\item{PolicyArn}{An IAM policy Amazon Resource Name (ARN) that will be applied to
specified QuickSight users and groups in this assignment.}

\item{Identities}{QuickSight users and/or groups that you want to assign to the specified
IAM policy.}
}
\description{
Updates an existing assignment. This operation updates only the optional
parameter or parameters that are specified in the request.
}
\section{Request syntax}{
\preformatted{svc$update_iam_policy_assignment(
  AwsAccountId = "string",
  AssignmentName = "string",
  Namespace = "string",
  AssignmentStatus = "ENABLED"|"DRAFT"|"DISABLED",
  PolicyArn = "string",
  Identities = list(
    list(
      "string"
    )
  )
)
}
}

\keyword{internal}
