% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_update_identity_propagation_config}
\alias{quicksight_update_identity_propagation_config}
\title{Adds or updates services and authorized targets to configure what the
Quick Sight IAM Identity Center application can access}
\usage{
quicksight_update_identity_propagation_config(
  AwsAccountId,
  Service,
  AuthorizedTargets = NULL
)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the identity
propagation configuration that you want to update.}

\item{Service}{[required] The name of the Amazon Web Services service that contains the authorized
targets that you want to add or update.}

\item{AuthorizedTargets}{Specifies a list of application ARNs that represent the authorized
targets for a service.}
}
\description{
Adds or updates services and authorized targets to configure what the Quick Sight IAM Identity Center application can access.

See \url{https://www.paws-r-sdk.com/docs/quicksight_update_identity_propagation_config/} for full documentation.
}
\keyword{internal}
