% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_update_key_registration}
\alias{quicksight_update_key_registration}
\title{Updates a customer managed key in a Quick Sight account}
\usage{
quicksight_update_key_registration(AwsAccountId, KeyRegistration)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the customer
managed key registration that you want to update.}

\item{KeyRegistration}{[required] A list of \code{RegisteredCustomerManagedKey} objects to be updated to the
Quick Sight account.}
}
\description{
Updates a customer managed key in a Quick Sight account.

See \url{https://www.paws-r-sdk.com/docs/quicksight_update_key_registration/} for full documentation.
}
\keyword{internal}
