% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_update_knowledge_base}
\alias{quicksight_update_knowledge_base}
\title{Updates the properties of an existing knowledge base}
\usage{
quicksight_update_knowledge_base(
  AwsAccountId,
  KnowledgeBaseId,
  Name = NULL,
  Description = NULL,
  KnowledgeBaseConfiguration = NULL,
  MediaExtractionConfiguration = NULL,
  IsEmailNotificationOptedForIngestionFailures = NULL,
  AccessControlConfiguration = NULL
)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the knowledge base.}

\item{KnowledgeBaseId}{[required] The unique identifier for the knowledge base.}

\item{Name}{The name of the knowledge base. If you don't specify a name, the existing name is retained.}

\item{Description}{A description for the knowledge base. If you don't specify a description, the existing description is retained.}

\item{KnowledgeBaseConfiguration}{The configuration settings for a knowledge base.}

\item{MediaExtractionConfiguration}{The configuration for media extraction from knowledge base documents.}

\item{IsEmailNotificationOptedForIngestionFailures}{Specifies whether email notifications are enabled for ingestion failures.}

\item{AccessControlConfiguration}{The access control configuration for the knowledge base. If you don't specify this parameter, the existing setting is retained.}
}
\description{
Updates the properties of an existing knowledge base.

See \url{https://www.paws-r-sdk.com/docs/quicksight_update_knowledge_base/} for full documentation.
}
\keyword{internal}
