% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_update_knowledge_base_permissions}
\alias{quicksight_update_knowledge_base_permissions}
\title{Updates the resource permissions for a knowledge base}
\usage{
quicksight_update_knowledge_base_permissions(
  AwsAccountId,
  KnowledgeBaseId,
  GrantPermissions = NULL,
  RevokePermissions = NULL
)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the knowledge base.}

\item{KnowledgeBaseId}{[required] The unique identifier for the knowledge base.}

\item{GrantPermissions}{The resource permissions that you want to grant on the knowledge base.}

\item{RevokePermissions}{The resource permissions that you want to revoke from the knowledge base.}
}
\description{
Updates the resource permissions for a knowledge base.

See \url{https://www.paws-r-sdk.com/docs/quicksight_update_knowledge_base_permissions/} for full documentation.
}
\keyword{internal}
