% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_update_limits_profile}
\alias{quicksight_update_limits_profile}
\title{Updates the properties of an existing limits profile}
\usage{
quicksight_update_limits_profile(
  profileId,
  accountId,
  profileName = NULL,
  description = NULL,
  resourceLimits = NULL
)
}
\arguments{
\item{profileId}{[required] The unique identifier for the limits profile to update.}

\item{accountId}{[required] The ID of the Amazon Web Services account that contains the limits profile.}

\item{profileName}{A new display name for the limits profile.}

\item{description}{A new description for the limits profile.}

\item{resourceLimits}{A map of resource types to their updated limit values.}
}
\description{
Updates the properties of an existing limits profile.

See \url{https://www.paws-r-sdk.com/docs/quicksight_update_limits_profile/} for full documentation.
}
\keyword{internal}
