% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_update_o_auth_client_application}
\alias{quicksight_update_o_auth_client_application}
\title{Updates an OAuthClientApplication}
\usage{
quicksight_update_o_auth_client_application(
  AwsAccountId,
  OAuthClientApplicationId,
  Name,
  ClientId = NULL,
  ClientSecret = NULL,
  OAuthTokenEndpointUrl = NULL,
  OAuthAuthorizationEndpointUrl = NULL,
  OAuthScopes = NULL,
  DataSourceType = NULL,
  IdentityProviderVpcConnectionProperties = NULL
)
}
\arguments{
\item{AwsAccountId}{[required] The Amazon Web Services account ID.}

\item{OAuthClientApplicationId}{[required] The ID of the OAuthClientApplication that you want to update.}

\item{Name}{[required] The display name for the OAuthClientApplication.}

\item{ClientId}{The client ID of the OAuth application that is registered with the identity provider.}

\item{ClientSecret}{The client secret of the OAuth application that is registered with the identity provider.}

\item{OAuthTokenEndpointUrl}{The token endpoint URL of the identity provider that is used to obtain access tokens.}

\item{OAuthAuthorizationEndpointUrl}{The authorization endpoint URL of the identity provider that is used to obtain authorization codes.}

\item{OAuthScopes}{The OAuth scopes that are requested when the OAuthClientApplication obtains an access token from the identity provider.}

\item{DataSourceType}{The type of data source that the OAuthClientApplication is used with. Valid values are \code{SNOWFLAKE}.}

\item{IdentityProviderVpcConnectionProperties}{VPC connection properties.}
}
\description{
Updates an OAuthClientApplication.

See \url{https://www.paws-r-sdk.com/docs/quicksight_update_o_auth_client_application/} for full documentation.
}
\keyword{internal}
