% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_update_public_sharing_settings}
\alias{quicksight_update_public_sharing_settings}
\title{Use the UpdatePublicSharingSettings operation to turn on or turn off the
public sharing settings of an Amazon QuickSight dashboard}
\usage{
quicksight_update_public_sharing_settings(
  AwsAccountId,
  PublicSharingEnabled = NULL
)
}
\arguments{
\item{AwsAccountId}{[required] The Amazon Web Services account ID associated with your Amazon
QuickSight subscription.}

\item{PublicSharingEnabled}{A Boolean value that indicates whether public sharing is turned on for
an Amazon QuickSight account.}
}
\description{
Use the \code{\link[=quicksight_update_public_sharing_settings]{update_public_sharing_settings}} operation to turn on or turn off the public sharing settings of an Amazon QuickSight dashboard.

See \url{https://www.paws-r-sdk.com/docs/quicksight_update_public_sharing_settings/} for full documentation.
}
\keyword{internal}
