% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_update_q_personalization_configuration}
\alias{quicksight_update_q_personalization_configuration}
\title{Updates a personalization configuration}
\usage{
quicksight_update_q_personalization_configuration(
  AwsAccountId,
  PersonalizationMode
)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account account that contains the
personalization configuration that the user wants to update.}

\item{PersonalizationMode}{[required] An option to allow Amazon Quick Sight to customize data stories with
user specific metadata, specifically location and job information, in
your IAM Identity Center instance.}
}
\description{
Updates a personalization configuration.

See \url{https://www.paws-r-sdk.com/docs/quicksight_update_q_personalization_configuration/} for full documentation.
}
\keyword{internal}
