% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_update_quick_sight_q_search_configuration}
\alias{quicksight_update_quick_sight_q_search_configuration}
\title{Updates the state of a Quick Sight Q Search configuration}
\usage{
quicksight_update_quick_sight_q_search_configuration(
  AwsAccountId,
  QSearchStatus
)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the Quick Sight
Q Search configuration that you want to update.}

\item{QSearchStatus}{[required] The status of the Quick Sight Q Search configuration that the user wants
to update.}
}
\description{
Updates the state of a Quick Sight Q Search configuration.

See \url{https://www.paws-r-sdk.com/docs/quicksight_update_quick_sight_q_search_configuration/} for full documentation.
}
\keyword{internal}
