% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_update_refresh_schedule}
\alias{quicksight_update_refresh_schedule}
\title{Updates a refresh schedule for a dataset}
\usage{
quicksight_update_refresh_schedule(DataSetId, AwsAccountId, Schedule)
}
\arguments{
\item{DataSetId}{[required] The ID of the dataset.}

\item{AwsAccountId}{[required] The Amazon Web Services account ID.}

\item{Schedule}{[required] The refresh schedule.}
}
\description{
Updates a refresh schedule for a dataset.

See \url{https://paws-r.github.io/docs/quicksight/update_refresh_schedule.html} for full documentation.
}
\keyword{internal}
