% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_update_role_custom_permission}
\alias{quicksight_update_role_custom_permission}
\title{Updates the custom permissions that are associated with a role}
\usage{
quicksight_update_role_custom_permission(
  CustomPermissionsName,
  Role,
  AwsAccountId,
  Namespace
)
}
\arguments{
\item{CustomPermissionsName}{[required] The name of the custom permission that you want to update the role with.}

\item{Role}{[required] The name of role tht you want to update.}

\item{AwsAccountId}{[required] The ID for the Amazon Web Services account that you want to create a
group in. The Amazon Web Services account ID that you provide must be
the same Amazon Web Services account that contains your Amazon Quick
Sight account.}

\item{Namespace}{[required] The namespace that contains the role that you want to update.}
}
\description{
Updates the custom permissions that are associated with a role.

See \url{https://www.paws-r-sdk.com/docs/quicksight_update_role_custom_permission/} for full documentation.
}
\keyword{internal}
