% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_update_self_upgrade}
\alias{quicksight_update_self_upgrade}
\title{Updates a self-upgrade request for a Quick Suite user by approving,
denying, or verifying the request}
\usage{
quicksight_update_self_upgrade(
  AwsAccountId,
  Namespace,
  UpgradeRequestId,
  Action
)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the self-upgrade
request.}

\item{Namespace}{[required] The Quick Suite namespace for the self-upgrade request.}

\item{UpgradeRequestId}{[required] The ID of the self-upgrade request to update.}

\item{Action}{[required] The action to perform on the self-upgrade request. Valid values are
\code{APPROVE}, \code{DENY}, or \code{VERIFY}.}
}
\description{
Updates a self-upgrade request for a Quick Suite user by approving, denying, or verifying the request.

See \url{https://www.paws-r-sdk.com/docs/quicksight_update_self_upgrade/} for full documentation.
}
\keyword{internal}
