% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_update_space}
\alias{quicksight_update_space}
\title{Updates the metadata of an Amazon QuickSight space}
\usage{
quicksight_update_space(AwsAccountId, SpaceId, Name = NULL, Description = NULL)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the space.}

\item{SpaceId}{[required] The ID of the space that you want to update.}

\item{Name}{A new display name for the space.}

\item{Description}{A new description for the space.}
}
\description{
Updates the metadata of an Amazon QuickSight space.

See \url{https://www.paws-r-sdk.com/docs/quicksight_update_space/} for full documentation.
}
\keyword{internal}
