% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_update_space_permissions}
\alias{quicksight_update_space_permissions}
\title{Updates the permissions for an Amazon QuickSight space}
\usage{
quicksight_update_space_permissions(
  AwsAccountId,
  SpaceId,
  GrantPermissions = NULL,
  RevokePermissions = NULL
)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the space.}

\item{SpaceId}{[required] The ID of the space that you want to update permissions for.}

\item{GrantPermissions}{The permissions that you want to grant on the space.}

\item{RevokePermissions}{The permissions that you want to revoke from the space.}
}
\description{
Updates the permissions for an Amazon QuickSight space.

See \url{https://www.paws-r-sdk.com/docs/quicksight_update_space_permissions/} for full documentation.
}
\keyword{internal}
