% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_update_space_resources}
\alias{quicksight_update_space_resources}
\title{Adds or removes resources from an Amazon QuickSight space}
\usage{
quicksight_update_space_resources(
  AwsAccountId,
  SpaceId,
  AddResources = NULL,
  RemoveResources = NULL
)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the space.}

\item{SpaceId}{[required] The ID of the space that you want to update resources for.}

\item{AddResources}{A list of resources to add to the space.}

\item{RemoveResources}{A list of resources to remove from the space.}
}
\description{
Adds or removes resources from an Amazon QuickSight space.

See \url{https://www.paws-r-sdk.com/docs/quicksight_update_space_resources/} for full documentation.
}
\keyword{internal}
