% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_update_spice_capacity_configuration}
\alias{quicksight_update_spice_capacity_configuration}
\title{Updates the SPICE capacity configuration for a Quick Sight account}
\usage{
quicksight_update_spice_capacity_configuration(AwsAccountId, PurchaseMode)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the SPICE
configuration that you want to update.}

\item{PurchaseMode}{[required] Determines how SPICE capacity can be purchased. The following options
are available.
\itemize{
\item \code{MANUAL}: SPICE capacity can only be purchased manually.
\item \code{AUTO_PURCHASE}: Extra SPICE capacity is automatically purchased on
your behalf as needed. SPICE capacity can also be purchased manually
with this option.
}}
}
\description{
Updates the SPICE capacity configuration for a Quick Sight account.

See \url{https://www.paws-r-sdk.com/docs/quicksight_update_spice_capacity_configuration/} for full documentation.
}
\keyword{internal}
