% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_update_theme_alias}
\alias{quicksight_update_theme_alias}
\title{Updates an alias of a theme}
\usage{
quicksight_update_theme_alias(
  AwsAccountId,
  ThemeId,
  AliasName,
  ThemeVersionNumber
)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the theme alias
that you're updating.}

\item{ThemeId}{[required] The ID for the theme.}

\item{AliasName}{[required] The name of the theme alias that you want to update.}

\item{ThemeVersionNumber}{[required] The version number of the theme that the alias should reference.}
}
\description{
Updates an alias of a theme.

See \url{https://paws-r.github.io/docs/quicksight/update_theme_alias.html} for full documentation.
}
\keyword{internal}
