% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_update_theme_permissions}
\alias{quicksight_update_theme_permissions}
\title{Updates the resource permissions for a theme}
\usage{
quicksight_update_theme_permissions(
  AwsAccountId,
  ThemeId,
  GrantPermissions = NULL,
  RevokePermissions = NULL
)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the theme.}

\item{ThemeId}{[required] The ID for the theme.}

\item{GrantPermissions}{A list of resource permissions to be granted for the theme.}

\item{RevokePermissions}{A list of resource permissions to be revoked from the theme.}
}
\description{
Updates the resource permissions for a theme. Permissions apply to the action to grant or revoke permissions on, for example \code{"quicksight:DescribeTheme"}.

See \url{https://paws-r.github.io/docs/quicksight/update_theme_permissions.html} for full documentation.
}
\keyword{internal}
