% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_update_topic_permissions}
\alias{quicksight_update_topic_permissions}
\title{Updates the permissions of a topic}
\usage{
quicksight_update_topic_permissions(
  AwsAccountId,
  TopicId,
  GrantPermissions = NULL,
  RevokePermissions = NULL
)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the topic that
you want to update the permissions for.}

\item{TopicId}{[required] The ID of the topic that you want to modify. This ID is unique per
Amazon Web Services Region for each Amazon Web Services account.}

\item{GrantPermissions}{The resource permissions that you want to grant to the topic.}

\item{RevokePermissions}{The resource permissions that you want to revoke from the topic.}
}
\description{
Updates the permissions of a topic.

See \url{https://paws-r.github.io/docs/quicksight/update_topic_permissions.html} for full documentation.
}
\keyword{internal}
