% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_update_topic_permissions_v2}
\alias{quicksight_update_topic_permissions_v2}
\title{Updates the permissions of a topic}
\usage{
quicksight_update_topic_permissions_v2(
  AwsAccountId,
  TopicId,
  GrantPermissions = NULL,
  RevokePermissions = NULL
)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the topic that you want to update the permissions for.}

\item{TopicId}{[required] The ID of the topic that you want to modify. This ID is unique per Amazon Web Services Region for each Amazon Web Services account.}

\item{GrantPermissions}{The resource permissions that you want to grant to the topic.}

\item{RevokePermissions}{The resource permissions that you want to revoke from the topic.}
}
\description{
Updates the permissions of a topic.

See \url{https://www.paws-r-sdk.com/docs/quicksight_update_topic_permissions_v2/} for full documentation.
}
\keyword{internal}
