% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_update_topic_v2}
\alias{quicksight_update_topic_v2}
\title{Updates the definition of a Q topic}
\usage{
quicksight_update_topic_v2(
  AwsAccountId,
  TopicId,
  Topic,
  CustomInstructions = NULL,
  PublishOption = NULL
)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the topic that you want to update.}

\item{TopicId}{[required] The ID of the topic that you want to modify. This ID is unique per Amazon Web Services Region for each Amazon Web Services account.}

\item{Topic}{[required] The definition of the topic that you want to update.}

\item{CustomInstructions}{Instructions that provide additional guidance and context for response generation.}

\item{PublishOption}{The publish option for the topic that you want to update.}
}
\description{
Updates the definition of a Q topic.

See \url{https://www.paws-r-sdk.com/docs/quicksight_update_topic_v2/} for full documentation.
}
\keyword{internal}
