% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_update_user_custom_permission}
\alias{quicksight_update_user_custom_permission}
\title{Updates a custom permissions profile for a user}
\usage{
quicksight_update_user_custom_permission(
  UserName,
  AwsAccountId,
  Namespace,
  CustomPermissionsName
)
}
\arguments{
\item{UserName}{[required] The username of the user that you want to update custom permissions for.}

\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the custom
permission configuration that you want to update.}

\item{Namespace}{[required] The namespace that the user belongs to.}

\item{CustomPermissionsName}{[required] The name of the custom permissions that you want to update.}
}
\description{
Updates a custom permissions profile for a user.

See \url{https://www.paws-r-sdk.com/docs/quicksight_update_user_custom_permission/} for full documentation.
}
\keyword{internal}
