% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_operations.R
\name{athena_delete_data_catalog}
\alias{athena_delete_data_catalog}
\title{Deletes a data catalog}
\usage{
athena_delete_data_catalog(Name)
}
\arguments{
\item{Name}{[required] The name of the data catalog to delete.}
}
\value{
An empty list.
}
\description{
Deletes a data catalog.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_data_catalog(
  Name = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
