% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_operations.R
\name{athena_delete_data_catalog}
\alias{athena_delete_data_catalog}
\title{Deletes a data catalog}
\usage{
athena_delete_data_catalog(Name, DeleteCatalogOnly = NULL)
}
\arguments{
\item{Name}{[required] The name of the data catalog to delete.}

\item{DeleteCatalogOnly}{Deletes the Athena Data Catalog. You can only use this with the
\code{FEDERATED} catalogs. You usually perform this before registering the
connector with Glue Data Catalog. After deletion, you will have to
manage the Glue Connection and Lambda function.}
}
\description{
Deletes a data catalog.

See \url{https://www.paws-r-sdk.com/docs/athena_delete_data_catalog/} for full documentation.
}
\keyword{internal}
