% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_operations.R
\name{athena_delete_prepared_statement}
\alias{athena_delete_prepared_statement}
\title{Deletes the prepared statement with the specified name from the
specified workgroup}
\usage{
athena_delete_prepared_statement(StatementName, WorkGroup)
}
\arguments{
\item{StatementName}{[required] The name of the prepared statement to delete.}

\item{WorkGroup}{[required] The workgroup to which the statement to be deleted belongs.}
}
\description{
Deletes the prepared statement with the specified name from the specified workgroup.

See \url{https://paws-r.github.io/docs/athena/delete_prepared_statement.html} for full documentation.
}
\keyword{internal}
