% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_operations.R
\name{athena_get_calculation_execution_status}
\alias{athena_get_calculation_execution_status}
\title{Gets the status of a current calculation}
\usage{
athena_get_calculation_execution_status(CalculationExecutionId)
}
\arguments{
\item{CalculationExecutionId}{[required] The calculation execution UUID.}
}
\description{
Gets the status of a current calculation.

See \url{https://paws-r.github.io/docs/athena/get_calculation_execution_status.html} for full documentation.
}
\keyword{internal}
