% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_operations.R
\name{athena_get_database}
\alias{athena_get_database}
\title{Returns a database object for the specfied database and data catalog}
\usage{
athena_get_database(CatalogName, DatabaseName)
}
\arguments{
\item{CatalogName}{[required] The name of the data catalog that contains the database to return.}

\item{DatabaseName}{[required] The name of the database to return.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Database = list(
    Name = "string",
    Description = "string",
    Parameters = list(
      "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Returns a database object for the specfied database and data catalog.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_database(
  CatalogName = "string",
  DatabaseName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
