% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_operations.R
\name{athena_get_prepared_statement}
\alias{athena_get_prepared_statement}
\title{Retrieves the prepared statement with the specified name from the
specified workgroup}
\usage{
athena_get_prepared_statement(StatementName, WorkGroup)
}
\arguments{
\item{StatementName}{[required] The name of the prepared statement to retrieve.}

\item{WorkGroup}{[required] The workgroup to which the statement to be retrieved belongs.}
}
\value{
A list with the following syntax:\preformatted{list(
  PreparedStatement = list(
    StatementName = "string",
    QueryStatement = "string",
    WorkGroupName = "string",
    Description = "string",
    LastModifiedTime = as.POSIXct(
      "2015-01-01"
    )
  )
)
}
}
\description{
Retrieves the prepared statement with the specified name from the
specified workgroup.
}
\section{Request syntax}{
\preformatted{svc$get_prepared_statement(
  StatementName = "string",
  WorkGroup = "string"
)
}
}

\keyword{internal}
