% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_operations.R
\name{athena_get_query_runtime_statistics}
\alias{athena_get_query_runtime_statistics}
\title{Returns query execution runtime statistics related to a single execution
of a query if you have access to the workgroup in which the query ran}
\usage{
athena_get_query_runtime_statistics(QueryExecutionId)
}
\arguments{
\item{QueryExecutionId}{[required] The unique ID of the query execution.}
}
\description{
Returns query execution runtime statistics related to a single execution of a query if you have access to the workgroup in which the query ran. The query execution runtime statistics is returned only when QueryExecutionStatus$State is in a SUCCEEDED or FAILED state.

See \url{https://paws-r.github.io/docs/athena/get_query_runtime_statistics.html} for full documentation.
}
\keyword{internal}
