% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_operations.R
\name{athena_get_table_metadata}
\alias{athena_get_table_metadata}
\title{Returns table metadata for the specified catalog, database, and table}
\usage{
athena_get_table_metadata(
  CatalogName,
  DatabaseName,
  TableName,
  WorkGroup = NULL
)
}
\arguments{
\item{CatalogName}{[required] The name of the data catalog that contains the database and table
metadata to return.}

\item{DatabaseName}{[required] The name of the database that contains the table metadata to return.}

\item{TableName}{[required] The name of the table for which metadata is returned.}

\item{WorkGroup}{The name of the workgroup for which the metadata is being fetched.
Required if requesting an IAM Identity Center enabled Glue Data Catalog.}
}
\description{
Returns table metadata for the specified catalog, database, and table.

See \url{https://www.paws-r-sdk.com/docs/athena_get_table_metadata/} for full documentation.
}
\keyword{internal}
