% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_operations.R
\name{athena_list_engine_versions}
\alias{athena_list_engine_versions}
\title{Returns a list of engine versions that are available to choose from,
including the Auto option}
\usage{
athena_list_engine_versions(NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{NextToken}{A token generated by the Athena service that specifies where to continue
pagination if a previous request was truncated. To obtain the next set
of pages, pass in the \code{NextToken} from the response object of the
previous page call.}

\item{MaxResults}{The maximum number of engine versions to return in this request.}
}
\description{
Returns a list of engine versions that are available to choose from, including the Auto option.

See \url{https://www.paws-r-sdk.com/docs/athena_list_engine_versions/} for full documentation.
}
\keyword{internal}
