% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_operations.R
\name{athena_list_named_queries}
\alias{athena_list_named_queries}
\title{Provides a list of available query IDs only for queries saved in the
specified workgroup}
\usage{
athena_list_named_queries(
  NextToken = NULL,
  MaxResults = NULL,
  WorkGroup = NULL
)
}
\arguments{
\item{NextToken}{A token generated by the Athena service that specifies where to continue
pagination if a previous request was truncated. To obtain the next set
of pages, pass in the \code{NextToken} from the response object of the
previous page call.}

\item{MaxResults}{The maximum number of queries to return in this request.}

\item{WorkGroup}{The name of the workgroup from which the named queries are being
returned. If a workgroup is not specified, the saved queries for the
primary workgroup are returned.}
}
\description{
Provides a list of available query IDs only for queries saved in the specified workgroup. Requires that you have access to the specified workgroup. If a workgroup is not specified, lists the saved queries for the primary workgroup.

See \url{https://paws-r.github.io/docs/athena/list_named_queries.html} for full documentation.
}
\keyword{internal}
