% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_operations.R
\name{athena_list_prepared_statements}
\alias{athena_list_prepared_statements}
\title{Lists the prepared statements in the specified workgroup}
\usage{
athena_list_prepared_statements(WorkGroup, NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{WorkGroup}{[required] The workgroup to list the prepared statements for.}

\item{NextToken}{A token generated by the Athena service that specifies where to continue
pagination if a previous request was truncated. To obtain the next set
of pages, pass in the \code{NextToken} from the response object of the
previous page call.}

\item{MaxResults}{The maximum number of results to return in this request.}
}
\description{
Lists the prepared statements in the specified workgroup.

See \url{https://paws-r.github.io/docs/athena/list_prepared_statements.html} for full documentation.
}
\keyword{internal}
