% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_operations.R
\name{athena_list_work_groups}
\alias{athena_list_work_groups}
\title{Lists available workgroups for the account}
\usage{
athena_list_work_groups(NextToken, MaxResults)
}
\arguments{
\item{NextToken}{A token generated by the Athena service that specifies where to continue
pagination if a previous request was truncated. To obtain the next set
of pages, pass in the \code{NextToken} from the response object of the
previous page call.}

\item{MaxResults}{The maximum number of workgroups to return in this request.}
}
\value{
A list with the following syntax:\preformatted{list(
  WorkGroups = list(
    list(
      Name = "string",
      State = "ENABLED"|"DISABLED",
      Description = "string",
      CreationTime = as.POSIXct(
        "2015-01-01"
      ),
      EngineVersion = list(
        SelectedEngineVersion = "string",
        EffectiveEngineVersion = "string"
      )
    )
  ),
  NextToken = "string"
)
}
}
\description{
Lists available workgroups for the account.
}
\section{Request syntax}{
\preformatted{svc$list_work_groups(
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
