% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_operations.R
\name{athena_list_work_groups}
\alias{athena_list_work_groups}
\title{Lists available workgroups for the account}
\usage{
athena_list_work_groups(NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{NextToken}{A token generated by the Athena service that specifies where to continue
pagination if a previous request was truncated. To obtain the next set
of pages, pass in the \code{NextToken} from the response object of the
previous page call.}

\item{MaxResults}{The maximum number of workgroups to return in this request.}
}
\description{
Lists available workgroups for the account.

See \url{https://www.paws-r-sdk.com/docs/athena_list_work_groups/} for full documentation.
}
\keyword{internal}
