% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_operations.R
\name{athena_stop_calculation_execution}
\alias{athena_stop_calculation_execution}
\title{Requests the cancellation of a calculation}
\usage{
athena_stop_calculation_execution(CalculationExecutionId)
}
\arguments{
\item{CalculationExecutionId}{[required] The calculation execution UUID.}
}
\description{
Requests the cancellation of a calculation. A \code{\link[=athena_stop_calculation_execution]{stop_calculation_execution}} call on a calculation that is already in a terminal state (for example, \code{STOPPED}, \code{FAILED}, or \code{COMPLETED}) succeeds but has no effect.

See \url{https://paws-r.github.io/docs/athena/stop_calculation_execution.html} for full documentation.
}
\keyword{internal}
