% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_operations.R
\name{athena_untag_resource}
\alias{athena_untag_resource}
\title{Removes one or more tags from an Athena resource}
\usage{
athena_untag_resource(ResourceARN, TagKeys)
}
\arguments{
\item{ResourceARN}{[required] Specifies the ARN of the resource from which tags are to be removed.}

\item{TagKeys}{[required] A comma-separated list of one or more tag keys whose tags are to be
removed from the specified resource.}
}
\description{
Removes one or more tags from an Athena resource.

See \url{https://www.paws-r-sdk.com/docs/athena_untag_resource/} for full documentation.
}
\keyword{internal}
