% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_operations.R
\name{athena_update_work_group}
\alias{athena_update_work_group}
\title{Updates the workgroup with the specified name}
\usage{
athena_update_work_group(
  WorkGroup,
  Description = NULL,
  ConfigurationUpdates = NULL,
  State = NULL
)
}
\arguments{
\item{WorkGroup}{[required] The specified workgroup that will be updated.}

\item{Description}{The workgroup description.}

\item{ConfigurationUpdates}{Contains configuration updates for an Athena SQL workgroup.}

\item{State}{The workgroup state that will be updated for the given workgroup.}
}
\description{
Updates the workgroup with the specified name. The workgroup's name cannot be changed. Only \code{ConfigurationUpdates} can be specified.

See \url{https://www.paws-r-sdk.com/docs/athena_update_work_group/} for full documentation.
}
\keyword{internal}
