% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudsearch_operations.R
\name{cloudsearch_build_suggesters}
\alias{cloudsearch_build_suggesters}
\title{Indexes the search suggestions}
\usage{
cloudsearch_build_suggesters(DomainName)
}
\arguments{
\item{DomainName}{[required]}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  FieldNames = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Indexes the search suggestions. For more information, see \href{https://docs.aws.amazon.com/cloudsearch/latest/developerguide/getting-suggestions.html#configuring-suggesters}{Configuring Suggesters}
in the \emph{Amazon CloudSearch Developer Guide}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$build_suggesters(
  DomainName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
