% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudsearch_operations.R
\name{cloudsearch_build_suggesters}
\alias{cloudsearch_build_suggesters}
\title{Indexes the search suggestions}
\usage{
cloudsearch_build_suggesters(DomainName)
}
\arguments{
\item{DomainName}{[required]}
}
\description{
Indexes the search suggestions. For more information, see \href{http://docs.aws.amazon.com/cloudsearch/latest/developerguide/getting-suggestions.html#configuring-suggesters}{Configuring Suggesters}
in the \emph{Amazon CloudSearch Developer Guide}.
}
\section{Request syntax}{
\preformatted{svc$build_suggesters(
  DomainName = "string"
)
}
}

\keyword{internal}
