% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudsearch_operations.R
\name{cloudsearch_create_domain}
\alias{cloudsearch_create_domain}
\title{Creates a new search domain}
\usage{
cloudsearch_create_domain(DomainName)
}
\arguments{
\item{DomainName}{[required] A name for the domain you are creating. Allowed characters are a-z
(lower-case letters), 0-9, and hyphen (-). Domain names must start with
a letter or number and be at least 3 and no more than 28 characters
long.}
}
\description{
Creates a new search domain. For more information, see <a href="https://docs.aws.amazon.com/cloudsearch/latest/developerguide/creating-domains.html" target="_blank">Creating a Search Domain</a> in the \emph{Amazon CloudSearch Developer Guide}.

See \url{https://paws-r.github.io/docs/cloudsearch/create_domain.html} for full documentation.
}
\keyword{internal}
