% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudsearch_operations.R
\name{cloudsearch_define_analysis_scheme}
\alias{cloudsearch_define_analysis_scheme}
\title{Configures an analysis scheme that can be applied to a text or
text-array field to define language-specific text processing options}
\usage{
cloudsearch_define_analysis_scheme(DomainName, AnalysisScheme)
}
\arguments{
\item{DomainName}{[required]}

\item{AnalysisScheme}{[required]}
}
\description{
Configures an analysis scheme that can be applied to a \code{text} or \code{text-array} field to define language-specific text processing options. For more information, see \if{html}{\out{<a href="https://docs.aws.amazon.com/cloudsearch/latest/developerguide/configuring-analysis-schemes.html" target="_blank">}}Configuring Analysis Schemes\if{html}{\out{</a>}} in the \emph{Amazon CloudSearch Developer Guide}.

See \url{https://paws-r.github.io/docs/cloudsearch/define_analysis_scheme.html} for full documentation.
}
\keyword{internal}
