% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudsearch_operations.R
\name{cloudsearch_define_suggester}
\alias{cloudsearch_define_suggester}
\title{Configures a suggester for a domain}
\usage{
cloudsearch_define_suggester(DomainName, Suggester)
}
\arguments{
\item{DomainName}{[required]}

\item{Suggester}{[required]}
}
\description{
Configures a suggester for a domain. A suggester enables you to display possible matches before users finish typing their queries. When you configure a suggester, you must specify the name of the text field you want to search for possible matches and a unique name for the suggester. For more information, see \if{html}{\out{<a href="https://docs.aws.amazon.com/cloudsearch/latest/developerguide/getting-suggestions.html" target="_blank">}}Getting Search Suggestions\if{html}{\out{</a>}} in the \emph{Amazon CloudSearch Developer Guide}.

See \url{https://www.paws-r-sdk.com/docs/cloudsearch_define_suggester/} for full documentation.
}
\keyword{internal}
