% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudsearch_operations.R
\name{cloudsearch_delete_analysis_scheme}
\alias{cloudsearch_delete_analysis_scheme}
\title{Deletes an analysis scheme}
\usage{
cloudsearch_delete_analysis_scheme(DomainName, AnalysisSchemeName)
}
\arguments{
\item{DomainName}{[required]}

\item{AnalysisSchemeName}{[required] The name of the analysis scheme you want to delete.}
}
\description{
Deletes an analysis scheme. For more information, see \href{https://docs.aws.amazon.com/cloudsearch/latest/developerguide/configuring-analysis-schemes.html}{Configuring Analysis Schemes}
in the \emph{Amazon CloudSearch Developer Guide}.
}
\section{Request syntax}{
\preformatted{svc$delete_analysis_scheme(
  DomainName = "string",
  AnalysisSchemeName = "string"
)
}
}

\keyword{internal}
