% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudsearch_operations.R
\name{cloudsearch_delete_domain}
\alias{cloudsearch_delete_domain}
\title{Permanently deletes a search domain and all of its data}
\usage{
cloudsearch_delete_domain(DomainName)
}
\arguments{
\item{DomainName}{[required] The name of the domain you want to permanently delete.}
}
\description{
Permanently deletes a search domain and all of its data. Once a domain has been deleted, it cannot be recovered. For more information, see \if{html}{\out{<a href="https://docs.aws.amazon.com/cloudsearch/latest/developerguide/deleting-domains.html" target="_blank">}}Deleting a Search Domain\if{html}{\out{</a>}} in the \emph{Amazon CloudSearch Developer Guide}.

See \url{https://www.paws-r-sdk.com/docs/cloudsearch_delete_domain/} for full documentation.
}
\keyword{internal}
