% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudsearch_operations.R
\name{cloudsearch_delete_index_field}
\alias{cloudsearch_delete_index_field}
\title{Removes an IndexField from the search domain}
\usage{
cloudsearch_delete_index_field(DomainName, IndexFieldName)
}
\arguments{
\item{DomainName}{[required]}

\item{IndexFieldName}{[required] The name of the index field your want to remove from the domain's
indexing options.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  IndexField = list(
    Options = list(
      IndexFieldName = "string",
      IndexFieldType = "int"|"double"|"literal"|"text"|"date"|"latlon"|"int-array"|"double-array"|"literal-array"|"text-array"|"date-array",
      IntOptions = list(
        DefaultValue = 123,
        SourceField = "string",
        FacetEnabled = TRUE|FALSE,
        SearchEnabled = TRUE|FALSE,
        ReturnEnabled = TRUE|FALSE,
        SortEnabled = TRUE|FALSE
      ),
      DoubleOptions = list(
        DefaultValue = 123.0,
        SourceField = "string",
        FacetEnabled = TRUE|FALSE,
        SearchEnabled = TRUE|FALSE,
        ReturnEnabled = TRUE|FALSE,
        SortEnabled = TRUE|FALSE
      ),
      LiteralOptions = list(
        DefaultValue = "string",
        SourceField = "string",
        FacetEnabled = TRUE|FALSE,
        SearchEnabled = TRUE|FALSE,
        ReturnEnabled = TRUE|FALSE,
        SortEnabled = TRUE|FALSE
      ),
      TextOptions = list(
        DefaultValue = "string",
        SourceField = "string",
        ReturnEnabled = TRUE|FALSE,
        SortEnabled = TRUE|FALSE,
        HighlightEnabled = TRUE|FALSE,
        AnalysisScheme = "string"
      ),
      DateOptions = list(
        DefaultValue = "string",
        SourceField = "string",
        FacetEnabled = TRUE|FALSE,
        SearchEnabled = TRUE|FALSE,
        ReturnEnabled = TRUE|FALSE,
        SortEnabled = TRUE|FALSE
      ),
      LatLonOptions = list(
        DefaultValue = "string",
        SourceField = "string",
        FacetEnabled = TRUE|FALSE,
        SearchEnabled = TRUE|FALSE,
        ReturnEnabled = TRUE|FALSE,
        SortEnabled = TRUE|FALSE
      ),
      IntArrayOptions = list(
        DefaultValue = 123,
        SourceFields = "string",
        FacetEnabled = TRUE|FALSE,
        SearchEnabled = TRUE|FALSE,
        ReturnEnabled = TRUE|FALSE
      ),
      DoubleArrayOptions = list(
        DefaultValue = 123.0,
        SourceFields = "string",
        FacetEnabled = TRUE|FALSE,
        SearchEnabled = TRUE|FALSE,
        ReturnEnabled = TRUE|FALSE
      ),
      LiteralArrayOptions = list(
        DefaultValue = "string",
        SourceFields = "string",
        FacetEnabled = TRUE|FALSE,
        SearchEnabled = TRUE|FALSE,
        ReturnEnabled = TRUE|FALSE
      ),
      TextArrayOptions = list(
        DefaultValue = "string",
        SourceFields = "string",
        ReturnEnabled = TRUE|FALSE,
        HighlightEnabled = TRUE|FALSE,
        AnalysisScheme = "string"
      ),
      DateArrayOptions = list(
        DefaultValue = "string",
        SourceFields = "string",
        FacetEnabled = TRUE|FALSE,
        SearchEnabled = TRUE|FALSE,
        ReturnEnabled = TRUE|FALSE
      )
    ),
    Status = list(
      CreationDate = as.POSIXct(
        "2015-01-01"
      ),
      UpdateDate = as.POSIXct(
        "2015-01-01"
      ),
      UpdateVersion = 123,
      State = "RequiresIndexDocuments"|"Processing"|"Active"|"FailedToValidate",
      PendingDeletion = TRUE|FALSE
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Removes an \code{IndexField} from the search domain. For more information,
see <a
href="https://docs.aws.amazon.com/cloudsearch/latest/developerguide/configuring-index-fields.html"
target="_blank">Configuring Index Fields</a> in the \emph{Amazon CloudSearch
Developer Guide}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_index_field(
  DomainName = "string",
  IndexFieldName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
