% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudsearch_operations.R
\name{cloudsearch_delete_index_field}
\alias{cloudsearch_delete_index_field}
\title{Removes an IndexField from the search domain}
\usage{
cloudsearch_delete_index_field(DomainName, IndexFieldName)
}
\arguments{
\item{DomainName}{[required]}

\item{IndexFieldName}{[required] The name of the index field your want to remove from the domain's
indexing options.}
}
\description{
Removes an \verb{<a>IndexField</a>} from the search domain. For more
information, see \href{http://docs.aws.amazon.com/cloudsearch/latest/developerguide/configuring-index-fields.html}{Configuring Index Fields}
in the \emph{Amazon CloudSearch Developer Guide}.
}
\section{Request syntax}{
\preformatted{svc$delete_index_field(
  DomainName = "string",
  IndexFieldName = "string"
)
}
}

\keyword{internal}
