% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudsearch_operations.R
\name{cloudsearch_index_documents}
\alias{cloudsearch_index_documents}
\title{Tells the search domain to start indexing its documents using the latest
indexing options}
\usage{
cloudsearch_index_documents(DomainName)
}
\arguments{
\item{DomainName}{[required]}
}
\description{
Tells the search domain to start indexing its documents using the latest indexing options. This operation must be invoked to activate options whose OptionStatus is \code{RequiresIndexDocuments}.

See \url{https://paws-r.github.io/docs/cloudsearch/index_documents.html} for full documentation.
}
\keyword{internal}
