% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudsearch_operations.R
\name{cloudsearch_index_documents}
\alias{cloudsearch_index_documents}
\title{Tells the search domain to start indexing its documents using the latest
indexing options}
\usage{
cloudsearch_index_documents(DomainName)
}
\arguments{
\item{DomainName}{[required]}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  FieldNames = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Tells the search domain to start indexing its documents using the latest
indexing options. This operation must be invoked to activate options
whose OptionStatus is \code{RequiresIndexDocuments}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$index_documents(
  DomainName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
