% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudsearch_operations.R
\name{cloudsearch_list_domain_names}
\alias{cloudsearch_list_domain_names}
\title{Lists all search domains owned by an account}
\usage{
cloudsearch_list_domain_names()
}
\description{
Lists all search domains owned by an account.

See \url{https://paws-r.github.io/docs/cloudsearch/list_domain_names.html} for full documentation.
}
\keyword{internal}
