% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudsearch_operations.R
\name{cloudsearch_update_service_access_policies}
\alias{cloudsearch_update_service_access_policies}
\title{Configures the access rules that control access to the domain's document
and search endpoints}
\usage{
cloudsearch_update_service_access_policies(DomainName, AccessPolicies)
}
\arguments{
\item{DomainName}{[required]}

\item{AccessPolicies}{[required] The access rules you want to configure. These rules replace any existing
rules.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  AccessPolicies = list(
    Options = "string",
    Status = list(
      CreationDate = as.POSIXct(
        "2015-01-01"
      ),
      UpdateDate = as.POSIXct(
        "2015-01-01"
      ),
      UpdateVersion = 123,
      State = "RequiresIndexDocuments"|"Processing"|"Active"|"FailedToValidate",
      PendingDeletion = TRUE|FALSE
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Configures the access rules that control access to the domain's document
and search endpoints. For more information, see <a
href="https://docs.aws.amazon.com/cloudsearch/latest/developerguide/configuring-access.html"
target="_blank">Configuring Access for an Amazon CloudSearch Domain</a>.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_service_access_policies(
  DomainName = "string",
  AccessPolicies = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
