% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudsearchdomain_operations.R
\name{cloudsearchdomain_suggest}
\alias{cloudsearchdomain_suggest}
\title{Retrieves autocomplete suggestions for a partial query string}
\usage{
cloudsearchdomain_suggest(query, suggester, size = NULL)
}
\arguments{
\item{query}{[required] Specifies the string for which you want to get suggestions.}

\item{suggester}{[required] Specifies the name of the suggester to use to find suggested matches.}

\item{size}{Specifies the maximum number of suggestions to return.}
}
\description{
Retrieves autocomplete suggestions for a partial query string. You can use suggestions enable you to display likely matches before users finish typing. In Amazon CloudSearch, suggestions are based on the contents of a particular text field. When you request suggestions, Amazon CloudSearch finds all of the documents whose values in the suggester field start with the specified query string. The beginning of the field must match the query string to be considered a match.

See \url{https://www.paws-r-sdk.com/docs/cloudsearchdomain_suggest/} for full documentation.
}
\keyword{internal}
