% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapipeline_operations.R
\name{datapipeline_activate_pipeline}
\alias{datapipeline_activate_pipeline}
\title{Validates the specified pipeline and starts processing pipeline tasks}
\usage{
datapipeline_activate_pipeline(
  pipelineId,
  parameterValues = NULL,
  startTimestamp = NULL
)
}
\arguments{
\item{pipelineId}{[required] The ID of the pipeline.}

\item{parameterValues}{A list of parameter values to pass to the pipeline at activation.}

\item{startTimestamp}{The date and time to resume the pipeline. By default, the pipeline
resumes from the last completed execution.}
}
\description{
Validates the specified pipeline and starts processing pipeline tasks. If the pipeline does not pass validation, activation fails.

See \url{https://www.paws-r-sdk.com/docs/datapipeline_activate_pipeline/} for full documentation.
}
\keyword{internal}
