% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapipeline_operations.R
\name{datapipeline_evaluate_expression}
\alias{datapipeline_evaluate_expression}
\title{Task runners call EvaluateExpression to evaluate a string in the context
of the specified object}
\usage{
datapipeline_evaluate_expression(pipelineId, objectId, expression)
}
\arguments{
\item{pipelineId}{[required] The ID of the pipeline.}

\item{objectId}{[required] The ID of the object.}

\item{expression}{[required] The expression to evaluate.}
}
\value{
A list with the following syntax:\preformatted{list(
  evaluatedExpression = "string"
)
}
}
\description{
Task runners call
\code{\link[=datapipeline_evaluate_expression]{evaluate_expression}} to evaluate a
string in the context of the specified object. For example, a task
runner can evaluate SQL queries stored in Amazon S3.
}
\section{Request syntax}{
\preformatted{svc$evaluate_expression(
  pipelineId = "string",
  objectId = "string",
  expression = "string"
)
}
}

\keyword{internal}
