% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapipeline_operations.R
\name{datapipeline_put_pipeline_definition}
\alias{datapipeline_put_pipeline_definition}
\title{Adds tasks, schedules, and preconditions to the specified pipeline}
\usage{
datapipeline_put_pipeline_definition(
  pipelineId,
  pipelineObjects,
  parameterObjects = NULL,
  parameterValues = NULL
)
}
\arguments{
\item{pipelineId}{[required] The ID of the pipeline.}

\item{pipelineObjects}{[required] The objects that define the pipeline. These objects overwrite the
existing pipeline definition.}

\item{parameterObjects}{The parameter objects used with the pipeline.}

\item{parameterValues}{The parameter values used with the pipeline.}
}
\description{
Adds tasks, schedules, and preconditions to the specified pipeline. You can use \code{\link[=datapipeline_put_pipeline_definition]{put_pipeline_definition}} to populate a new pipeline.

See \url{https://paws-r.github.io/docs/datapipeline/put_pipeline_definition.html} for full documentation.
}
\keyword{internal}
