% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_create_account_pool}
\alias{datazone_create_account_pool}
\title{Creates an account pool}
\usage{
datazone_create_account_pool(
  accountSource,
  description = NULL,
  domainIdentifier,
  name,
  resolutionStrategy
)
}
\arguments{
\item{accountSource}{[required] The source of accounts for the account pool. In the current release,
it's either a static list of accounts provided by the customer or a
custom Amazon Web Services Lambda handler.}

\item{description}{The description of the account pool.}

\item{domainIdentifier}{[required] The ID of the domain where the account pool is created.}

\item{name}{[required] The name of the account pool.}

\item{resolutionStrategy}{[required] The mechanism used to resolve the account selection from the account
pool.}
}
\description{
Creates an account pool.

See \url{https://www.paws-r-sdk.com/docs/datazone_create_account_pool/} for full documentation.
}
\keyword{internal}
