% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_create_asset_filter}
\alias{datazone_create_asset_filter}
\title{Creates a data asset filter}
\usage{
datazone_create_asset_filter(
  assetIdentifier,
  clientToken = NULL,
  configuration,
  description = NULL,
  domainIdentifier,
  name
)
}
\arguments{
\item{assetIdentifier}{[required] The ID of the data asset.}

\item{clientToken}{A unique, case-sensitive identifier that is provided to ensure the
idempotency of the request.}

\item{configuration}{[required] The configuration of the asset filter.}

\item{description}{The description of the asset filter.}

\item{domainIdentifier}{[required] The ID of the domain in which you want to create an asset filter.}

\item{name}{[required] The name of the asset filter.}
}
\description{
Creates a data asset filter.

See \url{https://www.paws-r-sdk.com/docs/datazone_create_asset_filter/} for full documentation.
}
\keyword{internal}
