% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_create_data_product}
\alias{datazone_create_data_product}
\title{Creates a data product}
\usage{
datazone_create_data_product(
  clientToken = NULL,
  description = NULL,
  domainIdentifier,
  formsInput = NULL,
  glossaryTerms = NULL,
  items = NULL,
  name,
  owningProjectIdentifier
)
}
\arguments{
\item{clientToken}{A unique, case-sensitive identifier that is provided to ensure the
idempotency of the request.}

\item{description}{The description of the data product.}

\item{domainIdentifier}{[required] The ID of the domain where the data product is created.}

\item{formsInput}{The metadata forms of the data product.}

\item{glossaryTerms}{The glossary terms of the data product.}

\item{items}{The data assets of the data product.}

\item{name}{[required] The name of the data product.}

\item{owningProjectIdentifier}{[required] The ID of the owning project of the data product.}
}
\description{
Creates a data product.

See \url{https://www.paws-r-sdk.com/docs/datazone_create_data_product/} for full documentation.
}
\keyword{internal}
