% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_create_environment_blueprint}
\alias{datazone_create_environment_blueprint}
\title{Creates a Amazon DataZone blueprint}
\usage{
datazone_create_environment_blueprint(
  description = NULL,
  domainIdentifier,
  name,
  provisioningProperties,
  userParameters = NULL
)
}
\arguments{
\item{description}{The description of the Amazon DataZone blueprint.}

\item{domainIdentifier}{[required] The identifier of the domain in which this blueprint is created.}

\item{name}{[required] The name of this Amazon DataZone blueprint.}

\item{provisioningProperties}{[required] The provisioning properties of this Amazon DataZone blueprint.}

\item{userParameters}{The user parameters of this Amazon DataZone blueprint.}
}
\description{
Creates a Amazon DataZone blueprint.

See \url{https://www.paws-r-sdk.com/docs/datazone_create_environment_blueprint/} for full documentation.
}
\keyword{internal}
