% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_create_glossary_term}
\alias{datazone_create_glossary_term}
\title{Creates a business glossary term}
\usage{
datazone_create_glossary_term(
  clientToken = NULL,
  domainIdentifier,
  glossaryIdentifier,
  longDescription = NULL,
  name,
  shortDescription = NULL,
  status = NULL,
  termRelations = NULL
)
}
\arguments{
\item{clientToken}{A unique, case-sensitive identifier that is provided to ensure the
idempotency of the request.}

\item{domainIdentifier}{[required] The ID of the Amazon DataZone domain in which this business glossary
term is created.}

\item{glossaryIdentifier}{[required] The ID of the business glossary in which this term is created.}

\item{longDescription}{The long description of this business glossary term.}

\item{name}{[required] The name of this business glossary term.}

\item{shortDescription}{The short description of this business glossary term.}

\item{status}{The status of this business glossary term.}

\item{termRelations}{The term relations of this business glossary term.}
}
\description{
Creates a business glossary term.

See \url{https://www.paws-r-sdk.com/docs/datazone_create_glossary_term/} for full documentation.
}
\keyword{internal}
