% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_create_group_profile}
\alias{datazone_create_group_profile}
\title{Creates a group profile in Amazon DataZone}
\usage{
datazone_create_group_profile(
  clientToken = NULL,
  domainIdentifier,
  groupIdentifier
)
}
\arguments{
\item{clientToken}{A unique, case-sensitive identifier that is provided to ensure the
idempotency of the request.}

\item{domainIdentifier}{[required] The identifier of the Amazon DataZone domain in which the group profile
is created.}

\item{groupIdentifier}{[required] The identifier of the group for which the group profile is created.}
}
\description{
Creates a group profile in Amazon DataZone.

See \url{https://www.paws-r-sdk.com/docs/datazone_create_group_profile/} for full documentation.
}
\keyword{internal}
