% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_create_listing_change_set}
\alias{datazone_create_listing_change_set}
\title{Create listing change set}
\usage{
datazone_create_listing_change_set(
  action,
  clientToken = NULL,
  domainIdentifier,
  entityIdentifier,
  entityRevision = NULL,
  entityType
)
}
\arguments{
\item{action}{[required]}

\item{clientToken}{}

\item{domainIdentifier}{[required]}

\item{entityIdentifier}{[required]}

\item{entityRevision}{}

\item{entityType}{[required]}
}
\description{
Create listing change set

See \url{https://www.paws-r-sdk.com/docs/datazone_create_listing_change_set/} for full documentation.
}
\keyword{internal}
