% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_create_project}
\alias{datazone_create_project}
\title{Creates an Amazon DataZone project}
\usage{
datazone_create_project(
  description = NULL,
  domainIdentifier,
  glossaryTerms = NULL,
  name
)
}
\arguments{
\item{description}{The description of the Amazon DataZone project.}

\item{domainIdentifier}{[required] The ID of the Amazon DataZone domain in which this project is created.}

\item{glossaryTerms}{The glossary terms that can be used in this Amazon DataZone project.}

\item{name}{[required] The name of the Amazon DataZone project.}
}
\description{
Creates an Amazon DataZone project.

See \url{https://www.paws-r-sdk.com/docs/datazone_create_project/} for full documentation.
}
\keyword{internal}
