% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_create_project_membership}
\alias{datazone_create_project_membership}
\title{Creates a project membership in Amazon DataZone}
\usage{
datazone_create_project_membership(
  designation,
  domainIdentifier,
  member,
  projectIdentifier
)
}
\arguments{
\item{designation}{[required] The designation of the project membership.}

\item{domainIdentifier}{[required] The ID of the Amazon DataZone domain in which project membership is
created.}

\item{member}{[required] The project member whose project membership was created.}

\item{projectIdentifier}{[required] The ID of the project for which this project membership was created.}
}
\description{
Creates a project membership in Amazon DataZone.

See \url{https://www.paws-r-sdk.com/docs/datazone_create_project_membership/} for full documentation.
}
\keyword{internal}
