% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_create_project_profile}
\alias{datazone_create_project_profile}
\title{Creates a project profile}
\usage{
datazone_create_project_profile(
  allowCustomProjectResourceTags = NULL,
  description = NULL,
  domainIdentifier,
  domainUnitIdentifier = NULL,
  environmentConfigurations = NULL,
  name,
  projectResourceTags = NULL,
  projectResourceTagsDescription = NULL,
  status = NULL
)
}
\arguments{
\item{allowCustomProjectResourceTags}{Specifies whether custom project resource tags are supported.}

\item{description}{A description of a project profile.}

\item{domainIdentifier}{[required] A domain ID of the project profile.}

\item{domainUnitIdentifier}{A domain unit ID of the project profile.}

\item{environmentConfigurations}{Environment configurations of the project profile.}

\item{name}{[required] Project profile name.}

\item{projectResourceTags}{The resource tags of the project profile.}

\item{projectResourceTagsDescription}{Field viewable through the UI that provides a project user with the
allowed resource tag specifications.}

\item{status}{Project profile status.}
}
\description{
Creates a project profile.

See \url{https://www.paws-r-sdk.com/docs/datazone_create_project_profile/} for full documentation.
}
\keyword{internal}
