% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_delete_domain}
\alias{datazone_delete_domain}
\title{Deletes a Amazon DataZone domain}
\usage{
datazone_delete_domain(
  clientToken = NULL,
  identifier,
  skipDeletionCheck = NULL
)
}
\arguments{
\item{clientToken}{A unique, case-sensitive identifier that is provided to ensure the
idempotency of the request.}

\item{identifier}{[required] The identifier of the Amazon Web Services domain that is to be deleted.}

\item{skipDeletionCheck}{Specifies the optional flag to delete all child entities within the
domain.}
}
\description{
Deletes a Amazon DataZone domain.

See \url{https://www.paws-r-sdk.com/docs/datazone_delete_domain/} for full documentation.
}
\keyword{internal}
