% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_delete_listing}
\alias{datazone_delete_listing}
\title{Deletes a listing (a record of an asset at a given time)}
\usage{
datazone_delete_listing(domainIdentifier, identifier)
}
\arguments{
\item{domainIdentifier}{[required] The ID of the Amazon DataZone domain.}

\item{identifier}{[required] The ID of the listing to be deleted.}
}
\description{
Deletes a listing (a record of an asset at a given time).

See \url{https://www.paws-r-sdk.com/docs/datazone_delete_listing/} for full documentation.
}
\keyword{internal}
