% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_get_asset_filter}
\alias{datazone_get_asset_filter}
\title{Gets an asset filter}
\usage{
datazone_get_asset_filter(assetIdentifier, domainIdentifier, identifier)
}
\arguments{
\item{assetIdentifier}{[required] The ID of the data asset.}

\item{domainIdentifier}{[required] The ID of the domain where you want to get an asset filter.}

\item{identifier}{[required] The ID of the asset filter.}
}
\description{
Gets an asset filter.

See \url{https://www.paws-r-sdk.com/docs/datazone_get_asset_filter/} for full documentation.
}
\keyword{internal}
