% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_get_notebook_run}
\alias{datazone_get_notebook_run}
\title{Gets the details of a notebook run in an Amazon DataZone domain}
\usage{
datazone_get_notebook_run(domainIdentifier, identifier)
}
\arguments{
\item{domainIdentifier}{[required] The identifier of the Amazon DataZone domain in which the notebook run
exists.}

\item{identifier}{[required] The identifier of the notebook run.}
}
\description{
Gets the details of a notebook run in an Amazon DataZone domain.

See \url{https://www.paws-r-sdk.com/docs/datazone_get_notebook_run/} for full documentation.
}
\keyword{internal}
