% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_get_user_profile}
\alias{datazone_get_user_profile}
\title{Gets a user profile in Amazon DataZone}
\usage{
datazone_get_user_profile(domainIdentifier, type = NULL, userIdentifier)
}
\arguments{
\item{domainIdentifier}{[required] the ID of the Amazon DataZone domain the data portal of which you want
to get.}

\item{type}{The type of the user profile.}

\item{userIdentifier}{[required] The identifier of the user for which you want to get the user profile.}
}
\description{
Gets a user profile in Amazon DataZone.

See \url{https://www.paws-r-sdk.com/docs/datazone_get_user_profile/} for full documentation.
}
\keyword{internal}
