% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_start_data_source_run}
\alias{datazone_start_data_source_run}
\title{Start the run of the specified data source in Amazon DataZone}
\usage{
datazone_start_data_source_run(
  clientToken = NULL,
  dataSourceIdentifier,
  domainIdentifier
)
}
\arguments{
\item{clientToken}{A unique, case-sensitive identifier that is provided to ensure the
idempotency of the request.}

\item{dataSourceIdentifier}{[required] The identifier of the data source.}

\item{domainIdentifier}{[required] The identifier of the Amazon DataZone domain in which to start a data
source run.}
}
\description{
Start the run of the specified data source in Amazon DataZone.

See \url{https://www.paws-r-sdk.com/docs/datazone_start_data_source_run/} for full documentation.
}
\keyword{internal}
