% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_update_data_source}
\alias{datazone_update_data_source}
\title{Updates the specified data source in Amazon DataZone}
\usage{
datazone_update_data_source(
  assetFormsInput = NULL,
  configuration = NULL,
  description = NULL,
  domainIdentifier,
  enableSetting = NULL,
  identifier,
  name = NULL,
  publishOnImport = NULL,
  recommendation = NULL,
  retainPermissionsOnRevokeFailure = NULL,
  schedule = NULL
)
}
\arguments{
\item{assetFormsInput}{The asset forms to be updated as part of the
\code{\link[=datazone_update_data_source]{update_data_source}} action.}

\item{configuration}{The configuration to be updated as part of the
\code{\link[=datazone_update_data_source]{update_data_source}} action.}

\item{description}{The description to be updated as part of the
\code{\link[=datazone_update_data_source]{update_data_source}} action.}

\item{domainIdentifier}{[required] The identifier of the domain in which to update a data source.}

\item{enableSetting}{The enable setting to be updated as part of the
\code{\link[=datazone_update_data_source]{update_data_source}} action.}

\item{identifier}{[required] The identifier of the data source to be updated.}

\item{name}{The name to be updated as part of the
\code{\link[=datazone_update_data_source]{update_data_source}} action.}

\item{publishOnImport}{The publish on import setting to be updated as part of the
\code{\link[=datazone_update_data_source]{update_data_source}} action.}

\item{recommendation}{The recommendation to be updated as part of the
\code{\link[=datazone_update_data_source]{update_data_source}} action.}

\item{retainPermissionsOnRevokeFailure}{Specifies that the granted permissions are retained in case of a
self-subscribe functionality failure for a data source.}

\item{schedule}{The schedule to be updated as part of the
\code{\link[=datazone_update_data_source]{update_data_source}} action.}
}
\description{
Updates the specified data source in Amazon DataZone.

See \url{https://www.paws-r-sdk.com/docs/datazone_update_data_source/} for full documentation.
}
\keyword{internal}
