% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_update_environment_blueprint}
\alias{datazone_update_environment_blueprint}
\title{Updates an environment blueprint in Amazon DataZone}
\usage{
datazone_update_environment_blueprint(
  description = NULL,
  domainIdentifier,
  identifier,
  provisioningProperties = NULL,
  userParameters = NULL
)
}
\arguments{
\item{description}{The description to be updated as part of the
\code{\link[=datazone_update_environment_blueprint]{update_environment_blueprint}}
action.}

\item{domainIdentifier}{[required] The identifier of the Amazon DataZone domain in which an environment
blueprint is to be updated.}

\item{identifier}{[required] The identifier of the environment blueprint to be updated.}

\item{provisioningProperties}{The provisioning properties to be updated as part of the
\code{\link[=datazone_update_environment_blueprint]{update_environment_blueprint}}
action.}

\item{userParameters}{The user parameters to be updated as part of the
\code{\link[=datazone_update_environment_blueprint]{update_environment_blueprint}}
action.}
}
\description{
Updates an environment blueprint in Amazon DataZone.

See \url{https://www.paws-r-sdk.com/docs/datazone_update_environment_blueprint/} for full documentation.
}
\keyword{internal}
