% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_update_notebook}
\alias{datazone_update_notebook}
\title{Updates a notebook in Amazon SageMaker Unified Studio}
\usage{
datazone_update_notebook(
  domainIdentifier,
  identifier,
  description = NULL,
  status = NULL,
  name = NULL,
  cellOrder = NULL,
  metadata = NULL,
  parameters = NULL,
  environmentConfiguration = NULL,
  clientToken = NULL
)
}
\arguments{
\item{domainIdentifier}{[required] The identifier of the Amazon SageMaker Unified Studio domain in which the notebook exists.}

\item{identifier}{[required] The identifier of the notebook to update.}

\item{description}{The updated description of the notebook.}

\item{status}{The updated status of the notebook.}

\item{name}{The updated name of the notebook.}

\item{cellOrder}{The updated ordered list of cells in the notebook.}

\item{metadata}{The updated metadata for the notebook, specified as key-value pairs.}

\item{parameters}{The updated sensitive parameters for the notebook, specified as key-value pairs.}

\item{environmentConfiguration}{The updated environment configuration for the notebook.}

\item{clientToken}{A unique, case-sensitive identifier to ensure idempotency of the request. This field is automatically populated if not provided.}
}
\description{
Updates a \href{https://docs.aws.amazon.com/sagemaker-unified-studio/latest/userguide/notebooks.html}{notebook} in Amazon SageMaker Unified Studio.

See \url{https://www.paws-r-sdk.com/docs/datazone_update_notebook/} for full documentation.
}
\keyword{internal}
