% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_update_project}
\alias{datazone_update_project}
\title{Updates the specified project in Amazon DataZone}
\usage{
datazone_update_project(
  description = NULL,
  domainIdentifier,
  glossaryTerms = NULL,
  identifier,
  name = NULL
)
}
\arguments{
\item{description}{The description to be updated as part of the
\code{\link[=datazone_update_project]{update_project}} action.}

\item{domainIdentifier}{[required] The ID of the Amazon DataZone domain where a project is being updated.}

\item{glossaryTerms}{The glossary terms to be updated as part of the
\code{\link[=datazone_update_project]{update_project}} action.}

\item{identifier}{[required] The identifier of the project that is to be updated.}

\item{name}{The name to be updated as part of the
\code{\link[=datazone_update_project]{update_project}} action.}
}
\description{
Updates the specified project in Amazon DataZone.

See \url{https://www.paws-r-sdk.com/docs/datazone_update_project/} for full documentation.
}
\keyword{internal}
