% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datazone_operations.R
\name{datazone_update_project_profile}
\alias{datazone_update_project_profile}
\title{Updates a project profile}
\usage{
datazone_update_project_profile(
  description = NULL,
  domainIdentifier,
  domainUnitIdentifier = NULL,
  environmentConfigurations = NULL,
  identifier,
  name = NULL,
  status = NULL
)
}
\arguments{
\item{description}{The description of a project profile.}

\item{domainIdentifier}{[required] The ID of the domain where a project profile is to be updated.}

\item{domainUnitIdentifier}{The ID of the domain unit where a project profile is to be updated.}

\item{environmentConfigurations}{The environment configurations of a project profile.}

\item{identifier}{[required] The ID of a project profile that is to be updated.}

\item{name}{The name of a project profile.}

\item{status}{The status of a project profile.}
}
\description{
Updates a project profile.

See \url{https://www.paws-r-sdk.com/docs/datazone_update_project_profile/} for full documentation.
}
\keyword{internal}
