% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticsearchservice_operations.R
\name{elasticsearchservice_add_tags}
\alias{elasticsearchservice_add_tags}
\title{Attaches tags to an existing Elasticsearch domain}
\usage{
elasticsearchservice_add_tags(ARN, TagList)
}
\arguments{
\item{ARN}{[required] Specify the \code{ARN} for which you want to add the tags.}

\item{TagList}{[required] List of \code{Tag} that need to be added for the Elasticsearch domain.}
}
\description{
Attaches tags to an existing Elasticsearch domain. Tags are a set of case-sensitive key value pairs. An Elasticsearch domain may have up to 10 tags. See \if{html}{\out{<a href="https://docs.aws.amazon.com/opensearch-service/latest/developerguide/?trkcampaign=aws-summit#es-managedomains-awsresorcetagging" target="_blank">}}Tagging Amazon Elasticsearch Service Domains for more information.\if{html}{\out{</a>}}

See \url{https://www.paws-r-sdk.com/docs/elasticsearchservice_add_tags/} for full documentation.
}
\keyword{internal}
