% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticsearchservice_operations.R
\name{elasticsearchservice_associate_package}
\alias{elasticsearchservice_associate_package}
\title{Associates a package with an Amazon ES domain}
\usage{
elasticsearchservice_associate_package(PackageID, DomainName)
}
\arguments{
\item{PackageID}{[required] Internal ID of the package that you want to associate with a domain. Use
\code{\link[=elasticsearchservice_describe_packages]{describe_packages}} to find
this value.}

\item{DomainName}{[required] Name of the domain that you want to associate the package with.}
}
\description{
Associates a package with an Amazon ES domain.

See \url{https://www.paws-r-sdk.com/docs/elasticsearchservice_associate_package/} for full documentation.
}
\keyword{internal}
