% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticsearchservice_operations.R
\name{elasticsearchservice_create_elasticsearch_domain}
\alias{elasticsearchservice_create_elasticsearch_domain}
\title{Creates a new Elasticsearch domain}
\usage{
elasticsearchservice_create_elasticsearch_domain(DomainName,
  ElasticsearchVersion, ElasticsearchClusterConfig, EBSOptions,
  AccessPolicies, SnapshotOptions, VPCOptions, CognitoOptions,
  EncryptionAtRestOptions, NodeToNodeEncryptionOptions, AdvancedOptions,
  LogPublishingOptions)
}
\arguments{
\item{DomainName}{[required] The name of the Elasticsearch domain that you are creating. Domain names
are unique across the domains owned by an account within an AWS region.
Domain names must start with a letter or number and can contain the
following characters: a-z (lowercase), 0-9, and - (hyphen).}

\item{ElasticsearchVersion}{String of format X.Y to specify version for the Elasticsearch domain eg.
"1.5" or "2.3". For more information, see \href{http://docs.aws.amazon.com/elasticsearch-service/latest/developerguide/es-createupdatedomains.html#es-createdomains}{Creating Elasticsearch Domains}
in the \emph{Amazon Elasticsearch Service Developer Guide}.}

\item{ElasticsearchClusterConfig}{Configuration options for an Elasticsearch domain. Specifies the
instance type and number of instances in the domain cluster.}

\item{EBSOptions}{Options to enable, disable and specify the type and size of EBS storage
volumes.}

\item{AccessPolicies}{IAM access policy as a JSON-formatted string.}

\item{SnapshotOptions}{Option to set time, in UTC format, of the daily automated snapshot.
Default value is 0 hours.}

\item{VPCOptions}{Options to specify the subnets and security groups for VPC endpoint. For
more information, see \href{http://docs.aws.amazon.com/elasticsearch-service/latest/developerguide/es-vpc.html#es-creating-vpc}{Creating a VPC}
in \emph{VPC Endpoints for Amazon Elasticsearch Service Domains}}

\item{CognitoOptions}{Options to specify the Cognito user and identity pools for Kibana
authentication. For more information, see \href{http://docs.aws.amazon.com/elasticsearch-service/latest/developerguide/es-cognito-auth.html}{Amazon Cognito Authentication for Kibana}.}

\item{EncryptionAtRestOptions}{Specifies the Encryption At Rest Options.}

\item{NodeToNodeEncryptionOptions}{Specifies the NodeToNodeEncryptionOptions.}

\item{AdvancedOptions}{Option to allow references to indices in an HTTP request body. Must be
\code{false} when configuring access to individual sub-resources. By default,
the value is \code{true}. See \href{http://docs.aws.amazon.com/elasticsearch-service/latest/developerguide/es-createupdatedomains.html#es-createdomain-configure-advanced-options}{Configuration Advanced Options}
for more information.}

\item{LogPublishingOptions}{Map of \code{LogType} and \code{LogPublishingOption}, each containing options to
publish a given type of Elasticsearch log.}
}
\description{
Creates a new Elasticsearch domain. For more information, see \href{http://docs.aws.amazon.com/elasticsearch-service/latest/developerguide/es-createupdatedomains.html#es-createdomains}{Creating Elasticsearch Domains}
in the \emph{Amazon Elasticsearch Service Developer Guide}.
}
\section{Request syntax}{
\preformatted{svc$create_elasticsearch_domain(
  DomainName = "string",
  ElasticsearchVersion = "string",
  ElasticsearchClusterConfig = list(
    InstanceType = "m3.medium.elasticsearch"|"m3.large.elasticsearch"|"m3.xlarge.elasticsearch"|"m3.2xlarge.elasticsearch"|"m4.large.elasticsearch"|"m4.xlarge.elasticsearch"|"m4.2xlarge.elasticsearch"|"m4.4xlarge.elasticsearch"|"m4.10xlarge.elasticsearch"|"t2.micro.elasticsearch"|"t2.small.elasticsearch"|"t2.medium.elasticsearch"|"r3.large.elasticsearch"|"r3.xlarge.elasticsearch"|"r3.2xlarge.elasticsearch"|"r3.4xlarge.elasticsearch"|"r3.8xlarge.elasticsearch"|"i2.xlarge.elasticsearch"|"i2.2xlarge.elasticsearch"|"d2.xlarge.elasticsearch"|"d2.2xlarge.elasticsearch"|"d2.4xlarge.elasticsearch"|"d2.8xlarge.elasticsearch"|"c4.large.elasticsearch"|"c4.xlarge.elasticsearch"|"c4.2xlarge.elasticsearch"|"c4.4xlarge.elasticsearch"|"c4.8xlarge.elasticsearch"|"r4.large.elasticsearch"|"r4.xlarge.elasticsearch"|"r4.2xlarge.elasticsearch"|"r4.4xlarge.elasticsearch"|"r4.8xlarge.elasticsearch"|"r4.16xlarge.elasticsearch"|"i3.large.elasticsearch"|"i3.xlarge.elasticsearch"|"i3.2xlarge.elasticsearch"|"i3.4xlarge.elasticsearch"|"i3.8xlarge.elasticsearch"|"i3.16xlarge.elasticsearch",
    InstanceCount = 123,
    DedicatedMasterEnabled = TRUE|FALSE,
    ZoneAwarenessEnabled = TRUE|FALSE,
    ZoneAwarenessConfig = list(
      AvailabilityZoneCount = 123
    ),
    DedicatedMasterType = "m3.medium.elasticsearch"|"m3.large.elasticsearch"|"m3.xlarge.elasticsearch"|"m3.2xlarge.elasticsearch"|"m4.large.elasticsearch"|"m4.xlarge.elasticsearch"|"m4.2xlarge.elasticsearch"|"m4.4xlarge.elasticsearch"|"m4.10xlarge.elasticsearch"|"t2.micro.elasticsearch"|"t2.small.elasticsearch"|"t2.medium.elasticsearch"|"r3.large.elasticsearch"|"r3.xlarge.elasticsearch"|"r3.2xlarge.elasticsearch"|"r3.4xlarge.elasticsearch"|"r3.8xlarge.elasticsearch"|"i2.xlarge.elasticsearch"|"i2.2xlarge.elasticsearch"|"d2.xlarge.elasticsearch"|"d2.2xlarge.elasticsearch"|"d2.4xlarge.elasticsearch"|"d2.8xlarge.elasticsearch"|"c4.large.elasticsearch"|"c4.xlarge.elasticsearch"|"c4.2xlarge.elasticsearch"|"c4.4xlarge.elasticsearch"|"c4.8xlarge.elasticsearch"|"r4.large.elasticsearch"|"r4.xlarge.elasticsearch"|"r4.2xlarge.elasticsearch"|"r4.4xlarge.elasticsearch"|"r4.8xlarge.elasticsearch"|"r4.16xlarge.elasticsearch"|"i3.large.elasticsearch"|"i3.xlarge.elasticsearch"|"i3.2xlarge.elasticsearch"|"i3.4xlarge.elasticsearch"|"i3.8xlarge.elasticsearch"|"i3.16xlarge.elasticsearch",
    DedicatedMasterCount = 123
  ),
  EBSOptions = list(
    EBSEnabled = TRUE|FALSE,
    VolumeType = "standard"|"gp2"|"io1",
    VolumeSize = 123,
    Iops = 123
  ),
  AccessPolicies = "string",
  SnapshotOptions = list(
    AutomatedSnapshotStartHour = 123
  ),
  VPCOptions = list(
    SubnetIds = list(
      "string"
    ),
    SecurityGroupIds = list(
      "string"
    )
  ),
  CognitoOptions = list(
    Enabled = TRUE|FALSE,
    UserPoolId = "string",
    IdentityPoolId = "string",
    RoleArn = "string"
  ),
  EncryptionAtRestOptions = list(
    Enabled = TRUE|FALSE,
    KmsKeyId = "string"
  ),
  NodeToNodeEncryptionOptions = list(
    Enabled = TRUE|FALSE
  ),
  AdvancedOptions = list(
    "string"
  ),
  LogPublishingOptions = list(
    list(
      CloudWatchLogsLogGroupArn = "string",
      Enabled = TRUE|FALSE
    )
  )
)
}
}

\keyword{internal}
