% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticsearchservice_operations.R
\name{elasticsearchservice_delete_inbound_cross_cluster_search_connection}
\alias{elasticsearchservice_delete_inbound_cross_cluster_search_connection}
\title{Allows the destination domain owner to delete an existing inbound
cross-cluster search connection}
\usage{
elasticsearchservice_delete_inbound_cross_cluster_search_connection(
  CrossClusterSearchConnectionId)
}
\arguments{
\item{CrossClusterSearchConnectionId}{[required] The id of the inbound connection that you want to permanently delete.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  CrossClusterSearchConnection = list(
    SourceDomainInfo = list(
      OwnerId = "string",
      DomainName = "string",
      Region = "string"
    ),
    DestinationDomainInfo = list(
      OwnerId = "string",
      DomainName = "string",
      Region = "string"
    ),
    CrossClusterSearchConnectionId = "string",
    ConnectionStatus = list(
      StatusCode = "PENDING_ACCEPTANCE"|"APPROVED"|"REJECTING"|"REJECTED"|"DELETING"|"DELETED",
      Message = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Allows the destination domain owner to delete an existing inbound
cross-cluster search connection.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_inbound_cross_cluster_search_connection(
  CrossClusterSearchConnectionId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
