% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticsearchservice_operations.R
\name{elasticsearchservice_delete_vpc_endpoint}
\alias{elasticsearchservice_delete_vpc_endpoint}
\title{Deletes an Amazon OpenSearch Service-managed interface VPC endpoint}
\usage{
elasticsearchservice_delete_vpc_endpoint(VpcEndpointId)
}
\arguments{
\item{VpcEndpointId}{[required] The unique identifier of the endpoint to be deleted.}
}
\description{
Deletes an Amazon OpenSearch Service-managed interface VPC endpoint.

See \url{https://paws-r.github.io/docs/elasticsearchservice/delete_vpc_endpoint.html} for full documentation.
}
\keyword{internal}
