% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticsearchservice_operations.R
\name{elasticsearchservice_describe_domain_auto_tunes}
\alias{elasticsearchservice_describe_domain_auto_tunes}
\title{Provides scheduled Auto-Tune action details for the Elasticsearch
domain, such as Auto-Tune action type, description, severity, and
scheduled date}
\usage{
elasticsearchservice_describe_domain_auto_tunes(
  DomainName,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{DomainName}{[required] Specifies the domain name for which you want Auto-Tune action details.}

\item{MaxResults}{Set this value to limit the number of results returned. If not
specified, defaults to 100.}

\item{NextToken}{NextToken is sent in case the earlier API call results contain the
NextToken. It is used for pagination.}
}
\description{
Provides scheduled Auto-Tune action details for the Elasticsearch domain, such as Auto-Tune action type, description, severity, and scheduled date.

See \url{https://www.paws-r-sdk.com/docs/elasticsearchservice_describe_domain_auto_tunes/} for full documentation.
}
\keyword{internal}
