% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticsearchservice_operations.R
\name{elasticsearchservice_dissociate_package}
\alias{elasticsearchservice_dissociate_package}
\title{Dissociates a package from the Amazon ES domain}
\usage{
elasticsearchservice_dissociate_package(PackageID, DomainName)
}
\arguments{
\item{PackageID}{[required] Internal ID of the package that you want to associate with a domain. Use
\code{\link[=elasticsearchservice_describe_packages]{describe_packages}} to find
this value.}

\item{DomainName}{[required] Name of the domain that you want to associate the package with.}
}
\description{
Dissociates a package from the Amazon ES domain.
}
\section{Request syntax}{
\preformatted{svc$dissociate_package(
  PackageID = "string",
  DomainName = "string"
)
}
}

\keyword{internal}
