% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticsearchservice_operations.R
\name{elasticsearchservice_get_compatible_elasticsearch_versions}
\alias{elasticsearchservice_get_compatible_elasticsearch_versions}
\title{Returns a list of upgrade compatible Elastisearch versions}
\usage{
elasticsearchservice_get_compatible_elasticsearch_versions(DomainName)
}
\arguments{
\item{DomainName}{}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  CompatibleElasticsearchVersions = list(
    list(
      SourceVersion = "string",
      TargetVersions = list(
        "string"
      )
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Returns a list of upgrade compatible Elastisearch versions. You can
optionally pass a \code{DomainName} to get all upgrade compatible
Elasticsearch versions for that specific domain.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_compatible_elasticsearch_versions(
  DomainName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
