% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticsearchservice_operations.R
\name{elasticsearchservice_get_compatible_elasticsearch_versions}
\alias{elasticsearchservice_get_compatible_elasticsearch_versions}
\title{Returns a list of upgrade compatible Elastisearch versions}
\usage{
elasticsearchservice_get_compatible_elasticsearch_versions(DomainName = NULL)
}
\arguments{
\item{DomainName}{}
}
\description{
Returns a list of upgrade compatible Elastisearch versions. You can optionally pass a \code{DomainName} to get all upgrade compatible Elasticsearch versions for that specific domain.

See \url{https://www.paws-r-sdk.com/docs/elasticsearchservice_get_compatible_elasticsearch_versions/} for full documentation.
}
\keyword{internal}
