% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticsearchservice_operations.R
\name{elasticsearchservice_get_upgrade_history}
\alias{elasticsearchservice_get_upgrade_history}
\title{Retrieves the complete history of the last 10 upgrades that were
performed on the domain}
\usage{
elasticsearchservice_get_upgrade_history(
  DomainName,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{DomainName}{[required]}

\item{MaxResults}{}

\item{NextToken}{}
}
\description{
Retrieves the complete history of the last 10 upgrades that were performed on the domain.

See \url{https://www.paws-r-sdk.com/docs/elasticsearchservice_get_upgrade_history/} for full documentation.
}
\keyword{internal}
