% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticsearchservice_operations.R
\name{elasticsearchservice_get_upgrade_status}
\alias{elasticsearchservice_get_upgrade_status}
\title{Retrieves the latest status of the last upgrade or upgrade eligibility
check that was performed on the domain}
\usage{
elasticsearchservice_get_upgrade_status(DomainName)
}
\arguments{
\item{DomainName}{[required]}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  UpgradeStep = "PRE_UPGRADE_CHECK"|"SNAPSHOT"|"UPGRADE",
  StepStatus = "IN_PROGRESS"|"SUCCEEDED"|"SUCCEEDED_WITH_ISSUES"|"FAILED",
  UpgradeName = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves the latest status of the last upgrade or upgrade eligibility
check that was performed on the domain.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_upgrade_status(
  DomainName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
