% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticsearchservice_operations.R
\name{elasticsearchservice_get_upgrade_status}
\alias{elasticsearchservice_get_upgrade_status}
\title{Retrieves the latest status of the last upgrade or upgrade eligibility
check that was performed on the domain}
\usage{
elasticsearchservice_get_upgrade_status(DomainName)
}
\arguments{
\item{DomainName}{[required]}
}
\description{
Retrieves the latest status of the last upgrade or upgrade eligibility check that was performed on the domain.

See \url{https://www.paws-r-sdk.com/docs/elasticsearchservice_get_upgrade_status/} for full documentation.
}
\keyword{internal}
