% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticsearchservice_operations.R
\name{elasticsearchservice_list_domain_names}
\alias{elasticsearchservice_list_domain_names}
\title{Returns the name of all Elasticsearch domains owned by the current
user's account}
\usage{
elasticsearchservice_list_domain_names(EngineType = NULL)
}
\arguments{
\item{EngineType}{Optional parameter to filter the output by domain engine type.
Acceptable values are 'Elasticsearch' and 'OpenSearch'.}
}
\description{
Returns the name of all Elasticsearch domains owned by the current user's account.

See \url{https://paws-r.github.io/docs/elasticsearchservice/list_domain_names.html} for full documentation.
}
\keyword{internal}
