% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticsearchservice_operations.R
\name{elasticsearchservice_list_domain_names}
\alias{elasticsearchservice_list_domain_names}
\title{Returns the name of all Elasticsearch domains owned by the current
user's account}
\usage{
elasticsearchservice_list_domain_names()
}
\description{
Returns the name of all Elasticsearch domains owned by the current
user's account.
}
\keyword{internal}
