% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticsearchservice_operations.R
\name{elasticsearchservice_list_elasticsearch_versions}
\alias{elasticsearchservice_list_elasticsearch_versions}
\title{List all supported Elasticsearch versions}
\usage{
elasticsearchservice_list_elasticsearch_versions(
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{MaxResults}{Set this value to limit the number of results returned. Value provided
must be greater than 10 else it wont be honored.}

\item{NextToken}{}
}
\description{
List all supported Elasticsearch versions

See \url{https://paws-r.github.io/docs/elasticsearchservice/list_elasticsearch_versions.html} for full documentation.
}
\keyword{internal}
