% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticsearchservice_operations.R
\name{elasticsearchservice_list_elasticsearch_versions}
\alias{elasticsearchservice_list_elasticsearch_versions}
\title{List all supported Elasticsearch versions}
\usage{
elasticsearchservice_list_elasticsearch_versions(MaxResults, NextToken)
}
\arguments{
\item{MaxResults}{Set this value to limit the number of results returned. Value provided
must be greater than 10 else it wont be honored.}

\item{NextToken}{}
}
\value{
A list with the following syntax:\preformatted{list(
  ElasticsearchVersions = list(
    "string"
  ),
  NextToken = "string"
)
}
}
\description{
List all supported Elasticsearch versions
}
\section{Request syntax}{
\preformatted{svc$list_elasticsearch_versions(
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
