% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticsearchservice_operations.R
\name{elasticsearchservice_purchase_reserved_elasticsearch_instance_offering}
\alias{elasticsearchservice_purchase_reserved_elasticsearch_instance_offering}
\title{Allows you to purchase reserved Elasticsearch instances}
\usage{
elasticsearchservice_purchase_reserved_elasticsearch_instance_offering(
  ReservedElasticsearchInstanceOfferingId,
  ReservationName,
  InstanceCount = NULL
)
}
\arguments{
\item{ReservedElasticsearchInstanceOfferingId}{[required] The ID of the reserved Elasticsearch instance offering to purchase.}

\item{ReservationName}{[required] A customer-specified identifier to track this reservation.}

\item{InstanceCount}{The number of Elasticsearch instances to reserve.}
}
\description{
Allows you to purchase reserved Elasticsearch instances.

See \url{https://paws-r.github.io/docs/elasticsearchservice/purchase_reserved_elasticsearch_instance_offering.html} for full documentation.
}
\keyword{internal}
