% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticsearchservice_operations.R
\name{elasticsearchservice_update_vpc_endpoint}
\alias{elasticsearchservice_update_vpc_endpoint}
\title{Modifies an Amazon OpenSearch Service-managed interface VPC endpoint}
\usage{
elasticsearchservice_update_vpc_endpoint(VpcEndpointId, VpcOptions)
}
\arguments{
\item{VpcEndpointId}{[required] Unique identifier of the VPC endpoint to be updated.}

\item{VpcOptions}{[required] The security groups and/or subnets to add, remove, or modify.}
}
\description{
Modifies an Amazon OpenSearch Service-managed interface VPC endpoint.

See \url{https://www.paws-r-sdk.com/docs/elasticsearchservice_update_vpc_endpoint/} for full documentation.
}
\keyword{internal}
