% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr_operations.R
\name{emr_delete_studio}
\alias{emr_delete_studio}
\title{Removes an Amazon EMR Studio from the Studio metadata store}
\usage{
emr_delete_studio(StudioId)
}
\arguments{
\item{StudioId}{[required] The ID of the Amazon EMR Studio.}
}
\value{
An empty list.
}
\description{
Removes an Amazon EMR Studio from the Studio metadata store.
}
\section{Request syntax}{
\preformatted{svc$delete_studio(
  StudioId = "string"
)
}
}

\keyword{internal}
