% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr_operations.R
\name{emr_describe_cluster}
\alias{emr_describe_cluster}
\title{Provides cluster-level details including status, hardware and software
configuration, VPC settings, and so on}
\usage{
emr_describe_cluster(ClusterId)
}
\arguments{
\item{ClusterId}{[required] The identifier of the cluster to describe.}
}
\description{
Provides cluster-level details including status, hardware and software configuration, VPC settings, and so on.

See \url{https://paws-r.github.io/docs/emr/describe_cluster.html} for full documentation.
}
\keyword{internal}
