% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr_operations.R
\name{emr_describe_release_label}
\alias{emr_describe_release_label}
\title{Provides Amazon EMR release label details, such as the releases
available the Region where the API request is run, and the available
applications for a specific Amazon EMR release label}
\usage{
emr_describe_release_label(
  ReleaseLabel = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{ReleaseLabel}{The target release label to be described.}

\item{NextToken}{The pagination token. Reserved for future use. Currently set to null.}

\item{MaxResults}{Reserved for future use. Currently set to null.}
}
\description{
Provides Amazon EMR release label details, such as the releases available the Region where the API request is run, and the available applications for a specific Amazon EMR release label. Can also list Amazon EMR releases that support a specified version of Spark.

See \url{https://www.paws-r-sdk.com/docs/emr_describe_release_label/} for full documentation.
}
\keyword{internal}
