% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr_operations.R
\name{emr_describe_step}
\alias{emr_describe_step}
\title{Provides more detail about the cluster step}
\usage{
emr_describe_step(ClusterId, StepId)
}
\arguments{
\item{ClusterId}{[required] The identifier of the cluster with steps to describe.}

\item{StepId}{[required] The identifier of the step to describe.}
}
\description{
Provides more detail about the cluster step.

See \url{https://www.paws-r-sdk.com/docs/emr_describe_step/} for full documentation.
}
\keyword{internal}
