% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr_operations.R
\name{emr_describe_studio}
\alias{emr_describe_studio}
\title{Returns details for the specified Amazon EMR Studio including ID, Name,
VPC, Studio access URL, and so on}
\usage{
emr_describe_studio(StudioId)
}
\arguments{
\item{StudioId}{[required] The Amazon EMR Studio ID.}
}
\description{
Returns details for the specified Amazon EMR Studio including ID, Name, VPC, Studio access URL, and so on.

See \url{https://paws-r.github.io/docs/emr/describe_studio.html} for full documentation.
}
\keyword{internal}
