% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr_operations.R
\name{emr_list_instance_groups}
\alias{emr_list_instance_groups}
\title{Provides all available details about the instance groups in a cluster}
\usage{
emr_list_instance_groups(ClusterId, Marker)
}
\arguments{
\item{ClusterId}{[required] The identifier of the cluster for which to list the instance groups.}

\item{Marker}{The pagination token that indicates the next set of results to retrieve.}
}
\description{
Provides all available details about the instance groups in a cluster.
}
\section{Request syntax}{
\preformatted{svc$list_instance_groups(
  ClusterId = "string",
  Marker = "string"
)
}
}

\keyword{internal}
