% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr_operations.R
\name{emr_list_studios}
\alias{emr_list_studios}
\title{The Amazon EMR Studio APIs are in preview release for Amazon EMR and are
subject to change}
\usage{
emr_list_studios(Marker)
}
\arguments{
\item{Marker}{The pagination token that indicates the set of results to retrieve.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Studios = list(
    list(
      StudioId = "string",
      Name = "string",
      VpcId = "string",
      Description = "string",
      Url = "string",
      CreationTime = as.POSIXct(
        "2015-01-01"
      )
    )
  ),
  Marker = "string"
)
}\if{html}{\out{</div>}}
}
\description{
The Amazon EMR Studio APIs are in preview release for Amazon EMR and are
subject to change.

Returns a list of all Amazon EMR Studios associated with the AWS
account. The list includes details such as ID, Studio Access URL, and
creation time for each Studio.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_studios(
  Marker = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
