% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr_operations.R
\name{emr_remove_managed_scaling_policy}
\alias{emr_remove_managed_scaling_policy}
\title{Removes a managed scaling policy from a specified EMR cluster}
\usage{
emr_remove_managed_scaling_policy(ClusterId)
}
\arguments{
\item{ClusterId}{[required] Specifies the ID of the cluster from which the managed scaling policy
will be removed.}
}
\value{
An empty list.
}
\description{
Removes a managed scaling policy from a specified EMR cluster.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$remove_managed_scaling_policy(
  ClusterId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
