% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr_operations.R
\name{emr_remove_tags}
\alias{emr_remove_tags}
\title{Removes tags from an Amazon EMR resource, such as a cluster or Amazon
EMR Studio}
\usage{
emr_remove_tags(ResourceId, TagKeys)
}
\arguments{
\item{ResourceId}{[required] The Amazon EMR resource identifier from which tags will be removed. For
example, a cluster identifier or an Amazon EMR Studio ID.}

\item{TagKeys}{[required] A list of tag keys to remove from the resource.}
}
\description{
Removes tags from an Amazon EMR resource, such as a cluster or Amazon EMR Studio. Tags make it easier to associate resources in various ways, such as grouping clusters to track your Amazon EMR resource allocation costs. For more information, see \href{https://docs.aws.amazon.com/emr/latest/ManagementGuide/emr-plan-tags.html}{Tag Clusters}.

See \url{https://www.paws-r-sdk.com/docs/emr_remove_tags/} for full documentation.
}
\keyword{internal}
