% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr_operations.R
\name{emr_stop_notebook_execution}
\alias{emr_stop_notebook_execution}
\title{Stops a notebook execution}
\usage{
emr_stop_notebook_execution(NotebookExecutionId)
}
\arguments{
\item{NotebookExecutionId}{[required] The unique identifier of the notebook execution.}
}
\value{
An empty list.
}
\description{
Stops a notebook execution.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$stop_notebook_execution(
  NotebookExecutionId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
