% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entityresolution_operations.R
\name{entityresolution_batch_delete_unique_id}
\alias{entityresolution_batch_delete_unique_id}
\title{Deletes multiple unique IDs in a matching workflow}
\usage{
entityresolution_batch_delete_unique_id(
  workflowName,
  inputSource = NULL,
  uniqueIds
)
}
\arguments{
\item{workflowName}{[required] The name of the workflow.}

\item{inputSource}{The input source for the batch delete unique ID operation.}

\item{uniqueIds}{[required] The unique IDs to delete.}
}
\description{
Deletes multiple unique IDs in a matching workflow.

See \url{https://www.paws-r-sdk.com/docs/entityresolution_batch_delete_unique_id/} for full documentation.
}
\keyword{internal}
