% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entityresolution_operations.R
\name{entityresolution_delete_id_namespace}
\alias{entityresolution_delete_id_namespace}
\title{Deletes the IdNamespace with a given name}
\usage{
entityresolution_delete_id_namespace(idNamespaceName)
}
\arguments{
\item{idNamespaceName}{[required] The name of the ID namespace.}
}
\description{
Deletes the \code{IdNamespace} with a given name.

See \url{https://www.paws-r-sdk.com/docs/entityresolution_delete_id_namespace/} for full documentation.
}
\keyword{internal}
