% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entityresolution_operations.R
\name{entityresolution_delete_matching_workflow}
\alias{entityresolution_delete_matching_workflow}
\title{Deletes the MatchingWorkflow with a given name}
\usage{
entityresolution_delete_matching_workflow(workflowName)
}
\arguments{
\item{workflowName}{[required] The name of the workflow to be retrieved.}
}
\description{
Deletes the \code{MatchingWorkflow} with a given name. This operation will succeed even if a workflow with the given name does not exist.

See \url{https://www.paws-r-sdk.com/docs/entityresolution_delete_matching_workflow/} for full documentation.
}
\keyword{internal}
