% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entityresolution_operations.R
\name{entityresolution_delete_schema_mapping}
\alias{entityresolution_delete_schema_mapping}
\title{Deletes the SchemaMapping with a given name}
\usage{
entityresolution_delete_schema_mapping(schemaName)
}
\arguments{
\item{schemaName}{[required] The name of the schema to delete.}
}
\description{
Deletes the \code{SchemaMapping} with a given name. This operation will succeed even if a schema with the given name does not exist. This operation will fail if there is a \code{MatchingWorkflow} object that references the \code{SchemaMapping} in the workflow's \code{InputSourceConfig}.

See \url{https://www.paws-r-sdk.com/docs/entityresolution_delete_schema_mapping/} for full documentation.
}
\keyword{internal}
