% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entityresolution_operations.R
\name{entityresolution_put_policy}
\alias{entityresolution_put_policy}
\title{Updates the resource-based policy}
\usage{
entityresolution_put_policy(arn, policy, token = NULL)
}
\arguments{
\item{arn}{[required] The Amazon Resource Name (ARN) of the resource for which the policy
needs to be updated.}

\item{policy}{[required] The resource-based policy.}

\item{token}{A unique identifier for the current revision of the policy.}
}
\description{
Updates the resource-based policy.

See \url{https://www.paws-r-sdk.com/docs/entityresolution_put_policy/} for full documentation.
}
\keyword{internal}
