% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entityresolution_operations.R
\name{entityresolution_update_matching_workflow}
\alias{entityresolution_update_matching_workflow}
\title{Updates an existing matching workflow}
\usage{
entityresolution_update_matching_workflow(
  workflowName,
  description = NULL,
  inputSourceConfig,
  outputSourceConfig,
  resolutionTechniques,
  incrementalRunConfig = NULL,
  roleArn
)
}
\arguments{
\item{workflowName}{[required] The name of the workflow to be retrieved.}

\item{description}{A description of the workflow.}

\item{inputSourceConfig}{[required] A list of \code{InputSource} objects, which have the fields \code{InputSourceARN}
and \code{SchemaName}.}

\item{outputSourceConfig}{[required] A list of \code{OutputSource} objects, each of which contains fields
\code{outputS3Path}, \code{applyNormalization}, \code{KMSArn}, and \code{output}.}

\item{resolutionTechniques}{[required] An object which defines the \code{resolutionType} and the
\code{ruleBasedProperties}.}

\item{incrementalRunConfig}{Optional. An object that defines the incremental run type. This object
contains only the \code{incrementalRunType} field, which appears as
"Automatic" in the console.

For workflows where \code{resolutionType} is \code{ML_MATCHING} or \code{PROVIDER},
incremental processing is not supported.}

\item{roleArn}{[required] The Amazon Resource Name (ARN) of the IAM role. Entity Resolution
assumes this role to create resources on your behalf as part of workflow
execution.}
}
\description{
Updates an existing matching workflow. The workflow must already exist for this operation to succeed.

See \url{https://www.paws-r-sdk.com/docs/entityresolution_update_matching_workflow/} for full documentation.
}
\keyword{internal}
