% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firehose_operations.R
\name{firehose_put_record_batch}
\alias{firehose_put_record_batch}
\title{Writes multiple data records into a Firehose stream in a single call,
which can achieve higher throughput per producer than when writing
single records}
\usage{
firehose_put_record_batch(DeliveryStreamName, Records)
}
\arguments{
\item{DeliveryStreamName}{[required] The name of the Firehose stream.}

\item{Records}{[required] One or more records.}
}
\description{
Writes multiple data records into a Firehose stream in a single call, which can achieve higher throughput per producer than when writing single records. To write single data records into a Firehose stream, use \code{\link[=firehose_put_record]{put_record}}. Applications using these operations are referred to as producers.

See \url{https://www.paws-r-sdk.com/docs/firehose_put_record_batch/} for full documentation.
}
\keyword{internal}
