% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firehose_operations.R
\name{firehose_stop_delivery_stream_encryption}
\alias{firehose_stop_delivery_stream_encryption}
\title{Disables server-side encryption (SSE) for the delivery stream}
\usage{
firehose_stop_delivery_stream_encryption(DeliveryStreamName)
}
\arguments{
\item{DeliveryStreamName}{[required] The name of the delivery stream for which you want to disable
server-side encryption (SSE).}
}
\description{
Disables server-side encryption (SSE) for the delivery stream.

See \url{https://paws-r.github.io/docs/firehose/stop_delivery_stream_encryption.html} for full documentation.
}
\keyword{internal}
