% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firehose_operations.R
\name{firehose_tag_delivery_stream}
\alias{firehose_tag_delivery_stream}
\title{Adds or updates tags for the specified Firehose stream}
\usage{
firehose_tag_delivery_stream(DeliveryStreamName, Tags)
}
\arguments{
\item{DeliveryStreamName}{[required] The name of the Firehose stream to which you want to add the tags.}

\item{Tags}{[required] A set of key-value pairs to use to create the tags.}
}
\description{
Adds or updates tags for the specified Firehose stream. A tag is a key-value pair that you can define and assign to Amazon Web Services resources. If you specify a tag that already exists, the tag value is replaced with the value that you specify in the request. Tags are metadata. For example, you can add friendly names and descriptions or other types of information that can help you distinguish the Firehose stream. For more information about tags, see \href{https://docs.aws.amazon.com/awsaccountbilling/latest/aboutv2/cost-alloc-tags.html}{Using Cost Allocation Tags} in the \emph{Amazon Web Services Billing and Cost Management User Guide}.

See \url{https://www.paws-r-sdk.com/docs/firehose_tag_delivery_stream/} for full documentation.
}
\keyword{internal}
