% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_batch_get_crawlers}
\alias{glue_batch_get_crawlers}
\title{Returns a list of resource metadata for a given list of crawler names}
\usage{
glue_batch_get_crawlers(CrawlerNames)
}
\arguments{
\item{CrawlerNames}{[required] A list of crawler names, which might be the names returned from the
\code{\link[=glue_list_crawlers]{list_crawlers}} operation.}
}
\description{
Returns a list of resource metadata for a given list of crawler names. After calling the \code{\link[=glue_list_crawlers]{list_crawlers}} operation, you can call this operation to access the data to which you have been granted permissions. This operation supports all IAM permissions, including permission conditions that uses tags.

See \url{https://paws-r.github.io/docs/glue/batch_get_crawlers.html} for full documentation.
}
\keyword{internal}
