% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_batch_get_dev_endpoints}
\alias{glue_batch_get_dev_endpoints}
\title{Returns a list of resource metadata for a given list of development
endpoint names}
\usage{
glue_batch_get_dev_endpoints(DevEndpointNames)
}
\arguments{
\item{DevEndpointNames}{[required] The list of \code{DevEndpoint} names, which might be the names returned from
the \code{ListDevEndpoint} operation.}
}
\description{
Returns a list of resource metadata for a given list of development endpoint names. After calling the \code{\link[=glue_list_dev_endpoints]{list_dev_endpoints}} operation, you can call this operation to access the data to which you have been granted permissions. This operation supports all IAM permissions, including permission conditions that uses tags.

See \url{https://paws-r.github.io/docs/glue/batch_get_dev_endpoints.html} for full documentation.
}
\keyword{internal}
