% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_batch_get_jobs}
\alias{glue_batch_get_jobs}
\title{Returns a list of resource metadata for a given list of job names}
\usage{
glue_batch_get_jobs(JobNames)
}
\arguments{
\item{JobNames}{[required] A list of job names, which might be the names returned from the
\code{\link[=glue_list_jobs]{list_jobs}} operation.}
}
\description{
Returns a list of resource metadata for a given list of job names. After calling the \code{\link[=glue_list_jobs]{list_jobs}} operation, you can call this operation to access the data to which you have been granted permissions. This operation supports all IAM permissions, including permission conditions that uses tags.

See \url{https://paws-r.github.io/docs/glue/batch_get_jobs.html} for full documentation.
}
\keyword{internal}
