% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_batch_stop_job_run}
\alias{glue_batch_stop_job_run}
\title{Stops one or more job runs for a specified job definition}
\usage{
glue_batch_stop_job_run(JobName, JobRunIds)
}
\arguments{
\item{JobName}{[required] The name of the job definition for which to stop job runs.}

\item{JobRunIds}{[required] A list of the \code{JobRunIds} that should be stopped for that job
definition.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  SuccessfulSubmissions = list(
    list(
      JobName = "string",
      JobRunId = "string"
    )
  ),
  Errors = list(
    list(
      JobName = "string",
      JobRunId = "string",
      ErrorDetail = list(
        ErrorCode = "string",
        ErrorMessage = "string"
      )
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Stops one or more job runs for a specified job definition.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$batch_stop_job_run(
  JobName = "string",
  JobRunIds = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
