% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_create_classifier}
\alias{glue_create_classifier}
\title{Creates a classifier in the user's account}
\usage{
glue_create_classifier(
  GrokClassifier = NULL,
  XMLClassifier = NULL,
  JsonClassifier = NULL,
  CsvClassifier = NULL
)
}
\arguments{
\item{GrokClassifier}{A \code{GrokClassifier} object specifying the classifier to create.}

\item{XMLClassifier}{An \code{XMLClassifier} object specifying the classifier to create.}

\item{JsonClassifier}{A \code{JsonClassifier} object specifying the classifier to create.}

\item{CsvClassifier}{A \code{CsvClassifier} object specifying the classifier to create.}
}
\description{
Creates a classifier in the user's account. This can be a \code{GrokClassifier}, an \code{XMLClassifier}, a \code{JsonClassifier}, or a \code{CsvClassifier}, depending on which field of the request is present.

See \url{https://paws-r.github.io/docs/glue/create_classifier.html} for full documentation.
}
\keyword{internal}
