% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_create_database}
\alias{glue_create_database}
\title{Creates a new database in a Data Catalog}
\usage{
glue_create_database(CatalogId = NULL, DatabaseInput, Tags = NULL)
}
\arguments{
\item{CatalogId}{The ID of the Data Catalog in which to create the database. If none is
provided, the Amazon Web Services account ID is used by default.}

\item{DatabaseInput}{[required] The metadata for the database.}

\item{Tags}{The tags you assign to the database.}
}
\description{
Creates a new database in a Data Catalog.

See \url{https://www.paws-r-sdk.com/docs/glue_create_database/} for full documentation.
}
\keyword{internal}
