% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_delete_column_statistics_for_partition}
\alias{glue_delete_column_statistics_for_partition}
\title{Delete the partition column statistics of a column}
\usage{
glue_delete_column_statistics_for_partition(CatalogId, DatabaseName,
  TableName, PartitionValues, ColumnName)
}
\arguments{
\item{CatalogId}{The ID of the Data Catalog where the partitions in question reside. If
none is supplied, the AWS account ID is used by default.}

\item{DatabaseName}{[required] The name of the catalog database where the partitions reside.}

\item{TableName}{[required] The name of the partitions' table.}

\item{PartitionValues}{[required] A list of partition values identifying the partition.}

\item{ColumnName}{[required] Name of the column.}
}
\value{
An empty list.
}
\description{
Delete the partition column statistics of a column.

The Identity and Access Management (IAM) permission required for this
operation is \code{\link[=glue_delete_partition]{delete_partition}}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_column_statistics_for_partition(
  CatalogId = "string",
  DatabaseName = "string",
  TableName = "string",
  PartitionValues = list(
    "string"
  ),
  ColumnName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
