% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_delete_column_statistics_task_settings}
\alias{glue_delete_column_statistics_task_settings}
\title{Deletes settings for a column statistics task}
\usage{
glue_delete_column_statistics_task_settings(DatabaseName, TableName)
}
\arguments{
\item{DatabaseName}{[required] The name of the database where the table resides.}

\item{TableName}{[required] The name of the table for which to delete column statistics.}
}
\description{
Deletes settings for a column statistics task.

See \url{https://www.paws-r-sdk.com/docs/glue_delete_column_statistics_task_settings/} for full documentation.
}
\keyword{internal}
