% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_delete_dev_endpoint}
\alias{glue_delete_dev_endpoint}
\title{Deletes a specified development endpoint}
\usage{
glue_delete_dev_endpoint(EndpointName)
}
\arguments{
\item{EndpointName}{[required] The name of the \code{DevEndpoint}.}
}
\description{
Deletes a specified development endpoint.

See \url{https://paws-r.github.io/docs/glue/delete_dev_endpoint.html} for full documentation.
}
\keyword{internal}
