% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_delete_job}
\alias{glue_delete_job}
\title{Deletes a specified job definition}
\usage{
glue_delete_job(JobName)
}
\arguments{
\item{JobName}{[required] The name of the job definition to delete.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  JobName = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Deletes a specified job definition. If the job definition is not found,
no exception is thrown.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_job(
  JobName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
