% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_delete_partition}
\alias{glue_delete_partition}
\title{Deletes a specified partition}
\usage{
glue_delete_partition(CatalogId, DatabaseName, TableName,
  PartitionValues)
}
\arguments{
\item{CatalogId}{The ID of the Data Catalog where the partition to be deleted resides. If
none is supplied, the AWS account ID is used by default.}

\item{DatabaseName}{[required] The name of the catalog database in which the table in question resides.}

\item{TableName}{[required] The name of the table where the partition to be deleted is located.}

\item{PartitionValues}{[required] The values that define the partition.}
}
\description{
Deletes a specified partition.
}
\section{Request syntax}{
\preformatted{svc$delete_partition(
  CatalogId = "string",
  DatabaseName = "string",
  TableName = "string",
  PartitionValues = list(
    "string"
  )
)
}
}

\keyword{internal}
