% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_delete_registry}
\alias{glue_delete_registry}
\title{Delete the entire registry including schema and all of its versions}
\usage{
glue_delete_registry(RegistryId)
}
\arguments{
\item{RegistryId}{[required] This is a wrapper structure that may contain the registry name and
Amazon Resource Name (ARN).}
}
\description{
Delete the entire registry including schema and all of its versions. To get the status of the delete operation, you can call the \code{\link[=glue_get_registry]{get_registry}} API after the asynchronous call. Deleting a registry will deactivate all online operations for the registry such as the \code{\link[=glue_update_registry]{update_registry}}, \code{\link[=glue_create_schema]{create_schema}}, \code{\link[=glue_update_schema]{update_schema}}, and \code{\link[=glue_register_schema_version]{register_schema_version}} APIs.

See \url{https://www.paws-r-sdk.com/docs/glue_delete_registry/} for full documentation.
}
\keyword{internal}
