% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_delete_registry}
\alias{glue_delete_registry}
\title{Delete the entire registry including schema and all of its versions}
\usage{
glue_delete_registry(RegistryId)
}
\arguments{
\item{RegistryId}{[required] This is a wrapper structure that may contain the registry name and
Amazon Resource Name (ARN).}
}
\description{
Delete the entire registry including schema and all of its versions. To
get the status of the delete operation, you can call the \code{GetRegistry}
API after the asynchronous call. Deleting a registry will disable all
online operations for the registry such as the \code{UpdateRegistry},
\code{CreateSchema}, \code{UpdateSchema}, and \code{RegisterSchemaVersion} APIs.
}
\section{Request syntax}{
\preformatted{svc$delete_registry(
  RegistryId = list(
    RegistryName = "string",
    RegistryArn = "string"
  )
)
}
}

\keyword{internal}
