% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_delete_resource_policy}
\alias{glue_delete_resource_policy}
\title{Deletes a specified policy}
\usage{
glue_delete_resource_policy(PolicyHashCondition, ResourceArn)
}
\arguments{
\item{PolicyHashCondition}{The hash value returned when this policy was set.}

\item{ResourceArn}{The ARN of the AWS Glue resource for the resource policy to be deleted.}
}
\value{
An empty list.
}
\description{
Deletes a specified policy.
}
\section{Request syntax}{
\preformatted{svc$delete_resource_policy(
  PolicyHashCondition = "string",
  ResourceArn = "string"
)
}
}

\keyword{internal}
