% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_delete_schema_versions}
\alias{glue_delete_schema_versions}
\title{Remove versions from the specified schema}
\usage{
glue_delete_schema_versions(SchemaId, Versions)
}
\arguments{
\item{SchemaId}{[required] This is a wrapper structure that may contain the schema name and Amazon
Resource Name (ARN).}

\item{Versions}{[required] A version range may be supplied which may be of the format:
\itemize{
\item a single version number, 5
\item a range, 5-8 : deletes versions 5, 6, 7, 8
}}
}
\description{
Remove versions from the specified schema. A version number or range may be supplied. If the compatibility mode forbids deleting of a version that is necessary, such as BACKWARDS_FULL, an error is returned. Calling the \code{GetSchemaVersions} API after this call will list the status of the deleted versions.

See \url{https://www.paws-r-sdk.com/docs/glue_delete_schema_versions/} for full documentation.
}
\keyword{internal}
