% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_catalog_import_status}
\alias{glue_get_catalog_import_status}
\title{Retrieves the status of a migration operation}
\usage{
glue_get_catalog_import_status(CatalogId = NULL)
}
\arguments{
\item{CatalogId}{The ID of the catalog to migrate. Currently, this should be the Amazon
Web Services account ID.}
}
\description{
Retrieves the status of a migration operation.

See \url{https://paws-r.github.io/docs/glue/get_catalog_import_status.html} for full documentation.
}
\keyword{internal}
