% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_column_statistics_for_partition}
\alias{glue_get_column_statistics_for_partition}
\title{Retrieves partition statistics of columns}
\usage{
glue_get_column_statistics_for_partition(CatalogId, DatabaseName,
  TableName, PartitionValues, ColumnNames)
}
\arguments{
\item{CatalogId}{The ID of the Data Catalog where the partitions in question reside. If
none is supplied, the AWS account ID is used by default.}

\item{DatabaseName}{[required] The name of the catalog database where the partitions reside.}

\item{TableName}{[required] The name of the partitions\' table.}

\item{PartitionValues}{[required] A list of partition values identifying the partition.}

\item{ColumnNames}{[required] A list of the column names.}
}
\description{
Retrieves partition statistics of columns.
}
\section{Request syntax}{
\preformatted{svc$get_column_statistics_for_partition(
  CatalogId = "string",
  DatabaseName = "string",
  TableName = "string",
  PartitionValues = list(
    "string"
  ),
  ColumnNames = list(
    "string"
  )
)
}
}

\keyword{internal}
