% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_crawler}
\alias{glue_get_crawler}
\title{Retrieves metadata for a specified crawler}
\usage{
glue_get_crawler(Name)
}
\arguments{
\item{Name}{[required] The name of the crawler to retrieve metadata for.}
}
\description{
Retrieves metadata for a specified crawler.

See \url{https://paws-r.github.io/docs/glue/get_crawler.html} for full documentation.
}
\keyword{internal}
