% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_dev_endpoint}
\alias{glue_get_dev_endpoint}
\title{Retrieves information about a specified development endpoint}
\usage{
glue_get_dev_endpoint(EndpointName)
}
\arguments{
\item{EndpointName}{[required] Name of the \code{DevEndpoint} to retrieve information for.}
}
\description{
Retrieves information about a specified development endpoint.

See \url{https://www.paws-r-sdk.com/docs/glue_get_dev_endpoint/} for full documentation.
}
\keyword{internal}
