% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_dev_endpoints}
\alias{glue_get_dev_endpoints}
\title{Retrieves all the development endpoints in this AWS account}
\usage{
glue_get_dev_endpoints(MaxResults, NextToken)
}
\arguments{
\item{MaxResults}{The maximum size of information to return.}

\item{NextToken}{A continuation token, if this is a continuation call.}
}
\description{
Retrieves all the development endpoints in this AWS account.

When you create a development endpoint in a virtual private cloud (VPC),
AWS Glue returns only a private IP address and the public IP address
field is not populated. When you create a non-VPC development endpoint,
AWS Glue returns only a public IP address.
}
\section{Request syntax}{
\preformatted{svc$get_dev_endpoints(
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
