% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_integration_table_properties}
\alias{glue_get_integration_table_properties}
\title{This API is used to retrieve optional override properties for the tables
that need to be replicated}
\usage{
glue_get_integration_table_properties(ResourceArn, TableName)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the target table for which to retrieve
integration table properties. Currently, this API only supports
retrieving properties for target tables, and the provided ARN should be
the ARN of the target table in the Glue Data Catalog. Support for
retrieving integration table properties for source connections (using
the connection ARN) is not yet implemented and will be added in a future
release.}

\item{TableName}{[required] The name of the table to be replicated.}
}
\description{
This API is used to retrieve optional override properties for the tables that need to be replicated. These properties can include properties for filtering and partition for source and target tables.

See \url{https://www.paws-r-sdk.com/docs/glue_get_integration_table_properties/} for full documentation.
}
\keyword{internal}
