% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_job_bookmark}
\alias{glue_get_job_bookmark}
\title{Returns information on a job bookmark entry}
\usage{
glue_get_job_bookmark(JobName, RunId = NULL)
}
\arguments{
\item{JobName}{[required] The name of the job in question.}

\item{RunId}{The unique run identifier associated with this job run.}
}
\description{
Returns information on a job bookmark entry.

See \url{https://www.paws-r-sdk.com/docs/glue_get_job_bookmark/} for full documentation.
}
\keyword{internal}
