% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_job_run}
\alias{glue_get_job_run}
\title{Retrieves the metadata for a given job run}
\usage{
glue_get_job_run(JobName, RunId, PredecessorsIncluded = NULL)
}
\arguments{
\item{JobName}{[required] Name of the job definition being run.}

\item{RunId}{[required] The ID of the job run.}

\item{PredecessorsIncluded}{True if a list of predecessor runs should be returned.}
}
\description{
Retrieves the metadata for a given job run.

See \url{https://paws-r.github.io/docs/glue/get_job_run.html} for full documentation.
}
\keyword{internal}
