% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_mapping}
\alias{glue_get_mapping}
\title{Creates mappings}
\usage{
glue_get_mapping(Source, Sinks = NULL, Location = NULL)
}
\arguments{
\item{Source}{[required] Specifies the source table.}

\item{Sinks}{A list of target tables.}

\item{Location}{Parameters for the mapping.}
}
\description{
Creates mappings.

See \url{https://paws-r.github.io/docs/glue/get_mapping.html} for full documentation.
}
\keyword{internal}
