% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_partition}
\alias{glue_get_partition}
\title{Retrieves information about a specified partition}
\usage{
glue_get_partition(CatalogId = NULL, DatabaseName, TableName, PartitionValues)
}
\arguments{
\item{CatalogId}{The ID of the Data Catalog where the partition in question resides. If
none is provided, the Amazon Web Services account ID is used by default.}

\item{DatabaseName}{[required] The name of the catalog database where the partition resides.}

\item{TableName}{[required] The name of the partition's table.}

\item{PartitionValues}{[required] The values that define the partition.}
}
\description{
Retrieves information about a specified partition.

See \url{https://www.paws-r-sdk.com/docs/glue_get_partition/} for full documentation.
}
\keyword{internal}
