% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_plan}
\alias{glue_get_plan}
\title{Gets code to perform a specified mapping}
\usage{
glue_get_plan(Mapping, Source, Sinks, Location, Language,
  AdditionalPlanOptionsMap)
}
\arguments{
\item{Mapping}{[required] The list of mappings from a source table to target tables.}

\item{Source}{[required] The source table.}

\item{Sinks}{The target tables.}

\item{Location}{The parameters for the mapping.}

\item{Language}{The programming language of the code to perform the mapping.}

\item{AdditionalPlanOptionsMap}{A map to hold additional optional key-value parameters.

Currently, these key-value pairs are supported:
\itemize{
\item \code{inferSchema}  —  Specifies whether to set \code{inferSchema} to true or
false for the default script generated by an AWS Glue job. For
example, to set \code{inferSchema} to true, pass the following key value
pair:

\verb{--additional-plan-options-map '\{"inferSchema":"true"\}'}
}}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  PythonScript = "string",
  ScalaCode = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Gets code to perform a specified mapping.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_plan(
  Mapping = list(
    list(
      SourceTable = "string",
      SourcePath = "string",
      SourceType = "string",
      TargetTable = "string",
      TargetPath = "string",
      TargetType = "string"
    )
  ),
  Source = list(
    DatabaseName = "string",
    TableName = "string"
  ),
  Sinks = list(
    list(
      DatabaseName = "string",
      TableName = "string"
    )
  ),
  Location = list(
    Jdbc = list(
      list(
        Name = "string",
        Value = "string",
        Param = TRUE|FALSE
      )
    ),
    S3 = list(
      list(
        Name = "string",
        Value = "string",
        Param = TRUE|FALSE
      )
    ),
    DynamoDB = list(
      list(
        Name = "string",
        Value = "string",
        Param = TRUE|FALSE
      )
    )
  ),
  Language = "PYTHON"|"SCALA",
  AdditionalPlanOptionsMap = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
