% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_registry}
\alias{glue_get_registry}
\title{Describes the specified registry in detail}
\usage{
glue_get_registry(RegistryId)
}
\arguments{
\item{RegistryId}{[required] This is a wrapper structure that may contain the registry name and
Amazon Resource Name (ARN).}
}
\description{
Describes the specified registry in detail.
}
\section{Request syntax}{
\preformatted{svc$get_registry(
  RegistryId = list(
    RegistryName = "string",
    RegistryArn = "string"
  )
)
}
}

\keyword{internal}
