% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_resource_policies}
\alias{glue_get_resource_policies}
\title{Retrieves the resource policies set on individual resources by Resource
Access Manager during cross-account permission grants}
\usage{
glue_get_resource_policies(NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{NextToken}{A continuation token, if this is a continuation request.}

\item{MaxResults}{The maximum size of a list to return.}
}
\description{
Retrieves the resource policies set on individual resources by Resource Access Manager during cross-account permission grants. Also retrieves the Data Catalog resource policy.

See \url{https://www.paws-r-sdk.com/docs/glue_get_resource_policies/} for full documentation.
}
\keyword{internal}
