% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_resource_policies}
\alias{glue_get_resource_policies}
\title{Retrieves the security configurations for the resource policies set on
individual resources, and also the account-level policy}
\usage{
glue_get_resource_policies(NextToken, MaxResults)
}
\arguments{
\item{NextToken}{A continuation token, if this is a continuation request.}

\item{MaxResults}{The maximum size of a list to return.}
}
\description{
Retrieves the security configurations for the resource policies set on
individual resources, and also the account-level policy.
}
\section{Request syntax}{
\preformatted{svc$get_resource_policies(
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
