% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_resource_policy}
\alias{glue_get_resource_policy}
\title{Retrieves a specified resource policy}
\usage{
glue_get_resource_policy(ResourceArn)
}
\arguments{
\item{ResourceArn}{The ARN of the AWS Glue resource for which to retrieve the resource
policy. If not supplied, the Data Catalog resource policy is returned.
Use \code{\link[=glue_get_resource_policies]{get_resource_policies}} to view all
existing resource policies. For more information see \href{https://docs.aws.amazon.com/glue/latest/dg/glue-specifying-resource-arns.html}{Specifying AWS Glue Resource ARNs}.}
}
\value{
A list with the following syntax:\preformatted{list(
  PolicyInJson = "string",
  PolicyHash = "string",
  CreateTime = as.POSIXct(
    "2015-01-01"
  ),
  UpdateTime = as.POSIXct(
    "2015-01-01"
  )
)
}
}
\description{
Retrieves a specified resource policy.
}
\section{Request syntax}{
\preformatted{svc$get_resource_policy(
  ResourceArn = "string"
)
}
}

\keyword{internal}
