% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_schema}
\alias{glue_get_schema}
\title{Describes the specified schema in detail}
\usage{
glue_get_schema(SchemaId)
}
\arguments{
\item{SchemaId}{[required] This is a wrapper structure to contain schema identity fields. The
structure contains:
\itemize{
\item SchemaId$SchemaArn: The Amazon Resource Name (ARN) of the schema.
Either \code{SchemaArn} or \code{SchemaName} and \code{RegistryName} has to be
provided.
\item SchemaId$SchemaName: The name of the schema. Either \code{SchemaArn} or
\code{SchemaName} and \code{RegistryName} has to be provided.
}}
}
\description{
Describes the specified schema in detail.

See \url{https://www.paws-r-sdk.com/docs/glue_get_schema/} for full documentation.
}
\keyword{internal}
