% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_security_configuration}
\alias{glue_get_security_configuration}
\title{Retrieves a specified security configuration}
\usage{
glue_get_security_configuration(Name)
}
\arguments{
\item{Name}{[required] The name of the security configuration to retrieve.}
}
\description{
Retrieves a specified security configuration.

See \url{https://paws-r.github.io/docs/glue/get_security_configuration.html} for full documentation.
}
\keyword{internal}
