% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_trigger}
\alias{glue_get_trigger}
\title{Retrieves the definition of a trigger}
\usage{
glue_get_trigger(Name)
}
\arguments{
\item{Name}{[required] The name of the trigger to retrieve.}
}
\description{
Retrieves the definition of a trigger.

See \url{https://paws-r.github.io/docs/glue/get_trigger.html} for full documentation.
}
\keyword{internal}
