% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_triggers}
\alias{glue_get_triggers}
\title{Gets all the triggers associated with a job}
\usage{
glue_get_triggers(NextToken = NULL, DependentJobName = NULL, MaxResults = NULL)
}
\arguments{
\item{NextToken}{A continuation token, if this is a continuation call.}

\item{DependentJobName}{The name of the job to retrieve triggers for. The trigger that can start
this job is returned, and if there is no such trigger, all triggers are
returned.}

\item{MaxResults}{The maximum size of the response.}
}
\description{
Gets all the triggers associated with a job.

See \url{https://paws-r.github.io/docs/glue/get_triggers.html} for full documentation.
}
\keyword{internal}
