% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_unfiltered_table_metadata}
\alias{glue_get_unfiltered_table_metadata}
\title{Get unfiltered table metadata}
\usage{
glue_get_unfiltered_table_metadata(
  CatalogId,
  DatabaseName,
  Name,
  AuditContext = NULL,
  SupportedPermissionTypes
)
}
\arguments{
\item{CatalogId}{[required]}

\item{DatabaseName}{[required]}

\item{Name}{[required]}

\item{AuditContext}{}

\item{SupportedPermissionTypes}{[required]}
}
\description{
Get unfiltered table metadata

See \url{https://paws-r.github.io/docs/glue/get_unfiltered_table_metadata.html} for full documentation.
}
\keyword{internal}
