% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_user_defined_function}
\alias{glue_get_user_defined_function}
\title{Retrieves a specified function definition from the Data Catalog}
\usage{
glue_get_user_defined_function(CatalogId = NULL, DatabaseName, FunctionName)
}
\arguments{
\item{CatalogId}{The ID of the Data Catalog where the function to be retrieved is
located. If none is provided, the Amazon Web Services account ID is used
by default.}

\item{DatabaseName}{[required] The name of the catalog database where the function is located.}

\item{FunctionName}{[required] The name of the function.}
}
\description{
Retrieves a specified function definition from the Data Catalog.

See \url{https://www.paws-r-sdk.com/docs/glue_get_user_defined_function/} for full documentation.
}
\keyword{internal}
