% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_workflow}
\alias{glue_get_workflow}
\title{Retrieves resource metadata for a workflow}
\usage{
glue_get_workflow(Name, IncludeGraph)
}
\arguments{
\item{Name}{[required] The name of the workflow to retrieve.}

\item{IncludeGraph}{Specifies whether to include a graph when returning the workflow
resource metadata.}
}
\description{
Retrieves resource metadata for a workflow.
}
\section{Request syntax}{
\preformatted{svc$get_workflow(
  Name = "string",
  IncludeGraph = TRUE|FALSE
)
}
}

\keyword{internal}
