% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_workflow_run_properties}
\alias{glue_get_workflow_run_properties}
\title{Retrieves the workflow run properties which were set during the run}
\usage{
glue_get_workflow_run_properties(Name, RunId)
}
\arguments{
\item{Name}{[required] Name of the workflow which was run.}

\item{RunId}{[required] The ID of the workflow run whose run properties should be returned.}
}
\description{
Retrieves the workflow run properties which were set during the run.

See \url{https://paws-r.github.io/docs/glue/get_workflow_run_properties.html} for full documentation.
}
\keyword{internal}
