% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_import_catalog_to_glue}
\alias{glue_import_catalog_to_glue}
\title{Imports an existing Amazon Athena Data Catalog to Glue}
\usage{
glue_import_catalog_to_glue(CatalogId = NULL)
}
\arguments{
\item{CatalogId}{The ID of the catalog to import. Currently, this should be the Amazon
Web Services account ID.}
}
\description{
Imports an existing Amazon Athena Data Catalog to Glue.

See \url{https://paws-r.github.io/docs/glue/import_catalog_to_glue.html} for full documentation.
}
\keyword{internal}
