% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_list_blueprints}
\alias{glue_list_blueprints}
\title{Lists all the blueprint names in an account}
\usage{
glue_list_blueprints(NextToken = NULL, MaxResults = NULL, Tags = NULL)
}
\arguments{
\item{NextToken}{A continuation token, if this is a continuation request.}

\item{MaxResults}{The maximum size of a list to return.}

\item{Tags}{Filters the list by an Amazon Web Services resource tag.}
}
\description{
Lists all the blueprint names in an account.

See \url{https://www.paws-r-sdk.com/docs/glue_list_blueprints/} for full documentation.
}
\keyword{internal}
