% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_list_data_quality_rule_recommendation_runs}
\alias{glue_list_data_quality_rule_recommendation_runs}
\title{Lists the recommendation runs meeting the filter criteria}
\usage{
glue_list_data_quality_rule_recommendation_runs(
  Filter = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{Filter}{The filter criteria.}

\item{NextToken}{A paginated token to offset the results.}

\item{MaxResults}{The maximum number of results to return.}
}
\description{
Lists the recommendation runs meeting the filter criteria.

See \url{https://www.paws-r-sdk.com/docs/glue_list_data_quality_rule_recommendation_runs/} for full documentation.
}
\keyword{internal}
