% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_list_materialized_view_refresh_task_runs}
\alias{glue_list_materialized_view_refresh_task_runs}
\title{List all task runs for a particular account}
\usage{
glue_list_materialized_view_refresh_task_runs(
  CatalogId,
  DatabaseName = NULL,
  TableName = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{CatalogId}{[required] The ID of the Data Catalog where the table resides. If none is supplied,
the account ID is used by default.}

\item{DatabaseName}{The database where the table resides.}

\item{TableName}{The name of the table for which statistics is generated.}

\item{MaxResults}{The maximum size of the response.}

\item{NextToken}{A continuation token, if this is a continuation call.}
}
\description{
List all task runs for a particular account.

See \url{https://www.paws-r-sdk.com/docs/glue_list_materialized_view_refresh_task_runs/} for full documentation.
}
\keyword{internal}
