% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_list_registries}
\alias{glue_list_registries}
\title{Returns a list of registries that you have created, with minimal
registry information}
\usage{
glue_list_registries(MaxResults, NextToken)
}
\arguments{
\item{MaxResults}{Maximum number of results required per page. If the value is not
supplied, this will be defaulted to 25 per page.}

\item{NextToken}{A continuation token, if this is a continuation call.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Registries = list(
    list(
      RegistryName = "string",
      RegistryArn = "string",
      Description = "string",
      Status = "AVAILABLE"|"DELETING",
      CreatedTime = "string",
      UpdatedTime = "string"
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Returns a list of registries that you have created, with minimal
registry information. Registries in the \code{Deleting} status will not be
included in the results. Empty results will be returned if there are no
registries available.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_registries(
  MaxResults = 123,
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
