% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_list_triggers}
\alias{glue_list_triggers}
\title{Retrieves the names of all trigger resources in this Amazon Web Services
account, or the resources with the specified tag}
\usage{
glue_list_triggers(
  NextToken = NULL,
  DependentJobName = NULL,
  MaxResults = NULL,
  Tags = NULL
)
}
\arguments{
\item{NextToken}{A continuation token, if this is a continuation request.}

\item{DependentJobName}{The name of the job for which to retrieve triggers. The trigger that can
start this job is returned. If there is no such trigger, all triggers
are returned.}

\item{MaxResults}{The maximum size of a list to return.}

\item{Tags}{Specifies to return only these tagged resources.}
}
\description{
Retrieves the names of all trigger resources in this Amazon Web Services account, or the resources with the specified tag. This operation allows you to see which resources are available in your account, and their names.

See \url{https://paws-r.github.io/docs/glue/list_triggers.html} for full documentation.
}
\keyword{internal}
