% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_reset_job_bookmark}
\alias{glue_reset_job_bookmark}
\title{Resets a bookmark entry}
\usage{
glue_reset_job_bookmark(JobName, RunId = NULL)
}
\arguments{
\item{JobName}{[required] The name of the job in question.}

\item{RunId}{The unique run identifier associated with this job run.}
}
\description{
Resets a bookmark entry.

See \url{https://paws-r.github.io/docs/glue/reset_job_bookmark.html} for full documentation.
}
\keyword{internal}
