% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_start_import_labels_task_run}
\alias{glue_start_import_labels_task_run}
\title{Enables you to provide additional labels (examples of truth) to be used
to teach the machine learning transform and improve its quality}
\usage{
glue_start_import_labels_task_run(
  TransformId,
  InputS3Path,
  ReplaceAllLabels = NULL
)
}
\arguments{
\item{TransformId}{[required] The unique identifier of the machine learning transform.}

\item{InputS3Path}{[required] The Amazon Simple Storage Service (Amazon S3) path from where you import
the labels.}

\item{ReplaceAllLabels}{Indicates whether to overwrite your existing labels.}
}
\description{
Enables you to provide additional labels (examples of truth) to be used to teach the machine learning transform and improve its quality. This API operation is generally used as part of the active learning workflow that starts with the \code{\link[=glue_start_ml_labeling_set_generation_task_run]{start_ml_labeling_set_generation_task_run}} call and that ultimately results in improving the quality of your machine learning transform.

See \url{https://paws-r.github.io/docs/glue/start_import_labels_task_run.html} for full documentation.
}
\keyword{internal}
