% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_start_workflow_run}
\alias{glue_start_workflow_run}
\title{Starts a new run of the specified workflow}
\usage{
glue_start_workflow_run(Name, RunProperties = NULL)
}
\arguments{
\item{Name}{[required] The name of the workflow to start.}

\item{RunProperties}{The workflow run properties for the new workflow run.

Run properties may be logged. Do not pass plaintext secrets as
properties. Retrieve secrets from a Glue Connection, Amazon Web Services
Secrets Manager or other secret management mechanism if you intend to
use them within the workflow run.}
}
\description{
Starts a new run of the specified workflow.

See \url{https://www.paws-r-sdk.com/docs/glue_start_workflow_run/} for full documentation.
}
\keyword{internal}
