% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_stop_crawler_schedule}
\alias{glue_stop_crawler_schedule}
\title{Sets the schedule state of the specified crawler to NOT_SCHEDULED, but
does not stop the crawler if it is already running}
\usage{
glue_stop_crawler_schedule(CrawlerName)
}
\arguments{
\item{CrawlerName}{[required] Name of the crawler whose schedule state to set.}
}
\value{
An empty list.
}
\description{
Sets the schedule state of the specified crawler to \code{NOT_SCHEDULED}, but
does not stop the crawler if it is already running.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$stop_crawler_schedule(
  CrawlerName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
