% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_stop_materialized_view_refresh_task_run}
\alias{glue_stop_materialized_view_refresh_task_run}
\title{Stops a materialized view refresh task run, for a specified table and
columns}
\usage{
glue_stop_materialized_view_refresh_task_run(
  CatalogId,
  DatabaseName,
  TableName
)
}
\arguments{
\item{CatalogId}{[required] The ID of the Data Catalog where the table reside. If none is supplied,
the account ID is used by default.}

\item{DatabaseName}{[required] The name of the database where the table resides.}

\item{TableName}{[required] The name of the table to generate statistics.}
}
\description{
Stops a materialized view refresh task run, for a specified table and columns.

See \url{https://www.paws-r-sdk.com/docs/glue_stop_materialized_view_refresh_task_run/} for full documentation.
}
\keyword{internal}
