% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_stop_session}
\alias{glue_stop_session}
\title{Stops the session}
\usage{
glue_stop_session(Id, RequestOrigin = NULL)
}
\arguments{
\item{Id}{[required] The ID of the session to be stopped.}

\item{RequestOrigin}{The origin of the request.}
}
\description{
Stops the session.

See \url{https://paws-r.github.io/docs/glue/stop_session.html} for full documentation.
}
\keyword{internal}
