% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_stop_trigger}
\alias{glue_stop_trigger}
\title{Stops a specified trigger}
\usage{
glue_stop_trigger(Name)
}
\arguments{
\item{Name}{[required] The name of the trigger to stop.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Name = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Stops a specified trigger.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$stop_trigger(
  Name = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
