% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_update_data_quality_ruleset}
\alias{glue_update_data_quality_ruleset}
\title{Updates the specified data quality ruleset}
\usage{
glue_update_data_quality_ruleset(Name, Description = NULL, Ruleset = NULL)
}
\arguments{
\item{Name}{[required] The name of the data quality ruleset.}

\item{Description}{A description of the ruleset.}

\item{Ruleset}{A Data Quality Definition Language (DQDL) ruleset. For more information,
see the Glue developer guide.}
}
\description{
Updates the specified data quality ruleset.

See \url{https://paws-r.github.io/docs/glue/update_data_quality_ruleset.html} for full documentation.
}
\keyword{internal}
