% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_update_trigger}
\alias{glue_update_trigger}
\title{Updates a trigger definition}
\usage{
glue_update_trigger(Name, TriggerUpdate)
}
\arguments{
\item{Name}{[required] The name of the trigger to update.}

\item{TriggerUpdate}{[required] The new values with which to update the trigger.}
}
\description{
Updates a trigger definition.

See \url{https://paws-r.github.io/docs/glue/update_trigger.html} for full documentation.
}
\keyword{internal}
