% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluedatabrew_operations.R
\name{gluedatabrew_create_dataset}
\alias{gluedatabrew_create_dataset}
\title{Creates a new DataBrew dataset}
\usage{
gluedatabrew_create_dataset(
  Name,
  Format = NULL,
  FormatOptions = NULL,
  Input,
  PathOptions = NULL,
  Tags = NULL
)
}
\arguments{
\item{Name}{[required] The name of the dataset to be created. Valid characters are alphanumeric
(A-Z, a-z, 0-9), hyphen (-), period (.), and space.}

\item{Format}{The file format of a dataset that is created from an Amazon S3 file or
folder.}

\item{FormatOptions}{}

\item{Input}{[required]}

\item{PathOptions}{A set of options that defines how DataBrew interprets an Amazon S3 path
of the dataset.}

\item{Tags}{Metadata tags to apply to this dataset.}
}
\description{
Creates a new DataBrew dataset.

See \url{https://www.paws-r-sdk.com/docs/gluedatabrew_create_dataset/} for full documentation.
}
\keyword{internal}
