% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluedatabrew_operations.R
\name{gluedatabrew_describe_project}
\alias{gluedatabrew_describe_project}
\title{Returns the definition of a specific DataBrew project}
\usage{
gluedatabrew_describe_project(Name)
}
\arguments{
\item{Name}{[required] The name of the project to be described.}
}
\description{
Returns the definition of a specific DataBrew project.

See \url{https://www.paws-r-sdk.com/docs/gluedatabrew_describe_project/} for full documentation.
}
\keyword{internal}
