% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluedatabrew_operations.R
\name{gluedatabrew_describe_schedule}
\alias{gluedatabrew_describe_schedule}
\title{Returns the definition of a specific DataBrew schedule}
\usage{
gluedatabrew_describe_schedule(Name)
}
\arguments{
\item{Name}{[required] The name of the schedule to be described.}
}
\description{
Returns the definition of a specific DataBrew schedule.

See \url{https://www.paws-r-sdk.com/docs/gluedatabrew_describe_schedule/} for full documentation.
}
\keyword{internal}
