% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluedatabrew_operations.R
\name{gluedatabrew_list_datasets}
\alias{gluedatabrew_list_datasets}
\title{Lists all of the DataBrew datasets}
\usage{
gluedatabrew_list_datasets(MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{The maximum number of results to return in this request.}

\item{NextToken}{The token returned by a previous call to retrieve the next set of
results.}
}
\description{
Lists all of the DataBrew datasets.

See \url{https://www.paws-r-sdk.com/docs/gluedatabrew_list_datasets/} for full documentation.
}
\keyword{internal}
