% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluedatabrew_operations.R
\name{gluedatabrew_list_tags_for_resource}
\alias{gluedatabrew_list_tags_for_resource}
\title{Lists all the tags for a DataBrew resource}
\usage{
gluedatabrew_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) string that uniquely identifies the
DataBrew resource.}
}
\description{
Lists all the tags for a DataBrew resource.

See \url{https://paws-r.github.io/docs/gluedatabrew/list_tags_for_resource.html} for full documentation.
}
\keyword{internal}
