% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluedatabrew_operations.R
\name{gluedatabrew_start_job_run}
\alias{gluedatabrew_start_job_run}
\title{Runs a DataBrew job}
\usage{
gluedatabrew_start_job_run(Name)
}
\arguments{
\item{Name}{[required] The name of the job to be run.}
}
\description{
Runs a DataBrew job.

See \url{https://paws-r.github.io/docs/gluedatabrew/start_job_run.html} for full documentation.
}
\keyword{internal}
