% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthlake_operations.R
\name{healthlake_delete_fhir_datastore}
\alias{healthlake_delete_fhir_datastore}
\title{Delete a FHIR-enabled data store}
\usage{
healthlake_delete_fhir_datastore(DatastoreId)
}
\arguments{
\item{DatastoreId}{[required] The AWS-generated identifier for the data store to be deleted.}
}
\description{
Delete a FHIR-enabled data store.

See \url{https://www.paws-r-sdk.com/docs/healthlake_delete_fhir_datastore/} for full documentation.
}
\keyword{internal}
