% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthlake_operations.R
\name{healthlake_describe_fhir_datastore}
\alias{healthlake_describe_fhir_datastore}
\title{Gets the properties associated with the FHIR Data Store, including the
Data Store ID, Data Store ARN, Data Store name, Data Store status,
created at, Data Store type version, and Data Store endpoint}
\usage{
healthlake_describe_fhir_datastore(DatastoreId)
}
\arguments{
\item{DatastoreId}{[required] The AWS-generated Data Store ID.}
}
\description{
Gets the properties associated with the FHIR Data Store, including the Data Store ID, Data Store ARN, Data Store name, Data Store status, created at, Data Store type version, and Data Store endpoint.

See \url{https://www.paws-r-sdk.com/docs/healthlake_describe_fhir_datastore/} for full documentation.
}
\keyword{internal}
