% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthlake_operations.R
\name{healthlake_describe_fhir_datastore}
\alias{healthlake_describe_fhir_datastore}
\title{Gets the properties associated with the FHIR data store, including the
data store ID, data store ARN, data store name, data store status, when
the data store was created, data store type version, and the data
store's endpoint}
\usage{
healthlake_describe_fhir_datastore(DatastoreId)
}
\arguments{
\item{DatastoreId}{[required] The AWS-generated data store ID.}
}
\description{
Gets the properties associated with the FHIR data store, including the data store ID, data store ARN, data store name, data store status, when the data store was created, data store type version, and the data store's endpoint.

See \url{https://www.paws-r-sdk.com/docs/healthlake_describe_fhir_datastore/} for full documentation.
}
\keyword{internal}
