% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthlake_operations.R
\name{healthlake_describe_fhir_datastore}
\alias{healthlake_describe_fhir_datastore}
\title{Get properties for a FHIR-enabled data store}
\usage{
healthlake_describe_fhir_datastore(DatastoreId)
}
\arguments{
\item{DatastoreId}{[required] The data store identifier.}
}
\description{
Get properties for a FHIR-enabled data store.

See \url{https://www.paws-r-sdk.com/docs/healthlake_describe_fhir_datastore/} for full documentation.
}
\keyword{internal}
