% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthlake_operations.R
\name{healthlake_describe_fhir_export_job}
\alias{healthlake_describe_fhir_export_job}
\title{Get FHIR export job properties}
\usage{
healthlake_describe_fhir_export_job(DatastoreId, JobId)
}
\arguments{
\item{DatastoreId}{[required] The data store identifier from which FHIR data is being exported from.}

\item{JobId}{[required] The export job identifier.}
}
\description{
Get FHIR export job properties.

See \url{https://www.paws-r-sdk.com/docs/healthlake_describe_fhir_export_job/} for full documentation.
}
\keyword{internal}
