% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthlake_operations.R
\name{healthlake_describe_fhir_import_job}
\alias{healthlake_describe_fhir_import_job}
\title{Get the import job properties to learn more about the job or job
progress}
\usage{
healthlake_describe_fhir_import_job(DatastoreId, JobId)
}
\arguments{
\item{DatastoreId}{[required] The data store identifier.}

\item{JobId}{[required] The import job identifier.}
}
\description{
Get the import job properties to learn more about the job or job progress.

See \url{https://www.paws-r-sdk.com/docs/healthlake_describe_fhir_import_job/} for full documentation.
}
\keyword{internal}
