% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthlake_operations.R
\name{healthlake_list_fhir_datastores}
\alias{healthlake_list_fhir_datastores}
\title{Lists all FHIR data stores that are in the user’s account, regardless of
data store status}
\usage{
healthlake_list_fhir_datastores(
  Filter = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{Filter}{Lists all filters associated with a FHIR data store request.}

\item{NextToken}{Fetches the next page of data stores when results are paginated.}

\item{MaxResults}{The maximum number of data stores returned in a single page of a
ListFHIRDatastoresRequest call.}
}
\description{
Lists all FHIR data stores that are in the user’s account, regardless of data store status.

See \url{https://www.paws-r-sdk.com/docs/healthlake_list_fhir_datastores/} for full documentation.
}
\keyword{internal}
