% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthlake_operations.R
\name{healthlake_list_fhir_datastores}
\alias{healthlake_list_fhir_datastores}
\title{List all FHIR-enabled data stores in a user’s account, regardless of
data store status}
\usage{
healthlake_list_fhir_datastores(
  Filter = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{Filter}{List all filters associated with a FHIR data store request.}

\item{NextToken}{The token used to retrieve the next page of data stores when results are
paginated.}

\item{MaxResults}{The maximum number of data stores returned on a page.}
}
\description{
List all FHIR-enabled data stores in a user’s account, regardless of data store status.

See \url{https://www.paws-r-sdk.com/docs/healthlake_list_fhir_datastores/} for full documentation.
}
\keyword{internal}
