% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthlake_operations.R
\name{healthlake_start_fhir_export_job}
\alias{healthlake_start_fhir_export_job}
\title{Start a FHIR export job}
\usage{
healthlake_start_fhir_export_job(
  JobName = NULL,
  OutputDataConfig,
  DatastoreId,
  DataAccessRoleArn,
  ClientToken = NULL
)
}
\arguments{
\item{JobName}{The export job name.}

\item{OutputDataConfig}{[required] The output data configuration supplied when the export job was started.}

\item{DatastoreId}{[required] The data store identifier from which files are being exported.}

\item{DataAccessRoleArn}{[required] The Amazon Resource Name (ARN) used during initiation of the export job.}

\item{ClientToken}{An optional user provided token used for ensuring API idempotency.}
}
\description{
Start a FHIR export job.

See \url{https://www.paws-r-sdk.com/docs/healthlake_start_fhir_export_job/} for full documentation.
}
\keyword{internal}
