% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthlake_operations.R
\name{healthlake_start_fhir_import_job}
\alias{healthlake_start_fhir_import_job}
\title{Begins a FHIR Import job}
\usage{
healthlake_start_fhir_import_job(
  JobName = NULL,
  InputDataConfig,
  JobOutputDataConfig,
  DatastoreId,
  DataAccessRoleArn,
  ClientToken
)
}
\arguments{
\item{JobName}{The name of the FHIR Import job in the StartFHIRImport job request.}

\item{InputDataConfig}{[required] The input properties of the FHIR Import job in the StartFHIRImport job
request.}

\item{JobOutputDataConfig}{[required]}

\item{DatastoreId}{[required] The AWS-generated Data Store ID.}

\item{DataAccessRoleArn}{[required] The Amazon Resource Name (ARN) that gives Amazon HealthLake access
permission.}

\item{ClientToken}{[required] Optional user provided token used for ensuring idempotency.}
}
\description{
Begins a FHIR Import job.

See \url{https://www.paws-r-sdk.com/docs/healthlake_start_fhir_import_job/} for full documentation.
}
\keyword{internal}
