% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthlake_operations.R
\name{healthlake_start_fhir_import_job}
\alias{healthlake_start_fhir_import_job}
\title{Start importing bulk FHIR data into an ACTIVE data store}
\usage{
healthlake_start_fhir_import_job(
  JobName = NULL,
  InputDataConfig,
  JobOutputDataConfig,
  DatastoreId,
  DataAccessRoleArn,
  ClientToken = NULL,
  ValidationLevel = NULL
)
}
\arguments{
\item{JobName}{The import job name.}

\item{InputDataConfig}{[required] The input properties for the import job request.}

\item{JobOutputDataConfig}{[required]}

\item{DatastoreId}{[required] The data store identifier.}

\item{DataAccessRoleArn}{[required] The Amazon Resource Name (ARN) that grants access permission to AWS
HealthLake.}

\item{ClientToken}{The optional user-provided token used for ensuring API idempotency.}

\item{ValidationLevel}{The validation level of the import job.}
}
\description{
Start importing bulk FHIR data into an ACTIVE data store. The import job imports FHIR data found in the \code{InputDataConfig} object and stores processing results in the \code{JobOutputDataConfig} object.

See \url{https://www.paws-r-sdk.com/docs/healthlake_start_fhir_import_job/} for full documentation.
}
\keyword{internal}
