% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthlake_operations.R
\name{healthlake_untag_resource}
\alias{healthlake_untag_resource}
\title{Remove a user-specifed key and value tag from a data store}
\usage{
healthlake_untag_resource(ResourceARN, TagKeys)
}
\arguments{
\item{ResourceARN}{[required] The Amazon Resource Name (ARN) of the data store from which tags are
being removed.}

\item{TagKeys}{[required] The keys for the tags to be removed from the data store.}
}
\description{
Remove a user-specifed key and value tag from a data store.

See \url{https://www.paws-r-sdk.com/docs/healthlake_untag_resource/} for full documentation.
}
\keyword{internal}
