% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivs_operations.R
\name{ivs_create_ad_configuration}
\alias{ivs_create_ad_configuration}
\title{Creates a new ad configuration to be used for server-side ad insertion}
\usage{
ivs_create_ad_configuration(
  name = NULL,
  mediaTailorPlaybackConfigurations,
  tags = NULL
)
}
\arguments{
\item{name}{Ad configuration name. Defaults to “”.}

\item{mediaTailorPlaybackConfigurations}{[required] List of integration configurations with media tailor resources.}

\item{tags}{Array of 1-50 maps, each of the form \code{string:string (key:value)}. See
\href{https://docs.aws.amazon.com/tag-editor/latest/userguide/best-practices-and-strats.html}{Best practices and strategies}
in \emph{Tagging Amazon Web Services Resources and Tag Editor} for details,
including restrictions that apply to tags and "Tag naming limits and
requirements"; Amazon IVS has no service-specific constraints beyond
what is documented there.}
}
\description{
Creates a new ad configuration to be used for server-side ad insertion.

See \url{https://www.paws-r-sdk.com/docs/ivs_create_ad_configuration/} for full documentation.
}
\keyword{internal}
