% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivs_operations.R
\name{ivs_create_channel}
\alias{ivs_create_channel}
\title{Creates a new channel and an associated stream key to start streaming}
\usage{
ivs_create_channel(
  authorized = NULL,
  insecureIngest = NULL,
  latencyMode = NULL,
  name = NULL,
  recordingConfigurationArn = NULL,
  tags = NULL,
  type = NULL
)
}
\arguments{
\item{authorized}{Whether the channel is private (enabled for playback authorization).
Default: \code{false}.}

\item{insecureIngest}{Whether the channel allows insecure RTMP ingest. Default: \code{false}.}

\item{latencyMode}{Channel latency mode. Use \code{NORMAL} to broadcast and deliver live video
up to Full HD. Use \code{LOW} for near-real-time interaction with viewers.
(Note: In the Amazon IVS console, \code{LOW} and \code{NORMAL} correspond to
Ultra-low and Standard, respectively.) Default: \code{LOW}.}

\item{name}{Channel name.}

\item{recordingConfigurationArn}{Recording-configuration ARN. Default: "" (empty string, recording is
disabled).}

\item{tags}{Array of 1-50 maps, each of the form \code{string:string (key:value)}. See
\href{https://docs.aws.amazon.com/tag-editor/latest/userguide/tagging.html}{Tagging Amazon Web Services Resources}
for more information, including restrictions that apply to tags and "Tag
naming limits and requirements"; Amazon IVS has no service-specific
constraints beyond what is documented there.}

\item{type}{Channel type, which determines the allowable resolution and bitrate. \emph{If
you exceed the allowable resolution or bitrate, the stream probably will
disconnect immediately.} Default: \code{STANDARD}. Valid values:
\itemize{
\item \code{STANDARD}: Video is transcoded: multiple qualities are generated
from the original input, to automatically give viewers the best
experience for their devices and network conditions. Transcoding
allows higher playback quality across a range of download speeds.
Resolution can be up to 1080p and bitrate can be up to 8.5 Mbps.
Audio is transcoded only for renditions 360p and below; above that,
audio is passed through. This is the default.
\item \code{BASIC}: Video is transmuxed: Amazon IVS delivers the original input
to viewers. The viewer’s video-quality choice is limited to the
original input. Resolution can be up to 1080p and bitrate can be up
to 1.5 Mbps for 480p and up to 3.5 Mbps for resolutions between 480p
and 1080p.
}}
}
\description{
Creates a new channel and an associated stream key to start streaming.

See \url{https://paws-r.github.io/docs/ivs/create_channel.html} for full documentation.
}
\keyword{internal}
