% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivs_operations.R
\name{ivs_get_stream_key}
\alias{ivs_get_stream_key}
\title{Gets stream-key information for a specified ARN}
\usage{
ivs_get_stream_key(arn)
}
\arguments{
\item{arn}{[required] ARN for the stream key to be retrieved.}
}
\description{
Gets stream-key information for a specified ARN.

See \url{https://www.paws-r-sdk.com/docs/ivs_get_stream_key/} for full documentation.
}
\keyword{internal}
