% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivs_operations.R
\name{ivs_insert_ad_break}
\alias{ivs_insert_ad_break}
\title{Inserts an ad marker in the playlist for the specified channel and
duration using the ad configuration associated with the channel}
\usage{
ivs_insert_ad_break(channelArn, durationSeconds)
}
\arguments{
\item{channelArn}{[required] ARN of the channel into which the ad break is inserted.}

\item{durationSeconds}{[required] Maximum duration of the ad break, in seconds.}
}
\description{
Inserts an ad marker in the playlist for the specified channel and duration using the ad configuration associated with the channel.

See \url{https://www.paws-r-sdk.com/docs/ivs_insert_ad_break/} for full documentation.
}
\keyword{internal}
