% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivs_operations.R
\name{ivs_list_playback_key_pairs}
\alias{ivs_list_playback_key_pairs}
\title{Gets summary information about playback key pairs}
\usage{
ivs_list_playback_key_pairs(maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{maxResults}{Maximum number of key pairs to return. Default: your service quota or
100, whichever is smaller.}

\item{nextToken}{The first key pair to retrieve. This is used for pagination; see the
\code{nextToken} response field.}
}
\description{
Gets summary information about playback key pairs. For more information, see \href{https://docs.aws.amazon.com/ivs/latest/LowLatencyUserGuide/private-channels.html}{Setting Up Private Channels} in the \emph{Amazon IVS User Guide}.

See \url{https://www.paws-r-sdk.com/docs/ivs_list_playback_key_pairs/} for full documentation.
}
\keyword{internal}
