% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivs_operations.R
\name{ivs_list_recording_configurations}
\alias{ivs_list_recording_configurations}
\title{Gets summary information about all recording configurations in your
account, in the Amazon Web Services region where the API request is
processed}
\usage{
ivs_list_recording_configurations(maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{maxResults}{Maximum number of recording configurations to return. Default: your
service quota or 100, whichever is smaller.}

\item{nextToken}{The first recording configuration to retrieve. This is used for
pagination; see the \code{nextToken} response field.}
}
\description{
Gets summary information about all recording configurations in your account, in the Amazon Web Services region where the API request is processed.

See \url{https://www.paws-r-sdk.com/docs/ivs_list_recording_configurations/} for full documentation.
}
\keyword{internal}
