% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivs_operations.R
\name{ivs_tag_resource}
\alias{ivs_tag_resource}
\title{Adds or updates tags for the Amazon Web Services resource with the
specified ARN}
\usage{
ivs_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] ARN of the resource for which tags are to be added or updated. The ARN
must be URL-encoded.}

\item{tags}{[required] Array of tags to be added or updated. Array of maps, each of the form
\code{string:string (key:value)}. See \href{https://docs.aws.amazon.com/tag-editor/latest/userguide/tagging.html}{Tagging Amazon Web Services Resources}
for more information, including restrictions that apply to tags and "Tag
naming limits and requirements"; Amazon IVS has no service-specific
constraints beyond what is documented there.}
}
\description{
Adds or updates tags for the Amazon Web Services resource with the specified ARN.

See \url{https://paws-r.github.io/docs/ivs/tag_resource.html} for full documentation.
}
\keyword{internal}
