% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivs_operations.R
\name{ivs_untag_resource}
\alias{ivs_untag_resource}
\title{Removes tags from the resource with the specified ARN}
\usage{
ivs_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] ARN of the resource for which tags are to be removed. The ARN must be
URL-encoded.}

\item{tagKeys}{[required] Array of tags to be removed. Array of maps, each of the form
s\code{tring:string (key:value)}. See \href{https://docs.aws.amazon.com/tag-editor/latest/userguide/tagging.html}{Tagging Amazon Web Services Resources}
for more information, including restrictions that apply to tags and "Tag
naming limits and requirements"; Amazon IVS has no service-specific
constraints beyond what is documented there.}
}
\description{
Removes tags from the resource with the specified ARN.

See \url{https://www.paws-r-sdk.com/docs/ivs_untag_resource/} for full documentation.
}
\keyword{internal}
