% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivsrealtime_operations.R
\name{ivsrealtime_list_compositions}
\alias{ivsrealtime_list_compositions}
\title{Gets summary information about all Compositions in your account, in the
AWS region where the API request is processed}
\usage{
ivsrealtime_list_compositions(
  filterByStageArn = NULL,
  filterByEncoderConfigurationArn = NULL,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{filterByStageArn}{Filters the Composition list to match the specified Stage ARN.}

\item{filterByEncoderConfigurationArn}{Filters the Composition list to match the specified EncoderConfiguration
attached to at least one of its output.}

\item{nextToken}{The first Composition to retrieve. This is used for pagination; see the
\code{nextToken} response field.}

\item{maxResults}{Maximum number of results to return. Default: 100.}
}
\description{
Gets summary information about all Compositions in your account, in the AWS region where the API request is processed.

See \url{https://www.paws-r-sdk.com/docs/ivsrealtime_list_compositions/} for full documentation.
}
\keyword{internal}
