% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivsrealtime_operations.R
\name{ivsrealtime_list_stages}
\alias{ivsrealtime_list_stages}
\title{Gets summary information about all stages in your account, in the AWS
region where the API request is processed}
\usage{
ivsrealtime_list_stages(nextToken = NULL, maxResults = NULL)
}
\arguments{
\item{nextToken}{The first stage to retrieve. This is used for pagination; see the
\code{nextToken} response field.}

\item{maxResults}{Maximum number of results to return. Default: 50.}
}
\description{
Gets summary information about all stages in your account, in the AWS region where the API request is processed.

See \url{https://www.paws-r-sdk.com/docs/ivsrealtime_list_stages/} for full documentation.
}
\keyword{internal}
