% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivsrealtime_operations.R
\name{ivsrealtime_update_stage}
\alias{ivsrealtime_update_stage}
\title{Updates a stage’s configuration}
\usage{
ivsrealtime_update_stage(arn, name = NULL)
}
\arguments{
\item{arn}{[required] ARN of the stage to be updated.}

\item{name}{Name of the stage to be updated.}
}
\description{
Updates a stage’s configuration.

See \url{https://www.paws-r-sdk.com/docs/ivsrealtime_update_stage/} for full documentation.
}
\keyword{internal}
