% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_batch_associate_scram_secret}
\alias{kafka_batch_associate_scram_secret}
\title{Associates one or more Scram Secrets with an Amazon MSK cluster}
\usage{
kafka_batch_associate_scram_secret(ClusterArn, SecretArnList)
}
\arguments{
\item{ClusterArn}{[required] The Amazon Resource Name (ARN) of the cluster to be updated.}

\item{SecretArnList}{[required] List of AWS Secrets Manager secret ARNs.}
}
\description{
Associates one or more Scram Secrets with an Amazon MSK cluster.

See \url{https://paws-r.github.io/docs/kafka/batch_associate_scram_secret.html} for full documentation.
}
\keyword{internal}
