% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_delete_cluster}
\alias{kafka_delete_cluster}
\title{Deletes the MSK cluster specified by the Amazon Resource Name (ARN) in
the request}
\usage{
kafka_delete_cluster(ClusterArn, CurrentVersion)
}
\arguments{
\item{ClusterArn}{[required] The Amazon Resource Name (ARN) that uniquely identifies the cluster.}

\item{CurrentVersion}{The current version of the MSK cluster.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ClusterArn = "string",
  State = "ACTIVE"|"CREATING"|"DELETING"|"FAILED"|"HEALING"|"MAINTENANCE"|"REBOOTING_BROKER"|"UPDATING"
)
}\if{html}{\out{</div>}}
}
\description{
Deletes the MSK cluster specified by the Amazon Resource Name (ARN) in
the request.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_cluster(
  ClusterArn = "string",
  CurrentVersion = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
