% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_delete_cluster}
\alias{kafka_delete_cluster}
\title{Deletes the MSK cluster specified by the Amazon Resource Name (ARN) in
the request}
\usage{
kafka_delete_cluster(ClusterArn, CurrentVersion)
}
\arguments{
\item{ClusterArn}{[required] The Amazon Resource Name (ARN) that uniquely identifies the cluster.}

\item{CurrentVersion}{The current version of the MSK cluster.}
}
\description{
Deletes the MSK cluster specified by the Amazon Resource Name (ARN) in
the request.
}
\section{Request syntax}{
\preformatted{svc$delete_cluster(
  ClusterArn = "string",
  CurrentVersion = "string"
)
}
}

\keyword{internal}
