% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_delete_configuration}
\alias{kafka_delete_configuration}
\title{Deletes an MSK Configuration}
\usage{
kafka_delete_configuration(Arn)
}
\arguments{
\item{Arn}{[required]
<p>The Amazon Resource Name (ARN) that uniquely identifies an MSK configuration.</p>}
}
\description{
\preformatted{        <p>Deletes an MSK Configuration.</p>
}
}
\section{Request syntax}{
\preformatted{svc$delete_configuration(
  Arn = "string"
)
}
}

\keyword{internal}
