% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_delete_configuration}
\alias{kafka_delete_configuration}
\title{Deletes an MSK Configuration}
\usage{
kafka_delete_configuration(Arn)
}
\arguments{
\item{Arn}{[required] The Amazon Resource Name (ARN) that uniquely identifies an MSK
configuration.}
}
\description{
Deletes an MSK Configuration.

See \url{https://www.paws-r-sdk.com/docs/kafka_delete_configuration/} for full documentation.
}
\keyword{internal}
