% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_delete_vpc_connection}
\alias{kafka_delete_vpc_connection}
\title{Deletes a MSK VPC connection}
\usage{
kafka_delete_vpc_connection(Arn)
}
\arguments{
\item{Arn}{[required] The Amazon Resource Name (ARN) that uniquely identifies an MSK VPC
connection.}
}
\description{
Deletes a MSK VPC connection.

See \url{https://www.paws-r-sdk.com/docs/kafka_delete_vpc_connection/} for full documentation.
}
\keyword{internal}
