% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_describe_cluster}
\alias{kafka_describe_cluster}
\title{Returns a description of the MSK cluster whose Amazon Resource Name
(ARN) is specified in the request}
\usage{
kafka_describe_cluster(ClusterArn)
}
\arguments{
\item{ClusterArn}{[required] The Amazon Resource Name (ARN) that uniquely identifies the cluster.}
}
\description{
Returns a description of the MSK cluster whose Amazon Resource Name (ARN) is specified in the request.

See \url{https://paws-r.github.io/docs/kafka/describe_cluster.html} for full documentation.
}
\keyword{internal}
