% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_describe_configuration}
\alias{kafka_describe_configuration}
\title{Returns a description of this MSK configuration}
\usage{
kafka_describe_configuration(Arn)
}
\arguments{
\item{Arn}{[required]
<p>The Amazon Resource Name (ARN) that uniquely identifies an MSK configuration and all of its revisions.</p>}
}
\description{
<p>Returns a description of this MSK configuration.</p>
}
\section{Request syntax}{
\preformatted{svc$describe_configuration(
  Arn = "string"
)
}
}

\keyword{internal}
