% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_describe_vpc_connection}
\alias{kafka_describe_vpc_connection}
\title{Returns a description of this MSK VPC connection}
\usage{
kafka_describe_vpc_connection(Arn)
}
\arguments{
\item{Arn}{[required] The Amazon Resource Name (ARN) that uniquely identifies a MSK VPC
connection.}
}
\description{
Returns a description of this MSK VPC connection.

See \url{https://www.paws-r-sdk.com/docs/kafka_describe_vpc_connection/} for full documentation.
}
\keyword{internal}
