% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_get_bootstrap_brokers}
\alias{kafka_get_bootstrap_brokers}
\title{A list of brokers that a client application can use to bootstrap}
\usage{
kafka_get_bootstrap_brokers(ClusterArn)
}
\arguments{
\item{ClusterArn}{[required] The Amazon Resource Name (ARN) that uniquely identifies the cluster.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  BootstrapBrokerString = "string",
  BootstrapBrokerStringTls = "string",
  BootstrapBrokerStringSaslScram = "string"
)
}\if{html}{\out{</div>}}
}
\description{
A list of brokers that a client application can use to bootstrap.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_bootstrap_brokers(
  ClusterArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
