% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_get_compatible_kafka_versions}
\alias{kafka_get_compatible_kafka_versions}
\title{Gets the Apache Kafka versions to which you can update the MSK cluster}
\usage{
kafka_get_compatible_kafka_versions(ClusterArn)
}
\arguments{
\item{ClusterArn}{The Amazon Resource Name (ARN) of the cluster check.}
}
\value{
A list with the following syntax:\preformatted{list(
  CompatibleKafkaVersions = list(
    list(
      SourceVersion = "string",
      TargetVersions = list(
        "string"
      )
    )
  )
)
}
}
\description{
Gets the Apache Kafka versions to which you can update the MSK cluster.
}
\section{Request syntax}{
\preformatted{svc$get_compatible_kafka_versions(
  ClusterArn = "string"
)
}
}

\keyword{internal}
