% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_list_configuration_revisions}
\alias{kafka_list_configuration_revisions}
\title{Returns a list of all the MSK configurations in this Region}
\usage{
kafka_list_configuration_revisions(Arn, MaxResults, NextToken)
}
\arguments{
\item{Arn}{[required]
<p>The Amazon Resource Name (ARN) that uniquely identifies an MSK configuration and all of its revisions.</p>}

\item{MaxResults}{<p>The maximum number of results to return in the response. If there are more results, the response includes a NextToken parameter.</p>}

\item{NextToken}{<p>The paginated results marker. When the result of the operation is truncated, the call returns NextToken in the response. 
To get the next batch, provide this token in your next request.</p>}
}
\description{
<p>Returns a list of all the MSK configurations in this Region.</p>
}
\section{Request syntax}{
\preformatted{svc$list_configuration_revisions(
  Arn = "string",
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
