% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_list_scram_secrets}
\alias{kafka_list_scram_secrets}
\title{Returns a list of the Scram Secrets associated with an Amazon MSK
cluster}
\usage{
kafka_list_scram_secrets(ClusterArn, MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{ClusterArn}{[required] The arn of the cluster.}

\item{MaxResults}{The maxResults of the query.}

\item{NextToken}{The nextToken of the query.}
}
\description{
Returns a list of the Scram Secrets associated with an Amazon MSK cluster.

See \url{https://paws-r.github.io/docs/kafka/list_scram_secrets.html} for full documentation.
}
\keyword{internal}
