% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_list_tags_for_resource}
\alias{kafka_list_tags_for_resource}
\title{Returns a list of the tags associated with the specified resource}
\usage{
kafka_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) that uniquely identifies the resource
that's associated with the tags.}
}
\value{
A list with the following syntax:\preformatted{list(
  Tags = list(
    "string"
  )
)
}
}
\description{
Returns a list of the tags associated with the specified resource.
}
\section{Request syntax}{
\preformatted{svc$list_tags_for_resource(
  ResourceArn = "string"
)
}
}

\keyword{internal}
