% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_reboot_broker}
\alias{kafka_reboot_broker}
\title{Reboots brokers}
\usage{
kafka_reboot_broker(BrokerIds, ClusterArn)
}
\arguments{
\item{BrokerIds}{[required]
<p>The list of broker IDs to be rebooted. The reboot-broker operation supports rebooting one broker at a time.</p>}

\item{ClusterArn}{[required]
<p>The Amazon Resource Name (ARN) of the cluster to be updated.</p>}
}
\description{
Reboots brokers.
}
\section{Request syntax}{
\preformatted{svc$reboot_broker(
  BrokerIds = list(
    "string"
  ),
  ClusterArn = "string"
)
}
}

\keyword{internal}
