% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_reboot_broker}
\alias{kafka_reboot_broker}
\title{Reboots brokers}
\usage{
kafka_reboot_broker(BrokerIds, ClusterArn)
}
\arguments{
\item{BrokerIds}{[required] The list of broker IDs to be rebooted. The reboot-broker operation
supports rebooting one broker at a time.}

\item{ClusterArn}{[required] The Amazon Resource Name (ARN) of the cluster to be updated.}
}
\description{
Reboots brokers.

See \url{https://www.paws-r-sdk.com/docs/kafka_reboot_broker/} for full documentation.
}
\keyword{internal}
