% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_reject_client_vpc_connection}
\alias{kafka_reject_client_vpc_connection}
\title{Returns empty response}
\usage{
kafka_reject_client_vpc_connection(ClusterArn, VpcConnectionArn)
}
\arguments{
\item{ClusterArn}{[required] The Amazon Resource Name (ARN) of the cluster.}

\item{VpcConnectionArn}{[required] The VPC connection ARN.}
}
\description{
Returns empty response.

See \url{https://www.paws-r-sdk.com/docs/kafka_reject_client_vpc_connection/} for full documentation.
}
\keyword{internal}
