% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_tag_resource}
\alias{kafka_tag_resource}
\title{Adds tags to the specified MSK resource}
\usage{
kafka_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) that uniquely identifies the resource
that's associated with the tags.}

\item{Tags}{[required] The key-value pair for the resource tag.}
}
\value{
An empty list.
}
\description{
Adds tags to the specified MSK resource.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$tag_resource(
  ResourceArn = "string",
  Tags = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
