% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_update_broker_count}
\alias{kafka_update_broker_count}
\title{Updates the number of broker nodes in the cluster}
\usage{
kafka_update_broker_count(
  ClusterArn,
  CurrentVersion,
  TargetNumberOfBrokerNodes
)
}
\arguments{
\item{ClusterArn}{[required] The Amazon Resource Name (ARN) that uniquely identifies the cluster.}

\item{CurrentVersion}{[required] The version of cluster to update from. A successful operation will then
generate a new version.}

\item{TargetNumberOfBrokerNodes}{[required] The number of broker nodes that you want the cluster to have after this
operation completes successfully.}
}
\description{
Updates the number of broker nodes in the cluster.

See \url{https://paws-r.github.io/docs/kafka/update_broker_count.html} for full documentation.
}
\keyword{internal}
