% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_update_broker_count}
\alias{kafka_update_broker_count}
\title{Updates the number of broker nodes in the cluster}
\usage{
kafka_update_broker_count(ClusterArn, CurrentVersion,
  TargetNumberOfBrokerNodes)
}
\arguments{
\item{ClusterArn}{[required]
<p>The Amazon Resource Name (ARN) that uniquely identifies the cluster.</p>}

\item{CurrentVersion}{[required]
<p>The version of cluster to update from. A successful operation will then generate a new version.</p>}

\item{TargetNumberOfBrokerNodes}{[required]
<p>The number of broker nodes that you want the cluster to have after this operation completes successfully.</p>}
}
\description{
\preformatted{        <p>Updates the number of broker nodes in the cluster.</p>
}
}
\section{Request syntax}{
\preformatted{svc$update_broker_count(
  ClusterArn = "string",
  CurrentVersion = "string",
  TargetNumberOfBrokerNodes = 123
)
}
}

\keyword{internal}
