% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_update_cluster_kafka_version}
\alias{kafka_update_cluster_kafka_version}
\title{Updates the Apache Kafka version for the cluster}
\usage{
kafka_update_cluster_kafka_version(
  ClusterArn,
  ConfigurationInfo = NULL,
  CurrentVersion,
  TargetKafkaVersion
)
}
\arguments{
\item{ClusterArn}{[required] The Amazon Resource Name (ARN) of the cluster to be updated.}

\item{ConfigurationInfo}{The custom configuration that should be applied on the new version of
cluster.}

\item{CurrentVersion}{[required] Current cluster version.}

\item{TargetKafkaVersion}{[required] Target Kafka version.}
}
\description{
Updates the Apache Kafka version for the cluster.

See \url{https://paws-r.github.io/docs/kafka/update_cluster_kafka_version.html} for full documentation.
}
\keyword{internal}
