% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafkaconnect_operations.R
\name{kafkaconnect_list_worker_configurations}
\alias{kafkaconnect_list_worker_configurations}
\title{Returns a list of all of the worker configurations in this account and
Region}
\usage{
kafkaconnect_list_worker_configurations(
  maxResults = NULL,
  namePrefix = NULL,
  nextToken = NULL
)
}
\arguments{
\item{maxResults}{The maximum number of worker configurations to list in one response.}

\item{namePrefix}{Lists worker configuration names that start with the specified text
string.}

\item{nextToken}{If the response of a ListWorkerConfigurations operation is truncated, it
will include a NextToken. Send this NextToken in a subsequent request to
continue listing from where the previous operation left off.}
}
\description{
Returns a list of all of the worker configurations in this account and Region.

See \url{https://www.paws-r-sdk.com/docs/kafkaconnect_list_worker_configurations/} for full documentation.
}
\keyword{internal}
