% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendra_operations.R
\name{kendra_clear_query_suggestions}
\alias{kendra_clear_query_suggestions}
\title{Clears existing query suggestions from an index}
\usage{
kendra_clear_query_suggestions(IndexId)
}
\arguments{
\item{IndexId}{[required] The identifier of the index you want to clear query suggestions from.}
}
\description{
Clears existing query suggestions from an index.

See \url{https://www.paws-r-sdk.com/docs/kendra_clear_query_suggestions/} for full documentation.
}
\keyword{internal}
