% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendra_operations.R
\name{kendra_delete_faq}
\alias{kendra_delete_faq}
\title{Removes a FAQ from an index}
\usage{
kendra_delete_faq(Id, IndexId)
}
\arguments{
\item{Id}{[required] The identifier of the FAQ you want to remove.}

\item{IndexId}{[required] The identifier of the index for the FAQ.}
}
\description{
Removes a FAQ from an index.

See \url{https://www.paws-r-sdk.com/docs/kendra_delete_faq/} for full documentation.
}
\keyword{internal}
