% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendra_operations.R
\name{kendra_describe_query_suggestions_config}
\alias{kendra_describe_query_suggestions_config}
\title{Gets information on the settings of query suggestions for an index}
\usage{
kendra_describe_query_suggestions_config(IndexId)
}
\arguments{
\item{IndexId}{[required] The identifier of the index with query suggestions that you want to get
information on.}
}
\description{
Gets information on the settings of query suggestions for an index.

See \url{https://www.paws-r-sdk.com/docs/kendra_describe_query_suggestions_config/} for full documentation.
}
\keyword{internal}
