% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendra_operations.R
\name{kendra_disassociate_personas_from_entities}
\alias{kendra_disassociate_personas_from_entities}
\title{Removes the specific permissions of users or groups in your IAM Identity
Center identity source with access to your Amazon Kendra experience}
\usage{
kendra_disassociate_personas_from_entities(Id, IndexId, EntityIds)
}
\arguments{
\item{Id}{[required] The identifier of your Amazon Kendra experience.}

\item{IndexId}{[required] The identifier of the index for your Amazon Kendra experience.}

\item{EntityIds}{[required] The identifiers of users or groups in your IAM Identity Center identity
source. For example, user IDs could be user emails.}
}
\description{
Removes the specific permissions of users or groups in your IAM Identity Center identity source with access to your Amazon Kendra experience. You can create an Amazon Kendra experience such as a search application. For more information on creating a search application experience, see \href{https://docs.aws.amazon.com/kendra/latest/dg/deploying-search-experience-no-code.html}{Building a search experience with no code}.

See \url{https://paws-r.github.io/docs/kendra/disassociate_personas_from_entities.html} for full documentation.
}
\keyword{internal}
