% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendra_operations.R
\name{kendra_list_data_source_sync_jobs}
\alias{kendra_list_data_source_sync_jobs}
\title{Gets statistics about synchronizing Amazon Kendra with a data source}
\usage{
kendra_list_data_source_sync_jobs(
  Id,
  IndexId,
  NextToken = NULL,
  MaxResults = NULL,
  StartTimeFilter = NULL,
  StatusFilter = NULL
)
}
\arguments{
\item{Id}{[required] The identifier of the data source.}

\item{IndexId}{[required] The identifier of the index used with the data source.}

\item{NextToken}{If the previous response was incomplete (because there is more data to
retrieve), Amazon Kendra returns a pagination token in the response. You
can use this pagination token to retrieve the next set of jobs.}

\item{MaxResults}{The maximum number of synchronization jobs to return in the response. If
there are fewer results in the list, this response contains only the
actual results.}

\item{StartTimeFilter}{When specified, the synchronization jobs returned in the list are
limited to jobs between the specified dates.}

\item{StatusFilter}{When specified, only returns synchronization jobs with the \code{Status}
field equal to the specified status.}
}
\description{
Gets statistics about synchronizing Amazon Kendra with a data source.

See \url{https://paws-r.github.io/docs/kendra/list_data_source_sync_jobs.html} for full documentation.
}
\keyword{internal}
