% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendra_operations.R
\name{kendra_list_tags_for_resource}
\alias{kendra_list_tags_for_resource}
\title{Gets a list of tags associated with a resource}
\usage{
kendra_list_tags_for_resource(ResourceARN)
}
\arguments{
\item{ResourceARN}{[required] The Amazon Resource Name (ARN) of the index, FAQ, data source, or other
resource to get a list of tags for. For example, the ARN of an index is
constructed as follows:
\emph{arn:aws:kendra:your-region:your-account-id:index/index-id} For
information on how to construct an ARN for all types of Amazon Kendra
resources, see \href{https://docs.aws.amazon.com/service-authorization/latest/reference/list_amazonkendra.html#amazonkendra-resources-for-iam-policies}{Resource types}.}
}
\description{
Gets a list of tags associated with a resource. Indexes, FAQs, data sources, and other resources can have tags associated with them.

See \url{https://www.paws-r-sdk.com/docs/kendra_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
