% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendra_operations.R
\name{kendra_start_data_source_sync_job}
\alias{kendra_start_data_source_sync_job}
\title{Starts a synchronization job for a data source}
\usage{
kendra_start_data_source_sync_job(Id, IndexId)
}
\arguments{
\item{Id}{[required] The identifier of the data source to synchronize.}

\item{IndexId}{[required] The identifier of the index that contains the data source.}
}
\description{
Starts a synchronization job for a data source. If a synchronization job is already in progress, Amazon Kendra returns a \code{ResourceInUseException} exception.

See \url{https://paws-r.github.io/docs/kendra/start_data_source_sync_job.html} for full documentation.
}
\keyword{internal}
