% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendra_operations.R
\name{kendra_tag_resource}
\alias{kendra_tag_resource}
\title{Adds the specified tag to the specified index, FAQ, or data source
resource}
\usage{
kendra_tag_resource(ResourceARN, Tags)
}
\arguments{
\item{ResourceARN}{[required] The Amazon Resource Name (ARN) of the index, FAQ, or data source to tag.}

\item{Tags}{[required] A list of tag keys to add to the index, FAQ, or data source. If a tag
already exists, the existing value is replaced with the new value.}
}
\description{
Adds the specified tag to the specified index, FAQ, or data source resource. If the tag already exists, the existing value is replaced with the new value.

See \url{https://paws-r.github.io/docs/kendra/tag_resource.html} for full documentation.
}
\keyword{internal}
