% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendra_operations.R
\name{kendra_untag_resource}
\alias{kendra_untag_resource}
\title{Removes a tag from an index, FAQ, data source, or other resource}
\usage{
kendra_untag_resource(ResourceARN, TagKeys)
}
\arguments{
\item{ResourceARN}{[required] The Amazon Resource Name (ARN) of the index, FAQ, data source, or other
resource to remove a tag. For example, the ARN of an index is
constructed as follows:
\emph{arn:aws:kendra:your-region:your-account-id:index/index-id} For
information on how to construct an ARN for all types of Amazon Kendra
resources, see \href{https://docs.aws.amazon.com/service-authorization/latest/reference/list_amazonkendra.html#amazonkendra-resources-for-iam-policies}{Resource types}.}

\item{TagKeys}{[required] A list of tag keys to remove from the index, FAQ, data source, or other
resource. If a tag key doesn't exist for the resource, it is ignored.}
}
\description{
Removes a tag from an index, FAQ, data source, or other resource.

See \url{https://www.paws-r-sdk.com/docs/kendra_untag_resource/} for full documentation.
}
\keyword{internal}
