% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesis_operations.R
\name{kinesis_add_tags_to_stream}
\alias{kinesis_add_tags_to_stream}
\title{Adds or updates tags for the specified Kinesis data stream}
\usage{
kinesis_add_tags_to_stream(StreamName, Tags)
}
\arguments{
\item{StreamName}{[required] The name of the stream.}

\item{Tags}{[required] A set of up to 10 key-value pairs to use to create the tags.}
}
\description{
Adds or updates tags for the specified Kinesis data stream. You can assign up to 50 tags to a data stream.

See \url{https://paws-r.github.io/docs/kinesis/add_tags_to_stream.html} for full documentation.
}
\keyword{internal}
