% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesis_operations.R
\name{kinesis_create_channel}
\alias{kinesis_create_channel}
\title{Creates a channel that delivers records from a Kinesis data stream to a
destination}
\usage{
kinesis_create_channel(
  ChannelName,
  ServiceExecutionRoleARN,
  StreamConfigurationList,
  S3DestinationConfiguration = NULL,
  S3TablesDestinationConfiguration = NULL,
  EncryptionConfiguration = NULL,
  Tags = NULL,
  LoggingConfiguration = NULL
)
}
\arguments{
\item{ChannelName}{[required] The name of the channel. The name is unique within your Amazon Web Services account and Amazon Web Services Region.}

\item{ServiceExecutionRoleARN}{[required] The Amazon Resource Name (ARN) of the IAM role that Amazon Kinesis Data Streams assumes to write records to the destination.}

\item{StreamConfigurationList}{[required] The source stream configuration for the channel. Currently, one stream is supported per channel.}

\item{S3DestinationConfiguration}{The configuration for delivery to a general purpose Amazon S3 bucket. You must specify either \code{S3DestinationConfiguration} or \code{S3TablesDestinationConfiguration}, but not both.}

\item{S3TablesDestinationConfiguration}{The configuration for delivery to streaming tables on Apache Iceberg in Amazon S3 Tables. You must specify either \code{S3DestinationConfiguration} or \code{S3TablesDestinationConfiguration}, but not both.}

\item{EncryptionConfiguration}{The server-side encryption configuration that uses an Amazon Web Services KMS key to encrypt data delivered to the destination.}

\item{Tags}{A set of key-value pairs to assign to the channel. A tag consists of a required key and an optional value.}

\item{LoggingConfiguration}{The Amazon CloudWatch Logs configuration for the channel.}
}
\description{
Creates a channel that delivers records from a Kinesis data stream to a destination. A channel reads records from the specified stream and writes them to streaming tables on Apache Iceberg (Amazon S3 Tables) or to a general purpose Amazon S3 bucket.

See \url{https://www.paws-r-sdk.com/docs/kinesis_create_channel/} for full documentation.
}
\keyword{internal}
