% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesis_operations.R
\name{kinesis_describe_limits}
\alias{kinesis_describe_limits}
\title{Describes the shard limits and usage for the account}
\usage{
kinesis_describe_limits()
}
\description{
Describes the shard limits and usage for the account.

See \url{https://www.paws-r-sdk.com/docs/kinesis_describe_limits/} for full documentation.
}
\keyword{internal}
