% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesis_operations.R
\name{kinesis_describe_limits}
\alias{kinesis_describe_limits}
\title{Describes the shard limits and usage for the account}
\usage{
kinesis_describe_limits()
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ShardLimit = 123,
  OpenShardCount = 123
)
}\if{html}{\out{</div>}}
}
\description{
Describes the shard limits and usage for the account.

If you update your account limits, the old limits might be returned for
a few minutes.

This operation has a limit of one transaction per second per account.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_limits()
}\if{html}{\out{</div>}}
}

\keyword{internal}
