% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesis_operations.R
\name{kinesis_describe_limits}
\alias{kinesis_describe_limits}
\title{Describes the shard limits and usage for the account}
\usage{
kinesis_describe_limits()
}
\description{
Describes the shard limits and usage for the account.
}
\details{
If you update your account limits, the old limits might be returned for
a few minutes.

This operation has a limit of one transaction per second per account.
}
\section{Request syntax}{
\preformatted{svc$describe_limits()
}
}

\keyword{internal}
