% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesis_operations.R
\name{kinesis_subscribe_to_shard}
\alias{kinesis_subscribe_to_shard}
\title{This operation establishes an HTTP/2 connection between the consumer you
specify in the ConsumerARN parameter and the shard you specify in the
ShardId parameter}
\usage{
kinesis_subscribe_to_shard(ConsumerARN, ShardId, StartingPosition)
}
\arguments{
\item{ConsumerARN}{[required] For this parameter, use the value you obtained when you called
\code{\link[=kinesis_register_stream_consumer]{register_stream_consumer}}.}

\item{ShardId}{[required] The ID of the shard you want to subscribe to. To see a list of all the
shards for a given stream, use \code{\link[=kinesis_list_shards]{list_shards}}.}

\item{StartingPosition}{[required] The starting position in the data stream from which to start streaming.}
}
\description{
This operation establishes an HTTP/2 connection between the consumer you specify in the \code{ConsumerARN} parameter and the shard you specify in the \code{ShardId} parameter. After the connection is successfully established, Kinesis Data Streams pushes records from the shard to the consumer over this connection. Before you call this operation, call \code{\link[=kinesis_register_stream_consumer]{register_stream_consumer}} to register the consumer with Kinesis Data Streams.

See \url{https://www.paws-r-sdk.com/docs/kinesis_subscribe_to_shard/} for full documentation.
}
\keyword{internal}
