% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesis_operations.R
\name{kinesis_update_stream_mode}
\alias{kinesis_update_stream_mode}
\title{Updates the capacity mode of the data stream}
\usage{
kinesis_update_stream_mode(
  StreamARN,
  StreamModeDetails,
  WarmThroughputMiBps = NULL
)
}
\arguments{
\item{StreamARN}{[required] Specifies the ARN of the data stream whose capacity mode you want to
update.}

\item{StreamModeDetails}{[required] Specifies the capacity mode to which you want to set your data stream.
Currently, in Kinesis Data Streams, you can choose between an
\strong{on-demand} capacity mode and a \strong{provisioned} capacity mode for your
data streams.}

\item{WarmThroughputMiBps}{The target warm throughput in MB/s that the stream should be scaled to
handle. This represents the throughput capacity that will be immediately
available for write operations. This field is only valid when the stream
mode is being updated to on-demand.}
}
\description{
Updates the capacity mode of the data stream. Currently, in Kinesis Data Streams, you can choose between an \strong{on-demand} capacity mode and a \strong{provisioned} capacity mode for your data stream.

See \url{https://www.paws-r-sdk.com/docs/kinesis_update_stream_mode/} for full documentation.
}
\keyword{internal}
