% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesis_operations.R
\name{kinesis_update_stream_warm_throughput}
\alias{kinesis_update_stream_warm_throughput}
\title{Updates the warm throughput configuration for the specified Amazon
Kinesis Data Streams on-demand data stream}
\usage{
kinesis_update_stream_warm_throughput(
  StreamARN = NULL,
  StreamName = NULL,
  WarmThroughputMiBps
)
}
\arguments{
\item{StreamARN}{The ARN of the stream to be updated.}

\item{StreamName}{The name of the stream to be updated.}

\item{WarmThroughputMiBps}{[required] The target warm throughput in MB/s that the stream should be scaled to
handle. This represents the throughput capacity that will be immediately
available for write operations.}
}
\description{
Updates the warm throughput configuration for the specified Amazon Kinesis Data Streams on-demand data stream. This operation allows you to proactively scale your on-demand data stream to a specified throughput level, enabling better performance for sudden traffic spikes.

See \url{https://www.paws-r-sdk.com/docs/kinesis_update_stream_warm_throughput/} for full documentation.
}
\keyword{internal}
