% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisanalytics_operations.R
\name{kinesisanalytics_add_application_input}
\alias{kinesisanalytics_add_application_input}
\title{This documentation is for version 1 of the Amazon Kinesis Data Analytics
API, which only supports SQL applications}
\usage{
kinesisanalytics_add_application_input(
  ApplicationName,
  CurrentApplicationVersionId,
  Input
)
}
\arguments{
\item{ApplicationName}{[required] Name of your existing Amazon Kinesis Analytics application to which you
want to add the streaming source.}

\item{CurrentApplicationVersionId}{[required] Current version of your Amazon Kinesis Analytics application. You can
use the \code{\link[=kinesisanalytics_describe_application]{describe_application}}
operation to find the current application version.}

\item{Input}{[required] The
\href{https://docs.aws.amazon.com/kinesisanalytics/latest/dev/API_Input.html}{Input}
to add.}
}
\description{
This documentation is for version 1 of the Amazon Kinesis Data Analytics API, which only supports SQL applications. Version 2 of the API supports SQL and Java applications. For more information about version 2, see Amazon Kinesis Data Analytics API V2 Documentation.

See \url{https://www.paws-r-sdk.com/docs/kinesisanalytics_add_application_input/} for full documentation.
}
\keyword{internal}
