% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisanalytics_operations.R
\name{kinesisanalytics_describe_application}
\alias{kinesisanalytics_describe_application}
\title{This documentation is for version 1 of the Amazon Kinesis Data Analytics
API, which only supports SQL applications}
\usage{
kinesisanalytics_describe_application(ApplicationName)
}
\arguments{
\item{ApplicationName}{[required] Name of the application.}
}
\description{
This documentation is for version 1 of the Amazon Kinesis Data Analytics API, which only supports SQL applications. Version 2 of the API supports SQL and Java applications. For more information about version 2, see Amazon Kinesis Data Analytics API V2 Documentation.

See \url{https://www.paws-r-sdk.com/docs/kinesisanalytics_describe_application/} for full documentation.
}
\keyword{internal}
