% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisanalytics_operations.R
\name{kinesisanalytics_list_applications}
\alias{kinesisanalytics_list_applications}
\title{This documentation is for version 1 of the Amazon Kinesis Data Analytics
API, which only supports SQL applications}
\usage{
kinesisanalytics_list_applications(
  Limit = NULL,
  ExclusiveStartApplicationName = NULL
)
}
\arguments{
\item{Limit}{Maximum number of applications to list.}

\item{ExclusiveStartApplicationName}{Name of the application to start the list with. When using pagination to
retrieve the list, you don't need to specify this parameter in the first
request. However, in subsequent requests, you add the last application
name from the previous response to get the next page of applications.}
}
\description{
This documentation is for version 1 of the Amazon Kinesis Data Analytics API, which only supports SQL applications. Version 2 of the API supports SQL and Java applications. For more information about version 2, see Amazon Kinesis Data Analytics API V2 Documentation.

See \url{https://paws-r.github.io/docs/kinesisanalytics/list_applications.html} for full documentation.
}
\keyword{internal}
