% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisanalytics_operations.R
\name{kinesisanalytics_list_tags_for_resource}
\alias{kinesisanalytics_list_tags_for_resource}
\title{Retrieves the list of key-value tags assigned to the application}
\usage{
kinesisanalytics_list_tags_for_resource(ResourceARN)
}
\arguments{
\item{ResourceARN}{[required] The ARN of the application for which to retrieve tags.}
}
\description{
Retrieves the list of key-value tags assigned to the application. For more information, see \href{https://docs.aws.amazon.com/kinesisanalytics/latest/dev/how-tagging.html}{Using Tagging}.

See \url{https://www.paws-r-sdk.com/docs/kinesisanalytics_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
