% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisanalytics_operations.R
\name{kinesisanalytics_untag_resource}
\alias{kinesisanalytics_untag_resource}
\title{Removes one or more tags from a Kinesis Analytics application}
\usage{
kinesisanalytics_untag_resource(ResourceARN, TagKeys)
}
\arguments{
\item{ResourceARN}{[required] The ARN of the Kinesis Analytics application from which to remove the
tags.}

\item{TagKeys}{[required] A list of keys of tags to remove from the specified application.}
}
\description{
Removes one or more tags from a Kinesis Analytics application. For more information, see \href{https://docs.aws.amazon.com/kinesisanalytics/latest/dev/how-tagging.html}{Using Tagging}.

See \url{https://paws-r.github.io/docs/kinesisanalytics/untag_resource.html} for full documentation.
}
\keyword{internal}
