% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisanalyticsv2_operations.R
\name{kinesisanalyticsv2_delete_application_vpc_configuration}
\alias{kinesisanalyticsv2_delete_application_vpc_configuration}
\title{Removes a VPC configuration from a Kinesis Data Analytics application}
\usage{
kinesisanalyticsv2_delete_application_vpc_configuration(ApplicationName,
  CurrentApplicationVersionId, VpcConfigurationId)
}
\arguments{
\item{ApplicationName}{[required] The name of an existing application.}

\item{CurrentApplicationVersionId}{[required] The current application version ID. You can retrieve the application
version ID using
\code{\link[=kinesisanalyticsv2_describe_application]{describe_application}}.}

\item{VpcConfigurationId}{[required] The ID of the VPC configuration to delete.}
}
\value{
A list with the following syntax:\preformatted{list(
  ApplicationARN = "string",
  ApplicationVersionId = 123
)
}
}
\description{
Removes a VPC configuration from a Kinesis Data Analytics application.
}
\section{Request syntax}{
\preformatted{svc$delete_application_vpc_configuration(
  ApplicationName = "string",
  CurrentApplicationVersionId = 123,
  VpcConfigurationId = "string"
)
}
}

\keyword{internal}
