% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisanalyticsv2_operations.R
\name{kinesisanalyticsv2_describe_application_operation}
\alias{kinesisanalyticsv2_describe_application_operation}
\title{Provides a detailed description of a specified application operation}
\usage{
kinesisanalyticsv2_describe_application_operation(ApplicationName, OperationId)
}
\arguments{
\item{ApplicationName}{[required]}

\item{OperationId}{[required]}
}
\description{
Provides a detailed description of a specified application operation. To see a list of all the operations of an application, invoke the \code{\link[=kinesisanalyticsv2_list_application_operations]{list_application_operations}} operation.

See \url{https://www.paws-r-sdk.com/docs/kinesisanalyticsv2_describe_application_operation/} for full documentation.
}
\keyword{internal}
