% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisanalyticsv2_operations.R
\name{kinesisanalyticsv2_describe_application_snapshot}
\alias{kinesisanalyticsv2_describe_application_snapshot}
\title{Returns information about a snapshot of application state data}
\usage{
kinesisanalyticsv2_describe_application_snapshot(ApplicationName,
  SnapshotName)
}
\arguments{
\item{ApplicationName}{[required] The name of an existing application.}

\item{SnapshotName}{[required] The identifier of an application snapshot. You can retrieve this value
using .}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  SnapshotDetails = list(
    SnapshotName = "string",
    SnapshotStatus = "CREATING"|"READY"|"DELETING"|"FAILED",
    ApplicationVersionId = 123,
    SnapshotCreationTimestamp = as.POSIXct(
      "2015-01-01"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Returns information about a snapshot of application state data.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_application_snapshot(
  ApplicationName = "string",
  SnapshotName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
