% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisanalyticsv2_operations.R
\name{kinesisanalyticsv2_discover_input_schema}
\alias{kinesisanalyticsv2_discover_input_schema}
\title{Infers a schema for a SQL-based Kinesis Data Analytics application by
evaluating sample records on the specified streaming source (Kinesis
data stream or Kinesis Data Firehose delivery stream) or Amazon S3
object}
\usage{
kinesisanalyticsv2_discover_input_schema(
  ResourceARN = NULL,
  ServiceExecutionRole,
  InputStartingPositionConfiguration = NULL,
  S3Configuration = NULL,
  InputProcessingConfiguration = NULL
)
}
\arguments{
\item{ResourceARN}{The Amazon Resource Name (ARN) of the streaming source.}

\item{ServiceExecutionRole}{[required] The ARN of the role that is used to access the streaming source.}

\item{InputStartingPositionConfiguration}{The point at which you want Kinesis Data Analytics to start reading
records from the specified streaming source discovery purposes.}

\item{S3Configuration}{Specify this parameter to discover a schema from data in an Amazon S3
object.}

\item{InputProcessingConfiguration}{The InputProcessingConfiguration to use to preprocess the records before
discovering the schema of the records.}
}
\description{
Infers a schema for a SQL-based Kinesis Data Analytics application by evaluating sample records on the specified streaming source (Kinesis data stream or Kinesis Data Firehose delivery stream) or Amazon S3 object. In the response, the operation returns the inferred schema and also the sample records that the operation used to infer the schema.

See \url{https://paws-r.github.io/docs/kinesisanalyticsv2/discover_input_schema.html} for full documentation.
}
\keyword{internal}
