% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisanalyticsv2_operations.R
\name{kinesisanalyticsv2_list_application_operations}
\alias{kinesisanalyticsv2_list_application_operations}
\title{Lists information about operations performed on a Managed Service for
Apache Flink application}
\usage{
kinesisanalyticsv2_list_application_operations(
  ApplicationName,
  Limit = NULL,
  NextToken = NULL,
  Operation = NULL,
  OperationStatus = NULL
)
}
\arguments{
\item{ApplicationName}{[required]}

\item{Limit}{}

\item{NextToken}{}

\item{Operation}{}

\item{OperationStatus}{}
}
\description{
Lists information about operations performed on a Managed Service for Apache Flink application

See \url{https://www.paws-r-sdk.com/docs/kinesisanalyticsv2_list_application_operations/} for full documentation.
}
\keyword{internal}
