% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisanalyticsv2_operations.R
\name{kinesisanalyticsv2_list_application_snapshots}
\alias{kinesisanalyticsv2_list_application_snapshots}
\title{Lists information about the current application snapshots}
\usage{
kinesisanalyticsv2_list_application_snapshots(
  ApplicationName,
  Limit = NULL,
  NextToken = NULL
)
}
\arguments{
\item{ApplicationName}{[required] The name of an existing application.}

\item{Limit}{The maximum number of application snapshots to list.}

\item{NextToken}{Use this parameter if you receive a \code{NextToken} response in a previous
request that indicates that there is more output available. Set it to
the value of the previous call's \code{NextToken} response to indicate where
the output should continue from.}
}
\description{
Lists information about the current application snapshots.

See \url{https://paws-r.github.io/docs/kinesisanalyticsv2/list_application_snapshots.html} for full documentation.
}
\keyword{internal}
