% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisanalyticsv2_operations.R
\name{kinesisanalyticsv2_list_tags_for_resource}
\alias{kinesisanalyticsv2_list_tags_for_resource}
\title{Retrieves the list of key-value tags assigned to the application}
\usage{
kinesisanalyticsv2_list_tags_for_resource(ResourceARN)
}
\arguments{
\item{ResourceARN}{[required] The ARN of the application for which to retrieve tags.}
}
\description{
Retrieves the list of key-value tags assigned to the application. For more information, see \href{https://docs.aws.amazon.com/kinesisanalytics/latest/java/how-tagging.html}{Using Tagging}.

See \url{https://paws-r.github.io/docs/kinesisanalyticsv2/list_tags_for_resource.html} for full documentation.
}
\keyword{internal}
