% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisanalyticsv2_operations.R
\name{kinesisanalyticsv2_rollback_application}
\alias{kinesisanalyticsv2_rollback_application}
\title{Reverts the application to the previous running version}
\usage{
kinesisanalyticsv2_rollback_application(
  ApplicationName,
  CurrentApplicationVersionId
)
}
\arguments{
\item{ApplicationName}{[required] The name of the application.}

\item{CurrentApplicationVersionId}{[required] The current application version ID. You can retrieve the application
version ID using
\code{\link[=kinesisanalyticsv2_describe_application]{describe_application}}.}
}
\description{
Reverts the application to the previous running version. You can roll back an application if you suspect it is stuck in a transient status or in the running status.

See \url{https://www.paws-r-sdk.com/docs/kinesisanalyticsv2_rollback_application/} for full documentation.
}
\keyword{internal}
