% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisanalyticsv2_operations.R
\name{kinesisanalyticsv2_tag_resource}
\alias{kinesisanalyticsv2_tag_resource}
\title{Adds one or more key-value tags to a Kinesis Data Analytics application}
\usage{
kinesisanalyticsv2_tag_resource(ResourceARN, Tags)
}
\arguments{
\item{ResourceARN}{[required] The ARN of the application to assign the tags.}

\item{Tags}{[required] The key-value tags to assign to the application.}
}
\value{
An empty list.
}
\description{
Adds one or more key-value tags to a Kinesis Data Analytics application.
Note that the maximum number of application tags includes system tags.
The maximum number of user-defined application tags is 50. For more
information, see \href{https://docs.aws.amazon.com/kinesisanalytics/latest/java/how-tagging.html}{Using Tagging}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$tag_resource(
  ResourceARN = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
