% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisanalyticsv2_operations.R
\name{kinesisanalyticsv2_update_application_maintenance_configuration}
\alias{kinesisanalyticsv2_update_application_maintenance_configuration}
\title{Updates the maintenance configuration of the Kinesis Data Analytics
application}
\usage{
kinesisanalyticsv2_update_application_maintenance_configuration(
  ApplicationName,
  ApplicationMaintenanceConfigurationUpdate
)
}
\arguments{
\item{ApplicationName}{[required] The name of the application for which you want to update the maintenance
configuration.}

\item{ApplicationMaintenanceConfigurationUpdate}{[required] Describes the application maintenance configuration update.}
}
\description{
Updates the maintenance configuration of the Kinesis Data Analytics application.

See \url{https://paws-r.github.io/docs/kinesisanalyticsv2/update_application_maintenance_configuration.html} for full documentation.
}
\keyword{internal}
