% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mturk_operations.R
\name{mturk_accept_qualification_request}
\alias{mturk_accept_qualification_request}
\title{The AcceptQualificationRequest operation approves a Worker's request for
a Qualification}
\usage{
mturk_accept_qualification_request(QualificationRequestId, IntegerValue)
}
\arguments{
\item{QualificationRequestId}{[required] The ID of the Qualification request, as returned by the
\code{GetQualificationRequests} operation.}

\item{IntegerValue}{The value of the Qualification. You can omit this value if you are using
the presence or absence of the Qualification as the basis for a HIT
requirement.}
}
\value{
An empty list.
}
\description{
The \code{\link[=mturk_accept_qualification_request]{accept_qualification_request}}
operation approves a Worker's request for a Qualification.

Only the owner of the Qualification type can grant a Qualification
request for that type.

A successful request for the
\code{\link[=mturk_accept_qualification_request]{accept_qualification_request}}
operation returns with no errors and an empty body.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$accept_qualification_request(
  QualificationRequestId = "string",
  IntegerValue = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
