% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mturk_operations.R
\name{mturk_get_qualification_score}
\alias{mturk_get_qualification_score}
\title{The GetQualificationScore operation returns the value of a Worker's
Qualification for a given Qualification type}
\usage{
mturk_get_qualification_score(QualificationTypeId, WorkerId)
}
\arguments{
\item{QualificationTypeId}{[required] The ID of the QualificationType.}

\item{WorkerId}{[required] The ID of the Worker whose Qualification is being updated.}
}
\description{
The \code{\link[=mturk_get_qualification_score]{get_qualification_score}} operation returns the value of a Worker's Qualification for a given Qualification type.

See \url{https://paws-r.github.io/docs/mturk/get_qualification_score.html} for full documentation.
}
\keyword{internal}
