% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mturk_operations.R
\name{mturk_get_qualification_type}
\alias{mturk_get_qualification_type}
\title{The GetQualificationTypeoperation retrieves information about a
Qualification type using its ID}
\usage{
mturk_get_qualification_type(QualificationTypeId)
}
\arguments{
\item{QualificationTypeId}{[required] The ID of the QualificationType.}
}
\description{
The \code{\link[=mturk_get_qualification_type]{get_qualification_type}}operation retrieves information about a Qualification type using its ID.

See \url{https://www.paws-r-sdk.com/docs/mturk_get_qualification_type/} for full documentation.
}
\keyword{internal}
