% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mturk_operations.R
\name{mturk_list_hi_ts}
\alias{mturk_list_hi_ts}
\title{The ListHITs operation returns all of a Requester's HITs}
\usage{
mturk_list_hi_ts(NextToken, MaxResults)
}
\arguments{
\item{NextToken}{Pagination token}

\item{MaxResults}{}
}
\description{
The \code{ListHITs} operation returns all of a Requester\'s HITs. The
operation returns HITs of any status, except for HITs that have been
deleted of with the DeleteHIT operation or that have been auto-deleted.
}
\section{Request syntax}{
\preformatted{svc$list_hi_ts(
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
