% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mturk_operations.R
\name{mturk_list_hi_ts}
\alias{mturk_list_hi_ts}
\title{The ListHITs operation returns all of a Requester's HITs}
\usage{
mturk_list_hi_ts(NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{NextToken}{Pagination token}

\item{MaxResults}{}
}
\description{
The \code{\link[=mturk_list_hi_ts]{list_hi_ts}} operation returns all of a Requester's HITs. The operation returns HITs of any status, except for HITs that have been deleted of with the DeleteHIT operation or that have been auto-deleted.

See \url{https://www.paws-r-sdk.com/docs/mturk_list_hi_ts/} for full documentation.
}
\keyword{internal}
