% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mturk_operations.R
\name{mturk_list_qualification_requests}
\alias{mturk_list_qualification_requests}
\title{The ListQualificationRequests operation retrieves requests for
Qualifications of a particular Qualification type}
\usage{
mturk_list_qualification_requests(QualificationTypeId, NextToken,
  MaxResults)
}
\arguments{
\item{QualificationTypeId}{The ID of the QualificationType.}

\item{NextToken}{}

\item{MaxResults}{The maximum number of results to return in a single call.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  NumResults = 123,
  NextToken = "string",
  QualificationRequests = list(
    list(
      QualificationRequestId = "string",
      QualificationTypeId = "string",
      WorkerId = "string",
      Test = "string",
      Answer = "string",
      SubmitTime = as.POSIXct(
        "2015-01-01"
      )
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
The \code{\link[=mturk_list_qualification_requests]{list_qualification_requests}}
operation retrieves requests for Qualifications of a particular
Qualification type. The owner of the Qualification type calls this
operation to poll for pending requests, and accepts them using the
AcceptQualification operation.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_qualification_requests(
  QualificationTypeId = "string",
  NextToken = "string",
  MaxResults = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
