% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mturk_operations.R
\name{mturk_list_reviewable_hi_ts}
\alias{mturk_list_reviewable_hi_ts}
\title{The ListReviewableHITs operation retrieves the HITs with Status equal to
Reviewable or Status equal to Reviewing that belong to the Requester
calling the operation}
\usage{
mturk_list_reviewable_hi_ts(HITTypeId, Status, NextToken, MaxResults)
}
\arguments{
\item{HITTypeId}{The ID of the HIT type of the HITs to consider for the query. If not
specified, all HITs for the Reviewer are considered}

\item{Status}{Can be either \code{Reviewable} or \code{Reviewing}. Reviewable is the default
value.}

\item{NextToken}{Pagination Token}

\item{MaxResults}{Limit the number of results returned.}
}
\description{
The \code{\link[=mturk_list_reviewable_hi_ts]{list_reviewable_hi_ts}} operation
retrieves the HITs with Status equal to Reviewable or Status equal to
Reviewing that belong to the Requester calling the operation.
}
\section{Request syntax}{
\preformatted{svc$list_reviewable_hi_ts(
  HITTypeId = "string",
  Status = "Reviewable"|"Reviewing",
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
