% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mturk_operations.R
\name{mturk_update_expiration_for_hit}
\alias{mturk_update_expiration_for_hit}
\title{The UpdateExpirationForHIT operation allows you update the expiration
time of a HIT}
\usage{
mturk_update_expiration_for_hit(HITId, ExpireAt)
}
\arguments{
\item{HITId}{[required] The HIT to update.}

\item{ExpireAt}{[required] The date and time at which you want the HIT to expire}
}
\description{
The \code{\link[=mturk_update_expiration_for_hit]{update_expiration_for_hit}} operation allows you update the expiration time of a HIT. If you update it to a time in the past, the HIT will be immediately expired.

See \url{https://paws-r.github.io/docs/mturk/update_expiration_for_hit.html} for full documentation.
}
\keyword{internal}
