% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_add_tags}
\alias{opensearchservice_add_tags}
\title{Attaches tags to an existing Amazon OpenSearch Service domain, data
source, or application}
\usage{
opensearchservice_add_tags(ARN, TagList)
}
\arguments{
\item{ARN}{[required] Amazon Resource Name (ARN) for the OpenSearch Service domain, data
source, or application to which you want to attach resource tags.}

\item{TagList}{[required] List of resource tags.}
}
\description{
Attaches tags to an existing Amazon OpenSearch Service domain, data source, or application.

See \url{https://www.paws-r-sdk.com/docs/opensearchservice_add_tags/} for full documentation.
}
\keyword{internal}
