% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_authorize_vpc_endpoint_access}
\alias{opensearchservice_authorize_vpc_endpoint_access}
\title{Provides access to an Amazon OpenSearch Service domain through the use
of an interface VPC endpoint}
\usage{
opensearchservice_authorize_vpc_endpoint_access(DomainName, Account)
}
\arguments{
\item{DomainName}{[required] The name of the OpenSearch Service domain to provide access to.}

\item{Account}{[required] The Amazon Web Services account ID to grant access to.}
}
\description{
Provides access to an Amazon OpenSearch Service domain through the use of an interface VPC endpoint.

See \url{https://paws-r.github.io/docs/opensearchservice/authorize_vpc_endpoint_access.html} for full documentation.
}
\keyword{internal}
