% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_cancel_domain_config_change}
\alias{opensearchservice_cancel_domain_config_change}
\title{Cancels a pending configuration change on an Amazon OpenSearch Service
domain}
\usage{
opensearchservice_cancel_domain_config_change(DomainName, DryRun = NULL)
}
\arguments{
\item{DomainName}{[required]}

\item{DryRun}{When set to \code{True}, returns the list of change IDs and properties that
will be cancelled without actually cancelling the change.}
}
\description{
Cancels a pending configuration change on an Amazon OpenSearch Service domain.

See \url{https://www.paws-r-sdk.com/docs/opensearchservice_cancel_domain_config_change/} for full documentation.
}
\keyword{internal}
