% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_create_application}
\alias{opensearchservice_create_application}
\title{Creates an OpenSearch UI application}
\usage{
opensearchservice_create_application(
  clientToken = NULL,
  name,
  dataSources = NULL,
  iamIdentityCenterOptions = NULL,
  appConfigs = NULL,
  tagList = NULL,
  kmsKeyArn = NULL
)
}
\arguments{
\item{clientToken}{Unique, case-sensitive identifier to ensure idempotency of the request.}

\item{name}{[required] The unique name of the OpenSearch application. Names must be unique
within an Amazon Web Services Region for each account.}

\item{dataSources}{The data sources to link to the OpenSearch application.}

\item{iamIdentityCenterOptions}{Configuration settings for integrating Amazon Web Services IAM Identity
Center with the OpenSearch application.}

\item{appConfigs}{Configuration settings for the OpenSearch application, including
administrative options.}

\item{tagList}{}

\item{kmsKeyArn}{The Amazon Resource Name (ARN) of the KMS key used to encrypt the
application's data at rest. If provided, the application uses your
customer-managed key for encryption. If omitted, the application uses an
AWS-managed key. The KMS key must be in the same region as the
application.}
}
\description{
Creates an OpenSearch UI application. For more information, see \href{https://docs.aws.amazon.com/opensearch-service/latest/developerguide/application.html}{Using the OpenSearch user interface in Amazon OpenSearch Service}.

See \url{https://www.paws-r-sdk.com/docs/opensearchservice_create_application/} for full documentation.
}
\keyword{internal}
