% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_create_outbound_connection}
\alias{opensearchservice_create_outbound_connection}
\title{Creates a new cross-cluster connection from a local OpenSearch domain to
a remote OpenSearch domain}
\usage{
opensearchservice_create_outbound_connection(
  LocalDomainInfo,
  RemoteDomainInfo,
  ConnectionAlias
)
}
\arguments{
\item{LocalDomainInfo}{[required] The \code{AWSDomainInformation} for the local OpenSearch domain.}

\item{RemoteDomainInfo}{[required] The \code{AWSDomainInformation} for the remote OpenSearch domain.}

\item{ConnectionAlias}{[required] The connection alias used used by the customer for this cross-cluster
connection.}
}
\description{
Creates a new cross-cluster connection from a local OpenSearch domain to a remote OpenSearch domain.

See \url{https://paws-r.github.io/docs/opensearchservice/create_outbound_connection.html} for full documentation.
}
\keyword{internal}
