% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_create_package}
\alias{opensearchservice_create_package}
\title{Create a package for use with Amazon OpenSearch Service domains}
\usage{
opensearchservice_create_package(
  PackageName,
  PackageType,
  PackageDescription = NULL,
  PackageSource
)
}
\arguments{
\item{PackageName}{[required] Unique identifier for the package.}

\item{PackageType}{[required] Type of package. Currently supports only TXT-DICTIONARY.}

\item{PackageDescription}{Description of the package.}

\item{PackageSource}{[required] The Amazon S3 location from which to import the package.}
}
\description{
Create a package for use with Amazon OpenSearch Service domains.

See \url{https://paws-r.github.io/docs/opensearchservice/create_package.html} for full documentation.
}
\keyword{internal}
