% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_create_package}
\alias{opensearchservice_create_package}
\title{Creates a package for use with Amazon OpenSearch Service domains}
\usage{
opensearchservice_create_package(
  PackageName,
  PackageType,
  PackageDescription = NULL,
  PackageSource
)
}
\arguments{
\item{PackageName}{[required] Unique name for the package.}

\item{PackageType}{[required] The type of package.}

\item{PackageDescription}{Description of the package.}

\item{PackageSource}{[required] The Amazon S3 location from which to import the package.}
}
\description{
Creates a package for use with Amazon OpenSearch Service domains. For more information, see \href{https://docs.aws.amazon.com/opensearch-service/latest/developerguide/custom-packages.html}{Custom packages for Amazon OpenSearch Service}.

See \url{https://www.paws-r-sdk.com/docs/opensearchservice_create_package/} for full documentation.
}
\keyword{internal}
