% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_delete_application}
\alias{opensearchservice_delete_application}
\title{Deletes a specified OpenSearch application}
\usage{
opensearchservice_delete_application(id)
}
\arguments{
\item{id}{[required] The unique identifier of the OpenSearch application to delete.}
}
\description{
Deletes a specified OpenSearch application.

See \url{https://www.paws-r-sdk.com/docs/opensearchservice_delete_application/} for full documentation.
}
\keyword{internal}
