% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_delete_domain}
\alias{opensearchservice_delete_domain}
\title{Permanently deletes the specified domain and all of its data}
\usage{
opensearchservice_delete_domain(DomainName)
}
\arguments{
\item{DomainName}{[required] The name of the domain you want to permanently delete.}
}
\description{
Permanently deletes the specified domain and all of its data. Once a domain is deleted, it cannot be recovered.

See \url{https://paws-r.github.io/docs/opensearchservice/delete_domain.html} for full documentation.
}
\keyword{internal}
