% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_delete_outbound_connection}
\alias{opensearchservice_delete_outbound_connection}
\title{Allows the source Amazon OpenSearch Service domain owner to delete an
existing outbound cross-cluster search connection}
\usage{
opensearchservice_delete_outbound_connection(ConnectionId)
}
\arguments{
\item{ConnectionId}{[required] The ID of the outbound connection you want to permanently delete.}
}
\description{
Allows the source Amazon OpenSearch Service domain owner to delete an existing outbound cross-cluster search connection. For more information, see \href{https://docs.aws.amazon.com/opensearch-service/latest/developerguide/cross-cluster-search.html}{Cross-cluster search for Amazon OpenSearch Service}.

See \url{https://www.paws-r-sdk.com/docs/opensearchservice_delete_outbound_connection/} for full documentation.
}
\keyword{internal}
