% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_delete_package}
\alias{opensearchservice_delete_package}
\title{Deletes an Amazon OpenSearch Service package}
\usage{
opensearchservice_delete_package(PackageID)
}
\arguments{
\item{PackageID}{[required] The internal ID of the package you want to delete. Use
\code{\link[=opensearchservice_describe_packages]{describe_packages}} to find this
value.}
}
\description{
Deletes an Amazon OpenSearch Service package. For more information, see \href{https://docs.aws.amazon.com/opensearch-service/latest/developerguide/custom-packages.html}{Custom packages for Amazon OpenSearch Service}.

See \url{https://www.paws-r-sdk.com/docs/opensearchservice_delete_package/} for full documentation.
}
\keyword{internal}
