% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_delete_vpc_endpoint}
\alias{opensearchservice_delete_vpc_endpoint}
\title{Deletes an Amazon OpenSearch Service-managed interface VPC endpoint}
\usage{
opensearchservice_delete_vpc_endpoint(VpcEndpointId)
}
\arguments{
\item{VpcEndpointId}{[required] The unique identifier of the endpoint.}
}
\description{
Deletes an Amazon OpenSearch Service-managed interface VPC endpoint.

See \url{https://paws-r.github.io/docs/opensearchservice/delete_vpc_endpoint.html} for full documentation.
}
\keyword{internal}
