% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_describe_domain}
\alias{opensearchservice_describe_domain}
\title{Describes the domain configuration for the specified Amazon OpenSearch
Service domain, including the domain ID, domain service endpoint, and
domain ARN}
\usage{
opensearchservice_describe_domain(DomainName)
}
\arguments{
\item{DomainName}{[required] The name of the domain that you want information about.}
}
\description{
Describes the domain configuration for the specified Amazon OpenSearch Service domain, including the domain ID, domain service endpoint, and domain ARN.

See \url{https://www.paws-r-sdk.com/docs/opensearchservice_describe_domain/} for full documentation.
}
\keyword{internal}
