% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_describe_domain_config}
\alias{opensearchservice_describe_domain_config}
\title{Returns the configuration of an Amazon OpenSearch Service domain}
\usage{
opensearchservice_describe_domain_config(DomainName)
}
\arguments{
\item{DomainName}{[required] Name of the OpenSearch Service domain configuration that you want to
describe.}
}
\description{
Returns the configuration of an Amazon OpenSearch Service domain.

See \url{https://www.paws-r-sdk.com/docs/opensearchservice_describe_domain_config/} for full documentation.
}
\keyword{internal}
