% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_describe_domain_health}
\alias{opensearchservice_describe_domain_health}
\title{Returns information about domain and node health, the standby
Availability Zone, number of nodes per Availability Zone, and shard
count per node}
\usage{
opensearchservice_describe_domain_health(DomainName)
}
\arguments{
\item{DomainName}{[required] The name of the domain.}
}
\description{
Returns information about domain and node health, the standby Availability Zone, number of nodes per Availability Zone, and shard count per node.

See \url{https://www.paws-r-sdk.com/docs/opensearchservice_describe_domain_health/} for full documentation.
}
\keyword{internal}
