% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_describe_domain_nodes}
\alias{opensearchservice_describe_domain_nodes}
\title{Returns information about domain and nodes, including data nodes, master
nodes, ultrawarm nodes, Availability Zone(s), standby nodes, node
configurations, and node states}
\usage{
opensearchservice_describe_domain_nodes(DomainName)
}
\arguments{
\item{DomainName}{[required] The name of the domain.}
}
\description{
Returns information about domain and nodes, including data nodes, master nodes, ultrawarm nodes, Availability Zone(s), standby nodes, node configurations, and node states.

See \url{https://www.paws-r-sdk.com/docs/opensearchservice_describe_domain_nodes/} for full documentation.
}
\keyword{internal}
