% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_describe_domains}
\alias{opensearchservice_describe_domains}
\title{Returns domain configuration information about the specified Amazon
OpenSearch Service domains}
\usage{
opensearchservice_describe_domains(DomainNames)
}
\arguments{
\item{DomainNames}{[required] Array of OpenSearch Service domain names that you want information
about. If you don't specify any domains, OpenSearch Service returns
information about all domains owned by the account.}
}
\description{
Returns domain configuration information about the specified Amazon OpenSearch Service domains.

See \url{https://www.paws-r-sdk.com/docs/opensearchservice_describe_domains/} for full documentation.
}
\keyword{internal}
