% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_describe_inbound_connections}
\alias{opensearchservice_describe_inbound_connections}
\title{Lists all the inbound cross-cluster connections for a remote domain}
\usage{
opensearchservice_describe_inbound_connections(
  Filters = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{Filters}{A list of filters used to match properties for inbound cross-cluster
connections. Available \code{Filter} values are:
\itemize{
\item connection-id
\item local-domain-info.domain-name
\item local-domain-info.owner-id
\item local-domain-info.region
\item remote-domain-info.domain-name
}}

\item{MaxResults}{Set this value to limit the number of results returned. If not
specified, defaults to 100.}

\item{NextToken}{If more results are available and NextToken is present, make the next
request to the same API with the received NextToken to paginate the
remaining results.}
}
\description{
Lists all the inbound cross-cluster connections for a remote domain.

See \url{https://paws-r.github.io/docs/opensearchservice/describe_inbound_connections.html} for full documentation.
}
\keyword{internal}
