% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_describe_instance_type_limits}
\alias{opensearchservice_describe_instance_type_limits}
\title{Describes the instance count, storage, and master node limits for a
given OpenSearch or Elasticsearch version and instance type}
\usage{
opensearchservice_describe_instance_type_limits(
  DomainName = NULL,
  InstanceType,
  EngineVersion
)
}
\arguments{
\item{DomainName}{The name of the domain. Only specify if you need the limits for an
existing domain.}

\item{InstanceType}{[required] The OpenSearch Service instance type for which you need limit
information.}

\item{EngineVersion}{[required] Version of OpenSearch or Elasticsearch, in the format Elasticsearch_X.Y
or OpenSearch_X.Y. Defaults to the latest version of OpenSearch.}
}
\description{
Describes the instance count, storage, and master node limits for a given OpenSearch or Elasticsearch version and instance type.

See \url{https://www.paws-r-sdk.com/docs/opensearchservice_describe_instance_type_limits/} for full documentation.
}
\keyword{internal}
