% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_dissociate_package}
\alias{opensearchservice_dissociate_package}
\title{Dissociates a package from the Amazon OpenSearch Service domain}
\usage{
opensearchservice_dissociate_package(PackageID, DomainName)
}
\arguments{
\item{PackageID}{[required] The internal ID of the package to associate with a domain. Use
\code{\link[=opensearchservice_describe_packages]{describe_packages}} to find this
value.}

\item{DomainName}{[required] The name of the domain to associate the package with.}
}
\description{
Dissociates a package from the Amazon OpenSearch Service domain.

See \url{https://paws-r.github.io/docs/opensearchservice/dissociate_package.html} for full documentation.
}
\keyword{internal}
