% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_get_application}
\alias{opensearchservice_get_application}
\title{Retrieves the configuration and status of an existing OpenSearch
application}
\usage{
opensearchservice_get_application(id)
}
\arguments{
\item{id}{[required] The unique identifier of the OpenSearch application to retrieve.}
}
\description{
Retrieves the configuration and status of an existing OpenSearch application.

See \url{https://www.paws-r-sdk.com/docs/opensearchservice_get_application/} for full documentation.
}
\keyword{internal}
