% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_get_compatible_versions}
\alias{opensearchservice_get_compatible_versions}
\title{Returns a list of upgrade-compatible versions of
OpenSearch/Elasticsearch}
\usage{
opensearchservice_get_compatible_versions(DomainName = NULL)
}
\arguments{
\item{DomainName}{}
}
\description{
Returns a list of upgrade-compatible versions of OpenSearch/Elasticsearch. You can optionally pass a \code{DomainName} to get all upgrade-compatible versions of OpenSearch/Elasticsearch for that specific domain.

See \url{https://paws-r.github.io/docs/opensearchservice/get_compatible_versions.html} for full documentation.
}
\keyword{internal}
