% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_get_package_version_history}
\alias{opensearchservice_get_package_version_history}
\title{Returns a list of package versions, along with their creation time and
commit message}
\usage{
opensearchservice_get_package_version_history(
  PackageID,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{PackageID}{[required] Returns an audit history of package versions.}

\item{MaxResults}{Limits results to a maximum number of package versions.}

\item{NextToken}{Used for pagination. Only necessary if a previous API call includes a
non-null NextToken value. If provided, returns results for the next
page.}
}
\description{
Returns a list of package versions, along with their creation time and commit message.

See \url{https://paws-r.github.io/docs/opensearchservice/get_package_version_history.html} for full documentation.
}
\keyword{internal}
