% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_get_upgrade_history}
\alias{opensearchservice_get_upgrade_history}
\title{Retrieves the complete history of the last 10 upgrades performed on the
domain}
\usage{
opensearchservice_get_upgrade_history(
  DomainName,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{DomainName}{[required]}

\item{MaxResults}{}

\item{NextToken}{}
}
\description{
Retrieves the complete history of the last 10 upgrades performed on the domain.

See \url{https://paws-r.github.io/docs/opensearchservice/get_upgrade_history.html} for full documentation.
}
\keyword{internal}
