% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_get_upgrade_status}
\alias{opensearchservice_get_upgrade_status}
\title{Retrieves the latest status of the last upgrade or upgrade eligibility
check performed on the domain}
\usage{
opensearchservice_get_upgrade_status(DomainName)
}
\arguments{
\item{DomainName}{[required]}
}
\description{
Retrieves the latest status of the last upgrade or upgrade eligibility check performed on the domain.

See \url{https://paws-r.github.io/docs/opensearchservice/get_upgrade_status.html} for full documentation.
}
\keyword{internal}
