% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_list_applications}
\alias{opensearchservice_list_applications}
\title{Lists all OpenSearch applications under your account}
\usage{
opensearchservice_list_applications(
  nextToken = NULL,
  statuses = NULL,
  maxResults = NULL
)
}
\arguments{
\item{nextToken}{}

\item{statuses}{Filters the list of OpenSearch applications by status. Possible values:
\code{CREATING}, \code{UPDATING}, \code{DELETING}, \code{FAILED}, \code{ACTIVE}, and \code{DELETED}.}

\item{maxResults}{}
}
\description{
Lists all OpenSearch applications under your account.

See \url{https://www.paws-r-sdk.com/docs/opensearchservice_list_applications/} for full documentation.
}
\keyword{internal}
