% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_list_domain_names}
\alias{opensearchservice_list_domain_names}
\title{Returns the names of all domains owned by the current user's account}
\usage{
opensearchservice_list_domain_names(EngineType = NULL)
}
\arguments{
\item{EngineType}{Optional parameter to filter the output by domain engine type.
Acceptable values are 'Elasticsearch' and 'OpenSearch'.}
}
\description{
Returns the names of all domains owned by the current user's account.

See \url{https://paws-r.github.io/docs/opensearchservice/list_domain_names.html} for full documentation.
}
\keyword{internal}
