% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_list_domains_for_package}
\alias{opensearchservice_list_domains_for_package}
\title{Lists all Amazon OpenSearch Service domains associated with the package}
\usage{
opensearchservice_list_domains_for_package(
  PackageID,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{PackageID}{[required] The package for which to list associated domains.}

\item{MaxResults}{Limits the results to a maximum number of domains.}

\item{NextToken}{Used for pagination. Only necessary if a previous API call includes a
non-null NextToken value. If provided, returns results for the next
page.}
}
\description{
Lists all Amazon OpenSearch Service domains associated with the package.

See \url{https://paws-r.github.io/docs/opensearchservice/list_domains_for_package.html} for full documentation.
}
\keyword{internal}
