% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_list_tags}
\alias{opensearchservice_list_tags}
\title{Returns all tags for the given domain}
\usage{
opensearchservice_list_tags(ARN)
}
\arguments{
\item{ARN}{[required] Specify the \code{ARN} of the domain that the tags you want to view are
attached to.}
}
\description{
Returns all tags for the given domain.

See \url{https://paws-r.github.io/docs/opensearchservice/list_tags.html} for full documentation.
}
\keyword{internal}
