% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_list_versions}
\alias{opensearchservice_list_versions}
\title{Lists all versions of OpenSearch and Elasticsearch that Amazon
OpenSearch Service supports}
\usage{
opensearchservice_list_versions(MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{An optional parameter that specifies the maximum number of results to
return. You can use \code{nextToken} to get the next page of results.}

\item{NextToken}{If your initial \code{\link[=opensearchservice_list_versions]{list_versions}}
operation returns a \code{nextToken}, you can include the returned
\code{nextToken} in subsequent
\code{\link[=opensearchservice_list_versions]{list_versions}} operations, which
returns results in the next page.}
}
\description{
Lists all versions of OpenSearch and Elasticsearch that Amazon OpenSearch Service supports.

See \url{https://paws-r.github.io/docs/opensearchservice/list_versions.html} for full documentation.
}
\keyword{internal}
