% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_remove_tags}
\alias{opensearchservice_remove_tags}
\title{Removes the specified set of tags from the given domain}
\usage{
opensearchservice_remove_tags(ARN, TagKeys)
}
\arguments{
\item{ARN}{[required] The \code{ARN} of the domain from which you want to delete the specified
tags.}

\item{TagKeys}{[required] The \code{TagKey} list you want to remove from the domain.}
}
\description{
Removes the specified set of tags from the given domain.

See \url{https://paws-r.github.io/docs/opensearchservice/remove_tags.html} for full documentation.
}
\keyword{internal}
