% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_revoke_vpc_endpoint_access}
\alias{opensearchservice_revoke_vpc_endpoint_access}
\title{Revokes access to an Amazon OpenSearch Service domain that was provided
through an interface VPC endpoint}
\usage{
opensearchservice_revoke_vpc_endpoint_access(
  DomainName,
  Account = NULL,
  Service = NULL
)
}
\arguments{
\item{DomainName}{[required] The name of the OpenSearch Service domain.}

\item{Account}{The account ID to revoke access from.}

\item{Service}{The service SP to revoke access from.}
}
\description{
Revokes access to an Amazon OpenSearch Service domain that was provided through an interface VPC endpoint.

See \url{https://www.paws-r-sdk.com/docs/opensearchservice_revoke_vpc_endpoint_access/} for full documentation.
}
\keyword{internal}
