% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_start_service_software_update}
\alias{opensearchservice_start_service_software_update}
\title{Schedules a service software update for an Amazon OpenSearch Service
domain}
\usage{
opensearchservice_start_service_software_update(DomainName)
}
\arguments{
\item{DomainName}{[required] The name of the domain that you want to update to the latest service
software.}
}
\description{
Schedules a service software update for an Amazon OpenSearch Service domain.

See \url{https://paws-r.github.io/docs/opensearchservice/start_service_software_update.html} for full documentation.
}
\keyword{internal}
