% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_update_application}
\alias{opensearchservice_update_application}
\title{Update the OpenSearch Application}
\usage{
opensearchservice_update_application(id, dataSources = NULL, appConfigs = NULL)
}
\arguments{
\item{id}{[required] Unique identifier of the OpenSearch Application to be updated.}

\item{dataSources}{Data sources to be associated with the OpenSearch Application.}

\item{appConfigs}{Configurations to be changed for the OpenSearch Application.}
}
\description{
Update the OpenSearch Application.

See \url{https://www.paws-r-sdk.com/docs/opensearchservice_update_application/} for full documentation.
}
\keyword{internal}
