% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_update_direct_query_data_source}
\alias{opensearchservice_update_direct_query_data_source}
\title{Updates the configuration or properties of an existing direct query data
source in Amazon OpenSearch Service}
\usage{
opensearchservice_update_direct_query_data_source(
  DataSourceName,
  DataSourceType,
  Description = NULL,
  OpenSearchArns
)
}
\arguments{
\item{DataSourceName}{[required] A unique, user-defined label to identify the data source within your
OpenSearch Service environment.}

\item{DataSourceType}{[required] The supported Amazon Web Services service that you want to use as the
source for direct queries in OpenSearch Service.}

\item{Description}{An optional text field for providing additional context and details
about the data source.}

\item{OpenSearchArns}{[required] A list of Amazon Resource Names (ARNs) for the OpenSearch collections
that are associated with the direct query data source.}
}
\description{
Updates the configuration or properties of an existing direct query data source in Amazon OpenSearch Service.

See \url{https://www.paws-r-sdk.com/docs/opensearchservice_update_direct_query_data_source/} for full documentation.
}
\keyword{internal}
