% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchservice_operations.R
\name{opensearchservice_update_package}
\alias{opensearchservice_update_package}
\title{Updates a package for use with Amazon OpenSearch Service domains}
\usage{
opensearchservice_update_package(
  PackageID,
  PackageSource,
  PackageDescription = NULL,
  CommitMessage = NULL,
  PackageConfiguration = NULL,
  PackageEncryptionOptions = NULL
)
}
\arguments{
\item{PackageID}{[required] The unique identifier for the package.}

\item{PackageSource}{[required] Amazon S3 bucket and key for the package.}

\item{PackageDescription}{A new description of the package.}

\item{CommitMessage}{Commit message for the updated file, which is shown as part of
\code{GetPackageVersionHistoryResponse}.}

\item{PackageConfiguration}{The updated configuration details for a package.}

\item{PackageEncryptionOptions}{Encryption options for a package.}
}
\description{
Updates a package for use with Amazon OpenSearch Service domains. For more information, see \href{https://docs.aws.amazon.com/opensearch-service/latest/developerguide/custom-packages.html}{Custom packages for Amazon OpenSearch Service}.

See \url{https://www.paws-r-sdk.com/docs/opensearchservice_update_package/} for full documentation.
}
\keyword{internal}
