% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchserviceserverless_operations.R
\name{opensearchserviceserverless_create_collection}
\alias{opensearchserviceserverless_create_collection}
\title{Creates a new OpenSearch Serverless collection}
\usage{
opensearchserviceserverless_create_collection(
  clientToken = NULL,
  description = NULL,
  name,
  standbyReplicas = NULL,
  tags = NULL,
  type = NULL
)
}
\arguments{
\item{clientToken}{Unique, case-sensitive identifier to ensure idempotency of the request.}

\item{description}{Description of the collection.}

\item{name}{[required] Name of the collection.}

\item{standbyReplicas}{Indicates whether standby replicas should be used for a collection.}

\item{tags}{An arbitrary set of tags (key–value pairs) to associate with the
OpenSearch Serverless collection.}

\item{type}{The type of collection.}
}
\description{
Creates a new OpenSearch Serverless collection. For more information, see \href{https://docs.aws.amazon.com/opensearch-service/latest/developerguide/serverless-manage.html}{Creating and managing Amazon OpenSearch Serverless collections}.

See \url{https://www.paws-r-sdk.com/docs/opensearchserviceserverless_create_collection/} for full documentation.
}
\keyword{internal}
