% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchserviceserverless_operations.R
\name{opensearchserviceserverless_delete_lifecycle_policy}
\alias{opensearchserviceserverless_delete_lifecycle_policy}
\title{Deletes an OpenSearch Serverless lifecycle policy}
\usage{
opensearchserviceserverless_delete_lifecycle_policy(
  type,
  name,
  clientToken = NULL
)
}
\arguments{
\item{type}{[required] The type of lifecycle policy.}

\item{name}{[required] The name of the policy to delete.}

\item{clientToken}{Unique, case-sensitive identifier to ensure idempotency of the request.}
}
\description{
Deletes an OpenSearch Serverless lifecycle policy. For more information, see \href{https://docs.aws.amazon.com/opensearch-service/latest/developerguide/serverless-lifecycle.html#serverless-lifecycle-delete}{Deleting data lifecycle policies}.

See \url{https://www.paws-r-sdk.com/docs/opensearchserviceserverless_delete_lifecycle_policy/} for full documentation.
}
\keyword{internal}
