% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchserviceserverless_operations.R
\name{opensearchserviceserverless_delete_security_policy}
\alias{opensearchserviceserverless_delete_security_policy}
\title{Deletes an OpenSearch Serverless security policy}
\usage{
opensearchserviceserverless_delete_security_policy(
  clientToken = NULL,
  name,
  type
)
}
\arguments{
\item{clientToken}{Unique, case-sensitive identifier to ensure idempotency of the request.}

\item{name}{[required] The name of the policy to delete.}

\item{type}{[required] The type of policy.}
}
\description{
Deletes an OpenSearch Serverless security policy.

See \url{https://www.paws-r-sdk.com/docs/opensearchserviceserverless_delete_security_policy/} for full documentation.
}
\keyword{internal}
