% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchserviceserverless_operations.R
\name{opensearchserviceserverless_list_access_policies}
\alias{opensearchserviceserverless_list_access_policies}
\title{Returns information about a list of OpenSearch Serverless access
policies}
\usage{
opensearchserviceserverless_list_access_policies(
  maxResults = NULL,
  nextToken = NULL,
  resource = NULL,
  type
)
}
\arguments{
\item{maxResults}{An optional parameter that specifies the maximum number of results to
return. You can use \code{nextToken} to get the next page of results. The
default is 20.}

\item{nextToken}{If your initial
\code{\link[=opensearchserviceserverless_list_access_policies]{list_access_policies}}
operation returns a \code{nextToken}, you can include the returned
\code{nextToken} in subsequent
\code{\link[=opensearchserviceserverless_list_access_policies]{list_access_policies}}
operations, which returns results in the next page.}

\item{resource}{Resource filters (can be collections or indexes) that policies can apply
to.}

\item{type}{[required] The type of access policy.}
}
\description{
Returns information about a list of OpenSearch Serverless access policies.

See \url{https://www.paws-r-sdk.com/docs/opensearchserviceserverless_list_access_policies/} for full documentation.
}
\keyword{internal}
