% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensearchserviceserverless_operations.R
\name{opensearchserviceserverless_update_collection}
\alias{opensearchserviceserverless_update_collection}
\title{Updates an OpenSearch Serverless collection}
\usage{
opensearchserviceserverless_update_collection(
  clientToken = NULL,
  description = NULL,
  id
)
}
\arguments{
\item{clientToken}{Unique, case-sensitive identifier to ensure idempotency of the request.}

\item{description}{A description of the collection.}

\item{id}{[required] The unique identifier of the collection.}
}
\description{
Updates an OpenSearch Serverless collection.

See \url{https://www.paws-r-sdk.com/docs/opensearchserviceserverless_update_collection/} for full documentation.
}
\keyword{internal}
