% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_cancel_ingestion}
\alias{quicksight_cancel_ingestion}
\title{Cancels an ongoing ingestion of data into SPICE}
\usage{
quicksight_cancel_ingestion(AwsAccountId, DataSetId, IngestionId)
}
\arguments{
\item{AwsAccountId}{[required] The AWS account ID.}

\item{DataSetId}{[required] The ID of the dataset used in the ingestion.}

\item{IngestionId}{[required] An ID for the ingestion.}
}
\description{
Cancels an ongoing ingestion of data into SPICE.
}
\section{Request syntax}{
\preformatted{svc$cancel_ingestion(
  AwsAccountId = "string",
  DataSetId = "string",
  IngestionId = "string"
)
}
}

\keyword{internal}
