% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_create_account_customization}
\alias{quicksight_create_account_customization}
\title{Creates Amazon Quick Sight customizations}
\usage{
quicksight_create_account_customization(
  AwsAccountId,
  Namespace = NULL,
  AccountCustomization,
  Tags = NULL
)
}
\arguments{
\item{AwsAccountId}{[required] The ID for the Amazon Web Services account that you want to customize
Quick Sight for.}

\item{Namespace}{The Quick Sight namespace that you want to add customizations to.}

\item{AccountCustomization}{[required] The Quick Sight customizations you're adding. You can add these to an
Amazon Web Services account and a QuickSight namespace.

For example, you can add a default theme by setting
\code{AccountCustomization} to the midnight theme:
\code{"AccountCustomization": { "DefaultTheme": "arn:aws:quicksight::aws:theme/MIDNIGHT" }}.
Or, you can add a custom theme by specifying
\code{"AccountCustomization": { "DefaultTheme": "arn:aws:quicksight:us-west-2:111122223333:theme/bdb844d0-0fe9-4d9d-b520-0fe602d93639" }}.}

\item{Tags}{A list of the tags that you want to attach to this resource.}
}
\description{
Creates Amazon Quick Sight customizations. Currently, you can add a custom default theme by using the \code{\link[=quicksight_create_account_customization]{create_account_customization}} or \code{\link[=quicksight_update_account_customization]{update_account_customization}} API operation. To further customize Amazon Quick Sight by removing Amazon Quick Sight sample assets and videos for all new users, see \href{https://docs.aws.amazon.com/quicksuite/latest/userguide/customizations.html}{Customizing Quick Sight} in the \emph{Amazon Quick Sight User Guide.}

See \url{https://www.paws-r-sdk.com/docs/quicksight_create_account_customization/} for full documentation.
}
\keyword{internal}
