% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_create_folder_membership}
\alias{quicksight_create_folder_membership}
\title{Adds an asset, such as a dashboard, analysis, or dataset into a folder}
\usage{
quicksight_create_folder_membership(
  AwsAccountId,
  FolderId,
  MemberId,
  MemberType
)
}
\arguments{
\item{AwsAccountId}{[required] The ID for the Amazon Web Services account that contains the folder.}

\item{FolderId}{[required] The ID of the folder.}

\item{MemberId}{[required] The ID of the asset (the dashboard, analysis, or dataset).}

\item{MemberType}{[required] The type of the member, including \code{DASHBOARD}, \code{ANALYSIS}, and
\code{DATASET}.}
}
\description{
Adds an asset, such as a dashboard, analysis, or dataset into a folder.

See \url{https://www.paws-r-sdk.com/docs/quicksight_create_folder_membership/} for full documentation.
}
\keyword{internal}
