% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_create_group_membership}
\alias{quicksight_create_group_membership}
\title{Adds an Amazon Quick Sight user to an Amazon Quick Sight group}
\usage{
quicksight_create_group_membership(
  MemberName,
  GroupName,
  AwsAccountId,
  Namespace
)
}
\arguments{
\item{MemberName}{[required] The name of the user that you want to add to the group membership.}

\item{GroupName}{[required] The name of the group that you want to add the user to.}

\item{AwsAccountId}{[required] The ID for the Amazon Web Services account that the group is in.
Currently, you use the ID for the Amazon Web Services account that
contains your Amazon Quick Sight account.}

\item{Namespace}{[required] The namespace that you want the user to be a part of.}
}
\description{
Adds an Amazon Quick Sight user to an Amazon Quick Sight group.

See \url{https://www.paws-r-sdk.com/docs/quicksight_create_group_membership/} for full documentation.
}
\keyword{internal}
