% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_create_refresh_schedule}
\alias{quicksight_create_refresh_schedule}
\title{Creates a refresh schedule for a dataset}
\usage{
quicksight_create_refresh_schedule(DataSetId, AwsAccountId, Schedule)
}
\arguments{
\item{DataSetId}{[required] The ID of the dataset.}

\item{AwsAccountId}{[required] The Amazon Web Services account ID.}

\item{Schedule}{[required] The refresh schedule.}
}
\description{
Creates a refresh schedule for a dataset. You can create up to 5 different schedules for a single dataset.

See \url{https://www.paws-r-sdk.com/docs/quicksight_create_refresh_schedule/} for full documentation.
}
\keyword{internal}
