% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_create_template_alias}
\alias{quicksight_create_template_alias}
\title{Creates a template alias for a template}
\usage{
quicksight_create_template_alias(
  AwsAccountId,
  TemplateId,
  AliasName,
  TemplateVersionNumber
)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the template
that you creating an alias for.}

\item{TemplateId}{[required] An ID for the template.}

\item{AliasName}{[required] The name that you want to give to the template alias that you're
creating. Don't start the alias name with the \code{$} character. Alias names
that start with \code{$} are reserved by Quick Sight.}

\item{TemplateVersionNumber}{[required] The version number of the template.}
}
\description{
Creates a template alias for a template.

See \url{https://www.paws-r-sdk.com/docs/quicksight_create_template_alias/} for full documentation.
}
\keyword{internal}
