% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_create_theme_alias}
\alias{quicksight_create_theme_alias}
\title{Creates a theme alias for a theme}
\usage{
quicksight_create_theme_alias(
  AwsAccountId,
  ThemeId,
  AliasName,
  ThemeVersionNumber
)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the theme for
the new theme alias.}

\item{ThemeId}{[required] An ID for the theme alias.}

\item{AliasName}{[required] The name that you want to give to the theme alias that you are creating.
The alias name can't begin with a \code{$}. Alias names that start with \code{$}
are reserved by Amazon Quick Sight.}

\item{ThemeVersionNumber}{[required] The version number of the theme.}
}
\description{
Creates a theme alias for a theme.

See \url{https://www.paws-r-sdk.com/docs/quicksight_create_theme_alias/} for full documentation.
}
\keyword{internal}
