% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_delete_account_customization}
\alias{quicksight_delete_account_customization}
\title{This API permanently deletes all Quick Sight customizations for the
specified Amazon Web Services account and namespace}
\usage{
quicksight_delete_account_customization(AwsAccountId, Namespace = NULL)
}
\arguments{
\item{AwsAccountId}{[required] The ID for the Amazon Web Services account that you want to delete Quick
Sight customizations from.}

\item{Namespace}{The Quick Sight namespace that you're deleting the customizations from.}
}
\description{
This API permanently deletes all Quick Sight customizations for the specified Amazon Web Services account and namespace. When you delete account customizations:

See \url{https://www.paws-r-sdk.com/docs/quicksight_delete_account_customization/} for full documentation.
}
\keyword{internal}
