% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_delete_dashboard}
\alias{quicksight_delete_dashboard}
\title{Deletes a dashboard}
\usage{
quicksight_delete_dashboard(AwsAccountId, DashboardId, VersionNumber = NULL)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the dashboard
that you're deleting.}

\item{DashboardId}{[required] The ID for the dashboard.}

\item{VersionNumber}{The version number of the dashboard. If the version number property is
provided, only the specified version of the dashboard is deleted.}
}
\description{
Deletes a dashboard.

See \url{https://www.paws-r-sdk.com/docs/quicksight_delete_dashboard/} for full documentation.
}
\keyword{internal}
