% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_delete_data_set}
\alias{quicksight_delete_data_set}
\title{Deletes a dataset}
\usage{
quicksight_delete_data_set(AwsAccountId, DataSetId)
}
\arguments{
\item{AwsAccountId}{[required] The Amazon Web Services account ID.}

\item{DataSetId}{[required] The ID for the dataset that you want to delete. This ID is unique per
Amazon Web Services Region for each Amazon Web Services account.}
}
\description{
Deletes a dataset.

See \url{https://www.paws-r-sdk.com/docs/quicksight_delete_data_set/} for full documentation.
}
\keyword{internal}
