% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_delete_data_source}
\alias{quicksight_delete_data_source}
\title{Deletes the data source permanently}
\usage{
quicksight_delete_data_source(AwsAccountId, DataSourceId)
}
\arguments{
\item{AwsAccountId}{[required] The Amazon Web Services account ID.}

\item{DataSourceId}{[required] The ID of the data source. This ID is unique per Amazon Web Services
Region for each Amazon Web Services account.}
}
\description{
Deletes the data source permanently. This operation breaks all the datasets that reference the deleted data source.

See \url{https://paws-r.github.io/docs/quicksight/delete_data_source.html} for full documentation.
}
\keyword{internal}
