% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_delete_folder}
\alias{quicksight_delete_folder}
\title{Deletes an empty folder}
\usage{
quicksight_delete_folder(AwsAccountId, FolderId)
}
\arguments{
\item{AwsAccountId}{[required] The ID for the Amazon Web Services account that contains the folder.}

\item{FolderId}{[required] The ID of the folder.}
}
\description{
Deletes an empty folder.

See \url{https://paws-r.github.io/docs/quicksight/delete_folder.html} for full documentation.
}
\keyword{internal}
