% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_delete_folder_membership}
\alias{quicksight_delete_folder_membership}
\title{Removes an asset, such as a dashboard, analysis, or dataset, from a
folder}
\usage{
quicksight_delete_folder_membership(
  AwsAccountId,
  FolderId,
  MemberId,
  MemberType
)
}
\arguments{
\item{AwsAccountId}{[required] The ID for the Amazon Web Services account that contains the folder.}

\item{FolderId}{[required] The Folder ID.}

\item{MemberId}{[required] The ID of the asset (the dashboard, analysis, or dataset) that you want
to delete.}

\item{MemberType}{[required] The type of the member, including \code{DASHBOARD}, \code{ANALYSIS}, and \code{DATASET}}
}
\description{
Removes an asset, such as a dashboard, analysis, or dataset, from a folder.

See \url{https://paws-r.github.io/docs/quicksight/delete_folder_membership.html} for full documentation.
}
\keyword{internal}
