% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_delete_iam_policy_assignment}
\alias{quicksight_delete_iam_policy_assignment}
\title{Deletes an existing IAM policy assignment}
\usage{
quicksight_delete_iam_policy_assignment(
  AwsAccountId,
  AssignmentName,
  Namespace
)
}
\arguments{
\item{AwsAccountId}{[required] The Amazon Web Services account ID where you want to delete the IAM
policy assignment.}

\item{AssignmentName}{[required] The name of the assignment.}

\item{Namespace}{[required] The namespace that contains the assignment.}
}
\description{
Deletes an existing IAM policy assignment.

See \url{https://paws-r.github.io/docs/quicksight/delete_iam_policy_assignment.html} for full documentation.
}
\keyword{internal}
