% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_delete_template}
\alias{quicksight_delete_template}
\title{Deletes a template}
\usage{
quicksight_delete_template(AwsAccountId, TemplateId, VersionNumber)
}
\arguments{
\item{AwsAccountId}{[required] AWS account ID that contains the template you are deleting.}

\item{TemplateId}{[required] An ID for the template you want to delete.}

\item{VersionNumber}{Specifies the version of the template that you want to delete. If you
don\'t provide a version number, \code{DeleteTemplate} deletes all versions
of the template.}
}
\description{
Deletes a template.
}
\section{Request syntax}{
\preformatted{svc$delete_template(
  AwsAccountId = "string",
  TemplateId = "string",
  VersionNumber = 123
)
}
}

\keyword{internal}
