% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_delete_template_alias}
\alias{quicksight_delete_template_alias}
\title{Deletes the item that the specified template alias points to}
\usage{
quicksight_delete_template_alias(AwsAccountId, TemplateId, AliasName)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the AWS account that contains the item to delete.}

\item{TemplateId}{[required] The ID for the template that the specified alias is for.}

\item{AliasName}{[required] The name for the template alias. To delete a specific alias, you delete
the version that the alias points to. You can specify the alias name, or
specify the latest version of the template by providing the keyword
\verb{$LATEST} in the \code{AliasName} parameter.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Status = 123,
  TemplateId = "string",
  AliasName = "string",
  Arn = "string",
  RequestId = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Deletes the item that the specified template alias points to. If you
provide a specific alias, you delete the version of the template that
the alias points to.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_template_alias(
  AwsAccountId = "string",
  TemplateId = "string",
  AliasName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
