% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_delete_theme}
\alias{quicksight_delete_theme}
\title{Deletes a theme}
\usage{
quicksight_delete_theme(AwsAccountId, ThemeId, VersionNumber)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the AWS account that contains the theme that you're deleting.}

\item{ThemeId}{[required] An ID for the theme that you want to delete.}

\item{VersionNumber}{The version of the theme that you want to delete.

\strong{Note:} If you don't provide a version number, you're using this call
to \code{DeleteTheme} to delete all versions of the theme.}
}
\description{
Deletes a theme.
}
\section{Request syntax}{
\preformatted{svc$delete_theme(
  AwsAccountId = "string",
  ThemeId = "string",
  VersionNumber = 123
)
}
}

\keyword{internal}
