% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_delete_user_by_principal_id}
\alias{quicksight_delete_user_by_principal_id}
\title{Deletes a user identified by its principal ID}
\usage{
quicksight_delete_user_by_principal_id(PrincipalId, AwsAccountId, Namespace)
}
\arguments{
\item{PrincipalId}{[required] The principal ID of the user.}

\item{AwsAccountId}{[required] The ID for the Amazon Web Services account that the user is in.
Currently, you use the ID for the Amazon Web Services account that
contains your Amazon Quick Sight account.}

\item{Namespace}{[required] The namespace. Currently, you should set this to \code{default}.}
}
\description{
Deletes a user identified by its principal ID.

See \url{https://www.paws-r-sdk.com/docs/quicksight_delete_user_by_principal_id/} for full documentation.
}
\keyword{internal}
