% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_account_settings}
\alias{quicksight_describe_account_settings}
\title{Describes the settings that were used when your Quick Sight subscription
was first created in this Amazon Web Services account}
\usage{
quicksight_describe_account_settings(AwsAccountId)
}
\arguments{
\item{AwsAccountId}{[required] The ID for the Amazon Web Services account that contains the settings
that you want to list.}
}
\description{
Describes the settings that were used when your Quick Sight subscription was first created in this Amazon Web Services account.

See \url{https://www.paws-r-sdk.com/docs/quicksight_describe_account_settings/} for full documentation.
}
\keyword{internal}
