% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_dashboard}
\alias{quicksight_describe_dashboard}
\title{Provides a summary for a dashboard}
\usage{
quicksight_describe_dashboard(
  AwsAccountId,
  DashboardId,
  VersionNumber = NULL,
  AliasName = NULL
)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the dashboard
that you're describing.}

\item{DashboardId}{[required] The ID for the dashboard.}

\item{VersionNumber}{The version number for the dashboard. If a version number isn't passed,
the latest published dashboard version is described.}

\item{AliasName}{The alias name.}
}
\description{
Provides a summary for a dashboard.

See \url{https://paws-r.github.io/docs/quicksight/describe_dashboard.html} for full documentation.
}
\keyword{internal}
