% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_data_source}
\alias{quicksight_describe_data_source}
\title{Describes a data source}
\usage{
quicksight_describe_data_source(AwsAccountId, DataSourceId)
}
\arguments{
\item{AwsAccountId}{[required] The AWS account ID.}

\item{DataSourceId}{[required] The ID of the data source. This ID is unique per AWS Region for each AWS
account.}
}
\description{
Describes a data source.
}
\section{Request syntax}{
\preformatted{svc$describe_data_source(
  AwsAccountId = "string",
  DataSourceId = "string"
)
}
}

\keyword{internal}
