% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_folder_permissions}
\alias{quicksight_describe_folder_permissions}
\title{Describes permissions for a folder}
\usage{
quicksight_describe_folder_permissions(AwsAccountId, FolderId)
}
\arguments{
\item{AwsAccountId}{[required] The ID for the Amazon Web Services account that contains the folder.}

\item{FolderId}{[required] The ID of the folder.}
}
\description{
Describes permissions for a folder.

See \url{https://www.paws-r-sdk.com/docs/quicksight_describe_folder_permissions/} for full documentation.
}
\keyword{internal}
