% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_folder_resolved_permissions}
\alias{quicksight_describe_folder_resolved_permissions}
\title{Describes the folder resolved permissions}
\usage{
quicksight_describe_folder_resolved_permissions(
  AwsAccountId,
  FolderId,
  Namespace = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{AwsAccountId}{[required] The ID for the Amazon Web Services account that contains the folder.}

\item{FolderId}{[required] The ID of the folder.}

\item{Namespace}{The namespace of the folder whose permissions you want described.}

\item{MaxResults}{The maximum number of results to be returned per request.}

\item{NextToken}{A pagination token for the next set of results.}
}
\description{
Describes the folder resolved permissions. Permissions consists of both folder direct permissions and the inherited permissions from the ancestor folders.

See \url{https://www.paws-r-sdk.com/docs/quicksight_describe_folder_resolved_permissions/} for full documentation.
}
\keyword{internal}
