% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_group}
\alias{quicksight_describe_group}
\title{Returns an Amazon QuickSight group's description and Amazon Resource
Name (ARN)}
\usage{
quicksight_describe_group(GroupName, AwsAccountId, Namespace)
}
\arguments{
\item{GroupName}{[required] The name of the group that you want to describe.}

\item{AwsAccountId}{[required] The ID for the Amazon Web Services account that the group is in.
Currently, you use the ID for the Amazon Web Services account that
contains your Amazon QuickSight account.}

\item{Namespace}{[required] The namespace of the group that you want described.}
}
\description{
Returns an Amazon QuickSight group's description and Amazon Resource Name (ARN).

See \url{https://paws-r.github.io/docs/quicksight/describe_group.html} for full documentation.
}
\keyword{internal}
