% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_group}
\alias{quicksight_describe_group}
\title{Returns an Amazon Quick Sight group's description and Amazon Resource
Name (ARN)}
\usage{
quicksight_describe_group(GroupName, AwsAccountId, Namespace)
}
\arguments{
\item{GroupName}{[required] The name of the group that you want to describe.}

\item{AwsAccountId}{[required] The ID for the Amazon Web Services account that the group is in.
Currently, you use the ID for the Amazon Web Services account that
contains your Amazon Quick Sight account.}

\item{Namespace}{[required] The namespace of the group that you want described.}
}
\description{
Returns an Amazon Quick Sight group's description and Amazon Resource Name (ARN).

See \url{https://www.paws-r-sdk.com/docs/quicksight_describe_group/} for full documentation.
}
\keyword{internal}
