% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_ingestion}
\alias{quicksight_describe_ingestion}
\title{Describes a SPICE ingestion}
\usage{
quicksight_describe_ingestion(AwsAccountId, DataSetId, IngestionId)
}
\arguments{
\item{AwsAccountId}{[required] The AWS account ID.}

\item{DataSetId}{[required] The ID of the dataset used in the ingestion.}

\item{IngestionId}{[required] An ID for the ingestion.}
}
\description{
Describes a SPICE ingestion.
}
\section{Request syntax}{
\preformatted{svc$describe_ingestion(
  AwsAccountId = "string",
  DataSetId = "string",
  IngestionId = "string"
)
}
}

\keyword{internal}
