% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_template_permissions}
\alias{quicksight_describe_template_permissions}
\title{Describes read and write permissions on a template}
\usage{
quicksight_describe_template_permissions(AwsAccountId, TemplateId)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the template
that you're describing.}

\item{TemplateId}{[required] The ID for the template.}
}
\description{
Describes read and write permissions on a template.

See \url{https://paws-r.github.io/docs/quicksight/describe_template_permissions.html} for full documentation.
}
\keyword{internal}
