% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_theme_alias}
\alias{quicksight_describe_theme_alias}
\title{Describes the alias for a theme}
\usage{
quicksight_describe_theme_alias(AwsAccountId, ThemeId, AliasName)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the theme alias
that you're describing.}

\item{ThemeId}{[required] The ID for the theme.}

\item{AliasName}{[required] The name of the theme alias that you want to describe.}
}
\description{
Describes the alias for a theme.

See \url{https://www.paws-r-sdk.com/docs/quicksight_describe_theme_alias/} for full documentation.
}
\keyword{internal}
