% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_theme_permissions}
\alias{quicksight_describe_theme_permissions}
\title{Describes the read and write permissions for a theme}
\usage{
quicksight_describe_theme_permissions(AwsAccountId, ThemeId)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the theme that
you're describing.}

\item{ThemeId}{[required] The ID for the theme that you want to describe permissions for.}
}
\description{
Describes the read and write permissions for a theme.

See \url{https://www.paws-r-sdk.com/docs/quicksight_describe_theme_permissions/} for full documentation.
}
\keyword{internal}
