% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_describe_user}
\alias{quicksight_describe_user}
\title{Returns information about a user, given the user name}
\usage{
quicksight_describe_user(UserName, AwsAccountId, Namespace)
}
\arguments{
\item{UserName}{[required] The name of the user that you want to describe.}

\item{AwsAccountId}{[required] The ID for the Amazon Web Services account that the user is in.
Currently, you use the ID for the Amazon Web Services account that
contains your Amazon Quick Sight account.}

\item{Namespace}{[required] The namespace. Currently, you should set this to \code{default}.}
}
\description{
Returns information about a user, given the user name.

See \url{https://www.paws-r-sdk.com/docs/quicksight_describe_user/} for full documentation.
}
\keyword{internal}
