% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_list_analyses}
\alias{quicksight_list_analyses}
\title{Lists Amazon Quick Sight analyses that exist in the specified Amazon Web
Services account}
\usage{
quicksight_list_analyses(AwsAccountId, NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the analyses.}

\item{NextToken}{A pagination token that can be used in a subsequent request.}

\item{MaxResults}{The maximum number of results to return.}
}
\description{
Lists Amazon Quick Sight analyses that exist in the specified Amazon Web Services account.

See \url{https://www.paws-r-sdk.com/docs/quicksight_list_analyses/} for full documentation.
}
\keyword{internal}
