% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_list_dashboard_versions}
\alias{quicksight_list_dashboard_versions}
\title{Lists all the versions of the dashboards in the Quicksight subscription}
\usage{
quicksight_list_dashboard_versions(AwsAccountId, DashboardId, NextToken,
  MaxResults)
}
\arguments{
\item{AwsAccountId}{[required] AWS account ID that contains the dashboard you are listing.}

\item{DashboardId}{[required] The ID for the dashboard.}

\item{NextToken}{The token for the next set of results, or null if there are no more
results.}

\item{MaxResults}{The maximum number of results to be returned per request.}
}
\description{
Lists all the versions of the dashboards in the Quicksight subscription.
}
\section{Request syntax}{
\preformatted{svc$list_dashboard_versions(
  AwsAccountId = "string",
  DashboardId = "string",
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
