% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_list_dashboard_versions}
\alias{quicksight_list_dashboard_versions}
\title{Lists all the versions of the dashboards in the Amazon Quick Sight
subscription}
\usage{
quicksight_list_dashboard_versions(
  AwsAccountId,
  DashboardId,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the dashboard
that you're listing versions for.}

\item{DashboardId}{[required] The ID for the dashboard.}

\item{NextToken}{The token for the next set of results, or null if there are no more
results.}

\item{MaxResults}{The maximum number of results to be returned per request.}
}
\description{
Lists all the versions of the dashboards in the Amazon Quick Sight subscription.

See \url{https://www.paws-r-sdk.com/docs/quicksight_list_dashboard_versions/} for full documentation.
}
\keyword{internal}
