% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_list_dashboards}
\alias{quicksight_list_dashboards}
\title{Lists dashboards in an Amazon Web Services account}
\usage{
quicksight_list_dashboards(AwsAccountId, NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the dashboards
that you're listing.}

\item{NextToken}{The token for the next set of results, or null if there are no more
results.}

\item{MaxResults}{The maximum number of results to be returned per request.}
}
\description{
Lists dashboards in an Amazon Web Services account.

See \url{https://paws-r.github.io/docs/quicksight/list_dashboards.html} for full documentation.
}
\keyword{internal}
