% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_list_data_sets}
\alias{quicksight_list_data_sets}
\title{Lists all of the datasets belonging to the current Amazon Web Services
account in an Amazon Web Services Region}
\usage{
quicksight_list_data_sets(AwsAccountId, NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{AwsAccountId}{[required] The Amazon Web Services account ID.}

\item{NextToken}{The token for the next set of results, or null if there are no more
results.}

\item{MaxResults}{The maximum number of results to be returned per request.}
}
\description{
Lists all of the datasets belonging to the current Amazon Web Services account in an Amazon Web Services Region.

See \url{https://paws-r.github.io/docs/quicksight/list_data_sets.html} for full documentation.
}
\keyword{internal}
