% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_list_data_sets}
\alias{quicksight_list_data_sets}
\title{Lists all of the datasets belonging to the current AWS account in an AWS
Region}
\usage{
quicksight_list_data_sets(AwsAccountId, NextToken, MaxResults)
}
\arguments{
\item{AwsAccountId}{[required] The AWS account ID.}

\item{NextToken}{The token for the next set of results, or null if there are no more
results.}

\item{MaxResults}{The maximum number of results to be returned per request.}
}
\description{
Lists all of the datasets belonging to the current AWS account in an AWS
Region.

The permissions resource is
\verb{arn:aws:quicksight:region:aws-account-id:dataset/*}.
}
\section{Request syntax}{
\preformatted{svc$list_data_sets(
  AwsAccountId = "string",
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
