% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_list_group_memberships}
\alias{quicksight_list_group_memberships}
\title{Lists member users in a group}
\usage{
quicksight_list_group_memberships(
  GroupName,
  NextToken = NULL,
  MaxResults = NULL,
  AwsAccountId,
  Namespace
)
}
\arguments{
\item{GroupName}{[required] The name of the group that you want to see a membership list of.}

\item{NextToken}{A pagination token that can be used in a subsequent request.}

\item{MaxResults}{The maximum number of results to return from this request.}

\item{AwsAccountId}{[required] The ID for the Amazon Web Services account that the group is in.
Currently, you use the ID for the Amazon Web Services account that
contains your Amazon Quick Sight account.}

\item{Namespace}{[required] The namespace of the group that you want a list of users from.}
}
\description{
Lists member users in a group.

See \url{https://www.paws-r-sdk.com/docs/quicksight_list_group_memberships/} for full documentation.
}
\keyword{internal}
