% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_list_groups}
\alias{quicksight_list_groups}
\title{Lists all user groups in Amazon Quick Sight}
\usage{
quicksight_list_groups(
  AwsAccountId,
  NextToken = NULL,
  MaxResults = NULL,
  Namespace
)
}
\arguments{
\item{AwsAccountId}{[required] The ID for the Amazon Web Services account that the group is in.
Currently, you use the ID for the Amazon Web Services account that
contains your Amazon Quick Sight account.}

\item{NextToken}{A pagination token that can be used in a subsequent request.}

\item{MaxResults}{The maximum number of results to return.}

\item{Namespace}{[required] The namespace that you want a list of groups from.}
}
\description{
Lists all user groups in Amazon Quick Sight.

See \url{https://www.paws-r-sdk.com/docs/quicksight_list_groups/} for full documentation.
}
\keyword{internal}
