% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_list_iam_policy_assignments}
\alias{quicksight_list_iam_policy_assignments}
\title{Lists the IAM policy assignments in the current Amazon Quick Sight
account}
\usage{
quicksight_list_iam_policy_assignments(
  AwsAccountId,
  AssignmentStatus = NULL,
  Namespace,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains these IAM policy
assignments.}

\item{AssignmentStatus}{The status of the assignments.}

\item{Namespace}{[required] The namespace for the assignments.}

\item{NextToken}{The token for the next set of results, or null if there are no more
results.}

\item{MaxResults}{The maximum number of results to be returned per request.}
}
\description{
Lists the IAM policy assignments in the current Amazon Quick Sight account.

See \url{https://www.paws-r-sdk.com/docs/quicksight_list_iam_policy_assignments/} for full documentation.
}
\keyword{internal}
