% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_list_ingestions}
\alias{quicksight_list_ingestions}
\title{Lists the history of SPICE ingestions for a dataset}
\usage{
quicksight_list_ingestions(
  DataSetId,
  NextToken = NULL,
  AwsAccountId,
  MaxResults = NULL
)
}
\arguments{
\item{DataSetId}{[required] The ID of the dataset used in the ingestion.}

\item{NextToken}{The token for the next set of results, or null if there are no more
results.}

\item{AwsAccountId}{[required] The Amazon Web Services account ID.}

\item{MaxResults}{The maximum number of results to be returned per request.}
}
\description{
Lists the history of SPICE ingestions for a dataset. Limited to 5 TPS per user and 25 TPS per account.

See \url{https://www.paws-r-sdk.com/docs/quicksight_list_ingestions/} for full documentation.
}
\keyword{internal}
