% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_list_tags_for_resource}
\alias{quicksight_list_tags_for_resource}
\title{Lists the tags assigned to a resource}
\usage{
quicksight_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource that you want a list of
tags for.}
}
\description{
Lists the tags assigned to a resource.

See \url{https://www.paws-r-sdk.com/docs/quicksight_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
