% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_list_template_aliases}
\alias{quicksight_list_template_aliases}
\title{Lists all the aliases of a template}
\usage{
quicksight_list_template_aliases(AwsAccountId, TemplateId, NextToken,
  MaxResults)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the AWS account that contains the template aliases that you're
listing.}

\item{TemplateId}{[required] The ID for the template.}

\item{NextToken}{The token for the next set of results, or null if there are no more
results.}

\item{MaxResults}{The maximum number of results to be returned per request.}
}
\description{
Lists all the aliases of a template.
}
\section{Request syntax}{
\preformatted{svc$list_template_aliases(
  AwsAccountId = "string",
  TemplateId = "string",
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
