% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_list_template_versions}
\alias{quicksight_list_template_versions}
\title{Lists all the versions of the templates in the current Amazon Quick
Sight account}
\usage{
quicksight_list_template_versions(
  AwsAccountId,
  TemplateId,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the templates
that you're listing.}

\item{TemplateId}{[required] The ID for the template.}

\item{NextToken}{The token for the next set of results, or null if there are no more
results.}

\item{MaxResults}{The maximum number of results to be returned per request.}
}
\description{
Lists all the versions of the templates in the current Amazon Quick Sight account.

See \url{https://www.paws-r-sdk.com/docs/quicksight_list_template_versions/} for full documentation.
}
\keyword{internal}
