% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_restore_analysis}
\alias{quicksight_restore_analysis}
\title{Restores an analysis}
\usage{
quicksight_restore_analysis(AwsAccountId, AnalysisId)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the AWS account that contains the analysis.}

\item{AnalysisId}{[required] The ID of the analysis that you're restoring.}
}
\description{
Restores an analysis.
}
\section{Request syntax}{
\preformatted{svc$restore_analysis(
  AwsAccountId = "string",
  AnalysisId = "string"
)
}
}

\keyword{internal}
