% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_search_dashboards}
\alias{quicksight_search_dashboards}
\title{Searches for dashboards that belong to a user}
\usage{
quicksight_search_dashboards(
  AwsAccountId,
  Filters,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the user whose
dashboards you're searching for.}

\item{Filters}{[required] The filters to apply to the search. Currently, you can search only by
user name, for example,
\verb{"Filters": [ \{ "Name": "QUICKSIGHT_USER", "Operator": "StringEquals", "Value": "arn:aws:quicksight:us-east-1:1:user/default/UserName1" \} ]}}

\item{NextToken}{The token for the next set of results, or null if there are no more
results.}

\item{MaxResults}{The maximum number of results to be returned per request.}
}
\description{
Searches for dashboards that belong to a user.

See \url{https://www.paws-r-sdk.com/docs/quicksight_search_dashboards/} for full documentation.
}
\keyword{internal}
