% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_untag_resource}
\alias{quicksight_untag_resource}
\title{Removes a tag or tags from a resource}
\usage{
quicksight_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource that you want to untag.}

\item{TagKeys}{[required] The keys of the key-value pairs for the resource tag or tags assigned to
the resource.}
}
\description{
Removes a tag or tags from a resource.

See \url{https://www.paws-r-sdk.com/docs/quicksight_untag_resource/} for full documentation.
}
\keyword{internal}
