% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_update_account_customization}
\alias{quicksight_update_account_customization}
\title{Updates Amazon Quick Sight customizations}
\usage{
quicksight_update_account_customization(
  AwsAccountId,
  Namespace = NULL,
  AccountCustomization
)
}
\arguments{
\item{AwsAccountId}{[required] The ID for the Amazon Web Services account that you want to update Quick
Sight customizations for.}

\item{Namespace}{The namespace that you want to update Quick Sight customizations for.}

\item{AccountCustomization}{[required] The Quick Sight customizations you're updating.}
}
\description{
Updates Amazon Quick Sight customizations. Currently, the only customization that you can use is a theme.

See \url{https://www.paws-r-sdk.com/docs/quicksight_update_account_customization/} for full documentation.
}
\keyword{internal}
