% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_update_dashboard_permissions}
\alias{quicksight_update_dashboard_permissions}
\title{Updates read and write permissions on a dashboard}
\usage{
quicksight_update_dashboard_permissions(
  AwsAccountId,
  DashboardId,
  GrantPermissions = NULL,
  RevokePermissions = NULL,
  GrantLinkPermissions = NULL,
  RevokeLinkPermissions = NULL
)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the dashboard
whose permissions you're updating.}

\item{DashboardId}{[required] The ID for the dashboard.}

\item{GrantPermissions}{The permissions that you want to grant on this resource.}

\item{RevokePermissions}{The permissions that you want to revoke from this resource.}

\item{GrantLinkPermissions}{Grants link permissions to all users in a defined namespace.}

\item{RevokeLinkPermissions}{Revokes link permissions from all users in a defined namespace.}
}
\description{
Updates read and write permissions on a dashboard.

See \url{https://www.paws-r-sdk.com/docs/quicksight_update_dashboard_permissions/} for full documentation.
}
\keyword{internal}
