% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_update_data_source_permissions}
\alias{quicksight_update_data_source_permissions}
\title{Updates the permissions to a data source}
\usage{
quicksight_update_data_source_permissions(
  AwsAccountId,
  DataSourceId,
  GrantPermissions = NULL,
  RevokePermissions = NULL
)
}
\arguments{
\item{AwsAccountId}{[required] The Amazon Web Services account ID.}

\item{DataSourceId}{[required] The ID of the data source. This ID is unique per Amazon Web Services
Region for each Amazon Web Services account.}

\item{GrantPermissions}{A list of resource permissions that you want to grant on the data
source.}

\item{RevokePermissions}{A list of resource permissions that you want to revoke on the data
source.}
}
\description{
Updates the permissions to a data source.

See \url{https://www.paws-r-sdk.com/docs/quicksight_update_data_source_permissions/} for full documentation.
}
\keyword{internal}
