% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_update_folder_permissions}
\alias{quicksight_update_folder_permissions}
\title{Updates permissions of a folder}
\usage{
quicksight_update_folder_permissions(
  AwsAccountId,
  FolderId,
  GrantPermissions = NULL,
  RevokePermissions = NULL
)
}
\arguments{
\item{AwsAccountId}{[required] The ID for the Amazon Web Services account that contains the folder to
update.}

\item{FolderId}{[required] The ID of the folder.}

\item{GrantPermissions}{The permissions that you want to grant on a resource.}

\item{RevokePermissions}{The permissions that you want to revoke from a resource.}
}
\description{
Updates permissions of a folder.

See \url{https://paws-r.github.io/docs/quicksight/update_folder_permissions.html} for full documentation.
}
\keyword{internal}
