% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_update_group}
\alias{quicksight_update_group}
\title{Changes a group description}
\usage{
quicksight_update_group(GroupName, Description = NULL, AwsAccountId, Namespace)
}
\arguments{
\item{GroupName}{[required] The name of the group that you want to update.}

\item{Description}{The description for the group that you want to update.}

\item{AwsAccountId}{[required] The ID for the Amazon Web Services account that the group is in.
Currently, you use the ID for the Amazon Web Services account that
contains your Amazon Quick Sight account.}

\item{Namespace}{[required] The namespace of the group that you want to update.}
}
\description{
Changes a group description.

See \url{https://www.paws-r-sdk.com/docs/quicksight_update_group/} for full documentation.
}
\keyword{internal}
