% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_update_ip_restriction}
\alias{quicksight_update_ip_restriction}
\title{Updates the content and status of IP rules}
\usage{
quicksight_update_ip_restriction(
  AwsAccountId,
  IpRestrictionRuleMap = NULL,
  Enabled = NULL
)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the IP rules.}

\item{IpRestrictionRuleMap}{A map that describes the updated IP rules with CIDR ranges and
descriptions.}

\item{Enabled}{A value that specifies whether IP rules are turned on.}
}
\description{
Updates the content and status of IP rules. To use this operation, you need to provide the entire map of rules. You can use the \code{\link[=quicksight_describe_ip_restriction]{describe_ip_restriction}} operation to get the current rule map.

See \url{https://paws-r.github.io/docs/quicksight/update_ip_restriction.html} for full documentation.
}
\keyword{internal}
