% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_update_public_sharing_settings}
\alias{quicksight_update_public_sharing_settings}
\title{This API controls public sharing settings for your entire Quick Sight
account, affecting data security and access}
\usage{
quicksight_update_public_sharing_settings(
  AwsAccountId,
  PublicSharingEnabled = NULL
)
}
\arguments{
\item{AwsAccountId}{[required] The Amazon Web Services account ID associated with your Amazon Quick
Sight subscription.}

\item{PublicSharingEnabled}{A Boolean value that indicates whether public sharing is turned on for
an Quick Suite account.}
}
\description{
This API controls public sharing settings for your entire Quick Sight account, affecting data security and access. When you enable public sharing:

See \url{https://www.paws-r-sdk.com/docs/quicksight_update_public_sharing_settings/} for full documentation.
}
\keyword{internal}
