% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_update_template}
\alias{quicksight_update_template}
\title{Updates a template from an existing QuickSight analysis}
\usage{
quicksight_update_template(AwsAccountId, TemplateId, SourceEntity,
  VersionDescription, Name)
}
\arguments{
\item{AwsAccountId}{[required] AWS account ID that contains the template you are updating.}

\item{TemplateId}{[required] The ID for the template.}

\item{SourceEntity}{[required] The source QuickSight entity from which this template is being created.
Templates can be currently created from an Analysis or another template.}

\item{VersionDescription}{A description of the current template version being updated. Every time
you cal \code{UpdateTemplate} you create a new version. Each version of the
template maintains a description of the version in the
\code{VersionDescription} field.}

\item{Name}{The name for the template.}
}
\description{
Updates a template from an existing QuickSight analysis.
}
\section{Request syntax}{
\preformatted{svc$update_template(
  AwsAccountId = "string",
  TemplateId = "string",
  SourceEntity = list(
    SourceAnalysis = list(
      Arn = "string",
      DataSetReferences = list(
        list(
          DataSetPlaceholder = "string",
          DataSetArn = "string"
        )
      )
    ),
    SourceTemplate = list(
      Arn = "string"
    )
  ),
  VersionDescription = "string",
  Name = "string"
)
}
}

\keyword{internal}
