% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_update_template_permissions}
\alias{quicksight_update_template_permissions}
\title{Updates the resource permissions for a template}
\usage{
quicksight_update_template_permissions(
  AwsAccountId,
  TemplateId,
  GrantPermissions = NULL,
  RevokePermissions = NULL
)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the template.}

\item{TemplateId}{[required] The ID for the template.}

\item{GrantPermissions}{A list of resource permissions to be granted on the template.}

\item{RevokePermissions}{A list of resource permissions to be revoked from the template.}
}
\description{
Updates the resource permissions for a template.

See \url{https://www.paws-r-sdk.com/docs/quicksight_update_template_permissions/} for full documentation.
}
\keyword{internal}
