% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_update_template_permissions}
\alias{quicksight_update_template_permissions}
\title{Updates the resource permissions for a template}
\usage{
quicksight_update_template_permissions(AwsAccountId, TemplateId,
  GrantPermissions, RevokePermissions)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the AWS account that contains the template.}

\item{TemplateId}{[required] The ID for the template.}

\item{GrantPermissions}{A list of resource permissions to be granted on the template.}

\item{RevokePermissions}{A list of resource permissions to be revoked from the template.}
}
\description{
Updates the resource permissions for a template.
}
\section{Request syntax}{
\preformatted{svc$update_template_permissions(
  AwsAccountId = "string",
  TemplateId = "string",
  GrantPermissions = list(
    list(
      Principal = "string",
      Actions = list(
        "string"
      )
    )
  ),
  RevokePermissions = list(
    list(
      Principal = "string",
      Actions = list(
        "string"
      )
    )
  )
)
}
}

\keyword{internal}
