% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksight_operations.R
\name{quicksight_update_topic}
\alias{quicksight_update_topic}
\title{Updates a topic}
\usage{
quicksight_update_topic(
  AwsAccountId,
  TopicId,
  Topic,
  CustomInstructions = NULL
)
}
\arguments{
\item{AwsAccountId}{[required] The ID of the Amazon Web Services account that contains the topic that
you want to update.}

\item{TopicId}{[required] The ID of the topic that you want to modify. This ID is unique per
Amazon Web Services Region for each Amazon Web Services account.}

\item{Topic}{[required] The definition of the topic that you want to update.}

\item{CustomInstructions}{Custom instructions for the topic.}
}
\description{
Updates a topic.

See \url{https://www.paws-r-sdk.com/docs/quicksight_update_topic/} for full documentation.
}
\keyword{internal}
